### 2026-04-09 #10387 @editorialbot
[PRE REVIEW]: StatsPAI: A Unified, Agent-Native Python Toolkit for Causal Inference and Applied Econometrics
**Submitting author:** @brycewang-stanford (<a href="http://orcid.org/0000-0002-1828-2208">Biaoyue Wang</a>)
**Repository:** https://github.com/brycewang-stanford/StatsPAI
**Branch with paper.md** (empty if default branch):
**Version:** v 0.6.0
