### 2026-04-10 3a3bebd
Add JOSS paper status badge to README
Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
