### 2026-04-10 5bec7ca
DeepIV: add optional unbiased paired-sample gradient + document trade-offs
Plan A — Add `n_gradient_samples` parameter (default 0, matching
Microsoft EconML's DeepIVEstimator behaviour). When set to >= 1, Stage 2
draws two independent sample sets per gradient step and uses the
Hartford et al. (2017) Section 3.2 paired-sample estimator

    E[(y - h(p,x)) * (y - h(p',x))] = (y - E[h(p,x)])^2

which is an unbiased estimator of both the loss and its gradient.

Plan B — Document the single-sample vs paired-sample trade-off,
clarify that the current default matches EconML (not a bug), add
"When to 
[…]
