+++ src/statspai/deepiv/deep_iv.py
@@ -10,14 +10,64 @@ The approach allows arbitrary non-linear treatment effects and handles
 continuous endogenous regressors without parametric assumptions on the
 first stage.
 
+Implementation notes
+--------------------
+This implementation follows the same defaults as Microsoft's EconML
+``DeepIVEstimator``: the Stage-2 gradient is computed from a single set
+of MC samples (``n_gradient_samples=0``). Per Hartford et al. Section
+3.2, the resulting gradient is a biased estimator of the true loss
+gradient, because the treatment sample appears in both the residual and
+the gradient path. As discussed in the EconML docs, the single-sample
+version *"is not guaranteed to lead to consistent estimates, but has the
+advantage of requiring only a single set of samples from the
+distribution, and can be interpreted as regularizing the loss with a
+variance penalty"*.
+
+To enable the **unbiased paired-sample gradient** described in the
+paper, set ``n_gradient_samples > 0``. This draws two independent sets
+of MC samples per gradient step and uses the identity
+``E[(y - h(p, x))(y - h(p', x))] = (y - E[h(p, x)])^2`` for ``p ⊥ p'``.
+
+When to use DeepIV
+------------------
+- Low-to-moderate dimensional X, continuous treatment, reasonably
+  strong instruments.
+- When you want a flexible non-parametric first stage but don't need
+  theoretical convergence guarantees.
+
+When NOT to use DeepIV
+----------------------
+- Weak instruments: the two-stage procedure compounds instability.
+- High-dimensional X with complex h(T, X): consider DeepGMM
+  (Bennett et al. 2019), DFIV (Xu et al. 2021) or DualIV (Muandet et
+  al. 2020), which are generally more stable in these settings.
+- When you need rigorous model selection / convergence guarantees.
+  See RegDeepIV (Dikkala et al. 2024) for a theoretically grounded
+  variant.
+
 Requires PyTorch (optional dependency):
     pip install statspai[deepiv]
 
 References
 ----------
 Hartford, J., Lewis, G., Leyton-Brown, K., & Taddy, M. (2017).
-"Deep IV: A Flexible Approach for Counterfactual Prediction."
-Proceedings of the 34th International Conference on Machine Learning.
+    "Deep IV: A Flexible Approach for Counterfactual Prediction."
+    Proceedings of the 34th International Conference on Machine Learning.
+
+Microsoft EconML. ``DeepIVEstimator`` — reference implementation whose
+    ``n_gradient_samples`` default of 0 this package follows.
+    https://github.com/py-why/EconML
+
+Bennett, A., Kallus, N., & Schnabel, T. (2019).
+    "Deep Generalized Method of Moments for Instrumental Variable
+    Analysis." NeurIPS 2019.
+
+Xu, L., Chen, Y., Srinivasan, S., de Freitas, N., Doucet, A., & Gretton,
+    A. (2021). "Learning Deep Features in Instrumental Variable
+    Regression." ICLR 2021.
+
+Muandet, K., Mehrjou, A., Lee, S. K., & Raj, A. (2020).
+    "Dual Instrumental Variable Regression." NeurIPS 2020.
 """
 
 from typing import Optional, List, Dict, Any, Tuple
@@ -44,6 +94,7 @@ def deepiv(
     first_stage_epochs: int = 100,
     second_stage_epochs: int = 100,
     n_samples: int = 1,
+    n_gradient_samples: int = 0,
     batch_size: int = 256,
     learning_rate: float = 1e-3,
     alpha: float = 0.05,
@@ -74,7 +125,19 @@ def deepiv(
     second_stage_epochs : int, default 100
         Training epochs for the response model.
     n_samples : int, default 1
[…]
