### 2026-04-13 83d4f77
Add numerical validation tests and humanize error messages
P0-1: 23 validation tests covering OLS, IV, Panel FE, DID, RD,
Matching, and DML — each verifies StatsPAI results against
statsmodels/linearmodels (Stata/R proxies) to 4+ decimal places,
plus cross-estimator consistency checks.

P0-2: Add clear, Stata-quality error messages to regress() and did():
- Bad column names → "Variable(s) not found: [...]. Available: ..."
- Non-DataFrame input → "must be a pandas DataFrame, got str"
- Empty DataFrame → "no observations to regress"
- All-NaN outcome → "entirely NaN — cannot estimate"
Previously
[…]
