+++ tests/test_validation_vs_stata_r.py
@@ -1,0 +1,361 @@
+"""
+Numerical validation: StatsPAI vs statsmodels / linearmodels (Stata/R proxies).
+
+These tests verify that StatsPAI's core estimators produce numerically
+identical results (to 4+ decimal places) to established backends. Since
+StatsPAI wraps statsmodels / linearmodels internally, exact agreement is
+expected — these tests guard against regressions from API-layer
+transformations, formula parsing, or output formatting that silently
+corrupt values.
+
+Each test uses a fixed random seed and a known DGP with analytically
+predictable parameters, so failures pinpoint coefficient-level deviations.
+
+Benchmark references:
+  - OLS / robust SE: statsmodels OLS (matches Stata `reg, r` and R `lm()`)
+  - IV / 2SLS: linearmodels IV2SLS (matches Stata `ivregress 2sls`)
+  - Panel FE: linearmodels PanelOLS (matches Stata `xtreg, fe`)
+  - DID: known DGP with exact ATT = 3.0
+  - RD: known DGP with exact jump = 2.0
+  - Matching: known DGP with true ATT = 2.0
+"""
+
+import sys
+import pytest
+import numpy as np
+import pandas as pd
+
+sys.path.insert(0, 'src')
+import statspai as sp
+
+
+# =====================================================================
+# Helpers
+# =====================================================================
+
+def _ols_data(seed=42, n=10000):
+    """DGP: y = 2 + 3*x1 - 1.5*x2 + N(0, 0.25)."""
+    np.random.seed(seed)
+    x1 = np.random.randn(n)
+    x2 = np.random.randn(n)
+    e = np.random.randn(n) * 0.5
+    y = 2.0 + 3.0 * x1 - 1.5 * x2 + e
+    return pd.DataFrame({'y': y, 'x1': x1, 'x2': x2})
+
+
+def _iv_data(seed=1, n=5000):
+    """DGP: y = 1 + 2.5*x + u, x = 0.8*z + 0.3*u + noise."""
+    np.random.seed(seed)
+    z = np.random.randn(n)
+    u = np.random.randn(n)
+    x = 0.8 * z + 0.3 * u + 0.2 * np.random.randn(n)
+    y = 1.0 + 2.5 * x + u + 0.3 * np.random.randn(n)
+    return pd.DataFrame({'y': y, 'x': x, 'z': z})
+
+
+def _panel_data(seed=2, n_id=200, n_t=10):
+    """DGP: y = 1.5*x + fe_i + N(0, 0.09)."""
+    np.random.seed(seed)
+    n_obs = n_id * n_t
+    fe = np.repeat(np.random.randn(n_id), n_t)
+    x = np.random.randn(n_obs)
+    y = 1.5 * x + fe + np.random.randn(n_obs) * 0.3
+    return pd.DataFrame({
+        'id': np.repeat(range(n_id), n_t),
+        'time': np.tile(range(n_t), n_id),
+        'x': x, 'y': y,
+    })
+
+
+def _did_data(seed=0, n=2000):
+    """DGP: ATT = 3.0 exactly."""
+    np.random.seed(seed)
+    n_units = n // 4
+    df = pd.DataFrame({
+        'id': np.repeat(range(n_units), 4),
+        'time': np.tile(range(4), n_units),
+        'treat': np.repeat(np.random.binomial(1, 0.5, n_units), 4),
+    })
+    df['post'] = (df['time'] >= 2).astype(int)
+    df['y'] = (1 + 2 * df['treat'] + 0.5 * df['time']
+               + 3.0 * df['treat'] * df['post']
[…]
