### 2026-04-13 debacb8
Implement predict() for OLS results and fix outreg2 format auto-detection
P2-1: result.predict() now works for both in-sample (returns
fitted values) and out-of-sample (X @ params) prediction. Previously
it raised NotImplementedError.

P2-2: outreg2() default format changed from 'excel' to 'auto',
which detects the output format from the filename extension (.xlsx,
.docx, .tex). Previously .docx and .tex files were silently exported
as .xlsx with an appended extension.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
