### 2026-04-13 afeea5d
Add 'State of the field' and 'Research impact statement' sections to JOSS paper
Address editorialbot paper check feedback: add the two missing required
sections for JOSS submission compliance.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
