### 2026-04-13 #10387 @brycewang-stanford
Thanks for the paper check feedback. I've added the two missing sections:

- **State of the field** — reviews existing Python packages (DoWhy, EconML, CausalML), R ecosystem fragmentation, and Stata's limitations, clarifying the gap StatsPAI fills.
- **Research impact statement** — describes how StatsPAI lowers barriers for applied researchers, enables AI-assisted empirical workflows, and has been adopted in Stanford REAP research projects.

The updated `paper.md` has been pushed. Could you please regenerate the PDF?
