### 2026-04-13 5d2cd37
Fix critical issues found in code review
Critical:
- Formula tokenizer now uses regex to strip function calls (C(), I(),
  np.log()) before checking column existence — previously raised
  false-positive ValueError on valid statsmodels syntax
- predict() gives clear error for models with formula transforms
  (interactions, categoricals) instead of cryptic KeyError

High:
- DID validation now checks weights column existence
- outreg2 docstring updated to match new default format='auto'

Medium:
- DML tests use shared fixture (no duplicate DGP, runs DML once)
- DID vs regression cross-check uses 
[…]
