### 2026-04-13 4b45fdd
Add verified DOIs to 11 bibliography entries
- 4 publisher DOIs via CrossRef: Spirtes (MIT Press), Hirano & Imbens
  (Wiley), Imbens & Rubin (Cambridge UP), Angrist & Pischke (Princeton UP)
- 6 arXiv DOIs (verified resolvable): NOTEARS, TARNet, DragonNet,
  DeepIV, DoWhy, CausalML
- Confirmed Angrist book DOI is 10.1515/9781400829828 (not the book
  review DOI suggested by editorialbot)
- 2 entries confirmed to have no DOI: Manski (1990, predates DOI system),
  EconML (software repo, no publication)

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
