+++ paper.bib
@@ -144,7 +144,8 @@
   author={Zheng, Xun and Aragam, Bryon and Ravikumar, Pradeep K. and Xing, Eric P.},
   booktitle={Advances in Neural Information Processing Systems},
   volume={31},
-  year={2018}
+  year={2018},
+  doi={10.48550/arXiv.1803.01422}
 }
 
 @article{hahn2020bayesian,
@@ -188,7 +189,8 @@
   author={Shalit, Uri and Johansson, Fredrik D. and Sontag, David},
   booktitle={Proceedings of the 34th International Conference on Machine Learning},
   pages={3076--3085},
-  year={2017}
+  year={2017},
+  doi={10.48550/arXiv.1606.03976}
 }
 
 @inproceedings{shi2019adapting,
@@ -196,7 +198,8 @@
   author={Shi, Claudia and Blei, David and Veitch, Victor},
   booktitle={Advances in Neural Information Processing Systems},
   volume={32},
-  year={2019}
+  year={2019},
+  doi={10.48550/arXiv.1906.02120}
 }
 
 @book{spirtes2000causation,
@@ -204,7 +207,8 @@
   author={Spirtes, Peter and Glymour, Clark and Scheines, Richard},
   year={2000},
   edition={2nd},
-  publisher={MIT Press}
+  publisher={MIT Press},
+  doi={10.7551/mitpress/1754.001.0001}
 }
 
 @article{lei2021conformal,
@@ -225,7 +229,8 @@
   booktitle={Applied Bayesian Modeling and Causal Inference from Incomplete-Data Perspectives},
   pages={73--84},
   year={2004},
-  publisher={Wiley}
+  publisher={Wiley},
+  doi={10.1002/0470090456.ch7}
 }
 
 @article{cattaneo2010efficient,
@@ -422,7 +427,8 @@
   author={Hartford, Jason and Lewis, Greg and Leyton-Brown, Kevin and Taddy, Matt},
   booktitle={Proceedings of the 34th International Conference on Machine Learning},
   pages={1414--1423},
-  year={2017}
+  year={2017},
+  doi={10.48550/arXiv.1612.09596}
 }
 
 @article{kennedy2023towards,
@@ -454,7 +460,8 @@
   author={Angrist, Joshua D. and Pischke, J{\"o}rn-Steffen},
   year={2009},
   publisher={Princeton University Press},
-  isbn={978-0-691-12035-5}
+  isbn={978-0-691-12035-5},
+  doi={10.1515/9781400829828}
 }
 
 @book{imbens2015causal,
@@ -462,14 +469,16 @@
   author={Imbens, Guido W. and Rubin, Donald B.},
   year={2015},
   publisher={Cambridge University Press},
-  isbn={978-0-521-88588-1}
+  isbn={978-0-521-88588-1},
+  doi={10.1017/CBO9781139025751}
 }
 
 @article{sharma2020dowhy,
   title={{DoWhy}: An end-to-end library for causal inference},
   author={Sharma, Amit and Kiciman, Emre},
   journal={arXiv preprint arXiv:2011.04216},
-  year={2020}
+  year={2020},
+  doi={10.48550/arXiv.2011.04216}
 }
 
 @misc{econml,
@@ -483,5 +492,6 @@
   title={{CausalML}: {Python} package for causal machine learning},
   author={Chen, Huigang and Harinen, Totte and Li, Jeong-Yoon and Yung, Liang and Zhao, Zhenyu},
   journal={arXiv preprint arXiv:2002.11631},
-  year={2020}
+  year={2020},
+  doi={10.48550/arXiv.2002.11631}
 }
