+++ src/statspai/synth/plots.py
@@ -11,6 +11,7 @@ Plot types
 * **both** — two-panel: trajectory on top, gap on bottom
 * **weights** — horizontal bar chart of donor weights
 * **placebo** — placebo distribution + treated rank
+* **rmspe** — post/pre RMSPE ratio histogram (Abadie et al. 2010)
 * **conformal** — period-level effects with conformal CI bands
 * **staggered** — cohort-specific ATT comparison
 * **factors** — latent factor loadings (gsynth only)
@@ -80,7 +81,8 @@ def synthplot(
         * ``'both'`` — two-panel: trajectory + gap.
         * ``'weights'`` — donor weight bar chart.
         * ``'placebo'`` — placebo ATT distribution.
-        * ``'placebo_gap'`` — placebo gap spaghetti plot.
+        * ``'placebo_gap'`` — placebo gap spaghetti plot (Abadie et al. 2010).
+        * ``'rmspe'`` — post/pre RMSPE ratio histogram (Abadie et al. 2010).
         * ``'conformal'`` — period-level effects + conformal CIs.
         * ``'staggered'`` — cohort-level ATT comparison.
         * ``'factors'`` — latent factor loadings (gsynth only).
@@ -164,6 +166,10 @@ def synthplot(
         return _plot_staggered(result, ax=ax,
                                figsize=figsize or (9, 6),
                                title=title, **kwargs)
+    if type_ in ("rmspe", "rmspe_ratio"):
+        return _plot_rmspe(result, ax=ax,
+                           figsize=figsize or (9, 6),
+                           title=title, **kwargs)
     if type_ in ("factors", "loadings"):
         return _plot_factors(result, ax=ax,
                              figsize=figsize or (10, 5),
@@ -171,7 +177,7 @@ def synthplot(
 
     raise ValueError(
         f"Unknown plot type {type_!r}. Choose from: 'trajectory', 'gap', "
-        f"'both', 'weights', 'placebo', 'placebo_gap', 'conformal', "
+        f"'both', 'weights', 'placebo', 'placebo_gap', 'rmspe', 'conformal', "
         f"'staggered', 'factors', 'compare'."
     )
 
@@ -470,12 +476,20 @@ def _plot_placebo(result, ax=None, figsize=(9, 6), title=None, **kw):
     return fig, ax
 
 
-def _plot_placebo_gap(result, ax=None, figsize=(10, 6), title=None, **kw):
+def _plot_placebo_gap(result, ax=None, figsize=(10, 6), title=None,
+                      rmspe_threshold=None, **kw):
     """
-    Spaghetti plot: placebo gap paths overlaid with treated gap.
+    Spaghetti plot: all placebo gap paths overlaid with treated gap.
 
-    Requires the full placebo SCM to have been stored — otherwise
-    we fall back to the treated gap only.
+    This is the standard visualisation from Abadie, Diamond & Hainmueller
+    (2010, Figure 4). Each grey line is one placebo unit's gap path;
+    the treated unit is highlighted in bold.
+
+    Parameters
+    ----------
+    rmspe_threshold : float, optional
+        Exclude placebos whose pre-treatment RMSPE exceeds this multiple
+        of the treated unit's pre-RMSPE. Abadie (2010) recommends 2.
     """
     import matplotlib.pyplot as plt
 
@@ -488,24 +502,42 @@ def _plot_placebo_gap(result, ax=None, figsize=(10, 6), title=None, **kw):
     else:
         fig = ax.get_figure()
 
-    # We don't typically store full placebo trajectories, so we show
[…]
