### 2026-04-13 #10387 @brycewang-stanford
I've also addressed the reference DOI gaps flagged in the check:

**Added 10 DOIs** (all verified via CrossRef API):
- 4 publisher DOIs: Spirtes 2000 (MIT Press), Hirano & Imbens 2004 (Wiley), Imbens & Rubin 2015 (Cambridge UP), Angrist & Pischke 2009 (Princeton UP)
- 6 arXiv DOIs for conference papers and preprints: Zheng/NOTEARS, Shalit/TARNet, Shi/DragonNet, Hartford/DeepIV, Sharma/DoWhy, Chen/CausalML

**2 entries without DOI** (confirmed unavailable):
- Manski (1990) — AER Papers & Proceedings, predates DOI adoption; cited with full journal coordinates
- EconML (Microsoft Research) — software repository with no associated journal publication; cited with URL

**Note on the MISSING DOI suggestion**: The suggested DOI `10.1111/j.1475-4932.2011.00742.x` for *Mostly Harmless Econometrics* is actually a book review by Harris & Skeels in *Economic Record*, not the book itself. The correct book DOI is `10.1515/9781400829828` (Princeton University Press), which I have added.
