### 2026-04-13 a25fb22
Add Contributor Covenant Code of Conduct v2.1
Required by JOSS reviewer checklist for community guidelines.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
