### 2026-04-13 2a16f7e
Update CODE_OF_CONDUCT contact email to Stanford address
Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
