### 2026-04-13 edde324
Add Chinese README with language switcher
Add README_CN.md with full Chinese translation of the project
documentation. Both READMEs have a language switcher link at the
top: [English](README.md) | [中文](README_CN.md)

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
