### 2026-04-13 a776496
Update README with team information and images
Added images and updated team member details in README.
