### 2026-04-14 5537b72
Add aggte() with Mammen multiplier bootstrap + uniform CI bands (CS2021 Sec 4)
Mirrors — and where possible surpasses — the aggregation layer of the
R `did` / Python `csdid` packages, reimplemented from the paper rather
than wrapping the upstream code:

* new module `statspai.did.aggte` providing `aggte(result, type=...)`
  with four aggregation schemes (simple / dynamic / group / calendar),
  all sharing a single weighted-influence-function engine.
* Mammen (1993) two-point multiplier bootstrap on the stacked ATT(g,t)
  influence functions, producing pointwise SEs (IQR-rescaled, matching
  `d
[…]
