+++ tests/test_aggte.py
@@ -1,0 +1,169 @@
+"""Tests for the aggte() aggregation + multiplier-bootstrap module."""
+import numpy as np
+import pandas as pd
+import pytest
+
+from statspai.did import callaway_santanna, aggte
+
+
+# --------------------------------------------------------------------------- #
+# Helpers                                                                     #
+# --------------------------------------------------------------------------- #
+
+def _make_staggered_panel(
+    n_per_cohort: int = 40,
+    cohorts=(3, 5, 7, 0),  # 0 = never treated
+    n_periods: int = 8,
+    effect_slope: float = 0.5,
+    seed: int = 0,
+) -> pd.DataFrame:
+    """Construct a staggered panel with dynamic treatment effects."""
+    rng = np.random.default_rng(seed)
+    rows = []
+    unit = 0
+    for g_val in cohorts:
+        for _ in range(n_per_cohort):
+            ui = rng.normal(scale=0.3)
+            for t in range(1, n_periods + 1):
+                te = max(0, t - g_val + 1) * effect_slope if g_val > 0 else 0.0
+                rows.append({
+                    'i': unit, 't': t, 'g': g_val,
+                    'y': ui + 0.2 * t + te + rng.normal(),
+                })
+            unit += 1
+    return pd.DataFrame(rows)
+
+
+@pytest.fixture(scope='module')
+def cs_result():
+    df = _make_staggered_panel(seed=42)
+    return callaway_santanna(df, y='y', g='g', t='t', i='i')
+
+
+# --------------------------------------------------------------------------- #
+# Basic sanity                                                                #
+# --------------------------------------------------------------------------- #
+
+def test_simple_matches_cs_overall(cs_result):
+    """aggte(type='simple') should reproduce the overall ATT of the base result."""
+    out = aggte(cs_result, type='simple', random_state=0, n_boot=400)
+    # Point estimate: simple aggregation uses same cohort-share weights.
+    assert out.estimate == pytest.approx(cs_result.estimate, rel=1e-6, abs=1e-6)
+    assert out.detail.shape[0] == 1
+    assert 'overall' in out.detail.columns
+
+
+def test_dynamic_recovers_positive_post(cs_result):
+    es = aggte(cs_result, type='dynamic', random_state=0, n_boot=400)
+    post = es.detail[es.detail['relative_time'] >= 0]
+    # With effect_slope=0.5 and a linear ramp, post-event coefficients are
+    # monotonically increasing on average.
+    assert (post['att'].iloc[-1] > post['att'].iloc[0])
+    # All post-event ATTs are positive (well-powered).
+    assert (post['att'] > 0).all()
+
+
+def test_group_aggregation_shape(cs_result):
+    grp = aggte(cs_result, type='group', random_state=0, n_boot=400)
+    cohorts = sorted(cs_result.detail['group'].unique())
+    assert list(grp.detail['group']) == cohorts
+    # Earlier cohorts experience the ramp longer → larger θ(g).
+    # (Not guaranteed in every simulation, so we just check finiteness.)
+    assert grp.detail['att'].notna().all()
+    assert (grp.detail['se'] > 0).all()
+
+
+def test_calendar_aggregation_shape(cs_result):
+    cal = aggte(cs_result, type='calendar', random_state=0, n_boot=400)
+    # Only post-treatment calendar times appear.
+    assert cal.detail['time'].min() >= cs_result.detail['group'].min()
+    assert (cal.detail['se'] > 0).all()
+
+
+# --------------------------------------------------------------------------- #
+# Uniform bands                                                               #
+# --------------------------------------------------------------------------- #
+
[…]
