+++ src/statspai/did/aggte.py
@@ -1,0 +1,452 @@
+"""
+Aggregated group-time ATTs for staggered DID with multiplier bootstrap.
+
+Implements the four aggregation schemes of Callaway & Sant'Anna (2021):
+
+- ``simple``   — cohort-size-weighted average over all post-treatment (g, t)
+- ``dynamic``  — event-study: average ATT by relative time e = t − g
+- ``group``    — average ATT per cohort g across its post-treatment periods
+- ``calendar`` — average ATT per calendar time t across already-treated cohorts
+
+Inference is by Mammen (1993) multiplier bootstrap applied to the influence
+functions of the underlying ATT(g, t) estimates.  This reproduces the
+uniform (simultaneous) confidence bands that are the signature of the R
+package ``did`` / Python package ``csdid``.
+
+References
+----------
+Callaway, B. and Sant'Anna, P.H.C. (2021).
+    "Difference-in-Differences with Multiple Time Periods."
+    *Journal of Econometrics*, 225(2), 200-230.
+    Section 4 (aggregated parameters) and Section 4.2
+    (uniform inference via multiplier bootstrap).
+
+Mammen, E. (1993).
+    "Bootstrap and Wild Bootstrap for High Dimensional Linear Models."
+    *Annals of Statistics*, 21(1), 255-285.
+    Two-point multiplier distribution used for the wild weights.
+"""
+
+from __future__ import annotations
+
+from typing import Optional, Tuple
+
+import numpy as np
+import pandas as pd
+from scipy import stats
+
+from ..core.results import CausalResult
+
+
+# ======================================================================
+# Public API
+# ======================================================================
+
+def aggte(
+    result: CausalResult,
+    type: str = 'simple',
+    balance_e: Optional[int] = None,
+    min_e: float = -np.inf,
+    max_e: float = np.inf,
+    na_rm: bool = True,
+    bstrap: bool = True,
+    boot_type: str = 'multiplier',
+    n_boot: int = 1000,
+    cband: bool = True,
+    alpha: float = 0.05,
+    random_state: Optional[int] = None,
+) -> CausalResult:
+    """
+    Aggregate group-time ATT(g, t) estimates from ``callaway_santanna``.
+
+    Parameters
+    ----------
+    result : CausalResult
+        Output of :func:`callaway_santanna`. Must contain influence
+        functions in ``result._influence_funcs``.
+    type : {'simple', 'dynamic', 'group', 'calendar'}, default 'simple'
+        Aggregation scheme. ``'dynamic'`` is the event study.
+    balance_e : int, optional
+        Only used when ``type='dynamic'``. If given, the event study
+        restricts attention to cohorts observed over the *balanced* set of
+        event times ``{-∞ ≤ e ≤ balance_e}``. This is the ``balance_e``
+        option of R's :func:`did::aggte`.
+    min_e, max_e : float, default (-inf, inf)
+        Truncate the reported event-time window.
+    na_rm : bool, default True
+        Drop ATT(g, t) with missing / infinite SE before aggregating.
+    bstrap : bool, default True
+        If ``True``, compute SE / CI by Mammen multiplier bootstrap on the
+        influence functions. If ``False``, fall back to the analytic
+        (delta-method) SE already attached to each ATT(g, t).
+    boot_type : {'multiplier'}, default 'multiplier'
+        Only ``'multiplier'`` is supported; kept for ``csdid`` parity.
[…]
