### 2026-04-14 2a278c6
Add pyegen-style row helpers + outlier_indicator, PyStataR migration guide
Ports pyegen (rowmean/rowtotal/rowmax/rowmin/rowsd/rowcount/rank) and
pywinsor2.outlier_indicator from PyStataR so users migrating from the
deprecated package keep a familiar API under sp.*. Adds MIGRATION.md
with the full API mapping table.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
