+++ src/statspai/utils/egen.py
@@ -1,0 +1,159 @@
+"""
+Stata-style egen row-wise and rank helpers.
+
+Ports the core ``pyegen`` convenience functions to StatsPAI so users
+migrating from PyStataR keep a familiar API. All functions accept a
+DataFrame plus a list of variables and return a pandas Series aligned
+to the DataFrame's index.
+
+Equivalent to Stata's::
+
+    egen newvar = rowmean(var1 var2 var3)
+    egen newvar = rowtotal(var1 var2 var3)
+    egen newvar = rowmax(var1 var2 var3)
+    egen newvar = rowmin(var1 var2 var3)
+    egen newvar = rowsd(var1 var2 var3)
+    egen newvar = rownonmiss(var1 var2 var3)
+    egen newvar = rank(var), [by(group)] [field|track|unique]
+"""
+
+from typing import List, Optional, Sequence, Union
+
+import numpy as np
+import pandas as pd
+
+
+def _select(data: pd.DataFrame, vars: Sequence[str]) -> pd.DataFrame:
+    missing = [v for v in vars if v not in data.columns]
+    if missing:
+        raise ValueError(f"Columns not found: {missing}")
+    return data[list(vars)]
+
+
+def rowmean(data: pd.DataFrame, vars: Sequence[str]) -> pd.Series:
+    """Row-wise mean across ``vars``, ignoring NaN (Stata ``egen rowmean``)."""
+    return _select(data, vars).mean(axis=1, skipna=True)
+
+
+def rowtotal(data: pd.DataFrame, vars: Sequence[str]) -> pd.Series:
+    """Row-wise sum across ``vars``, treating NaN as 0 (Stata ``egen rowtotal``)."""
+    return _select(data, vars).sum(axis=1, skipna=True)
+
+
+def rowmax(data: pd.DataFrame, vars: Sequence[str]) -> pd.Series:
+    """Row-wise max across ``vars``, ignoring NaN."""
+    return _select(data, vars).max(axis=1, skipna=True)
+
+
+def rowmin(data: pd.DataFrame, vars: Sequence[str]) -> pd.Series:
+    """Row-wise min across ``vars``, ignoring NaN."""
+    return _select(data, vars).min(axis=1, skipna=True)
+
+
+def rowsd(data: pd.DataFrame, vars: Sequence[str]) -> pd.Series:
+    """Row-wise standard deviation across ``vars`` (Stata ``egen rowsd``)."""
+    return _select(data, vars).std(axis=1, skipna=True, ddof=1)
+
+
+def rowcount(data: pd.DataFrame, vars: Sequence[str]) -> pd.Series:
+    """Row-wise count of non-missing values (Stata ``egen rownonmiss``)."""
+    return _select(data, vars).notna().sum(axis=1).astype(int)
+
+
+def rank(
+    data: pd.DataFrame,
+    var: str,
+    by: Optional[Union[str, Sequence[str]]] = None,
+    method: str = 'average',
+    ascending: bool = True,
+) -> pd.Series:
+    """
+    Rank a variable, optionally within groups.
+
+    Equivalent to Stata's ``egen newvar = rank(var), by(group)`` with
+    ``method`` mapping to Stata's tie-handling options:
+
+    - ``'average'`` → Stata default
+    - ``'min'``     → ``field``
+    - ``'dense'``   → ``track``
+    - ``'first'``   → ``unique``
+
+    Parameters
+    ----------
+    data : pd.DataFrame
+    var : str
+        Column to rank.
+    by : str or list of str, optional
+        Group variables. If given, ranks are computed within each group.
+    method : {'average', 'min', 'max', 'dense', 'first'}, default 'average'
+    ascending : bool, default True
+
+    Returns
+    -------
+    pd.Series aligned to ``data.index``.
+    """
[…]
