### 2026-04-14 d21faff
Guard outlier_indicator against all-NaN columns
Skip nanpercentile on empty slices to avoid RuntimeWarning; flag
defaults to 0 for columns with no valid observations.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
