### 2026-04-14 ec2c57e
Sun-Abraham cluster-robust sandwich SE + ggdid() aggte visualiser
Continue the csdid / differences parity effort by reimplementing the two
remaining high-impact pieces from their respective papers.

1. Rewrite `sun_abraham()` inference layer (src/statspai/did/sun_abraham.py)
   * Saturated cohort × event-time design estimated with proper two-way
     within transformation (iterative unit/time demeaning, correct on
     unbalanced panels).
   * Liang-Zeger cluster-robust sandwich covariance with small-sample
     (G/(G-1)) · (N-1)/(N-k) adjustment, defaulting to clustering on the
     unit id; 
[…]
