+++ src/statspai/did/sun_abraham.py
@@ -1,20 +1,32 @@
 """
-Sun & Abraham (2021) interaction-weighted estimator for staggered DID.
+Sun & Abraham (2021) interaction-weighted event-study estimator.
 
-Estimates dynamic treatment effects via cohort-specific event study
-coefficients that are free of contamination from heterogeneous effects
-across cohorts. The estimator re-weights the OLS event study
-coefficients using cohort shares.
+Fits a *saturated* regression that interacts every cohort dummy with
+every relative-time dummy, then aggregates the interaction coefficients
+across cohorts using empirical cohort shares to deliver the IW
+estimator δ̂^IW_ℓ that is robust to heterogeneous treatment effects
+(Sun & Abraham 2021, Theorem 1 / Corollary 1).
+
+Standard errors are computed from the classical OLS sandwich:
+
+    Var(β̂) = (X'X)⁻¹  ( Σ_c  X_c' u_c u_c' X_c )  (X'X)⁻¹
+
+clustered at the unit (or a user-supplied) level.  The IW estimator is
+a linear combination of β̂, so Var(δ̂^IW_ℓ) = w_ℓ' Var(β̂) w_ℓ by the
+delta method (shares treated as estimated but converging at parametric
+rate, following SA 2021 eq. (18)).
 
 References
 ----------
 Sun, L. and Abraham, S. (2021).
-"Estimating Dynamic Treatment Effects in Event Studies with
-Heterogeneous Treatment Effects."
-*Journal of Econometrics*, 225(2), 175-199.
+    "Estimating Dynamic Treatment Effects in Event Studies with
+     Heterogeneous Treatment Effects."
+    *Journal of Econometrics*, 225(2), 175-199.
 """
 
-from typing import Optional, List, Dict, Any, Tuple
+from __future__ import annotations
+
+from typing import Optional, List, Tuple
 
 import numpy as np
 import pandas as pd
@@ -22,6 +34,10 @@ from scipy import stats
 
 from ..core.results import CausalResult
 
+
+# ======================================================================
+# Public API
+# ======================================================================
 
 def sun_abraham(
     data: pd.DataFrame,
@@ -36,16 +52,12 @@ def sun_abraham(
     alpha: float = 0.05,
 ) -> CausalResult:
     """
-    Sun & Abraham (2021) interaction-weighted event study estimator.
-
-    Estimates cohort-specific treatment effects and aggregates them
-    using cohort-share weights to obtain heterogeneity-robust dynamic
-    treatment effect estimates.
+    Sun & Abraham (2021) interaction-weighted event-study estimator.
 
     Parameters
     ----------
     data : pd.DataFrame
-        Balanced panel data (long format).
+        Long-format panel data.
     y : str
         Outcome variable.
     g : str
@@ -55,12 +67,14 @@ def sun_abraham(
     i : str
         Unit identifier.
     event_window : tuple of (int, int), optional
-        (min_relative_time, max_relative_time) for event study.
-        Default: all available relative times.
+        (min_relative_time, max_relative_time).
+        Default: observed range in the data.
     control_group : str, default 'nevertreated'
-        ``'nevertreated'`` or ``'lastcohort'`` (last treated cohort).
+        ``'nevertreated'`` or ``'lastcohort'``.  When ``'lastcohort'``,
+        the latest treated cohort is used as the reference and dropped
+        from the IW aggregation.
     covariates : list of str, optional
-        Additional control variables.
+        Additional controls (time-varying; added linearly).
     cluster : str, optional
[…]
