### 2026-04-14 b90e8b4
dCDH: joint placebo Wald + avg cumulative dynamic effect (2024)
Extend the de Chaisemartin-D'Haultfoeuille estimator with two pieces of
joint inference that match the standard output of the upstream R /
Python `did_multiplegt_dyn` package (dCDH 2024 RESt §3):

* `_joint_placebo_test()` — Wald chi-squared test of
  H0: all placebo lags equal zero.  Covariance across lags is estimated
  directly from the cluster-bootstrap draws (so consecutive pre-periods,
  which share data, are not assumed independent).  Ridge-regularises the
  covariance and falls back to a pseudo-inverse if singular.
* `_avg
[…]
