+++ src/statspai/did/did_multiplegt.py
@@ -21,6 +21,11 @@ de Chaisemartin, C. and D'Haultfoeuille, X. (2020).
 de Chaisemartin, C. and D'Haultfoeuille, X. (2022).
 "Two-way fixed effects and differences-in-differences with heterogeneous
 treatment effects: A survey."  *The Econometrics Journal*, 26(3), C1-C30.
+
+de Chaisemartin, C. and D'Haultfoeuille, X. (2024).
+"Difference-in-Differences Estimators of Intertemporal Treatment Effects."
+*Review of Economics and Statistics*, forthcoming.  (Joint placebo test
+and average cumulative effect, Section 3.)
 """
 
 from typing import Optional, List, Dict, Any, Tuple
@@ -252,6 +257,12 @@ def did_multiplegt(
         })
     es_df = pd.DataFrame(es_rows) if es_rows else pd.DataFrame()
 
+    # ── Joint placebo test + avg cumulative dynamic effect ───────── #
+    joint_placebo = _joint_placebo_test(placebo_results, boot_placebo)
+    avg_cumulative = _avg_cumulative_effect(
+        dynamic_results, boot_dynamic, alpha,
+    )
+
     # ── Assemble model_info ──────────────────────────────────────── #
     model_info: Dict[str, Any] = {
         'estimator': 'de Chaisemartin-D\'Haultfoeuille (2020)',
@@ -264,6 +275,8 @@ def did_multiplegt(
         'placebo': placebo_out,
         'dynamic': dynamic_out,
         'event_study': es_df,
+        'joint_placebo_test': joint_placebo,
+        'avg_cumulative_effect': avg_cumulative,
     }
 
     return CausalResult(
@@ -564,3 +577,86 @@ def _estimate_dynamic(
     weighted_est = sum(e * w / total_w for e, w in zip(estimates, weights))
 
     return {'estimate': float(weighted_est), 'n_cells': len(estimates)}
+
+
+# ======================================================================
+# Joint inference
+# ======================================================================
+
+def _joint_placebo_test(
+    placebo_results: List[Dict[str, Any]],
+    boot_placebo: Optional[np.ndarray],
+) -> Optional[Dict[str, Any]]:
+    """Joint Wald test of H0: all placebo lags equal zero.
+
+    Uses the bootstrap draws (cluster-level) to estimate the covariance
+    across lags, so the test accounts for correlation between consecutive
+    pre-periods.  Follows dCDH (2024, §3.3) "joint placebo" recommendation.
+    """
+    if not placebo_results or boot_placebo is None:
+        return None
+
+    est = np.array([r['estimate'] for r in placebo_results], dtype=float)
+    boot = np.asarray(boot_placebo, dtype=float)  # (n_boot, k)
+    # Drop all-NaN columns / rows before covariance.
+    mask_cols = ~np.all(np.isnan(boot), axis=0)
+    if not mask_cols.any():
+        return None
+    est = est[mask_cols]
+    boot = boot[:, mask_cols]
+    valid_rows = ~np.any(np.isnan(boot), axis=1)
+    if valid_rows.sum() < boot.shape[1] + 1:
+        return None
+    boot = boot[valid_rows]
+
+    cov = np.cov(boot, rowvar=False, ddof=1)
+    if cov.ndim == 0:
+        cov = np.array([[float(cov)]])
+    k = est.shape[0]
+    cov_reg = cov + np.eye(k) * 1e-10
+
+    try:
+        W = float(est @ np.linalg.solve(cov_reg, est))
+    except np.linalg.LinAlgError:
[…]
