### 2026-04-14 cc6f25f
honest_did: accept aggte(type='dynamic') output (R-R ⇄ CS integration)
Close the Callaway-Sant'Anna → Rambachan-Roth sensitivity loop by
making `honest_did` and `breakdown_m` polymorphic over how an event
study is stored on a `CausalResult`.

Before this change the sensitivity functions only looked at
`result.model_info['event_study']`, so feeding them the output of
`aggte(cs, type='dynamic')` (where the event study lives in `detail`
with Mammen uniform bands attached) silently raised ValueError.
Now a new `_extract_event_study()` helper searches both locations:

1. `model_info['event_study']`
[…]
