### 2026-04-14 26c8818
Add cs_report() — one-call Callaway-Sant'Anna report card
Compose the full staggered-DID workflow — ATT(g,t) estimation, all
four aggte aggregations with Mammen multiplier-bootstrap uniform
bands, the pre-trend Wald test, and a Rambachan-Roth breakdown-M*
row for every post-treatment event time — into a single function
call that pretty-prints an end-to-end report.

Entry points
* `src/statspai/did/report.py::cs_report()` — accepts either a raw
  panel (runs callaway_santanna first) or a pre-fitted CausalResult;
  returns a `CSReport` dataclass with DataFrames for each component.
* `CSReport.to
[…]
