+++ src/statspai/did/report.py
@@ -1,0 +1,289 @@
+"""
+One-call staggered-DID report.
+
+``cs_report()`` composes the full Callaway-Sant'Anna workflow —
+ATT(g,t) estimation, all four :func:`aggte` aggregations with Mammen
+multiplier-bootstrap uniform bands, the pre-trend Wald test, and a
+Rambachan-Roth breakdown-value row for every post-treatment event
+time — into a single function call, and pretty-prints the result.
+
+The design mirrors the one-screen summaries that practitioners expect
+from ``did::summary()`` + ``HonestDiD`` + ``ggdid`` in R, so that a
+user can interpret a staggered-DID study at a glance.
+
+References
+----------
+Callaway, B. and Sant'Anna, P.H.C. (2021).
+    "Difference-in-Differences with Multiple Time Periods."
+    *Journal of Econometrics*, 225(2), 200-230.
+Rambachan, A. and Roth, J. (2023).
+    "A More Credible Approach to Parallel Trends."
+    *Review of Economic Studies*, 90(5), 2555-2591.
+"""
+
+from __future__ import annotations
+
+from dataclasses import dataclass, field
+from typing import Any, Dict, List, Optional
+
+import numpy as np
+import pandas as pd
+
+from ..core.results import CausalResult
+from .aggte import aggte
+from .callaway_santanna import callaway_santanna
+from .honest_did import breakdown_m
+
+
+# ======================================================================
+# Container
+# ======================================================================
+
+@dataclass
+class CSReport:
+    """Structured output of :func:`cs_report`.
+
+    Attributes are plain pandas objects so downstream users can export
+    to LaTeX, Markdown, or Excel without any custom converters.
+    """
+
+    overall: Dict[str, float]
+    simple: pd.DataFrame
+    dynamic: pd.DataFrame
+    group: pd.DataFrame
+    calendar: pd.DataFrame
+    pretrend: Dict[str, Any]
+    breakdown: pd.DataFrame
+    meta: Dict[str, Any] = field(default_factory=dict)
+
+    def __repr__(self) -> str:  # pragma: no cover - cosmetic
+        return self._format()
+
+    def to_text(self) -> str:
+        """Return the human-readable report as a single string."""
+        return self._format()
+
+    # ------------------------------------------------------------------
+    def _format(self) -> str:
+        w = 78
+        lines: List[str] = []
+        title = " Callaway–Sant'Anna Staggered-DID Report "
+        lines.append("=" * w)
+        lines.append(title.center(w, "="))
+        lines.append("=" * w)
+
+        # Header block
+        m = self.meta
+        lines.append(f"Units: {m.get('n_units', '?')}    "
+                     f"Periods: {m.get('n_periods', '?')}    "
+                     f"Cohorts: {m.get('n_cohorts', '?')}    "
+                     f"α = {m.get('alpha', 0.05)}")
+        lines.append(f"Estimator: {m.get('estimator', '?')}    "
+                     f"Control group: {m.get('control_group', '?')}    "
+                     f"Anticipation: {m.get('anticipation', 0)}")
+        lines.append(f"Multiplier bootstrap: B = {m.get('n_boot', '?')}, "
+                     f"seed = {m.get('random_state', '—')}")
+
+        # Overall ATT
+        lines.append("-" * w)
+        o = self.overall
+        lines.append(
+            f"Overall ATT  =  {o['estimate']:.4f}   "
+            f"SE = {o['se']:.4f}   "
+            f"{int(100*(1-m.get('alpha', 0.05)))}% CI = "
+            f"[{o['ci_lower']:.4f}, {o['ci_upper']:.4f}]   "
+            f"p = {o['pvalue']:.4g}"
+        )
+
+        # Pre-trend
[…]
