### 2026-04-14 03f08d9
docs: add 0.7.0 CHANGELOG + csdid/differences parity table in README
Make the last five rounds of DiD work externally visible.

CHANGELOG.md
* Add a 0.7.0 release block (2026-04-14) summarising: aggte() unified
  aggregation, Mammen multiplier bootstrap w/ uniform bands,
  balance_e / min_e / max_e, anticipation, cs_report(), ggdid(), dCDH
  joint placebo + avg cumulative effect, Sun-Abraham rewritten with
  cluster sandwich SEs, and honest_did polymorphism.
* Full reference list for every cited paper.

README.md
* Add `sp.aggte`, `sp.cs_report`, `sp.ggdid` rows to the DiD feature
  table.
* A
[…]
