### 2026-04-15 0d89afa
CSReport.to_markdown() + to_latex() — paper-ready export of cs_report
Add two export methods to the CSReport dataclass returned by cs_report().

* `to_markdown(float_format='%.4f') -> str`
    GitHub-flavoured Markdown: a header block followed by four
    sub-tables (event study / θ(g) / θ(t) / breakdown M*).  Integer
    columns (relative_time, group, time) render as integers instead
    of 3.0000.  Suitable for pasting directly into a PR, blog post,
    or Jupyter Markdown cell.

* `to_latex(float_format='%.4f', caption=None, label=None) -> str`
    Publication-ready LaTeX fragment wrapped i
[…]
