+++ src/statspai/did/report.py
@@ -64,6 +64,181 @@ class CSReport:
         return self._format()
 
     # ------------------------------------------------------------------
+    # Export: Markdown
+    # ------------------------------------------------------------------
+    def to_markdown(self, float_format: str = "%.4f") -> str:
+        """Render the report as GitHub-Flavoured Markdown.
+
+        Suitable for pasting directly into a pull request, blog post,
+        or Jupyter notebook Markdown cell.
+        """
+        m = self.meta
+        o = self.overall
+        alpha = m.get("alpha", 0.05)
+        ci_pct = int(100 * (1 - alpha))
+
+        lines: List[str] = []
+        lines.append("## Callaway–Sant'Anna Staggered-DID Report")
+        lines.append("")
+        lines.append(
+            f"- **Units / Periods / Cohorts**: "
+            f"{m.get('n_units', '?')} / {m.get('n_periods', '?')} / "
+            f"{m.get('n_cohorts', '?')}"
+        )
+        lines.append(
+            f"- **Estimator**: {m.get('estimator', '?')} · "
+            f"**Control group**: {m.get('control_group', '?')} · "
+            f"**Anticipation**: {m.get('anticipation', 0)}"
+        )
+        lines.append(
+            f"- **Multiplier bootstrap**: B = {m.get('n_boot', '?')}, "
+            f"seed = {m.get('random_state', '—')}"
+        )
+        lines.append(
+            f"- **Overall ATT** = {o['estimate']:.4f} "
+            f"(SE = {o['se']:.4f}) · "
+            f"{ci_pct}% CI [{o['ci_lower']:.4f}, {o['ci_upper']:.4f}] · "
+            f"p = {o['pvalue']:.4g}"
+        )
+        pt = self.pretrend or {}
+        if pt:
+            lines.append(
+                f"- **Pre-trend Wald**: χ²({pt.get('df', 0)}) = "
+                f"{pt.get('statistic', float('nan')):.3f}, "
+                f"p = {pt.get('pvalue', float('nan')):.4g}"
+            )
+        lines.append("")
+
+        def _md(df: pd.DataFrame, cols: List[str]) -> str:
+            present = [c for c in cols if c in df.columns]
+            sub = _format_numeric_columns(df[present], float_format)
+            try:
+                return sub.to_markdown(index=False)
+            except ImportError:  # tabulate missing — fall back to plain
+                return sub.to_string(index=False)
+
+        lines.append("### Event study (dynamic aggregation)")
+        lines.append("")
+        lines.append(_md(self.dynamic, [
+            "relative_time", "att", "se",
+            "ci_lower", "ci_upper",
+            "cband_lower", "cband_upper", "pvalue",
+        ]))
+        lines.append("")
+        lines.append("### θ(g) — per-cohort aggregation")
+        lines.append("")
+        lines.append(_md(self.group, [
+            "group", "att", "se",
+            "ci_lower", "ci_upper",
+            "cband_lower", "cband_upper", "pvalue",
+        ]))
+        lines.append("")
+        lines.append("### θ(t) — per-calendar-time aggregation")
+        lines.append("")
+        lines.append(_md(self.calendar, [
+            "time", "att", "se",
+            "ci_lower", "ci_upper",
+            "cband_lower", "cband_upper", "pvalue",
+        ]))
+        lines.append("")
+        lines.append("### Rambachan–Roth breakdown M\\*")
+        lines.append("")
+        if len(self.breakdown):
+            lines.append(_md(self.breakdown, [
+                "relative_time", "att", "se",
+                "breakdown_M_star", "robust_at_1_SE",
+            ]))
+        else:
[…]
