### 2026-04-15 105a1d2
CSReport.plot() 2×2 summary figure + end-to-end demo script
Add a one-call figure renderer to the CSReport container and a
self-contained demo script under test-notebooks/ that exercises the
full v0.7.0 staggered-DID pipeline.

CSReport.plot()  (src/statspai/did/report.py)
  Renders a 2×2 figure:
   - top-left  : event study (dynamic) with uniform band
   - top-right : θ(g) per-cohort aggregation with uniform band
   - bottom-L  : θ(t) per-calendar-time aggregation
   - bottom-R  : RR breakdown M* horizontal bars with 1σ reference
  Under the hood a tiny `_CSReportLike` shim exposes exactly th
[…]
