### 2026-04-15 2c2ff18
Add end-to-end CS report smoke test (doubles as runnable demo)
Convert the v0.7.0 walkthrough into a tracked pytest integration
test, since the user's `test-notebooks/` and `examples/` directories
are both gitignored by design.

tests/test_cs_report_smoke.py
* `simulate_staggered_panel()` — balanced 12-period panel with 4
  cohorts and a heterogeneous linear ramp (effect slope 0.4).
* `run_demo(outdir)` — public helper that imports can call to
  regenerate /tmp-style artefacts (Markdown, LaTeX, 2×2 PNG panel).
* Six pytest assertions that pin down the expected behaviour of the
  whole v0.7.0 p
[…]
