+++ tests/test_cs_report_smoke.py
@@ -1,0 +1,147 @@
+"""End-to-end smoke test for the v0.7.0 Callaway-Sant'Anna workflow.
+
+Exercises the full pipeline introduced in v0.7.0:
+
+    callaway_santanna() -> aggte() -> honest_did() -> cs_report()
+                        -> CSReport.plot() / .to_markdown() / .to_latex()
+
+Doubles as a runnable demo: import and call :func:`run_demo` to produce
+artifacts (PNG / Markdown / LaTeX) in a specified directory::
+
+    from tests.test_cs_report_smoke import run_demo
+    run_demo("/tmp/my_outputs")
+"""
+from __future__ import annotations
+
+import os
+from pathlib import Path
+
+import numpy as np
+import pandas as pd
+import pytest
+
+from statspai.did import cs_report
+
+
+# ---------------------------------------------------------------------------
+# Data-generating process
+# ---------------------------------------------------------------------------
+
+def simulate_staggered_panel(
+    n_per_cohort: int = 60,
+    cohorts=(4, 7, 10, 0),
+    n_periods: int = 12,
+    effect_slope: float = 0.4,
+    seed: int = 42,
+) -> pd.DataFrame:
+    """Balanced staggered panel with heterogeneous linear dynamic effects.
+
+    Cohort ``g`` is first treated at period ``g`` and the effect at event
+    time ``e = t - g`` is ``max(0, e + 1) * effect_slope``.  Cohort 0 is
+    never treated.
+    """
+    rng = np.random.default_rng(seed)
+    rows = []
+    unit = 0
+    for g_val in cohorts:
+        for _ in range(n_per_cohort):
+            alpha_i = rng.normal(scale=0.30)
+            for t in range(1, n_periods + 1):
+                te = max(0, t - g_val + 1) * effect_slope if g_val > 0 else 0
+                y = alpha_i + 0.15 * t + te + rng.normal()
+                rows.append({"unit": unit, "t": t, "g": g_val, "y": y})
+            unit += 1
+    return pd.DataFrame(rows)
+
+
+# ---------------------------------------------------------------------------
+# Runnable demo
+# ---------------------------------------------------------------------------
+
+def run_demo(outdir: str = "/tmp", n_boot: int = 500, seed: int = 0):
+    """Execute the v0.7.0 workflow end-to-end and save artifacts to ``outdir``.
+
+    Returns the :class:`statspai.did.CSReport` instance.
+    """
+    outdir_p = Path(outdir)
+    outdir_p.mkdir(parents=True, exist_ok=True)
+
+    df = simulate_staggered_panel()
+    rpt = cs_report(df, y="y", g="g", t="t", i="unit",
+                    n_boot=n_boot, random_state=seed, verbose=False)
+
+    (outdir_p / "cs_report_demo.md").write_text(
+        rpt.to_markdown(float_format="%.3f"))
+    (outdir_p / "cs_report_demo.tex").write_text(
+        rpt.to_latex(float_format="%.3f",
+                     caption="CS report on the simulated DGP.",
+                     label="tab:cs_demo"))
+
+    try:
+        import matplotlib
+        matplotlib.use("Agg")
+        fig, _ = rpt.plot(suptitle="Callaway-Sant'Anna report — simulated panel")
+        fig.savefig(outdir_p / "cs_report_demo_panel.png",
+                    dpi=110, bbox_inches="tight")
+    except ImportError:
+        pass
+
+    return rpt
+
+
+# ---------------------------------------------------------------------------
+# Assertions (pytest)
+# ---------------------------------------------------------------------------
+
[…]
