### 2026-04-15 3fd86c5
CSReport.to_excel() — multi-sheet workbook export
Add an Excel writer to the CSReport dataclass so policy-analysis
workflows that live in spreadsheets can consume the full staggered-DID
report directly.

Six sheets:
 * Summary    — header block (run configuration + overall ATT + CI +
                pre-trend Wald) as a two-column key/value table.
 * Dynamic    — event-study frame with uniform-band columns.
 * Group      — θ(g) per-cohort.
 * Calendar   — θ(t) per-calendar-time.
 * Breakdown  — Rambachan-Roth breakdown M* per post event time.
 * Meta       — full model_info dict round-tripped 
[…]
