### 2026-04-15 0ff401d
cs_report(save_to=prefix): one-call dump of all five export formats
Close the v0.7.0 export surface by making `cs_report()` itself take
a `save_to` path prefix and write every supported format in one call.
Users no longer have to chain `.to_markdown()` / `.to_latex()` /
`.to_excel()` / `.plot()` manually just to archive a run.

Behaviour: `save_to='/path/prefix'` writes

    /path/prefix.txt    fixed-width plain text (via .to_text)
    /path/prefix.md     GitHub Markdown
    /path/prefix.tex    booktabs LaTeX fragment
    /path/prefix.xlsx   6-sheet workbook   (skipped if openpyxl missing)
   
[…]
