### 2026-04-15 489b03b
callaway_santanna(panel=False): repeated cross-sections support
Close the last significant capability gap vs `csdid` / `differences`:
repeated-cross-sections (RCS) data where observations are not matched
across time.

Under `panel=False`, each (g, t) ATT is estimated as the unconditional
2×2 cell-mean DID

    ATT(g, t) = (Ȳ_{g, t} - Ȳ_{g, base}) - (Ȳ_{c, t} - Ȳ_{c, base})

with c = never-treated cohort, and observation-level influence
functions

    ψ_i =   1{G_i=g, T_i=t}    (Y_i - Ȳ_{g, t})    / p_{g, t}
          - 1{G_i=g, T_i=base} (Y_i - Ȳ_{g, base}) / p_{g, base}
          - 1{G_i=c, T
[…]
