+++ tests/test_cs_rcs.py
@@ -1,0 +1,167 @@
+"""Tests for the repeated-cross-sections (panel=False) path of
+callaway_santanna()."""
+import numpy as np
+import pandas as pd
+import pytest
+
+from statspai.did import aggte, callaway_santanna, cs_report, honest_did
+
+
+def _rcs_panel(
+    n_obs: int = 1200,
+    cohorts=(3, 5, 7, 0),
+    cohort_probs=(0.2, 0.2, 0.2, 0.4),
+    n_periods: int = 8,
+    effect_slope: float = 0.5,
+    seed: int = 0,
+) -> pd.DataFrame:
+    """Repeated cross-sections: each 'obs' appears at exactly one time."""
+    rng = np.random.default_rng(seed)
+    rows = []
+    for obs in range(n_obs):
+        g = int(rng.choice(cohorts, p=cohort_probs))
+        t = int(rng.integers(1, n_periods + 1))
+        ui = rng.normal(scale=0.3)
+        te = max(0, t - g + 1) * effect_slope if g > 0 else 0.0
+        y = ui + 0.2 * t + te + rng.normal()
+        rows.append({'obs': obs, 't': t, 'g': g, 'y': y})
+    return pd.DataFrame(rows)
+
+
+@pytest.fixture(scope='module')
+def rcs_result():
+    df = _rcs_panel(seed=42)
+    return callaway_santanna(
+        df, y='y', g='g', t='t', i='obs',
+        estimator='reg', panel=False,
+    )
+
+
+# --------------------------------------------------------------------------- #
+# Basic RCS behaviour                                                         #
+# --------------------------------------------------------------------------- #
+
+def test_rcs_returns_causal_result(rcs_result):
+    from statspai.core.results import CausalResult
+    assert isinstance(rcs_result, CausalResult)
+    assert rcs_result.model_info['panel'] is False
+    assert 'RCS' in rcs_result.model_info['estimator']
+
+
+def test_rcs_recovers_positive_att(rcs_result):
+    assert rcs_result.estimate > 0
+    # With n=1200 and effect_slope=0.5 the effect should be significant.
+    assert rcs_result.pvalue < 0.01
+
+
+def test_rcs_produces_full_detail_grid(rcs_result):
+    # 3 cohorts × (8 - 1) non-base periods each = 21 (g, t) pairs
+    assert len(rcs_result.detail) == 21
+    assert {'group', 'time', 'att', 'se',
+            'ci_lower', 'ci_upper', 'pvalue',
+            'relative_time'} <= set(rcs_result.detail.columns)
+
+
+def test_rcs_influence_functions_are_obs_level(rcs_result):
+    inf = rcs_result._influence_funcs
+    assert inf is not None
+    # One row per observation, one column per (g, t) pair.
+    assert inf.shape[0] == rcs_result.n_obs
+    assert inf.shape[1] == len(rcs_result.detail)
+
+
+# --------------------------------------------------------------------------- #
+# aggte / cs_report integration                                               #
+# --------------------------------------------------------------------------- #
+
+def test_aggte_dynamic_works_on_rcs(rcs_result):
+    es = aggte(rcs_result, type='dynamic', n_boot=200, random_state=0)
+    assert {'cband_lower', 'cband_upper'} <= set(es.detail.columns)
+    post = es.detail[es.detail['relative_time'] >= 0]
+    # Post-event ATTs should be positive (true DGP effect).
+    assert (post['att'] > 0).all()
+
+
+def test_cs_report_accepts_rcs_result(rcs_result):
+    rpt = cs_report(rcs_result, n_boot=200, random_state=0, verbose=False)
[…]
