+++ src/statspai/did/callaway_santanna.py
@@ -46,6 +46,7 @@ def callaway_santanna(
     base_period: str = 'universal',
     anticipation: int = 0,
     alpha: float = 0.05,
+    panel: bool = True,
 ) -> CausalResult:
     """
     Callaway & Sant'Anna (2021) estimator for staggered DID.
@@ -78,6 +79,15 @@ def callaway_santanna(
         Sant'Anna (2021), Section 3.2.
     alpha : float, default 0.05
         Significance level.
+    panel : bool, default True
+        If ``True`` (default), treat the data as a balanced panel and
+        estimate ATT(g, t) via within-unit first differences.
+        If ``False``, treat the data as *repeated cross-sections* —
+        observations are not matched across time.  In RCS mode the
+        estimator is the unconditional 2×2 cell-mean DID per (g, t)
+        pair (CS2021 §3.2, eqn 2.4, RCS version).  The covariate-free
+        ``estimator='reg'`` path is the only one currently supported
+        for RCS; IPW / DR can be added later.
 
     Returns
     -------
@@ -111,6 +121,31 @@ def callaway_santanna(
         )
     if anticipation < 0:
         raise ValueError(f"anticipation must be >= 0, got {anticipation}")
+
+    # ---- Repeated cross-sections branch --------------------------------
+    if not panel:
+        if x:
+            raise NotImplementedError(
+                "panel=False (repeated cross-sections) does not yet support "
+                "covariates.  Pass x=None or run panel=True if your data are "
+                "balanced across units."
+            )
+        if estimator != 'reg':
+            raise NotImplementedError(
+                "panel=False currently only supports estimator='reg' "
+                "(unconditional 2×2 cell-mean DID).  IPW / DR for RCS are "
+                "planned for a future release."
+            )
+        if control_group != 'nevertreated':
+            raise NotImplementedError(
+                "panel=False currently requires "
+                "control_group='nevertreated'."
+            )
+        return _callaway_santanna_rcs(
+            data=data, y=y, g=g, t=t,
+            base_period=base_period, anticipation=anticipation,
+            alpha=alpha,
+        )
 
     # 1. Prepare panel data
     y_wide, unit_info, time_periods, cohorts, n_units = _prepare_panel(
@@ -652,3 +687,176 @@ def _pretrend_test(
     pvalue = float(1 - stats.chi2.cdf(W, k))
 
     return {'statistic': W, 'df': k, 'pvalue': pvalue}
+
+
+# ======================================================================
+# Repeated cross-sections (panel=False) branch
+# ======================================================================
+
+def _callaway_santanna_rcs(
+    data: pd.DataFrame,
+    y: str,
+    g: str,
+    t: str,
+    base_period: str,
+    anticipation: int,
+    alpha: float,
+) -> CausalResult:
+    """Unconditional 2×2 cell-mean DID for repeated cross-sections.
+
+    For each (g, t) pair with base period b:
+
+        ATT(g, t) = (Ȳ_{g,t} - Ȳ_{g,b}) - (Ȳ_{c,t} - Ȳ_{c,b})
+
+    where c = never-treated cohort.  Observation-level influence
+    functions are assembled as
[…]
