### 2026-04-15 53a03a8
docs: 0.7.1 CHANGELOG + README parity matrix updated for RCS + exports
Bring the externally-visible documentation up to the actual state of
the package after rounds 7–11 (commits `0d89afa` through `489b03b`).

CHANGELOG.md
  Add a 0.7.1 section (2026-04-14) covering:
  * repeated-cross-sections support in callaway_santanna() via
    `panel=False`, with inf_matrix contract preserved so every
    downstream consumer (aggte / cs_report / ggdid / honest_did)
    gets RCS for free;
  * CSReport.plot() 2×2 summary figure;
  * CSReport.to_markdown / .to_latex / .to_excel export methods
    (plus the 
[…]
