### 2026-04-15 73c40de
sp.did(aggregation=...) — one-call CS estimation + aggte in the top-level API
Teach the top-level `sp.did()` dispatcher to pipe a Callaway-Sant'Anna
result straight through `aggte()` when the caller passes a new
`aggregation=` argument (one of 'simple', 'dynamic', 'group',
'calendar'), so users no longer have to hand-compose:

    cs = callaway_santanna(...)
    es = aggte(cs, type='dynamic', n_boot=..., random_state=...)

Instead:

    es = sp.did(df, y=..., treat=..., time=..., id=...,
                aggregation='dynamic',
                n_boot=500, random_state=0)

Also surface `panel=` a
[…]
