### 2026-04-15 81ea955
RCS covariate support: regression-adjusted 2×2 cell-mean DID
Extend `callaway_santanna(panel=False, estimator='reg')` to accept a
non-empty `x=[...]` list of covariates.  Implementation:

1. Fit Y = X β + period-FE on the never-treated (G=0) pool using OLS.
2. Residualise every observation's Y by subtracting X·β̂ (period FEs
   are absorbed by the control group; we intentionally *keep* the
   remaining treated-cell mean in Y since that carries the ATT signal).
3. Run the existing unconditional 2×2 cell-mean RCS DID on the
   residualised Y, re-using the observation-level influence-function
   
[…]
