+++ tests/test_cs_rcs.py
@@ -101,14 +101,16 @@ def test_honest_did_on_rcs_result(rcs_result):
 # Input validation                                                            #
 # --------------------------------------------------------------------------- #
 
-def test_rcs_rejects_covariates():
+def test_rcs_accepts_covariates():
+    """Covariates are now supported in RCS mode via residualisation."""
     df = _rcs_panel(seed=1)
-    df['x1'] = 0.0
-    with pytest.raises(NotImplementedError, match='covariates'):
-        callaway_santanna(
-            df, y='y', g='g', t='t', i='obs',
-            x=['x1'], panel=False,
-        )
+    df['x1'] = 0.0  # constant covariate — residualisation should no-op
+    r = callaway_santanna(
+        df, y='y', g='g', t='t', i='obs',
+        x=['x1'], panel=False, estimator='reg',
+    )
+    # Constant covariate → residualisation leaves Y unchanged up to numerics.
+    assert r.model_info['covariates'] == ['x1']
 
 
 def test_rcs_rejects_non_reg_estimator():
@@ -165,3 +167,52 @@ def test_rcs_and_panel_agree_when_balanced():
     np.testing.assert_allclose(
         merged['att_p'], merged['att_r'], atol=1e-10,
     )
+
+
+# --------------------------------------------------------------------------- #
+# RCS with covariates (regression adjustment)                                 #
+# --------------------------------------------------------------------------- #
+
+def _rcs_panel_with_covariate(n_obs=2000, seed=42):
+    rng = np.random.default_rng(seed)
+    rows = []
+    for obs in range(n_obs):
+        g = int(rng.choice([3, 5, 7, 0], p=[0.2, 0.2, 0.2, 0.4]))
+        t = int(rng.integers(1, 9))
+        xx = rng.normal()
+        ui = rng.normal(scale=0.3)
+        te = max(0, t - g + 1) * 0.5 if g > 0 else 0
+        y = ui + 0.2 * t + 0.8 * xx + te + rng.normal()
+        rows.append({'obs': obs, 't': t, 'g': g, 'x1': xx, 'y': y})
+    return pd.DataFrame(rows)
+
+
+def test_rcs_with_covariate_runs():
+    df = _rcs_panel_with_covariate(seed=7)
+    r = callaway_santanna(df, y='y', g='g', t='t', i='obs',
+                          x=['x1'], panel=False, estimator='reg')
+    assert r.model_info['panel'] is False
+    assert 'covariates' in r.model_info['estimator']
+    assert r.model_info['covariates'] == ['x1']
+    assert r.estimate > 0 and r.pvalue < 0.01
+
+
+def test_rcs_covariate_reduces_se():
+    """Adding a predictive covariate should narrow the SE (variance
+    reduction from residualisation)."""
+    df = _rcs_panel_with_covariate(seed=11, n_obs=3000)
+    r_no = callaway_santanna(df, y='y', g='g', t='t', i='obs',
+                             panel=False, estimator='reg')
+    r_cov = callaway_santanna(df, y='y', g='g', t='t', i='obs',
+                              x=['x1'], panel=False, estimator='reg')
+    assert r_cov.se < r_no.se
+
+
+def test_rcs_covariate_downstream_aggte_works():
+    from statspai.did import aggte
+    df = _rcs_panel_with_covariate(seed=3)
+    r = callaway_santanna(df, y='y', g='g', t='t', i='obs',
+                          x=['x1'], panel=False, estimator='reg')
+    es = aggte(r, type='dynamic', n_boot=200, random_state=0)
+    post = es.detail[es.detail['relative_time'] >= 0]
+    assert (post['att'] > 0).all()
