+++ src/statspai/did/callaway_santanna.py
@@ -124,17 +124,11 @@ def callaway_santanna(
 
     # ---- Repeated cross-sections branch --------------------------------
     if not panel:
-        if x:
-            raise NotImplementedError(
-                "panel=False (repeated cross-sections) does not yet support "
-                "covariates.  Pass x=None or run panel=True if your data are "
-                "balanced across units."
-            )
         if estimator != 'reg':
             raise NotImplementedError(
                 "panel=False currently only supports estimator='reg' "
-                "(unconditional 2×2 cell-mean DID).  IPW / DR for RCS are "
-                "planned for a future release."
+                "(unconditional / covariate-adjusted 2×2 cell-mean DID).  "
+                "IPW / DR for RCS are planned for a future release."
             )
         if control_group != 'nevertreated':
             raise NotImplementedError(
@@ -142,7 +136,7 @@ def callaway_santanna(
                 "control_group='nevertreated'."
             )
         return _callaway_santanna_rcs(
-            data=data, y=y, g=g, t=t,
+            data=data, y=y, g=g, t=t, x=x,
             base_period=base_period, anticipation=anticipation,
             alpha=alpha,
         )
@@ -701,8 +695,9 @@ def _callaway_santanna_rcs(
     base_period: str,
     anticipation: int,
     alpha: float,
+    x: Optional[List[str]] = None,
 ) -> CausalResult:
-    """Unconditional 2×2 cell-mean DID for repeated cross-sections.
+    """Unconditional (or regression-adjusted) 2×2 cell-mean DID for RCS.
 
     For each (g, t) pair with base period b:
 
@@ -723,12 +718,29 @@ def _callaway_santanna_rcs(
     for col in (y, g, t):
         if col not in df.columns:
             raise ValueError(f"Column '{col}' not found in data")
+    if x:
+        for col in x:
+            if col not in df.columns:
+                raise ValueError(f"Covariate '{col}' not found in data")
 
     df[g] = df[g].fillna(0).replace([np.inf, -np.inf], 0).astype(int)
-    df = df.dropna(subset=[y, t]).reset_index(drop=True)
+    drop_cols = [y, t] + (list(x) if x else [])
+    df = df.dropna(subset=drop_cols).reset_index(drop=True)
     n_obs = len(df)
     if n_obs == 0:
         raise ValueError("No observations after dropping NaNs.")
+
+    # Covariate adjustment: residualise Y on X using the never-treated
+    # pool with period fixed effects. Plug-in influence functions treat
+    # β̂ as known (asymptotically valid; see Sant'Anna & Zhao 2020).
+    y_series = df[y].astype(float).to_numpy().copy()
+    covariate_info: Optional[Dict[str, Any]] = None
+    if x:
+        y_series = _rcs_residualise_on_controls(y_series, df, g, t, x)
+        covariate_info = {
+            'covariates': list(x),
+            'approach': 'residualisation on never-treated with period FEs',
+        }
 
     time_periods = sorted(df[t].unique())
     t_max = max(time_periods)
@@ -740,7 +752,7 @@ def _callaway_santanna_rcs(
     if not gt_pairs:
         raise ValueError("No valid (group, time) pairs to estimate.")
 
-    y_arr = df[y].values.astype(float)
+    y_arr = y_series  # possibly residualised
     g_arr = df[g].values
     t_arr = df[t].values
 
@@ -788,7 +800,7 @@ def _callaway_santanna_rcs(
     pretrend = _pretrend_test(detail, inf_matrix, n_obs)
 
     model_info: Dict[str, Any] = {
-        'estimator': 'REG (RCS)',
[…]
