### 2026-04-15 9344c20
BJS pre-trend joint Wald via cluster bootstrap (dCDH-style covariance)
The existing `did_imputation()` reports a sum-of-z² pre-trend Wald
that treats the pre-period τ̂(k) estimates as uncorrelated across k.
That is a valid but conservative approximation: in practice
consecutive pre-periods share data (the same unit-FE estimates enter
both imputations) and are strongly correlated.  With only the
diagonal covariance, the test misses real violations that show up
as non-zero off-diagonal structure.

New: `src/statspai/did/bjs_inference.py::bjs_pretrend_joint()`.

* Cluster-resamples the data `n_bo
[…]
