+++ src/statspai/did/bjs_inference.py
@@ -1,0 +1,170 @@
+"""Opt-in joint inference upgrades for the BJS imputation estimator.
+
+The core ``did_imputation()`` function reports a sum-of-z² pre-trend
+Wald statistic that assumes the pre-period ATT(k) estimates are
+uncorrelated across ``k``.  This is a valid but *conservative*
+approximation when consecutive pre-periods share data (they typically
+do, because the imputed counterfactual for both k = −1 and k = −2
+leans on the same unit-FE estimates).
+
+This module adds :func:`bjs_pretrend_joint`, a cluster-bootstrap
+estimator of the full pre-period covariance matrix.  The resulting
+joint Wald test is size-correct in finite samples even when the
+off-diagonal covariance is non-zero.
+
+References
+----------
+Borusyak, K., Jaravel, X. and Spiess, J. (2024).
+    "Revisiting Event-Study Designs: Robust and Efficient Estimation."
+    *Review of Economic Studies*, 91(6), 3253-3285.
+    Appendix discussion of joint pre-trend testing.
+"""
+
+from __future__ import annotations
+
+from typing import Any, Dict, List, Optional
+
+import numpy as np
+import pandas as pd
+from scipy import stats
+
+from ..core.results import CausalResult
+from .did_imputation import did_imputation
+
+
+def bjs_pretrend_joint(
+    result: CausalResult,
+    data: pd.DataFrame,
+    y: str,
+    group: str,
+    time: str,
+    first_treat: str,
+    controls: Optional[List[str]] = None,
+    cluster: Optional[str] = None,
+    horizon: Optional[List[int]] = None,
+    n_boot: int = 300,
+    seed: Optional[int] = None,
+) -> Dict[str, Any]:
+    """Cluster-bootstrap joint Wald test for BJS pre-treatment coefficients.
+
+    Parameters
+    ----------
+    result : CausalResult
+        Output of :func:`did_imputation` on ``data`` with a non-trivial
+        ``horizon`` that covers negative values.  Only its
+        ``model_info['event_study']`` frame is consulted, to look up
+        the observed pre-period point estimates that we re-test with
+        a covariance-aware statistic.
+    data, y, group, time, first_treat, controls, cluster :
+        Same arguments you passed to the original
+        :func:`did_imputation` call.  Needed to re-run BJS on each
+        cluster-bootstrap resample.
+    horizon : list of int, optional
+        If omitted, inferred from ``result.model_info['event_study']``.
+    n_boot : int, default 300
+        Cluster-bootstrap replications.  Clusters are sampled with
+        replacement; unit ids are reassigned in the resampled frame
+        so BJS refits cleanly.
+    seed : int, optional
+        RNG seed for reproducibility.
+
+    Returns
+    -------
+    dict
+        ``{'statistic', 'df', 'pvalue', 'method', 'n_boot',
+           'pre_cov'}`` where ``pre_cov`` is the bootstrap covariance
+        matrix used for the Wald quadratic form.
+
+    Notes
+    -----
+    Cost:  ``n_boot`` full BJS re-fits.  On a 10 000-row balanced
+    panel with |horizon|=10, expect roughly
+    ``n_boot × 0.3 s = 90 s`` for the default ``n_boot=300`` — the
+    function is therefore opt-in, not run by default inside
+    :func:`did_imputation`.
+    """
+    es = result.model_info.get('event_study') if result.model_info else None
+    if es is None or len(es) == 0:
+        raise ValueError(
+            "result is missing an event-study table.  Re-run "
+            "did_imputation() with a non-empty `horizon=` argument."
+        )
+
+    if horizon is None:
[…]
