### 2026-04-15 2a9c135
Pinned numerical-fixture tests for CS + aggte (regression detection)
Lock the current numerical outputs of callaway_santanna() and
aggte(type='dynamic'/'simple') to tight tolerances so that any
future refactor which silently changes an ATT estimate or a
bootstrap-derived standard error fails loudly instead of drifting
unnoticed.

Approach:
  * fixed-seed 60-unit × 8-period staggered DGP (`_fixture_panel()`)
  * 21 pinned ATT(g, t) point estimates + cluster-level SEs
  * 11 pinned dynamic-aggregation (event study) rows
  * pinned overall ATT + SE
  * pytest.approx with abs=1e-4 on point estimat
[…]
