### 2026-04-15 6b2fdd3
Add mkdocs-material docs site scaffold + v0.7.x DiD guides
Introduce a proper documentation site for StatsPAI using mkdocs +
mkdocs-material, with a focused v0.7.x pass over the DiD module.
Build verified locally: `mkdocs build --clean --strict` succeeds in
~0.2 s with zero warnings.

* `mkdocs.yml` — site config (theme, nav, Pygments + MathJax extras).
* `docs/index.md` — project-level landing page; lead code block is the
  one-call `sp.cs_report(..., save_to=...)` pipeline.
* `docs/guides/callaway_santanna.md` — from-zero walkthrough of the
  CS2021 estimator: raw estimation, aggte with unif
[…]
