### 2026-04-15 56ef82a
fix(tests): write_text encoding='utf-8' for Windows CI (χ² in report)
Windows cp1252 default codec cannot encode U+03C7 (χ) emitted by
rpt.to_markdown()/to_latex(). Linux/macOS default to UTF-8 so the bug
was silent there. Explicit encoding='utf-8' fixes the smoke test on
windows-latest × py3.11/3.12.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
