### 2026-04-15 d03fbf3
Audit-round polish: 5 real bugs found in the v0.7.1 work, all fixed
Full audit of the 17 commits landed today surfaced five genuine issues
plus a dozen purely cosmetic ones.  This commit fixes the real five
and adds targeted regression tests for each.

1. `_df_to_booktabs` LaTeX escape was non-idempotent.
   Old code replaced `\` with `\textbackslash{}` first, then replaced
   `{` → `\{` and `}` → `\}` in a second pass — which re-escaped the
   braces just inserted, producing `\textbackslash\{\}` instead of the
   correct `\textbackslash{}`.  Same failure mode for `~` and `^`.
   Fix: single-p
[…]
