### 2026-04-15 e5d68ec
Audit round 2: 3 deeper semantic / UX bugs found and fixed
Second-pass audit of today's work — focused on algorithmic semantics
rather than mechanics.  Three genuine issues, all now fixed with
regression tests.

1. **aggte(type='dynamic').estimate averaged pre + post event times.**
   R `did::aggte(type='dynamic')` reports the overall ATT as a simple
   average of *post-treatment* event-study estimates (e ≥ 0).  Pre-
   treatment cells are placebos — averaging them into the headline
   number pollutes the causal summary with pre-trend signal.

   On the canonical demo DGP the bug produced over
[…]
