### 2026-04-15 086e597
Audit round 3: aggte() cryptic KeyError on non-CS input
Third-round audit focused on boundary conditions, data-flow edge
cases, and cross-module API consistency.  Only one genuine new
issue surfaced; everything else (Mammen math, RCS IF normalisation,
breakdown_m formula, ggdid ordering, save_to path handling, BJS
Optional import) verified correct on close re-reading.

Issue: `aggte(sa_result)` crashed with `KeyError: 'group'` deep
inside `_weights_dynamic`, for the same reason as the round-2
`cs_report(sa_result)` bug — Sun-Abraham / BJS / dCDH event-study
frames carry `relative_time` but no 
[…]
