### 2026-04-15 b76fdb7
fix(tests): expand tilde test — set USERPROFILE for Windows CI
ntpath.expanduser uses USERPROFILE on Windows; HOME is ignored.
Setting both keeps the test portable across POSIX and Windows runners.
