### 2026-04-15 953febf
chore(release): bump version to 0.7.0
Consolidates the v0.7.0 and previous v0.7.1 CHANGELOG sections into a
single v0.7.0 release, matching the bumped version in pyproject.toml
and the package __version__.  No code changes.

Contents of this release (summary):
  * `sp.callaway_santanna` with DR/IPW/REG, anticipation, and
    panel=False for repeated cross-sections (with optional
    covariate residualisation).
  * `sp.aggte` with four aggregations and Mammen multiplier-bootstrap
    uniform confidence bands.
  * `sp.cs_report` one-call report card + `CSReport.plot/.to_text/
    .to_markdown/.t
[…]
