+++ CHANGELOG.md
@@ -2,40 +2,67 @@
 
 All notable changes to StatsPAI will be documented in this file.
 
-## [0.7.1] - 2026-04-14
+## [0.7.0] - 2026-04-14
 
-Follow-up release to 0.7.0: export surface, visualisations, and the
-last remaining capability gap vs `csdid` / `differences` — repeated
-cross-sections.  Full DiD test suite: 114 → 142 (+28).
+Focused release reaching feature parity with the R `did` / `HonestDiD`
+packages and the Python `csdid` / `differences` packages for staggered
+Difference-in-Differences.  All core algorithms are reimplemented from
+the original papers — **no wrappers, no runtime dependencies on upstream
+DID packages**.  Full DiD test suite: 47 → 170+ (including three rounds
+of post-implementation audit that surfaced and fixed 9 bugs before
+release).
 
-### Added
+### Added — Core estimation
 
-- **Repeated cross-sections** support for `callaway_santanna()` via
-  the new `panel=False` argument.  Uses the unconditional 2×2
-  cell-mean DID with observation-level influence functions
-  (CS2021 eq. 2.4, RCS version), so `aggte` / `cs_report` / `ggdid`
-  / `honest_did` all work on RCS results with no code changes —
-  Mammen uniform bands, four aggregations, and Rambachan–Roth
-  breakdown M\* all compose downstream.  Scope of the initial
-  implementation: `estimator='reg'`, `control_group='nevertreated'`,
-  no covariates (other paths raise `NotImplementedError` with an
-  actionable message).
+- **`sp.aggte(result, type=...)`** — unified aggregation layer for
+  `callaway_santanna()` results.  Four aggregation schemes (`simple`,
+  `dynamic`, `group`, `calendar`) backed by a single weighted-
+  influence-function engine.  Callaway & Sant'Anna (2021) Section 4.
+- **Mammen (1993) multiplier bootstrap** — IQR-rescaled pointwise
+  standard errors *and* simultaneous (uniform / sup-t) confidence
+  bands over the aggregation dimension.  Matches the uniform-band
+  behaviour of the R `did::aggte` function.
+- **`balance_e` / `min_e` / `max_e`** — event-study cohort balancing
+  and window truncation (CS2021 eq. 3.8).
+- **`anticipation=δ`** parameter on `callaway_santanna()` — shifts
+  the base period back by δ periods per CS2021 §3.2.
+- **Repeated cross-sections** support via `callaway_santanna(panel=False)`
+  — unconditional 2×2 cell-mean DID with observation-level influence
+  functions (CS2021 eq. 2.4, RCS version).  Optional covariate
+  residualisation with `x=[...]` for regression adjustment.  All
+  downstream modules (`aggte`, `cs_report`, `ggdid`, `honest_did`)
+  work on RCS results with no code changes.
+- **dCDH joint inference** (`did_multiplegt`) — `joint_placebo_test`
+  (Wald χ² across placebo lags with bootstrap covariance, dCDH 2024
+  §3.3) and `avg_cumulative_effect` (mean of dynamic[0..L] with
+  SE preserving cross-horizon covariance, dCDH 2024 §3.4).
+- **`sp.bjs_pretrend_joint()`** — cluster-bootstrap joint Wald pre-
+  trend test for BJS imputation results.  Upgrades the default
+  sum-of-z² test (which assumes pre-period independence) to a full
+  covariance-aware statistic.
+
+### Added — Reporting & visualisation
+
[…]
