### 2026-04-15 dc040b0
docs: fix broken image and CN link on PyPI page
Use absolute GitHub raw/blob URLs so PyPI's isolated README renderer
can resolve the screenshot and the English/Chinese language switcher.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
