+++ README.md
@@ -1,4 +1,4 @@
-[English](README.md) | [中文](README_CN.md)
+[English](https://github.com/brycewang-stanford/statspai/blob/main/README.md) | [中文](https://github.com/brycewang-stanford/statspai/blob/main/README_CN.md)
 
 # StatsPAI: The Agent-Native Causal Inference & Econometrics Toolkit for Python
 
@@ -17,7 +17,7 @@ It brings R's [Causal Inference Task View](https://cran.r-project.org/web/views/
 
 **NEW in v0.6**: `sp.interactive(fig)` — a Stata Graph Editor-style WYSIWYG plot editor for Jupyter, with 29 academic themes, real-time preview, and auto-generated reproducible code.
 
-![alt text](image-1.png)
+![StatsPAI interactive plot editor](https://raw.githubusercontent.com/brycewang-stanford/statspai/main/image-1.png)
 
 > Built by the team behind [CoPaper.AI](https://copaper.ai) · Stanford REAP Program
 
