### 2026-04-15 2194ad2
docs: add StatsPAI ecosystem gap analysis & v1.0 roadmap
Top-level design anchor for the v0.7.0 → v1.0 upgrade cycle. Catalogues
54 modules against R + Python reference ecosystems across 24 domains,
flags three P0 fronts (spatial, ML-causal deepening, infrastructure),
and decomposes the upgrade into 10 sub-projects spanning 5 sprints.

Each sub-project will receive its own spec and plan before any code is
written. SP-01 (spatial econometrics full-stack) enters brainstorming
next.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
