+++ specs/2026-04-15-statspai-ecosystem-gap-analysis.md
@@ -1,0 +1,105 @@
+# StatsPAI Ecosystem Gap Analysis & Roadmap to v1.0
+
+**Date**: 2026-04-15
+**Status**: Top-level design — anchor for subsequent sub-specs
+**Baseline**: StatsPAI v0.7.0 (54 modules, ~85k LOC, 250+ public functions)
+
+---
+
+## 1. Purpose
+
+This document is the top-level design anchor for a comprehensive upgrade of StatsPAI
+to reach parity with — and in several dimensions exceed — the mainstream R and
+Python ecosystems for applied econometrics and causal inference. It catalogues
+the current capability, maps each domain to the reference ecosystem, quantifies
+the gap, and decomposes the upgrade into ten independent sub-projects.
+
+Each sub-project below will have its own spec (`YYYY-MM-DD-sp-NN-<topic>-design.md`)
+and its own implementation plan before any code is written.
+
+## 2. Reference Ecosystem (Benchmark)
+
+### R (mainstream)
+`stats`, `fixest`, `estimatr`, `did`, `didimputation`, `DIDmultiplegt`, `honestDiD`,
+`rdrobust`, `rddensity`, `rdlocrand`, `rdpower`, `Synth`, `augsynth`, `synthdid`,
+`gsynth`, `MatchIt`, `cobalt`, `WeightIt`, `ebal`, `plm`, `grf`, `DoubleML`,
+`mediation`, `EValue`, `sensemakr`, `spatialreg`, `spdep`, `splm`, `sphet`,
+`GWmodel`, `vars`, `urca`, `rugarch`, `bvartools`, `survival`, `flexsurv`,
+`pcalg`, `bnlearn`, `MendelianRandomization`, `TwoSampleMR`, `policytree`,
+`survey`, `frontier`, `mice`, `Amelia`, `randomizr`, `DeclareDesign`.
+
+### Python (mainstream)
+`statsmodels`, `linearmodels`, `pyfixest`, `EconML`, `CausalML`, `DoubleML`,
+`dowhy`, `causal-learn`, `lingam`, `causalnex`, `PySAL` (`libpysal`, `esda`,
+`spreg`, `mgwr`, `splot`), `pyblp`, `lifelines`, `scikit-survival`, `mapie`,
+`differences`, `pysyncon`, `SparseSC`.
+
+## 3. Gap Matrix (24 domains)
+
+Legend: 🟢 parity / 🟡 usable but gap / 🟠 large gap / 🔴 missing or placeholder
+
+| # | Domain | StatsPAI Current | R Benchmark | Python Benchmark | Gap | Priority |
+|---|---|---|---|---|---|---|
+| 1 | Linear / IV basics | `regress`, `iv`, `advanced_iv` | fixest, estimatr, ivreg | linearmodels, pyfixest | No OOS `predict()`; HDFE speed behind fixest | 🟡 **P0** |
+| 2 | GLM / discrete choice | logit/probit/poisson/zip/mlogit/fracreg | glm, VGAM, mlogit | statsmodels | Some GLM families raise `NotImplementedError` | 🟡 **P1** |
+| 3 | DID (static + staggered) | did/CS/SA/bacon/honest/sdid/DDD/imputation/stacked/continuous | did, didimputation, DIDmultiplegt, honestDiD, fixest::sunab | differences, pyfixest | 🟢 **industry-leading**; only missing explicit ETWFE (Wooldridge 2021) API | 🟢 **P2** |
+| 4 | RD / RKD / RDD | rdrobust/rdplot/rdbw/rkd/rd_honest/rdmc/rdms | rdrobust, rddensity, **rdlocrand**, **rdpower**, rdmulti | rdrobust-py | Missing `rdlocrand` (local randomization) and `rdpower` (power) | 🟡 **P1** |
[…]
