### 2026-04-15 36ae5bc
feat(did): add explicit etwfe() API aligned with R etwfe package
etwfe() is a thin alias of wooldridge_did() that matches the naming
convention of Grant McDermott's R 'etwfe' package. Same saturated TWFE
estimator underneath; the new name improves discoverability for users
coming from the R ecosystem and makes the Wooldridge (2021) connection
explicit in the public API.

Docstring includes argument mapping table (R etwfe -> sp.etwfe) and
notes that xvar (covariate heterogeneity) and emfx aggregations are
not yet mirrored.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
