+++ src/statspai/did/__init__.py
@@ -45,7 +45,7 @@ from .pretrends import (
     SensitivityResult,
     pretrends_summary,
 )
-from .wooldridge_did import wooldridge_did, drdid, twfe_decomposition
+from .wooldridge_did import wooldridge_did, etwfe, drdid, twfe_decomposition
 from .continuous_did import continuous_did
 from .plots import (
     parallel_trends_plot,
@@ -384,6 +384,7 @@ __all__ = [
     'cic',
     # Wooldridge / DR-DID / TWFE decomposition
     'wooldridge_did',
+    'etwfe',
     'drdid',
     'twfe_decomposition',
     # Pre-trends
