### 2026-04-15 4dc37c1
docs(spec): SP-01 spatial econometrics full-stack design
Three-phase design to close domain #16 of the ecosystem gap:
  S1.1 weights + ESDA + cross-section models (ML sparse + GMM + impacts)
  S1.2 GWR + MGWR
  S1.3 spatial panel

Pins key decisions: sparse-first W, dual log-det path, soft deps on
geopandas/libpysal, interop not wrap, LeSage-Pace impacts default,
<1e-4 cross-validation against PySAL + R spatialreg.

Backward compatibility preserved — sp.sar/sem/sdm keep current signature
while internals route to the sparse solver.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropi
[…]
