+++ specs/2026-04-15-sp-01-spatial-full-stack-design.md
@@ -1,0 +1,231 @@
+# SP-01: Spatial Econometrics Full-Stack — Design Spec
+
+**Date**: 2026-04-15
+**Parent**: [ecosystem gap analysis](2026-04-15-statspai-ecosystem-gap-analysis.md)
+**Scope**: Domain #16 of the gap matrix — close the largest current gap in StatsPAI.
+**Sprint**: S1 (sub-phases S1.1 / S1.2 / S1.3 — see §7)
+**Status**: Design approved — ready for `writing-plans`.
+
+## 1. Why this sub-project
+
+The existing `statspai.spatial` module ships only three ML estimators
+(SAR / SEM / SDM, dense matrices, 438 LOC). The R side ships
+`spatialreg + spdep + splm + sphet + GWmodel` (~60 functions). The Python
+side ships PySAL (`libpysal + esda + spreg + mgwr + splot`, ~100 public
+functions). The gap is the largest in StatsPAI, and the existing dense
+implementation does not scale beyond N ≈ 2000.
+
+Closing this gap makes StatsPAI the first unified Python package where a
+user can go weights → ESDA → regression → impacts → GWR → spatial panel
+without touching `libpysal`, `spreg`, or `mgwr` directly, while still
+interoperating with them.
+
+## 2. Success criteria
+
+A user can do each of the following with `import statspai as sp`:
+
+1. Build spatial weights from a `GeoDataFrame` or coordinate array
+   (queen, rook, KNN, distance-band, kernel, block) and row-normalise,
+   binary-transform, or variance-stabilise them.
+2. Run Moran's I (global and local), Geary's C, Getis-Ord G, and join
+   counts, with permutation-based inference and publication-ready plots.
+3. Estimate SAR / SEM / SDM / SLX / SAC models by ML (small-to-moderate N)
+   or Kelejian-Prucha GMM (large N, heteroscedasticity-robust), with
+   LM / Robust-LM diagnostics for model selection.
+4. Recover direct / indirect / total impacts with simulated standard
+   errors (LeSage-Pace 2009).
+5. Fit Geographically Weighted Regression (GWR) and Multiscale GWR
+   (MGWR) with cross-validation or AICc bandwidth selection.
+6. Fit spatial panel models (FE × {SAR, SEM, SDM}) with Baltagi-style
+   inference.
+
+Numerical tolerance vs PySAL / R `spatialreg` on the Columbus and Boston
+housing datasets: **< 1e-4 relative error on all estimated parameters**.
+
+## 3. Architecture
+
+```
+src/statspai/spatial/
+├─ __init__.py            # re-exports flat API (queen_weights, moran, sar, ...)
+├─ weights/
+│   ├─ core.py            # W class (sparse CSR backing)
+│   ├─ contiguity.py      # queen, rook (needs geopandas)
+│   ├─ distance.py        # knn, distance_band, kernel
+│   ├─ block.py           # block / regime weights
+│   └─ transform.py       # row / binary / variance-stabilise transforms
+├─ esda/
+│   ├─ moran.py           # global + local
+│   ├─ geary.py
+│   ├─ getis_ord.py       # G and G* (global + local)
+│   ├─ join_counts.py
+│   └─ plots.py           # moran_plot, lisa_cluster_map
+├─ models/
+│   ├─ base.py            # shared estimation machinery
+│   ├─ ml.py              # SAR/SEM/SDM/SLX/SAC via ML (concentrated LL)
+│   ├─ gmm.py             # Kelejian-Prucha GMM
[…]
