### 2026-04-15 a47634a
feat: expose pyfixest HDFE wrappers (feols/fepois/feglm/etable) at sp.*
The fixest/ module already wraps pyfixest for high-dimensional fixed
effects estimation, but the functions were only reachable via
statspai.fixest.feols — users reasonably expect sp.feols to work.

Top-level import is safe: fixest/wrapper.py defers the pyfixest import
until first call, so statspai still imports without pyfixest installed;
users just get a clear ImportError with install instructions when they
invoke sp.feols.

This unlocks the fastest-known Python path for two/three-way FE
regressions in DID-heavy workflows
[…]
