### 2026-04-15 70dccaf
docs(readme): add 'What StatsPAI is — and is not' positioning block
Preempts the 'isn't this just a wrapper?' critique with a concrete
statement of what StatsPAI contributes beyond R/Stata/existing Python:

- Independent re-implementation + citation transparency (not code copy)
- Unified CausalResult across 390+ functions (R has 20+ S3 classes)
- Scope breadth no single package matches (DID+RD+Synth+DML+Neural+...)
- Agent-native schemas (list_functions/describe_function/function_schema)
- Publication pipeline built in (Word/Excel/LaTeX/HTML/Markdown)

Also surfaces the flagship DID estimators
[…]
