### 2026-04-15 9c72f71
docs(plan): SP-01 S1.1 spatial core implementation plan
23 TDD-structured tasks across 7 phases (weights → ESDA → log-det →
ML models → GMM → diagnostics/impacts → integration). Each task
includes exact files, failing test, implementation, verification,
commit. Acceptance gate is Columbus cross-validation vs spreg at
rtol=1e-4.

Self-review confirms 24/24 spec API surface items covered; D1-D8
decisions land in specific tasks; backward-compat is enforced by
saved baseline fixture + regression test.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
