+++ plans/2026-04-15-sp-01-s1.1-spatial-core.md
@@ -1,0 +1,1867 @@
+# SP-01 S1.1 — Spatial Core Implementation Plan
+
+> **For agentic workers:** REQUIRED SUB-SKILL: Use `superpowers:subagent-driven-development` (recommended) or `superpowers:executing-plans` to implement this plan task-by-task. Steps use checkbox (`- [ ]`) syntax for tracking.
+
+**Goal:** Close the largest ecosystem gap in StatsPAI by shipping a sparse-backed spatial weights layer, the full PySAL-equivalent ESDA suite, spatial regression models (ML + GMM) with LeSage-Pace impacts, and LM diagnostics — all interoperable with libpysal but independently implemented.
+
+**Architecture:** Three-layer module under `src/statspai/spatial/`. Layer 1 (`weights/`) owns a sparse CSR-backed `W` class and six constructors. Layer 2 (`esda/`) owns six permutation-based spatial statistics. Layer 3 (`models/`) owns ML (concentrated likelihood, dual log-det path) and GMM (Kelejian-Prucha) estimators, plus LM tests and impacts. Existing `sp.sar/sem/sdm` are preserved as thin forwarders; backward-compat is enforced by a regression test against saved outputs.
+
+**Tech Stack:** numpy, scipy.sparse, scipy.stats, scipy.optimize, pandas, formulaic (already a dep via existing spatial module). Soft deps: geopandas, libpysal (lazy-imported, test-skipped if absent), matplotlib (for plots).
+
+**Parent spec:** [`specs/2026-04-15-sp-01-spatial-full-stack-design.md`](../specs/2026-04-15-sp-01-spatial-full-stack-design.md)
+
+---
+
+## File structure
+
+```
+src/statspai/spatial/
+├─ __init__.py                  # flat re-exports
+├─ models.py                    # EXISTING — becomes thin shim after refactor
+├─ weights/
+│   ├─ __init__.py
+│   ├─ core.py                  # class W
+│   ├─ contiguity.py            # queen_weights, rook_weights
+│   ├─ distance.py              # knn_weights, distance_band, kernel_weights
+│   └─ block.py                 # block_weights
+├─ esda/
+│   ├─ __init__.py
+│   ├─ _base.py                 # SpatialStatistic result class, permutation helper
+│   ├─ moran.py                 # moran, moran_local
+│   ├─ geary.py
+│   ├─ getis_ord.py             # getis_ord_g, getis_ord_local
+│   ├─ join_counts.py
+│   └─ plots.py                 # moran_plot, lisa_cluster_map
+├─ models/
+│   ├─ __init__.py
+│   ├─ _logdet.py               # exact + Barry-Pace approx
+│   ├─ _base.py                 # shared design-matrix + result plumbing
+│   ├─ ml.py                    # sar, sem, sdm, slx, sac (ML)
+│   ├─ gmm.py                   # sar_gmm, sem_gmm, sarar_gmm
+│   ├─ diagnostics.py           # lm_tests, moran_residuals
+│   └─ impacts.py               # LeSage-Pace direct/indirect/total
+
+tests/spatial/
+├─ __init__.py
+├─ fixtures/
+│   ├─ columbus.csv             # dataset copy (public)
+│   ├─ columbus_shp/            # geometries
+│   └─ reference_outputs.json   # saved PySAL / spatialreg numbers
+├─ test_weights_core.py
+├─ test_weights_contiguity.py
+├─ test_weights_distance.py
+├─ test_weights_block.py
+├─ test_esda_moran.py
+├─ test_esda_geary.py
+├─ test_esda_getis_ord.py
+├─ test_esda_join_counts.py
+├─ test_models_ml.py
[…]
