### 2026-04-15 2b263be
docs(readme_cn): sync positioning block and flagship DID/HDFE entries
Mirrors the 'What StatsPAI is — and is not' positioning block added to
the English README in 70dccaf, and fills the same function-table gaps:

- DID table: +etwfe, drdid, stacked_did, ddd, cic, twfe_decomposition,
  wooldridge_did
- Regression table: +feols, fepois, feglm, etable (pyfixest backend)

Keeps CN/EN READMEs aligned so Chinese users see the same narrative
and same feature surface. No code changes.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
