+++ README_CN.md
@@ -40,6 +40,19 @@ StatsPAI 是一个**面向 AI Agent** 的 Python 因果推断与应用计量经�
 
 ---
 
+## StatsPAI 是什么，不是什么
+
+StatsPAI **不是** R 的 wrapper。我们从原始论文独立重新实现每一个算法（通过 `.cite()` 暴露引用），少数成熟引擎（pyfixest、rdrobust）使用显式透明的绑定。StatsPAI 的真正差异化在于**上层的统一架构**：
+
+- **一个结果对象，一套 API。** 从 `regress()` 到 `callaway_santanna()` 到 `causal_forest()` 到 `notears()`，所有估计器返回同样的 `CausalResult`，共享 `.summary()` / `.plot()` / `.to_latex()` / `.cite()` 接口。R 用户要记 20+ 个互不兼容的 S3 类，StatsPAI 用户只需要记一个。
+- **单个 R / Python 包无法企及的广度。** DID + RD + Synth + Matching + DML + Meta-learners + TMLE + Neural Causal + Causal Discovery + Policy Learning + Conformal + Bunching + Spillover + Matrix Completion —— 全部风格统一，全部在 `sp.*` 命名空间下。
+- **Agent-native 原生设计。** 自描述 schema（`list_functions()`、`describe_function()`、`function_schema()`）让 StatsPAI 成为**首个为 LLM 驱动研究工作流而构建**的计量工具包——任何语言的包都没有这个能力。
+- **论文级输出流水线开箱即用。** 每个估计器都直接支持 Word + Excel + LaTeX + HTML + Markdown 导出，不需要额外跳一段 `modelsummary` 风格的舞。
+
+**原则**：R 里有的方法，我们在 Python 里做到同等或更强的功能覆盖；然后叠加 Python 独有的优势——sklearn 生态集成、JAX / PyTorch 后端、agent-native schema。
+
+---
+
 ## 完整功能列表
 
 ### 回归模型
@@ -59,6 +72,10 @@ StatsPAI 是一个**面向 AI Agent** 的 Python 因果推断与应用计量经�
 | `ologit()`, `oprobit()` | 有序 Logit / Probit | `ologit` / `oprobit` | `MASS::polr()` |
 | `poisson()`, `nbreg()` | 计数模型（泊松、负二项） | `poisson` / `nbreg` | `MASS::glm.nb()` |
 | `ppmlhdfe()` | 引力模型伪泊松 MLE | `ppmlhdfe` | `fixest::fepois()` |
+| `feols()` | OLS / IV，高维固定效应（pyfixest 后端） | `reghdfe` | `fixest::feols()` |
+| `fepois()` | 高维固定效应泊松 | `ppmlhdfe` | `fixest::fepois()` |
+| `feglm()` | 高维固定效应 GLM | — | `fixest::feglm()` |
+| `etable()` | 出版级回归表（LaTeX / Markdown / HTML） | `esttab` | `fixest::etable()` |
 | `gmm()` | 一般 GMM（任意矩条件） | `gmm` | `gmm::gmm()` |
 | `frontier()` | 随机前沿分析 | `frontier` | `sfa::sfa()` |
 
@@ -73,6 +90,12 @@ StatsPAI 是一个**面向 AI Agent** 的 Python 因果推断与应用计量经�
 | `honest_did()` | 平行趋势假设敏感性 | Rambachan & Roth (2023) |
 | `continuous_did()` | 连续处理 DID（剂量反应） | Callaway, Goodman-Bacon & Sant'Anna (2024) |
 | `did_imputation()` | 插补 DID 估计量 | Borusyak, Jaravel & Spiess (2024) |
[…]
