### 2026-04-15 c9ceb72
docs: add R-to-StatsPAI migration cheatsheet
New guide at docs/guides/migration-from-r.md mapping R causal-inference
packages to StatsPAI's unified sp.* API, covering:

- fixest (feols/fepois/feglm/etable)
- did, DIDmultiplegt, didimputation, HonestDiD, pretrends, bacondecomp
- rdrobust, rddensity, rdmulti, rdhonest
- Synth, gsynth, augsynth, synthdid
- MatchIt, ebal, cobalt
- AER, ivmodel, hdm, MendelianRandomization
- DoubleML, grf, SuperLearner, policytree, causalTree
- pcalg, EValue, specr

All 63 sp.* references and 5 class imports in the doc are verified
against the installed package (no
[…]
