+++ docs/guides/migration-from-r.md
@@ -1,0 +1,195 @@
+# Migrating from R to StatsPAI
+
+A practical one-page map for researchers moving from R's causal inference
+ecosystem to StatsPAI. Every R function listed here has an independent
+Python re-implementation in StatsPAI, following the same statistical
+methodology but exposed through a unified `sp.*` API.
+
+```python
+import statspai as sp
+```
+
+---
+
+## Regression & Fixed Effects (`fixest`)
+
+| R (`fixest`)                              | StatsPAI                                                          | Notes                                                        |
+| ----------------------------------------- | ----------------------------------------------------------------- | ------------------------------------------------------------ |
+| `feols(y ~ x1 + x2, data = df)`           | `sp.feols("y ~ x1 + x2", data=df)`                                | pyfixest-backed; same formula syntax                         |
+| `feols(y ~ x \| firm + year, data = df)`  | `sp.feols("y ~ x \| firm + year", data=df)`                       | Two-way fixed effects                                        |
+| `fepois(y ~ x \| firm, data = df)`        | `sp.fepois("y ~ x \| firm", data=df)`                             | HDFE Poisson / PPML for gravity models                       |
+| `feglm(..., family = "binomial")`         | `sp.feglm(..., family="binomial")`                                | HDFE GLM                                                     |
+| `etable(m1, m2, m3)`                      | `sp.etable([m1, m2, m3])`                                         | Publication-quality regression tables (LaTeX / MD / HTML)    |
+| `vcov(m, cluster = ~firm)`                | `vcov={"CRV1": "firm"}` in `sp.feols`                             | Cluster-robust SE                                            |
+| `fixef(m)`                                | `m.fixef()` (on pyfixest result)                                  | Extract fixed-effect estimates                               |
+
+For plain OLS without HDFE, `sp.regress("y ~ x", data=df)` gives a statsmodels-
+compatible interface returning a `CausalResult`.
+
+---
+
+## Staggered DID (`did`, `fixest::sunab`, `didimputation`, `DIDmultiplegt`)
+
+| R                                                       | StatsPAI                                                 |
+| ------------------------------------------------------- | -------------------------------------------------------- |
+| `did::att_gt(yname, tname, gname, idname, data)`        | `sp.callaway_santanna(data, y=, time=, first_treat=, group=)` |
+| `did::aggte(obj, type = "dynamic")`                     | `sp.aggte(cs_result, type="dynamic")`                    |
+| `did::aggte(obj, type = "group")`                       | `sp.aggte(cs_result, type="group")`                      |
+| `did::ggdid(agg)`                                       | `sp.ggdid(agg_result)`                                   |
+| `fixest::sunab(cohort, period)` inside `feols`          | `sp.sun_abraham(data, y=, time=, first_treat=, group=)`  |
+| `didimputation::did_imputation(...)`                    | `sp.did_imputation(data, y=, time=, first_treat=, group=)` |
+| `DIDmultiplegt::did_multiplegt(...)`                    | `sp.did_multiplegt(data, y=, group=, time=, treatment=)` |
+| `DIDmultiplegt::did_multiplegt_dyn(...)`                | `sp.did_multiplegt(..., dynamic=True)`                   |
+| `bacondecomp::bacon(...)`                               | `sp.bacon_decomposition(data, y=, time=, treat=, id=)`   |
+| —                                                       | `sp.etwfe(...)` — Wooldridge (2021) explicit API         |
[…]
