### 2026-04-15 fd97d56
feat(spatial): add sparse-backed W weights class
First brick of the SP-01 S1.1 spatial core refactor. Adds a
CSR-sparse-backed W class with neighbours dict construction,
transforms (B/R/V/D), islands detection, and libpysal interop
via .to_libpysal() / .from_libpysal().

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
