+++ src/statspai/spatial/weights/__init__.py
@@ -1,0 +1,3 @@
+from .core import W
+
+__all__ = ["W"]
