### 2026-04-15 900f393
feat(spatial): knn / distance-band / kernel weights
Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
