+++ src/statspai/spatial/weights/__init__.py
@@ -1,3 +1,4 @@
 from .core import W
+from .distance import knn_weights, distance_band, kernel_weights
 
-__all__ = ["W"]
+__all__ = ["W", "knn_weights", "distance_band", "kernel_weights"]
