### 2026-04-15 b261f6e
feat(spatial): queen/rook contiguity (soft geopandas dep)
Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
