### 2026-04-15 1954403
feat(spatial): block (regime) weights
Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
