+++ src/statspai/spatial/weights/__init__.py
@@ -1,8 +1,9 @@
 from .core import W
 from .distance import knn_weights, distance_band, kernel_weights
 from .contiguity import queen_weights, rook_weights
+from .block import block_weights
 
 __all__ = [
     "W", "knn_weights", "distance_band", "kernel_weights",
-    "queen_weights", "rook_weights",
+    "queen_weights", "rook_weights", "block_weights",
 ]
