### 2026-04-15 194449d
feat(weak-iv): Olea-Pflueger robust F + Lee et al. (2022) tF procedure
Replaces the placeholder 'effective F' stub in anderson_rubin_test
(which was just aliasing the first-stage F) with a proper implementation
of the weak-IV robust toolkit:

1. effective_f_test() — Olea-Pflueger (2013) heteroskedasticity-robust
   effective F, following Andrews-Stock-Sun (2019, Annual Review)
   formula. Supports 'classic', 'HC0', 'HC1' variance estimators.
   Verified: under homoskedasticity + single instrument, F_eff equals
   the standard first-stage F exactly (tolerance 1e-8).

2. tF_critical_value() — Le
[…]
