+++ src/statspai/diagnostics/weak_iv.py
@@ -4,6 +4,17 @@ Modern weak instrument diagnostics.
 Provides tests that remain valid even when instruments are weak,
 addressing the limitations of the standard first-stage F-statistic.
 
+Public API
+----------
+- ``anderson_rubin_test`` — Anderson-Rubin (1949) weak-IV-robust test
+  and confidence set (by grid inversion).
+- ``effective_f_test`` — Olea-Pflueger (2013) robust effective F with
+  heteroskedasticity-robust variance. Reduces exactly to first-stage F
+  under homoskedasticity and a single instrument.
+- ``tF_critical_value`` — Lee et al. (2022, AER) tF adjusted critical
+  value as a function of first-stage F; produces weak-IV-robust CIs
+  that use the 2SLS t-ratio directly.
+
 References
 ----------
 Anderson, T.W. and Rubin, H. (1949).
@@ -23,14 +34,301 @@ Olea, J.L.M. and Pflueger, C. (2013).
 Lee, D.S., McCrary, J., Moreira, M.J. and Porter, J. (2022).
 "Valid t-ratio Inference for IV."
 *American Economic Review*, 112(10), 3260-3290.
+
+Andrews, I., Stock, J.H. and Sun, L. (2019).
+"Weak Instruments in Instrumental Variables Regression: Theory and
+Practice." *Annual Review of Economics*, 11, 727-753.
 """
 
-from typing import Optional, List, Dict, Any
+from typing import Optional, List, Dict, Any, Tuple
 
 import numpy as np
 import pandas as pd
 from scipy import stats
 
+
+# ═══════════════════════════════════════════════════════════════════════
+#  Internal helpers
+# ═══════════════════════════════════════════════════════════════════════
+
+def _partial_out(X: np.ndarray, W: np.ndarray) -> np.ndarray:
+    """Residuals from regressing columns of X on W."""
+    if W.size == 0 or W.shape[1] == 0:
+        return X
+    beta = np.linalg.lstsq(W, X, rcond=None)[0]
+    return X - W @ beta
+
+
+def _prep_matrices(
+    data: pd.DataFrame,
+    y: str,
+    endog: str,
+    instruments: List[str],
+    exog: Optional[List[str]],
+) -> Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray, int]:
+    """Return (Y, D, Z, W, n) as numpy arrays, dropping NA rows."""
+    cols = [y, endog] + instruments + (exog or [])
+    df = data[cols].dropna()
+    n = len(df)
+    Y = df[y].values.astype(float)
+    D = df[endog].values.astype(float)
+    Z = df[instruments].values.astype(float)
+    if Z.ndim == 1:
+        Z = Z.reshape(-1, 1)
+    if exog:
+        W = np.column_stack([np.ones(n)] + [df[v].values.astype(float) for v in exog])
+    else:
+        W = np.ones((n, 1))
+    return Y, D, Z, W, n
+
+
+# ═══════════════════════════════════════════════════════════════════════
+#  1. Olea-Pflueger (2013) effective F — HC-robust
+# ═══════════════════════════════════════════════════════════════════════
+
+def effective_f_test(
+    data: pd.DataFrame,
+    endog: str,
+    instruments: List[str],
+    exog: Optional[List[str]] = None,
+    vcov: str = "HC1",
+) -> Dict[str, Any]:
+    """
+    Olea-Pflueger (2013) robust effective F statistic for weak instruments.
+
[…]
