### 2026-04-15 42e4d14
feat(did): add did_summary() — one-call method-robustness comparison
New sp.did_summary() fits five modern staggered-DID estimators to the
same data and returns a tidy comparison DataFrame — a common 'robustness
check' workflow collapsed into a single function call.

Methods included (shared y/group/time/first_treat interface):
- cs: Callaway & Sant'Anna (2021) + aggte(type='simple')
- sa: Sun & Abraham (2021)
- bjs: Borusyak, Jaravel & Spiess (2024) imputation
- etwfe: Wooldridge (2021) extended TWFE
- stacked: Cengiz et al. (2019) stacked event study

Returns a CausalResult whose:
- .detail 
[…]
