+++ src/statspai/did/summary.py
@@ -1,0 +1,268 @@
+"""
+One-call DID robustness summary across multiple staggered-DID estimators.
+
+``did_summary()`` fits a common set of modern staggered-DID estimators to
+the same data and returns a tidy comparison table with overall ATT, SE,
+95 % CI, p-value, and a short note for each method. The goal is to make
+method-robustness checks a single function call instead of hand-wiring
+five different estimators.
+
+Supported methods (all use the shared ``(y, group, time, first_treat)``
+interface — ``group`` is the unit identifier, ``first_treat`` the cohort
+variable):
+
+================  =====================================================
+Key               Underlying estimator
+================  =====================================================
+``'cs'``          Callaway & Sant'Anna (2021) + ``aggte(type='simple')``
+``'sa'``          Sun & Abraham (2021) interaction-weighted
+``'bjs'``         Borusyak, Jaravel & Spiess (2024) imputation
+``'etwfe'``       Wooldridge (2021) extended TWFE
+``'stacked'``     Cengiz et al. (2019) stacked event study
+================  =====================================================
+
+Example
+-------
+>>> import statspai as sp
+>>> df = sp.dgp_did(n_units=200, n_periods=10, staggered=True, seed=0)
+>>> summary = sp.did_summary(df, y='y', time='time',
+...                          first_treat='first_treat', group='unit')
+>>> print(summary.detail)
+"""
+
+from typing import Optional, List, Union
+
+import numpy as np
+import pandas as pd
+
+from ..core.results import CausalResult
+
+
+_DEFAULT_METHODS: List[str] = ["cs", "sa", "bjs", "etwfe", "stacked"]
+
+_METHOD_LABELS = {
+    "cs": "Callaway & Sant'Anna (2021)",
+    "sa": "Sun & Abraham (2021)",
+    "bjs": "Borusyak, Jaravel & Spiess (2024)",
+    "etwfe": "Wooldridge (2021) ETWFE",
+    "stacked": "Stacked DID (Cengiz et al. 2019)",
+}
+
+
+def _run_cs(data, y, group, time, first_treat, controls, cluster, alpha):
+    from .callaway_santanna import callaway_santanna
+    from .aggte import aggte
+
+    cs = callaway_santanna(
+        data, y=y, g=first_treat, t=time, i=group,
+        x=controls, alpha=alpha,
+    )
+    return aggte(cs, type="simple", alpha=alpha, bstrap=False)
+
+
+def _run_sa(data, y, group, time, first_treat, controls, cluster, alpha):
+    from .sun_abraham import sun_abraham
+
+    return sun_abraham(
+        data, y=y, g=first_treat, t=time, i=group,
+        covariates=controls, cluster=cluster, alpha=alpha,
+    )
+
+
+def _run_bjs(data, y, group, time, first_treat, controls, cluster, alpha):
+    from .did_imputation import did_imputation
+
+    return did_imputation(
+        data, y=y, group=group, time=time, first_treat=first_treat,
+        controls=controls, cluster=cluster, alpha=alpha,
+    )
+
+
+def _run_etwfe(data, y, group, time, first_treat, controls, cluster, alpha):
+    from .wooldridge_did import etwfe
+
+    return etwfe(
+        data, y=y, group=group, time=time, first_treat=first_treat,
+        controls=controls, cluster=cluster, alpha=alpha,
+    )
+
+
+def _run_stacked(data, y, group, time, first_treat, controls, cluster, alpha):
+    from .stacked_did import stacked_did
+
+    return stacked_did(
+        data, y=y, group=group, time=time, first_treat=first_treat,
+        controls=controls, cluster=cluster, alpha=alpha,
+    )
+
+
+_DISPATCH = {
+    "cs": _run_cs,
[…]
