### 2026-04-15 b35806f
feat(spatial.esda): Moran's I (global + local) with permutation inference
Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
