+++ src/statspai/spatial/esda/__init__.py
@@ -1,0 +1,3 @@
+from .moran import moran, moran_local
+
+__all__ = ["moran", "moran_local"]
