### 2026-04-15 512546e
feat(spatial.esda): Geary's C with permutation inference
Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
