+++ src/statspai/spatial/esda/__init__.py
@@ -1,3 +1,4 @@
 from .moran import moran, moran_local
+from .geary import geary
 
-__all__ = ["moran", "moran_local"]
+__all__ = ["moran", "moran_local", "geary"]
