### 2026-04-15 61fe483
feat(spatial.esda): Getis-Ord G and G_local / G*_local
Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
