+++ src/statspai/spatial/esda/__init__.py
@@ -1,4 +1,8 @@
 from .moran import moran, moran_local
 from .geary import geary
+from .getis_ord import getis_ord_g, getis_ord_local
 
-__all__ = ["moran", "moran_local", "geary"]
+__all__ = [
+    "moran", "moran_local", "geary",
+    "getis_ord_g", "getis_ord_local",
+]
