### 2026-04-15 27a37ec
feat(did): forest plot for did_summary — did_summary_plot()
Adds sp.did_summary_plot(result) — publication-ready forest plot of the
method-robustness comparison returned by sp.did_summary():

- One row per method, horizontal errorbar at the 95% CI
- Dashed red line at the cross-method mean (easy dispersion read-off)
- Dotted grey reference at 0 (null effect)
- Optional sort_by='estimate' for ordered display
- Rejects non-did_summary CausalResults with a clear ValueError

Verified: 43KB PNG rendered on 5-method dgp_did data; 2-method subset
also renders; malformed input correctly raises.

Expos
[…]
