+++ src/statspai/spatial/esda/__init__.py
@@ -1,8 +1,9 @@
 from .moran import moran, moran_local
 from .geary import geary
 from .getis_ord import getis_ord_g, getis_ord_local
+from .join_counts import join_counts
 
 __all__ = [
     "moran", "moran_local", "geary",
-    "getis_ord_g", "getis_ord_local",
+    "getis_ord_g", "getis_ord_local", "join_counts",
 ]
