### 2026-04-15 a53009a
feat(did): did_summary gains include_sensitivity — 3-way robustness
Extends sp.did_summary() with an include_sensitivity=True flag that
runs the Rambachan-Roth (2023) breakdown M* on top of the Callaway-
Sant'Anna fit (when 'cs' is among the requested methods). The value is
exposed both in model_info['breakdown_m'] and in a new breakdown_m
column on the detail DataFrame (populated for the CS row; NaN for
other methods since honest_did is CS-specific).

This gives researchers a one-call three-way robustness readout:
  method (CS/SA/BJS/ETWFE/Stacked) × estimate × sensitivity threshold

Backward
[…]
