### 2026-04-15 416a0fe
feat(spatial.esda): Moran scatter + LISA cluster map
Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
