+++ src/statspai/spatial/esda/__init__.py
@@ -2,8 +2,10 @@ from .moran import moran, moran_local
 from .geary import geary
 from .getis_ord import getis_ord_g, getis_ord_local
 from .join_counts import join_counts
+from .plots import moran_plot, lisa_cluster_map
 
 __all__ = [
     "moran", "moran_local", "geary",
     "getis_ord_g", "getis_ord_local", "join_counts",
+    "moran_plot", "lisa_cluster_map",
 ]
