### 2026-04-15 67b5cf9
feat(did): did_summary_to_markdown() and did_summary_to_latex() exports
Publication-ready table export helpers for sp.did_summary() results:

- did_summary_to_markdown(result): GitHub-flavoured Markdown table with
  Method / Estimate (with significance stars) / SE / 95% CI / p-value /
  optional Breakdown M* / Notes. Includes a footer with cross-method
  mean and SD.

- did_summary_to_latex(result): full booktabs LaTeX block with
  \toprule / \midrule / \bottomrule, auto-escaped ampersands in
  method names, and a footer matching the Markdown version.

Both helpers:
- Auto-detect whether break
[…]
