+++ src/statspai/did/__init__.py
@@ -46,7 +46,7 @@ from .pretrends import (
     pretrends_summary,
 )
 from .wooldridge_did import wooldridge_did, etwfe, drdid, twfe_decomposition
-from .summary import did_summary
+from .summary import did_summary, did_summary_to_markdown, did_summary_to_latex
 from .continuous_did import continuous_did
 from .plots import (
     parallel_trends_plot,
@@ -388,6 +388,8 @@ __all__ = [
     'wooldridge_did',
     'etwfe',
     'did_summary',
+    'did_summary_to_markdown',
+    'did_summary_to_latex',
     'drdid',
     'twfe_decomposition',
     # Pre-trends
