+++ src/statspai/did/summary.py
@@ -38,6 +38,19 @@ import pandas as pd
 from ..core.results import CausalResult
 
 
+def _stars(p: float) -> str:
+    """Significance stars for a p-value."""
+    if pd.isna(p):
+        return ""
+    if p < 0.01:
+        return "***"
+    if p < 0.05:
+        return "**"
+    if p < 0.1:
+        return "*"
+    return ""
+
+
 _DEFAULT_METHODS: List[str] = ["cs", "sa", "bjs", "etwfe", "stacked"]
 
 _METHOD_LABELS = {
@@ -309,3 +322,202 @@ def did_summary(
         },
         _citation_key="did_summary",
     )
+
+
+# ═══════════════════════════════════════════════════════════════════════
+#  Export helpers: publication-ready Markdown / LaTeX from a did_summary
+# ═══════════════════════════════════════════════════════════════════════
+
+def _ensure_did_summary(result: CausalResult) -> pd.DataFrame:
+    """Validate that result came from did_summary() and return its detail."""
+    det = getattr(result, "detail", None)
+    if not isinstance(det, pd.DataFrame) or "estimator" not in det.columns:
+        raise ValueError(
+            "did_summary_to_markdown / _to_latex require a CausalResult "
+            "produced by sp.did_summary()."
+        )
+    return det
+
+
+def did_summary_to_markdown(
+    result: CausalResult,
+    digits: int = 4,
+    include_ci: bool = True,
+    include_breakdown: bool = True,
+) -> str:
+    """
+    Render a :func:`did_summary` result as a GitHub-Flavoured Markdown table.
+
+    Columns shown (in order):
+    ``Method``, ``Estimate``, ``SE``, ``95 % CI``, ``p-value``, and
+    optionally ``Breakdown M*`` (when sensitivity was requested).
+
+    Parameters
+    ----------
+    result : CausalResult
+        Output of :func:`did_summary`.
+    digits : int, default 4
+        Decimal precision for numeric columns.
+    include_ci : bool, default True
+        Include the 95 % CI column.
+    include_breakdown : bool, default True
+        Include the Rambachan-Roth breakdown M* column (CS row only,
+        blank for others). Ignored if sensitivity was not requested.
+
+    Returns
+    -------
+    str
+        Multi-line Markdown table, ready to paste into notebooks or PRs.
+    """
+    det = _ensure_did_summary(result)
+
+    has_bd = include_breakdown and det["breakdown_m"].notna().any()
+    lines = []
+    header_cells = ["Method", "Estimate", "SE"]
+    if include_ci:
+        header_cells.append("95% CI")
+    header_cells += ["p"]
+    if has_bd:
+        header_cells.append("Breakdown M*")
+    header_cells.append("Notes")
+    lines.append("| " + " | ".join(header_cells) + " |")
+    lines.append("|" + "|".join([":---"] + ["---:"] * (len(header_cells) - 2) + [":---"]) + "|")
+
+    fmt = f"{{:.{digits}f}}"
+    for _, row in det.iterrows():
+        cells = [row["estimator"]]
+        if pd.isna(row["estimate"]):
+            cells.append("—")
+            cells.append("—")
+            if include_ci:
+                cells.append("—")
+            cells.append("—")
+            if has_bd:
+                cells.append("—")
+        else:
+            est_str = fmt.format(row["estimate"]) + _stars(row["pvalue"])
+            cells.append(est_str)
+            cells.append(fmt.format(row["se"]))
+            if include_ci:
+                cells.append(
+                    f"[{fmt.format(row['ci_low'])}, {fmt.format(row['ci_high'])}]"
[…]
