### 2026-04-15 711cf03
feat(spatial.models): log-det dual path (exact + Barry-Pace approx)
- models/ is now a package: __init__.py re-exports legacy sar/sem/sdm
  (the rename from models.py → models/_legacy.py was bundled into an
  earlier DID commit in parallel — this commit adjusts the relative
  import accordingly).
- _logdet.log_det_exact uses full complex eigendecomposition (real part
  alone is wrong when W is non-symmetric).
- _logdet.log_det_approx uses Hutchinson trace estimation over the
  Mercator series; hard-codes the exactly-known tr(W) = diag(W).sum()
  at k=1 to eliminate the dominant variance source
[…]
