### 2026-04-15 8d97da2
feat(spatial.models): design-matrix helper (formulaic-backed)
Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
