### 2026-04-15 9b5eb6e
feat(did): xvar heterogeneity + parity tests + custom summary + did_report
Four-in-one DID polish round, all scoped to the DID module:

1. etwfe(xvar=...) — covariate-moderated heterogeneity
   New _etwfe_with_xvar helper adds cohort × post × (x - mean(x))
   interactions so the baseline coefficient is ATT(g) evaluated at
   the sample mean of xvar and the slope coefficient measures how
   ATT(g) shifts per unit of xvar. Closes the last R-etwfe feature
   gap. Backward compatible: etwfe(xvar=None) is byte-identical to
   wooldridge_did().

2. tests/test_did_summary.py — DID consistency + parit
[…]
