+++ src/statspai/did/summary.py
@@ -303,7 +303,7 @@ def did_summary(
     else:
         avg_est, disp = np.nan, np.nan
 
-    return CausalResult(
+    result = CausalResult(
         method="DID Method-Robustness Summary",
         estimand="Overall ATT (mean across methods)",
         estimate=avg_est,
@@ -322,6 +322,28 @@ def did_summary(
         },
         _citation_key="did_summary",
     )
+
+    # Custom .summary() — prints the comparison table as a Markdown block
+    # instead of the base CausalResult's single-estimate format.
+    def _custom_summary(alpha=None):  # noqa: ARG001
+        header = "=" * 78 + "\n"
+        header += "  DID Method-Robustness Summary\n"
+        header += "=" * 78 + "\n\n"
+        if fit:
+            header += f"  Fitted methods: {', '.join(fit)}\n"
+        if failed:
+            header += f"  Failed:         {', '.join(failed.keys())}\n"
+        if not np.isnan(avg_est):
+            header += f"  Mean ATT:       {avg_est:+.4f}\n"
+            header += f"  Cross-method SD: {disp:+.4f}\n"
+        if breakdown_m_value is not None:
+            header += f"  Breakdown M*:   {breakdown_m_value:.4f} (CS row)\n"
+        header += "\n"
+        body = did_summary_to_markdown(result)
+        return header + body
+
+    result.summary = _custom_summary
+    return result
 
 
 # ═══════════════════════════════════════════════════════════════════════
@@ -521,3 +543,117 @@ def did_summary_to_latex(
         )
     lines.append("\\end{table}")
     return "\n".join(lines)
+
+
+# ═══════════════════════════════════════════════════════════════════════
+#  One-call DID bundle report — txt + md + tex + png in a single directory
+# ═══════════════════════════════════════════════════════════════════════
+
+def did_report(
+    data: pd.DataFrame,
+    y: str,
+    time: str,
+    first_treat: str,
+    group: str,
+    save_to: str,
+    methods: Union[str, List[str]] = "auto",
+    controls: Optional[List[str]] = None,
+    cluster: Optional[str] = None,
+    alpha: float = 0.05,
+    include_sensitivity: bool = True,
+    plot_sort_by: Optional[str] = "estimate",
+    verbose: bool = False,
+) -> CausalResult:
+    """
+    One-call comprehensive DID report: fits all methods + exports every format.
+
+    Writes the following files to ``save_to``:
+
+    - ``did_summary.txt`` : text dump of ``result.summary()``.
+    - ``did_summary.md``  : GitHub-Flavoured Markdown table.
+    - ``did_summary.tex`` : publication-ready LaTeX ``booktabs`` fragment.
+    - ``did_summary.png`` : forest plot (requires matplotlib).
+    - ``did_summary.json`` : detail table + model_info in JSON.
+
+    Parameters
+    ----------
+    data, y, time, first_treat, group, methods, controls, cluster, alpha
+        Same as :func:`did_summary`.
+    save_to : str
+        Directory path. Created if it does not exist.
+    include_sensitivity : bool, default True
+        Whether to run Rambachan-Roth breakdown M*. Defaults to ``True``
+        in ``did_report`` (vs ``False`` in ``did_summary``) because a
+        report is expected to be comprehensive.
+    plot_sort_by : {'estimate', None}, default 'estimate'
+        Sort the forest plot by estimate ascending.
+    verbose : bool, default False
+        Print progress lines.
+
+    Returns
+    -------
+    CausalResult
+        The underlying :func:`did_summary` output. All side-effect files
+        are written to ``save_to`` as a bundle.
+    """
+    import json
[…]
