+++ src/statspai/did/__init__.py
@@ -46,7 +46,7 @@ from .pretrends import (
     pretrends_summary,
 )
 from .wooldridge_did import wooldridge_did, etwfe, drdid, twfe_decomposition
-from .summary import did_summary, did_summary_to_markdown, did_summary_to_latex
+from .summary import did_summary, did_summary_to_markdown, did_summary_to_latex, did_report
 from .continuous_did import continuous_did
 from .plots import (
     parallel_trends_plot,
@@ -390,6 +390,7 @@ __all__ = [
     'did_summary',
     'did_summary_to_markdown',
     'did_summary_to_latex',
+    'did_report',
     'drdid',
     'twfe_decomposition',
     # Pre-trends
