+++ README_CN.md
@@ -86,6 +86,8 @@ StatsPAI **不是** R 的 wrapper。我们从原始论文独立重新实现每�
 | `did()` | 自动分派 DID（2×2 或交错） | — |
 | `did_summary()` | 一次调用跑 CS/SA/BJS/ETWFE/Stacked 五种方法并出对比表 | — |
 | `did_summary_plot()` | 方法稳健性森林图（点估计 + CI 并排） | — |
+| `did_summary_to_markdown()` / `_to_latex()` | 论文级稳健性对比表（GFM / booktabs） | — |
+| `did_report()` | 一次调用打包输出：txt + md + tex + png + json | — |
 | `callaway_santanna()` | 交错 DID，异质处理效应 | Callaway & Sant'Anna (2021) |
 | `sun_abraham()` | 交互加权事件研究 | Sun & Abraham (2021) |
 | `bacon_decomposition()` | TWFE 分解诊断 | Goodman-Bacon (2021) |

[…]
