### 2026-04-15 2c84157
feat(did): etwfe_emfx() — R etwfe::emfx-equivalent four aggregations
Completes ETWFE parity with R's etwfe package by adding the emfx()
marginal-effects aggregator. Takes an etwfe/wooldridge_did result and
returns one of four standard aggregations used in applied work:

  type='simple'   - overall cohort-size-weighted ATT
                    (verified: matches fit.estimate to 1e-10)
  type='group'    - ATT per treatment cohort (one row per g)
  type='event'    - ATT per event time e = t - g, averaged across
                    cohorts with cohort-size weights
  type='calendar' - ATT per calend
[…]
