+++ src/statspai/did/wooldridge_did.py
@@ -1100,3 +1100,198 @@ def twfe_decomposition(
         model_info=model_info,
         _citation_key="twfe_decomposition",
     )
+
+
+# ═══════════════════════════════════════════════════════════════════════
+#  4. etwfe_emfx — R etwfe-style marginal-effects aggregations
+# ═══════════════════════════════════════════════════════════════════════
+
+def etwfe_emfx(
+    result: CausalResult,
+    type: str = "simple",
+    alpha: float = 0.05,
+) -> CausalResult:
+    """
+    R ``etwfe::emfx``-style aggregated marginal effects for an ETWFE fit.
+
+    Takes the result of :func:`etwfe` / :func:`wooldridge_did` and returns
+    one of four aggregations used in applied work:
+
+    ================  ========================================================
+    ``type``          Aggregation
+    ================  ========================================================
+    ``'simple'``      Overall cohort-size-weighted ATT (same as ``result.estimate``).
+    ``'group'``       ATT per treatment cohort ``g``.
+    ``'event'``       ATT per event time ``e = t - g``, averaged across cohorts.
+    ``'calendar'``    ATT per calendar time ``t``, averaged across cohorts for
+                      which ``t >= g``.
+    ================  ========================================================
+
+    Parameters
+    ----------
+    result : CausalResult
+        Output of :func:`etwfe` or :func:`wooldridge_did`.
+    type : {'simple', 'group', 'event', 'calendar'}, default 'simple'
+        Aggregation type.
+    alpha : float, default 0.05
+        Significance level for confidence intervals.
+
+    Returns
+    -------
+    CausalResult
+        ``estimate`` is the overall ATT (for ``type='simple'``) or the
+        mean of the sub-aggregation (for the other types). ``detail``
+        contains one row per group/event-time/calendar-time with
+        (estimate, se, pvalue, ci_low, ci_high).
+
+    Notes
+    -----
+    For ``'event'`` and ``'calendar'``, the reported SE treats the
+    per-cohort coefficients as independent — a standard approximation
+    that matches R etwfe's default under classical vcov. Cluster-robust
+    or fully-general SEs require the full regression vcov, which can
+    be requested via ``sp.wooldridge_did`` + the ``model_info`` matrix
+    in a future release.
+
+    Examples
+    --------
+    >>> import statspai as sp
+    >>> df = sp.dgp_did(n_units=200, n_periods=10, staggered=True)
+    >>> fit = sp.etwfe(df, y='y', time='time',
+    ...                first_treat='first_treat', group='unit')
+    >>> evt = sp.etwfe_emfx(fit, type='event')
+    >>> print(evt.detail)   # ATT by event time
+    >>> grp = sp.etwfe_emfx(fit, type='group')
+    >>> cal = sp.etwfe_emfx(fit, type='calendar')
+    """
+    valid = {"simple", "group", "event", "calendar"}
+    if type not in valid:
+        raise ValueError(f"type must be one of {sorted(valid)}; got {type!r}")
+
+    if not isinstance(result.model_info, dict) or "cohorts" not in result.model_info:
+        raise ValueError(
+            "etwfe_emfx requires a result produced by sp.etwfe / "
+            "sp.wooldridge_did — missing 'cohorts' in model_info."
+        )
+
+    mi = result.model_info
+    cohorts = mi["cohorts"]
+    cohort_weights = mi.get("cohort_weights", {})
+    event_study = mi.get("event_study")
+
+    # ── simple ──
+    if type == "simple":
+        # Just return a tidy one-row result
+        df_resid = max(result.n_obs - len(cohorts), 1)
[…]
