### 2026-04-15 8ff253a
refactor(spatial): sparse-aware SAR/SEM/SDM ML — backward-compat at rtol<1e-4
Task 12-14 of SP-01 S1.1. Replaces the legacy 419-line dense implementation
(preserved at statspai.spatial.models._legacy) with a sparse-aware ML
estimator in statspai.spatial.models.ml:

- Accepts ndarray, scipy.sparse, OR statspai.spatial.weights.W as weights.
- Inner loop uses sparse matvec; dense eigendecomposition only for log-det
  at n < 2500, Barry-Pace approximation for larger problems.
- Matches legacy parameter vector, standard errors, sigma², and
  log-likelihood to within rtol=1e-4 on the pinned baseline
[…]
