+++ src/statspai/spatial/models/ml.py
@@ -1,0 +1,433 @@
+"""Sparse-backed ML estimation for spatial regression models (SAR / SEM / SDM).
+
+Drop-in replacement for the legacy dense implementation in :mod:`._legacy`.
+Accepts ndarray, scipy.sparse matrix, or :class:`statspai.spatial.weights.W`.
+
+Numerics are designed to match the legacy reference implementation to within
+``rtol < 1e-6`` for ``n < 5000`` (exact log-det path), and to scale to
+``n > 10000`` via a stochastic Barry-Pace log-det (pulled in from
+:mod:`._logdet`).
+
+References
+----------
+Anselin (1988); Ord (1975); LeSage & Pace (2009).
+"""
+from __future__ import annotations
+
+from typing import Any, Dict, List, Optional, Tuple, Union
+
+import numpy as np
+import pandas as pd
+from scipy import sparse, stats as sp_stats  # noqa: F401  (sp_stats reserved for future SE work)
+from scipy.optimize import minimize_scalar
+
+from ...core.results import EconometricResults
+from ..weights.core import W as _W
+from ._logdet import log_det_approx
+
+
+ArrayOrW = Union[np.ndarray, sparse.spmatrix, "_W"]
+
+# When the problem is small enough, prefer the exact eigenvalue-based
+# log-determinant — it matches the legacy implementation bit-for-bit and
+# allows sub-1e-6 numerical agreement on the backward-compat fixture.
+_EXACT_LOGDET_MAX_N = 5000
+
+
+# ====================================================================== #
+#  Internal helpers
+# ====================================================================== #
+
+def _coerce_W(W: ArrayOrW, n_expected: Optional[int],
+              row_normalize: bool) -> sparse.csr_matrix:
+    """Return W as a CSR matrix, optionally row-normalised.
+
+    A ``W`` object's ``transform`` attribute already encodes whether it was
+    row-standardised; we still re-normalise here when ``row_normalize=True``
+    so behaviour is identical regardless of input type. (Re-normalising an
+    already row-stochastic matrix is a no-op.)
+    """
+    if isinstance(W, _W):
+        M = W.sparse
+    elif sparse.issparse(W):
+        M = W.tocsr()
+    else:
+        M = sparse.csr_matrix(np.asarray(W, dtype=float))
+    if n_expected is not None and M.shape != (n_expected, n_expected):
+        raise ValueError(
+            f"W must be ({n_expected}, {n_expected}), got {M.shape}"
+        )
+    if row_normalize:
+        rs = np.asarray(M.sum(axis=1)).ravel()
+        rs = np.where(rs == 0, 1.0, rs)
+        M = sparse.diags(1.0 / rs) @ M
+    return M.tocsr()
+
+
+def _parse_formula(formula: str, data: pd.DataFrame
+                   ) -> Tuple[np.ndarray, np.ndarray, str, List[str]]:
+    """Legacy-compatible formula parser.
+
+    Returns ``(y, X, dep_var, indep_vars)`` where ``X`` includes a leading
+    constant column. We deliberately mirror the legacy parser (rather than
+    using formulaic) so column names match exactly — ``["const", x1, ...]``
+    — and the backward-compat parameter vectors line up.
+    """
+    if "~" not in formula:
[…]
