### 2026-04-15 9dfcb52
feat(did): etwfe final three — multi-xvar + repeated CS + cgroup
Completes R-etwfe parity by adding the last three features:

1. Multi-xvar heterogeneity: etwfe(xvar=['x1','x2',...])
   - xvar now accepts a str (backward-compat) or list of strs
   - For each cohort g and each x_j, fits a slope coefficient
   - detail DataFrame gains per-xvar columns:
       slope_<x>, slope_<x>_se, slope_<x>_pvalue
   - Backward-compat: when exactly one xvar is given, adds aliases
     slope_wrt_x / slope_se / slope_pvalue so existing callers keep working
   - model_info['xvar'] is always a list; model_info['x
[…]
