+++ src/statspai/did/wooldridge_did.py
@@ -358,7 +358,9 @@ def etwfe(
     controls: Optional[List[str]] = None,
     cluster: Optional[str] = None,
     alpha: float = 0.05,
-    xvar: Optional[str] = None,
+    xvar: Optional[Any] = None,  # Union[str, List[str]]
+    panel: bool = True,
+    cgroup: str = "notyet",
 ) -> CausalResult:
     """
     Extended Two-Way Fixed Effects (ETWFE) — Wooldridge (2021).
@@ -404,7 +406,7 @@ def etwfe(
     ``tvar = time``               ``time='time'``
     ``gvar = first_treat``        ``first_treat='first_treat'``
     ``ivar = unit``               ``group='unit'``
-    ``xvar`` (covariate het.)     ``xvar='x1'`` (single covariate)
+    ``xvar`` (covariate het.)     ``xvar='x1'`` or ``xvar=['x1','x2']``
     ``vcov = ~cluster``           ``cluster='cluster'``
     ============================  ========================================
 
@@ -434,14 +436,39 @@ def etwfe(
     callaway_santanna : CS (2021) group-time ATT estimator.
     aggte : Aggregation of group-time ATTs (event/group/calendar/simple).
     """
-    if xvar is None:
+    # Validate cgroup
+    if cgroup not in ("notyet", "nevertreated"):
+        raise ValueError(
+            f"cgroup must be 'notyet' or 'nevertreated'; got {cgroup!r}"
+        )
+
+    # Normalise xvar to a list (or None)
+    xvar_list: Optional[List[str]] = None
+    if xvar is not None:
+        xvar_list = [xvar] if isinstance(xvar, str) else list(xvar)
+
+    # Dispatch to the right implementation.
+    if not panel:
+        # Repeated cross-section: no unit FE, replace with cohort dummies.
+        return _etwfe_repeated_cs(
+            data=data, y=y, time=time, first_treat=first_treat,
+            xvar=xvar_list, controls=controls, cluster=cluster,
+            alpha=alpha, cgroup=cgroup,
+        )
+    if cgroup == "nevertreated":
+        # Per-cohort regressions restricted to cohort + never-treated.
+        return _etwfe_never_only(
+            data=data, y=y, group=group, time=time, first_treat=first_treat,
+            xvar=xvar_list, controls=controls, cluster=cluster, alpha=alpha,
+        )
+    if xvar_list is None:
         return wooldridge_did(
             data=data, y=y, group=group, time=time, first_treat=first_treat,
             controls=controls, cluster=cluster, alpha=alpha,
         )
     return _etwfe_with_xvar(
         data=data, y=y, group=group, time=time, first_treat=first_treat,
-        xvar=xvar, controls=controls, cluster=cluster, alpha=alpha,
+        xvar=xvar_list, controls=controls, cluster=cluster, alpha=alpha,
     )
 
 
@@ -451,18 +478,18 @@ def _etwfe_with_xvar(
     group: str,
     time: str,
     first_treat: str,
-    xvar: str,
+    xvar: List[str],
     controls: Optional[List[str]] = None,
     cluster: Optional[str] = None,
     alpha: float = 0.05,
 ) -> CausalResult:
     """ETWFE with covariate-moderated heterogeneity (R etwfe's ``xvar``).
 
-    For each cohort ``g``, adds an interaction between the
-    cohort × post dummy and ``(xvar - mean(xvar))``, so the slope
-    coefficient measures how ATT(g) shifts per unit of ``xvar``. The
[…]
