+++ README_CN.md
@@ -94,7 +94,7 @@ StatsPAI **不是** R 的 wrapper。我们从原始论文独立重新实现每�
 | `honest_did()` | 平行趋势假设敏感性 | Rambachan & Roth (2023) |
 | `continuous_did()` | 连续处理 DID（剂量反应） | Callaway, Goodman-Bacon & Sant'Anna (2024) |
 | `did_imputation()` | 插补 DID 估计量 | Borusyak, Jaravel & Spiess (2024) |
-| `wooldridge_did()` / `etwfe()` | 扩展 TWFE（cohort×post 饱和），支持 `xvar=` 异质性 | Wooldridge (2021) |
+| `wooldridge_did()` / `etwfe()` | 扩展 TWFE：`xvar=`（单/多协变量异质性）+ `panel=`（重复截面）+ `cgroup=`（never/notyet 控制组） | Wooldridge (2021) |
 | `etwfe_emfx()` | R `etwfe::emfx` 等价——simple/group/event/calendar 四种聚合边际效应 | McDermott (2023) |
 | `drdid()` | 2×2 双重稳健 DID（OR + IPW） | Sant'Anna & Zhao (2020) |
 | `stacked_did()` | 堆叠事件研究 DID | Cengiz et al. (2019); Baker, Larcker & Wang (2022) |
