### 2026-04-15 4bd718d
feat(spatial.models): SLX (lag of X) and SAC/SARAR estimators
Task 15 of SP-01 S1.1. Adds two more spatial regression flavours to ml.py:

- SLX: Y = Xβ + WXθ + ε. Ordinary least squares on the design matrix
  augmented with spatial lags of X (skip constant).
- SAC / SARAR: Y = ρWY + Xβ + u, u = λWu + ε. Concentrated ML jointly
  profiling (β, σ²) at each (ρ, λ); multi-start with differential_evolution
  global search + Nelder-Mead polish (single-start Nelder-Mead gets stuck
  near zero because the (ρ, λ) neighbourhood is flat).

Exposed at statspai.spatial.models.ml.slx and sac. Tests verify:

[…]
