+++ src/statspai/spatial/models/ml.py
@@ -138,6 +138,8 @@ def _make_results(
         "sar": "SAR (Spatial Lag)",
         "sem": "SEM (Spatial Error)",
         "sdm": "SDM (Spatial Durbin)",
+        "slx": "SLX (Spatial Lag of X)",
+        "sac": "SAC / SARAR (Spatial Lag + Error)",
     }.get(model_type, model_type)
     diagnostics: Dict[str, Any] = {
         "sigma2": round(float(sigma2), 6),
@@ -430,4 +432,138 @@ def sdm(W: ArrayOrW, data: pd.DataFrame, formula: str,
     )
 
 
-__all__ = ["sar", "sem", "sdm"]
+# --------------------------------------------------------------------- #
+#  SLX:  Y = X β + W X θ + ε   — OLS on augmented design
+# --------------------------------------------------------------------- #
+
+def slx(W: ArrayOrW, data: pd.DataFrame, formula: str,
+        row_normalize: bool = True, alpha: float = 0.05) -> EconometricResults:
+    """Spatial Lag of X (SLX) model: ``Y = Xβ + WXθ + ε``.
+
+    No autoregressive term. Estimated by ordinary least squares on the
+    design matrix augmented with spatially-lagged covariates (skip the
+    constant).
+    """
+    y, X, dep_var, indep = _parse_formula(formula, data)
+    n, k = X.shape
+    M = _coerce_W(W, n_expected=n, row_normalize=row_normalize)
+
+    WX = M @ X[:, 1:]
+    X_aug = np.column_stack([X, WX])
+    XtX_inv = np.linalg.inv(X_aug.T @ X_aug)
+    beta_aug = XtX_inv @ (X_aug.T @ y)
+    e = y - X_aug @ beta_aug
+    sigma2 = float(e @ e) / (n - X_aug.shape[1])
+
+    se_beta = np.sqrt(np.diag(sigma2 * XtX_inv))
+    lag_names = [f"W_{v}" for v in indep]
+    var_names = ["const"] + list(indep) + lag_names
+    fitted = y - e
+    log_lik = (
+        -n / 2 * np.log(2 * np.pi)
+        - n / 2 * np.log(sigma2)
+        - (e @ e) / (2 * sigma2)
+    )
+
+    return _make_results(
+        model_type="slx", spatial_param_name="", spatial_param_value=0.0,
+        var_names=var_names, params_vec=beta_aug, se_vec=se_beta,
+        sigma2=sigma2, resid=e, fitted=fitted, n=n, dep_var=dep_var,
+        log_lik=log_lik, extra_diag={},
+    )
+
+
+# --------------------------------------------------------------------- #
+#  SAC / SARAR:  Y = ρ W Y + X β + u,   u = λ W u + ε
+# --------------------------------------------------------------------- #
+
+def sac(W: ArrayOrW, data: pd.DataFrame, formula: str,
+        row_normalize: bool = True, alpha: float = 0.05) -> EconometricResults:
+    """SAC / SARAR: combined spatial lag + spatial error model.
+
+    Jointly estimates the autoregressive coefficient ρ (on the dependent
+    variable) and the spatial-error coefficient λ by concentrated
+    maximum likelihood, profiling out (β, σ²) at each (ρ, λ) candidate.
+    """
+    y, X, dep_var, indep = _parse_formula(formula, data)
+    n, k = X.shape
+    M = _coerce_W(W, n_expected=n, row_normalize=row_normalize)
+    Wy = M @ y
+    eigvals = _eigvals_for_bounds(M)
[…]
