### 2026-04-15 5574799
feat(spatial): LM spatial diagnostics (Anselin 1988) + LeSage-Pace impacts
Tasks 18-19 of SP-01 S1.1.

diagnostics.lm_tests() — the full Anselin (1988) / Anselin-Bera (1998)
Lagrange-multiplier battery: LM-err, LM-lag, Robust LM-err, Robust LM-lag,
and joint SARMA (df=2). Guides model selection between OLS, SAR, and SEM.
Rejects spatial lag DGPs at p<0.01 on the included 120-obs test; does not
over-reject iid data.

impacts.impacts() — LeSage-Pace (2009) direct / indirect / total impacts
for SAR, SAC, and SDM results. Point estimates from (I-ρW)^{-1} β (plus
θ term for SDM); simulated SEs via 
[…]
