+++ src/statspai/spatial/models/impacts.py
@@ -1,0 +1,152 @@
+"""LeSage-Pace (2009) direct / indirect / total impacts for SAR-family models.
+
+For the reduced form  Y = (I - ρW)^{-1} (X β + ε),
+the partial derivative matrix w.r.t. regressor k is
+
+    S_k(ρ, β_k) = (I - ρW)^{-1} β_k              (SAR, SAC)
+    S_k(ρ, β_k, θ_k) = (I - ρW)^{-1} (β_k I + θ_k W)   (SDM)
+
+Summary measures:
+    direct_k   = (1/n) trace(S_k)
+    total_k    = (1/n) sum(S_k)
+    indirect_k = total_k - direct_k
+
+Confidence intervals are produced by Monte-Carlo simulation from the
+estimator's asymptotic distribution (multivariate normal centred at the
+point estimate with covariance diag(se²)).
+"""
+from __future__ import annotations
+
+from typing import Optional
+
+import numpy as np
+import pandas as pd
+from scipy import sparse
+
+
+def impacts(result, n_sim: int = 1000, seed: Optional[int] = None) -> pd.DataFrame:
+    """Compute direct / indirect / total impacts + simulated SEs.
+
+    Parameters
+    ----------
+    result : EconometricResults
+        Output of ``sp.sar``, ``sp.sdm``, or ``sp.sac``.
+    n_sim : int, default 1000
+        Number of Monte-Carlo draws for simulated SEs.
+    seed : int, optional
+        RNG seed.
+
+    Returns
+    -------
+    DataFrame with one row per non-constant regressor and columns
+    ``Direct``, ``SE_Direct``, ``Indirect``, ``SE_Indirect``, ``Total``,
+    ``SE_Total``.
+    """
+    model_type = result.model_info.get("model_type", "").split(" ")[0].upper()
+    if model_type not in {"SAR", "SAC", "SDM"}:
+        raise ValueError(
+            f"impacts() only defined for SAR / SDM / SAC models; got {model_type!r}"
+        )
+
+    names = list(result.params.index)
+    params = np.asarray(result.params.values, dtype=float)
+    se = np.asarray(result.std_errors.values, dtype=float)
+
+    # Locate spatial parameter and β slice (and θ slice for SDM)
+    if model_type == "SAR":
+        rho = float(result.model_info["spatial_param_value"])
+        rho_idx = names.index("rho")
+        beta_idx = [i for i, nm in enumerate(names) if nm not in {"const", "rho"}]
+        theta_idx: list[int] = []
+        covariate_names = [names[i] for i in beta_idx]
+    elif model_type == "SAC":
+        rho = float(result.model_info["spatial_param_value"])
+        rho_idx = names.index("rho")
+        beta_idx = [i for i, nm in enumerate(names)
+                    if nm not in {"const", "rho", "lambda"}]
+        theta_idx = []
+        covariate_names = [names[i] for i in beta_idx]
+    else:    # SDM
+        rho = float(result.model_info["spatial_param_value"])
+        rho_idx = names.index("rho")
+        lag_cols = [i for i, nm in enumerate(names) if nm.startswith("W_")]
+        beta_idx = [i for i, nm in enumerate(names)
+                    if nm not in {"const", "rho"} and not nm.startswith("W_")]
+        theta_idx = lag_cols
+        covariate_names = [names[i] for i in beta_idx]
+
+    # Reconstruct W from result (attached by ml.sar/sem/sdm as result... actually not)
+    # The legacy result object does not carry W. Instead we require the caller
+    # to pass it explicitly via result.data_info["W"] OR we require that the
+    # model was fit with n small enough to keep a dense reference. For now,
[…]

[…]
