### 2026-04-15 9311eb9
feat(spatial): flat re-exports at sp.* and sp.spatial.*
Task 20 of SP-01 S1.1. Wires the spatial full-stack public surface
into the top-level statspai namespace so users can write

    import statspai as sp
    w   = sp.queen_weights(gdf)
    out = sp.moran(df["y"], w)
    res = sp.sar(w, data=df, formula="y ~ x1 + x2")
    imp = sp.impacts(res)
    lms = sp.lm_tests("y ~ x1 + x2", df, w)

without touching the intermediate sp.spatial.* paths. Backwards
compatibility preserved: the pre-existing sp.sar / sp.sem / sp.sdm /
sp.SpatialModel entry points continue to work unchanged. Also exposes
sp.s
[…]
