+++ src/statspai/__init__.py
@@ -71,7 +71,15 @@ from .output.regression_table import regtable, RegtableResult, mean_comparison, 
 from .postestimation import margins, marginsplot, margins_at, margins_at_plot, contrast, pwcompare, test, lincom
 from .diagnostics import oster_bounds, mccrary_test, diagnose, het_test, reset_test, vif, sensemakr, rddensity, hausman_test, anderson_rubin_test, effective_f_test, tF_critical_value, evalue, evalue_from_result, diagnose_result, estat, kitagawa_test, KitagawaResult
 from .inference import wild_cluster_bootstrap, aipw, ri_test, ipw, bootstrap, BootstrapResult, twoway_cluster, conley, pate, PATEEstimator, fisher_exact, FisherResult, jackknife_se, cr2_se, wild_cluster_boot
-from .spatial import sar, sem, sdm, SpatialModel
+from .spatial import (
+    sar, sem, sdm, slx, sac, SpatialModel,
+    W, queen_weights, rook_weights, knn_weights,
+    distance_band, kernel_weights, block_weights,
+    moran, moran_local, geary, getis_ord_g, getis_ord_local, join_counts,
+    moran_plot, lisa_cluster_map,
+    lm_tests, moran_residuals, impacts,
+)
+from . import spatial
 from .plots import binscatter, set_theme, list_themes, use_chinese, interactive, get_code
 from .utils import (
     label_var, label_vars, get_label, get_labels, describe, pwcorr, winsor, read_data,
