### 2026-04-15 8fbddae
test(spatial): Columbus cross-validation vs PySAL spreg — rtol<1e-4
Task 21 of SP-01 S1.1 — THE ACCEPTANCE GATE for the spatial core sprint.

Reference numbers generated once from libpysal 4.14.1 + spreg 1.9.0 on
the Columbus shapefile (standard spatial-econometrics benchmark):

  ML_Lag (SAR): rho = 0.42332541733, betas = [45.603, -1.049, -0.266]
  ML_Error (SEM): lam = 0.54675302707, betas = [60.279, -0.957, -0.305]

Our sparse-aware ML reproduces these to:
  SAR rho differs by ~1e-7; betas by ~1e-6 to 5e-6
  SEM lam differs by ~1e-6; betas by ~1e-6 to 2e-5
  sigma^2 matches at rtol=1e-4 for
[…]
