### 2026-04-15 d7c48f1
fix(did): address code review findings — 5 blockers + 3 high-severity
Independent code review flagged 5 blockers and 7 high-severity issues
before claiming DID module is release-ready. This commit closes all 5
blockers and the 3 highest-impact high-severity items.

Blockers:
C1. xvar=NaN-column silently returned n_obs=0 / estimate=0 with no
    warning. Dispatcher now validates each xvar: raises ValueError
    if fewer than 2 finite rows, naming the offending column.
C2. Constant xvar silently returned the no-heterogeneity baseline with
    all-zero slopes. Dispatcher now raises ValueError("..
[…]
