+++ src/statspai/did/summary.py
@@ -51,6 +51,50 @@ def _stars(p: float) -> str:
     return ""
 
 
+class DIDSummaryResult(CausalResult):
+    """
+    CausalResult subclass returned by :func:`did_summary`.
+
+    Defines ``.summary()`` as a real method (rather than a closure-bound
+    instance attribute) so the result object serialises cleanly for
+    caching (joblib / multiprocessing).
+
+    Notes
+    -----
+    The ``se`` attribute stores the **cross-method dispersion** (SD of
+    estimates across successfully-fit methods), *not* a sampling SE.
+    Read :attr:`model_info` ``['dispersion']`` for the same value.
+    """
+
+    _DID_SUMMARY_MARKER: bool = True
+
+    def summary(self, alpha: Optional[float] = None) -> str:  # noqa: ARG002
+        mi = self.model_info or {}
+        fit = mi.get("methods_fit", [])
+        failed = mi.get("methods_failed", {})
+        disp = mi.get("dispersion")
+        bd = mi.get("breakdown_m")
+
+        lines = ["=" * 78,
+                 "  DID Method-Robustness Summary",
+                 "=" * 78, ""]
+        if fit:
+            lines.append(f"  Fitted methods: {', '.join(fit)}")
+        if failed:
+            lines.append(f"  Failed:         {', '.join(failed.keys())}")
+        if len(fit) >= 2 and not np.isnan(self.estimate):
+            lines.append(f"  Mean ATT:       {self.estimate:+.4f}")
+            if disp is not None and not np.isnan(disp):
+                lines.append(f"  Cross-method SD: {disp:+.4f}")
+        elif len(fit) == 1 and not np.isnan(self.estimate):
+            lines.append(f"  ATT:            {self.estimate:+.4f}")
+        if bd is not None:
+            lines.append(f"  Breakdown M*:   {bd:.4f} (CS row)")
+        lines.append("")
+        lines.append(did_summary_to_markdown(self))
+        return "\n".join(lines)
+
+
 _DEFAULT_METHODS: List[str] = ["cs", "sa", "bjs", "etwfe", "stacked"]
 
 _METHOD_LABELS = {
@@ -231,10 +275,33 @@ def did_summary(
             f"Valid keys: {sorted(_DISPATCH)} (or 'auto' / 'all')."
         )
 
+    # C4: validate column names up front so user typos surface as
+    # crisp KeyErrors rather than 5 "method FAILED" rows.
+    required = [y, time, first_treat, group]
+    optional = list(controls or [])
+    if cluster:
+        optional.append(cluster)
+    missing = [c for c in required + optional if c not in data.columns]
+    if missing:
+        raise KeyError(
+            f"columns not found in data: {missing}. "
+            f"Available columns: {list(data.columns)[:15]}"
+            + (" ..." if len(data.columns) > 15 else "")
+        )
+
     rows: List[dict] = []
     failed: dict = {}
     fit: List[str] = []
     cs_raw = None  # raw CS result for sensitivity analysis
+
+    # Narrow exceptions that represent "the estimator couldn't fit this
+    # data" rather than a bug in user input. Other exceptions propagate.
+    _expected_exc = (
+        np.linalg.LinAlgError,
+        ValueError,
+        RuntimeError,
+        ZeroDivisionError,
+    )
 
     for name in methods_list:
         label = _METHOD_LABELS[name]
@@ -242,8 +309,6 @@ def did_summary(
             print(f"  running {name} ({label})...", flush=True)
         try:
             if name == "cs" and include_sensitivity:
-                # Run CS + aggte inline so we can hold on to the raw
-                # CS result for the subsequent breakdown_m call.
                 from .callaway_santanna import callaway_santanna
                 from .aggte import aggte as _aggte
                 cs_raw = callaway_santanna(
[…]
