### 2026-04-15 458632f
fix(did): close remaining review follow-ups — H3/H4/H6/H7
Finishes the polish round so the DID module can be declared release-ready.

H3. _etwfe_never_only no longer mutates the caller's DataFrame.
    Removed the _ft_cache helper column; the cohort membership is now
    computed into a local Series and consumed inline. Verified that
    df.columns is unchanged after the call. Docstring also notes the
    per-cohort cluster SE adjustment so users aren't surprised when
    the aggregated SE differs from a full-sample fit.

H4. Slope indexing in _etwfe_with_xvar is now name-keyed.
    Builds an 
[…]
