+++ src/statspai/did/wooldridge_did.py
@@ -183,15 +183,19 @@ def wooldridge_did(
         interaction_cols.append(col)
 
     # ── Also build cohort × relative-time dummies for event study ───
+    # H7 fix: build event dummies for BOTH leads and lags. Omit
+    # rel = -1 as the reference category so the design is identified.
+    # Post-only consumers can still filter via etwfe_emfx(include_leads=False).
     event_cols: List[str] = []
     rel_times = set()
     for g in cohorts:
         mask_g = df["_ft"] == g
         for t_val in periods:
             rel = int(t_val - g)
-            if rel < 0:
-                continue  # only post-treatment for basic spec
-            col = f"_coh{int(g)}_rel{rel}"
+            if rel == -1:
+                continue  # reference / omitted period
+            col = (f"_coh{int(g)}_rel{rel}" if rel >= 0
+                   else f"_coh{int(g)}_rel_neg{abs(rel)}")
             df[col] = ((mask_g) & (df[time] == t_val)).astype(float)
             event_cols.append(col)
             rel_times.add(rel)
@@ -308,9 +312,15 @@ def wooldridge_did(
 
             ev_rows = []
             for j, col in enumerate(event_cols):
-                parts = col.replace("_coh", "").replace("_rel", " ").split()
-                coh_val = int(parts[0])
-                rel_val = int(parts[1])
+                # Parse name: _coh{g}_rel{k}  or  _coh{g}_rel_neg{k}
+                stripped = col.replace("_coh", "")
+                if "_rel_neg" in stripped:
+                    coh_part, rel_part = stripped.split("_rel_neg")
+                    rel_val = -int(rel_part)
+                else:
+                    coh_part, rel_part = stripped.split("_rel")
+                    rel_val = int(rel_part)
+                coh_val = int(coh_part)
                 idx_j = j + 1
                 ev_rows.append({
                     "cohort": coh_val,
@@ -564,6 +574,13 @@ def _etwfe_with_xvar(
     all_slope = [c for v in slope_cols_by_cohort.values() for c in v]
     all_inter = base_cols + all_slope
 
+    # H4 fix: explicit name-to-index map so slope lookups never rely on
+    # implicit ordering. Column 0 is the constant; columns [1:] are
+    # X_cols = [f"{c}_dm" for c in all_inter] + ctrl_dm_cols in order.
+    coef_index: Dict[str, int] = {"_const": 0}
+    for i, col in enumerate(all_inter, start=1):
+        coef_index[col] = i
+
     # Demean interactions via FE projection
     for col in all_inter:
         u_m = df.groupby(group)[col].transform("mean")
@@ -597,11 +614,14 @@ def _etwfe_with_xvar(
     df_resid = max(n_obs - X.shape[1], 1)
 
     k = len(cohorts)
-    p_x = len(xvar)
+    p_x = len(xvar)  # retained for the single-xvar backward-compat block
 
     cohort_results = []
-    for i, g in enumerate(cohorts):
-        b_idx = 1 + i
+    for g in cohorts:
+        # H4 fix: look up baseline and slope indices by coefficient name
+        # rather than by arithmetic position, so future column-order
+        # changes cannot silently mis-attribute.
+        b_idx = coef_index[f"_coh{int(g)}_post"]
         att = float(beta[b_idx])
         att_se = float(se[b_idx])
         p = float(2 * (1 - stats.t.cdf(abs(att / att_se) if att_se > 0 else 0,
@@ -610,9 +630,8 @@ def _etwfe_with_xvar(
             "cohort": int(g),
             "att_at_xmean": att, "att_se": att_se, "att_pvalue": p,
[…]
