### 2026-04-15 80768a2
feat(spatial.gmm): Kelejian-Prucha GMM estimators (SEM/SAR/SARAR)
Tasks 16-17 of SP-01 S1.1. Adds sphet / spreg.GM_* equivalents:

- sem_gmm (Kelejian & Prucha 1999): 3-moment GMM for spatial-error λ,
  with optional heteroskedasticity-robust β SE (Arraiz et al. 2010).
- sar_gmm (Kelejian & Prucha 1998): 2SLS with spatial-lag instruments
  [X, WX, …, W^q X]; default q=1 matches spreg.GM_Lag.
- sarar_gmm / GM_Combo: SAR GMM → SEM GMM on residuals → Cochrane-Orcutt
  GLS re-estimation for final β (matches spreg.GM_Combo output).

Cross-validated against spreg 1.9.0 on the Columbus benchmark:
  G
[…]
