+++ src/statspai/spatial/models/gmm.py
@@ -1,0 +1,305 @@
+"""Generalised Method of Moments estimators for spatial regression models.
+
+Follows Kelejian & Prucha (1998, 1999) for SEM GMM and SAR 2SLS-GMM, and
+Arraiz et al. (2010) for the heteroskedasticity-robust variants used in
+``sphet`` / ``spreg.GM_*_Het``. Designed for large-N settings where dense
+eigenvalue-based ML is intractable.
+
+- :func:`sem_gmm`    — Kelejian-Prucha (1999) 3-moment GMM for ``λ``.
+- :func:`sar_gmm`    — Kelejian-Prucha (1998) 2SLS with spatial-lag
+  instruments (columns of ``[X, WX, W²X]``).
+- :func:`sarar_gmm`  — combined lag + error (``GM_Combo`` in spreg).
+"""
+from __future__ import annotations
+
+from typing import Optional
+
+import numpy as np
+import pandas as pd
+from scipy.optimize import minimize
+
+from ...core.results import EconometricResults
+from .ml import _coerce_W, _parse_formula
+
+
+# --------------------------------------------------------------------- #
+#  SEM GMM (Kelejian-Prucha 1999)
+# --------------------------------------------------------------------- #
+
+def _kp_moment_residuals(u: np.ndarray, W, lam: float, sigma2: float) -> np.ndarray:
+    """Three KP 1999 moment residuals given residuals ``u`` and candidate
+    ``(λ, σ²)``.
+
+    Uses the OLS residuals (``u``) to form ``v = W u`` and
+    ``w_bar = W² u`` once; returns the 3-vector of deviations from the
+    theoretical moments under ``u = ε + λ W ε``.
+    """
+    n = u.shape[0]
+    v = W @ u
+    w_bar = W @ v
+    tr_WtW_over_n = float((W.multiply(W)).sum()) / n
+    # Moment 1
+    m1 = (u @ u - 2 * lam * u @ v + lam ** 2 * v @ v) / n - sigma2
+    # Moment 2
+    m2 = (v @ v - 2 * lam * v @ w_bar + lam ** 2 * w_bar @ w_bar) / n - sigma2 * tr_WtW_over_n
+    # Moment 3
+    m3 = (u @ v - lam * (u @ w_bar + v @ v) + lam ** 2 * v @ w_bar) / n
+    return np.array([m1, m2, m3])
+
+
+def sem_gmm(W, data: pd.DataFrame, formula: str,
+            row_normalize: bool = True,
+            robust: Optional[str] = None) -> EconometricResults:
+    """Kelejian-Prucha (1999) GMM for the spatial-error parameter λ.
+
+    Stage 1 — OLS on ``y = Xβ + u`` ⇒ residuals ``u``.
+    Stage 2 — minimise sum of squared KP moment residuals over ``(λ, σ²)``.
+    Stage 3 — GLS with ``(I - λ̂ W)`` → final ``β̂``.
+
+    Parameters
+    ----------
+    robust : {None, "het"}
+        When ``"het"``, the final β covariance uses the heteroscedasticity-
+        robust sandwich (Arraiz et al. 2010 / ``spreg.GM_Error_Het``).
+    """
+    y, X, dep, indep = _parse_formula(formula, data)
+    n = len(y)
+    M = _coerce_W(W, n_expected=n, row_normalize=row_normalize)
+
+    # Stage 1 — OLS
+    XtX_inv = np.linalg.inv(X.T @ X)
+    beta_ols = XtX_inv @ (X.T @ y)
+    u = y - X @ beta_ols
+
+    # Stage 2 — minimise sum of squared moment residuals
+    def obj(theta):
+        lam, s2 = float(theta[0]), float(theta[1])
+        if not (-0.99 < lam < 0.99) or s2 <= 0:
+            return 1e20
[…]
