### 2026-04-15 73cf984
chore(release): bump version to 0.7.1
DID-focused polish release. See CHANGELOG.md [0.7.1] for the full
list of additions and fixes:

- ETWFE full parity with R etwfe (xvar multi, panel=False, cgroup,
  etwfe_emfx with four aggregations + include_leads).
- One-call method-robustness workflow (did_summary + plot + export
  bundle via did_report).
- Twelve review-round fixes across blocker (C1-C5) and high-severity
  (H1-H7) categories; regression tests locked in.

27 DID-focused pytest cases pass (tests/test_did_summary.py).

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
