### 2026-04-15 a5348ab
feat(spatial.gwr): GWR + MGWR with AICc/CV bandwidth selection
SP-01 S1.2: Layer 4 of the spatial full-stack. New subpackage
statspai.spatial.gwr providing:

- gwr(coords, y, X, bw, kernel, fixed) — local weighted least squares
  per observation with Gaussian / bisquare / exponential kernels and
  fixed-distance or adaptive-NN bandwidth. Returns per-obs (n × k)
  parameter matrix, predictions, residuals, local R², trace(S),
  AIC / AICc / BIC.

- gwr_bandwidth(…, criterion='AICc') — golden-section search over
  integer (adaptive) or continuous (fixed) bandwidth space,
  optimising AICc / AIC /
[…]
