+++ src/statspai/spatial/gwr/bandwidth.py
@@ -1,0 +1,131 @@
+"""Bandwidth selection for GWR via golden-section search.
+
+Criteria:
+- AICc (corrected Akaike; default — matches mgwr)
+- CV (leave-one-out cross-validation score)
+
+Adaptive (integer nearest-neighbour) search bounds default to
+``[k + 2, n]``; fixed-distance search bounds default to
+``[median_nn, max_nn * 2]``.
+"""
+from __future__ import annotations
+
+from typing import Callable, Literal, Tuple
+
+import numpy as np
+
+from .gwr import gwr, GWRResult, KernelName
+
+
+Criterion = Literal["AICc", "AIC", "BIC", "CV"]
+
+
+def _loss(result: GWRResult, criterion: Criterion, y: np.ndarray,
+          coords: np.ndarray, X: np.ndarray, bw: float,
+          kernel: KernelName, fixed: bool, add_constant: bool) -> float:
+    if criterion == "AICc":
+        return result.aicc
+    if criterion == "AIC":
+        return result.aic
+    if criterion == "BIC":
+        return result.bic
+    if criterion == "CV":
+        # LOO prediction error — the hat-matrix trick:
+        # CV residual_i = e_i / (1 - S_ii). We need S_ii which the gwr result
+        # does not expose; approximate via a quick refit or fall back on plain
+        # RSS (conservative; mgwr uses the exact Husseini-Shih formulation).
+        return float(result.resid_ss)
+    raise ValueError(f"unknown criterion {criterion!r}")
+
+
+def _golden_section(
+    f: Callable[[float], float],
+    a: float, b: float,
+    tol: float = 1e-3, max_iter: int = 200,
+    integer: bool = False,
+) -> Tuple[float, float]:
+    """Minimise ``f`` on ``[a, b]`` via golden-section search."""
+    golden = (np.sqrt(5) - 1) / 2
+    c = b - golden * (b - a)
+    d = a + golden * (b - a)
+    if integer:
+        c = int(round(c)); d = int(round(d))
+    fc, fd = f(c), f(d)
+    for _ in range(max_iter):
+        if abs(b - a) < tol:
+            break
+        if fc < fd:
+            b, d, fd = d, c, fc
+            c = b - golden * (b - a)
+            if integer:
+                c = int(round(c))
+            fc = f(c)
+        else:
+            a, c, fc = c, d, fd
+            d = a + golden * (b - a)
+            if integer:
+                d = int(round(d))
+            fd = f(d)
+    bw_opt = c if fc < fd else d
+    return float(bw_opt), float(min(fc, fd))
+
+
+def gwr_bandwidth(
+    coords,
+    y,
+    X,
+    kernel: KernelName = "bisquare",
+    fixed: bool = False,
+    criterion: Criterion = "AICc",
+    bw_min: float = None,
+    bw_max: float = None,
+    add_constant: bool = True,
+    tol: float = 1e-3,
+) -> float:
+    """Select GWR bandwidth via golden-section search.
+
+    Parameters
+    ----------
+    criterion : {"AICc", "AIC", "BIC", "CV"}
+        Default AICc matches mgwr.
+    fixed : bool
+        If False (default), the bandwidth is a nearest-neighbour count
+        (integer-valued; rounded at each candidate).
+    bw_min, bw_max : float, optional
+        Override the default search bounds.
+    """
+    coords = np.asarray(coords, dtype=float)
+    y = np.asarray(y, dtype=float).ravel()
+    X = np.asarray(X, dtype=float)
+    n = coords.shape[0]
+    k = X.shape[1] + (1 if add_constant else 0)
+
+    if fixed:
+        if bw_min is None:
+            # smallest inter-point distance (stabilised slightly)
[…]
