### 2026-04-15 d8cf177
feat(spatial.panel): spatial panel data estimators (SAR/SEM/SDM × FE/twoways)
SP-01 S1.3: Layer 5 of the spatial full-stack. New subpackage
statspai.spatial.panel providing splm-equivalent fixed-effects
spatial panel estimators via concentrated ML (Elhorst 2014,
Lee & Yu 2010):

- y_it = ρ W y_it + X_it β + μ_i + ε_it       (SAR-FE)
- y_it = X_it β + μ_i + u_it, u = λWu + ε     (SEM-FE)
- y_it = ρWy_it + X_it β + W X_it θ + μ_i + ε (SDM-FE)

Implementation:
- Within transform removes entity FE (and time FE with effects='twoways');
  constant is absorbed by the transform so formula must exclude
[…]
