+++ src/statspai/spatial/panel/__init__.py
@@ -1,0 +1,4 @@
+"""Spatial panel econometrics (R ``splm`` / ``spreg.ML_LagFE`` etc.)."""
+from .estimator import spatial_panel, SpatialPanelResult
+
+__all__ = ["spatial_panel", "SpatialPanelResult"]
