+++ src/statspai/spatial/panel/estimator.py
@@ -1,0 +1,262 @@
+"""Spatial panel data estimators — SAR / SEM / SDM with entity fixed effects.
+
+Balanced panel:  ``y_it = ρ W y_it + X_it β + μ_i + ε_it`` (SAR-FE)
+               ``y_it = X_it β + μ_i + u_it;  u_it = λ W u_it + ε_it`` (SEM-FE)
+               ``y_it = ρ W y_it + X_it β + W X_it θ + μ_i + ε_it`` (SDM-FE)
+
+Within-transform removes ``μ_i``; concentrated ML on the demeaned variables
+gives ρ (or λ), β, σ². Log-determinant is ``T * log|I_N − ρ W|`` (small-N
+eigenvalue path).
+
+Two-way fixed effects (``effects='twoways'``) demean along both dimensions.
+Random effects are not implemented here (splm's ``spreml``) — open a
+separate sub-spec if needed.
+
+References
+----------
+Elhorst, J.P. (2014). *Spatial Econometrics: From Cross-Sectional Data to
+  Spatial Panels*. Springer.
+Lee, L.-F. & Yu, J. (2010). "Estimation of spatial autoregressive panel
+  data models with fixed effects." *JoE*, 154(2), 165-185.
+"""
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Literal
+
+import numpy as np
+import pandas as pd
+from scipy import sparse
+from scipy.optimize import minimize_scalar
+
+from ..weights.core import W as _W
+from ..models.ml import _coerce_W
+
+
+EffectKind = Literal["fe", "twoways"]
+ModelKind = Literal["sar", "sem", "sdm"]
+
+
+# --------------------------------------------------------------------- #
+#  Balanced-panel reshaping
+# --------------------------------------------------------------------- #
+
+def _balanced_panel_matrix(
+    data: pd.DataFrame, entity: str, time: str, var: str
+) -> np.ndarray:
+    """Return (N, T) matrix: rows = entities (sorted), columns = time (sorted)."""
+    pivot = data.pivot(index=entity, columns=time, values=var)
+    if pivot.isna().any().any():
+        raise ValueError(
+            "Panel is unbalanced — every (entity, time) cell must be present."
+        )
+    return pivot.to_numpy(dtype=float)
+
+
+def _within_transform(arr: np.ndarray, effects: EffectKind) -> np.ndarray:
+    """Demean an (N, T) array within entities (and within time if twoways)."""
+    out = arr - arr.mean(axis=1, keepdims=True)          # entity demean
+    if effects == "twoways":
+        out = out - arr.mean(axis=0, keepdims=True) + arr.mean()
+    return out
+
+
+# --------------------------------------------------------------------- #
+#  Result dataclass
+# --------------------------------------------------------------------- #
+
+@dataclass
+class SpatialPanelResult:
+    params: pd.Series            # [x1, x2, …, ρ or λ]
+    std_errors: pd.Series
+    model: ModelKind
+    effects: EffectKind
+    spatial_param: str           # "rho" or "lambda"
+    spatial_param_value: float
+    sigma2: float
+    log_likelihood: float
+    residuals: np.ndarray        # (N, T)
+    N: int
+    T: int
+
+    def summary(self) -> str:
+        lines = [
+            f"Spatial Panel ({self.model.upper()}, {self.effects})",
+            "-" * 50,
+            f"Entities N : {self.N}",
+            f"Periods  T : {self.T}",
+            f"NT         : {self.N * self.T}",
[…]
