### 2026-04-16 8b6faff
feat(regression): out-of-sample predict() for OLS and IV
SP-03 Task 1 (infrastructure completion). Closes the most-touched gap
flagged in the ecosystem gap analysis — OLS/IV prediction used to raise
NotImplementedError on out-of-sample data.

OLS:
- predict(data, what='mean'|'confidence'|'prediction', alpha=0.05)
- 'mean'       → point forecast only
- 'confidence' → CI for E[y|x] using x' Σ_β x
- 'prediction' → wider interval adding σ² (new-observation variance)
- Design matrix for new data is rebuilt from the RHS of the formula
  via patsy.dmatrix() so the dependent variable is NOT required a
[…]
