### 2026-04-16 b06b389
feat(timeseries): local_projections (Jordà 2005) with Newey-West HAC
SP-04 Task 1. Adds the modern-macro workhorse impulse response
estimator:

    y_{t+h} - y_{t-1} = α_h + β_h·shock_t + γ_h·controls_{t-1} + ε

fitted horizon-by-horizon with Newey-West HAC standard errors
(truncation lag default = round(1.5·H)). Returns a result object
with horizons, IRF, SE, CI, ``.plot()``, ``.to_frame()``, ``.summary()``.

On an AR shock DGP with known IRF {1.0, 1.0, 0.7, 0.4, 0.1, 0, …}:
recovers h=0 impact ≈ 1.0, decay after peak to <0.3 by h=8, CI
properly brackets the point estimates.

Exposed at sp.lo
[…]
