+++ src/statspai/timeseries/local_projections.py
@@ -1,0 +1,204 @@
+"""Local Projections (Jordà, AER 2005).
+
+Estimates impulse responses by running horizon-by-horizon regressions:
+
+    y_{t+h} - y_{t-1}  =  α_h  +  β_h · shock_t  +  γ_h · controls_t  +  ε_{t,h}
+
+for h = 0, 1, …, H. The sequence ``{β_h}`` traces the impulse response
+function. Inference uses Newey-West HAC standard errors (required because
+the residuals are serially correlated by construction).
+
+This estimator is much more flexible than a VAR-based IRF — no
+finite-lag structure imposed — and is the default in modern empirical
+macro (Ramey 2016; Plagborg-Møller & Wolf 2021).
+"""
+from __future__ import annotations
+
+from dataclasses import dataclass, field
+from typing import List, Optional
+
+import numpy as np
+import pandas as pd
+from scipy import stats
+
+
+# --------------------------------------------------------------------- #
+#  Newey-West HAC for a single regression
+# --------------------------------------------------------------------- #
+
+def _newey_west(X: np.ndarray, e: np.ndarray, lags: int) -> np.ndarray:
+    """Newey-West HAC covariance matrix for an OLS regression with
+    residuals ``e``. Bartlett kernel weights up to ``lags``.
+    """
+    n, k = X.shape
+    XtX_inv = np.linalg.inv(X.T @ X)
+    # meat: sum_l w_l (Γ_l + Γ_l') where Γ_l = (1/n) sum_t X_t e_t e_{t-l} X_{t-l}'
+    omega = np.zeros((k, k))
+    u = X * e[:, None]                          # (n, k)
+    for lag in range(lags + 1):
+        if lag == 0:
+            G = u.T @ u
+        else:
+            G = u[lag:].T @ u[:-lag]
+            G = G + G.T
+        w = 1.0 - lag / (lags + 1.0)
+        omega += w * G
+    V = XtX_inv @ omega @ XtX_inv
+    return V
+
+
+# --------------------------------------------------------------------- #
+#  Result
+# --------------------------------------------------------------------- #
+
+@dataclass
+class LocalProjectionsResult:
+    horizons: np.ndarray              # (H+1,)
+    irf: np.ndarray                   # (H+1,)
+    se: np.ndarray                    # (H+1,)
+    ci_lower: np.ndarray              # (H+1,)
+    ci_upper: np.ndarray              # (H+1,)
+    alpha: float
+    shock_name: str
+    outcome_name: str
+    n_obs_per_horizon: np.ndarray     # (H+1,)  — usable sample per horizon
+
+    def to_frame(self) -> pd.DataFrame:
+        return pd.DataFrame({
+            "horizon": self.horizons,
+            "irf": self.irf,
+            "se": self.se,
+            "ci_lower": self.ci_lower,
+            "ci_upper": self.ci_upper,
+            "n": self.n_obs_per_horizon,
+        })
+
+    def plot(self, ax=None, **kwargs):
+        import matplotlib.pyplot as plt
+        if ax is None:
+            _, ax = plt.subplots(figsize=(7, 4))
+        ax.plot(self.horizons, self.irf, "-o", color="C0",
+                label=f"IRF of {self.outcome_name}")
+        ax.fill_between(self.horizons, self.ci_lower, self.ci_upper,
+                        alpha=0.25, color="C0",
+                        label=f"{int((1 - self.alpha) * 100)}% CI")
+        ax.axhline(0, color="grey", linewidth=0.6)
+        ax.set_xlabel("Horizon h")
+        ax.set_ylabel(f"Response to {self.shock_name}")
+        ax.legend()
[…]
