### 2026-04-16 332b705
fix(weak-iv): add lowercase aliases for first_stage_f / effective_f
Tests in test_weakiv_tobit.py expect lowercase keys but the dict only
exposed the capital-F versions. Add both for backward compatibility —
no existing callers of the capital-F keys are broken, and the 17
weak-IV tests now pass.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
