+++ src/statspai/diagnostics/weak_iv.py
@@ -479,7 +479,9 @@ def anderson_rubin_test(
         "ar_ci": ar_ci,
         "beta_2sls": beta_2sls,
         "first_stage_F": f_first,
+        "first_stage_f": f_first,
         "effective_F": f_eff,
+        "effective_f": f_eff,
         "tF_critical_value": tF_c,
         "n_instruments": k_z,
         "strength": f_result["strength"],
