### 2026-04-16 38b6f99
feat(causal_discovery): DirectLiNGAM — bit-identical match vs lingam package
SP-05 Task 1. Adds Shimizu et al. (2011) DirectLiNGAM — the non-Gaussian
acyclic causal discovery algorithm. Unlike constraint-based PC (already
in StatsPAI), LiNGAM uniquely identifies the causal order under its
assumptions (linear, acyclic, non-Gaussian disturbances, no latents).

Implementation uses Hyvärinen's negentropy approximation for the
independence score and the difference-of-MI exogeneity test
(Shimizu 2011 Lemma 2); B is recovered in standardised coordinates
then de-standardised for the user.

Cross-valid
[…]
