+++ src/statspai/causal_discovery/lingam.py
@@ -1,0 +1,225 @@
+"""LiNGAM — Linear Non-Gaussian Acyclic Model (Shimizu et al., JMLR 2006).
+
+Identifies a causal DAG under the assumptions:
+
+1. Causal structure is linear and acyclic.
+2. All disturbances are mutually independent and non-Gaussian.
+3. No latent common causes.
+
+Under these assumptions the causal order is **uniquely identifiable**
+from purely observational data (in contrast with constraint-based
+methods like PC which only identify a Markov-equivalence class).
+
+This implementation uses the DirectLiNGAM algorithm (Shimizu 2011):
+
+    1. Pick the variable most independent of the others' residuals
+       (Hilbert-Schmidt Independence Criterion) → declare it a source.
+    2. Regress the remaining variables on the source, take residuals.
+    3. Recurse on the residuals until all variables are ordered.
+    4. The adjacency matrix B[i, j] of direct effects is then recovered
+       by regressing each variable on its predecessors in the ordering.
+"""
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import List, Optional
+
+import numpy as np
+import pandas as pd
+
+
+# --------------------------------------------------------------------- #
+#  Helper: pairwise independence measure
+# --------------------------------------------------------------------- #
+
+def _negentropy_hyvarinen(x: np.ndarray) -> float:
+    """Hyvärinen's approximation of differential entropy / negentropy.
+
+    Uses the two non-polynomial non-linearities recommended in Hyvärinen
+    (1998, *Neural Networks*) eq. (7). Larger values indicate stronger
+    departure from Gaussianity (which DirectLiNGAM exploits to
+    identify direction).
+    """
+    z = (x - x.mean()) / (x.std() + 1e-12)
+    k1 = 36.0 / (8.0 * np.sqrt(3.0) - 9.0)
+    k2a = 1.0 / (2.0 - 6.0 / np.pi)
+    gamma = 0.37457
+    t1 = k1 * (float(np.mean(np.log(np.cosh(z)))) - gamma) ** 2
+    t2 = k2a * (float(np.mean(z * np.exp(-z ** 2 / 2.0)))) ** 2
+    return t1 + t2
+
+
+def _mi_quadratic(x: np.ndarray, y: np.ndarray) -> float:
+    """Approximation of mutual information via Hyvärinen's negentropy.
+
+    Follows Shimizu (2011) eq. (5):
+
+        M(x, y) ≈ H(x) + H(y) - H(x, y)
+               ≈ -J(x) - J(y) + J((x+y)/√2)
+
+    but because J is non-negative and we only use this score for
+    *ordering* candidates, we use the simpler pairwise form:
+
+        dep(x, y)  =  J(x_std) + J(y_std) - J_joint_proxy
+
+    Concretely, we return a non-negative score that is 0 for
+    independent x, y and grows with non-Gaussian dependence.
+    """
+    x = x - x.mean()
+    y = y - y.mean()
+    nx = _negentropy_hyvarinen(x)
+    ny = _negentropy_hyvarinen(y)
+    # Proxy: non-Gaussianity of the projection (x+y)/√2
+    nxy = _negentropy_hyvarinen((x + y) / np.sqrt(2.0))
+    return max(nx + ny - nxy, 0.0)
+
+
+# --------------------------------------------------------------------- #
+#  Result
+# --------------------------------------------------------------------- #
+
+@dataclass
+class LiNGAMResult:
+    order: List[int]                 # causal order (most exogenous first)
+    adjacency: np.ndarray            # B[i, j] = direct effect of j on i
+    names: List[str]
+    residuals: np.ndarray            # (n, k) final residuals
+
[…]
