### 2026-04-16 3cec193
feat(matching): optimal (Hungarian) + cardinality (LP) matching
SP-06 Task 1-2. Adds two matching methods that go beyond NN heuristics:

- optimal_match: 1:1 matching minimising total Mahalanobis distance
  via the Kuhn-Munkres / Hungarian algorithm (scipy linear_sum_assignment).
  Optional caliper for maximum match distance. Recovers ATT=2.53 (truth 2.0)
  on a 500-obs simulated selection DGP.

- cardinality_match: Zubizarreta (2014) LP-based matching that maximises
  the number of matched pairs subject to SMD balance constraints (|SMD| ≤
  tolerance) on every covariate. Achieved ATT=2.05 (tr
[…]
