+++ src/statspai/matching/optimal.py
@@ -1,0 +1,253 @@
+"""Optimal and cardinality matching.
+
+Two approaches that go beyond the "nearest-neighbour" heuristic:
+
+- :func:`optimal_match` — 1:1 matching that minimises the **total**
+  distance across all matched pairs, solved exactly via the
+  Hungarian / Kuhn-Munkres algorithm
+  (``scipy.optimize.linear_sum_assignment``).
+
+- :func:`cardinality_match` — Zubizarreta et al. (2014) cardinality
+  matching: keep as many treated units matched to at least one control
+  as possible, subject to covariate-balance constraints solved by
+  linear programming.
+"""
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import List, Optional
+
+import numpy as np
+import pandas as pd
+from scipy import optimize
+
+
+@dataclass
+class OptimalMatchResult:
+    pairs: pd.DataFrame              # (n_matched, 2) treated_idx + control_idx
+    distances: np.ndarray            # (n_matched,) matching distances
+    ate: float                       # average treatment effect (ATT)
+    se: float                        # (rough) analytic SE on matched pairs
+    n_treated: int
+    n_matched: int
+
+    def summary(self) -> str:
+        lines = [
+            "Optimal 1:1 Matching (Hungarian algorithm)",
+            "-" * 40,
+            f"Treated (T=1) : {self.n_treated}",
+            f"Matched       : {self.n_matched}",
+            f"Mean distance : {self.distances.mean():.4f}",
+            f"ATT           : {self.ate:.4f}  (SE = {self.se:.4f})",
+        ]
+        return "\n".join(lines)
+
+    def __repr__(self) -> str:
+        return self.summary()
+
+
+def _distance_matrix(
+    X_treat: np.ndarray, X_ctrl: np.ndarray, metric: str
+) -> np.ndarray:
+    if metric == "mahalanobis":
+        X_all = np.vstack([X_treat, X_ctrl])
+        cov = np.cov(X_all, rowvar=False) + 1e-8 * np.eye(X_all.shape[1])
+        cov_inv = np.linalg.inv(cov)
+        D = np.empty((X_treat.shape[0], X_ctrl.shape[0]))
+        for i in range(X_treat.shape[0]):
+            diff = X_ctrl - X_treat[i]
+            D[i] = np.sqrt(np.einsum("ij,jk,ik->i", diff, cov_inv, diff))
+        return D
+    if metric == "euclidean":
+        diff = X_treat[:, None, :] - X_ctrl[None, :, :]
+        return np.linalg.norm(diff, axis=2)
+    raise ValueError(f"unknown metric {metric!r}")
+
+
+def optimal_match(
+    data: pd.DataFrame,
+    treatment: str,
+    outcome: str,
+    covariates: List[str],
+    metric: str = "mahalanobis",
+    caliper: Optional[float] = None,
+) -> OptimalMatchResult:
+    """Optimal 1:1 matching via the Hungarian algorithm.
+
+    Each treated unit is matched to exactly one control; the total
+    sum of matched distances is globally minimised. Requires
+    ``n_treated ≤ n_control``.
+
+    Parameters
+    ----------
+    caliper : float, optional
+        Drop any pair with distance greater than ``caliper``.
+    """
+    df = data.dropna(subset=[treatment, outcome] + covariates).reset_index(drop=True)
+    t = df[treatment].to_numpy().astype(int)
+    y = df[outcome].to_numpy(dtype=float)
+    X = df[covariates].to_numpy(dtype=float)
+    treated_idx = np.where(t == 1)[0]
[…]
