### 2026-04-16 197ce18
feat(decomposition): RIF regression + Oaxaca-Blinder (Firpo-Fortin-Lemieux 2009)
SP-08 Task 1. Adds unconditional quantile partial effects and distributional
decomposition — two features that R's dineq package provides but no Python
package had in a unified interface:

- rifreg(formula, data, statistic, tau) — regress Recentered Influence
  Function values on covariates. Supported statistics: quantile(τ),
  variance, Gini. The coefficients estimate how a covariate shift affects
  the distributional statistic of the unconditional outcome, not the
  conditional (unlike standard quantile regressi
[…]
