+++ src/statspai/decomposition/rif.py
@@ -1,0 +1,250 @@
+"""Recentered Influence Function (RIF) regression and decomposition
+(Firpo, Fortin & Lemieux, *Econometrica* 2009).
+
+The RIF of a distributional statistic ``ν(F)`` evaluated at observation
+``y_i`` is
+
+    RIF(y_i; ν, F) = ν(F) + IF(y_i; ν, F)
+
+where ``IF`` is the (classical) influence function.
+
+**RIF-OLS** regresses RIF values on ``X`` to estimate the *unconditional
+quantile partial effect* (UQPE) — how a small shift in ``X`` affects the
+quantile of the unconditional distribution, not the conditional distribution.
+
+**RIF-Oaxaca-Blinder** decomposes the between-group difference in ``ν`` into
+explained (covariate endowment) and unexplained (coefficient) components at
+any chosen distributional statistic, not just the mean.
+
+Supported statistics: ``quantile(τ)``, ``variance``, ``gini``.
+"""
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Literal
+
+import numpy as np
+import pandas as pd
+from scipy import stats as sp_stats
+
+
+StatisticKind = Literal["quantile", "variance", "gini"]
+
+
+# --------------------------------------------------------------------- #
+#  Influence function construction
+# --------------------------------------------------------------------- #
+
+def _kernel_density_at(y: np.ndarray, point: float, bw: str = "silverman") -> float:
+    """Kernel density estimate at a single point."""
+    try:
+        kde = sp_stats.gaussian_kde(y, bw_method=bw)
+        return float(kde(point))
+    except Exception:
+        # fallback: histogram-based density
+        h = 1.06 * y.std() * len(y) ** (-0.2)
+        return float(np.mean(np.exp(-0.5 * ((y - point) / h) ** 2)) / (h * np.sqrt(2 * np.pi)))
+
+
+def rif_values(y: np.ndarray, statistic: StatisticKind = "quantile",
+               tau: float = 0.5) -> np.ndarray:
+    """Compute the RIF of each observation.
+
+    Parameters
+    ----------
+    y : (n,) array
+    statistic : {"quantile", "variance", "gini"}
+    tau : float
+        Quantile level (only used when ``statistic="quantile"``).
+    """
+    y = np.asarray(y, dtype=float).ravel()
+    n = y.size
+    if statistic == "quantile":
+        q = float(np.quantile(y, tau))
+        f_q = _kernel_density_at(y, q)
+        f_q = max(f_q, 1e-12)
+        return q + (tau - (y <= q).astype(float)) / f_q
+    if statistic == "variance":
+        mu = y.mean()
+        return (y - mu) ** 2
+    if statistic == "gini":
+        # Gini influence function (Firpo et al. 2009 Appendix A)
+        mu = y.mean()
+        n_obs = len(y)
+        ranks = sp_stats.rankdata(y, method="average") / n_obs
+        G = float(2 * np.cov(y, ranks)[0, 1] / mu)    # Gini coefficient
+        rif = (
+            -G
+            + (2 / mu) * (ranks * y - np.cumsum(np.sort(y)) / n_obs)
+            - (2 * y / mu) * (ranks - (G + 1) / 2)
+        )
+        return G + rif
+    raise ValueError(f"unknown statistic {statistic!r}")
+
+
+# --------------------------------------------------------------------- #
+#  RIF-OLS
+# --------------------------------------------------------------------- #
+
+@dataclass
+class RIFResult:
+    params: pd.Series
+    std_errors: pd.Series
+    statistic: str
[…]
