### 2026-04-16 60bb1cb
feat(timeseries): GARCH(p,q) volatility model with MLE
SP-04 Task 2. Adds the most-requested volatility estimator for
financial / macro time series:

  σ²_t = ω + Σ α_i ε²_{t-i} + Σ β_j σ²_{t-j}

Estimated by conditional Gaussian MLE (Nelder-Mead). Returns GARCHResult
with: omega/alpha/beta params, conditional variance path σ²_t,
standardised residuals z_t, persistence, multi-step forecast(),
AIC/BIC, summary().

On a GARCH(1,1) DGP with (ω,α,β) = (0.01, 0.10, 0.85), T=2000:
recovered persistence 0.91 (true 0.95), α ≈ 0.12 (true 0.10),
std residuals variance near 1.0. Finite-sample bias is exp
[…]
