### 2026-04-16 be2a1a9
feat(rd): rdpower + rdsampsi — power/sample-size for RD designs
SP-07 Task 1. Ports the core functionality of Cattaneo-Titiunik-
Vazquez-Bare (2019) RD power calculations — previously R-only via
the rdpower package — to StatsPAI.

- rdpower(tau, var_left, var_right, n_left, n_right, …): compute power
  given sample size and hypothesised effect; returns MDE at 80%.
- rdsampsi(tau, target_power, …): minimum sample size for a target power.

Both based on the asymptotic distribution of the local-polynomial
estimator with MSE-optimal bandwidth calibration. Kernel constant C_K
= 0.35 (triangular). V
[…]
