+++ src/statspai/rd/rdpower.py
@@ -1,0 +1,177 @@
+"""RD Power and Sample-Size Calculations (Cattaneo, Titiunik & Vazquez-Bare, 2019).
+
+Provides two entry points:
+
+- :func:`rdpower`  — compute power of an RD design given a sample size
+  (or a MDE given target power).
+- :func:`rdsampsi` — compute minimum sample size for a target power.
+
+Both are based on the asymptotic distribution of the local-polynomial
+point estimator (as in ``rdrobust``). They assume MSE-optimal bandwidth
+for the variance calibration and a normal-approximation critical region.
+
+References
+----------
+Cattaneo, M.D., Titiunik, R. & Vazquez-Bare, G. (2019).
+  "Power calculations for regression-discontinuity designs."
+  *Stata Journal*, 19(1), 210–245.
+"""
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Optional
+
+import numpy as np
+from scipy import stats as sp_stats
+
+
+@dataclass
+class RDPowerResult:
+    power: float
+    mde: float
+    alpha: float
+    n_left: int
+    n_right: int
+    se: float
+    tau: float                       # assumed or hypothesised effect
+
+    def summary(self) -> str:
+        lines = [
+            "RD Power Calculation",
+            "-" * 40,
+            f"Hypothesised τ : {self.tau: .4f}",
+            f"SE(τ̂)         : {self.se: .4f}",
+            f"Alpha          : {self.alpha}",
+            f"Power          : {self.power: .4f}",
+            f"MDE (at 80%)   : {self.mde: .4f}",
+            f"n (left)       : {self.n_left}",
+            f"n (right)      : {self.n_right}",
+        ]
+        return "\n".join(lines)
+
+    def __repr__(self) -> str:
+        return self.summary()
+
+
+@dataclass
+class RDSampSiResult:
+    n_left: int
+    n_right: int
+    n_total: int
+    tau: float
+    target_power: float
+    alpha: float
+
+    def summary(self) -> str:
+        return (
+            f"RD Sample Size: N = {self.n_total} "
+            f"({self.n_left} left + {self.n_right} right) "
+            f"for τ = {self.tau:.4f} at power = {self.target_power:.0%}, "
+            f"α = {self.alpha}"
+        )
+
+    def __repr__(self) -> str:
+        return self.summary()
+
+
+def _rd_se(n_left: int, n_right: int,
+           var_left: float, var_right: float,
+           h_left: float, h_right: float) -> float:
+    """Approximate SE for the RD point estimator (local-linear).
+
+    SE ≈ sqrt(C_K * [σ²_L / (n_L h_L) + σ²_R / (n_R h_R)])
+
+    where C_K = ∫K²(u)du is the kernel constant (~0.35 for triangular).
+    """
+    Ck = 0.35
+    return float(np.sqrt(Ck * (var_left / max(n_left * h_left, 1.0) +
+                                var_right / max(n_right * h_right, 1.0))))
+
+
+def rdpower(
+    tau: float,
+    var_left: float = 1.0,
+    var_right: float = 1.0,
+    n_left: int = 500,
+    n_right: int = 500,
+    h_left: float = 0.5,
+    h_right: float = 0.5,
+    alpha: float = 0.05,
+    target_power: Optional[float] = None,
+) -> RDPowerResult:
+    """Power of an RD design given sample size and effect size.
+
+    Parameters
+    ----------
+    tau : float
+        Hypothesised treatment effect at the cutoff.
[…]
