### 2026-04-16 cfbd348
feat(mediation): sensitivity analysis (Imai, Keele & Yamamoto 2010)
SP-08 Task 2. Ports the ρ-sensitivity analysis from R's mediation
package — previously unavailable in Python:

  For each candidate ρ (mediator-outcome error correlation), compute a
  bias-adjusted ACME via omitted-variable algebra. The ρ at which ACME
  crosses zero is the "sensitivity point" — how much unobserved
  confounding would be needed to explain away the mediation effect.

On a DGP with true ACME = 0.40: baseline ACME recovered at 0.415;
sensitivity point ρ = 0.81 (i.e. very strong confounding needed to
nullify the e
[…]
