+++ src/statspai/mediation/sensitivity.py
@@ -1,0 +1,157 @@
+"""Mediation sensitivity analysis (Imai, Keele & Tingley, 2010).
+
+Under sequential ignorability (SI), the average causal mediation
+effect (ACME) is identified. But SI is untestable. The sensitivity
+parameter ``ρ`` quantifies the degree of unobserved confounding
+between the mediator and the outcome (the correlation between their
+error terms). When ρ = 0, SI holds; as |ρ| increases, the ACME may
+shrink to zero and even flip sign.
+
+- :func:`mediate_sensitivity` — over a grid of ρ values, re-estimate
+  the ACME with an imputed correlation between mediator and outcome
+  errors, and return the ACME as a function of ρ.
+- The ρ at which ACME = 0 is the *sensitivity point* (analogous to
+  the E-value in observational studies).
+
+References
+----------
+Imai, K., Keele, L. & Tingley, D. (2010). "A General Approach to
+  Causal Mediation Analysis." *Psych Methods*, 15(4), 309–334.
+Imai, K., Keele, L. & Yamamoto, T. (2010). "Identification, Inference
+  and Sensitivity Analysis for Causal Mediation Effects." *Stat Sci*,
+  25(1), 51–71.
+"""
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Optional
+
+import numpy as np
+import pandas as pd
+
+
+@dataclass
+class MediateSensitivityResult:
+    rho_grid: np.ndarray             # (n_grid,)
+    acme_at_rho: np.ndarray          # (n_grid,)
+    rho_at_zero: Optional[float]     # ρ where ACME crosses zero
+    acme_at_zero: float              # ACME at ρ=0 (the baseline)
+
+    def plot(self, ax=None, **kwargs):
+        import matplotlib.pyplot as plt
+        if ax is None:
+            _, ax = plt.subplots(figsize=(6, 4))
+        ax.plot(self.rho_grid, self.acme_at_rho, "-", color="C0")
+        ax.axhline(0, color="grey", linewidth=0.6)
+        ax.axvline(0, color="grey", linewidth=0.6, linestyle="--")
+        if self.rho_at_zero is not None:
+            ax.axvline(self.rho_at_zero, color="C3", linewidth=1, linestyle="--",
+                       label=f"ACME=0 at ρ={self.rho_at_zero:.2f}")
+            ax.legend()
+        ax.set_xlabel("ρ (mediator-outcome error correlation)")
+        ax.set_ylabel("ACME(ρ)")
+        ax.set_title("Mediation sensitivity analysis")
+        return ax
+
+    def summary(self) -> str:
+        lines = [
+            "Mediation Sensitivity (Imai et al. 2010)",
+            "-" * 45,
+            f"Baseline ACME (ρ=0) : {self.acme_at_zero: .4f}",
+        ]
+        if self.rho_at_zero is not None:
+            lines.append(f"ρ at which ACME = 0  : {self.rho_at_zero: .4f}")
+            lines.append(
+                f"Interpretation: unobserved confounding with |ρ| > "
+                f"{abs(self.rho_at_zero):.2f} would explain away the "
+                f"estimated mediation effect."
+            )
+        else:
+            lines.append("ACME does not cross zero in [-0.9, 0.9].")
+        return "\n".join(lines)
+
+    def __repr__(self) -> str:
+        return self.summary()
+
+
[…]
