### 2026-04-16 da306a5
feat(causal): GRF extensions — variable_importance, BLP, ate, att + leaf bugfix
SP-02 Task 1. Adds grf-inspired methods to CausalForest:

- variable_importance(): permutation-based feature importance for
  treatment-effect heterogeneity. Returns normalised scores (sum=1).
- best_linear_projection(): Chernozhukov et al. (2020) BLP — regress
  CATE on X to test for systematic heterogeneity (F-test on β₁=0).
- ate() / att(): convenience accessors for average effects.

BUGFIX: _replace_leaf_values_with_causal_effects was overwriting ALL
leaf nodes with each leaf's effect instead of only the matchi
[…]
