### 2026-04-16 de6db42
feat(survival): Cox proportional hazards with shared gamma frailty
SP-10 Task 1. Adds the frailty extension to the Cox model, previously
only available in R (survival::frailty, frailtypack):

    h(t|X,z) = z_i · h₀(t) · exp(X β)

where z_i ~ Gamma(θ,θ) is a cluster-level frailty (mean 1,
variance 1/θ). Estimated via EM: E-step updates frailties given β
and θ; M-step updates β via penalised partial likelihood (L-BFGS-B)
and θ via profile ML (bounded scalar optim).

On 300-obs DGP with 30 clusters, true β=0.5: recovered β=0.491
(concordance 0.65). Frailty variance estimation has known small-sam
[…]
