+++ src/statspai/survival/frailty.py
@@ -1,0 +1,183 @@
+"""Cox proportional hazards with shared (cluster) frailty.
+
+Extends the standard Cox model to account for unobserved heterogeneity
+(frailty) within clusters. The model is:
+
+    h(t | X_ij, z_i) = z_i · h₀(t) · exp(X_ij β)
+
+where z_i is a multiplicative frailty term, assumed iid from a
+gamma(θ, θ) distribution (mean 1, variance 1/θ). When θ → ∞ the
+frailties collapse to 1 and the model reduces to standard Cox.
+
+Estimation follows the penalised partial likelihood approach
+(Therneau & Grambsch 2000) — iteratively update β via Newton-Raphson
+on the penalised Cox partial log-likelihood and θ via profile ML.
+
+References
+----------
+Therneau, T.M. & Grambsch, P.M. (2000). *Modeling Survival Data:
+  Extending the Cox Model*. Springer.
+Duchateau, L. & Janssen, P. (2008). *The Frailty Model*. Springer.
+"""
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import List, Optional
+
+import numpy as np
+import pandas as pd
+from scipy.optimize import minimize_scalar
+
+from .models import _parse_formula, CoxResult
+
+
+@dataclass
+class FrailtyResult:
+    beta: np.ndarray
+    se: np.ndarray
+    var_names: List[str]
+    theta: float
+    frailties: np.ndarray
+    cluster_ids: np.ndarray
+    log_likelihood: float
+    n: int
+    n_events: int
+    n_clusters: int
+    concordance: float
+
+    def summary(self) -> str:
+        lines = [
+            "Cox Model with Shared Gamma Frailty",
+            "-" * 50,
+            f"Observations   : {self.n}",
+            f"Events         : {self.n_events}",
+            f"Clusters       : {self.n_clusters}",
+            f"Theta (frailty): {self.theta:.4f}  (frailty variance = {1/self.theta:.4f})",
+            f"Log-Lik        : {self.log_likelihood:.4f}",
+            f"Concordance    : {self.concordance:.4f}",
+            "",
+            "Coefficients:",
+        ]
+        from scipy import stats
+        for nm, b, s in zip(self.var_names, self.beta, self.se):
+            t = b / s if s > 0 else np.nan
+            p = 2 * (1 - stats.norm.cdf(abs(t)))
+            lines.append(
+                f"  {nm:<15s}  {b: .4f}  (SE {s: .4f}, z {t: .3f}, p {p: .4f})"
+            )
+        return "\n".join(lines)
+
+    def __repr__(self) -> str:
+        return self.summary()
+
+
+def cox_frailty(
+    formula: str,
+    data: pd.DataFrame,
+    cluster: str,
+    alpha: float = 0.05,
+    maxiter: int = 50,
+    tol: float = 1e-6,
+) -> FrailtyResult:
+    """Cox proportional hazards with shared gamma frailty.
+
+    Parameters
+    ----------
+    formula : str
+        ``"duration + event ~ x1 + x2"`` (like R's ``Surv(time, event) ~ x``).
+    data : pd.DataFrame
+    cluster : str
+        Column identifying clusters (e.g. hospital, site).
+    """
+    lhs_str, covariates = _parse_formula(formula)
+    # LHS is "duration + event" — split on "+"
+    lhs_parts = [s.strip() for s in lhs_str.split("+")]
+    if len(lhs_parts) != 2:
+        raise ValueError(
+            "formula LHS must be 'duration + event' (e.g. 'T + E ~ x1 + x2')"
+        )
+    duration_col, event_col = lhs_parts[0], lhs_parts[1]
[…]
