+++ src/statspai/survival/__init__.py
@@ -1,4 +1,8 @@
 """Survival and duration analysis models."""
 from .models import cox, kaplan_meier, survreg, CoxResult, KMResult, logrank_test
+from .frailty import cox_frailty, FrailtyResult
 
-__all__ = ["cox", "kaplan_meier", "survreg", "CoxResult", "KMResult", "logrank_test"]
+__all__ = [
+    "cox", "kaplan_meier", "survreg", "CoxResult", "KMResult", "logrank_test",
+    "cox_frailty", "FrailtyResult",
+]
