### 2026-04-16 147440e
feat(causal_discovery): GES — Greedy Equivalence Search (Chickering 2002)
SP-05 Task 2. Score-based structure learning via BIC-greedy search
through equivalence classes. Two phases: forward (add edges) + backward
(prune edges). Output is a CPDAG — the Markov equivalence class.

Correctly recovers chain skeleton x1—x2—x3 on 3-variable DGP; does not
hallucinate edges between independent variables. Complements the
existing constraint-based PC and non-Gaussian LiNGAM — StatsPAI now
has three complementary discovery algorithms covering the main families.

Exposed at sp.ges. 4 tests pass.

Co-Author
[…]
