### 2026-04-16 932bd37
feat(survey): raking + linear calibration (Deville-Särndal 1992)
SP-09 Task 1. Ports R's survey::calibrate() / survey::rake() to Python:

- rake(data, margins): iterative proportional fitting to match weighted
  marginal distributions to known population totals. Converges to
  minimum-entropy-distance weights.
- linear_calibration(data, totals): Deville-Särndal chi-squared distance
  calibration that exactly hits population totals for continuous auxiliaries.

On a biased sample (60% male, target 49%): raking hits the target within
0.01 tolerance; linear calibration hits income total exactly.


[…]
