+++ src/statspai/survey/calibration.py
@@ -1,0 +1,165 @@
+"""Survey calibration (raking / post-stratification / linear calibration).
+
+Adjusts design weights so that the weighted sample matches known
+population totals (margins) on a set of auxiliary variables. This is
+the ``survey::calibrate()`` and ``survey::rake()`` functionality from
+R's survey package — previously unavailable in a Python econometrics
+toolkit.
+
+Three calibration methods:
+
+- **Raking** (Deming & Stephan 1940) — iterative proportional fitting
+  that adjusts weights to match marginal distributions one variable
+  at a time. Converges to the minimum-entropy distance solution.
+- **Linear calibration** (Deville & Särndal 1992) — find weights
+  closest to the design weights (chi-squared distance) such that
+  weighted totals of X exactly equal population totals.
+- **Post-stratification** — special case where cells are defined
+  by the crossing of categorical variables.
+"""
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Dict, List, Optional
+
+import numpy as np
+import pandas as pd
+
+
+@dataclass
+class CalibrationResult:
+    calibrated_weights: np.ndarray
+    method: str
+    converged: bool
+    iterations: int
+    weight_summary: Dict[str, float]
+
+    def summary(self) -> str:
+        lines = [
+            f"Survey Calibration ({self.method})",
+            "-" * 40,
+            f"Converged : {self.converged} ({self.iterations} iterations)",
+            f"Min weight: {self.weight_summary['min']:.4f}",
+            f"Max weight: {self.weight_summary['max']:.4f}",
+            f"Mean weight: {self.weight_summary['mean']:.4f}",
+            f"CV(weight) : {self.weight_summary['cv']:.4f}",
+        ]
+        return "\n".join(lines)
+
+    def __repr__(self) -> str:
+        return self.summary()
+
+
+def rake(
+    data: pd.DataFrame,
+    margins: Dict[str, Dict],
+    weight: Optional[str] = None,
+    max_iter: int = 100,
+    tol: float = 1e-6,
+) -> CalibrationResult:
+    """Raking (iterative proportional fitting).
+
+    Parameters
+    ----------
+    data : pd.DataFrame
+    margins : dict
+        ``{column_name: {category: target_proportion}}``.
+        E.g. ``{"sex": {"M": 0.49, "F": 0.51}, "age_group": {"18-34": 0.3, ...}}``.
+    weight : str, optional
+        Existing design weight column. If ``None``, starts with equal weights.
+    """
+    df = data.copy()
+    n = len(df)
+    if weight is not None:
+        w = df[weight].to_numpy(dtype=float).copy()
+    else:
+        w = np.ones(n, dtype=float) / n
+
+    converged = False
+    for iteration in range(max_iter):
+        w_old = w.copy()
+        for col, targets in margins.items():
+            vals = df[col].to_numpy()
+            for cat, target in targets.items():
+                mask = vals == cat
+                if mask.sum() == 0:
+                    continue
+                current_total = w[mask].sum()
+                if current_total > 0:
+                    factor = target / current_total
+                    w[mask] *= factor
+        # Normalise
+        w = w / w.sum()
+        if np.max(np.abs(w - w_old)) < tol:
+            converged = True
+            break
+
+    return CalibrationResult(
+        calibrated_weights=w,
+        method="raking",
+        converged=converged,
+        iterations=iteration + 1,
+        weight_summary={
+            "min": float(w.min()),
[…]
