### 2026-04-16 975a3ab
feat(survey): raking + linear calibration (Deville-Särndal 1992)
SP-09 Task 1. Ports R's survey::calibrate() / survey::rake():

- rake(data, margins): IPF to match marginal distributions. Hits target
  within 0.01 tolerance on biased sample.
- linear_calibration(data, totals): chi-squared distance calibration
  that exactly hits population totals for continuous auxiliaries.

Exposed at sp.rake / sp.linear_calibration. 4 tests pass.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
