+++ src/statspai/iv/plausibly_exogenous.py
@@ -1,0 +1,288 @@
+"""
+Plausibly exogenous IV — Conley, Hansen and Rossi (2012).
+
+Sensitivity of 2SLS inference when the exclusion restriction is relaxed
+to allow the instrument to have a *small* direct effect on the outcome.
+
+Four approaches from CHR (2012):
+
+1. **UCI (Union of Confidence Intervals)** — for each candidate direct
+   effect γ in a user-supplied support, compute the 2SLS CI assuming
+   ``y = X*beta + Z*gamma + u`` with γ known, then take the union.
+2. **LTZ (Local-to-Zero)** — Bayesian-flavoured Gaussian prior on γ with
+   mean zero and variance ``Omega``; delivers a closed-form variance
+   inflation for beta.
+
+Both are exceptionally useful as headline sensitivity checks and are
+absent from Python's IV stack.
+
+References
+----------
+Conley, T.G., Hansen, C.B. and Rossi, P.E. (2012).
+    "Plausibly exogenous." *Review of Economics and Statistics*,
+    94(1), 260-272.
+
+van Kippersluis, H. and Rietveld, C.A. (2018).
+    "Beyond plausibly exogenous." *Econometrics Journal*, 21(3),
+    316-331.  [pragmatic guidance]
+"""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Iterable, List, Optional, Tuple, Union
+
+import numpy as np
+import pandas as pd
+from scipy import stats
+
+
+@dataclass
+class PlausiblyExogenousResult:
+    method: str
+    beta_hat: float
+    se_hat: float
+    gamma_grid: np.ndarray
+    beta_at_gamma: np.ndarray
+    se_at_gamma: np.ndarray
+    ci_lower: float
+    ci_upper: float
+    ci_level: float
+    extra: dict
+
+    def summary(self) -> str:
+        lines = [
+            f"Plausibly exogenous IV — {self.method}",
+            "-" * 48,
+            f"  Point estimate (γ=0) : {self.beta_hat:>10.4f}   SE={self.se_hat:.4f}",
+            f"  Sensitivity CI ({int(self.ci_level*100)}%): "
+            f"[{self.ci_lower:.4f}, {self.ci_upper:.4f}]",
+            f"  γ grid size          : {len(self.gamma_grid)}",
+        ]
+        if 'gamma_mean' in self.extra:
+            lines.append(f"  γ prior mean         : {self.extra['gamma_mean']}")
+            lines.append(f"  γ prior variance     : {self.extra['gamma_var']}")
+        return "\n".join(lines)
+
+
+def _as_matrix(x) -> np.ndarray:
+    a = np.asarray(x, dtype=float)
+    return a.reshape(-1, 1) if a.ndim == 1 else a
+
+
+def _prep(
+    y, endog, instruments, exog, data, add_const,
+) -> Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray]:
+    def grab(v, prefix):
+        if isinstance(v, str):
+            return data[v].values.astype(float)
+        if isinstance(v, list) and all(isinstance(x, str) for x in v):
+            return data[v].values.astype(float)
+        return np.asarray(v, dtype=float)
+
+    Y = grab(y, "y").reshape(-1)
+    D = grab(endog, "d")
+    if D.ndim == 1:
+        D = D.reshape(-1, 1)
+    Z = grab(instruments, "z")
[…]
