+++ src/statspai/iv/weak_iv_ci.py
@@ -1,0 +1,403 @@
+"""
+Weak-IV-robust confidence *sets* for a single endogenous regressor.
+
+These routines invert weak-IV-robust tests across a grid of candidate
+values of β₀ to yield confidence sets that remain valid regardless of
+instrument strength. Three variants:
+
+- :func:`anderson_rubin_ci` — AR (1949) test by grid inversion.
+- :func:`conditional_lr_ci` — Moreira (2003) CLR test by grid inversion
+  (uniformly most powerful invariant in the single-endogenous case).
+- :func:`k_test_ci`         — Kleibergen (2002, 2005) K/K-J test by grid
+  inversion; faster than CLR and also weak-IV-robust.
+
+When the first-stage F is large, all three sets collapse to the usual
+normal Wald CI; when identification is weak, they can be much wider
+(or even unbounded / disconnected), which is the *correct* behaviour.
+
+References
+----------
+Anderson, T.W. and Rubin, H. (1949).
+    "Estimation of the Parameters of a Single Equation in a Complete
+    System of Stochastic Equations." *AMS*, 20(1), 46-63.
+
+Moreira, M.J. (2003). "A conditional likelihood ratio test for structural
+    models." *Econometrica*, 71(4), 1027-1048.
+
+Kleibergen, F. (2002). "Pivotal statistics for testing structural
+    parameters in instrumental variables regression."
+    *Econometrica*, 70(5), 1781-1803.
+
+Kleibergen, F. (2005). "Testing parameters in GMM without assuming
+    that they are identified." *Econometrica*, 73(4), 1103-1123.
+
+Andrews, I., Stock, J.H. and Sun, L. (2019). "Weak Instruments in IV
+    Regression: Theory and Practice." *Annual Review of Economics*, 11.
+"""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import List, Optional, Tuple, Union
+
+import numpy as np
+import pandas as pd
+from scipy import stats
+
+
+@dataclass
+class WeakIVConfidenceSet:
+    method: str
+    level: float
+    beta_grid: np.ndarray
+    statistic: np.ndarray
+    critical_value: np.ndarray
+    in_set: np.ndarray
+    lower: float
+    upper: float
+    is_empty: bool
+    is_connected: bool
+    is_unbounded: bool
+    extra: dict
+
+    def as_intervals(self) -> List[Tuple[float, float]]:
+        """Return the CI as a list of (lo, hi) intervals (handles disconnection)."""
+        if self.is_empty:
+            return []
+        intervals = []
+        in_set = self.in_set
+        grid = self.beta_grid
+        i = 0
+        n = len(grid)
+        while i < n:
+            if in_set[i]:
+                j = i
+                while j + 1 < n and in_set[j + 1]:
+                    j += 1
+                intervals.append((float(grid[i]), float(grid[j])))
+                i = j + 1
+            else:
+                i += 1
+        return intervals
+
+    def summary(self) -> str:
+        intervals = self.as_intervals()
+        lines = [
+            f"{self.method} — weak-IV-robust confidence set",
+            "-" * 60,
+            f"  level                : {int(self.level * 100)}%",
[…]
