+++ src/statspai/iv/mte.py
@@ -1,0 +1,466 @@
+"""
+Marginal Treatment Effects (MTE) via polynomial MTR specification.
+
+Implements the Brinch-Mogstad-Wiswall (2017, JoPE) closed-form polynomial
+approach: parameterise the marginal treatment response functions
+
+    m_0(u, X) = X' * phi_0(u),
+    m_1(u, X) = X' * phi_1(u),
+
+with polynomials ``phi_d(u) = sum_{k=0}^K theta_{d,k} u^k``, and invert
+the observed conditional means ``E[Y | P(Z)=p, X, D=1]`` and
+``E[Y | P(Z)=p, X, D=0]`` — each linear in the theta's — via GLS.
+
+From the fitted MTR functions the skill derives:
+
+- **MTE(u | X)** — u ∈ (0, 1)
+- **ATE(X)** — ∫ MTE(u|X) du
+- **ATT(X), ATU(X)**  — weighted integrals
+- **LATE(p, p')** — ∫_{p}^{p'} MTE(u) du / (p'-p)
+- **Policy-Relevant TE** for user-supplied policy shifts in P(Z)
+
+This fills a long-standing Python gap (R has ``ivmte``, Stata ``mtefe``,
+Python had nothing).
+
+References
+----------
+Brinch, C.N., Mogstad, M. and Wiswall, M. (2017). "Beyond LATE with a
+    Discrete Instrument." *Journal of Political Economy*, 125(4), 985-1039.
+
+Heckman, J.J. and Vytlacil, E.J. (2005). "Structural Equations,
+    Treatment Effects, and Econometric Policy Evaluation."
+    *Econometrica*, 73(3), 669-738.
+
+Mogstad, M., Santos, A. and Torgovitsky, A. (2018). "Using Instrumental
+    Variables for Inference About Policy Relevant Treatment Parameters."
+    *Econometrica*, 86(5), 1589-1619.
+"""
+
+from __future__ import annotations
+
+from dataclasses import dataclass, field
+from typing import Callable, Dict, List, Optional, Tuple, Union
+
+import numpy as np
+import pandas as pd
+from scipy import stats
+
+
+@dataclass
+class MTEResult:
+    """Marginal treatment effects result."""
+    poly_degree: int
+    theta0: np.ndarray  # (K+1, dim_X) coefficients for control arm MTR
+    theta1: np.ndarray  # (K+1, dim_X) coefficients for treated arm MTR
+    theta0_se: np.ndarray
+    theta1_se: np.ndarray
+    ate: float
+    ate_se: float
+    att: float
+    atu: float
+    late_2sls: float
+    propensity_range: Tuple[float, float]
+    mte_curve: pd.DataFrame  # u, mte, se columns
+    x_bar: np.ndarray
+    n_obs: int
+    treated_share: float
+    extra: Dict
+
+    def summary(self) -> str:
+        u_min, u_max = self.propensity_range
+        lines = [
+            "Marginal Treatment Effects (BMW 2017 polynomial MTR)",
+            "-" * 60,
+            f"  Polynomial degree    : {self.poly_degree}",
+            f"  Observations         : {self.n_obs}  (treated share = {self.treated_share:.3f})",
+            f"  Propensity support   : [{u_min:.3f}, {u_max:.3f}]",
+            "",
+            "  Aggregate parameters (at mean X)",
+            f"    ATE   = {self.ate:>8.4f}   SE={self.ate_se:.4f}",
+            f"    ATT   = {self.att:>8.4f}",
[…]

[…]
