### 2026-04-16 e69ea01
feat(timeseries): ARIMA/SARIMAX with auto order selection (AICc grid)
SP-04 Task 3. Wraps statsmodels SARIMAX with a StatsPAI-style result
object: .summary(), .forecast(horizon), .plot(), auto=(p,d,q) grid
search by AICc.

On a random-walk series: ARIMA(1,1,0) fits correctly; auto mode selects
ARIMA(0,1,2) with lower AICc. Forecast returns DataFrame with CI.

Exposed at sp.arima. 4 tests pass.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
