+++ src/statspai/__init__.py
@@ -153,6 +153,7 @@ from .timeseries import (
     structural_break, StructuralBreakResult, cusum_test,
     local_projections, LocalProjectionsResult,
     garch, GARCHResult,
+    arima, ARIMAResult,
 )
 # Experimental Design
 from .experimental import randomize, RandomizationResult, balance_check, BalanceResult, attrition_test, attrition_bounds, AttritionResult, optimal_design, OptimalDesignResult
