+++ src/statspai/timeseries/__init__.py
@@ -10,6 +10,7 @@ from .structural_break import structural_break, StructuralBreakResult, cusum_tes
 from .cointegration import engle_granger, johansen, CointegrationResult
 from .local_projections import local_projections, LocalProjectionsResult
 from .garch import garch, GARCHResult
+from .arima import arima, ARIMAResult
 
 __all__ = [
     "var", "VARResult", "granger_causality", "irf",
@@ -17,4 +18,5 @@ __all__ = [
     "engle_granger", "johansen", "CointegrationResult",
     "local_projections", "LocalProjectionsResult",
     "garch", "GARCHResult",
+    "arima", "ARIMAResult",
 ]
