### 2026-04-16 ecb0870
feat(timeseries): Bayesian VAR with Minnesota (Litterman) prior
SP-04 Task 4. Closed-form posterior (normal-inverse-Wishart, no MCMC)
with configurable tightness (λ₁) and cross-variable shrinkage (λ₂).

- bvar(data, lags, lambda1, lambda2): fit BVAR, returns posterior mean
  B, Sigma, fitted, residuals.
- .forecast(horizon): multi-step ahead conditional forecast.
- .irf(shock_var, horizon): orthogonalised IRFs via Cholesky.

Tight prior (λ₁→0) shrinks toward random-walk; loose prior (λ₁→∞)
recovers OLS-VAR. Verified on VAR(1) DGP: own-lag coefficient with
λ₁=10 recovers 0.49 (true 0.5); with λ
[…]
