+++ src/statspai/timeseries/bvar.py
@@ -1,0 +1,174 @@
+"""Bayesian VAR with Minnesota (Litterman) prior.
+
+The Minnesota prior shrinks towards a random-walk model: each variable's
+own first lag coefficient is centred at 1, all other coefficients at 0,
+with tightness controlled by hyperparameter λ₁ (overall tightness) and
+λ₂ (cross-variable shrinkage).
+
+The posterior is analytically tractable (normal-inverse-Wishart) and
+computed in closed form — no MCMC required for the posterior mean and
+credible intervals.
+
+References
+----------
+Litterman, R.B. (1986). "Forecasting with Bayesian Vector Autoregressions."
+  *JBE&S*, 4(1), 25–38.
+Doan, T., Litterman, R. & Sims, C. (1984). "Forecasting and Conditional
+  Projection Using Realistic Prior Distributions."
+  *ECREV*, 3(1), 1–100.
+Kilian, L. & Lütkepohl, H. (2017). *Structural Vector Autoregressive
+  Analysis*. Cambridge.
+"""
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Optional
+
+import numpy as np
+import pandas as pd
+
+
+@dataclass
+class BVARResult:
+    coef: np.ndarray                 # (K*p + 1, K)  posterior mean B
+    sigma: np.ndarray                # (K, K) posterior mean of Sigma
+    fitted: np.ndarray               # (T-p, K)
+    residuals: np.ndarray            # (T-p, K)
+    var_names: list
+    lags: int
+    n: int
+    lambda1: float
+    lambda2: float
+
+    def forecast(self, horizon: int = 8) -> pd.DataFrame:
+        K = self.sigma.shape[0]
+        p = self.lags
+        B = self.coef
+        last_obs = self.fitted[-1:] + self.residuals[-1:]  # last actual
+        # gather last p observations
+        history = np.zeros((p, K))
+        T_data = self.fitted.shape[0]
+        for lag in range(p):
+            idx = T_data - 1 - lag
+            if idx >= 0:
+                history[lag] = self.fitted[idx] + self.residuals[idx]
+        fc = np.empty((horizon, K))
+        for h in range(horizon):
+            x = np.concatenate([history.ravel(), [1.0]])
+            fc[h] = x @ B
+            history = np.roll(history, 1, axis=0)
+            history[0] = fc[h]
+        cols = self.var_names
+        return pd.DataFrame(fc, columns=cols,
+                           index=np.arange(1, horizon + 1))
+
+    def irf(self, shock_var: int = 0, horizon: int = 20) -> np.ndarray:
+        """Orthogonalised impulse responses (Cholesky decomposition)."""
+        K = self.sigma.shape[0]
+        p = self.lags
+        B = self.coef[:-1]           # exclude constant row
+        # Companion form
+        A_mats = [B[k * K:(k + 1) * K].T for k in range(p)]
+        chol = np.linalg.cholesky(self.sigma)
+        irfs = np.zeros((horizon, K))
+        phi = np.eye(K)
+        irfs[0] = chol[:, shock_var]
+        for h in range(1, horizon):
+            phi_new = np.zeros((K, K))
+            for j in range(min(h, p)):
+                phi_new += A_mats[j] @ (irfs[h - 1 - j][:, None] @ np.eye(1, K, 0)
+                                         if h - 1 - j == 0 else np.diag(irfs[h - 1 - j]))
+            # Simplified: multiply lag coefficients
+            contrib = np.zeros(K)
[…]
