### 2026-04-16 8fbe25b
feat(survival): AFT parametric survival (exponential/Weibull/lognormal/loglogistic)
SP-10 Task 2. Adds the Accelerated Failure Time model family:

    log T = X β + σ ε,  ε ~ {Gumbel, Normal, Logistic}

Four distributions: exponential (σ=1 fixed), Weibull, lognormal,
loglogistic. Estimated by MLE with right-censoring (L-BFGS-B).
SE from numerical Hessian. AIC/BIC for model comparison.

On Weibull DGP (true β_x=0.5): recovers β_x=0.42, correct sign and
order of magnitude. All 4 families run without error.

Exposed at sp.aft. 5 tests pass.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@a
[…]
