+++ src/statspai/survival/aft.py
@@ -1,0 +1,165 @@
+"""Accelerated Failure Time (AFT) parametric survival models.
+
+AFT models the **log survival time** directly as a linear function:
+
+    log T_i = X_i β + σ ε_i
+
+where the distribution of ε determines the family:
+
+- **exponential**: ε ~ standard extreme-value (Gumbel)
+- **weibull**: ε ~ standard extreme-value (Gumbel)
+- **lognormal**: ε ~ N(0,1)
+- **loglogistic**: ε ~ standard logistic
+
+Estimation is by maximum likelihood with right-censoring.
+
+References
+----------
+Kalbfleisch, J.D. & Prentice, R.L. (2002). *The Statistical Analysis
+  of Failure Time Data*, 2nd ed. Wiley.
+Klein, J.P. & Moeschberger, M.L. (2003). *Survival Analysis:
+  Techniques for Censored and Truncated Data*, 2nd ed. Springer.
+"""
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import List, Literal
+
+import numpy as np
+import pandas as pd
+from scipy import stats as sp_stats
+from scipy.optimize import minimize
+
+from .models import _parse_formula
+
+
+AFTFamily = Literal["exponential", "weibull", "lognormal", "loglogistic"]
+
+
+def _aft_log_likelihood(
+    beta: np.ndarray, sigma: float,
+    X: np.ndarray, logT: np.ndarray, E: np.ndarray,
+    family: AFTFamily,
+) -> float:
+    """AFT log-likelihood with right-censoring."""
+    eta = X @ beta
+    z = (logT - eta) / sigma
+    if family in ("exponential", "weibull"):
+        # Gumbel (minimum): f(z) = exp(z - exp(z)), S(z) = exp(-exp(z))
+        log_f = z - np.exp(z) - np.log(sigma)
+        log_S = -np.exp(z)
+    elif family == "lognormal":
+        log_f = sp_stats.norm.logpdf(z) - np.log(sigma)
+        log_S = sp_stats.norm.logsf(z)
+    elif family == "loglogistic":
+        log_f = sp_stats.logistic.logpdf(z) - np.log(sigma)
+        log_S = sp_stats.logistic.logsf(z)
+    else:
+        raise ValueError(f"unknown family {family!r}")
+    ll = float(np.sum(E * log_f + (1 - E) * log_S))
+    return ll
+
+
+@dataclass
+class AFTResult:
+    beta: np.ndarray
+    se: np.ndarray
+    sigma: float
+    var_names: List[str]
+    family: AFTFamily
+    log_likelihood: float
+    aic: float
+    bic: float
+    n: int
+    n_events: int
+
+    def summary(self) -> str:
+        lines = [
+            f"AFT Model ({self.family})",
+            "-" * 45,
+            f"n = {self.n}, events = {self.n_events}",
+            f"log σ = {np.log(self.sigma):.4f}  (σ = {self.sigma:.4f})",
+            f"Log-Lik = {self.log_likelihood:.4f}",
+            f"AIC = {self.aic:.2f}, BIC = {self.bic:.2f}",
+            "",
+            "Coefficients (on log-time scale):",
+        ]
+        for nm, b, s in zip(self.var_names, self.beta, self.se):
+            t = b / s if s > 0 else np.nan
+            p = 2 * (1 - sp_stats.norm.cdf(abs(t)))
+            lines.append(
+                f"  {nm:<15s}  {b: .4f}  (SE {s: .4f}, z {t: .3f}, p {p: .4f})"
+            )
+        return "\n".join(lines)
+
[…]
