+++ src/statspai/survival/__init__.py
@@ -1,8 +1,10 @@
 """Survival and duration analysis models."""
 from .models import cox, kaplan_meier, survreg, CoxResult, KMResult, logrank_test
 from .frailty import cox_frailty, FrailtyResult
+from .aft import aft, AFTResult
 
 __all__ = [
     "cox", "kaplan_meier", "survreg", "CoxResult", "KMResult", "logrank_test",
     "cox_frailty", "FrailtyResult",
+    "aft", "AFTResult",
 ]
