### 2026-04-16 c36362d
fix(spatial): 3 critical + 2 high-priority bugs from pre-release review
Code review findings addressed:

Critical:
1. panel/estimator.py: two-way FE demeaning formula was wrong — used
   already-demeaned array for the time mean. Fixed to compute both
   entity and time means from the original array.
2. ml.py sac(): log-likelihood was inconsistently reconstructed from
   neg_conc_ll. Fixed to use _full_loglik() consistently with other
   models.

High:
3. gmm.py: np.diag(e**2) created O(n²) dense matrix in het-robust SE
   computation. Replaced with element-wise multiply (O(n) memory).
4. diagn
[…]
