### 2026-04-16 2d8ae3e
docs: update README + CHANGELOG for v0.8.0 release
- README: add v0.8.0 headline, update API-at-a-Glance with all new
  modules (spatial 38 symbols, time series, discovery, matching, etc.),
  update comparison table (spatial no longer an R advantage), add
  spatial row to pain-point table
- CHANGELOG: full v0.8.0 release notes covering all 10 sub-projects
- Fix function count: 390+ → 450+

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
