### 2026-04-16 594461d
fix(iv): 3 critical + 6 high-priority bugs from code review
CRITICAL fixes:
- KP rk: Kronecker order inverted in robust meat (kron(v⊗v,z⊗z) → kron(z⊗z,v⊗v))
- KP rk LM: use SVD min singular value instead of tr(A²) for correct chi² df
- ivmte_lp: remove dead-code tautology in ATT weight path

HIGH fixes:
- IJIVE: remove incorrect D-shrinkage (AD09 differs only in SE, not point est)
- NPIV: multi-Z now uses linear projection index instead of ‖Z‖₂
- Bayesian IV: avoid n×n PZ matrix, use ZtZ_inv for O(nk) computation
- weak_iv_ci: avoid n×n M_Z in CLR/K-test loops via YD'YD - (Zs'YD)'(Zs'YD)
- KP 
[…]
