+++ src/statspai/iv/weak_identification.py
@@ -235,7 +235,11 @@ def kleibergen_paap_rk(
     Z_tilde = _residualize(Z, W)
 
     # Reduced form coefficients: D_tilde = Z_tilde @ Pi + V
-    ZZ_inv = np.linalg.inv(Z_tilde.T @ Z_tilde)
+    ZtZ = Z_tilde.T @ Z_tilde
+    try:
+        ZZ_inv = np.linalg.inv(ZtZ)
+    except np.linalg.LinAlgError:
+        ZZ_inv = np.linalg.pinv(ZtZ)
     Pi = ZZ_inv @ (Z_tilde.T @ D_tilde)  # k x p
     V = D_tilde - Z_tilde @ Pi
 
@@ -246,13 +250,16 @@ def kleibergen_paap_rk(
         cov_vec = np.kron(Sigma, ZZ_inv)
         cov_label = "nonrobust"
     elif cov_type == "robust":
-        # Meat: sum_i (Z_i Z_i') ⊗ (V_i V_i')
+        # Meat: Σ_i  kron(z_i z_i', v_i v_i')  — KP (2006) eq. 13
+        # Convention: vec(Pi) stacks columns of Pi (k×p), so Var is (kp × kp)
+        # with blocks (Z'Z)^{-1} ⊗ Σ_VV under homoskedasticity.
+        # Robust sandwich: bread = kron(ZZ_inv, I_p) on both sides.
         meat = np.zeros((k * p, k * p))
         for i in range(n):
-            zi = Z_tilde[i:i + 1].T  # k x 1
-            vi = V[i:i + 1].T  # p x 1
-            meat += np.kron(vi @ vi.T, zi @ zi.T)
-        bread = np.kron(np.eye(p), ZZ_inv)
+            zi = Z_tilde[i]   # (k,)
+            vi = V[i]         # (p,)
+            meat += np.kron(np.outer(zi, zi), np.outer(vi, vi))
+        bread = np.kron(ZZ_inv, np.eye(p))
         cov_vec = bread @ meat @ bread
         cov_label = "HC robust"
     elif cov_type == "cluster":
@@ -264,17 +271,14 @@ def kleibergen_paap_rk(
         meat = np.zeros((k * p, k * p))
         for cid in groups:
             idx = np.where((g == cid).values)[0]
-            Zc = Z_tilde[idx]
-            Vc = V[idx]
-            score = np.kron(Vc, Zc).sum(axis=0)
-            # score is p*k (row-major kron of per-obs kron sum); rebuild properly
+            # Cluster score: Σ_{t∈cluster} kron(z_t, v_t) → vectorised form
             score_mat = np.zeros((k, p))
             for t in idx:
                 score_mat += np.outer(Z_tilde[t], V[t])
-            v = score_mat.flatten(order='F')  # matches vec(Pi) stacking cols
+            v = score_mat.flatten(order='F')  # vec(score) with col-stacking
             meat += np.outer(v, v)
         meat *= G / max(G - 1, 1)
-        bread = np.kron(np.eye(p), ZZ_inv)
+        bread = np.kron(ZZ_inv, np.eye(p))
         cov_vec = bread @ meat @ bread
         cov_label = f"cluster ({G} groups)"
     else:
@@ -291,19 +295,26 @@ def kleibergen_paap_rk(
     rk_f = rk_wald / df_num
     rk_wald_pvalue = float(1 - stats.chi2.cdf(rk_wald, df=k - p + 1))
 
-    # KP rk LM statistic (Kleibergen 2002 / Kleibergen-Paap 2006 eq. 17)
[…]
