### 2026-04-16 #10387 @brycewang-stanford
**A note on the repository history flag (85% code in a 48-hour window)**

I'd like to provide context on this. StatsPAI has been under active local development since mid-2025 as part of my research at Stanford REAP. The April 3–5 window reflects a bulk push of locally developed and tested code to the public repository — not code written in 48 hours.

Each module went through statistical validation before being committed: cross-validation against established packages (PySAL, statsmodels, R packages) with documented tolerance thresholds (e.g., `rtol < 1e-7` for spatial regression coefficients against PySAL spreg). The CHANGELOG and test suite document these validation benchmarks.

I use Claude Code as a development assistant for code generation and testing, but all statistical methodology decisions, algorithm choices, and numerical validation are my own. The development workflow is: I specify the econometric method and acceptance criteria → iterate on implementation → verify against reference implementations before merging.
