### 2026-04-16 2104cf1
feat(rd): comprehensive RDD mega-upgrade — 6 new modules + major enhancements
New modules (6 files, ~6000 lines):
- bandwidth.py: Full bandwidth selection suite (8 methods: MSE/CER × common/separate/comb1/comb2)
  with fuzzy-specific and covariate-adjusted support (Calonico-Cattaneo-Farrell 2020)
- locrand.py: Local randomization inference framework (rdrandinf, rdwinselect, rdsensitivity,
  rdrbounds) with Fisher exact tests and Rosenbaum bounds (Cattaneo-Titiunik-VB 2016)
- hte.py: Heterogeneous treatment effects via fully interacted local linear model (rdhte,
  rdbwhte, rdhte_lincom) with Wa
[…]
