+++ src/statspai/rd/rdml.py
@@ -1,0 +1,990 @@
+"""
+Machine Learning + Regression Discontinuity estimation.
+
+Combines modern ML methods (random forests, gradient boosting, LASSO) with
+RD designs for flexible CATE estimation, heterogeneity detection, and
+precision improvement through automated covariate selection.
+
+Functions
+---------
+rd_forest   : Causal Forest for RD (adapted Athey-Wager)
+rd_boost    : Gradient Boosting for RD
+rd_lasso    : LASSO-assisted RD (post-double-selection)
+rd_cate_summary : Unified multi-method CATE comparison
+
+References
+----------
+Athey, S., Tibshirani, J. and Wager, S. (2019).
+"Generalized Random Forests." *Annals of Statistics*, 47(2), 1148-1178.
+
+Belloni, A., Chernozhukov, V. and Hansen, C. (2014).
+"Inference on Treatment Effects after Selection among High-Dimensional
+Controls." *Review of Economic Studies*, 81(2), 608-650.
+"""
+
+from __future__ import annotations
+
+from typing import Optional, List, Dict, Any, Tuple
+
+import numpy as np
+import pandas as pd
+from scipy import stats
+
+from ..core.results import CausalResult
+
+# ======================================================================
+# Citations
+# ======================================================================
+
+CausalResult._CITATIONS['rd_forest'] = (
+    "@article{athey2019generalized,\n"
+    "  title={Generalized random forests},\n"
+    "  author={Athey, Susan and Tibshirani, Julie and Wager, Stefan},\n"
+    "  journal={Annals of Statistics},\n"
+    "  volume={47},\n"
+    "  number={2},\n"
+    "  pages={1148--1178},\n"
+    "  year={2019}\n"
+    "}"
+)
+
+CausalResult._CITATIONS['rd_boost'] = (
+    "@article{athey2019generalized,\n"
+    "  title={Generalized random forests},\n"
+    "  author={Athey, Susan and Tibshirani, Julie and Wager, Stefan},\n"
+    "  journal={Annals of Statistics},\n"
+    "  volume={47},\n"
+    "  number={2},\n"
+    "  pages={1148--1178},\n"
+    "  year={2019}\n"
+    "}"
+)
+
+CausalResult._CITATIONS['rd_lasso'] = (
+    "@article{belloni2014inference,\n"
+    "  title={Inference on treatment effects after selection among "
+    "high-dimensional controls},\n"
+    "  author={Belloni, Alexandre and Chernozhukov, Victor "
+    "and Hansen, Christian},\n"
+    "  journal={Review of Economic Studies},\n"
+    "  volume={81},\n"
+    "  number={2},\n"
+    "  pages={608--650},\n"
+    "  year={2014},\n"
+    "  publisher={Oxford University Press}\n"
+    "}"
+)
+
+
+# ======================================================================
+# Internal helpers
+# ======================================================================
+
+def _ik_bandwidth_simple(y: np.ndarray, x: np.ndarray, c: float) -> float:
+    """
+    Simple IK-style bandwidth: Silverman pilot scaled by curvature.
+
+    Used as a default when the user does not supply *h*.
+    """
+    x_c = x - c
+    n = len(x_c)
+    sd = np.std(x_c)
+    if sd < 1e-12 or n < 10:
+        return float(np.ptp(x_c) * 0.5)
+
+    h_pilot = 1.06 * sd * n ** (-1 / 5)
+
+    # Estimate curvature difference via local quadratic on each side
+    left = x_c < 0
+    right = x_c >= 0
+
+    def _quad_curv(yy, xx, hp):
+        mask = np.abs(xx) <= hp
[…]
