+++ src/statspai/rd/bandwidth.py
@@ -1,0 +1,969 @@
+"""
+Comprehensive bandwidth selection for local polynomial RD estimation.
+
+Implements all eight bandwidth selection methods from Calonico, Cattaneo,
+and Farrell (2020) including MSE-optimal and CER-optimal variants:
+
+MSE-optimal (minimize mean squared error of the RD estimator):
+  - mserd    : common bandwidth
+  - msetwo   : separate left/right
+  - msecomb1 : min(mserd, mseleft, mseright)
+  - msecomb2 : median(mserd, mseleft, mseright)
+
+CER-optimal (minimize coverage error rate of confidence intervals):
+  - cerrd    : common bandwidth
+  - certwo   : separate left/right
+  - cercomb1 : min(cerrd, cerleft, cerright)
+  - cercomb2 : median(cerrd, cerleft, cerright)
+
+Supports sharp RD, fuzzy RD, covariate adjustment, and cluster-robust
+variance estimation.
+
+References
+----------
+Calonico, S., Cattaneo, M.D. and Farrell, M.H. (2020).
+"Optimal Bandwidth Choice for Robust Bias-Corrected Inference in
+Regression Discontinuity Designs." *Econometrics Journal*, 23(2), 192-210.
+
+Calonico, S., Cattaneo, M.D. and Titiunik, R. (2014).
+"Robust Nonparametric Confidence Intervals for Regression-Discontinuity
+Designs." *Econometrica*, 82(6), 2295-2326.
+
+Imbens, G. and Kalyanaraman, K. (2012).
+"Optimal Bandwidth Choice for the Regression Discontinuity Estimator."
+*Review of Economic Studies*, 79(3), 933-959.
+"""
+
+from typing import Optional, List, Tuple
+
+import numpy as np
+import pandas as pd
+from scipy import stats
+
+
+# ======================================================================
+# Kernel helpers
+# ======================================================================
+
+def _kernel_fn(u: np.ndarray, kernel: str) -> np.ndarray:
+    """Kernel K(u) supported on |u| <= 1."""
+    u = np.asarray(u, dtype=float)
+    if kernel == 'triangular':
+        return np.maximum(1 - np.abs(u), 0)
+    elif kernel == 'uniform':
+        return 0.5 * (np.abs(u) <= 1).astype(float)
+    elif kernel == 'epanechnikov':
+        return 0.75 * np.maximum(1 - u ** 2, 0)
+    raise ValueError(f"Unknown kernel: {kernel}")
+
+
+def _kernel_constants(kernel: str) -> dict:
+    """
+    Return kernel-specific constants for bandwidth selection.
+
+    C_K  : MSE-optimal bandwidth constant for local linear (p=1).
+    mu_2 : second moment of the kernel, int u^2 K(u) du.
+    nu_0 : int K(u)^2 du (roughness).
+    """
+    table = {
+        'triangular':   {'C_K': 3.4375, 'mu_2': 1 / 6, 'nu_0': 2 / 3},
+        'epanechnikov': {'C_K': 3.0,    'mu_2': 1 / 5, 'nu_0': 3 / 5},
+        'uniform':      {'C_K': 2.7,    'mu_2': 1 / 3, 'nu_0': 1 / 2},
+    }
+    return table[kernel]
+
+
+# ======================================================================
+# Internal estimation helpers
+# ======================================================================
+
+def _local_poly_fit(
+    y: np.ndarray,
+    x: np.ndarray,
+    h: float,
+    p: int,
[…]
