### 2026-04-16 b66f312
fix(rd): density test bug + DGP functions + validation tests
Bug fix:
- rddensity: fixed density estimation using correct ECDF scaling and
  proper sandwich SE with F(x)(1-F(x))/n variance. False rejection
  rate was 75% → now ~2-5% (properly calibrated).

New features:
- 5 new RD DGP functions: dgp_rd_kink, dgp_rd_multi, dgp_rd_hte,
  dgp_rd_2d, dgp_rdit — all with known true effects in df.attrs
- Exported at sp.dgp_rd_kink() etc.

Validation:
- 22 new validation tests covering numerical accuracy, CI coverage,
  bandwidth properties, density test calibration, integration between
  modules, an
[…]
