+++ src/statspai/synth/multi_outcome.py
@@ -1,0 +1,526 @@
+"""
+Multiple Outcomes Synthetic Control Method.
+
+When multiple outcome variables are available, find a single set of
+donor weights that simultaneously matches ALL outcomes in the
+pre-treatment period.  Under a low-rank factor structure, using K
+outcomes reduces bias by a factor of 1/sqrt(K) compared to a single
+outcome.
+
+Two approaches
+--------------
+- **concatenated**: Stack standardised pre-treatment outcomes vertically
+  and solve a single quadratic programme for shared weights.
+- **averaged**: Standardise each outcome, average across K outcomes,
+  then run standard SCM on the averaged series.
+
+References
+----------
+Sun, L. (2023).
+"Using Multiple Outcomes to Improve the Synthetic Control Method."
+*Review of Economics and Statistics*, forthcoming.
+"""
+
+from __future__ import annotations
+
+from typing import Any, Dict, List, Optional
+
+import numpy as np
+import pandas as pd
+from scipy import stats as sp_stats
+from scipy.optimize import minimize
+
+from ..core.results import CausalResult
+
+
+# ====================================================================== #
+#  Public API
+# ====================================================================== #
+
+def multi_outcome_synth(
+    data: pd.DataFrame,
+    outcomes: List[str],
+    unit: str,
+    time: str,
+    treated_unit: Any,
+    treatment_time: Any,
+    method: str = "concatenated",
+    standardize: bool = True,
+    penalization: float = 0.0,
+    placebo: bool = True,
+    alpha: float = 0.05,
+) -> CausalResult:
+    """
+    Multiple Outcomes Synthetic Control Method (Sun 2023).
+
+    Finds a *single* set of donor weights that simultaneously matches
+    the treated unit across all K outcomes in the pre-treatment period.
+
+    Parameters
+    ----------
+    data : pd.DataFrame
+        Long-format panel data containing all outcome columns.
+    outcomes : list of str
+        Column names for the K outcome variables.
+    unit : str
+        Unit identifier column.
+    time : str
+        Time period column.
+    treated_unit : any
+        Value identifying the treated unit.
+    treatment_time : any
+        First treatment period (inclusive).
+    method : {'concatenated', 'averaged'}, default 'concatenated'
+        Weight-estimation strategy.
+
+        * ``'concatenated'`` -- stack all K standardised outcome panels
+          vertically and solve one quadratic programme.
+        * ``'averaged'`` -- standardise each outcome, average across K,
+          then solve SCM on the mean series.
+    standardize : bool, default True
+        Standardise each outcome to zero mean / unit variance before
+        stacking or averaging (strongly recommended when outcome
+        scales differ).
+    penalization : float, default 0.0
+        Ridge-type penalty added to the diagonal of the donor
+        cross-product matrix (``penalization * I``).  Helps when donors
+        are collinear.
+    placebo : bool, default True
+        Run in-space placebo permutations for inference (each donor is
+        pretended to be treated in turn).
+    alpha : float, default 0.05
+        Significance level for confidence intervals and joint test.
+
+    Returns
+    -------
+    CausalResult
+        Unified result object with:
+
+        - ``estimate`` : average treatment effect across outcomes
+          (mean of per-outcome ATTs).
[…]
