+++ src/statspai/synth/scpi.py
@@ -1,0 +1,832 @@
+"""
+Prediction Intervals for Synthetic Control Methods.
+
+Constructs valid prediction intervals for SCM that account for two
+sources of uncertainty:
+
+1. **In-sample uncertainty** -- estimation error in the donor weights
+   (from finite pre-treatment periods).
+2. **Out-of-sample uncertainty** -- prediction error even with known
+   weights (noise in post-treatment outcomes).
+
+Standard SCM only provides point estimates.  This method provides
+prediction intervals with formal coverage guarantees.
+
+References
+----------
+Cattaneo, M.D., Feng, Y. and Titiunik, R. (2021).
+"Prediction Intervals for Synthetic Control Methods."
+*Journal of the American Statistical Association*, 116(536), 1865-1880.
+"""
+
+from __future__ import annotations
+
+from typing import Any, Dict, List, Optional, Tuple, Union
+
+import numpy as np
+import pandas as pd
+from scipy import optimize, stats
+
+from ..core.results import CausalResult
+
+
+# ====================================================================== #
+#  Public API: scdata, scest, scpi
+# ====================================================================== #
+
+def scdata(
+    data: pd.DataFrame,
+    outcome: str,
+    unit: str,
+    time: str,
+    treated_unit: Any,
+    treatment_time: Any,
+) -> Dict[str, Any]:
+    """
+    Prepare data matrices for synthetic control estimation.
+
+    Reshapes a long-format panel into the matrices needed by ``scest``
+    and ``scpi``.  Mirrors the R package's ``scdata()`` function.
+
+    Parameters
+    ----------
+    data : pd.DataFrame
+        Long-format panel data.
+    outcome : str
+        Outcome variable column name.
+    unit : str
+        Unit identifier column name.
+    time : str
+        Time period column name.
+    treated_unit : scalar
+        Identifier of the treated unit.
+    treatment_time : scalar
+        First treatment period.
+
+    Returns
+    -------
+    dict
+        Keys:
+
+        - ``Y_pre``  : treated unit pre-treatment outcomes (T0,)
+        - ``Y_post`` : treated unit post-treatment outcomes (T1,)
+        - ``Y_donors_pre``  : donor pre-treatment matrix (T0, J)
+        - ``Y_donors_post`` : donor post-treatment matrix (T1, J)
+        - ``donor_names``   : list of donor unit labels
+        - ``pre_times``     : array of pre-treatment time values
+        - ``post_times``    : array of post-treatment time values
+        - ``times``         : full array of time values
+        - ``treated_unit``  : echo of the treated unit label
+        - ``treatment_time``: echo of the first treatment period
+
+    Examples
+    --------
+    >>> prepared = sp.scdata(df, outcome='gdp', unit='state',
+    ...     time='year', treated_unit='California', treatment_time=1989)
+    >>> prepared['Y_pre'].shape
+    (19,)
+    """
+    pivot = data.pivot_table(index=time, columns=unit, values=outcome)
+    times = pivot.index.values
+    pre_mask = times < treatment_time
+    post_mask = times >= treatment_time
+
+    if pre_mask.sum() < 2:
+        raise ValueError("Need at least 2 pre-treatment periods.")
+    if post_mask.sum() < 1:
+        raise ValueError("Need at least 1 post-treatment period.")
+
+    if treated_unit not in pivot.columns:
[…]
