+++ src/statspai/synth/mc.py
@@ -1,0 +1,504 @@
+"""
+Matrix Completion for Synthetic Control (MC-SCM).
+
+Treats the treated unit's post-treatment outcomes as missing entries in
+a panel matrix Y (N x T) and imputes them via nuclear norm regularisation
+(low-rank matrix completion).  The counterfactual is the imputed value;
+the treatment effect is observed minus imputed.
+
+Algorithm -- Soft-Impute / SVT
+-------------------------------
+1. Build Y matrix (units x time).  Mask treated unit's post-treatment entries.
+2. Initialise M = row/column means.
+3. Iterate:  M = SVT_lambda( P_obs(Y) + P_miss(M) )
+   where SVT_lambda soft-thresholds singular values by lambda.
+4. Counterfactual for treated = M[treated, post_periods].
+5. Effects = Y_observed - M_imputed.
+
+Cross-validation for lambda: hold out random entries from observed cells,
+pick lambda minimising reconstruction error.
+
+References
+----------
+Athey, S., Bayati, M., Doudchenko, N., Imbens, G. and Khosravi, A. (2021).
+"Matrix Completion Methods for Causal Panel Data Models."
+*Journal of the American Statistical Association*, 116(536), 1716-1730.
+"""
+
+from __future__ import annotations
+
+from typing import Any, List, Optional
+
+import numpy as np
+import pandas as pd
+from scipy import stats
+
+from ..core.results import CausalResult
+
+
+# ====================================================================== #
+#  Public API
+# ====================================================================== #
+
+def mc_synth(
+    data: pd.DataFrame,
+    outcome: str,
+    unit: str,
+    time: str,
+    treated_unit: Any,
+    treatment_time: Any,
+    covariates: Optional[List[str]] = None,
+    lambda_reg: Optional[float] = None,
+    max_iter: int = 500,
+    tol: float = 1e-6,
+    cv_folds: int = 5,
+    alpha: float = 0.05,
+    placebo: bool = True,
+    seed: Optional[int] = None,
+) -> CausalResult:
+    """
+    Matrix Completion Synthetic Control Method.
+
+    Imputes the treated unit's post-treatment counterfactual by solving
+    a nuclear-norm-penalised matrix completion problem on the full
+    panel, following Athey et al. (2021).
+
+    Parameters
+    ----------
+    data : pd.DataFrame
+        Long-format panel data.
+    outcome : str
+        Outcome variable name.
+    unit : str
+        Unit identifier column.
+    time : str
+        Time period column.
+    treated_unit : any
+        Identifier of the treated unit.
+    treatment_time : any
+        First treatment period (inclusive).
+    covariates : list of str, optional
+        Time-varying covariates to partial out before matrix completion.
+    lambda_reg : float, optional
+        Nuclear norm penalty.  If ``None`` (default), selected
+        automatically via cross-validation on observed entries.
+    max_iter : int, default 500
+        Maximum Soft-Impute iterations.
+    tol : float, default 1e-6
+        Convergence tolerance (relative change in Frobenius norm).
+    cv_folds : int, default 5
+        Number of CV folds for automatic lambda selection.
+    alpha : float, default 0.05
+        Significance level for confidence intervals.
+    placebo : bool, default True
+        Run placebo (permutation) inference by treating each control
+        unit as if it were treated.
+    seed : int, optional
+        Random seed for reproducibility.
+
+    Returns
+    -------
+    CausalResult
+        With ``.estimate`` equal to the average post-treatment effect (ATT),
[…]

[…]
