### 2026-04-16 052594a
fix(rd): extrapolate _ols_fit singular matrix fallback
_ols_fit now falls back to lstsq + pinv when the QR decomposition
encounters a singular design matrix (e.g., collinear covariates).
Prevents crash in rd_extrapolate with degenerate covariate sets.

79 RD tests pass. 13 end-to-end functional checks verified.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
