### 2026-04-16 6dca2dd
feat(synth): 7 new SCM methods — Bayesian, BSTS/CausalImpact, Penalized (A&L 2021), FDID, Cluster, Sparse, Kernel
Expand synthetic control module from 13 to 20 methods, making StatsPAI
the most comprehensive SCM toolkit in any language (Python, R, or Stata).

New modules (5,009 LOC):
- bayesian.py: Full Bayesian SCM with MCMC posterior inference
- bsts.py: Bayesian Structural Time Series / CausalImpact (Brodersen et al. 2015)
- penscm.py: Penalized SCM with pairwise discrepancy (Abadie & L'Hour, JASA 2021)
- fdid.py: Forward DID with optimal donor selection (Li 2024)
- cluster.py: Cluster SCM 
[…]
