+++ src/statspai/synth/power.py
@@ -1,0 +1,487 @@
+"""
+Power Analysis and Sample Size Planning for Synthetic Control Designs.
+
+A **novel** feature with no existing package equivalent.  Helps researchers
+plan their SCM study by answering:
+
+    *"Given my donor pool and pre-treatment periods, what is the minimum
+    detectable effect size?"*
+
+Functions
+---------
+- ``synth_power``   — full power curve across a grid of effect sizes
+- ``synth_mde``     — quick MDE at a target power level
+- ``synth_power_plot`` — visualise the power curve with MDE annotation
+
+Algorithm
+---------
+1. Fit SCM on actual data to obtain baseline weights and the null
+   (placebo) distribution of RMSPE ratios.
+2. For each hypothetical effect size *delta*:
+   a. Inject *delta* into the treated unit's post-treatment outcomes.
+   b. Re-compute the treated RMSPE ratio under the augmented data.
+   c. Compare against the pre-computed null distribution.
+   d. Record whether H0 is rejected at level *alpha*.
+3. Repeat over *n_simulations* with optional noise perturbation.
+4. Power = fraction of simulations where H0 is rejected.
+5. MDE = smallest *delta* where power >= 0.80.
+
+References
+----------
+Abadie, A., Diamond, A. and Hainmueller, J. (2010).
+"Synthetic Control Methods for Comparative Case Studies."
+*JASA*, 105(490), 493-505.
+
+Firpo, S. and Possebom, V. (2018).
+"Synthetic Control Method: Inference, Sensitivity Analysis and
+Confidence Sets."
+*Journal of Causal Inference*, 6(2).
+"""
+
+from __future__ import annotations
+
+from typing import Any, List, Optional, Sequence, Union
+
+import numpy as np
+import pandas as pd
+from scipy import stats
+
+from .scm import synth, SyntheticControl
+from ..core.results import CausalResult
+
+
+# ======================================================================
+# Internal helpers
+# ======================================================================
+
+def _build_null_distribution(
+    model: SyntheticControl,
+    weights: np.ndarray,
+) -> np.ndarray:
+    """
+    Compute the placebo RMSPE-ratio distribution (null) from donor units.
+
+    Parameters
+    ----------
+    model : SyntheticControl
+        Fitted model with matrices already prepared.
+    weights : np.ndarray
+        Donor weights from the baseline fit.
+
+    Returns
+    -------
+    np.ndarray
+        Array of RMSPE ratios for each valid placebo unit.
+    """
+    all_units = np.column_stack([
+        model.Y_treated[:, np.newaxis], model.Y_donors,
+    ])
+
+    ratios: list[float] = []
+
+    for i in range(len(model.donor_units)):
+        idx_placebo = i + 1  # treated is at index 0
+        Y_placebo = all_units[:, idx_placebo]
+        donor_idx = [j for j in range(all_units.shape[1]) if j != idx_placebo]
+        Y_placebo_donors = all_units[:, donor_idx]
+
+        Y_pre_p = Y_placebo[model.pre_mask]
+        Y_pre_d = Y_placebo_donors[model.pre_mask]
+
+        try:
+            w = model._solve_weights(Y_pre_p, Y_pre_d)
+            synth_p = Y_placebo_donors @ w
+            gap_p = Y_placebo - synth_p
+
+            pre_mspe = float(np.mean(gap_p[model.pre_mask] ** 2))
+            post_mspe = float(np.mean(gap_p[model.post_mask] ** 2))
+
+            if pre_mspe > 1e-10:
+                ratios.append(np.sqrt(post_mspe) / np.sqrt(pre_mspe))
+        except Exception:
+            continue
+
+    return np.asarray(ratios, dtype=float)
+
+
+def _treated_ratio_with_effect(
[…]
