+++ src/statspai/synth/bayesian.py
@@ -1,0 +1,755 @@
+"""
+Bayesian Synthetic Control Method.
+
+Places a Dirichlet prior on donor weights and uses Metropolis-Hastings
+MCMC to sample from the joint posterior of (weights, sigma).  The
+posterior distribution of treatment effects yields credible intervals
+that incorporate both estimation uncertainty and model uncertainty —
+avoiding the need for asymptotic approximations or permutation-based
+inference.
+
+Model
+-----
+Prior:      w ~ Dirichlet(alpha)
+            sigma ~ InverseGamma(a0, b0)  [weakly informative]
+Likelihood: Y1_pre | w, sigma ~ N(Y0_pre.T @ w, sigma^2 I)
+Posterior:  p(w, sigma | Y1_pre) via MCMC
+Effect:     tau_t = Y1_post_t - Y0_post_t.T @ w   for each posterior draw
+
+References
+----------
+Vives, J. and Martinez, A. (2024). "Bayesian Synthetic Control Methods."
+*Journal of Computational and Graphical Statistics*.
+
+Brodersen, K. H., Gallusser, F., Koehler, J., Remy, N. and Scott, S. L.
+(2015). "Inferring causal impact using Bayesian structural time-series
+models." *The Annals of Applied Statistics*, 9(1), 247-274.
+"""
+
+from __future__ import annotations
+
+from typing import Dict, List, Optional, Tuple
+
+import numpy as np
+import pandas as pd
+from scipy import stats as sp_stats
+from scipy.special import gammaln
+
+from ..core.results import CausalResult
+
+
+# ====================================================================== #
+#  Public API
+# ====================================================================== #
+
+
+def bayesian_synth(
+    data: pd.DataFrame,
+    outcome: str,
+    unit: str,
+    time: str,
+    treated_unit,
+    treatment_time,
+    covariates: Optional[List[str]] = None,
+    n_iter: int = 2000,
+    n_warmup: int = 1000,
+    n_chains: int = 2,
+    dirichlet_alpha: float = 1.0,
+    seed: Optional[int] = None,
+    alpha: float = 0.05,
+) -> CausalResult:
+    """
+    Bayesian Synthetic Control Method.
+
+    Estimates the ATT by placing a Dirichlet prior on donor weights and
+    sampling from the posterior via Metropolis-Hastings MCMC.  Returns
+    full posterior credible intervals for the treatment effect.
+
+    Parameters
+    ----------
+    data : pd.DataFrame
+        Panel data in long format with columns for unit, time, and outcome.
+    outcome : str
+        Name of the outcome variable column.
+    unit : str
+        Name of the unit identifier column.
+    time : str
+        Name of the time period column.
+    treated_unit : scalar
+        Value in *unit* that identifies the treated unit.
+    treatment_time : scalar
+        First period of treatment (inclusive).
+    covariates : list of str, optional
+        Additional pre-treatment predictors to include in the matching
+        objective.  Covariates are appended to the pre-treatment outcome
+        series for each unit before fitting.
+    n_iter : int, default 2000
+        Total MCMC iterations per chain (including warmup).
+    n_warmup : int, default 1000
+        Number of warmup (burn-in) iterations for adaptation.  Must be
+        strictly less than *n_iter*.
+    n_chains : int, default 2
+        Number of independent MCMC chains.  Multiple chains enable the
+        R-hat convergence diagnostic.
+    dirichlet_alpha : float, default 1.0
+        Concentration parameter for the symmetric Dirichlet prior on
[…]
