+++ src/statspai/__init__.py
@@ -60,6 +60,10 @@ from .synth import (
     discos, qqsynth, discos_test, discos_plot, stochastic_dominance,
     synth_loo, synth_time_placebo, synth_donor_sensitivity,
     synth_rmspe_filter, synth_sensitivity, synth_sensitivity_plot,
+    synth_power, synth_mde, synth_power_plot,
+    synth_compare, synth_recommend, SynthComparison,
+    synth_report, synth_report_to_file,
+    german_reunification, basque_terrorism, california_tobacco,
     synthdid_estimate, sc_estimate, did_estimate,
     synthdid_placebo, synthdid_plot, synthdid_units_plot, synthdid_rmse_plot,
     california_prop99,

[…]
