### 2026-04-16 88923b5
fix(synth): Bayesian MCMC Jacobian + fdid deprecation + cross-method benchmark + tutorial
- bayesian: add missing log-Jacobian log(sigma_prop/sigma) for log-normal
  random walk on sigma — corrects MH acceptance ratio (posterior was
  slightly biased toward lower sigma without it)
- fdid: fix NumPy 1.25 DeprecationWarning (ndim>0 scalar conversion) by
  extracting scalar before power operation
- tests: add TestCrossMethodConsistency — 11 methods × recovery-of-known-ATT
  + 4-method agreement benchmark (12 tests, all passing)
- docs: comprehensive synth.md tutorial covering all 20 methods, 6
  
[…]
