+++ tests/test_synth_extras.py
@@ -1,0 +1,281 @@
+"""
+Tests for synth extras: compare, report, datasets, power.
+"""
+
+import pytest
+import numpy as np
+import pandas as pd
+from statspai.core.results import CausalResult
+
+
+@pytest.fixture
+def panel_data():
+    """Standard test panel: 1 treated + 10 donors, 20 periods, effect=5."""
+    rng = np.random.default_rng(42)
+    n_units = 11
+    n_periods = 20
+    treatment_time = 11
+    records = []
+    alphas = rng.normal(10, 2, n_units)
+    betas = rng.normal(0.5, 0.1, n_units)
+    for i in range(n_units):
+        for t in range(1, n_periods + 1):
+            y = alphas[i] + betas[i] * t + rng.normal(0, 0.3)
+            if i == 0 and t >= treatment_time:
+                y += 5.0
+            records.append({'unit': f'unit_{i}', 'time': t, 'outcome': y})
+    return pd.DataFrame(records)
+
+
+# ====================================================================== #
+#  synth_compare
+# ====================================================================== #
+
+class TestSynthCompare:
+
+    def test_compare_runs(self, panel_data):
+        from statspai.synth.compare import synth_compare, SynthComparison
+        comp = synth_compare(
+            panel_data, outcome='outcome', unit='unit', time='time',
+            treated_unit='unit_0', treatment_time=11,
+            methods=['classic', 'penalized', 'demeaned'],
+            placebo=False,
+        )
+        assert isinstance(comp, SynthComparison)
+        assert hasattr(comp, 'comparison_table')
+        assert hasattr(comp, 'recommended')
+
+    def test_compare_table_structure(self, panel_data):
+        from statspai.synth.compare import synth_compare
+        comp = synth_compare(
+            panel_data, outcome='outcome', unit='unit', time='time',
+            treated_unit='unit_0', treatment_time=11,
+            methods=['classic', 'demeaned', 'mc'],
+            placebo=False,
+        )
+        df = comp.comparison_table
+        assert 'method' in df.columns
+        assert 'att' in df.columns
+        assert 'pre_rmspe' in df.columns
+        assert len(df) >= 2
+
+    def test_recommend(self, panel_data):
+        from statspai.synth.compare import synth_recommend
+        rec = synth_recommend(
+            panel_data, outcome='outcome', unit='unit', time='time',
+            treated_unit='unit_0', treatment_time=11,
+        )
+        assert isinstance(rec, str)
+        assert len(rec) > 0
+
+    def test_all_methods_compared(self, panel_data):
+        from statspai.synth.compare import synth_compare
+        comp = synth_compare(
+            panel_data, outcome='outcome', unit='unit', time='time',
+            treated_unit='unit_0', treatment_time=11,
+            placebo=False,
+        )
+        assert len(comp.comparison_table) >= 8
+
+
+# ====================================================================== #
+#  synth_report
+# ====================================================================== #
+
+class TestSynthReport:
+
+    def test_text_report(self, panel_data):
+        from statspai.synth.report import synth_report
+        report = synth_report(
+            panel_data, outcome='outcome', unit='unit', time='time',
+            treated_unit='unit_0', treatment_time=11,
+            sensitivity=False,
+        )
+        assert isinstance(report, str)
+        assert 'ATT' in report or 'Estimate' in report
+        assert 'unit_0' in report
+
+    def test_markdown_report(self, panel_data):
+        from statspai.synth.report import synth_report
+        report = synth_report(
+            panel_data, outcome='outcome', unit='unit', time='time',
+            treated_unit='unit_0', treatment_time=11,
+            output='markdown', sensitivity=False,
+        )
+        assert isinstance(report, str)
[…]
