+++ docs/guides/synth.md
@@ -1,0 +1,155 @@
+# Synthetic Control — 20 methods, one package
+
+StatsPAI bundles the most comprehensive synthetic control toolkit
+available in any language: 20 estimators, 6 inference strategies, plus
+a research workflow of automatic comparison, power analysis,
+sensitivity diagnostics, and one-click reports.
+
+## Quick start
+
+```python
+import statspai as sp
+
+# Load the California Proposition 99 dataset (Abadie et al. 2010 style)
+df = sp.synth.california_tobacco()
+
+# Classic SCM (Abadie, Diamond & Hainmueller 2010)
+res = sp.synth(
+    df, outcome='cigsale', unit='state', time='year',
+    treated_unit='California', treatment_time=1989,
+    method='classic',
+)
+print(res.summary())
+```
+
+`sp.synth()` is the unified dispatcher — switch `method=` to run any
+of the 20 variants with the same API and return type.
+
+## The 20 methods
+
+| `method=` | Paper | When to use |
+|---|---|---|
+| `classic` | Abadie, Diamond & Hainmueller (2010) | Baseline. One treated unit, few pre-periods, good donor fit. |
+| `penalized` / `ridge` | Doudchenko & Imbens (2016) | When classic weights are unstable; ridge improves extrapolation. |
+| `demeaned` | Ferman & Pinto (2021) | When unit fixed effects are large; demean before matching. |
+| `unconstrained` / `elastic_net` | Doudchenko & Imbens (2016) | When negative/unbounded weights are theoretically justified. |
+| `augmented` / `ascm` | Ben-Michael, Feller & Rothstein (2021) | When no exact pre-treatment fit exists; bias-corrects classic SCM. |
+| `sdid` | Arkhangelsky et al. (2021) | Combines DID & SCM; robust to effect heterogeneity. |
+| `factor` / `gsynth` | Xu (2017) | Multiple treated units + interactive fixed effects. |
+| `staggered` | Ben-Michael, Feller & Rothstein (2022) | Heterogeneous adoption times. |
+| `mc` | Athey et al. (2021) | Many missing cells; matrix completion via nuclear norm. |
+| `discos` | Gunsilius (2023) | When treatment affects the whole distribution, not just the mean. |
+| `multi_outcome` | Sun (2023) | Multiple correlated outcomes; borrows strength across. |
+| `scpi` | Cattaneo, Feng & Titiunik (2021) | Need honest prediction intervals, not placebo p-values. |
+| `bayesian` | Vives & Martinez (2024) | Want full posterior; Dirichlet prior + MCMC credible intervals. |
+| `bsts` / `causal_impact` | Brodersen et al. (2015) | Structural TS with level/trend; the Google CausalImpact model. |
+| `penscm` | Abadie & L'Hour (2021) | Many similar donors; pairwise discrepancy penalty ensures interpretability. |
+| `fdid` | Li (2024) | Want a *subset* of donors, not weights — simple to communicate. |
+| `cluster` | Rho (2024) | Large donor pool; cluster first, then SCM within best cluster. |
+| `sparse` / `lasso` | Amjad, Shah & Shen (2018) | High-dimensional donors; L1 picks a handful. |
+| `kernel` | RKHS / MMD matching | Nonlinear donor relationships. |
+| `kernel_ridge` | Kernel ridge regression | Nonlinear, unconstrained, regularised. |
+
+## Inference strategies
+
+| Strategy | Call | When |
+|---|---|---|
+| Placebo (in-space) | default | Always available; exchangeable donor pool. |
[…]
