### 2026-04-16 fe47ccf
feat(decomposition): tier-C mega-expansion — 18 methods + unified sp.decompose()
World-class decomposition analysis toolkit covering mean, distributional,
inequality, demographic, and causal decomposition. Python's first
comprehensive decomposition library, beating Stata ddecompose / cdeco
and R Counterfactual in scope.

New modules (13 files, ~4800 LOC)
---------------------------------
- dfl.py           — DiNardo-Fortin-Lemieux (1996) reweighting
- ffl.py           — Firpo-Fortin-Lemieux (2009, 2018) two-step detailed
- machado_mata.py  — Machado-Mata (2005) quantile regression decompositio
[…]
