+++ src/statspai/decomposition/dfl.py
@@ -1,0 +1,376 @@
+"""
+DiNardo-Fortin-Lemieux (1996) reweighting decomposition.
+
+Reweights one group's sample so that its covariate distribution matches
+the other group's, creating a counterfactual distribution. The gap
+between a statistic (quantile, mean, variance, Gini, CDF, density) in
+the observed vs counterfactual distribution identifies the *composition
+effect*; the remainder is the *structural* (wage-structure) effect.
+
+References
+----------
+DiNardo, J., Fortin, N., & Lemieux, T. (1996). "Labor Market Institutions
+and the Distribution of Wages, 1973-1992: A Semiparametric Approach."
+*Econometrica*, 64(5), 1001-1044.
+
+Fortin, N., Lemieux, T., & Firpo, S. (2011). "Decomposition Methods in
+Economics." In *Handbook of Labor Economics*, Vol. 4A, Ch. 1.
+"""
+from __future__ import annotations
+
+from dataclasses import dataclass, field
+from typing import Any, Callable, Dict, Optional, Sequence, Tuple, Union
+
+import numpy as np
+import pandas as pd
+
+from ._common import (
+    add_constant,
+    bootstrap_ci,
+    bootstrap_stat,
+    logit_fit,
+    logit_predict,
+    parse_formula,
+    prepare_frame,
+    sig_stars,
+    weighted_ecdf,
+    weighted_quantile,
+)
+
+
+# ════════════════════════════════════════════════════════════════════════
+# Statistic functions
+# ════════════════════════════════════════════════════════════════════════
+
+def _weighted_gini(y: np.ndarray, w: np.ndarray) -> float:
+    """Weighted Gini (Lerman-Yitzhaki 1989 formula)."""
+    order = np.argsort(y)
+    y_s = y[order]
+    w_s = w[order]
+    W = w_s.sum()
+    cum_w = np.cumsum(w_s)
+    F = (cum_w - 0.5 * w_s) / W  # fractional rank
+    mu = np.average(y_s, weights=w_s)
+    if mu <= 0:
+        return float("nan")
+    return float(2.0 * np.cov(y_s, F, aweights=w_s)[0, 1] / mu)
+
+
+def _statistic_value(
+    y: np.ndarray,
+    w: np.ndarray,
+    stat: str,
+    tau: float = 0.5,
+) -> float:
+    """Evaluate distributional statistic."""
+    if stat == "mean":
+        return float(np.average(y, weights=w))
+    if stat == "variance":
+        return float(np.cov(y, aweights=w))
+    if stat == "std":
+        return float(np.sqrt(np.cov(y, aweights=w)))
+    if stat == "quantile":
+        return float(weighted_quantile(y, tau, w=w))
+    if stat == "iqr":
+        return float(
+            weighted_quantile(y, 0.75, w=w) - weighted_quantile(y, 0.25, w=w)
+        )
+    if stat == "gini":
+        return _weighted_gini(y, w)
+    if stat == "log_var":
+        return float(np.cov(np.log(np.clip(y, 1e-12, None)), aweights=w))
+    raise ValueError(f"unknown statistic {stat!r}")
+
+
+# ════════════════════════════════════════════════════════════════════════
+# Core DFL reweighter
+# ════════════════════════════════════════════════════════════════════════
+
+@dataclass
+class DFLResult:
+    """Container for DFL reweighting decomposition results."""
+    gap: float
+    composition: float
+    structure: float
+    stat: str
+    tau: float
+    stat_a: float
+    stat_b: float
+    stat_cf: float            # counterfactual: A's X distribution, B's Y conditional
[…]
