+++ src/statspai/__init__.py
@@ -141,7 +141,17 @@ from .robustness import spec_curve, SpecCurveResult, robustness_report, Robustne
 from .survey import svydesign, SurveyDesign, svymean, svytotal, svyglm, rake, linear_calibration
 from .dag import dag, DAG, dag_example, dag_examples, dag_example_positions, dag_simulate
 from .power import power, PowerResult, power_rct, power_did, power_rd, power_iv, power_cluster_rct, power_ols, mde
-from .decomposition import oaxaca, gelbach, OaxacaResult, GelbachResult, rifreg, rif_decomposition
+from .decomposition import (
+    oaxaca, gelbach, OaxacaResult, GelbachResult, rifreg, rif_decomposition,
+    # Tier C additions
+    dfl_decompose, ffl_decompose, machado_mata, melly_decompose, cfm_decompose,
+    fairlie, bauer_sinning, yun_nonlinear,
+    inequality_index, subgroup_decompose, source_decompose, shapley_inequality,
+    kitagawa_decompose, das_gupta,
+    gap_closing, mediation_decompose, disparity_decompose,
+    decompose, available_methods,
+    cps_wage, chilean_households, mincer_wage_panel, disparity_panel,
+)
 from .selection import stepwise, lasso_select, SelectionResult
 from .qte import qdid, qte, QTEResult
 from .mht import romano_wolf, RomanoWolfResult, adjust_pvalues, bonferroni, holm, benjamini_hochberg
