+++ src/statspai/decomposition/machado_mata.py
@@ -1,0 +1,327 @@
+"""
+Machado-Mata (2005) quantile regression decomposition.
+
+Simulate counterfactual outcome distributions by combining one group's
+quantile regression coefficients with another group's covariate
+distribution. Decompose quantile gaps into composition (X distribution)
+and coefficient (price / structural) effects at each τ.
+
+References
+----------
+Machado, J.A.F. & Mata, J. (2005). "Counterfactual Decomposition of Changes
+in Wage Distributions Using Quantile Regression." *Journal of Applied
+Econometrics*, 20(4), 445-465.
+
+Albrecht, Björklund, Vroman (2003). "Is There a Glass Ceiling in Sweden?"
+*Journal of Labor Economics*, 21(1), 145-177.
+"""
+from __future__ import annotations
+
+from dataclasses import dataclass, field
+from typing import Any, Callable, Dict, Optional, Sequence, Tuple, Union
+import warnings
+
+import numpy as np
+import pandas as pd
+from scipy.optimize import linprog
+
+from ._common import (
+    add_constant,
+    bootstrap_ci,
+    bootstrap_stat,
+    parse_formula,
+    prepare_frame,
+    sig_stars,
+    weighted_quantile,
+)
+
+
+# ════════════════════════════════════════════════════════════════════════
+# Quantile regression via IRLS (Koenker)
+# ════════════════════════════════════════════════════════════════════════
+
+def _qreg_irls(
+    y: np.ndarray,
+    X: np.ndarray,
+    tau: float,
+    max_iter: int = 50,
+    tol: float = 1e-6,
+) -> np.ndarray:
+    """
+    Quantile regression via Iteratively Reweighted Least Squares.
+
+    Uses the weight w_i = 1/max(|resid_i|, eps) and adjusts residuals
+    for asymmetry (Koenker's algorithm, adequate but not optimal).
+    For higher accuracy one could wire to scipy.optimize or statsmodels.
+    """
+    n, k = X.shape
+    # initialise with OLS
+    beta = np.linalg.lstsq(X, y, rcond=None)[0]
+    eps = 1e-6
+    for _ in range(max_iter):
+        resid = y - X @ beta
+        # asymmetric check function weights
+        w = np.where(resid >= 0, tau, 1 - tau) / np.maximum(np.abs(resid), eps)
+        # Weighted LS
+        WX = X * w[:, None]
+        try:
+            beta_new = np.linalg.solve(X.T @ WX, X.T @ (w * y))
+        except np.linalg.LinAlgError:
+            beta_new = np.linalg.lstsq(X.T @ WX, X.T @ (w * y), rcond=None)[0]
+        if np.max(np.abs(beta_new - beta)) < tol:
+            beta = beta_new
+            break
+        beta = beta_new
+    return beta
+
+
+def _qreg_grid(
+    y: np.ndarray,
+    X: np.ndarray,
+    tau_grid: np.ndarray,
+) -> np.ndarray:
+    """Run quantile regression at each τ, return (n_tau, k) coefficient matrix."""
+    out = np.empty((len(tau_grid), X.shape[1]))
+    for i, t in enumerate(tau_grid):
+        out[i] = _qreg_irls(y, X, t)
+    return out
+
+
+# ════════════════════════════════════════════════════════════════════════
+# Result
+# ════════════════════════════════════════════════════════════════════════
+
+@dataclass
+class MachadoMataResult:
+    """Container for Machado-Mata decomposition."""
+    quantile_grid: pd.DataFrame   # τ, q_a, q_b, q_cf, gap, composition, structure
+    overall: Dict[str, float]     # aggregated across grid
[…]

[…]
