### 2026-04-16 56a8103
test(decomposition): comprehensive tier-C test suite (41 tests, all pass)
- 41 tests covering all 18 decomposition methods + unified dispatcher
- Fixes dispatcher method-arg clash (positional-only /)
- Smoke-tests: datasets, Oaxaca, Gelbach, DFL (mean/quantile/Gini),
  FFL (mean/quantile/variance), Machado-Mata, Melly, CFM,
  Fairlie (logit/probit), Bauer-Sinning, RIF, inequality
  (Theil T/L, Gini, Atkinson, CV²), Shapley, Gini source,
  Kitagawa, Das Gupta, gap-closing (regression/IPW/AIPW),
  mediation, Jackson-VanderWeele disparity
- Regression checks: result .summary() / .to_latex() / ._r
[…]
