+++ tests/test_decomposition_tier_c.py
@@ -1,0 +1,476 @@
+"""
+Comprehensive tests for the tier-C decomposition suite.
+
+Covers: OB, Gelbach, DFL, FFL, Machado-Mata, Melly, CFM, Fairlie,
+Bauer-Sinning, RIF, Shapley inequality, subgroup inequality,
+Lerman-Yitzhaki source, Kitagawa, Das Gupta, gap-closing,
+mediation, Jackson-VanderWeele disparity, and the unified
+``sp.decompose`` dispatcher.
+"""
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+from statspai.decomposition import _common
+
+
+SEED = 42
+
+
+@pytest.fixture(scope="module")
+def cps():
+    return sp.cps_wage(n=1200, seed=SEED)
+
+
+@pytest.fixture(scope="module")
+def disparity():
+    return sp.disparity_panel(n=1200, seed=SEED)
+
+
+@pytest.fixture(scope="module")
+def mincer():
+    return sp.mincer_wage_panel(n=1500, seed=SEED)
+
+
+# ════════════════════════════════════════════════════════════════════════
+# Datasets
+# ════════════════════════════════════════════════════════════════════════
+
+def test_datasets_shapes(cps, disparity, mincer):
+    assert cps.shape[0] == 1200
+    assert "female" in cps.columns
+    assert "log_wage" in cps.columns
+    assert disparity.shape[0] == 1200
+    assert "group" in disparity.columns
+    assert mincer.shape[0] == 1500
+
+
+def test_cps_wage_has_gap(cps):
+    gap = cps[cps.female == 0].log_wage.mean() - cps[cps.female == 1].log_wage.mean()
+    assert gap > 0.3, f"gender gap should be positive & large, got {gap}"
+
+
+# ════════════════════════════════════════════════════════════════════════
+# Existing: Oaxaca + Gelbach (regression guard)
+# ════════════════════════════════════════════════════════════════════════
+
+def test_oaxaca_basic(cps):
+    r = sp.oaxaca(data=cps, y="log_wage", group="female",
+                  x=["education", "experience", "tenure"])
+    assert r.overall["gap"] > 0
+    assert abs(r.overall["gap"] - (r.overall["explained"] + r.overall["unexplained"])) < 1e-6
+    assert "education" in r.detailed["variable"].values
+
+
+def test_oaxaca_pooled(cps):
+    r = sp.oaxaca(data=cps, y="log_wage", group="female",
+                  x=["education", "experience", "tenure"],
+                  reference="pooled")
+    assert r.reference == "pooled"
+
+
+def test_gelbach_basic(cps):
+    r = sp.gelbach(data=cps, y="log_wage",
+                   base_x=["education"],
+                   added_x=["experience", "tenure", "union"])
+    # Deltas should sum to total change (up to floating error)
+    assert abs(r.decomposition["delta"].sum() - r.total_change) < 1e-6
+
+
+# ════════════════════════════════════════════════════════════════════════
+# DFL
+# ════════════════════════════════════════════════════════════════════════
+
+def test_dfl_mean(cps):
+    r = sp.dfl_decompose(data=cps, y="log_wage", group="female",
+                         x=["education", "experience", "tenure", "union", "married"],
+                         stat="mean")
+    assert abs(r.composition + r.structure - r.gap) < 1e-6
+    # In this DGP, β-differences dominate
+    assert abs(r.structure) > abs(r.composition)
+
+
+def test_dfl_quantile(cps):
+    r = sp.dfl_decompose(data=cps, y="log_wage", group="female",
+                         x=["education", "experience", "tenure"],
+                         stat="quantile", tau=0.5)
+    assert abs(r.composition + r.structure - r.gap) < 1e-6
+
+
[…]
