### 2026-04-16 44f7529
refactor(rd): centralize kernel/lpoly/sandwich primitives into _core.py
Consolidate low-level RD primitives that were duplicated across 9 files
into a single private module (rd/_core.py) so numerical definitions
cannot drift. All 97 RD tests pass with zero regression.

New _core.py (191 lines) is the canonical home for:
  - _kernel_fn: triangular / epanechnikov / uniform / gaussian (4 previously
    duplicated definitions + rdit's lambda dict collapsed to one)
  - _kernel_constants / _kernel_mse_constant: MSE-optimal bandwidth
    constants (was 4 separate tables with identical values)
  - _lo
[…]
