### 2026-04-16 c2d1730
fix(synth): 9 v0.9.0 release blockers from comprehensive module review
Driven by a 5-parallel-agent code review of all 26 synth/ files for the
v0.9.0 release decision. Full synth suite: 144 tests pass.

Statistical correctness
- augsynth: ASCM bias correction now follows Ben-Michael, Feller &
  Rothstein (2021) Eq. 3 — per-period bias via ridge outcome model
  m̂: Y0_pre → Y0_post, applied as (Y1_pre − Y0_pre'γ) @ β(T0, T1).
  Replaces the previous scalar mean-of-residuals placeholder and the
  latent `_ridge_fit` solve(XtX, X.T@X) RHS bug. Same correction
  applied in the placebo loop for exc
[…]
