+++ src/statspai/synth/augsynth.py
@@ -45,46 +45,61 @@ def augsynth(
     treatment_time,
     covariates: Optional[List[str]] = None,
     ridge_lambda: Optional[float] = None,
+    placebo: bool = True,
     alpha: float = 0.05,
+    **kwargs,
 ) -> CausalResult:
     """
-    Augmented Synthetic Control Method.
+    Augmented Synthetic Control Method (Ben-Michael, Feller & Rothstein 2021).
 
-    Fits a standard synthetic control, then applies a ridge-regression
-    bias correction using the pre-treatment residuals.
+    Fits a standard SCM then adds a ridge-outcome-model bias correction.
+    Per-period correction is
+
+        bias(t) = m̂_t(X1_pre) − Σ_j γ_j m̂_t(X_j,pre),
+
+    where m̂_t is a ridge regression of donor post-period outcomes on
+    donor pre-period outcomes. Collapses to standard SCM when
+    ``ridge_lambda → ∞`` and to pure outcome-model imputation when
+    ``ridge_lambda → 0``.
 
     Parameters
     ----------
     data : pd.DataFrame
         Panel data in long format.
-    outcome : str
-        Outcome variable column.
-    unit : str
-        Unit identifier column.
-    time : str
-        Time period column.
-    treated_unit : scalar
-        Value in *unit* that identifies the treated unit.
-    treatment_time : scalar
-        First period of treatment.
+    outcome, unit, time : str
+        Column names.
+    treated_unit, treatment_time : scalar
+        Treated-unit identifier and first treatment period.
     covariates : list of str, optional
-        Additional predictors (used in the ridge outcome model).
+        Additional predictors (currently informational; main adjustment
+        comes from pre-treatment outcomes).
     ridge_lambda : float, optional
-        Ridge penalty. If None, selected via leave-one-out CV on
-        pre-treatment donor data.
+        Ridge penalty. When ``None``, selected by leave-one-donor-out CV.
+    placebo : bool, default True
+        Run in-space placebo permutation tests for SE / p-value.
     alpha : float, default 0.05
         Significance level.
+    **kwargs
+        Ignored — accepted for dispatcher compatibility.
 
     Returns
     -------
     CausalResult
-        With `.estimate` (average post-treatment effect), event-study-style
-        dynamic effects in `model_info['effects_by_period']`, and SCM
-        diagnostics in `model_info`.
+        ``detail`` has one row per post-treatment period with columns
+        ``time, treated, counterfactual, effect``. ``model_info`` includes
+        ``pre_rmspe, post_rmspe, weights, ridge_lambda, n_donors,
+        n_pre_periods, n_post_periods, placebo_distribution``.
+
+    References
+    ----------
+    Ben-Michael, E., Feller, A. and Rothstein, J. (2021). "The Augmented
+    Synthetic Control Method." *JASA*, 116(536), 1789-1803.
 
     Examples
     --------
-    >>> result = sp.augsynth(df, outcome='gdp', unit='state', time='year',
+    >>> import statspai as sp
+    >>> df = sp.synth.california_tobacco()
+    >>> result = sp.augsynth(df, outcome='cigsale', unit='state', time='year',
[…]
