### 2026-04-16 c15cbec
chore(release): v0.9.0 — synth module comprehensive expansion
Version bump 0.8.0 -> 0.9.0 coinciding with the synth module
reaching 20 estimators + 6 inference strategies + full research
workflow (compare / power / sensitivity / one-click reports).

Includes 9 release-blocker fixes from a 5-agent code review (see
CHANGELOG.md for details). Other modules (decomposition tier-C,
RD _core.py refactor) will be highlighted in follow-up release notes.
