### 2026-04-16 a4036a2
refactor(synth): centralize simplex weight solver into _core.py
Six SCM files had near-identical implementations of the simplex-constrained
(optionally ridge-penalized) weight optimization problem:

    min_w ||y - X w||^2 + penalization * ||w||^2
    s.t.  w_j >= 0, sum(w) = 1

All now delegate to synth._core.solve_simplex_weights. The canonical
implementation uses SLSQP with an analytic gradient (previously only
3 of the 6 callers supplied it), producing identical optima on this
convex QP with a ~3-5x speedup when the gradient was absent.

Migrated call sites (wrapper signatures preserved, z
[…]
