### 2026-04-16 dc93ada
fix(decomposition): address code-review findings — 3 critical + 6 high fixed
Critical bugs fixed
-------------------
1. Gini RIF formula was wrong (factor-of-2 error in E[RIF]):
   - Both ffl._rif_for_sample(stat='gini') and rif.rif_values(statistic='gini')
     re-derived from G = (2/μ)·E[y·F(y)] − 1:
     RIF(y_i) = 1 + (2/μ)[y_i·F(y_i) − GL(F(y_i))] − ((G+1)/μ)·y_i
   - Converges to G as n→∞ (O(1/n) discrete-ECDF correction only).
2. _gini_subgroup was unweighted and O(n²):
   - New _weighted_pairwise_mad gives O(n log n) via sorted-ECDF identity
     E|X−Y| = E_X[y·(2·F_Y(y)−1)] + E_Y[y·(1
[…]
