### 2026-04-16 f4426a8
refactor(decomposition): address 3 medium code-review findings
10. Factor _statistic_value into _common.statistic_value so that dfl.py,
    ffl.py (and future methods) import from one source. Removes the
    hidden coupling where ffl._statistic_value_generic silently
    extended dfl._statistic_value via fall-through. ffl.py keeps
    _statistic_value_generic as a backwards-compat alias.
11. Add test_qreg_irls_recovers_true_beta — verifies the custom QR
    IRLS solver used by machado_mata/melly recovers the true
    coefficients on a known location-scale DGP at τ ∈ {0.25, 0.5, 0.75}.
    Prev
[…]
