### 2026-04-16 d64a4a6
docs(readme): highlight v0.9.0 synth module expansion
