### 2026-04-16 0789223
refactor(decomposition): route oaxaca through existing _common primitives
oaxaca.py had three helpers that duplicated functions already available
in decomposition/_common.py:

  _ols(y, X)              -> _common.wls(y, X, w=None, robust=True)
  _add_constant(X)        -> _common.add_constant(X)
  _significance_stars(p)  -> _common.sig_stars(p)

Both _ols and _common.wls use QR decomposition with HC1 robust variance
and the n/(n-k) small-sample correction; with w=None and robust=True
the two are numerically identical. Verified: 48 decomposition tests
pass with zero regression.

Net: -36 lines 
[…]
