### 2026-04-16 5569fd0
refactor(decomposition): centralize influence functions and inequality atoms
Three complementary consolidations into decomposition/_common.py:

1. New _common.influence_function(y, stat, tau, w)
   — 9-stat weighted RIF kernel: quantile, mean, variance, std, log_var,
   iqr, gini, theil_t, theil_l, atkinson(eps=1). Canonical implementation
   shared by ffl._rif_for_sample (unweighted and weighted calls) and
   rif.rif_values. Previously these lived as near-duplicate 94-line and
   58-line blocks with subtly different coverage (ffl had 9 stats; rif
   had only 3).

2. rif.rif_values now delegat
[…]
