### 2026-04-17 c1f6fc6
ci: add skip-existing to PyPI publish to handle duplicate uploads
The v0.9.2 release workflow failed with '400 File already exists' because
the package was uploaded manually via twine before the GH Actions release
publish step ran. skip-existing makes the publish step idempotent.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
