+++ .github/workflows/ci-cd.yml
@@ -120,6 +120,8 @@ jobs:
 
     - name: Publish to PyPI
       uses: pypa/gh-action-pypi-publish@release/v1
+      with:
+        skip-existing: true
 
   security:
     runs-on: ubuntu-latest
