### 2026-04-18 bcfd0f3
feat(econ-trinity): P0 pillars — DML-PLIV, Mixed Logit, IV-QR
Extends three foundational econometrics tools that were identified as the
highest-ROI gaps vs. Stata, R, and existing Python packages:

1. DML-PLIV (src/statspai/dml/double_ml.py)
   Partially Linear IV per Chernozhukov et al. (2018, §4.2). Adds
   `model='pliv'` with a Neyman-orthogonal score and cross-fitted
   nuisance functions g(X)=E[Y|X], m(X)=E[D|X], r(X)=E[Z|X]. Returns
   LATE with influence-function-based standard errors.

2. Mixed Logit (src/statspai/regression/mixed_logit.py, 520 LOC)
   Random-coefficient MNL via simula
[…]
