+++ tests/test_econ_trinity.py
@@ -1,0 +1,170 @@
+"""
+Tests for the ``feat/econ-trinity`` P0 trio:
+
+    1. DML-PLIV   (Partially Linear IV, Chernozhukov et al. 2018)
+    2. Mixed Logit (Simulated MLE, Train 2009)
+    3. IV-QR       (Inverse-QR, Chernozhukov-Hansen 2006/2008)
+
+Each test validates against a controlled DGP — estimates must land within
+reasonable bands of the true parameters (looser than vs. reference packages
+because these are Monte-Carlo style tests, not bit-matched benchmarks).
+"""
+
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+
+
+# =============================================================================
+# DML-PLIV
+# =============================================================================
+
+class TestDMLPLIV:
+
+    @pytest.fixture(scope='class')
+    def dgp_pliv(self):
+        rng = np.random.default_rng(42)
+        n = 2000
+        X = rng.normal(size=(n, 5))
+        Z = 0.8 * X[:, 0] + rng.normal(size=n)
+        u = rng.normal(size=n)
+        D = 0.5 * Z + 0.3 * X[:, 0] ** 2 + 0.4 * u + 0.3 * rng.normal(size=n)
+        Y = 1.5 * D + np.sin(X[:, 0]) + 0.5 * X[:, 1] ** 2 + 2.0 * u
+        df = pd.DataFrame({
+            'y': Y, 'd': D, 'z': Z,
+            **{f'x{i}': X[:, i] for i in range(5)},
+        })
+        return df, ['x0', 'x1', 'x2', 'x3', 'x4']
+
+    def test_pliv_recovers_true_theta(self, dgp_pliv):
+        df, X_cols = dgp_pliv
+        r = sp.dml(df, y='y', treat='d', covariates=X_cols,
+                   model='pliv', instrument='z', n_folds=5)
+        assert abs(r.estimate - 1.5) < 0.2
+        assert r.se > 0 and r.se < 0.2
+        assert r.estimand == 'LATE'
+        assert r.model_info['dml_model'] == 'PLIV'
+
+    def test_pliv_ci_covers_truth(self, dgp_pliv):
+        df, X_cols = dgp_pliv
+        r = sp.dml(df, y='y', treat='d', covariates=X_cols,
+                   model='pliv', instrument='z', n_folds=5)
+        lo, hi = r.ci
+        assert lo <= 1.5 <= hi
+
+    def test_pliv_requires_instrument(self, dgp_pliv):
+        df, X_cols = dgp_pliv
+        with pytest.raises(ValueError, match='pliv'):
+            sp.dml(df, y='y', treat='d', covariates=X_cols, model='pliv')
+
+    def test_instrument_rejected_for_plr(self, dgp_pliv):
+        df, X_cols = dgp_pliv
+        with pytest.raises(ValueError):
+            sp.dml(df, y='y', treat='d', covariates=X_cols,
+                   model='plr', instrument='z')
+
+
+# =============================================================================
+# Mixed Logit
+# =============================================================================
+
+class TestMixedLogit:
+
+    @pytest.fixture(scope='class')
+    def panel_choice_data(self):
+        rng = np.random.default_rng(0)
+        N, T, J = 500, 4, 3
+        beta_price = -1.0
+        mean_q, sd_q = 1.2, 0.6
+        beta_q_ind = rng.normal(mean_q, sd_q, size=N)
+        rows = []
+        for n in range(N):
+            for t in range(T):
+                prices = rng.uniform(0.5, 2.0, J)
+                quality = rng.uniform(0, 3, J)
+                u = (beta_price * prices
+                     + beta_q_ind[n] * quality
[…]
