+++ src/statspai/dml/double_ml.py
@@ -1,7 +1,7 @@
 """
 Double/Debiased Machine Learning (Chernozhukov et al. 2018).
 
-Implements the partially linear model:
+Implements the partially linear model family:
     Y = theta * D + g(X) + eps
     D = m(X) + v
 
@@ -9,9 +9,16 @@ where g(.) and m(.) are estimated via cross-fitting with any sklearn-
 compatible learner, and theta is the causal parameter of interest.
 
 Supports:
-- Partially linear regression (PLR)
-- Interactive regression model (IRM) for binary treatment
+- Partially linear regression (PLR)               — continuous/binary D, no IV
+- Interactive regression model (IRM)              — binary D, ATE via AIPW
+- Partially linear IV (PLIV)                      — endogenous D with instrument Z
 - Multiple cross-fitting splits for median aggregation
+
+References
+----------
+Chernozhukov, V., Chetverikov, D., Demirer, M., Duflo, E., Hansen, C.,
+Newey, W., and Robins, J. (2018). "Double/Debiased Machine Learning for
+Treatment and Structural Parameters." *Econometrics Journal*, 21(1), C1-C68.
 """
 
 from typing import Optional, List, Dict, Any, Union
@@ -28,8 +35,10 @@ def dml(
     treat: str,
     covariates: List[str],
     model: str = 'plr',
+    instrument: Optional[Union[str, List[str]]] = None,
     ml_g: Optional[Any] = None,
     ml_m: Optional[Any] = None,
+    ml_r: Optional[Any] = None,
     n_folds: int = 5,
     n_rep: int = 1,
     alpha: float = 0.05,
@@ -44,20 +53,36 @@ def dml(
     y : str
         Outcome variable.
     treat : str
-        Treatment variable.
+        Treatment variable (continuous or binary).
     covariates : list of str
-        Control variables (high-dimensional confounders).
+        High-dimensional controls (nuisance covariates X).
     model : str, default 'plr'
-        DML model: 'plr' (partially linear) or 'irm' (interactive).
+        DML model:
+
+        - ``'plr'`` : partially linear (continuous or binary D)
+        - ``'irm'`` : interactive regression model (binary D; ATE via AIPW)
+        - ``'pliv'`` : partially linear IV (endogenous D with instrument(s) Z)
+
+    instrument : str or list of str, optional
+        Instrument variable(s) Z. **Required when** ``model='pliv'``.
+        If a list is passed, the first-stage reduced form is a
+        scalar projection ``z_bar = E[Z_1 | X]`` (use a single strong
+        instrument; for many-instrument settings project into a scalar
+        before passing).
     ml_g : sklearn estimator, optional
-        ML model for outcome nuisance (default: GradientBoostingRegressor).
+        ML model for outcome nuisance ``E[Y|X]``
+        (default: ``GradientBoostingRegressor``).
     ml_m : sklearn estimator, optional
-        ML model for treatment nuisance (default: same as ml_g, or
-        GradientBoostingClassifier for binary treatment in IRM).
+        ML model for treatment nuisance ``E[D|X]``. For IRM, a classifier
+        is used.
+    ml_r : sklearn estimator, optional
+        ML model for instrument reduced form ``E[Z|X]`` (PLIV only).
+        Defaults to ``ml_g`` type.
     n_folds : int, default 5
-        Number of cross-fitting folds.
+        K-fold cross-fitting.
     n_rep : int, default 1
-        Number of repeated cross-fitting splits (median aggregation).
+        Repeated splits (median aggregation — robustness against
+        split randomness, Chernozhukov et al. 2018 §3.5).
     alpha : float, default 0.05
[…]

[…]
