### 2026-04-18 2aa709b
fix(econ-trinity): address 4 BLOCKER + 7 HIGH from self-audit
Self-audit + code-reviewer agent surfaced critical bugs in the first
commit. This commit fixes all of them and adds adversarial tests.

BLOCKERS
--------
B1/B2  Mixed Logit chid indexing:
       np.unique returned lex-sorted codes; DataFrame was sorted by
       value order. For string chids like 'c0','c1','c10','c100',...
       the two orderings diverge and situation blocks silently mix.
       Replaced with pd.factorize(sort=False) after a mergesort over
       [panel_id, chid] so situation codes are monotonically
       non-decr
[…]
