+++ tests/test_econ_trinity.py
@@ -168,3 +168,149 @@ class TestIVQR:
             sp.ivqreg(dgp_ivqr, y='y',
                       endog=['d', 'x'], instruments='z',
                       tau=0.5, bootstrap=0)
+
+
+# =============================================================================
+# Adversarial / regression tests (bugs caught in code review)
+# =============================================================================
+
+class TestAdversarial:
+    """Tests that target specific bugs discovered during the first review."""
+
+    def test_mixlogit_string_chid_matches_numeric(self):
+        """
+        B1/B2 regression: Mixed Logit must produce identical results whether
+        chid values sort lexicographically or numerically. Prior to the fix,
+        using string chids like 'c0','c1','c10','c100',... would silently
+        misalign situation blocks and biased all parameters.
+        """
+        import numpy as np
+        rng = np.random.default_rng(0)
+        N, T, J = 150, 3, 3
+        beta_price, mean_q, sd_q = -1.0, 1.5, 0.6
+        beta_q_ind = rng.normal(mean_q, sd_q, size=N)
+        rows_num, rows_str = [], []
+        for n in range(N):
+            for t in range(T):
+                prices = rng.uniform(0.5, 2.0, J)
+                quality = rng.uniform(0, 3, J)
+                u = (beta_price * prices + beta_q_ind[n] * quality
+                     + rng.gumbel(0, 1, J))
+                chosen = int(np.argmax(u))
+                for j in range(J):
+                    base = {
+                        'price': prices[j], 'quality': quality[j],
+                        'y': 1.0 if j == chosen else 0.0,
+                    }
+                    rows_num.append({'pid': n, 'chid': n * T + t, **base})
+                    rows_str.append({'pid': f'p{n:03d}',
+                                     'chid': f'c{n * T + t}', **base})
+        df_num = pd.DataFrame(rows_num)
+        df_str = pd.DataFrame(rows_str)
+
+        r_num = sp.mixlogit(df_num, y='y', chid='chid',
+                            x_fixed=['price'], x_random=['quality'],
+                            panel_id='pid', n_draws=100, maxiter=30)
+        r_str = sp.mixlogit(df_str, y='y', chid='chid',
+                            x_fixed=['price'], x_random=['quality'],
+                            panel_id='pid', n_draws=100, maxiter=30)
+        # Bit-identical parameters (tolerance for float noise)
+        for name in ['price', 'mean_quality', 'sd_quality']:
+            assert abs(r_num.params[name] - r_str.params[name]) < 1e-8, (
+                f"{name}: numeric vs string chid differ by "
+                f"{abs(r_num.params[name] - r_str.params[name]):.2e}"
+            )
+
+    def test_mixlogit_missing_chosen_alternative_raises(self):
+        """B-extra: situations without exactly one y==1 must raise."""
+        df = pd.DataFrame({
+            'chid': [1, 1, 1, 2, 2, 2],
+            'y':    [0, 0, 0, 1, 0, 0],  # chid=1 has no chosen
+            'x':    [1.0, 2.0, 3.0, 1.5, 2.5, 3.5],
+        })
+        with pytest.raises(ValueError, match='chosen alternative'):
+            sp.mixlogit(df, y='y', chid='chid', x_random=['x'],
+                        n_draws=50, maxiter=10)
+
+    def test_mixlogit_correlated_rejects_non_normal(self):
[…]
