### 2026-04-18 fbbc5e8
docs(readme): update to v0.9.2 with 0.9.0/0.9.1/0.9.2 release notes
- Refresh headline banner from v0.9.0 synth to v0.9.2 decomposition
- Add 'Previously in' chain: v0.9.1 RD + v0.9.0 synth + v0.8.0 spatial
- Prepend three new Release Notes entries:
  * v0.9.2 — decomposition (18 methods, 13 modules, 54 tests)
  * v0.9.1 — RD mega-upgrade (18+ estimators, 14 modules)
  * v0.9.0 — synth mega-expansion (20 SCM + workflow)
- Bump citation version 0.6.0 -> 0.9.2, year 2025 -> 2026
- Remove stale local image-1.png / image.png (README uses raw GitHub URL)

Co-Authored-By: Claude Opus 4.7 (1M context
[…]
