+++ README.md
@@ -15,40 +15,33 @@ StatsPAI is the **agent-native** Python package for causal inference and applied
 
 It brings R's [Causal Inference Task View](https://cran.r-project.org/web/views/CausalInference.html) (fixest, did, rdrobust, gsynth, DoubleML, MatchIt, CausalImpact, ...) and Stata's core econometrics commands into a single, consistent Python API.
 
-**🎉 NEW in v0.9.0 — Synthetic Control: Most Comprehensive SCM Toolkit in Any Language**
+**🎉 NEW in v0.9.2 — Decomposition Analysis: Most Comprehensive Decomposition Toolkit in Python**
 
-**20 SCM estimators + 6 inference strategies + full research workflow**, all behind a unified `sp.synth(method=...)` dispatcher. No competing package in Python, R, or Stata offers this breadth.
+**18 first-class decomposition methods across 13 modules (~6,200 LOC, 54 tests)**, unified under one `sp.decompose(method=...)` dispatcher. Python's first (and most complete) implementation of the full decomposition stack — mean, distributional, inequality, demographic, and causal. Beats Stata (`oaxaca`/`ddecompose`/`cdeco`/`rifhdreg`/`mvdcmp`/`fairlie`) and R (`Counterfactual`/`ddecompose`/`oaxaca`/`dineq`) in scope with one API, one result class, three inference modes (analytical / bootstrap / none).
 
-| What | Coverage |
+| Family | Methods |
 |---|---|
-| **20 estimators** | `classic`, `penalized`, `demeaned`, `unconstrained`, `augmented` (ASCM), `sdid`, `gsynth`, `staggered`, `mc` (matrix completion), `discos` (distributional), `multi_outcome`, `scpi` (prediction intervals), `bayesian` (MCMC), `bsts` / `causal_impact`, `penscm` (Abadie-L'Hour), `fdid` (Forward DID), `cluster`, `sparse` (LASSO), `kernel`, `kernel_ridge` |
-| **6 inference** | placebo · conformal · bootstrap · jackknife · Bayesian posterior · BSTS posterior |
-| **Research workflow** | `synth_compare()` runs all 20 · `synth_recommend()` auto-selects · `synth_power()` + `synth_mde()` · `synth_sensitivity()` (LOO + time placebos + donor + RMSPE) · `synth_report(format='latex')` one-click publication |
-| **Canonical datasets** | `california_tobacco()`, `german_reunification()`, `basque_terrorism()` |
-| **Correctness** | ASCM re-implemented to Ben-Michael et al. (2021) Eq. 3; Bayesian MCMC Jacobian corrected; 9 release-blocker fixes from a 5-agent code review; 144 synth tests passing |
+| **Mean** | `oaxaca` (Blinder-Oaxaca threefold, 5 reference coefficients), `gelbach` (sequential OVB, Gelbach 2016), `fairlie` (nonlinear logit/probit), `bauer_sinning` / `yun_nonlinear` (detailed nonlinear) |
+| **Distributional** | `rifreg` / `rif_decomposition` (Firpo-Fortin-Lemieux 2009), `ffl_decompose` (FFL 2018 two-step), `dfl_decompose` (DiNardo-Fortin-Lemieux 1996 reweighting), `machado_mata` (MM 2005 simulation), `melly_decompose` (Melly 2005 analytical), `cfm_decompose` (Chernozhukov-Fernández-Val-Melly 2013) |
+| **Inequality** | `subgroup_decompose` (Theil T/L, GE(α), Dagum Gini, Atkinson, CV²), `shapley_inequality` (Shorrocks 2013), `source_decompose` (Lerman-Yitzhaki 1985) |
[…]
