### 2026-04-18 52d3fbf
ci: add TestPyPI publish job + workflow_dispatch trigger
- New publish-test job mirrors publish-prod, publishes to TestPyPI on
  release-published (parallel to PyPI) and supports manual re-runs.
- workflow_dispatch with publish_target choice (none|testpypi-only|
  pypi-only|both) lets us retrigger a stuck release without re-tagging.
- Both publish jobs use skip-existing:true, so replayed runs no longer
  400 on File-already-exists (the v0.9.2 release failure mode).

Unblocks: v0.9.2 TestPyPI gap (already backfilled via manual twine),
and prevents the same red-X deployment status for future rel
[…]
