### 2026-04-18 b212e01
feat(multilevel): overhaul to lme4/Stata-grade mixed-models toolkit
Rewrite `statspai.multilevel` from a 400-LOC two-level-LMM single
file into a proper sub-package (~2,500 LOC across `_core.py`,
`lmm.py`, `glmm.py`, `diagnostics.py`, `comparison.py`) with feature
parity against R `lme4` and Stata `mixed`, and several additions.

Linear mixed model (`sp.mixed`) — new capabilities:
  * Unstructured random-effect covariance (default), Cholesky-
    parameterised for unconstrained optimisation. `diagonal` and
    `identity` remain available.
  * Three-level nested models via `group=['school','cla
[…]
