+++ src/statspai/multilevel/mixed.py
@@ -1,400 +1,10 @@
-"""
-Linear mixed effects (hierarchical linear) models.
+"""Back-compat shim — the real implementation now lives in ``lmm.py``.
 
-y_ij = X_ij'beta + Z_ij'u_j + eps_ij
-where u_j ~ N(0, G),  eps_ij ~ N(0, sigma^2)
-
-Equivalent to Stata ``mixed`` and R ``lme4::lmer()``.
-
-Estimation via profiled (RE)ML with Newton-type optimisation over
-variance components, analytical GLS for fixed effects.
+Older code and external notebooks import from
+``statspai.multilevel.mixed``; re-export the symbols they need so that
+path keeps working.
 """
 
-from __future__ import annotations
+from .lmm import MixedResult, mixed  # noqa: F401
 
-from dataclasses import dataclass, field
-from typing import Dict, List, Optional
-
-import numpy as np
-import pandas as pd
-from scipy import stats
-from scipy.optimize import minimize
-
-from ..core.results import EconometricResults
-
-
-# ---------------------------------------------------------------------------
-# Result container
-# ---------------------------------------------------------------------------
-
-@dataclass
-class MixedResult:
-    """Container for mixed model estimation results."""
-
-    fixed_effects: pd.Series
-    random_effects: pd.DataFrame  # BLUPs per group
-    variance_components: Dict[str, float]
-    blups: Dict[str, np.ndarray]
-    n_obs: int
-    n_groups: int
-    icc: float
-    log_likelihood: float
-
-    # internal bookkeeping (set after __init__)
-    _se_fixed: pd.Series = field(default=None, repr=False)
-    _groups: np.ndarray = field(default=None, repr=False)
-    _method: str = field(default="reml", repr=False)
-    _converged: bool = field(default=True, repr=False)
-    _lr_test: Optional[Dict[str, float]] = field(default=None, repr=False)
-    _alpha: float = field(default=0.05, repr=False)
-
-    # ------------------------------------------------------------------
-    def ranef(self) -> pd.DataFrame:
-        """Return random effects (BLUPs) as a DataFrame."""
-        return self.random_effects.copy()
-
-    # ------------------------------------------------------------------
-    def summary(self) -> str:
-        """Formatted summary table (Stata-style)."""
-        lines: List[str] = []
-        w = 72
-        lines.append("=" * w)
-        lines.append("Mixed-effects model".center(w))
-        lines.append("=" * w)
-
-        lines.append(f"  Method:          {self._method.upper()}")
-        lines.append(f"  No. obs:         {self.n_obs}")
-        lines.append(f"  No. groups:      {self.n_groups}")
-        lines.append(f"  Log-likelihood:  {self.log_likelihood:.4f}")
-        lines.append(f"  ICC:             {self.icc:.4f}")
-        lines.append(f"  Converged:       {self._converged}")
-        lines.append("-" * w)
-
-        # Fixed effects table
-        lines.append("Fixed effects:")
-        alpha = self._alpha
-        t_crit = stats.norm.ppf(1 - alpha / 2)
-        hdr = f"{'':>16s} {'Coef':>10s} {'Std.Err':>10s} {'z':>8s} {'P>|z|':>8s}  [{100*(1-alpha):.0f}% CI]"
-        lines.append(hdr)
-        lines.append("-" * w)
-        for var in self.fixed_effects.index:
-            b = self.fixed_effects[var]
[…]
