+++ src/statspai/multilevel/glmm.py
@@ -1,0 +1,828 @@
+"""
+Generalized linear mixed models (GLMM) via Laplace approximation.
+
+Supported families and links (canonical only, to keep the Laplace
+approximation clean and the observed information analytic):
+
+    * ``gaussian`` — identity link.  Mostly for cross-checks against
+      ``mixed()``; for production Gaussian fits use the closed-form
+      linear mixed model.
+    * ``binomial`` — logit link (``sp.melogit``).  Accepts both
+      0/1 Bernoulli and count/trials responses via ``trials=``.
+    * ``poisson``  — log link (``sp.mepoisson``).  Optional exposure
+      offsets via ``offset=``.
+
+Model:
+
+    η_ij = x_ij' β + z_ij' u_j,    u_j ~ N(0, G),
+    y_ij | u_j ~ Family( g⁻¹(η_ij) ).
+
+The integrated log-likelihood
+
+    ℓ(β, θ) = Σ_j log ∫ f(y_j | β, u_j) φ(u_j; 0, G) du_j
+
+has no closed form for non-Gaussian families; we approximate each
+integral by a first-order Laplace expansion around the conditional
+mode û_j, giving
+
+    ℓ_j(β, θ) ≈ log f(y_j | β, û_j) − ½ û_j' G⁻¹ û_j
+               − ½ log |G| − ½ log |Z_j' W_j Z_j + G⁻¹|,
+
+where W_j is the diagonal GLM weight matrix at û_j.  The outer
+optimisation over (β, θ) is done with L-BFGS; the inner optimisation
+for û_j is a few Newton steps.
+
+Fixed-effect covariance uses the standard GLMM formula
+
+    Cov(β̂) = ( Σ_j X_j' W_j X_j
+              − Σ_j (X_j' W_j Z_j) H_j⁻¹ (Z_j' W_j X_j) )⁻¹,
+
+which is the observed information implied by the Laplace approximation.
+"""
+
+from __future__ import annotations
+
+from dataclasses import dataclass, field
+from typing import Any, Dict, List, Optional, Sequence, Tuple
+
+import numpy as np
+import pandas as pd
+from scipy import stats
+from scipy.optimize import minimize
+
+from ._core import (
+    _GroupBlock,
+    _as_str_list,
+    _group_blocks,
+    _initial_theta,
+    _n_cov_params,
+    _prepare_frame,
+    _unpack_G,
+)
+
+
+# ---------------------------------------------------------------------------
+# Exponential-family specs
+# ---------------------------------------------------------------------------
+
+class _Family:
+    """Protocol for canonical-link exponential families."""
+
+    name: str
+
+    @staticmethod
+    def inv_link(eta: np.ndarray) -> np.ndarray:          # g⁻¹(η)
+        raise NotImplementedError
+
+    @staticmethod
+    def variance(mu: np.ndarray) -> np.ndarray:           # V(μ) for canonical link
+        raise NotImplementedError
+
+    @staticmethod
+    def log_lik(y: np.ndarray, mu: np.ndarray, weight: np.ndarray) -> float:
+        raise NotImplementedError
+
+
+class _Gaussian(_Family):
+    name = "gaussian"
+
+    @staticmethod
+    def inv_link(eta):
+        return eta
+
+    @staticmethod
+    def variance(mu):
+        return np.ones_like(mu)
+
+    @staticmethod
+    def log_lik(y, mu, weight):
+        # weight carries the residual inverse-variance (set by the caller).
+        return float(-0.5 * np.sum(weight * (y - mu) ** 2))
+
+
+class _Binomial(_Family):
+    name = "binomial"
+
+    @staticmethod
+    def inv_link(eta):
[…]
