### 2026-04-18 273db22
fix(multilevel): address oracle + code-reviewer audit of b212e01
Spawned two independent review agents after the overnight LMM/GLMM
rewrite. They surfaced 2 BLOCKER + 4 HIGH + 3 MEDIUM + 3 LOW findings
(one HIGH claim — REML→ML sign flip — was a false positive verified
against statsmodels to 4 decimals).

Concrete fixes landing in this commit (file:line references follow
the reviewer reports):

[BLOCKER] predict(data=None) was returning fitted values in
group-iteration order rather than the original DataFrame row order,
silently mis-aligning `df["yhat"] = result.predict()`. `_GroupBlock`
now c
[…]
