+++ tests/test_multilevel.py
@@ -105,6 +105,32 @@ class TestRandomIntercept:
         # Residual variance ≈ σ²_ε = 0.25 (same sample)
         assert 0.2 < np.var(resid) < 0.32
 
+    def test_predict_is_row_aligned_with_training_frame(self):
+        """Regression test: predict(data=None) must not reorder rows."""
+        rng = np.random.default_rng(0)
+        n_g, n_per = 6, 5
+        group = np.repeat(np.arange(n_g), n_per)
+        # Scramble row order so groups are interleaved.
+        idx = rng.permutation(len(group))
+        group = group[idx]
+        x = rng.normal(0, 1, len(group))
+        u = np.array([rng.normal(0, 1, n_g)[g] for g in group])
+        y = 2.0 + 0.5 * x + u + rng.normal(0, 0.3, len(group))
+        df = pd.DataFrame({"y": y, "x": x, "g": group}).reset_index(drop=True)
+        r = sp.mixed(df, "y", ["x"], "g")
+        p_null = r.predict().values
+        p_df = r.predict(df).values
+        np.testing.assert_allclose(p_null, p_df)
+        # Independent alignment check: the correlation with y should be
+        # positive and not ~0 (which is what mis-alignment produces).
+        assert np.corrcoef(df["y"].values, p_null)[0, 1] > 0.4
+
+    def test_rejects_unhashable_group_values(self):
+        df = _random_intercept_panel()
+        df["g_bad"] = [list(range(3)) for _ in range(len(df))]  # unhashable
+        with pytest.raises(TypeError):
+            sp.mixed(df, "y", ["x"], "g_bad")
+
     def test_predict_marginal_vs_conditional(self):
         df = _random_intercept_panel()
         r = sp.mixed(df, "y", ["x"], "g")
@@ -210,6 +236,34 @@ class TestThreeLevelNested:
         with pytest.raises(NotImplementedError):
             sp.mixed(df, "y", ["x"], group=["school", "klass"], x_random=["x"])
 
+    def test_variance_components_keys(self):
+        df = _three_level_panel(n_schools=30)
+        r = sp.mixed(df, "y", ["x"], group=["school", "klass"])
+        # Three-level fit must expose all three variance components and
+        # both ICCs via the documented key names.
+        assert "var(_cons|school)" in r.variance_components
+        assert "var(_cons|klass)" in r.variance_components
+        assert "var(Residual)" in r.variance_components
+        assert "icc(school)" in r.variance_components
+        assert "icc(school+klass)" in r.variance_components
+
+    def test_singleton_inner_warns(self):
+        # Build a panel where some schools have only one class.
+        rng = np.random.default_rng(0)
+        records = []
+        cid = 0
+        for s in range(20):
+            n_classes = 1 if s < 5 else 4
+            for _ in range(n_classes):
+                for _ in range(12):
+                    x = rng.normal()
+                    y = 1 + 0.4 * x + rng.normal(0, 0.5)
+                    records.append({"school": s, "klass": cid, "x": x, "y": y})
+                cid += 1
+        df = pd.DataFrame(records)
+        with pytest.warns(RuntimeWarning):
+            sp.mixed(df, "y", ["x"], group=["school", "klass"])
+
 
 # ---------------------------------------------------------------------------
 # GLMM — logit and Poisson
@@ -249,6 +303,35 @@ class TestMELogit:
         assert (orr["upper"] > orr["OR"]).all()
 
 
+class TestMEGLMResultContract:
[…]
