+++ src/statspai/multilevel/lmm.py
@@ -197,16 +197,21 @@ class MixedResult:
         new cluster.
         """
         if data is None:
-            # Use the training blocks.
+            # Use the training blocks; scatter results back to the
+            # cleaned-frame row order so the returned Series lines up
+            # with ``df.dropna(subset=<used columns>)``.
             out = np.empty(self.n_obs)
-            idx_start = 0
             for block in self._blocks:
                 mu = block.X @ self.fixed_effects.values
                 if include_random:
-                    u = self.blups[block.key]
-                    mu = mu + block.Z @ u
-                out[idx_start : idx_start + block.n] = mu
-                idx_start += block.n
+                    u = self.blups.get(block.key)
+                    if u is not None:
+                        mu = mu + block.Z @ u
+                if block.row_idx is not None:
+                    out[block.row_idx] = mu
+                else:  # pragma: no cover — legacy path without row_idx
+                    idx_start = int(getattr(block, "_idx_start", 0))
+                    out[idx_start : idx_start + block.n] = mu
             return pd.Series(out, name="yhat")
 
         # Arbitrary new data ------------------------------------------------
@@ -644,7 +649,7 @@ def _three_level_nll(
     XtVinvy = np.zeros(p_fixed)
     logdet_sum = 0.0
 
-    for y_s, X_s, inner_ids, inner_unique in blocks_outer:
+    for y_s, X_s, inner_ids, inner_unique, _row_idx in blocks_outer:
         n_s = len(y_s)
         V = sigma2_e * np.eye(n_s) + sigma2_s * np.ones((n_s, n_s))
         # Add block-diagonal contribution from inner clusters.
@@ -669,7 +674,7 @@ def _three_level_nll(
         return 1e12
 
     quad = 0.0
-    for y_s, X_s, inner_ids, inner_unique in blocks_outer:
+    for y_s, X_s, inner_ids, inner_unique, _row_idx in blocks_outer:
         n_s = len(y_s)
         V = sigma2_e * np.eye(n_s) + sigma2_s * np.ones((n_s, n_s))
         for c in inner_unique:
@@ -706,6 +711,8 @@ def _fit_three_level_intercept(
     alpha: float,
 ) -> MixedResult:
     """Fit a school > class > student nested random-intercept LMM."""
+    import warnings as _warnings
+
     fixed_names = ["_cons"] + list(x_fixed)
     random_names = ["_cons"]
     p_fixed = len(fixed_names)
@@ -713,12 +720,28 @@ def _fit_three_level_intercept(
 
     # Build outer blocks.
     blocks_outer = []
+    singleton_outers = 0
+    positions = np.arange(len(df))
     for key, sub in df.groupby(outer_col, sort=False):
+        row_idx = positions[df.index.get_indexer(sub.index)]
         y_s = sub[y].to_numpy(dtype=float)
         X_s = sub[["__intercept__"] + list(x_fixed)].to_numpy(dtype=float)
         inner_ids = sub[inner_col].to_numpy()
         inner_unique = np.unique(inner_ids)
-        blocks_outer.append((y_s, X_s, inner_ids, inner_unique))
+        if len(inner_unique) < 2:
+            singleton_outers += 1
+        blocks_outer.append((y_s, X_s, inner_ids, inner_unique, row_idx))
+
+    if singleton_outers > 0:
+        _warnings.warn(
+            f"three-level fit: {singleton_outers}/{len(blocks_outer)} "
+            f"outer groups ({outer_col!r}) contain only one inner group "
[…]
