### 2026-04-18 3f64b9f
feat(frontier): overhaul to Stata/R-grade stochastic frontier toolkit
Rewrites sp.frontier and adds sp.xtfrontier to match or exceed Stata's
frontier/xtfrontier and R's frontier/sfaR. Fixes two latent bugs in the
previous 270-LOC single-file implementation.

Correctness fixes:
- Jondrow posterior mu*: the old code's -sign*eps*sigma_u^2/sigma^2 was
  sign-flipped vs Jondrow-Lovell-Materov-Schmidt (1982); corrected to
  +sign*eps*sigma_u^2/sigma^2. All prior efficiency scores were biased.
- Exponential log-density: the old code had the wrong sign on the eps
  coefficient and Phi argument (verifi
[…]
