+++ tests/test_frontier.py
@@ -1,0 +1,490 @@
+"""Comprehensive tests for the stochastic frontier module.
+
+Covers:
+* Cross-sectional half-normal, exponential, truncated-normal recovery.
+* Heteroskedastic ``sigma_u`` and ``sigma_v``.
+* BC95-style inefficiency determinants (``emean``).
+* Cost-frontier sign conventions.
+* Panel: Pitt-Lee TI (half-normal & truncated-normal), Battese-Coelli
+  TVD decay, and BC95 inefficiency-effects.
+* Efficiency scores (Battese-Coelli vs JLMS) internal consistency.
+* Specification tests (LR against OLS) and skewness diagnostic.
+* Bootstrap CIs, ranking, and descriptive summaries.
+"""
+
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+from scipy import stats as sst
+
+import warnings
+
+# Suppress noisy runtime warnings from fringe optimizer steps.
+warnings.filterwarnings("ignore", category=RuntimeWarning)
+
+from statspai.frontier import (
+    frontier,
+    xtfrontier,
+    FrontierResult,
+    te_summary,
+    te_rank,
+)
+from statspai.frontier import _core as _fc
+
+
+# ---------------------------------------------------------------------------
+# Simulated data helpers
+# ---------------------------------------------------------------------------
+
+
+def _simulate_hn_production(n, sigma_v, sigma_u, seed=0):
+    rng = np.random.default_rng(seed)
+    x1 = rng.normal(0, 1, n)
+    x2 = rng.normal(0, 1, n)
+    u = np.abs(rng.normal(0, sigma_u, n))
+    v = rng.normal(0, sigma_v, n)
+    y = 1.0 + 0.5 * x1 + 0.3 * x2 + v - u
+    return pd.DataFrame({"y": y, "x1": x1, "x2": x2})
+
+
+def _simulate_cost(n, sigma_v, sigma_u, seed=1):
+    rng = np.random.default_rng(seed)
+    x1 = rng.normal(0, 1, n)
+    u = np.abs(rng.normal(0, sigma_u, n))
+    v = rng.normal(0, sigma_v, n)
+    y = 1.0 + 0.5 * x1 + v + u  # cost: + u
+    return pd.DataFrame({"y": y, "x1": x1})
+
+
+def _simulate_panel_ti(N, T, sigma_v, sigma_u, seed=2):
+    rng = np.random.default_rng(seed)
+    id_ = np.repeat(np.arange(N), T)
+    t_ = np.tile(np.arange(T), N)
+    n = N * T
+    x1 = rng.normal(0, 1, n)
+    x2 = rng.normal(0, 1, n)
+    u_i = np.abs(rng.normal(0, sigma_u, N))
+    u_it = np.repeat(u_i, T)
+    v = rng.normal(0, sigma_v, n)
+    y = 1.0 + 0.6 * x1 + 0.4 * x2 + v - u_it
+    return pd.DataFrame({"y": y, "x1": x1, "x2": x2, "id": id_, "t": t_})
+
+
+def _simulate_panel_tvd(N, T, sigma_v, sigma_u, mu, eta, seed=3):
+    rng = np.random.default_rng(seed)
+    id_ = np.repeat(np.arange(N), T)
+    t_ = np.tile(np.arange(T), N)
+    n = N * T
+    x1 = rng.normal(0, 1, n)
+    u_i = sst.truncnorm.rvs(-mu / sigma_u, np.inf, loc=mu, scale=sigma_u,
+                            size=N, random_state=rng)
+    a_it = np.exp(-eta * (t_ - (T - 1)))
+    u_it = a_it * np.repeat(u_i, T)
[…]
