+++ src/statspai/frontier/_core.py
@@ -1,0 +1,507 @@
+"""
+Shared primitives for stochastic frontier analysis (SFA).
+
+Provides log-density kernels, Jondrow-style posterior moments of u|eps,
+Battese-Coelli exp(-u) technical efficiency, numerical Hessian / design-matrix
+helpers, and sign-convention utilities used by both cross-sectional
+(``sfa.py``) and panel (``panel.py``) estimators.
+
+Sign convention
+---------------
+Composed error: ``y = x'beta + v + sign * u``, with ``u >= 0``,
+``sign = -1`` for production (output frontier), ``sign = +1`` for cost frontier.
+
+Distributions supported
+-----------------------
+* Normal / half-normal (Aigner-Lovell-Schmidt 1977)
+* Normal / exponential (Meeusen-van den Broeck 1977; Greene 2008)
+* Normal / truncated-normal (Stevenson 1980)
+
+Heteroskedasticity
+------------------
+All kernels accept vector-valued ``sigma_u`` and ``sigma_v`` so that the
+caller can plug in ``sigma_u_i = exp(w_i' gamma)`` (Caudill-Ford-Gropper
+1995, Hadri 1999) or ``sigma_v_i = exp(r_i' eta)`` (Wang 2002).  The mean
+parameter ``mu`` of the truncated-normal may likewise be either a scalar
+or a vector ``mu_i = z_i' delta`` (Battese-Coelli 1995, Kumbhakar-Ghosh-
+McGuckin 1991).
+
+References
+----------
+Aigner, D., Lovell, C.A.K. & Schmidt, P. (1977). "Formulation and Estimation
+    of Stochastic Frontier Production Function Models." J. Econometrics 6, 21-37.
+Battese, G.E. & Coelli, T.J. (1988). "Prediction of Firm-level Technical
+    Efficiencies with a Generalized Frontier Production Function and Panel
+    Data." J. Econometrics 38, 387-399.
+Battese, G.E. & Coelli, T.J. (1992). "Frontier Production Functions,
+    Technical Efficiency and Panel Data: With Application to Paddy Farmers
+    in India." J. Productivity Analysis 3, 153-169.
+Battese, G.E. & Coelli, T.J. (1995). "A Model for Technical Inefficiency
+    Effects in a Stochastic Frontier Production Function for Panel Data."
+    Empirical Economics 20, 325-332.
+Caudill, S.B., Ford, J.M. & Gropper, D.M. (1995). "Frontier Estimation and
+    Firm-specific Inefficiency Measures in the Presence of Heteroscedasticity."
+    J. Business & Economic Statistics 13, 105-111.
+Greene, W.H. (2008). "The Econometric Approach to Efficiency Analysis." In
+    Fried, H., Lovell, C.A.K. & Schmidt, S. (eds), The Measurement of
+    Productive Efficiency and Productivity Growth, Oxford U.P.
+Jondrow, J., Lovell, C.A.K., Materov, I.S. & Schmidt, P. (1982). "On the
+    Estimation of Technical Inefficiency in the Stochastic Frontier
+    Production Function Model." J. Econometrics 19, 233-238.
+Meeusen, W. & van den Broeck, J. (1977). "Efficiency Estimation from
+    Cobb-Douglas Production Functions With Composed Error." IER 18, 435-444.
+Pitt, M.M. & Lee, L.-F. (1981). "The Measurement and Sources of Technical
+    Inefficiency in the Indonesian Weaving Industry." J. Development
+    Economics 9, 43-64.
[…]
