### 2026-04-18 0d9aa46
fix(frontier): self-audit — summary dump, TVD loop, Mills tail, cost panel
Self-review of 3f64b9f surfaced four issues. Fixed in this commit with
seven new regression tests (40 total, up from 33). Bulk of the complaint
came from summary() being unusable on real data.

Issues fixed:

1. summary() dumped per-observation diagnostic arrays — calling
   .summary() on n=800 data printed 37KB of noise (sigma_u_i, eps,
   efficiency_bc, efficiency_index etc rendered as raw numpy arrays).
   Override now hides array-valued diagnostics during the parent render
   and appends a Stata-style variance-decom
[…]
