### 2026-04-18 edc3141
feat(frontier): add predict(), marginal_effects(), TN multi-start
Continues the audit closure — three paper-grade features I flagged as
"touched but not yet built" in the previous review.

1. FrontierResult.predict(new_data, what=...):
   * what='frontier'              -> x_new' beta (deterministic frontier)
   * what='expected_inefficiency' -> marginal E[u_new] using fitted
                                      usigma / emean coefficients
   * what='expected_efficiency'   -> marginal E[exp(-u_new)] with
                                      closed-form MGF for all three
                      
[…]
