+++ tests/test_frontier.py
@@ -447,6 +447,148 @@ class TestPanelBC95:
 # ---------------------------------------------------------------------------
 
 
+class TestPredict:
+    def test_predict_frontier_matches_xbeta(self):
+        df = _simulate_hn_production(500, 0.2, 0.4, seed=201)
+        res = frontier(df, y="y", x=["x1", "x2"], dist="half-normal")
+        new = pd.DataFrame({"x1": [0.0, 1.0], "x2": [0.0, -0.5]})
+        fr = res.predict(new, what="frontier")
+        # Manually: cons + 1*x1 coef + (-0.5)*x2 coef
+        cons = res.params["_cons"]
+        b1 = res.params["x1"]
+        b2 = res.params["x2"]
+        expected = np.array([cons, cons + b1 - 0.5 * b2])
+        assert np.allclose(fr.values, expected, atol=1e-12)
+
+    def test_predict_expected_efficiency_in_bounds(self):
+        df = _simulate_hn_production(200, 0.2, 0.5, seed=202)
+        res = frontier(df, y="y", x=["x1"], dist="half-normal")
+        new = pd.DataFrame({"x1": np.linspace(-2, 2, 10)})
+        te = res.predict(new, what="expected_efficiency")
+        assert (te > 0).all() and (te <= 1).all()
+
+    def test_predict_expected_inefficiency_matches_closed_form(self):
+        df = _simulate_hn_production(300, 0.2, 0.5, seed=203)
+        res = frontier(df, y="y", x=["x1"], dist="half-normal")
+        new = pd.DataFrame({"x1": [0.0]})
+        E_u = res.predict(new, what="expected_inefficiency").iloc[0]
+        sigma_u = np.exp(res.params["ln_sigma_u"])
+        expected = sigma_u * np.sqrt(2.0 / np.pi)  # marginal HN
+        assert abs(E_u - expected) < 1e-10
+
+    def test_predict_handles_usigma(self):
+        """Heteroskedastic u: new data with different w should give different E[u]."""
+        rng = np.random.default_rng(204)
+        n = 1500
+        x1 = rng.normal(0, 1, n)
+        w1 = rng.normal(0, 1, n)
+        sigma_u_i = np.exp(-1.0 + 0.3 * w1)
+        u = np.abs(rng.normal(0, sigma_u_i))
+        v = rng.normal(0, 0.2, n)
+        y = 1.0 + 0.5 * x1 + v - u
+        df = pd.DataFrame({"y": y, "x1": x1, "w1": w1})
+        res = frontier(df, y="y", x=["x1"], dist="half-normal", usigma=["w1"])
+        new = pd.DataFrame({"x1": [0.0, 0.0], "w1": [-2.0, 2.0]})
+        E_u = res.predict(new, what="expected_inefficiency").values
+        # Different w → different E[u]; w=2 should give higher u (positive coef).
+        assert E_u[1] > E_u[0]
+
+    def test_predict_rejects_missing_columns(self):
+        df = _simulate_hn_production(200, 0.2, 0.4, seed=205)
+        res = frontier(df, y="y", x=["x1", "x2"], dist="half-normal")
+        with pytest.raises(KeyError, match="missing"):
+            res.predict(pd.DataFrame({"x1": [0.0]}))  # missing x2
+
+    def test_predict_rejects_unknown_what(self):
+        df = _simulate_hn_production(100, 0.2, 0.4, seed=206)
+        res = frontier(df, y="y", x=["x1"], dist="half-normal")
+        with pytest.raises(ValueError):
+            res.predict(df, what="nonsense")
+
+
+class TestMarginalEffects:
+    def test_bc95_marginal_effects_match_finite_difference(self):
[…]
