+++ src/statspai/frontier/sfa.py
@@ -150,6 +150,193 @@ class FrontierResult(EconometricResults):
             return "efficiency_jlms"
         raise ValueError(f"Unknown TE method: {method!r}")
 
+    # ------------------------------------------------------------------
+    # Out-of-sample prediction
+    # ------------------------------------------------------------------
+
+    def predict(
+        self,
+        new_data: pd.DataFrame,
+        what: str = "frontier",
+    ) -> pd.Series:
+        """Out-of-sample prediction.
+
+        Parameters
+        ----------
+        new_data : pandas.DataFrame
+            Must contain the frontier regressors and, if the model has
+            ``usigma`` / ``vsigma`` / ``emean`` covariates, those columns too.
+            Rows with any missing value are dropped.
+        what : {'frontier', 'expected_inefficiency', 'expected_efficiency'}
+            * ``'frontier'`` — deterministic frontier ``x_new' beta``.  This
+              is the maximum attainable output (production) or minimum cost.
+            * ``'expected_inefficiency'`` — marginal ``E[u_new]`` using the
+              fitted variance / mean parameters (no residual conditioning).
+            * ``'expected_efficiency'`` — marginal ``E[exp(-u_new)]``.
+
+        Returns
+        -------
+        pandas.Series
+            Indexed by the (post-dropna) rows of ``new_data``.
+        """
+        what = what.lower()
+        if what not in {"frontier", "expected_inefficiency", "expected_efficiency"}:
+            raise ValueError(f"Unknown what={what!r}.")
+
+        regressors = self.data_info.get("regressors", [])
+        usigma_cols = self.data_info.get("usigma_cols") or []
+        vsigma_cols = self.data_info.get("vsigma_cols") or []
+        emean_cols = self.data_info.get("emean_cols") or []
+
+        required = list(regressors) + list(usigma_cols) + list(vsigma_cols) + list(emean_cols)
+        missing = [c for c in required if c not in new_data.columns]
+        if missing:
+            raise KeyError(f"new_data is missing required columns: {missing}")
+        df_new = new_data[required].dropna().copy()
+        idx = df_new.index
+        n_new = len(df_new)
+        if n_new == 0:
+            raise ValueError("All rows dropped after removing missing values.")
+
+        # --- Rebuild design and pull fitted coefficients ---
+        params = self.params
+        beta = params.loc[["_cons"] + list(regressors)].to_numpy()
+        const = np.ones((n_new, 1))
+        X_new = np.concatenate([const, df_new[regressors].to_numpy()], axis=1)
+        frontier_hat = X_new @ beta
+
+        if what == "frontier":
+            return pd.Series(frontier_hat, index=idx, name="frontier")
+
+        # Need sigma_u / sigma_v / mu for new rows.
+        sigma_u_new = self._eval_sigma(
+            df_new, usigma_cols,
+            log_sigma_const_name="u__cons" if usigma_cols else "ln_sigma_u",
+            coef_prefix="u_",
+        )
+        mu_new = self._eval_mu(df_new, emean_cols)
+        dist = self.model_info.get("inefficiency_dist", "half-normal")
+
+        if dist == "half-normal":
+            E_u = sigma_u_new * np.sqrt(2.0 / np.pi)
+        elif dist == "exponential":
+            E_u = sigma_u_new
+        elif dist == "truncated-normal":
+            ratio = mu_new / sigma_u_new
[…]
