### 2026-04-18 bc418b1
feat(frontier): vce='opg'|'robust'|'cluster' + panel predict()
Closes the remaining gaps flagged in the previous review. Three
additions, all mirroring Stata's frontier/xtfrontier vce() semantics.

1. Variance-covariance options (sp.frontier):
   vce='oim'    — observed information (inverse numerical Hessian, default)
   vce='opg'    — outer product of gradients (BHHH)
   vce='robust' — sandwich H^-1 (S'S) H^-1 (White 1982)
   cluster='g'  — cluster-robust (Liang-Zeger 1986; implies vce='robust'
                  with per-cluster score aggregation).

2. Panel SE aggregation (sp.xtfrontier):
  
[…]
