### 2026-04-18 0a48a85
feat(frontier): Greene 2005 TFE/TRE + vce='bootstrap' + MC coverage check
Closes the last three gaps flagged from the audit review. Frontier module
now mirrors Stata frontier/xtfrontier feature-for-feature across core
workflows, with TRE as the recommended panel model when heterogeneity
should be separated from inefficiency.

1. Greene (2005) True Random Effects (model='tre'):
   y_it = alpha_i + x_it' beta + v_it + sign * u_it
   with alpha_i ~ N(0, sigma_alpha^2), v_it ~ N(0, sigma_v^2),
   u_it ~ N+(0, sigma_u^2).  alpha is integrated out by vectorised
   Gauss-Hermite quadrature (default n
[…]
