+++ src/statspai/frontier/panel.py
@@ -1,23 +1,20 @@
 """
 Panel stochastic frontier estimation — :func:`xtfrontier`.
 
-Implements three panel SFA designs:
+Implements five panel SFA designs:
 
 * ``model='ti'``  — Pitt-Lee (1981) time-invariant: ``u_it = u_i``.
-  With ``dist='half-normal'`` u_i ~ N^+(0, sigma_u^2); with
-  ``dist='truncated-normal'`` u_i ~ N^+(mu, sigma_u^2).
-
-* ``model='tvd'`` — Battese-Coelli (1992) time-varying decay:
-  ``u_it = exp(-eta (t - T_i)) * u_i``  with u_i ~ N^+(mu, sigma_u^2).
-  ``eta > 0`` means inefficiency declines over time; ``eta < 0`` rises.
-  ``eta = 0`` collapses to Pitt-Lee TI.
-
-* ``model='bc95'`` — Battese-Coelli (1995) inefficiency effects:
-  u_it ~ N^+(z_it' delta, sigma_u^2), independent across (i, t).  The
-  ``z`` covariates shift the mean of the truncated normal.  This is
-  equivalent to the cross-sectional :func:`frontier` call with
-  ``dist='truncated-normal'`` and ``emean=z``, but returns a panel-aware
-  result with per-unit efficiency aggregation.
+* ``model='tvd'`` — Battese-Coelli (1992) time-varying decay.
+* ``model='bc95'`` — Battese-Coelli (1995) inefficiency effects.
+* ``model='tfe'`` — Greene (2005) **True Fixed Effects**:
+  ``y_it = alpha_i + x_it' beta + v_it + s * u_it``, with ``alpha_i`` as
+  firm fixed effects (estimated via dummies) separately from inefficiency
+  ``u_it ~ N^+(0, sigma_u^2)``.  Resolves the Pitt-Lee confound where
+  unobserved firm heterogeneity was absorbed into ``u_i``.
+* ``model='tre'`` — Greene (2005) **True Random Effects**:
+  ``alpha_i ~ N(0, sigma_alpha^2)`` integrated out by Gauss-Hermite
+  quadrature; ``u_it`` independent inefficiency.  Same identification
+  as TFE but efficient when the RE assumption is reasonable.
 
 Equivalent to Stata's::
 
@@ -26,8 +23,9 @@ Equivalent to Stata's::
     xtfrontier y x, tvd
     frontier y x, dist(tnormal) emean(z)      (BC95)
 
-and R's ``frontier::sfa`` with panel data, or ``sfaR::sfacross`` /
-``sfaR::sfapanel``.
+and R's ``frontier::sfa`` / ``sfaR::sfacross``.  TFE/TRE mirror Greene
+(2005) "Reconsidering heterogeneity in panel data estimators of the
+stochastic frontier model", *J. Econometrics* 126, 269-303.
 """
 
 from __future__ import annotations
@@ -94,10 +92,25 @@ def xtfrontier(
     model = model.lower()
     dist = dist.lower().replace("_", "-")
 
-    if model not in {"ti", "tvd", "bc95"}:
+    if model not in {"ti", "tvd", "bc95", "tfe", "tre"}:
         raise ValueError(f"Unknown panel model: {model!r}.")
     if model == "tvd" and time is None:
         raise ValueError("model='tvd' requires a time variable.")
+
+    if model == "tfe":
+        return _fit_tfe(
+            data, y, x, id_col=id, time_col=time,
+            dist=dist, cost=cost,
+            vce=vce, cluster=cluster,
+            maxiter=maxiter, tol=tol, alpha=alpha,
+        )
+    if model == "tre":
+        return _fit_tre(
[…]
