+++ tests/test_frontier.py
@@ -602,7 +602,7 @@ class TestVarianceEstimators:
     def test_unknown_vce_raises(self):
         df = _simulate_hn_production(100, 0.2, 0.4, seed=504)
         with pytest.raises(ValueError, match="vce"):
-            frontier(df, y="y", x=["x1"], vce="bootstrap")
+            frontier(df, y="y", x=["x1"], vce="jackknife")
 
     def test_cluster_implies_robust(self):
         df = _simulate_hn_production(500, 0.2, 0.4, seed=505)
@@ -707,6 +707,153 @@ class TestTruncatedNormalRobustness:
         # TN has one more parameter (mu), so its LL must be >= HN's LL - epsilon.
         assert (res.diagnostics["log_likelihood"]
                 >= res_hn.diagnostics["log_likelihood"] - 0.5)
+
+
+class TestGreeneTrueEffects:
+    """Greene (2005) True Fixed/Random Effects SFA models."""
+
+    def test_tre_recovers_three_variance_components(self):
+        """TRE should separate alpha_i, v_it, and u_it."""
+        rng = np.random.default_rng(701)
+        N, T = 80, 6
+        id_ = np.repeat(np.arange(N), T)
+        t_ = np.tile(np.arange(T), N)
+        n = N * T
+        x1 = rng.normal(0, 1, n)
+        sigma_alpha_true = 0.4
+        sigma_v_true = 0.15
+        sigma_u_true = 0.3
+        alpha_i = rng.normal(0, sigma_alpha_true, N)
+        u_it = np.abs(rng.normal(0, sigma_u_true, n))
+        v = rng.normal(0, sigma_v_true, n)
+        y = 1.0 + 0.5 * x1 + np.repeat(alpha_i, T) + v - u_it
+        df = pd.DataFrame({"y": y, "x1": x1, "id": id_, "t": t_})
+
+        res = xtfrontier(df, y="y", x=["x1"], id="id", time="t", model="tre")
+        assert res.model_info["converged"]
+        assert abs(res.params["x1"] - 0.5) < 0.05
+        assert abs(np.exp(res.params["ln_sigma_v"]) - sigma_v_true) < 0.08
+        assert abs(np.exp(res.params["ln_sigma_u"]) - sigma_u_true) < 0.08
+        assert abs(np.exp(res.params["ln_sigma_alpha"]) - sigma_alpha_true) < 0.15
+
+    def test_tre_exposes_sigma_alpha_in_model_info(self):
+        df = _simulate_panel_ti(50, 5, 0.15, 0.3, seed=702)
+        res = xtfrontier(df, y="y", x=["x1"], id="id", time="t", model="tre")
+        assert "sigma_alpha" in res.model_info
+        assert res.model_info["sigma_alpha"] > 0
+
+    def test_tre_rejects_truncated_normal(self):
+        df = _simulate_panel_ti(30, 4, 0.2, 0.3, seed=703)
+        with pytest.raises(ValueError, match="tre currently|TRE currently"):
+            xtfrontier(df, y="y", x=["x1"], id="id", time="t",
+                       model="tre", dist="truncated-normal")
+
+    def test_tfe_recovers_with_long_t(self):
+        """TFE (brute-force firm dummies) works well when T is not too short."""
+        rng = np.random.default_rng(704)
+        N, T = 30, 20  # long T to mitigate incidental-parameters bias
+        id_ = np.repeat(np.arange(N), T)
+        t_ = np.tile(np.arange(T), N)
+        n = N * T
+        x1 = rng.normal(0, 1, n)
+        alpha_i = rng.normal(2.0, 0.5, N)
[…]

[…]
