### 2026-04-18 7c5e859
feat(frontier): conditional predict, RTS, usigma margins, metafrontier
Four additions covering the practitioner-oriented gaps:

1. FrontierResult.predict(what='conditional_inefficiency'|'conditional_efficiency'):
   When the dependent variable is present in new_data, compute the Jondrow
   posterior E[u|eps] / Battese-Coelli E[exp(-u)|eps] on held-out rows.
   Confirmed by test to exactly reproduce the in-sample efficiency()
   values when called with the training data.

2. FrontierResult.returns_to_scale(inputs=...):
   Stata-style RTS helper for Cobb-Douglas / translog frontiers.  Uses
   th
[…]
