+++ src/statspai/frontier/sfa.py
@@ -166,13 +166,20 @@ class FrontierResult(EconometricResults):
         new_data : pandas.DataFrame
             Must contain the frontier regressors and, if the model has
             ``usigma`` / ``vsigma`` / ``emean`` covariates, those columns too.
+            For ``what='conditional_*'`` the dependent variable ``y`` must
+            also be present so that the composed-error residual can be
+            computed and conditioned on.
             Rows with any missing value are dropped.
-        what : {'frontier', 'expected_inefficiency', 'expected_efficiency'}
-            * ``'frontier'`` — deterministic frontier ``x_new' beta``.  This
-              is the maximum attainable output (production) or minimum cost.
-            * ``'expected_inefficiency'`` — marginal ``E[u_new]`` using the
-              fitted variance / mean parameters (no residual conditioning).
+        what : {'frontier', 'expected_inefficiency', 'expected_efficiency',
+                'conditional_inefficiency', 'conditional_efficiency'}
+            * ``'frontier'`` — deterministic frontier ``x_new' beta``.
+            * ``'expected_inefficiency'`` — marginal ``E[u_new]``.
             * ``'expected_efficiency'`` — marginal ``E[exp(-u_new)]``.
+            * ``'conditional_inefficiency'`` — Jondrow posterior
+              ``E[u | eps_new]`` where ``eps_new = y_new - x_new'beta``;
+              requires ``y`` column in ``new_data``.
+            * ``'conditional_efficiency'`` — Battese-Coelli
+              ``E[exp(-u) | eps_new]``; requires ``y``.
 
         Returns
         -------
@@ -180,8 +187,12 @@ class FrontierResult(EconometricResults):
             Indexed by the (post-dropna) rows of ``new_data``.
         """
         what = what.lower()
-        if what not in {"frontier", "expected_inefficiency", "expected_efficiency"}:
-            raise ValueError(f"Unknown what={what!r}.")
+        valid = {
+            "frontier", "expected_inefficiency", "expected_efficiency",
+            "conditional_inefficiency", "conditional_efficiency",
+        }
+        if what not in valid:
+            raise ValueError(f"Unknown what={what!r}.  Valid: {sorted(valid)}.")
 
         regressors = self.data_info.get("regressors", [])
         usigma_cols = self.data_info.get("usigma_cols") or []
@@ -189,6 +200,15 @@ class FrontierResult(EconometricResults):
         emean_cols = self.data_info.get("emean_cols") or []
 
         required = list(regressors) + list(usigma_cols) + list(vsigma_cols) + list(emean_cols)
+        needs_y = what.startswith("conditional_")
+        if needs_y:
+            y_col = self.data_info.get("dep_var")
+            if y_col is None or y_col not in new_data.columns:
+                raise KeyError(
+                    f"what={what!r} requires the dependent variable "
+                    f"{y_col!r} to be present in new_data."
+                )
+            required = [y_col] + required
         missing = [c for c in required if c not in new_data.columns]
         if missing:
             raise KeyError(f"new_data is missing required columns: {missing}")
@@ -230,6 +250,35 @@ class FrontierResult(EconometricResults):
 
         if what == "expected_inefficiency":
             return pd.Series(E_u, index=idx, name="expected_inefficiency")
+
+        if what.startswith("conditional_"):
+            y_col = self.data_info.get("dep_var")
[…]
