+++ tests/test_frontier.py
@@ -814,6 +814,137 @@ class TestBootstrap:
         assert np.isclose(r1.std_errors["x1"], r2.std_errors["x1"])
 
 
+class TestConditionalPredict:
+    def test_conditional_efficiency_matches_training_efficiency(self):
+        """Posterior E[exp(-u)|eps] on held-out data should agree with efficiency() on same rows."""
+        df = _simulate_hn_production(1200, 0.2, 0.4, seed=1001)
+        # Fit on all data.
+        res = frontier(df, y="y", x=["x1", "x2"], dist="half-normal")
+        # predict with y present on the same rows — should match efficiency() output.
+        te_train = res.efficiency(method="bc").values
+        te_pred = res.predict(df, what="conditional_efficiency").values
+        assert np.allclose(te_train, te_pred, atol=1e-10)
+
+    def test_conditional_requires_y(self):
+        df = _simulate_hn_production(300, 0.2, 0.4, seed=1002)
+        res = frontier(df, y="y", x=["x1"], dist="half-normal")
+        no_y = df[["x1"]]
+        with pytest.raises(KeyError, match="requires the dependent"):
+            res.predict(no_y, what="conditional_efficiency")
+
+    def test_conditional_inefficiency_bounded(self):
+        df = _simulate_hn_production(400, 0.2, 0.5, seed=1003)
+        res = frontier(df, y="y", x=["x1"], dist="half-normal")
+        u_hat = res.predict(df, what="conditional_inefficiency").values
+        assert (u_hat >= -1e-6).all()
+
+
+class TestReturnsToScale:
+    def test_rts_correct_for_crs_dgp(self):
+        rng = np.random.default_rng(1101)
+        n = 3000
+        x1 = rng.normal(0, 1, n)
+        x2 = rng.normal(0, 1, n)
+        v = rng.normal(0, 0.2, n)
+        u = np.abs(rng.normal(0, 0.3, n))
+        y = 1.0 + 0.6 * x1 + 0.4 * x2 + v - u  # sum = 1 (CRS)
+        df = pd.DataFrame({"y": y, "x1": x1, "x2": x2})
+        res = frontier(df, y="y", x=["x1", "x2"])
+        rts = res.returns_to_scale()
+        assert abs(rts["rts"] - 1.0) < 0.05
+        assert rts["pvalue"] > 0.05  # fail to reject CRS
+        assert "CRS" in rts["interpretation"]
+
+    def test_rts_detects_irs(self):
+        rng = np.random.default_rng(1102)
+        n = 3000
+        x1 = rng.normal(0, 1, n)
+        x2 = rng.normal(0, 1, n)
+        v = rng.normal(0, 0.2, n)
+        u = np.abs(rng.normal(0, 0.3, n))
+        y = 1.0 + 0.7 * x1 + 0.6 * x2 + v - u  # sum = 1.3
+        df = pd.DataFrame({"y": y, "x1": x1, "x2": x2})
+        res = frontier(df, y="y", x=["x1", "x2"])
+        rts = res.returns_to_scale()
+        assert rts["rts"] > 1.1
+        assert rts["pvalue"] < 0.01
+        assert "IRS" in rts["interpretation"]
+
+
+class TestUsigmaMarginalEffects:
+    def test_usigma_me_matches_finite_difference(self):
+        rng = np.random.default_rng(1201)
+        n = 3000
+        x1 = rng.normal(0, 1, n)
+        w1 = rng.normal(0, 1, n)
+        sigma_u_i = np.exp(-1.0 + 0.4 * w1)
+        u = np.abs(rng.normal(0, sigma_u_i))
+        v = rng.normal(0, 0.2, n)
+        y = 1.0 + 0.5 * x1 + v - u
[…]
