### 2026-04-19 e876937
feat(frontier): ZISF + Latent-Class SFA + TFE jackknife bias correction
Closes the three frontier-frontier extensions left on the whiteboard:

1. sp.zisf() — Zero-Inefficiency SFA (Kumbhakar-Parmeter-Tsionas 2013):
   Mixture of fully-efficient (u=0, pure noise) and standard composed-
   error regimes.  Mixing probability p_i parameterised via logit on
   optional zprob covariates (constant if zprob=None).  Posterior
   P(efficient|eps) is exposed in diagnostics['p_efficient_posterior'].
   Recovery test: 30% true efficient share recovered as 0.286 on n=2000.

2. sp.lcsf() — 2-class Latent-Cla
[…]
