+++ tests/test_frontier.py
@@ -987,6 +987,148 @@ class TestMonteCarloCoverage:
         assert coverage_su >= 0.80, f"sigma_u coverage: {coverage_su}"
 
 
+class TestZeroInefficiency:
+    """Zero-Inefficiency SFA (Kumbhakar-Parmeter-Tsionas 2013)."""
+
+    def test_zisf_recovers_mixture_probability(self):
+        from statspai.frontier import zisf
+        rng = np.random.default_rng(1401)
+        n = 2000
+        x1 = rng.normal(0, 1, n)
+        eff_mask = rng.uniform(0, 1, n) < 0.3  # 30% efficient
+        u = np.where(eff_mask, 0.0, np.abs(rng.normal(0, 0.5, n)))
+        v = rng.normal(0, 0.15, n)
+        y = 1.0 + 0.5 * x1 + v - u
+        df = pd.DataFrame({"y": y, "x1": x1})
+        res = zisf(df, y="y", x=["x1"])
+        assert res.model_info["converged"]
+        assert abs(res.params["x1"] - 0.5) < 0.05
+        assert abs(np.exp(res.params["ln_sigma_u"]) - 0.5) < 0.1
+        assert abs(res.model_info["mean_p_efficient"] - 0.3) < 0.05
+
+    def test_zisf_with_zprob_covariate(self):
+        """Class probability should respond to a covariate."""
+        from statspai.frontier import zisf
+        rng = np.random.default_rng(1402)
+        n = 2000
+        x1 = rng.normal(0, 1, n)
+        z1 = rng.normal(0, 1, n)
+        # Higher z1 → higher chance of being efficient.
+        p_eff = 1.0 / (1.0 + np.exp(-z1))
+        eff_mask = rng.uniform(0, 1, n) < p_eff
+        u = np.where(eff_mask, 0.0, np.abs(rng.normal(0, 0.4, n)))
+        v = rng.normal(0, 0.15, n)
+        y = 1.0 + 0.5 * x1 + v - u
+        df = pd.DataFrame({"y": y, "x1": x1, "z1": z1})
+        res = zisf(df, y="y", x=["x1"], zprob=["z1"])
+        assert res.model_info["converged"]
+        # Logit coef on z1 should be clearly positive.
+        assert res.params["p_z1"] > 0.3
+
+    def test_zisf_rejects_unsupported_dist(self):
+        from statspai.frontier import zisf
+        df = _simulate_hn_production(200, 0.15, 0.4, seed=1403)
+        with pytest.raises(ValueError, match="half-normal"):
+            zisf(df, y="y", x=["x1"], dist="exponential")
+
+
+class TestLatentClassSFA:
+    """Latent-Class SFA (Orea-Kumbhakar 2004)."""
+
+    def test_lcsf_recovers_two_classes(self):
+        from statspai.frontier import lcsf
+        rng = np.random.default_rng(1501)
+        n = 2500
+        x1 = rng.normal(0, 1, n)
+        class_1 = rng.uniform(0, 1, n) < 0.5
+        u = np.abs(rng.normal(0, 0.3, n))
+        v = rng.normal(0, 0.15, n)
+        y = np.where(
+            class_1,
+            1.0 + 0.8 * x1 + v - u,
+            2.0 + 0.3 * x1 + v - u,
+        )
+        df = pd.DataFrame({"y": y, "x1": x1})
+        res = lcsf(df, y="y", x=["x1"])
+        assert res.model_info["converged"]
+        # Class recovery: one class should have x1 ≈ 0.8, the other ≈ 0.3.
+        c1_x1 = res.params["c1:x1"]
+        c2_x1 = res.params["c2:x1"]
[…]
