### 2026-04-19 be59260
feat(frontier+smart): Malmquist TFP index + translog helper + verify engine
Dual-topic commit that landed together via pre-commit hook auto-staging.

=== frontier additions ===

1. sp.malmquist() — Fare-Grosskopf-Lindgren-Roos (1994) Malmquist TFP
   index via period-by-period parametric frontier fits.  Returns per-
   transition decomposition M = EC x TC:
     * EC = D^{t+1}(x^{t+1}, y^{t+1}) / D^t(x^t, y^t)  (efficiency change)
     * TC = geometric mean of two cross-period frontier-shift ratios
            (technical change)
   Row-wise M == EC * TC verified to rtol=1e-8.  Cost frontiers
  
[…]
