+++ src/statspai/smart/recommend.py
@@ -103,6 +103,121 @@ class RecommendationResult:
         lines.append("\n" + "=" * 70)
         return "\n".join(lines)
 
+    def to_latex(self, caption: Optional[str] = None,
+                 label: str = "tab:recommendation") -> str:
+        r"""Export recommendations as a booktabs LaTeX table.
+
+        If ``verify=True`` was used when calling ``recommend()``, the
+        table includes the empirical verification columns (score,
+        bootstrap stability, placebo pass-rate, subsample agreement) —
+        suitable for dropping into a paper appendix.
+
+        Parameters
+        ----------
+        caption : str, optional
+            Table caption. Defaults to the detected design.
+        label : str
+            LaTeX label for cross-referencing.
+
+        Returns
+        -------
+        str
+            LaTeX source (booktabs + threeparttable).
+        """
+        has_verify = any(
+            isinstance(r.get("verify"), dict)
+            and np.isfinite(r["verify"].get("score", np.nan))
+            for r in self.recommendations
+        )
+        if caption is None:
+            caption = (
+                f"StatsPAI recommended estimators for "
+                f"{self.design.replace('_', ' ')} design"
+                + (" (with empirical verification)" if has_verify else "")
+            )
+
+        def _esc(s: str) -> str:
+            return (str(s).replace("\\", r"\textbackslash{}")
+                         .replace("_", r"\_")
+                         .replace("&", r"\&")
+                         .replace("%", r"\%")
+                         .replace("#", r"\#"))
+
+        if has_verify:
+            header = (r"Rank & Method & Function & "
+                      r"Score & Stab. & Plac. & Subs. \\")
+            col_spec = "rllrrrr"
+        else:
+            header = r"Rank & Method & Function & Reason \\"
+            col_spec = "rllp{6cm}"
+
+        lines = [
+            r"\begin{table}[!htbp]",
+            r"\centering",
+            r"\begin{threeparttable}",
+            rf"\caption{{{_esc(caption)}}}",
+            rf"\label{{{label}}}",
+            rf"\begin{{tabular}}{{{col_spec}}}",
+            r"\toprule",
+            header,
+            r"\midrule",
+        ]
+
+        for i, rec in enumerate(self.recommendations, 1):
+            method = _esc(rec["method"])
+            func = rf"\texttt{{sp.{_esc(rec['function'])}()}}"
+            if has_verify:
+                v = rec.get("verify") or {}
+                score = v.get("score", np.nan)
+                stab = (v.get("stability") or {}).get("score", np.nan)
+                plac = (v.get("placebo") or {}).get("score", np.nan)
+                subs = (v.get("subsample") or {}).get("score", np.nan)
+                err = v.get("error")
+
+                def _fmt(x):
+                    return f"{x:.0f}" if isinstance(x, float) and np.isfinite(x) else "--"
+
+                marker = f" {{\\tiny\\textit{{({_esc(err)})}}}}" if err else ""
+                lines.append(
+                    f"{i} & {method}{marker} & {func} & "
+                    f"{_fmt(score)} & {_fmt(stab)} & {_fmt(plac)} & {_fmt(subs)} \\\\"
+                )
+            else:
+                reason = _esc(rec.get("reason", ""))
+                lines.append(f"{i} & {method} & {func} & {reason} \\\\")
+
+        lines.extend([
+            r"\bottomrule",
+            r"\end{tabular}",
+            r"\begin{tablenotes}",
+            r"\footnotesize",
+        ])
+        if has_verify:
+            lines.extend([
+                r"\item \textbf{Score}: weighted composite in [0, 100] "
+                r"(higher = more empirical support).",
+                r"\item \textbf{Stab.}: bootstrap stability "
[…]
