+++ src/statspai/frontier/malmquist.py
@@ -1,0 +1,294 @@
+"""
+Malmquist productivity index for parametric stochastic frontiers.
+
+Given panel data, fit a separate frontier ``F^t`` for each period ``t``
+and compute the Fare-Grosskopf-Lindgren-Roos (1994) Malmquist index
+between adjacent periods::
+
+    M_O^{t -> t+1}
+        = [ D^t(x^{t+1}, y^{t+1}) / D^t(x^t, y^t) ]
+        x [ D^{t+1}(x^{t+1}, y^{t+1}) / D^{t+1}(x^t, y^t) ]^{1/2}
+
+The canonical output-oriented decomposition is ``M = EC x TC``:
+
+* **EC** (efficiency change) = ``D^{t+1}(x^{t+1}, y^{t+1}) / D^t(x^t, y^t)``
+* **TC** (technical change) = ``[ D^t(x^{t+1}, y^{t+1}) / D^{t+1}(x^{t+1}, y^{t+1})
+        x D^t(x^t, y^t) / D^{t+1}(x^t, y^t) ]^{1/2}``
+
+We use the frontier-distance definition ``D^s(x, y) = y / exp(x' beta_s)``
+(deterministic approximation standard in parametric Malmquist work;
+Coelli-Rao 2005, "Total Factor Productivity Growth in Agriculture").
+
+``M > 1`` = productivity growth, ``EC > 1`` = catch-up to own-period
+frontier, ``TC > 1`` = outward frontier shift.
+
+References
+----------
+Fare, R., Grosskopf, S., Lindgren, B. & Roos, P. (1994).  "Productivity
+    changes in Swedish pharmacies 1980-1989: A non-parametric Malmquist
+    approach."  J. Productivity Analysis 3, 85-101.
+Coelli, T.J. & Rao, D.S.P. (2005).  "Total factor productivity growth in
+    agriculture: a Malmquist index analysis of 93 countries, 1980-2000."
+    Agricultural Economics 32, 115-134.
+Kumbhakar, S.C., Wang, H.J. & Horncastle, A.P. (2015).  A Practitioner's
+    Guide to Stochastic Frontier Analysis, Cambridge U.P., Chapter 14.
+"""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Any, Dict, List, Optional
+
+import numpy as np
+import pandas as pd
+
+from .sfa import FrontierResult, frontier as _frontier
+
+
+@dataclass
+class MalmquistResult:
+    """Container for Malmquist productivity index decomposition."""
+
+    index_table: pd.DataFrame
+    """Wide table: one row per (id, period pair) with columns
+    ``['m_index', 'ec', 'tc']`` plus the original id / period columns."""
+
+    period_frontiers: Dict[Any, FrontierResult]
+    """Frontier fit per period."""
+
+    summary_by_period: pd.DataFrame
+    """Mean M / EC / TC per period transition."""
+
+    data_info: Dict[str, Any]
+
+    def summary(self) -> str:
+        lines = [
+            "=" * 80,
+            "Malmquist Productivity Index (Fare-Grosskopf-Lindgren-Roos 1994)",
+            "=" * 80,
+            f"Periods : {self.data_info['periods']}",
+            f"N units : {self.data_info['n_units']}",
+            f"Total transitions: {len(self.index_table)}",
+            "",
+            "Mean Malmquist components by period transition:",
+            self.summary_by_period.round(4).to_string(),
+            "",
+            "Interpretation: M>1 = productivity growth, EC>1 = catch-up,",
+            "TC>1 = frontier moved outward.",
+        ]
+        return "\n".join(lines)
+
+
+def malmquist(
[…]
