+++ src/statspai/smart/__init__.py
@@ -18,6 +18,7 @@ from .sensitivity import sensitivity_dashboard, SensitivityDashboard
 from .publication import pub_ready, PubReadyResult
 from .replicate import replicate, list_replications
 from .verify import verify_recommendation
+from .benchmark import verify_benchmark
 
 __all__ = [
     "recommend", "RecommendationResult",
@@ -26,5 +27,5 @@ __all__ = [
     "sensitivity_dashboard", "SensitivityDashboard",
     "pub_ready", "PubReadyResult",
     "replicate", "list_replications",
-    "verify_recommendation",
+    "verify_recommendation", "verify_benchmark",
 ]

[…]
