### 2026-04-19 bd2828a
chore(release): finalize 0.9.3 — CHANGELOG + frontier/smart polish
Final polish on top of feat/econ-trinity merge to close the 0.9.3 release:

- CHANGELOG.md: add 0.9.3 release section covering frontier overhaul
  (Jondrow sign fix + Stata/R-grade rewrite), multilevel overhaul,
  econometric trinity P0 pillars, and smart verify engine.
- README.md: update to 0.9.3 headline + version bump.
- frontier/{_core,sfa,mixture,panel,malmquist}.py: tail-end correctness
  touch-ups identified during release prep.
- smart/{recommend,verify}.py: verify engine wiring + recommend()
  verify=True opt-in.
- te
[…]
