+++ src/statspai/smart/verify.py
@@ -1,33 +1,42 @@
 """
-Empirical verification for estimator recommendations.
+Resampling-stability checks for estimator recommendations.
 
-The ``recommend()`` engine is a rule-based decision tree: given data shape
-and research design, it returns ranked methods as a *prior*. This module
-adds a *posterior* signal by running each recommended method on resamples
-of the actual data and scoring:
+The ``recommend()`` engine is a rule-based decision tree: given data
+shape and research design, it returns ranked methods as a *prior*.
+This module runs each recommended method on resamples of the observed
+data and produces a composite **stability_score** in [0, 100]:
 
 1. **Bootstrap stability** — coefficient of variation across B bootstraps.
-   Low CV → method gives consistent estimates under resampling.
-2. **Placebo pass** — p-value under permuted treatment. Should be >α
-   (no spurious effect when treatment is randomly shuffled).
+   Low CV → estimate is stable under row/cluster resampling.
+2. **Placebo pass rate** — share of permuted-treatment runs with p>0.10.
+   Note: unconditional permutation destroys confounder structure and
+   therefore has limited power for selection-on-observables designs.
 3. **Subsample agreement** — sign agreement across K random 50% splits.
    High agreement → robust to sample composition.
 
-These are combined into a ``verify_score`` in [0, 100] used only to
-re-rank the top-k recommendations. The rule-based design detection is
-untouched — this is *additive* empirical evidence, not a replacement.
+WHAT THIS SCORE IS NOT
+----------------------
+The composite is a **stability** score, not a validity score. It does
+NOT establish:
+
+- that identifying assumptions hold (unconfoundedness, parallel trends,
+  exclusion restrictions, continuity at the cutoff, ...);
+- that the method is unbiased for the estimand of interest;
+- that the method beats alternatives on MSE against ground truth.
+
+A biased estimator on observational data can be perfectly stable, pass
+the (unconditional) placebo test, and agree across subsamples — and
+score near 100. Treat this score as "does the method behave predictably
+on this dataset," not as empirical evidence that it is correct.
 
 Design notes
 ------------
 - Opt-in only (``verify=False`` default in ``recommend()``).
 - Time budget enforced; B automatically reduced if base method is slow.
 - Failures per-method are caught — one bad fit does not poison the run.
-- No ground-truth required: all signals are derivable from the data
-  itself, so this works for any dataset the user brings.
-
-This is NOT a claim of "which method has lowest RMSE" — without a true
-counterfactual we cannot measure accuracy. It measures *behavior under
-resampling* on the data you have.
+- Independent RNG streams for bootstrap / placebo / subsample so that
+  results are reproducible even when the bootstrap budget is exhausted
+  early.
 """
 
 from __future__ import annotations
@@ -228,10 +237,24 @@ def _placebo_pass(
     """Permutation test: shuffle treatment, expect null.
 
     Returns share of placebo runs where p>0.10 (correctly non-significant).
+
+    Three distinct outcomes:
+    - ``applicable=False`` (score=NaN): treatment column not identifiable
+      from the recommendation — placebo cannot run at all. This drops
+      the component from the composite average (NaN-safe).
+    - ``applicable=True, total=0`` (score=0): placebo ran but no p-value
+      could be extracted from any rep. We score 0 rather than NaN so
+      methods that silently lack a p-value are penalised rather than
+      escaping the placebo check via NaN renormalisation.
[…]

[…]
