### 2026-04-20 85c589f
chore(release): 0.9.3 polish — P2 hardening + self-audit follow-ups
P2 source fixes (11):
- A1: __all__ completeness — add ~80 missing public exports incl. W,
  impactplot, mediate_sensitivity, synthplot, synth_*, moran_*, gwr_*,
  rd2d_*, decompose/rif/ffl/fairlie, lingam/ges, arima/bvar/garch,
  dataset helpers. `from statspai import *` now pulls 554 names.
- A2: metafrontier gains lp_tol parameter with automatic near-infeasible
  retry at 1e-5; failure error message dynamically suggests next tol.
- A3: predict() decoupled from hardcoded "u__cons"/"v__cons"/"mu__cons";
  _core.const_name_for
[…]
