### 2026-04-20 23269d9
feat(multilevel): 0.9.4 — AGHQ + Gamma/NegBin/Ordinal GLMMs
Closes the three GLMM gaps flagged in the 0.9.3 self-audit. Additive
only — every existing meglm/melogit/mepoisson call returns the same
fit (β, var components unchanged).

AGHQ
- New `nAGQ` argument on meglm + wrappers; nAGQ=1 ≡ Laplace (verified
  identical to 1e-10), nAGQ>1 invokes adaptive Gauss-Hermite quadrature
  recentered at û_j with curvature 1/H_j. Restricted to scalar random
  effects (matches lme4::glmer). Fixes Laplace bias on small clusters
  with binary outcomes (var(u): 1.34 → 1.75 vs truth 2.25 on n_per=3).

New fami
[…]
