+++ src/statspai/multilevel/__init__.py
@@ -4,9 +4,11 @@ Multilevel / mixed-effects models.
 Exports
 -------
 mixed, MixedResult                 Linear mixed effects (Gaussian response)
-meglm, melogit, mepoisson, MEGLMResult
-                                   Generalised linear mixed models via
-                                   Laplace approximation
+meglm, melogit, mepoisson,
+menbreg, megamma, meologit,
+MEGLMResult                        Generalised linear mixed models via
+                                   Laplace approximation or adaptive
+                                   Gauss-Hermite quadrature (``nAGQ``).
 icc                                Intra-class correlation with 95% CI
 lrtest                             Likelihood-ratio test between two fitted
                                    mixed models (with chi-bar² boundary
@@ -15,7 +17,8 @@ lrtest                             Likelihood-ratio test between two fitted
 """
 
 from .lmm import MixedResult, mixed
-from .glmm import MEGLMResult, meglm, melogit, mepoisson
+from .glmm import MEGLMResult, meglm, melogit, mepoisson, menbreg, megamma
+from ._ordinal import meologit
 from .diagnostics import icc
 from .comparison import lrtest
 
@@ -25,6 +28,9 @@ __all__ = [
     "meglm",
     "melogit",
     "mepoisson",
+    "menbreg",
+    "megamma",
+    "meologit",
     "MEGLMResult",
     "icc",
     "lrtest",
