+++ tests/test_multilevel.py
@@ -353,6 +353,230 @@ class TestMEPoisson:
 
 
 # ---------------------------------------------------------------------------
+# Adaptive Gauss-Hermite quadrature (nAGQ > 1)
+# ---------------------------------------------------------------------------
+
+
+class TestAGHQ:
+    """AGHQ should match Laplace at nAGQ=1, beat it on small clusters."""
+
+    def _small_cluster_binary(self, seed: int = 2026,
+                              n_groups: int = 200, n_per: int = 3,
+                              sigma_u: float = 1.5, beta_x: float = 0.8):
+        rng = np.random.default_rng(seed)
+        g = np.repeat(np.arange(n_groups), n_per)
+        x = rng.normal(0, 1, n_groups * n_per)
+        u = np.repeat(rng.normal(0, sigma_u, n_groups), n_per)
+        eta = -0.5 + beta_x * x + u
+        p = 1.0 / (1.0 + np.exp(-eta))
+        y = rng.binomial(1, p)
+        return pd.DataFrame({"y": y, "x": x, "g": g}), sigma_u, beta_x
+
+    def test_nagq1_equals_laplace(self):
+        df, _, _ = self._small_cluster_binary()
+        r1 = sp.melogit(df, "y", ["x"], "g", nAGQ=1)
+        # nAGQ=1 path uses the Laplace branch — verify by method label.
+        assert r1._method == "laplace"
+
+    def test_nagq_label(self):
+        df, _, _ = self._small_cluster_binary()
+        r7 = sp.melogit(df, "y", ["x"], "g", nAGQ=7)
+        assert r7._method == "AGHQ(nAGQ=7)"
+
+    def test_aghq_logL_at_least_laplace(self):
+        # The integrated log-likelihood under AGHQ must be >= Laplace
+        # (more accurate quadrature, higher likelihood at the optimum).
+        df, _, _ = self._small_cluster_binary()
+        r1 = sp.melogit(df, "y", ["x"], "g", nAGQ=1)
+        r7 = sp.melogit(df, "y", ["x"], "g", nAGQ=7)
+        # Allow tiny numerical slack — AGHQ should not be materially worse.
+        assert r7.log_likelihood >= r1.log_likelihood - 1e-3
+
+    def test_aghq_var_less_biased(self):
+        # On small clusters, Laplace severely underestimates var(u).
+        # AGHQ should be closer to truth than Laplace.
+        df, sigma_u, _ = self._small_cluster_binary()
+        true_var = sigma_u ** 2
+        r1 = sp.melogit(df, "y", ["x"], "g", nAGQ=1)
+        r7 = sp.melogit(df, "y", ["x"], "g", nAGQ=7)
+        v1 = r1.variance_components["var(_cons)"]
+        v7 = r7.variance_components["var(_cons)"]
+        assert abs(v7 - true_var) < abs(v1 - true_var)
+
+    def test_aghq_converges_in_nodes(self):
+        # nAGQ=15 should be ~indistinguishable from nAGQ=7 (4 sig figs).
+        df, _, _ = self._small_cluster_binary()
+        r7 = sp.melogit(df, "y", ["x"], "g", nAGQ=7)
+        r15 = sp.melogit(df, "y", ["x"], "g", nAGQ=15)
+        # log-likelihood and beta should agree to 3 decimals.
+        assert abs(r7.log_likelihood - r15.log_likelihood) < 1e-2
+        assert abs(r7.fixed_effects["x"] - r15.fixed_effects["x"]) < 1e-2
+
+    def test_aghq_rejects_random_slopes(self):
+        # AGHQ is restricted to scalar random effects.
+        df = _random_slope_panel(n_groups=20, n_per=20)
[…]
