+++ src/statspai/multilevel/glmm.py
@@ -1,60 +1,72 @@
 """
-Generalized linear mixed models (GLMM) via Laplace approximation.
+Generalized linear mixed models (GLMM).
 
-Supported families and links (canonical only, to keep the Laplace
-approximation clean and the observed information analytic):
+Estimation
+----------
+Two estimation paths are exposed via the ``nAGQ`` argument:
 
-    * ``gaussian`` — identity link.  Mostly for cross-checks against
-      ``mixed()``; for production Gaussian fits use the closed-form
-      linear mixed model.
-    * ``binomial`` — logit link (``sp.melogit``).  Accepts both
-      0/1 Bernoulli and count/trials responses via ``trials=``.
-    * ``poisson``  — log link (``sp.mepoisson``).  Optional exposure
-      offsets via ``offset=``.
+    nAGQ = 1   →  Laplace approximation (default).  Fast; equivalent to
+                  lme4's ``glmer(nAGQ = 1)`` and Stata ``meglm`` default.
+    nAGQ ≥ 2   →  Adaptive Gauss-Hermite quadrature (AGHQ) with ``nAGQ``
+                  nodes per scalar random effect.  Required for unbiased
+                  small-cluster binary outcomes; matches lme4
+                  ``glmer(nAGQ = k)`` and Stata ``meglm intpoints(k)``.
+                  Currently restricted to single scalar random effects
+                  (``q = 1``, e.g. random intercept only) — same
+                  restriction lme4 imposes; tensor-product AGHQ over
+                  random slopes is not enabled because the cost grows
+                  as nAGQ^q.
 
-Model:
+Families
+--------
+Five exponential / dispersion families are supported; each has its
+canonical link plus the most common practical link reported by Stata
+``meglm``:
 
-    η_ij = x_ij' β + z_ij' u_j,    u_j ~ N(0, G),
-    y_ij | u_j ~ Family( g⁻¹(η_ij) ).
+    * ``gaussian``  identity link (``meglm`` Gaussian-with-RE).
+    * ``binomial``  logit link  (``melogit``).  Bernoulli or counts/trials.
+    * ``poisson``   log link    (``mepoisson``).
+    * ``gamma``     log link    — dispersion ``φ`` estimated by ML.
+    * ``nbinomial`` log link    — NB-2 (mean-dispersion ``α`` estimated).
+
+Ordinal-logit GLMM (``meologit``) lives in :mod:`._ordinal`; it shares
+the result class but uses a dedicated fitter because of its threshold
+parameters.
 
 The integrated log-likelihood
 
-    ℓ(β, θ) = Σ_j log ∫ f(y_j | β, u_j) φ(u_j; 0, G) du_j
+    ℓ(β, θ, ψ) = Σ_j log ∫ f(y_j | β, ψ, u_j) φ(u_j; 0, G(θ)) du_j
 
-has no closed form for non-Gaussian families; we approximate each
-integral by a first-order Laplace expansion around the conditional
-mode û_j, giving
+has no closed form for non-Gaussian families.  Each integral is
+approximated either by a first-order Laplace expansion around the
+conditional mode û_j (``nAGQ=1``) or by adaptive Gauss-Hermite
+quadrature recentered at û_j with curvature 1/H_j (``nAGQ>1``).
 
-    ℓ_j(β, θ) ≈ log f(y_j | β, û_j) − ½ û_j' G⁻¹ û_j
-               − ½ log |G| − ½ log |Z_j' W_j Z_j + G⁻¹|,
-
-where W_j is the diagonal GLM weight matrix at û_j.  The outer
-optimisation over (β, θ) is done with L-BFGS; the inner optimisation
-for û_j is a few Newton steps.
-
-Fixed-effect covariance uses the standard GLMM formula
[…]
