### 2026-04-20 39d7aa8
fix(synth): migrate 3 stale _solve_weights call sites (8 tests)
The a4036a2 simplex-solver centralization changed
SyntheticControl._solve_weights to require X_treated, X_donors,
run_nested and to return a dict {w, v, loss, ...} rather than a
flat ndarray. Three downstream call sites weren't migrated and
crashed with TypeError on every invocation:

- sensitivity.py:92 (_fit_scm_core, hit by all synth_loo /
  synth_time_placebo / synth_donor_pool / synth_pre_rmspe paths)
- power.py:265 (synth_power main fit)
- power.py:92 (placebo loop inside _build_null_distribution)

Fixes:
- sensitivity.py + 
[…]
