### 2026-04-20 abd543f
fix(synth): align power placebo with scm.py:888 + lock LOO ↔ main consistency
Self-audit follow-ups on commit 39d7aa8:

(1) power.py:92 placebo loop now mirrors SyntheticControl's own
in-space placebo (scm.py:888) instead of hard-coding equal-V Y-as-X.
This is data-equivalent under power's current public API (which doesn't
expose covariates/v_method, so _has_predictors=False and
_should_run_nested()=False ⇒ original code path), but eliminates the
divergence that would silently bite the moment a covariate-aware power
function is added — predictor columns now follow the placebo unit just
like in
[…]
