### 2026-04-20 36b0555
chore(release): 0.9.3 — merge 0.9.4 GLMM content + fix numpy 2.x trapz
- Fold 0.9.4 GLMM hardening (AGHQ nAGQ>1; new families megamma/menbreg/
  meologit; cross-family AIC comparability; 18 new multilevel tests)
  into the 0.9.3 CHANGELOG as pillar #3; bump pyproject + __init__
  back to 0.9.3 so the release tag is a single canonical 0.9.3.
- tests/test_frontier.py: np.trapz → np.trapezoid (numpy 2.x removed
  trapz on Windows CI). Closes 3 TestKernelMath failures.
- Also documents the two synth _solve_weights migrations
  (commits 39d7aa8 + abd543f) under 'Synth API-drift fixes'.

Local pytes
[…]
