### 2026-04-20 221067f
docs: sync paper.md / docs / READMEs to v0.9.3 (P0+P1+P2)
Closes the documentation drift that left README/docs/JOSS pointing at
pre-0.9.0 state despite four mega-releases (0.9.0 Synth, 0.9.1 RD,
0.9.2 Decomposition, 0.9.3 Frontier+Multilevel+GLMM+Trinity).

P0 — JOSS + site landing
  * paper.md: rewrite Summary, Statement of Need, State of the Field,
    Methodological Coverage, Unique Features, and Validation to cover
    all four mega-releases; update function count 390+ → 550+; date
    2026-04-06 → 2026-04-20.
  * paper.bib: add @bach2022doubleml, @bates2015lme4 (cited from
    rewritten S
[…]
