+++ paper.md
@@ -7,6 +7,9 @@ tags:
   - difference-in-differences
   - regression discontinuity
   - synthetic control
+  - stochastic frontier
+  - mixed-effects models
+  - decomposition analysis
   - machine learning
   - heterogeneous treatment effects
 authors:
@@ -20,60 +23,280 @@ affiliations:
   - name: Stanford REAP Program, Stanford University, United States
     index: 2
     ror: 00f54p054
-date: 6 April 2026
+date: 20 April 2026
 bibliography: paper.bib
 ---
 
 # Summary
 
-`StatsPAI` is an open-source Python package that provides a unified API for causal inference and applied econometrics. With a single import (`import statspai as sp`), researchers access over 390 functions spanning classical econometric models, modern ML-based causal methods, and publication-ready output generation. The package consolidates functionality that previously required dozens of separate R packages or proprietary software such as Stata into one coherent library. Uniquely, `StatsPAI` is *agent-native*: every function exposes structured result objects and machine-readable schemas (`list_functions()`, `describe_function()`, `function_schema()`), making it the first econometrics toolkit purpose-built for LLM-driven research workflows while remaining fully ergonomic for human researchers.
+`StatsPAI` is an open-source Python package providing a unified API for
+causal inference and applied econometrics. A single import
+(`import statspai as sp`) exposes **550+ public functions** spanning
+classical econometric models, modern ML-based causal methods, stochastic
+frontier analysis, multilevel mixed-effects models, decomposition
+analysis, and publication-ready output generation. The package
+consolidates functionality that previously required dozens of separate
+R packages (`did`, `rdrobust`, `Synth`, `sfaR`, `lme4`, `oaxaca`,
+`ddecompose`, `mlogit`, `grf`, …) or proprietary software such as
+Stata (`frontier`, `xtfrontier`, `mixed`, `meglm`, `mixlogit`, `ivqreg`,
+…) into one coherent library. Uniquely, `StatsPAI` is *agent-native*:
+every function exposes structured result objects and machine-readable
+schemas (`list_functions()`, `describe_function()`, `function_schema()`),
+making it the first econometrics toolkit purpose-built for LLM-driven
+research workflows while remaining fully ergonomic for human researchers.
 
 # Statement of Need
 
-Empirical researchers face a fragmented software landscape for causal inference. Stata lacks modern ML causal methods; R scatters them across 20+ packages with incompatible APIs; Python's existing libraries occupy non-overlapping niches. `DoWhy` [@sharma2020dowhy] emphasizes causal graphs and assumption refutation. `EconML` [@econml] focuses on heterogeneous treatment effects. `CausalML` [@chen2020causalml] specializes in uplift modeling. None provides the full empirical workflow---from OLS and panel models through DML and causal forests to publication tables in Word, Excel, and LaTeX---in a single interface.
+Empirical researchers face a fragmented software landscape for causal
+inference and applied econometrics. Stata offers a broad but proprietary
+and closed ecosystem that lacks modern ML causal methods and exposes no
+programmatic schema for AI agent integration. R scatters equivalent
+functionality across twenty or more packages with incompatible APIs,
+data structures, and output conventions. Python's existing libraries
[…]
