+++ paper.bib
@@ -495,3 +495,24 @@
   year={2020},
   doi={10.48550/arXiv.2002.11631}
 }
+
+@article{bach2022doubleml,
+  title={{DoubleML} -- An object-oriented {DoubleML} package for the {Python} ecosystem},
+  author={Bach, Philipp and Chernozhukov, Victor and Kurz, Malte S. and Spindler, Martin},
+  journal={Journal of Machine Learning Research},
+  volume={23},
+  number={53},
+  pages={1--6},
+  year={2022}
+}
+
+@article{bates2015lme4,
+  title={Fitting linear mixed-effects models using {lme4}},
+  author={Bates, Douglas and M{\"a}chler, Martin and Bolker, Benjamin and Walker, Steve},
+  journal={Journal of Statistical Software},
+  volume={67},
+  number={1},
+  pages={1--48},
+  year={2015},
+  doi={10.18637/jss.v67.i01}
+}
