### 2026-04-20 43f1928
fix: 0.9.3 post-release bugfixes — use_chinese Linux, regtable output, did docstring
Four user-reported regressions from the 0.9.3 smoke test. No version bump.

* plots/themes.py: use_chinese() auto-detect now covers all Linux CJK
  fonts (Noto CJK JP/TC/KR, WenQuanYi Zen Hei, Source Han) plus a
  substring fallback; warning message documents the apt install recipe.
* output/regression_table.py, output/estimates.py: regtable / esttab /
  mean_comparison no longer auto-print (removes REPL/Jupyter double
  display). RegtableResult + MeanComparisonResult now store _output and
  dispatch __str__ o
[…]
