+++ tests/test_v093_bugfixes.py
@@ -1,0 +1,246 @@
+"""Regression tests for 0.9.3 bug fixes reported in production.
+
+Covers:
+1. `use_chinese()` auto-detects Linux-common Chinese fonts
+   (Noto Sans/Serif CJK JP/TC/KR, WenQuanYi Zen Hei, Source Han Sans/Serif).
+2. `RegtableResult` / `MeanComparisonResult` / `EstimateTableResult`
+   do not auto-print during construction (no double-print in REPL/Jupyter).
+3. `regtable(..., output="latex")` actually returns LaTeX from `str(result)`.
+4. `did()` docstring spells out the `treat` semantics (first_treat period
+   for staggered; 0/1 indicator for 2x2) with a runnable example.
+"""
+
+from __future__ import annotations
+
+import contextlib
+import io
+import textwrap
+from types import SimpleNamespace
+from unittest.mock import patch
+
+import matplotlib
+matplotlib.use("Agg")
+
+import numpy as np
+import pandas as pd
+import pytest
+
+
+# ---------------------------------------------------------------------------
+# Bug 1 — use_chinese() on Linux with JP/TC/Zen Hei / Source Han fonts
+# ---------------------------------------------------------------------------
+
+
+def _fake_fontmanager(names):
+    """Return an object mimicking matplotlib's fontManager with given names."""
+    return SimpleNamespace(ttflist=[SimpleNamespace(name=n) for n in names])
+
+
+def test_use_chinese_picks_noto_jp_on_linux():
+    """Noto Sans CJK JP covers CJK; auto-detect must pick it up on Linux.
+
+    Also pins the priority order: sans is preferred over serif in 'auto' mode.
+    """
+    from statspai.plots import themes
+    import matplotlib.font_manager as _fm
+
+    fake = _fake_fontmanager([
+        "DejaVu Sans", "Liberation Serif",
+        "Noto Sans CJK JP", "Noto Serif CJK JP",
+        "WenQuanYi Zen Hei",
+    ])
+    with patch.object(_fm, "fontManager", fake):
+        chosen = themes.use_chinese()
+    # Sans-preferred in auto mode, and Noto Sans CJK JP is listed before
+    # WenQuanYi Zen Hei in sans_priority — so we pin Noto wins.
+    assert chosen == "Noto Sans CJK JP", (
+        f"auto mode should prefer Noto Sans CJK JP (sans, higher priority) "
+        f"when both are available; got {chosen!r}"
+    )
+
+
+def test_use_chinese_picks_wenquanyi_zen_hei():
+    """WenQuanYi Zen Hei is the classic Linux free Chinese font."""
+    from statspai.plots import themes
+    import matplotlib.font_manager as _fm
+
+    fake = _fake_fontmanager(["DejaVu Sans", "WenQuanYi Zen Hei"])
+    with patch.object(_fm, "fontManager", fake):
+        chosen = themes.use_chinese()
+    assert chosen == "WenQuanYi Zen Hei"
+
+
+def test_use_chinese_picks_source_han_sans():
+    """Source Han Sans (Adobe/Google) is Linux/cloud-common."""
+    from statspai.plots import themes
+    import matplotlib.font_manager as _fm
+
+    fake = _fake_fontmanager(["DejaVu Sans", "Source Han Sans"])
+    with patch.object(_fm, "fontManager", fake):
+        chosen = themes.use_chinese()
+    assert chosen == "Source Han Sans"
+
+
+def test_use_chinese_warns_when_no_chinese_font():
+    """Still warns + returns '' when no CJK font is present."""
+    from statspai.plots import themes
+    import matplotlib.font_manager as _fm
+
[…]
