### 2026-04-20 d03a34d
feat(hdfe+inference): native HDFE absorber + multiway cluster + subcluster wild bootstrap
Foundation layer to compete with reghdfe/fixest on the causal-inference
workflow that applied econ lives in:

panel/hdfe.py — native alternating-projection absorber
- Irons-Tuck / vector-Aitken Δ² acceleration
- Iterative singleton pruning (Correia 2015)
- 1 / 2 / N-way FE support; weighted means
- Reusable Absorber class for event-study coefficient paths

panel/feols.py — reghdfe-style one-call API
- Formula `"y ~ x | fe1 + fe2 + fe3"` parsing
- Classical / cluster / multiway-cluster / wild-cluster SE
- 
[…]
