+++ tests/test_hdfe_native.py
@@ -1,0 +1,158 @@
+"""Native HDFE absorber + hdfe_ols tests.
+
+Verifies numerical agreement with dummy-variable OLS on small samples.
+"""
+
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+from statspai.panel.hdfe import Absorber, absorb_ols, demean
+
+
+RNG = np.random.default_rng(42)
+
+
+# ---------------------------------------------------------------------------
+# Data generator
+# ---------------------------------------------------------------------------
+
+
+def _make_panel(n_firms=20, n_years=8, beta=(1.0, -0.5), seed=0):
+    rng = np.random.default_rng(seed)
+    firm = np.repeat(np.arange(n_firms), n_years)
+    year = np.tile(np.arange(n_years), n_firms)
+    n = n_firms * n_years
+    x1 = rng.standard_normal(n)
+    x2 = rng.standard_normal(n)
+    alpha_firm = rng.standard_normal(n_firms)[firm]
+    alpha_year = rng.standard_normal(n_years)[year]
+    y = alpha_firm + alpha_year + beta[0] * x1 + beta[1] * x2 + rng.standard_normal(n) * 0.5
+    return pd.DataFrame({"firm": firm, "year": year, "x1": x1, "x2": x2, "y": y})
+
+
+# ---------------------------------------------------------------------------
+# Demean correctness
+# ---------------------------------------------------------------------------
+
+
+def test_demean_one_way_matches_dummy_regression():
+    df = _make_panel(n_firms=10, n_years=6)
+    ab = Absorber(df[["firm"]].values, drop_singletons=False)
+    yw = ab.demean(df["y"].values)
+
+    # Compare to manual group-demean
+    expected = df["y"].values - df.groupby("firm")["y"].transform("mean").values
+    np.testing.assert_allclose(yw, expected, atol=1e-10)
+
+
+def test_demean_two_way_converges():
+    df = _make_panel(n_firms=20, n_years=8)
+    ab = Absorber(df[["firm", "year"]].values, drop_singletons=False, tol=1e-10)
+    yw = ab.demean(df["y"].values)
+    # After two-way demeaning, any further group mean should be near-zero
+    resid_firm_mean = pd.Series(yw).groupby(df["firm"].values).mean().abs().max()
+    resid_year_mean = pd.Series(yw).groupby(df["year"].values).mean().abs().max()
+    assert resid_firm_mean < 1e-6
+    assert resid_year_mean < 1e-6
+
+
+def test_absorb_ols_matches_dummy_ols_two_way():
+    df = _make_panel(n_firms=12, n_years=10, beta=(2.0, -1.0), seed=7)
+    # Native HDFE
+    res_native = absorb_ols(
+        y=df["y"].values,
+        X=df[["x1", "x2"]].values,
+        fe=df[["firm", "year"]].values,
+        drop_singletons=False,
+    )
+    # Dummy-variable OLS for comparison
+    firm_dummies = pd.get_dummies(df["firm"], prefix="firm", drop_first=True, dtype=float)
+    year_dummies = pd.get_dummies(df["year"], prefix="year", drop_first=True, dtype=float)
+    X_full = np.column_stack([
+        np.ones(len(df)),
+        df[["x1", "x2"]].values,
+        firm_dummies.values,
+        year_dummies.values,
+    ])
+    beta_full, *_ = np.linalg.lstsq(X_full, df["y"].values, rcond=None)
+    # coefficients on x1, x2 are at positions 1 and 2
[…]
