+++ src/statspai/panel/feols.py
@@ -1,0 +1,481 @@
+"""
+Unified high-dimensional fixed-effects OLS estimator — ``sp.feols()``.
+
+Mirrors R's ``fixest::feols`` and Stata's ``reghdfe`` top-level API:
+
+    sp.feols("y ~ x1 + x2 | firm + year", data=df, cluster="firm")
+
+Formula grammar
+---------------
+The pipe ``|`` separates regressors from absorbed fixed effects.
+
+    "y ~ x1 + x2"                   → no FE, pure OLS (constant included)
+    "y ~ x1 | firm"                 → firm FE absorbed
+    "y ~ x1 + x2 | firm + year"     → two-way FE (firm, year)
+    "y ~ x1 | firm + year + state"  → three-way FE
+
+Variables on both sides are bare column names (no Patsy syntax).
+
+Inference
+---------
+- Default: classical OLS SE (appropriate only when errors i.i.d.).
+- ``cluster='firm'``: one-way cluster-robust (Liang-Zeger CR1).
+- ``cluster=['firm', 'year']``: N-way CGM via inclusion-exclusion.
+- ``cluster='firm', wild=True``: wild cluster bootstrap on top of CR1.
+
+Returns a :class:`FEOLSResult` with coef / se / vcov / R²-within, plus
+reference to the reusable ``Absorber`` (enables re-running on subsamples
+or for event-study path estimation).
+"""
+
+from __future__ import annotations
+
+import re
+from dataclasses import dataclass
+from typing import Any, Dict, List, Optional, Union
+
+import numpy as np
+import pandas as pd
+from scipy import stats
+
+from .hdfe import Absorber, absorb_ols
+
+
+@dataclass
+class FEOLSResult:
+    """Result of ``sp.feols()``.
+
+    Attributes
+    ----------
+    params : pd.Series
+        Coefficient estimates indexed by regressor name.
+    std_errors : pd.Series
+        Standard errors indexed by regressor name.
+    vcov : np.ndarray
+        Variance-covariance matrix of the coefficients.
+    tvalues, pvalues : pd.Series
+    conf_int_lower, conf_int_upper : pd.Series
+    residuals : np.ndarray
+        In-sample residuals (after FE absorption).
+    fitted_within : np.ndarray
+        Predicted values from X β (excludes FE contribution).
+    n_obs : int
+    n_singletons_dropped : int
+    n_fe : List[int]
+        Number of groups per absorbed FE dimension.
+    dof_fe : int
+        Degrees of freedom consumed by the FEs.
+    df_resid : int
+    r2_within : float
+    se_type : str
+        'iid' | 'cluster' | 'multiway_cluster' | 'wild_cluster'
+    cluster_info : dict
+        Metadata (cluster names, counts).
+    formula : str
+    absorber : Absorber
+        Reusable absorber (includes ``keep_mask`` to subset rows).
+    """
+
+    params: pd.Series
+    std_errors: pd.Series
+    vcov: np.ndarray
+    tvalues: pd.Series
+    pvalues: pd.Series
+    conf_int_lower: pd.Series
+    conf_int_upper: pd.Series
+    residuals: np.ndarray
+    fitted_within: np.ndarray
+    n_obs: int
+    n_singletons_dropped: int
+    n_fe: List[int]
+    dof_fe: int
+    df_resid: int
+    r2_within: float
+    se_type: str
+    cluster_info: Dict[str, Any]
+    formula: str
+    absorber: Absorber
+    converged: bool
+    iters: int
+
+    def summary(self) -> str:
+        lines: List[str] = []
+        lines.append(f"FEOLS (reghdfe-style)  |  {self.formula}")
+        lines.append("=" * max(60, len(self.formula) + 25))
[…]
