### 2026-04-20 dbfa537
feat(matching+causal): overlap weights, CBPS, GRF-style forest inference
Close the gap vs R's PSweight / CBPS / grf packages:

matching/overlap_weights.py — Li-Morgan-Zaslavsky 2018
- overlap_weights() with ATO/ATE/ATT/ATC/matching/entropy tilts
- Bounded weights so small pscores no longer blow up variance
- Effective-sample-size diagnostic

matching/cbps.py — Imai-Ratkovic 2014
- cbps() exact (balance-only) and over-identified GMM variants
- ATE / ATT targets
- Post-weighting standardised-mean-difference balance table

causal/forest_inference.py — GRF parity layer
- calibration_test() = CDDF 
[…]
