+++ src/statspai/matching/overlap_weights.py
@@ -1,0 +1,174 @@
+"""
+Overlap weights (Li, Morgan, Zaslavsky 2018).
+
+Overlap weights target the "average treatment effect among the overlap
+population" (ATO) — those observations with genuine equipoise. The
+weight function is
+
+    w_i = 1 - e(X_i)    for treated  i   (T_i=1)
+    w_i =     e(X_i)    for control  i   (T_i=0)
+
+which is proportional to the "tilting" that minimises the variance of
+the resulting weighted treatment-effect estimator subject to exact
+covariate balance on the moments used to fit ``e(·)`` (Li et al. 2018,
+Theorem 1 & 3). Overlap weights:
+
+- are bounded in [0, 1] — no extreme weights from small propensity
+  scores, so results are stable even with poor overlap;
+- exactly balance the log-odds covariates when ``e(X)`` is a logit fit;
+- target ATO, not ATE, and should be interpreted accordingly.
+
+References
+----------
+Li, F., Morgan, K.L., Zaslavsky, A.M. (2018). "Balancing Covariates via
+Propensity Score Weighting." JASA, 113(521), 390-400.
+
+Li, F., Thomas, L.E., Li, F. (2019). "Addressing Extreme Propensity
+Scores via the Overlap Weights." American Journal of Epidemiology,
+188(1), 250-257.
+"""
+
+from __future__ import annotations
+
+from typing import List, Optional
+
+import numpy as np
+import pandas as pd
+from scipy import stats as sp_stats
+
+from ..core.results import CausalResult
+
+
+def overlap_weights(
+    data: pd.DataFrame,
+    y: str,
+    treat: str,
+    covariates: List[str],
+    estimand: str = "ATO",
+    n_bootstrap: int = 500,
+    alpha: float = 0.05,
+    seed: Optional[int] = None,
+    trim: float = 0.0,
+) -> CausalResult:
+    """Overlap-weight (ATO) treatment effect estimator.
+
+    Parameters
+    ----------
+    data : DataFrame
+    y : str
+        Outcome column.
+    treat : str
+        Binary 0/1 treatment column.
+    covariates : list of str
+        Covariates for the logistic propensity-score model.
+    estimand : {'ATO', 'ATE', 'ATT', 'ATC', 'matching', 'entropy'}
+        Which generalized-weight scheme to use. All follow Li-Li-Li
+        (2019) Table 1; 'ATO' uses the overlap weights; 'matching'
+        uses the ``min(e, 1-e)`` weight; 'entropy' uses
+        ``-e·log(e) - (1-e)·log(1-e)``; 'ATE/ATT/ATC' reduce to
+        standard IPW for comparison.
+    n_bootstrap : int
+        Paired-sample bootstrap replications for SE.
+    alpha : float
+    seed : int, optional
+    trim : float
+        Optional clip of pscore to ``[trim, 1-trim]``. For overlap weights
+        this is rarely needed — set to 0 by default.
+
+    Returns
+    -------
+    CausalResult
+        ``.estimate`` targets the named estimand; ``.model_info`` stores
+        the weight summary, effective sample size, and pscore diagnostics.
+    """
+    estimand = estimand.upper()
+    valid = {"ATO", "ATE", "ATT", "ATC", "MATCHING", "ENTROPY"}
+    if estimand not in valid:
+        raise ValueError(f"estimand must be one of {valid}, got {estimand!r}")
+
[…]
