### 2026-04-20 9298095
feat(api): expose HDFE + matching + forest primitives at sp.* top level
All new surfaces in this session are now reachable via the canonical
sp.X / import statspai as sp convention that the docs and examples use:

sp.Absorber / sp.demean / sp.absorb_ols / sp.hdfe_ols / sp.FEOLSResult
sp.overlap_weights / sp.cbps
sp.calibration_test / sp.test_calibration (grf-alias) / sp.rate / sp.honest_variance
sp.subcluster_wild_bootstrap / sp.wild_cluster_ci_inv
sp.multiway_cluster_vcov / sp.cluster_robust_se / sp.cr3_jackknife_vcov

sp.feols remains the pyfixest-backed wrapper (unchanged); the new
native i
[…]
