### 2026-04-20 8aca56e
perf(hdfe): Numba-JIT the alternating-projection kernel (3x speed-up)
The hot loop in the HDFE absorber is a per-column triple pass:

    sums[g] += col[i]      # group accumulate
    sums[g] /= counts[g]   # group divide
    col[i] -= sums[codes[i]]  # subtract group mean

NumPy's np.bincount does pass 1 in C but then allocates two temporary
arrays for passes 2–3. Numba fuses all three into a single, cache-
friendly inner loop with SIMD, roughly 3x faster on a 3-way FE panel.

New file: panel/_hdfe_kernels.py (two @njit kernels with pragma-guarded
import so Numba is optional). panel/hdfe.py d
[…]
