### 2026-04-20 1721767
fix(cbps+rate): robustness + influence-function variance
CBPS (matching/cbps.py)
- Bootstrap now uses draw-until-success: degenerate resamples (no
  treated or no controls) are discarded, the optimiser is retried,
  and exactly n_bootstrap successes are returned (with a 10x retry
  cap to guard against pathological DGPs).
- Warm start switched from sklearn LogisticRegression (L2-regularised
  LBFGS) to statsmodels Logit (Newton-Raphson + exact Hessian). Falls
  back to sklearn, then to zeros. Closer to R `glm`; more stable on
  near-separable data.
- _fit_cbps now returns convergence flag + GM
[…]
