+++ src/statspai/causal/forest_inference.py
@@ -220,43 +220,63 @@ def rate(
 ) -> Dict[str, float]:
     """Rank-Average Treatment Effect (Yadlowsky et al. 2023).
 
-    Let ``S(x) = -τ̂(x)`` denote a prioritization score (here, the
-    negative predicted CATE — so "high priority" = large positive
-    benefit). For a quantile ``q ∈ (0, 1]``, define the TOC (targeting
-    operator characteristic) curve:
+    Let ``S(x) = τ̂(x)`` denote a prioritisation score (higher = higher
+    priority). Define the TOC (targeting operator characteristic) curve:
 
-        TOC(q) = E[τ(X) | S(X) ≤ Q_q(S)] - E[τ(X)]
+        TOC(q) = E[τ(X) | S(X) ≥ Q_{1-q}(S)] - E[τ(X)]
 
-    **AUTOC**: area under the TOC curve, ``∫₀¹ TOC(q) dq``. Large
-    positive values imply that prioritising by τ̂ yields larger average
-    benefit than random assignment.
+    i.e. the expected CATE among the top-``q`` fraction minus the
+    population ATE. Two scalar summaries are supported:
 
-    Inference uses a half-sample split: fit the forest on split A,
-    evaluate AUTOC on split B's AIPW pseudo-outcome, and vice versa.
-    The variance of the averaged estimate is computed from the two
-    half-sample estimates. This is the "DR-learner RATE" estimator
-    recommended by Yadlowsky et al.
+    - **AUTOC**: ``∫₀¹ TOC(q) dq`` — the *unweighted* area under the
+      TOC curve. Emphasises prioritisation performance uniformly
+      across the quantile range.
+    - **QINI**:  ``∫₀¹ q · TOC(q) dq`` — down-weights narrow top
+      fractions; closer to the classical uplift / Qini coefficient.
+
+    Estimation
+    ----------
+    The DR-RATE estimator from Yadlowsky et al. uses an AIPW pseudo-
+    outcome ``Ψ_i`` (computed from the forest's own cross-fitted
+    nuisance predictions) and reduces AUTOC / Qini to a weighted sum:
+
+        AUTOC_hat = (1/n) Σ_i Ψ_i · w_{AUTOC}(R_i / n) - Ψ̄
+        QINI_hat  = (1/n) Σ_i Ψ_i · w_{QINI}(R_i / n)  - (1/2) Ψ̄
+
+    where ``R_i`` is the descending rank of ``S(X_i)`` and the weights
+    are closed-form rank kernels. This representation makes the
+    estimator a sample mean of per-observation contributions φ_i, so
+    the variance admits the standard influence-function form
+
+        Var(AUTOC_hat) = (1/(n(n-1))) Σ_i (φ_i - φ̄)²
+
+    which replaces the conservative half-sample estimator used in the
+    earlier draft of this function.
 
     Parameters
     ----------
     forest : fitted CausalForest
-        Used only for the ranking τ̂(X). (In a strict version you would
-        refit on each half; here we treat the forest as fixed and rely
-        on the forest's own honest split for approximate orthogonality.)
-    X, Y, T : arrays
+    X, Y, T : arrays, optional
+        If omitted, falls back to the forest's stored training arrays.
     target : {'AUTOC', 'QINI'}
-        ``AUTOC`` integrates unweighted; ``QINI`` uses the Qini
-        weighting ``q·TOC(q)``. Both are reported as a single scalar.
     q_grid : int
-        Evaluation points for the TOC curve.
[…]
