+++ src/statspai/matching/cbps.py
@@ -107,26 +107,46 @@ def cbps(
     def _solve(X_, T_):
         return _fit_cbps(X_, T_, estimand=estimand, variant=variant)
 
-    beta_hat, ps = _solve(X, T)
+    beta_hat, ps, converged_pt, obj_pt = _solve(X, T)
     if trim > 0:
         ps = np.clip(ps, trim, 1 - trim)
     w1, w0 = _cbps_weights(T, ps, estimand)
     est = float(np.sum(w1 * Y) - np.sum(w0 * Y))
 
-    # Bootstrap
+    # Bootstrap with draw-until-success. Pathological resamples (all
+    # treated or all controls, optimizer failure, singular Hessian) are
+    # discarded and re-drawn so that we return exactly ``n_bootstrap``
+    # successful reps. A hard ceiling on retries guards against
+    # degenerate DGPs.
     boot = np.empty(n_bootstrap)
-    for b in range(n_bootstrap):
+    boot_converged = np.empty(n_bootstrap, dtype=bool)
+    max_retries = 10 * n_bootstrap
+    retries = 0
+    b = 0
+    while b < n_bootstrap and retries < max_retries:
         idx = rng.integers(0, n, size=n)
-        X_b, T_b, Y_b = X[idx], T[idx], Y[idx]
+        T_b = T[idx]
+        # Skip degenerate resamples where the treatment or control arm
+        # disappeared — CBPS has no solution in that subsample.
+        if T_b.sum() < 2 or (1 - T_b).sum() < 2:
+            retries += 1
+            continue
+        X_b, Y_b = X[idx], Y[idx]
         try:
-            _, ps_b = _solve(X_b, T_b)
+            _, ps_b, conv_b, _ = _solve(X_b, T_b)
             if trim > 0:
                 ps_b = np.clip(ps_b, trim, 1 - trim)
             w1b, w0b = _cbps_weights(T_b, ps_b, estimand)
             boot[b] = float(np.sum(w1b * Y_b) - np.sum(w0b * Y_b))
+            boot_converged[b] = conv_b
+            b += 1
         except Exception:
-            boot[b] = np.nan
-    boot = boot[~np.isnan(boot)]
+            retries += 1
+            continue
+    boot = boot[:b]
+    boot_converged = boot_converged[:b]
+    n_boot_success = b
+    n_boot_nonconv = int((~boot_converged).sum())
     se = float(np.std(boot, ddof=1)) if boot.size > 1 else np.nan
     z = sp_stats.norm.ppf(1 - alpha / 2)
     ci = (est - z * se, est + z * se) if np.isfinite(se) else (np.nan, np.nan)
@@ -151,7 +171,12 @@ def cbps(
             zip(["_intercept"] + list(covariates) if add_intercept else list(covariates),
                 smd.tolist())
         ),
+        "converged": converged_pt,
+        "gmm_objective": obj_pt,
         "n_bootstrap": n_bootstrap,
+        "n_bootstrap_success": n_boot_success,
+        "n_bootstrap_nonconverged": n_boot_nonconv,
+        "n_bootstrap_retries": retries,
     }
 
     return CausalResult(
@@ -215,36 +240,48 @@ def _fit_cbps(
     T: np.ndarray,
     estimand: str,
     variant: str,
-) -> tuple[np.ndarray, np.ndarray]:
-    """Fit CBPS by minimising ||g(β)||² (quadratic in moments).
+) -> tuple[np.ndarray, np.ndarray, bool, float]:
+    """Fit CBPS by minimising ``||g(β)||²`` (quadratic in moments).
 
     For ``variant == 'exact'``, solves the K balance equations directly.
     For ``variant == 'over'``, two-step GMM: step 1 uses identity
-    weighting, step 2 uses inverse of empirical second-moment matrix.
[…]
