### 2026-04-20 394f25c
test(parity): reference parity suite — 36 recovery + cross-estimator tests
Adds tests/reference_parity/ with design-level identification-validity
guards that sit above the existing pinned-fixture drift guards.

Coverage:
- DID: 2x2 + staggered homogeneous + heterogeneous DGPs; classic/CS/SA/
  Wooldridge all recover truth within 4 sigma; Bacon decomposition runs.
- RD: sharp/fuzzy recovery with known population jump; kernel & bandwidth
  choice agreement within combined SE; zero-effect CI coverage.
- IV: 2SLS recovers LATE; Wald ratio equals 2SLS; first-stage F > 10
  on strong-IV DGP; weak-IV
[…]
