+++ tests/reference_parity/test_rd_parity.py
@@ -1,0 +1,119 @@
+"""RD reference parity tests.
+
+Recovery + cross-method consistency for regression discontinuity.
+
+Validates:
+1. Sharp RD (CCT 2014 rdrobust) recovers the true jump at the cutoff.
+2. Fuzzy RD recovers the LATE via the Wald ratio.
+3. Different bandwidth selectors (MSE, CER) give similar point estimates.
+4. Kernel choice (triangular, uniform, epanechnikov) agrees within noise.
+
+All DGPs are deterministic (seeded) with known population discontinuities.
+"""
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+
+from statspai.rd import rdrobust
+
+
+def _within_n_se(estimate, truth, se, n_sigma=4.0):
+    return abs(estimate - truth) <= n_sigma * se
+
+
+# ---------------------------------------------------------------------------
+# Sharp RD
+# ---------------------------------------------------------------------------
+
+class TestSharpRDRecovery:
+    """Sharp RD on a DGP with known jump = 1.0 must recover truth."""
+
+    def test_triangular_mserd_recovers_truth(self, rd_sharp_data):
+        truth = rd_sharp_data.attrs['true_effect']
+        r = rdrobust(rd_sharp_data, y='y', x='x', c=0.0,
+                     kernel='triangular', bwselect='mserd')
+        assert _within_n_se(r.estimate, truth, r.se), (
+            f"Sharp RD triangular/mserd: {r.estimate:.4f} vs {truth} "
+            f"(SE {r.se:.4f})"
+        )
+
+    def test_uniform_kernel_close_to_triangular(self, rd_sharp_data):
+        """Point estimates under different kernels must agree within combined SE."""
+        r_t = rdrobust(rd_sharp_data, y='y', x='x', c=0.0,
+                       kernel='triangular', bwselect='mserd')
+        r_u = rdrobust(rd_sharp_data, y='y', x='x', c=0.0,
+                       kernel='uniform', bwselect='mserd')
+        combined_se = np.sqrt(r_t.se**2 + r_u.se**2)
+        assert abs(r_t.estimate - r_u.estimate) <= 4.0 * combined_se, (
+            f"triangular {r_t.estimate:.4f} vs uniform {r_u.estimate:.4f} "
+            f"(combined SE {combined_se:.4f})"
+        )
+
+    def test_epanechnikov_close_to_triangular(self, rd_sharp_data):
+        r_t = rdrobust(rd_sharp_data, y='y', x='x', c=0.0,
+                       kernel='triangular', bwselect='mserd')
+        r_e = rdrobust(rd_sharp_data, y='y', x='x', c=0.0,
+                       kernel='epanechnikov', bwselect='mserd')
+        combined_se = np.sqrt(r_t.se**2 + r_e.se**2)
+        assert abs(r_t.estimate - r_e.estimate) <= 4.0 * combined_se
+
+    def test_positive_effect_is_positive(self, rd_sharp_data):
+        r = rdrobust(rd_sharp_data, y='y', x='x', c=0.0)
+        assert r.estimate > 0, f"Expected positive, got {r.estimate}"
+
+    def test_ci_covers_truth(self, rd_sharp_data):
+        truth = rd_sharp_data.attrs['true_effect']
+        r = rdrobust(rd_sharp_data, y='y', x='x', c=0.0)
+        lo, hi = r.ci
+        assert lo <= truth <= hi, (
+            f"95% CI [{lo:.4f}, {hi:.4f}] does not cover truth {truth}"
+        )
+
+
+# ---------------------------------------------------------------------------
+# Fuzzy RD
[…]
