### 2026-04-20 e41408e
feat(core): broom-style tidy() + glance() on EconometricResults + CausalResult
Adds unified long-format coefficient tables and 1-row model-level
summaries, mirroring R's broom package.  Makes every estimator's
results programmatically comparable and feeds directly into
modelsummary / regtable pipelines.

EconometricResults.tidy() / glance():
- tidy: term, estimate, std_error, statistic, p_value, conf_low, conf_high
- glance: method, nobs, r_squared, adj_r_squared, f_statistic, aic, bic,
  log_likelihood, df_resid, df_model (present subset depends on model)
- tidy(conf_level=0.90) recomputes CI
[…]
