+++ src/statspai/core/results.py
@@ -134,6 +134,99 @@ class EconometricResults:
             f'{1-alpha/2:.3f}': upper
         }, index=self.params.index)
     
+    # ------------------------------------------------------------------
+    # Broom-style tidy interface (EconometricResults)
+    # ------------------------------------------------------------------
+
+    def tidy(self, conf_level: float = 0.95) -> pd.DataFrame:
+        """Return a long-format DataFrame of coefficients, broom-style.
+
+        Columns
+        -------
+        term : str
+            Variable / coefficient name.
+        estimate : float
+        std_error : float
+        statistic : float
+            t-statistic.
+        p_value : float
+        conf_low, conf_high : float
+            Two-sided conf_level CI.
+
+        Examples
+        --------
+        >>> result = sp.regress("y ~ x1 + x2", data=df)
+        >>> result.tidy()
+           term  estimate  std_error  statistic  p_value  conf_low  conf_high
+        0  Intercept     ...
+
+        See Also
+        --------
+        glance : 1-row model-level summary (R^2, F, N, AIC, BIC).
+        """
+        alpha = 1 - conf_level
+        df_resid = self.data_info.get('df_resid', np.inf)
+        t_crit = stats.t.ppf(1 - alpha/2, df_resid)
+        lo = self.params - t_crit * self.std_errors
+        hi = self.params + t_crit * self.std_errors
+
+        def _arr(x):
+            return x.values if hasattr(x, 'values') else np.asarray(x)
+
+        return pd.DataFrame({
+            'term': list(self.params.index),
+            'estimate': _arr(self.params),
+            'std_error': _arr(self.std_errors),
+            'statistic': _arr(self.tvalues),
+            'p_value': _arr(self.pvalues),
+            'conf_low': _arr(lo),
+            'conf_high': _arr(hi),
+        })
+
+    def glance(self) -> pd.DataFrame:
+        """Return a 1-row DataFrame of model-level statistics, broom-style.
+
+        Columns (present subset depends on the model type)
+        --------------------------------------------------
+        nobs : int
+        r_squared : float
+        adj_r_squared : float
+        f_statistic : float
+        f_p_value : float
+        aic, bic : float
+        df_resid, df_model : int
+        method : str
+            Estimation method label.
+
+        See Also
+        --------
+        tidy : long-format coefficient table.
+        """
+        g: Dict[str, Any] = {}
+        g['method'] = self.model_info.get('method', self.model_info.get('model_type', ''))
+        if 'nobs' in self.data_info:
+            g['nobs'] = int(self.data_info['nobs'])
+        key_map = {
+            'R-squared': 'r_squared',
+            'Adj R-squared': 'adj_r_squared',
+            'F-statistic': 'f_statistic',
+            'F p-value': 'f_p_value',
+            'Prob (F-statistic)': 'f_p_value',
+            'AIC': 'aic',
+            'BIC': 'bic',
+            'Log-Likelihood': 'log_likelihood',
+        }
+        for src, dst in key_map.items():
+            if src in self.diagnostics and dst not in g:
+                v = self.diagnostics[src]
+                if isinstance(v, (int, float, np.integer, np.floating)):
+                    g[dst] = float(v)
+        if 'df_resid' in self.data_info:
+            g['df_resid'] = int(self.data_info['df_resid'])
+        if 'df_model' in self.data_info:
+            g['df_model'] = int(self.data_info['df_model'])
+        return pd.DataFrame([g])
+
     def predict(self, data: Optional[pd.DataFrame] = None) -> np.ndarray:
         """
         Generate predictions from the fitted model.
@@ -739,6 +832,128 @@ class CausalResult:
         lines.append("  * p<0.1, ** p<0.05, *** p<0.01")
 
[…]
