+++ tests/test_tidy_glance.py
@@ -1,0 +1,261 @@
+"""Tests for broom-style ``tidy()`` and ``glance()`` methods.
+
+Validates that every major result class exposes the unified broom-style
+interface:
+
+- ``tidy()`` -> long-format DataFrame with required columns
+- ``glance()`` -> 1-row DataFrame with model-level stats
+
+Required columns on ``tidy()``:
+    term, estimate, std_error, statistic, p_value, conf_low, conf_high
+
+Required columns on ``glance()``:
+    method, nobs
+"""
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+
+
+TIDY_REQUIRED = {'term', 'estimate', 'std_error', 'statistic',
+                 'p_value', 'conf_low', 'conf_high'}
+
+GLANCE_REQUIRED = {'method', 'nobs'}
+
+
+# ---------------------------------------------------------------------------
+# OLS
+# ---------------------------------------------------------------------------
+
+@pytest.fixture
+def ols_result():
+    rng = np.random.default_rng(1)
+    df = pd.DataFrame({
+        'x1': rng.normal(size=300),
+        'x2': rng.normal(size=300),
+    })
+    df['y'] = 1.5 + 2 * df['x1'] - 0.5 * df['x2'] + rng.normal(size=300)
+    return sp.regress('y ~ x1 + x2', data=df)
+
+
+class TestOLSTidyGlance:
+    def test_tidy_returns_dataframe(self, ols_result):
+        t = ols_result.tidy()
+        assert isinstance(t, pd.DataFrame)
+
+    def test_tidy_has_required_columns(self, ols_result):
+        t = ols_result.tidy()
+        missing = TIDY_REQUIRED - set(t.columns)
+        assert not missing, f"tidy() missing columns: {missing}"
+
+    def test_tidy_correct_number_of_rows(self, ols_result):
+        t = ols_result.tidy()
+        assert len(t) == 3  # Intercept, x1, x2
+
+    def test_tidy_estimates_match_params(self, ols_result):
+        t = ols_result.tidy().set_index('term')
+        for name in ['Intercept', 'x1', 'x2']:
+            assert t.loc[name, 'estimate'] == pytest.approx(
+                ols_result.params[name]
+            )
+
+    def test_tidy_conf_level_affects_interval(self, ols_result):
+        t95 = ols_result.tidy(conf_level=0.95).set_index('term')
+        t90 = ols_result.tidy(conf_level=0.90).set_index('term')
+        # 90% CI must be strictly narrower than 95%
+        w95 = (t95['conf_high'] - t95['conf_low'])
+        w90 = (t90['conf_high'] - t90['conf_low'])
+        assert (w90 < w95).all(), f"90% CI wider than 95%: {w90} vs {w95}"
+
+    def test_glance_returns_one_row(self, ols_result):
+        g = ols_result.glance()
+        assert isinstance(g, pd.DataFrame)
+        assert len(g) == 1
+
+    def test_glance_has_required_columns(self, ols_result):
+        g = ols_result.glance()
+        missing = GLANCE_REQUIRED - set(g.columns)
+        assert not missing, f"glance() missing: {missing}"
+
+    def test_glance_nobs_matches_data(self, ols_result):
+        g = ols_result.glance()
+        assert g['nobs'].iloc[0] == 300
+
+    def test_glance_has_r_squared(self, ols_result):
+        g = ols_result.glance()
+        assert 'r_squared' in g.columns
+        assert 0 < g['r_squared'].iloc[0] <= 1
+
+
+# ---------------------------------------------------------------------------
[…]
