### 2026-04-20 3325d23
docs(guides): 5 decision-tree vignettes for choosing causal methods
Addresses the "18 DID variants, which one?" problem that makes users
bounce off the library.  Each guide is a flowchart: read the first
question, jump to the section it sends you to, stop at a recommendation.

- choosing_did_estimator.md: 2x2 vs staggered, homogeneous vs
  heterogeneous, Bacon decomposition gate, CS/SA/Wooldridge/imputation/
  stacked/continuous decision; default recommendation when in doubt.
- choosing_rd_estimator.md: sharp/fuzzy split, kernel/bandwidth/
  polynomial order defaults, mandatory diagnostics che
[…]
