+++ docs/guides/choosing_matching_estimator.md
@@ -1,0 +1,164 @@
+# Choosing a matching / weighting estimator
+
+When your design relies on selection-on-observables (CIA / unconfoundedness)
+and you have a binary treatment, you have 7+ estimators in StatsPAI.
+Here's how to choose.
+
+## 0. TL;DR flowchart
+
+```
+Is your covariate set high-dimensional (p > 20)?
+  YES -> Double ML (sp.dml), meta-learners (sp.S_Learner, etc.)
+  NO  -> continue
+
+Is your target the ATT (effect on the treated)?
+  YES -> sp.ebalance (entropy balancing) or sp.match(estimand='ATT')
+  NO  -> continue
+
+Is your target the ATE (population average)?
+  YES -> sp.cbps(estimand='ATE') or sp.aipw
+  NO  -> continue
+
+Do you need OVERLAP-weighted effect (avoiding extrapolation)?
+  YES -> sp.overlap_weights (ATO)
+  NO  -> rethink — what estimand do you actually want?
+```
+
+## 1. Entropy balancing (ebal) — the "just works" default for ATT
+
+Hainmueller (2012). Exact covariate balance by reweighting, no
+propensity-score modelling needed.
+
+```python
+r = sp.ebalance(df, y='y', treat='d',
+                covariates=['X1', 'X2', 'X3'],
+                moments=1)  # balance means; moments=2 adds variances
+```
+
+**Pros:** no PSM model specification; exact balance by construction;
+no King-Nielsen issue.
+**Cons:** targets ATT only; can be sensitive to extreme weights.
+
+## 2. Nearest-neighbor matching
+
+Beware: King & Nielsen (2019) show that PSM-based nearest-neighbor
+matching can **increase** imbalance. Prefer Mahalanobis or coarsened
+exact matching (CEM):
+
+```python
+r = sp.match(df, y='y', treat='d', covariates=[...],
+             distance='mahalanobis',  # NOT 'propensity'
+             method='nearest', n_matches=3)
+```
+
+## 3. Covariate Balancing Propensity Score (CBPS)
+
+Imai-Ratkovic (2014). Fits the propensity score to balance covariates
+directly, not to maximise likelihood.
+
+```python
+r = sp.cbps(df, y='y', treat='d', covariates=[...],
+            estimand='ATE',  # or 'ATT'
+            variant='over')   # 'over' (overidentified) is preferred
+```
+
+More robust to PS misspecification than IPW.
+
+## 4. Overlap weights (ATO)
+
+Li-Morgan-Zaslavsky (2018). Weights each unit by its propensity of
+receiving the "other" treatment, yielding effects on the **overlap
+population** — the subpopulation where both treatments are plausible.
+
+```python
+r = sp.overlap_weights(df, y='y', treat='d', covariates=[...],
+                       estimand='ATO')
+```
+
+Avoids extreme weights from near-zero / near-one propensities.
+
+## 5. Doubly-robust estimators
+
+AIPW combines an outcome model and a propensity-score model — correct
+if **either** is right.
+
+```python
+r = sp.aipw(df, y='y', treat='d', covariates=[...])
+```
+
+For high-dimensional covariates, use Double ML (Chernozhukov et al. 2018):
+
+```python
+r = sp.dml(df, y='y', treat='d', covariates=[...],
+           ml_model='lasso',         # or 'rf', 'xgb'
+           cross_fitting_folds=5)
+```
+
+DML is the state of the art for observational ATE with many controls.
+
+## 6. Meta-learners (for heterogeneous effects)
+
+If you want not just the ATE but a CATE function τ(X):
+
+```python
[…]
