+++ mkdocs.yml
@@ -56,6 +56,11 @@ nav:
       - "Report card: cs_report()": guides/cs_report.md
       - "Repeated cross-sections": guides/repeated_cross_sections.md
       - "Sensitivity analysis (Rambachan–Roth)": guides/honest_did.md
+      - "Choosing a DID estimator": guides/choosing_did_estimator.md
+      - "Choosing an RD estimator": guides/choosing_rd_estimator.md
+      - "Choosing an IV estimator": guides/choosing_iv_estimator.md
+      - "Choosing a matching estimator": guides/choosing_matching_estimator.md
+      - "Robustness workflow": guides/robustness_workflow.md
   - Reference:
       - "Overview": reference/index.md
       - "Difference-in-differences": reference/did.md

[…]
