### 2026-04-20 3f60d3b
feat(smart): sp.check_identification — design-level identification diagnostics
Reads DataFrame + design and outputs a priority-ordered list of
identification pitfalls BEFORE you fit any estimator.  Complements
sp.recommend() (which picks an estimator) and diagnose_result()
(which inspects a fitted model).

Checks:
- Bad controls: covariates with |corr| > 0.5 with treatment (mediator
  / selection-driver heuristic), flagged as warning with suggestion.
- Overlap: quick logistic PS fit; flags extreme propensities
  (<0.02 or >0.98) as warning; near-perfect separation as blocker.
- Treatment varia
[…]
