+++ tests/test_check_identification.py
@@ -1,0 +1,209 @@
+"""Tests for sp.check_identification — design-level identification diagnostics."""
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+from statspai.smart.identification import (
+    check_identification,
+    IdentificationReport,
+    DiagnosticFinding,
+)
+
+
+# ---------------------------------------------------------------------------
+# Clean DGP: should produce minimal findings
+# ---------------------------------------------------------------------------
+
+def test_clean_rct_has_no_blockers():
+    """A clean RCT with balanced treatment should have no blockers."""
+    rng = np.random.default_rng(1)
+    n = 500
+    df = pd.DataFrame({
+        'd': rng.binomial(1, 0.5, n),
+        'X1': rng.normal(size=n),
+        'X2': rng.normal(size=n),
+    })
+    df['y'] = 1.0 + 0.5 * df['X1'] + 1.0 * df['d'] + rng.normal(size=n)
+    rep = check_identification(df, y='y', treatment='d',
+                               covariates=['X1', 'X2'], design='observational')
+    assert rep.verdict in ('OK', 'WARNINGS'), (
+        f"Expected OK/WARNINGS, got {rep.verdict}: "
+        f"{[f.message for f in rep.findings if f.severity == 'blocker']}"
+    )
+    # No blockers expected on a well-balanced RCT
+    blockers = [f for f in rep.findings if f.severity == 'blocker']
+    assert len(blockers) == 0
+
+
+# ---------------------------------------------------------------------------
+# Overlap blocker: near-deterministic treatment
+# ---------------------------------------------------------------------------
+
+def test_near_deterministic_treatment_flagged_as_blocker():
+    """If treatment is nearly determined by covariates, must flag as blocker."""
+    rng = np.random.default_rng(2)
+    n = 500
+    X = rng.normal(size=n)
+    # Near-perfect separation: d = 1 iff X > 0 (no overlap)
+    d = (X > 0).astype(int)
+    y = 1.0 + X + d + rng.normal(scale=0.1, size=n)
+    df = pd.DataFrame({'y': y, 'd': d, 'X': X})
+    rep = check_identification(df, y='y', treatment='d',
+                               covariates=['X'], design='observational')
+    # Either blocker (perfect separation) or warning (extreme PS)
+    flagged = [f for f in rep.findings
+               if f.category == 'overlap'
+               and f.severity in ('blocker', 'warning')]
+    assert len(flagged) > 0, (
+        f"Near-deterministic treatment must flag overlap issue. "
+        f"Findings: {[(f.category, f.severity, f.message) for f in rep.findings]}"
+    )
+
+
+# ---------------------------------------------------------------------------
+# Bad control detection
+# ---------------------------------------------------------------------------
+
+def test_mediator_as_control_flagged():
+    """A covariate highly correlated with treatment (candidate mediator)
+    must be flagged."""
+    rng = np.random.default_rng(3)
+    n = 500
+    d = rng.binomial(1, 0.5, n)
+    # mediator M = 0.9 * d + noise
+    m = 0.9 * d + rng.normal(scale=0.2, size=n)
+    y = 1.0 + 2.0 * d + 0.5 * m + rng.normal(size=n)
+    df = pd.DataFrame({'y': y, 'd': d, 'm': m})
+    rep = check_identification(df, y='y', treatment='d',
+                               covariates=['m'], design='observational')
+    flagged = [f for f in rep.findings
+               if f.category == 'bad_controls']
+    assert len(flagged) > 0, "Should flag highly-correlated covariate"
+    # Should mention 'm'
+    assert any('m' in f.message for f in flagged)
+
+
+# ---------------------------------------------------------------------------
+# Treatment variation: too few treated
[…]
