+++ src/statspai/smart/identification.py
@@ -1,0 +1,577 @@
+"""Identification diagnostics: read the data, flag design pitfalls.
+
+``check_identification()`` reads a DataFrame + research design and
+outputs a prioritised list of *design-level* diagnostic warnings.
+
+This is different from ``recommend()`` (which picks an estimator) and
+from ``diagnose_result()`` (which inspects a fitted model).
+``check_identification()`` answers the question you should ask BEFORE
+running any estimator:
+
+    "Is my design capable of identifying a causal effect at all?"
+
+Checks performed (with literature references):
+
+1. Bad controls
+   - Post-treatment variables (candidate mediators/colliders) —
+     Angrist-Pischke MHE §3.2.
+   - Outcome-correlated covariates with variance dominated by the
+     treatment period — classic "controlling away the effect".
+
+2. Overlap / common support
+   - Propensity score distribution across treatment groups
+   - Extreme propensities (<0.01 or >0.99) that destabilise IPW
+
+3. Sample size and statistical power
+   - Minimum detectable effect (MDE) at 80% power
+   - Cluster-aware MDE if clustering is specified
+   - Unit-level vs observation-level power for panel data
+
+4. Variation in treatment
+   - Fraction treated (problematic if < 5% or > 95%)
+   - Cohort sizes (for DID)
+   - Variation in running variable density at cutoff (for RD)
+
+5. Clustering
+   - Recommendation: which level to cluster at
+   - Moulton factor warning if ignored
+
+Usage
+-----
+>>> import statspai as sp
+>>> diag = sp.check_identification(
+...     df, y='wage', treatment='training',
+...     covariates=['age', 'education'],
+...     id='worker_id', time='year',
+...     design='did',
+... )
+>>> print(diag.summary())
+>>> diag.verdict   # 'OK' | 'WARNINGS' | 'BLOCKERS'
+"""
+from __future__ import annotations
+
+from dataclasses import dataclass, field
+from typing import List, Optional, Dict, Any, Sequence
+
+import numpy as np
+import pandas as pd
+
+
+# ---------------------------------------------------------------------------
+# Result type
+# ---------------------------------------------------------------------------
+
+@dataclass
+class DiagnosticFinding:
+    """A single design-level finding."""
+    severity: str  # 'blocker' | 'warning' | 'info'
+    category: str  # 'bad_controls' | 'overlap' | 'power' | 'variation' | 'clustering'
+    message: str
+    suggestion: Optional[str] = None
+    evidence: Dict[str, Any] = field(default_factory=dict)
+
+    @property
+    def icon(self) -> str:
+        return {'blocker': '[X]', 'warning': '[!]', 'info': '[i]'}.get(
+            self.severity, '[?]')
+
+
+@dataclass
+class IdentificationReport:
+    """Report from ``check_identification``."""
+    findings: List[DiagnosticFinding] = field(default_factory=list)
+    design: str = ''
+    n_obs: int = 0
+    n_units: Optional[int] = None
+
+    @property
+    def verdict(self) -> str:
+        """Overall verdict: BLOCKERS | WARNINGS | OK."""
+        if any(f.severity == 'blocker' for f in self.findings):
+            return 'BLOCKERS'
+        if any(f.severity == 'warning' for f in self.findings):
+            return 'WARNINGS'
+        return 'OK'
+
+    def by_category(self, category: str) -> List[DiagnosticFinding]:
+        return [f for f in self.findings if f.category == category]
+
+    def summary(self) -> str:
+        lines = [
+            '=' * 70,
+            'Identification Diagnostics',
+            '=' * 70,
+            f'  Design:    {self.design}',
+            f'  N obs:     {self.n_obs:,}',
+        ]
+        if self.n_units is not None:
+            lines.append(f'  N units:   {self.n_units:,}')
[…]
