### 2026-04-20 dd22d69
fix(core): CausalResult.tidy(conf_level) uses t-dist when df_resid is recorded
Addresses trade-off #3 from the v0.9.3+1 retrospective.  When the
CausalResult carries ``model_info['df_resid']`` (finite, positive),
tidy(conf_level=...) now uses t.ppf for the critical value — correct
for small-sample OLS-style DID fits.  Falls back to normal.ppf only
when df_resid is missing (influence-function SEs: CS, synth, meta-
learners) where the asymptotic normal is the right call.

Previously the recomputation always used normal, which gave CIs that
were ~10% too narrow at df=10.  New behaviour matches th
[…]
