### 2026-04-20 f77be22
feat(smart): DAG-based bad controls + strict mode + IV first-stage F in check_identification
Closes trade-offs #1 and #2 from v0.9.3+1 retrospective.

New capabilities on sp.check_identification():

1. DAG-based bad-control detection (trade-off #2)
   - New ``dag=`` parameter accepting an sp.DAG.
   - When supplied, runs dag.bad_controls(treatment, outcome)
     (Cinelli-Forney-Pearl 2022) and flags mediators, descendants,
     colliders, and M-bias structural problems as BLOCKERS.
   - Also verifies the covariate set satisfies one of the DAG's
     adjustment sets; suggests the shortest valid
[…]
