+++ tests/test_check_identification.py
@@ -207,3 +207,272 @@ def test_top_level_api_exposed():
     """sp.check_identification should be importable from top-level."""
     assert hasattr(sp, 'check_identification')
     assert sp.check_identification is check_identification
+
+
+# ---------------------------------------------------------------------------
+# strict mode: IdentificationError
+# ---------------------------------------------------------------------------
+
+def test_strict_mode_raises_on_blocker():
+    """strict=True must raise IdentificationError when a blocker is present."""
+    from statspai.smart.identification import IdentificationError
+    rng = np.random.default_rng(11)
+    n = 400
+    X = rng.normal(size=n)
+    d = (X > 0).astype(int)  # near-perfect separation => blocker
+    y = 1.0 + X + d + rng.normal(scale=0.1, size=n)
+    df = pd.DataFrame({'y': y, 'd': d, 'X': X})
+    with pytest.raises(IdentificationError) as exc_info:
+        check_identification(df, y='y', treatment='d',
+                             covariates=['X'], design='observational',
+                             strict=True)
+    # Exception should carry the report and blocker messages
+    err = exc_info.value
+    assert err.report.verdict == 'BLOCKERS'
+    assert 'blocker' in str(err).lower() or 'identifi' in str(err).lower()
+
+
+def test_strict_mode_allows_warnings():
+    """strict=True must NOT raise when only warnings/info are present."""
+    rng = np.random.default_rng(12)
+    n = 500
+    df = pd.DataFrame({
+        'd': rng.binomial(1, 0.5, n),
+        'X1': rng.normal(size=n),
+    })
+    df['y'] = 1.0 + 0.5 * df['X1'] + 1.0 * df['d'] + rng.normal(size=n)
+    rep = check_identification(df, y='y', treatment='d', covariates=['X1'],
+                               design='observational', strict=True)
+    # Should reach here without raising
+    assert rep.verdict in ('OK', 'WARNINGS')
+
+
+def test_strict_mode_default_is_non_strict():
+    """Default strict=False must return a report even with blockers."""
+    rng = np.random.default_rng(13)
+    n = 400
+    X = rng.normal(size=n)
+    d = (X > 0).astype(int)
+    y = 1.0 + X + d + rng.normal(scale=0.1, size=n)
+    df = pd.DataFrame({'y': y, 'd': d, 'X': X})
+    rep = check_identification(df, y='y', treatment='d',
+                               covariates=['X'], design='observational')
+    # Returns, doesn't raise
+    assert rep.verdict == 'BLOCKERS'
+
+
+def test_identification_error_exposed_at_sp_top_level():
+    assert hasattr(sp, 'IdentificationError')
+    from statspai.smart.identification import IdentificationError as _IE
+    assert sp.IdentificationError is _IE
+
+
+# ---------------------------------------------------------------------------
+# IV strength check
+# ---------------------------------------------------------------------------
+
+def test_iv_weak_instrument_flagged():
+    """Near-zero correlation between instrument and treatment => weak IV warning/blocker."""
+    rng = np.random.default_rng(14)
+    n = 600
+    Z = rng.normal(size=n)
+    # D essentially unrelated to Z
+    D = rng.binomial(1, 0.5, n).astype(float)
+    Y = 1.0 + D + rng.normal(size=n)
+    df = pd.DataFrame({'y': Y, 'd': D, 'z': Z})
+    rep = check_identification(df, y='y', treatment='d', instrument='z',
+                               design='iv')
+    # Some finding about IV strength should exist
[…]
