+++ src/statspai/smart/identification.py
@@ -58,6 +58,27 @@ import pandas as pd
 
 
 # ---------------------------------------------------------------------------
+# Exception type for strict mode
+# ---------------------------------------------------------------------------
+
+class IdentificationError(Exception):
+    """Raised by ``check_identification(strict=True)`` when a blocker is found.
+
+    Carries the full :class:`IdentificationReport` on ``self.report`` so
+    downstream code can still inspect findings without re-running.
+    """
+
+    def __init__(self, report: 'IdentificationReport'):
+        self.report = report
+        blockers = [f for f in report.findings if f.severity == 'blocker']
+        header = (f"Identification has {len(blockers)} blocker(s) "
+                  f"({report.design} design, N={report.n_obs})")
+        body = '\n'.join(f'  - {f.category}: {f.message}'
+                         for f in blockers)
+        super().__init__(f"{header}\n{body}" if body else header)
+
+
+# ---------------------------------------------------------------------------
 # Result type
 # ---------------------------------------------------------------------------
 
@@ -427,6 +448,188 @@ def _check_clustering(
             ))
 
 
+def _check_dag_bad_controls(
+    dag,
+    treatment: str,
+    outcome: str,
+    covariates: Sequence[str],
+    findings: List[DiagnosticFinding],
+) -> None:
+    """DAG-based bad-control detection (Cinelli-Forney-Pearl 2022).
+
+    Unlike the correlation heuristic, this catches *M-bias* colliders,
+    mediators, and descendants of the treatment that look pre-treatment
+    but violate backdoor adjustment.
+    """
+    if dag is None:
+        return
+    # Flag any requested covariate that is itself a bad control
+    try:
+        dag_bad = dag.bad_controls(treatment, outcome)
+    except Exception as e:
+        findings.append(DiagnosticFinding(
+            severity='info',
+            category='bad_controls',
+            message=f'DAG bad-control analysis skipped ({e}).',
+        ))
+        return
+
+    requested = set(covariates or [])
+    hit = {v: r for v, r in dag_bad.items() if v in requested}
+    if hit:
+        for v, reasons in hit.items():
+            findings.append(DiagnosticFinding(
+                severity='blocker',
+                category='bad_controls',
+                message=(f"Covariate '{v}' is a DAG-flagged bad control: "
+                         f"{'; '.join(reasons)}."),
+                suggestion=f"Remove '{v}' from the covariate set; use "
+                           f"DAG.adjustment_sets('{treatment}', "
+                           f"'{outcome}') for a valid alternative.",
+                evidence={'covariate': v, 'reasons': reasons},
+            ))
+
+    # Also check that covariates form a valid adjustment set
+    try:
+        adj_sets = dag.adjustment_sets(treatment, outcome)
+    except Exception:
+        adj_sets = []
+    if adj_sets:
+        valid = any(set(a).issubset(requested) for a in adj_sets)
+        if not valid and requested:
+            shortest = min(adj_sets, key=len)
+            findings.append(DiagnosticFinding(
+                severity='warning',
+                category='bad_controls',
+                message=('Covariate set does not satisfy any DAG '
+                         'adjustment criterion; backdoor paths may be '
+                         'open.'),
+                suggestion=f'Use adjustment set: {sorted(shortest)}.',
+                evidence={'valid_adjustment_sets': [sorted(s)
+                                                    for s in adj_sets[:3]]},
+            ))
+
+
+def _check_iv_strength(
+    data: pd.DataFrame,
+    treatment: str,
[…]
