### 2026-04-20 c568b1f
test(parity): 2-sigma tight recovery at n=5000-8000 (test_paper_parity.py)
Escalates the baseline 4-sigma/n~2000 parity tests from
test_did_parity.py etc. to 2-sigma/n=5000-8000.  The population
parameters are derived from the DGPs analytically (not borrowed
from R or Stata output), so the claim this file validates is
purely mathematical: "every estimator recovers its target on a
clean DGP at this tolerance."

Coverage:
- High-N staggered DID (n=5000, homogeneous ATT=0.8):
  CS2021, Sun-Abraham, Wooldridge — each recovers truth within
  2 sigma; three-way pairwise agreement within 2 * combined
[…]
