+++ tests/reference_parity/test_paper_parity.py
@@ -1,0 +1,276 @@
+"""Paper-level parity: tight-tolerance recovery on high-N deterministic DGPs.
+
+This file strengthens the identification-validity guarantee from the
+rest of ``reference_parity/``.  The earlier tests use n<=3000 and
+4-sigma tolerance — that is deliberately loose so no single draw
+flakes.  Here we go the other way:
+
+- n = 5000-8000 units
+- 2-sigma tolerance (tight)
+- Each DGP's population parameter is DERIVED analytically (not
+  borrowed from R or Stata), so matching it is a purely mathematical
+  claim about estimator correctness.
+- Key numerical outputs are ADDITIONALLY pinned to 6 decimals as
+  drift guards: any refactor that shifts the 6th decimal must force
+  an explicit pin update, not a silent number change.
+
+The combination gives us:
+    (a) 2-sigma recovery on a derived truth -> no hidden bias,
+    (b) 6-decimal pin on a fixed seed    -> no numerical drift.
+
+References for the analytical population parameters are in
+``REFERENCES.md`` under the "High-N analytical parity" section.
+"""
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+from statspai.did import callaway_santanna, sun_abraham, wooldridge_did
+
+
+TIGHT_SIGMA = 2.0   # 95.45% one-sided coverage
+PIN_TOL = 1e-6      # 6 decimals for drift guards
+
+
+# ---------------------------------------------------------------------------
+# High-N staggered DID (homogeneous effect): analytical ATT = 0.8
+# ---------------------------------------------------------------------------
+
+@pytest.fixture(scope='module')
+def highn_staggered_homogeneous():
+    """dgp_did with n=5000, 10 periods, homogeneous effect=0.8.
+
+    Analytical population ATT: 0.8 (set by dgp_did).
+    Noise: Normal(0, 1); within-unit FE; staggered across 4 cohorts.
+    """
+    df = sp.dgp_did(n_units=5000, n_periods=10, effect=0.8,
+                    staggered=True, n_groups=4, seed=2026)
+    # dgp_did uses 'first_treat' as cohort; adapt to (g, t, i) for CS/SA.
+    # Never-treated units have first_treat = NaN or a sentinel.
+    out = df.rename(columns={'unit': 'i', 'time': 't',
+                             'first_treat': 'g', 'y': 'y'})
+    # Fill never-treated cohort (NaN) as 0 — the canonical "never" code
+    out['g'] = out['g'].fillna(0).astype(int)
+    return out
+
+
+class TestHighNStaggeredHomogeneous:
+    TRUTH = 0.8
+
+    def test_cs_recovers_truth_tight(self, highn_staggered_homogeneous):
+        """CS2021 on n=5000 homogeneous DGP must recover 0.8 within 2 sigma."""
+        r = callaway_santanna(highn_staggered_homogeneous,
+                              y='y', g='g', t='t', i='i',
+                              estimator='reg')
+        err = abs(r.estimate - self.TRUTH)
+        assert err <= TIGHT_SIGMA * r.se, (
+            f"CS high-N homogeneous: estimate={r.estimate:.6f}, "
+            f"truth={self.TRUTH}, se={r.se:.6f}, err/SE={err/r.se:.2f}"
+        )
+
+    def test_sa_recovers_truth_tight(self, highn_staggered_homogeneous):
[…]
