+++ tests/reference_parity/REFERENCES.md
@@ -20,31 +20,36 @@ errors.
 
 ## DGPs and their population parameters
 
-| DGP fixture                       | True parameter  | Source / derivation |
-|-----------------------------------|-----------------|---------------------|
-| `did_2x2_data`                    | ATT = 2.0       | `y = 1 + 0.3t + 0.5T + 2.0*T*t + u + e`; direct reading of the interaction coefficient gives ATT = 2.0 (Angrist-Pischke MHE Ch. 5). |
-| `did_staggered_homogeneous`       | ATT = 1.5       | Homogeneous effect (cohort- and time-invariant) so TWFE, CS, SA, Wooldridge all target the same estimand = post-period interaction coefficient = 1.5. |
-| `did_staggered_heterogeneous`     | Cohort-specific | Cohort effects 1.0 / 1.5 / 2.0; TWFE is biased (de Chaisemartin & D'Haultfoeuille 2020); CS-unit-uniform simple ATT is cohort-weighted mean ≈ 1.5. |
-| `rd_sharp_data`                   | ATE@c = 1.0     | Sharp RD with `m_1 - m_0 = 1.0` at `x=0` (Hahn-Todd-van der Klaauw 2001). |
-| `rd_fuzzy_data`                   | LATE = 0.8      | Fuzzy RD with complier share = 0.7; outcome jump = 0.56; ratio = 0.8 (Imbens-Lemieux 2008). |
-| `iv_strong_data`                  | LATE = 1.5      | Binary-Z IV with strong first stage and homogeneous complier effect = 1.5 (Imbens-Angrist 1994). |
-| `synth_factor_model_data`         | ATT = -5.0      | 2-factor DGP with treated loadings in the convex hull of donor loadings; SCM is exactly unbiased in the population (Abadie-Diamond-Hainmueller 2010 §3). |
-| `matching_cia_data`               | ATT = 2.0       | Selection on observables with homogeneous effect; CIA matching / IPW / CBPS / ebalance all target ATT = 2.0 (Imbens-Wooldridge 2009). |
+| DGP fixture | True parameter | Source / derivation |
+| --- | --- | --- |
+| `did_2x2_data` | ATT = 2.0 | `y = 1 + 0.3t + 0.5T + 2.0*T*t + u + e`; interaction coefficient = 2.0 (Angrist-Pischke MHE Ch. 5). |
+| `did_staggered_homogeneous` | ATT = 1.5 | Homogeneous effect; TWFE, CS, SA, Wooldridge target post-period interaction = 1.5. |
+| `did_staggered_heterogeneous` | Cohort-specific | Cohort effects 1.0 / 1.5 / 2.0; TWFE biased (de Chaisemartin-D'Haultfoeuille 2020); CS simple-ATT ≈ 1.5. |
+| `rd_sharp_data` | ATE@c = 1.0 | Sharp RD with `m_1 - m_0 = 1.0` at `x=0` (Hahn-Todd-van der Klaauw 2001). |
+| `rd_fuzzy_data` | LATE = 0.8 | Fuzzy RD, complier share 0.7, outcome jump 0.56, ratio 0.8 (Imbens-Lemieux 2008). |
+| `iv_strong_data` | LATE = 1.5 | Binary-Z IV, strong first stage, homogeneous complier effect = 1.5 (Imbens-Angrist 1994). |
+| `synth_factor_model_data` | ATT = -5.0 | 2-factor DGP, treated loadings in convex hull of donors; SCM exactly unbiased (Abadie-Diamond-Hainmueller 2010 §3). |
+| `matching_cia_data` | ATT = 2.0 | CIA with homogeneous effect; matching / IPW / CBPS / ebalance target ATT = 2.0 (Imbens-Wooldridge 2009). |
[…]
