### 2026-04-20 59415ac
feat: v0.9.4 — sp.auto_cate() multi-learner race + design spec
Adds sp.auto_cate() — one-line multi-learner CATE race.  Races
S/T/X/R/DR-Learners on the same data with honest cross-fitted
nuisances, scores each via the Nie-Wager (2021) held-out R-loss,
runs Chernozhukov et al. (2018) BLP calibration per learner, and
picks a winner by the rule:

    "lowest R-loss among BLP-beta1-calibrated learners"

Returns AutoCATEResult with:
- .leaderboard (sorted by R-loss; ATE, SE, CI, BLP, CATE spread)
- .best_learner / .best_result
- .results (full CausalResult per learner)
- .agreement (Pearson rho ma
[…]
