+++ src/statspai/metalearners/auto_cate.py
@@ -1,0 +1,506 @@
+"""
+``sp.auto_cate`` — one-line multi-learner CATE race with honest scoring.
+
+Fits several meta-learners on the same data, scores each on held-out
+data via the R-loss (Nie & Wager 2021), runs the BLP calibration test
+(Chernozhukov et al. 2018) on each, and picks a winner.
+
+Motivation
+----------
+Section 8 of StatsPAI's 0.9.3+1 retrospective called out "ML CATE
+scheduling isn't as good as econml" as a known gap. ``auto_cate``
+closes it: every meta-learner in this package is a first-class
+citizen, so racing them and reporting a leaderboard is a natural
+agent-native API.
+
+Why R-loss
+----------
+Unlike ATE bootstrap SE (which is about the variance of the mean),
+the R-loss directly measures held-out CATE prediction quality. It
+is uniform across S/T/X/R/DR because it only depends on held-out
+``tau_hat(X)``, ``m_hat(X)`` and ``e_hat(X)``:
+
+    R(tau_hat) = E[ ((Y - m_hat) - tau_hat(X) * (D - e_hat))^2 ]
+
+Selection rule
+--------------
+The winner is the learner with the lowest held-out R-loss. BLP
+**beta_1** (≈ the ATE) and **beta_2** (the heterogeneity signal,
+along with its p-value) are computed per learner and reported in the
+leaderboard as diagnostics — they help the analyst understand *what*
+the chosen learner is saying, but they are not part of the selection
+gate. A user who wants to pick by heterogeneity signal can inspect
+``result.leaderboard[['learner', 'r_loss', 'blp_beta2_pvalue']]``
+and override manually.
+
+Why not gate on beta_1: in this parametrization beta_1 equals the
+ATE in the units of Y, not a calibration factor around 1 — so a
+"beta_1 ≈ 1" gate would be testing "ATE is close to 1", which is
+not a sensible selector.
+"""
+
+from __future__ import annotations
+
+from dataclasses import dataclass, field
+from typing import Any, Dict, List, Optional, Sequence, Tuple
+
+import numpy as np
+import pandas as pd
+from scipy import stats
+from sklearn.base import clone
+from sklearn.model_selection import KFold
+
+from ..core.results import CausalResult
+from .metalearners import (
+    SLearner,
+    TLearner,
+    XLearner,
+    RLearner,
+    DRLearner,
+    metalearner,
+    _default_outcome_model,
+    _default_propensity_model,
+    _default_cate_model,
+    _get_propensity,
+    _prepare_data,
+)
+
+
+_LEARNER_NAMES = {
+    's': 'S-Learner', 't': 'T-Learner', 'x': 'X-Learner',
+    'r': 'R-Learner', 'dr': 'DR-Learner',
+}
+
+
+# ---------------------------------------------------------------------------
+# Result type
+# ---------------------------------------------------------------------------
+
+@dataclass
+class AutoCATEResult:
+    """Leaderboard + winner from ``sp.auto_cate``.
+
+    Attributes
+    ----------
+    leaderboard : pd.DataFrame
+        One row per learner with ATE, SE, CI, R-loss, BLP calibration
+        columns, and CATE dispersion. Sorted by R-loss ascending.
+    best_learner : str
+        Full name of the chosen winner (e.g. ``"DR-Learner"``).
+    best_result : CausalResult
+        Full fitted result for the winner — supports ``.summary()``,
+        ``.tidy()``, ``.glance()``.
+    results : dict[str, CausalResult]
+        All fitted learners keyed by short code (``'s'``, ``'t'``, ...).
+    agreement : pd.DataFrame
+        Pearson-rho matrix of CATE vectors across learners
[…]
