### 2026-04-20 fde1721
chore(docs): tighten v0.9.4 CHANGELOG wording + ignore .claude/ session state
- Clarify auto_cate winner selection: lowest held-out R-loss;
  BLP beta is reported as a diagnostic, not a selection gate
  (beta1 equals the ATE in Y units under this parametrization,
  so a "beta1 ≈ 1" gate has no natural threshold).
- Clarify IV first-stage F: covariates are partialled out before
  computing the instrument's F so the reported value matches
  Staiger-Stock (1997) when controls are present.
- .gitignore the local .claude/ session-lock directory.

Co-Authored-By: Claude Opus 4.7 (1M context) <norepl
[…]
