### 2026-04-20 46acc0c
feat(datasets): sp.datasets — canonical dataset bundle + external parity anchors
Consolidates the causal-inference teaching datasets under a single
import path sp.datasets, with every replica tuned so that the
canonical estimator recovers output in the neighbourhood of the
PUBLISHED NUMBERS on the original data.

New bundled replicas (src/statspai/datasets/_canonical.py):

- mpdta()              — Callaway-Sant'Anna (2021) county-year teen
                         employment, 500 counties × 5 years, target
                         simple ATT ≈ -0.04 (published R: -0.0454)
- card_1995()        
[…]
