+++ tests/external_parity/test_published_replications.py
@@ -1,0 +1,260 @@
+"""Pinned external-parity tests on canonical simulated replicas.
+
+Each test:
+
+1. Loads a bundled replica from ``sp.datasets``.
+2. Runs the canonical estimator for that paper.
+3. Asserts the output matches a pinned numerical value to 4 decimals.
+4. Asserts the output is in the neighbourhood of the published
+   estimate on the ORIGINAL data (proving the replica's structural
+   calibration is right).
+
+The pinned values come from running the current implementation on the
+simulated replica (seed=42) at the moment of test creation.  Updating
+a pinned value requires an explicit commit; this prevents silent
+numerical drift across refactors.
+"""
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+
+
+# =========================================================================
+# Pinned reference values (from current-main output on seed=42 replicas)
+# =========================================================================
+
+# mpdta — CS 2021 simple ATT
+PINNED_MPDTA_CS_ATT = -0.0330
+PINNED_MPDTA_CS_SE = 0.0027
+
+# Card 1995 — OLS and IV coefficients on educ
+PINNED_CARD_OLS_EDUC = 0.1100
+PINNED_CARD_IV_EDUC = 0.1418
+
+# NSW-Lalonde — naive OLS on re78
+PINNED_LALONDE_NAIVE_ATT = 1556.0
+
+# NSW-DW (Dehejia-Wahba) — naive OLS vs covariate-adjusted
+PINNED_NSW_DW_NAIVE_ATT = -8387.0
+PINNED_NSW_DW_ADJ_ATT = 2313.0
+
+# Lee 2008 — RD jump in Democratic voteshare
+PINNED_LEE_RD_JUMP = 0.0616
+
+# Tolerance for pinned matches: 1e-3 catches true drift while
+# allowing for BLAS-level floating-point variation (~1e-6).
+PIN_TOL = 1e-3
+
+
+# =========================================================================
+# Callaway-Sant'Anna (2021) — mpdta
+# =========================================================================
+
+class TestMpdtaParity:
+    """CS2021 on simulated mpdta replica."""
+
+    @pytest.fixture(scope='class')
+    def df(self):
+        return sp.datasets.mpdta()
+
+    def test_replica_shape_and_attrs(self, df):
+        assert df.shape == (2500, 5)
+        assert set(df.columns) == {'countyreal', 'year', 'lemp',
+                                   'first_treat', 'treat'}
+        assert df.attrs['expected_simple_att'] == pytest.approx(-0.04)
+        assert df.attrs['published_simple_att_original'] == pytest.approx(-0.0454)
+
+    def test_cs_simple_att_pinned(self, df):
+        """CS simple ATT on our replica matches the pinned value."""
+        r = sp.callaway_santanna(df, y='lemp', g='first_treat',
+                                 t='year', i='countyreal',
+                                 estimator='reg')
+        assert r.estimate == pytest.approx(PINNED_MPDTA_CS_ATT, abs=PIN_TOL)
+
+    def test_cs_att_in_published_neighbourhood(self, df):
+        """Estimate must be negative and within 50% of published R output."""
+        r = sp.callaway_santanna(df, y='lemp', g='first_treat',
+                                 t='year', i='countyreal',
+                                 estimator='reg')
+        # Published R did::att_gt simple ATT on original mpdta: -0.0454
+        # Our replica target: -0.04; accept anywhere in [-0.06, -0.02]
+        assert -0.06 <= r.estimate <= -0.02, (
+            f"CS simple ATT {r.estimate} outside expected range "
[…]
