### 2026-04-20 1e50956
test(coverage): Monte Carlo CI coverage suite + CS aggregation miscalibration finding
Adds tests/coverage_monte_carlo/ — the third layer of regression
testing (after recovery + pinned-fixture).  For each major
estimator, runs B=300 independent draws, computes empirical 95% CI
coverage, asserts empirical rate falls in the Wilson 99%-confidence
band around 0.95.

This is the only testing approach that catches SE MISCALIBRATION.
Recovery tests pass with underestimated SEs (point estimate is still
on-target on a single draw); only coverage tests surface the bug.

## Framework

- `@pytest.mark.slow
[…]
