+++ tests/coverage_monte_carlo/test_coverage.py
@@ -1,0 +1,266 @@
+"""Monte Carlo 95% CI coverage validation for every major estimator.
+
+For each estimator, we generate B independent draws of a deterministic
+DGP with known population parameter, build the 95% CI on each draw,
+and check that the CI covers the truth at least 92% of the time
+(Wilson 95% lower band for B=300 with nominal 0.95).
+
+Failures indicate SE miscalibration — a class of bug that recovery
+tests cannot detect.
+"""
+from __future__ import annotations
+
+import os
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+
+
+# Default draws per test.  Override via env STATSPAI_MC_DRAWS=N.
+# CI default: 300 (about 45-60s per test).  Set higher for deep audits.
+B_DEFAULT = int(os.environ.get('STATSPAI_MC_DRAWS', 300))
+
+# Nominal coverage
+NOMINAL_COVERAGE = 0.95
+
+
+def _coverage_rate(truths_in_ci: int, B: int) -> float:
+    return truths_in_ci / B
+
+
+def _wilson_bounds(B: int, conf: float = 0.99) -> tuple:
+    """Return (lo, hi) Wilson-score bounds around nominal 0.95 for B draws.
+
+    Uses ``conf`` confidence for the test-of-coverage itself (default
+    99% — want to be permissive so only gross SE bugs fail).  For B=30
+    this gives roughly [0.83, 0.99]; for B=300, [0.92, 0.98].
+    """
+    from scipy.stats import norm
+    p = NOMINAL_COVERAGE
+    z = norm.ppf((1 + conf) / 2)
+    denom = 1 + z**2 / B
+    centre = (p + z**2 / (2 * B)) / denom
+    half = z * (p * (1 - p) / B + z**2 / (4 * B**2))**0.5 / denom
+    return max(0.0, centre - half), min(1.0, centre + half)
+
+
+def _assert_calibrated(covered: int, B: int, label: str,
+                       upper_slack: float = 0.02) -> None:
+    """Assert empirical coverage is within the Wilson band of nominal 0.95."""
+    rate = _coverage_rate(covered, B)
+    lo, hi = _wilson_bounds(B)
+    # Most bugs manifest as under-coverage; upper bound is looser.
+    assert lo <= rate <= hi + upper_slack, (
+        f"{label} 95% CI coverage = {rate:.3f} outside Wilson band "
+        f"[{lo:.3f}, {hi+upper_slack:.3f}] (B={B})"
+    )
+
+
+# ---------------------------------------------------------------------------
+# OLS on RCT — baseline: should be essentially exactly calibrated
+# ---------------------------------------------------------------------------
+
+@pytest.mark.slow
+def test_ols_rct_ci_coverage():
+    """OLS on RCT data with robust SE: 95% CI must cover truth in [0.92, 0.98]."""
+    B = B_DEFAULT
+    truth = 1.5
+    covered = 0
+    for seed in range(B):
+        rng = np.random.default_rng(seed)
+        n = 500
+        d = rng.binomial(1, 0.5, n)
+        x = rng.normal(size=n)
+        y = 1.0 + 0.5 * x + truth * d + rng.normal(size=n)
+        df = pd.DataFrame({'y': y, 'd': d, 'x': x})
+        r = sp.regress('y ~ d + x', data=df, robust='hc1')
+        ci = r.conf_int()
+        # conf_int returns a DataFrame with 2 columns; row 'd' gives CI
+        lo, hi = ci.loc['d'].values
+        if lo <= truth <= hi:
+            covered += 1
+    _assert_calibrated(covered, B, 'OLS RCT')
+
+
+# ---------------------------------------------------------------------------
+# Classic 2x2 DID
[…]
