### 2026-04-20 fe6c780
feat: sp.causal() — end-to-end causal-inference workflow orchestrator
One call runs the full pipeline:
  1. sp.check_identification  → diagnostics
  2. sp.recommend             → pick an estimator
  3. fit the recommended model (auto-enriching with user covariates)
  4. design-specific robustness (pretrends for DID, first-stage F for IV)
  5. render an HTML or Markdown report

Usage:

    import statspai as sp
    w = sp.causal(df, y='wage', treatment='training',
                  id='worker', time='year', design='did')
    w.report('analysis.html')

Granular control (auto_run=False) lets the 
[…]
