+++ src/statspai/workflow/causal_workflow.py
@@ -1,0 +1,577 @@
+"""``sp.causal()`` — end-to-end causal-inference orchestrator.
+
+The workflow object runs the canonical pipeline:
+
+    1. ``.diagnose()``     — sp.check_identification (design-level blockers)
+    2. ``.recommend()``    — sp.recommend         (pick estimator)
+    3. ``.estimate()``     — fit the recommended model
+    4. ``.robustness()``   — method-specific robustness suite
+    5. ``.report(path)``   — one-page HTML summary with every output
+
+Every stage is cached; re-invoking ``.report()`` does not re-fit.
+Each stage is also independently callable so advanced users can skip
+or override steps.
+
+Usage
+-----
+>>> import statspai as sp
+>>> w = sp.causal(df, y='wage', treatment='training',
+...               id='worker', time='year', design='did')
+>>> w.report('analysis.html')
+>>> w.result      # the fitted CausalResult
+>>> w.diagnostics # IdentificationReport
+"""
+from __future__ import annotations
+
+from dataclasses import dataclass, field
+from html import escape
+from typing import Any, Dict, List, Optional
+
+import numpy as np
+import pandas as pd
+
+
+# ---------------------------------------------------------------------------
+# Workflow object
+# ---------------------------------------------------------------------------
+
+@dataclass
+class CausalWorkflow:
+    """Holds state across the diagnose -> estimate -> report pipeline."""
+
+    data: pd.DataFrame
+    y: str
+    treatment: Optional[str]
+    covariates: List[str]
+    id: Optional[str]
+    time: Optional[str]
+    running_var: Optional[str]
+    instrument: Optional[str]
+    cutoff: Optional[float]
+    cohort: Optional[str]
+    cluster: Optional[str]
+    design: Optional[str]
+    dag: Optional[Any]
+    strict: bool
+
+    # Outputs (filled as stages run)
+    diagnostics: Optional[Any] = None           # IdentificationReport
+    recommendation: Optional[Any] = None        # RecommendationResult
+    result: Optional[Any] = None                # CausalResult / EconometricResults
+    robustness_findings: Dict[str, Any] = field(default_factory=dict)
+
+    # Execution stats
+    stages_completed: List[str] = field(default_factory=list)
+
+    # ------------------------------------------------------------------
+    # Stage 1: diagnose
+    # ------------------------------------------------------------------
+
+    def diagnose(self):
+        """Run sp.check_identification, cache the report."""
+        from ..smart.identification import check_identification
+        self.diagnostics = check_identification(
+            self.data, y=self.y, treatment=self.treatment,
+            covariates=self.covariates, id=self.id, time=self.time,
+            running_var=self.running_var, instrument=self.instrument,
+            cluster=self.cluster, cutoff=self.cutoff,
+            design=self.design, cohort=self.cohort, dag=self.dag,
+            strict=self.strict,
+        )
+        # Auto-adopt design if detection was left to check_identification
+        if self.design is None:
+            self.design = self.diagnostics.design
+        self._mark('diagnose')
+        return self.diagnostics
+
+    # ------------------------------------------------------------------
+    # Stage 2: recommend
+    # ------------------------------------------------------------------
+
+    def recommend(self):
+        """Run sp.recommend() to pick an estimator."""
+        from ..smart.recommend import recommend as _rec
+        self.recommendation = _rec(
+            data=self.data, y=self.y, treatment=self.treatment,
+            covariates=self.covariates, id=self.id, time=self.time,
+            running_var=self.running_var, instrument=self.instrument,
+            cutoff=self.cutoff, design=self.design, dag=self.dag,
+        )
+        self._mark('recommend')
+        return self.recommendation
+
+    # ------------------------------------------------------------------
+    # Stage 3: estimate
[…]
