### 2026-04-20 b4eb4bd
feat: v0.9.5 — Bayesian causal + Optuna-tuned CATE + Rust HDFE spike
Three "诚实认怂" items from the v0.9.4 retrospective are now
actionable:

1. `sp.bayes_did` / `sp.bayes_rd` (PyMC) — 2x2 + panel DID with
   hierarchical random effects; sharp RD with local polynomial.
   Return `BayesianCausalResult` with HDI, prob_positive, R-hat,
   ESS, and ArviZ InferenceData on `.trace`. Broom-style
   `.tidy() / .glance() / .summary()`. PyMC + ArviZ are optional
   (`pip install 'statspai[bayes]'`).

2. `sp.auto_cate_tuned` (Optuna) — TPE sampler over nuisance GBM
   HP space against shared held-out R-loss
[…]
