+++ src/statspai/bayes/rd.py
@@ -1,0 +1,202 @@
+"""Bayesian sharp regression discontinuity via PyMC.
+
+Fits a local polynomial model around a known cutoff within a
+user-supplied bandwidth, with independent slope coefficients on each
+side of the cutoff and a normal prior on the discontinuity ``tau``.
+"""
+from __future__ import annotations
+
+from typing import Optional, Tuple
+
+import numpy as np
+import pandas as pd
+
+from ._base import (
+    BayesianCausalResult,
+    _require_pymc,
+    _summarise_posterior,
+)
+
+
+def _local_polynomial_design(
+    x: np.ndarray,
+    cutoff: float,
+    treated: np.ndarray,
+    poly: int,
+) -> np.ndarray:
+    """Build a design matrix with independent polynomial fits per side.
+
+    Columns: ``treated, (x-c), (x-c)^2, ..., (x-c)^p,
+              treated*(x-c), ..., treated*(x-c)^p``.
+    The intercept is handled separately in the PyMC model.
+    """
+    x_c = x - cutoff
+    cols = [treated.astype(float)]
+    for k in range(1, poly + 1):
+        cols.append(x_c ** k)
+    for k in range(1, poly + 1):
+        cols.append(treated * (x_c ** k))
+    return np.column_stack(cols)
+
+
+def bayes_rd(
+    data: pd.DataFrame,
+    y: str,
+    running: str,
+    cutoff: float = 0.0,
+    bandwidth: Optional[float] = None,
+    poly: int = 1,
+    *,
+    prior_tau: Tuple[float, float] = (0.0, 10.0),
+    prior_slope_sigma: float = 10.0,
+    prior_noise: float = 5.0,
+    rope: Optional[Tuple[float, float]] = None,
+    hdi_prob: float = 0.95,
+    draws: int = 2000,
+    tune: int = 1000,
+    chains: int = 4,
+    target_accept: float = 0.9,
+    random_state: int = 42,
+    progressbar: bool = False,
+) -> BayesianCausalResult:
+    """Bayesian sharp regression discontinuity.
+
+    Within the bandwidth ``[cutoff - bw, cutoff + bw]`` the outcome
+    is modelled as a polynomial (default order 1) in ``running - cutoff``
+    with independent slopes on each side. The causal parameter is the
+    jump ``tau`` at the cutoff, with a weakly informative Normal prior.
+
+    Parameters
+    ----------
+    data : pd.DataFrame
+    y : str
+        Outcome column.
+    running : str
+        Running / forcing variable column.
+    cutoff : float, default 0.0
+        Threshold above which treatment applies (``running >= cutoff``).
+    bandwidth : float, optional
+        Half-width of the local window around ``cutoff``. If ``None``,
+        uses the rule-of-thumb ``0.5 * std(running)``.
+    poly : int, default 1
+        Polynomial order (1 = local linear; 2 = local quadratic).
+        Stata's ``rdrobust`` defaults to 1.
+    prior_tau : (float, float), default ``(0.0, 10.0)``
+        Mean / SD of the Normal prior on the discontinuity.
+    prior_slope_sigma, prior_noise : float
+        Prior scales for the polynomial slopes and residual SD.
+    rope : (float, float), optional
+        Region of practical equivalence.
+    hdi_prob : float, default 0.95
+    draws, tune, chains, target_accept, random_state, progressbar : see
+        :func:`bayes_did`.
+
+    Returns
+    -------
+    BayesianCausalResult
+        Posterior summary of the local average treatment effect
+        (LATE) at the cutoff.
+    """
+    pm, _ = _require_pymc()
+
+    for c in (y, running):
[…]
