+++ src/statspai/bayes/_base.py
@@ -1,0 +1,234 @@
+"""Shared types + PyMC lazy-import guard for ``statspai.bayes``.
+
+PyMC is an **optional** dependency. Importing ``statspai.bayes`` never
+imports PyMC; it is resolved at call time inside each estimator. If
+PyMC is missing the estimator raises :class:`ImportError` with the
+install recipe, leaving the rest of the package unaffected.
+"""
+from __future__ import annotations
+
+from dataclasses import dataclass, field
+from typing import Any, Dict, Optional, Tuple
+
+import numpy as np
+import pandas as pd
+
+
+# ---------------------------------------------------------------------------
+# Optional-dependency guard
+# ---------------------------------------------------------------------------
+
+_PYMC_INSTALL_HINT = (
+    "The statspai.bayes module requires PyMC + ArviZ. Install with:\n"
+    "    pip install 'statspai[bayes]'\n"
+    "or directly:\n"
+    "    pip install pymc arviz"
+)
+
+
+def _require_pymc():
+    """Import PyMC and ArviZ, or raise a clear ImportError."""
+    try:
+        import pymc as pm  # noqa: F401
+        import arviz as az  # noqa: F401
+    except ImportError as err:
+        raise ImportError(_PYMC_INSTALL_HINT) from err
+    return pm, az
+
+
+# ---------------------------------------------------------------------------
+# Result type
+# ---------------------------------------------------------------------------
+
+@dataclass
+class BayesianCausalResult:
+    """Summary of a Bayesian causal fit.
+
+    A sibling of :class:`statspai.core.results.CausalResult` that
+    speaks posterior (HDI, posterior probabilities, convergence
+    diagnostics) instead of frequentist (CI, p-value).
+
+    Attributes
+    ----------
+    method : str
+        Human-readable method tag, e.g. ``"Bayesian DID (panel)"``.
+    estimand : str
+        Name of the causal estimand, e.g. ``"ATT"`` / ``"LATE"``.
+    posterior_mean, posterior_median, posterior_sd : float
+        Central tendency + dispersion of the posterior on the
+        causal parameter.
+    hdi_lower, hdi_upper : float
+        Endpoints of the 95 % highest density interval.
+    prob_positive : float
+        Posterior probability that the estimand is > 0.
+    prob_rope : float | None
+        Posterior probability that the estimand lies in a
+        user-supplied region of practical equivalence.
+    rhat : float
+        Gelman-Rubin potential scale reduction factor. Warn if
+        > 1.01.
+    ess : float
+        Effective sample size (bulk).
+    n_obs : int
+        Sample size used in the fit.
+    hdi_prob : float
+        Nominal HDI coverage (default 0.95).
+    trace : arviz.InferenceData | None
+        Full posterior trace for downstream plotting / diagnostics.
+    model_info : dict
+        Misc. fit metadata (draws, tune, chains, priors, ...).
+    """
+
+    method: str
+    estimand: str
+    posterior_mean: float
+    posterior_median: float
+    posterior_sd: float
+    hdi_lower: float
+    hdi_upper: float
+    prob_positive: float
+    rhat: float
+    ess: float
+    n_obs: int
+    hdi_prob: float = 0.95
+    prob_rope: Optional[float] = None
+    trace: Any = None  # arviz.InferenceData
+    model_info: Dict[str, Any] = field(default_factory=dict)
+
+    # ------------------------------------------------------------------
+    # Broom-style delegation (matches CausalResult.tidy / glance)
+    # ------------------------------------------------------------------
+
[…]
