### 2026-04-20 d35418a
feat(output): sp.paper_tables() — AER/QJE/Econometrica multi-panel tables
Given a list of fitted results + a template name, produce the
standard 4-panel layout economics journals expect:

- Main          (headline result, one column per spec)
- Heterogeneity (subsample / interaction results)
- Robustness    (alternative estimators on the same spec)
- Placebo       (outcomes that should NOT respond)

Templates ('aer', 'qje', 'econometrica', 'restat') set star
thresholds, SE style, and footer notes so the author hands off a
publication-ready LaTeX file without further formatting work.

sp.regtab
[…]
