+++ src/statspai/output/paper_tables.py
@@ -1,0 +1,271 @@
+"""``sp.paper_tables()`` — AER/QJE/Econometrica-style multi-panel paper tables.
+
+Given a list of fitted results + a template name, produce the
+standard multi-panel layout that top economics journals expect:
+
+- **Main** — headline result across specifications (baseline →
+  +covariates → +FE → +cluster).
+- **Heterogeneity** — same outcome, subsamples / interactions.
+- **Robustness** — alternative estimators on the same data.
+- **Placebo** — outcomes that should NOT respond.
+
+This is not a replacement for ``sp.regtable``; it is a convention
+layer that enforces journal-specific presentation rules (three-line
+rules, FE indicators, cluster notes, stars at (0.10, 0.05, 0.01)) so
+the authors can hand off a LaTeX file to a co-author or journal
+with no further formatting work.
+"""
+from __future__ import annotations
+
+from dataclasses import dataclass, field
+from typing import Any, Dict, List, Optional, Sequence
+
+from .regression_table import regtable, RegtableResult
+
+
+# ---------------------------------------------------------------------------
+# Journal templates
+# ---------------------------------------------------------------------------
+
+TEMPLATES = {
+    'aer': {
+        'label': 'AER style',
+        'star_levels': (0.10, 0.05, 0.01),
+        'se_type': 'se',
+        'stats': ('N', 'R-squared'),
+        'notes_default': [
+            "Standard errors in parentheses.",
+            "*** p<0.01, ** p<0.05, * p<0.10.",
+        ],
+    },
+    'qje': {
+        'label': 'QJE style',
+        'star_levels': (0.10, 0.05, 0.01),
+        'se_type': 'se',
+        'stats': ('N',),
+        'notes_default': [
+            "Robust standard errors in parentheses.",
+            "*** p<0.01, ** p<0.05, * p<0.10.",
+        ],
+    },
+    'econometrica': {
+        'label': 'Econometrica style',
+        'star_levels': (0.01, 0.05),     # only 5% and 1%
+        'se_type': 'se',
+        'stats': ('N', 'R-squared'),
+        'notes_default': [
+            "Standard errors in parentheses.",
+            "** p<0.05, *** p<0.01.",
+        ],
+    },
+    'restat': {
+        'label': 'Review of Economics and Statistics style',
+        'star_levels': (0.10, 0.05, 0.01),
+        'se_type': 'se',
+        'stats': ('N', 'R-squared'),
+        'notes_default': [
+            "Standard errors in parentheses.",
+            "*** p<0.01, ** p<0.05, * p<0.10.",
+        ],
+    },
+}
+
+
+# ---------------------------------------------------------------------------
+# Result container
+# ---------------------------------------------------------------------------
+
+@dataclass
+class PaperTables:
+    """Container for the multi-panel paper-table bundle.
+
+    Each attribute is a RegtableResult (has .text, .latex, .html,
+    .to_latex(), ...).  Iterate via ``.panels()`` or access by name:
+    ``pt.main`` / ``pt.heterogeneity`` / ``pt.robustness`` / ``pt.placebo``.
+    """
+    main: Optional[RegtableResult] = None
+    heterogeneity: Optional[RegtableResult] = None
+    robustness: Optional[RegtableResult] = None
+    placebo: Optional[RegtableResult] = None
+    template: str = 'aer'
+
+    def panels(self) -> Dict[str, RegtableResult]:
+        """Return {name: RegtableResult} for every populated panel."""
+        return {
+            name: getattr(self, name)
+            for name in ('main', 'heterogeneity', 'robustness', 'placebo')
+            if getattr(self, name) is not None
+        }
+
+    def to_latex(self, path: Optional[str] = None) -> str:
[…]
