### 2026-04-20 d487574
perf(benchmarks): benchmarks/ suite + HDFE 16-25x faster than linearmodels
Adds benchmarks/ directory with runnable scaling benchmarks and a
RESULTS.md reporter.  These live OUTSIDE tests/ because:
1. They take minutes (tests/ must stay fast).
2. They measure speed, not correctness, so assertions would flake
   on shared CI runners.
3. Results feed CHANGELOG / README, not pytest exit codes.

Contents:
- bench_regression.py — sp.regress vs statsmodels OLS
- bench_hdfe.py       — sp.absorb_ols vs linearmodels PanelOLS
- bench_did.py        — sp.callaway_santanna + wooldridge_did scaling
- run_al
[…]
