### 2026-04-20 a6d53b6
feat(agent): sp.agent — LLM tool-definition surface + error-remediation registry
Closes the "agent-native" branding promise with a concrete API layer
designed for LLMs consuming StatsPAI programmatically.

Three public functions:

1. sp.agent.tool_manifest() -> list[dict]
   JSON-schema tool specifications in Anthropic / OpenAI tool-use
   format.  Drop directly into client.messages.create(tools=...);
   the model can call regress / did / callaway_santanna / rdrobust /
   ivreg / ebalance / check_identification / causal as native tools.

2. sp.agent.execute_tool(name, arguments, data) -> dict

[…]
