+++ src/statspai/agent/remediation.py
@@ -1,0 +1,270 @@
+"""Error -> actionable-fix registry for agent self-repair loops.
+
+Given a Python exception raised from a StatsPAI tool call, ``remediate()``
+returns a structured suggestion the agent can use to repair its next
+attempt.  The registry covers the most common failure modes seen in
+practice:
+
+- Missing columns / typos
+- Non-binary treatment / wrong dtype
+- Too few observations per cohort
+- Identification blockers from ``strict=True`` mode
+- Formula-parse errors
+- Missing dependencies (sklearn, matplotlib, ...)
+"""
+from __future__ import annotations
+
+import re
+from typing import Any, Dict, Optional
+
+
+# ---------------------------------------------------------------------------
+# Registry
+# ---------------------------------------------------------------------------
+#
+# Each pattern:
+#   'match'       : regex matched against str(exception)
+#   'exception'   : optional type constraint (class name string)
+#   'category'    : short label
+#   'diagnosis'   : what went wrong (for the agent to understand)
+#   'fix'         : concrete next action (what the agent should try)
+
+REMEDIATIONS = [
+    # --- Column errors ---
+    {
+        # KeyError itself; any message inside (quoted column name);
+        # also catches "not in index" / "column not found" for other types.
+        'match': r".*",                  # matches when 'exception' filter hits
+        'exception': ['KeyError'],
+        'category': 'missing_column',
+        'diagnosis': (
+            "A requested column is not in the DataFrame. Most common "
+            "cause: typo, or column belongs to a different design."
+        ),
+        'fix': (
+            "Print df.columns and pick the correct name.  If the "
+            "estimator needs a cohort column (g=) but you only have "
+            "a binary treat, derive it first:\n"
+            "  treated = df[df[treat] == 1]; cohort_per_unit = "
+            "treated.groupby('id')['time'].min()"
+        ),
+    },
+    {
+        # Non-KeyError variants that still indicate missing columns
+        'match': r"(column not found|not in index|column.*does not exist)",
+        'category': 'missing_column',
+        'diagnosis': "A requested column is not in the DataFrame.",
+        'fix': "Print df.columns and pick the correct name.",
+    },
+
+    # --- Treatment variable errors ---
+    {
+        'match': r"(treatment must be binary|could not convert|invalid "
+                 r"treat)",
+        'category': 'non_binary_treatment',
+        'diagnosis': (
+            "Treatment column is not binary/integer-coded (0/1).  DID "
+            "and most matching estimators require 0/1 coding."
+        ),
+        'fix': (
+            "Recode: df['treat'] = (df['treat_original'] == 'yes').astype(int)."
+        ),
+    },
+
+    # --- Identification blockers ---
+    {
+        'match': r"Identification has .* blocker",
+        'exception': ['IdentificationError'],
+        'category': 'identification_blocker',
+        'diagnosis': (
+            "strict=True + check_identification found a design blocker "
+            "(no overlap / near-separation / no controls / etc.).  "
+            "Do NOT retry the same spec; fix the design first."
+        ),
+        'fix': (
+            "Re-run check_identification(strict=False) to see the full "
+            "report.  Common fixes: (a) drop bad controls, (b) trim "
+            "extreme propensity scores, (c) expand the sample, (d) "
+            "pick a different design (e.g. IV instead of observational)."
+        ),
+    },
+
+    # --- Singular matrix / collinearity ---
+    {
+        'match': r"(Singular matrix|LinAlgError|not.+invert|perfectly "
+                 r"collinear)",
+        'category': 'collinearity',
+        'diagnosis': (
+            "Design matrix is rank-deficient — two or more covariates "
[…]
