+++ src/statspai/agent/tools.py
@@ -1,0 +1,505 @@
+"""JSON-schema tool definitions + dispatch for StatsPAI estimators.
+
+Each tool specification follows the Anthropic / OpenAI tool-use
+format: ``{'name': ..., 'description': ..., 'input_schema': {...}}``.
+Agents built on top of Claude, GPT-4, etc. can use these directly
+without wrapping.
+"""
+from __future__ import annotations
+
+from typing import Any, Callable, Dict, List, Optional
+
+import pandas as pd
+
+
+# ---------------------------------------------------------------------------
+# Tool registry
+# ---------------------------------------------------------------------------
+#
+# Each entry:
+#   'name'            : canonical tool name (what the LLM sees)
+#   'description'     : short one-paragraph doc
+#   'input_schema'    : JSON schema for arguments (strict types)
+#   'statspai_fn'     : path resolved lazily
+#   'serializer'      : optional callable (result -> JSON-friendly dict)
+
+def _default_serializer(r) -> Dict[str, Any]:
+    """Serialise a ``CausalResult`` / ``EconometricResults`` to a JSON dict.
+
+    LLM tool-use protocol requires JSON-serialisable return values,
+    so convert floats / arrays / DataFrames to primitives.
+    """
+    out: Dict[str, Any] = {}
+    if hasattr(r, 'estimate'):
+        out['estimate'] = float(r.estimate)
+    if hasattr(r, 'se'):
+        out['std_error'] = float(r.se)
+    if hasattr(r, 'pvalue'):
+        out['p_value'] = float(r.pvalue)
+    if hasattr(r, 'ci') and r.ci is not None:
+        out['conf_low'] = float(r.ci[0])
+        out['conf_high'] = float(r.ci[1])
+    if hasattr(r, 'estimand'):
+        out['estimand'] = str(r.estimand)
+    if hasattr(r, 'method'):
+        out['method'] = str(r.method)
+    if hasattr(r, 'n_obs'):
+        out['n_obs'] = int(r.n_obs)
+    # Regression-style: extract coefficient table if no 'estimate'
+    if not out.get('estimate') and hasattr(r, 'params'):
+        try:
+            names = list(r.params.index)
+            # pvalues / std_errors can be either pandas Series (indexable
+            # by name) or numpy arrays (indexable by position); handle both.
+            def _get(obj, name, pos):
+                import pandas as pd
+                if obj is None:
+                    return None
+                if isinstance(obj, pd.Series):
+                    return float(obj[name])
+                try:
+                    return float(obj[pos])
+                except (TypeError, IndexError, KeyError):
+                    return None
+
+            coefs = {}
+            for pos, k in enumerate(names):
+                coefs[str(k)] = {
+                    'estimate': float(r.params.iloc[pos]),
+                    'std_error': _get(getattr(r, 'std_errors', None),
+                                       k, pos),
+                    'p_value': _get(getattr(r, 'pvalues', None), k, pos),
+                }
+            out['coefficients'] = coefs
+        except Exception:
+            pass
+    if hasattr(r, 'diagnostics'):
+        diag = {}
+        for k, v in r.diagnostics.items():
+            if isinstance(v, (int, float, str, bool)):
+                diag[str(k)] = v
+        if diag:
+            out['diagnostics'] = diag
+    return out
+
+
+def _identification_serializer(r) -> Dict[str, Any]:
+    """Serialise IdentificationReport."""
+    return {
+        'verdict': r.verdict,
+        'design': r.design,
+        'n_obs': r.n_obs,
+        'n_units': r.n_units,
+        'findings': [
+            {
+                'severity': f.severity,
+                'category': f.category,
+                'message': f.message,
+                'suggestion': f.suggestion,
[…]
