### 2026-04-20 037ddfe
feat: v0.9.6 — g-methods family + Proximal + Principal Stratification + Bayesian IV/fuzzy RD
Two independent sprints landed the same day.

Sprint B (this session):
- DML IIVM (binary D, binary Z, LATE via Wald-ratio EIF)
- DML internal split: _base + plr/irm/pliv/iivm files
- sp.g_computation (parametric g-formula; ATE/ATT/dose-response)
- sp.front_door (Pearl 95) + integrate_by='marginal'/'conditional'
- sp.msm (IPTW for time-varying confounders) with hand-rolled
  pweight IRLS (fixes statsmodels.freq_weights semantic mismatch)
- sp.mediate_interventional (VanderWeele et al. 2014), MC loop
  
[…]
