+++ src/statspai/bayes/fuzzy_rd.py
@@ -1,0 +1,259 @@
+"""Bayesian fuzzy regression discontinuity via PyMC.
+
+Models the intent-to-treat effect on Y and the first-stage effect on
+D at the cutoff jointly, then forms the LATE as the ratio
+``itt_Y / itt_D`` (Wald estimator). Using a deterministic ratio at
+the posterior level means the resulting posterior on ``late``
+automatically inherits both noise channels and becomes heavy-tailed
+under weak compliance (``itt_D ≈ 0``).
+"""
+from __future__ import annotations
+
+from typing import Optional, Tuple
+
+import numpy as np
+import pandas as pd
+
+from ._base import (
+    BayesianCausalResult,
+    _require_pymc,
+    _summarise_posterior,
+)
+
+
+def _local_poly_design(
+    x: np.ndarray, cutoff: float, treated: np.ndarray, poly: int,
+) -> np.ndarray:
+    """Design matrix for one-side-dependent polynomial in (x - cutoff)."""
+    x_c = x - cutoff
+    cols = []
+    for k in range(1, poly + 1):
+        cols.append(x_c ** k)
+    for k in range(1, poly + 1):
+        cols.append(treated * (x_c ** k))
+    if not cols:
+        return np.zeros((len(x), 0))
+    return np.column_stack(cols)
+
+
+def bayes_fuzzy_rd(
+    data: pd.DataFrame,
+    y: str,
+    treat: str,
+    running: str,
+    cutoff: float = 0.0,
+    bandwidth: Optional[float] = None,
+    poly: int = 1,
+    *,
+    prior_late: Tuple[float, float] = (0.0, 10.0),
+    prior_slope_sigma: float = 10.0,
+    prior_noise: float = 5.0,
+    rope: Optional[Tuple[float, float]] = None,
+    hdi_prob: float = 0.95,
+    draws: int = 2000,
+    tune: int = 1000,
+    chains: int = 4,
+    target_accept: float = 0.9,
+    random_state: int = 42,
+    progressbar: bool = False,
+) -> BayesianCausalResult:
+    """Bayesian fuzzy RD — LATE at the cutoff via joint ITT-Y / ITT-D posterior.
+
+    Models within-bandwidth observations:
+
+    .. code-block:: text
+
+        Y_i = a_Y + itt_Y * I(x_i >= c) + poly_Y(x_i - c) + eps_Y
+        D_i = a_D + itt_D * I(x_i >= c) + poly_D(x_i - c) + eps_D
+        LATE := itt_Y / itt_D   (deterministic posterior)
+
+    Each polynomial gets independent slopes on each side (standard
+    local-linear RD practice). Priors on the ITT jumps are
+    Normal(mean, ``prior_slope_sigma``) (weakly informative); LATE
+    inherits its prior from ``prior_late`` but the ratio form means
+    the posterior is data-driven once compliance is non-trivial.
+
+    Parameters
+    ----------
+    data : pd.DataFrame
+    y : str
+        Outcome column.
+    treat : str
+        Realised treatment uptake (0/1). May disagree with
+        ``running >= cutoff`` when compliance is partial.
+    running : str
+        Running variable.
+    cutoff : float, default 0.0
+    bandwidth : float, optional
+        Default: ``0.5 * std(running)``.
+    poly : int, default 1
+    prior_late : (float, float)
+        Normal prior on the LATE (used as a regularising prior on the
+        deterministic ratio — it is implemented as a prior on the
+        ``itt_Y`` magnitude so that the ratio inherits natural scale).
+    prior_slope_sigma, prior_noise : float
+    rope, hdi_prob, draws, tune, chains, target_accept, random_state,
[…]
