+++ src/statspai/inference/front_door.py
@@ -48,6 +48,7 @@ Fulcher, I.R., Shpitser, I., Marealle, S., Tchetgen Tchetgen, E.J.
 the generalized front-door criterion." *JRSS-B*, 82(1), 199-214.
 """
 
+import warnings
 from typing import Optional, List, Any
 import numpy as np
 import pandas as pd
@@ -63,6 +64,7 @@ def front_door(
     mediator: str,
     covariates: Optional[List[str]] = None,
     mediator_type: str = 'auto',
+    integrate_by: str = 'marginal',
     n_boot: int = 500,
     n_mc: int = 200,
     alpha: float = 0.05,
@@ -87,6 +89,24 @@ def front_door(
     mediator_type : {'auto', 'binary', 'continuous'}, default 'auto'
         How to model the mediator. ``'auto'`` detects binary if M takes
         only {0,1} values, else continuous.
+    integrate_by : {'marginal', 'conditional'}, default 'marginal'
+        How to integrate over the M-distribution when computing the
+        front-door ATE (continuous M only — ignored for binary M).
+
+        - ``'marginal'``: follows Pearl's (1995) aggregate formulation.
+          For each unit's X_base row, Monte Carlo samples of M are drawn
+          from the population marginal M|D=d distribution (covariate
+          rows are re-sampled inside the MC). Appropriate when the
+          causal estimand is population-averaged and you want variance
+          that reflects the full population of baseline-covariate rows.
+        - ``'conditional'``: follows Fulcher et al. (2020) generalized
+          front-door. For each unit i, M-samples are drawn from the
+          unit-specific conditional M|D=d,X_i distribution. Stricter
+          identification (requires the mediator model to be correct
+          conditional on X_i) but gives per-unit ATE contributions.
+
+        For binary M and for problems with no covariates the two
+        formulations coincide.
     n_boot : int, default 500
         Nonparametric bootstrap replications.
     n_mc : int, default 200
@@ -113,7 +133,18 @@ def front_door(
     front-door criterion is a DAG assumption. Use :func:`sp.dag` with
     ``front_door_adjustment_sets()`` to verify identifiability from your
     assumed DAG before calling this estimator.
+
+    References
+    ----------
+    Pearl, J. (1995). Causal diagrams for empirical research. *Biometrika*.
+    Fulcher et al. (2020). Robust inference on population indirect causal
+    effects: the generalized front-door criterion. *JRSS-B*.
     """
+    if integrate_by not in ('marginal', 'conditional'):
+        raise ValueError(
+            f"integrate_by must be 'marginal' or 'conditional'; "
+            f"got '{integrate_by}'"
+        )
     covariates = list(covariates or [])
     df = data[[y, treat, mediator] + covariates].dropna().reset_index(drop=True)
     Y = df[y].values.astype(float)
@@ -138,20 +169,44 @@ def front_door(
     rng = np.random.default_rng(seed)
 
     def _point(Y_, D_, M_, X_):
-        return _front_door_ate(Y_, D_, M_, X_, mediator_type, n_mc, rng)
+        return _front_door_ate(
[…]
