### 2026-04-20 30baa52
feat(rust): Phase 1 scaffold — PyO3 group_demean crate + Python bridge
Opens feat/rust-hdfe per the v0.9.5 Rust-HDFE spike spec
(docs/superpowers/specs/2026-04-20-v095-rust-hdfe-spike.md).

Contents on this branch only:
- rust/statspai_hdfe/ — Cargo crate with PyO3 bindings and a
  single-threaded group_demean reference implementation. Pinned
  to pyo3 0.21 (abi3-py39) + numpy 0.21 + rayon 1.10 + ndarray 0.15.
  Release profile: opt-level=3, LTO=fat, codegen-units=1, panic=abort.
- src/statspai/panel/hdfe_rust.py — Python bridge with graceful
  fallback: the bridge raises RuntimeError (not Imp
[…]
