### 2026-04-20 5625ecc
chore(bayes): add List import + v0.9.7 HTE-IV / ADVI design spec
Small prep changes for the upcoming v0.9.7 release:
- List typing import in bayes/_base.py (used by HTE-IV signatures)
- Design spec: Bayesian heterogeneous-effect IV + ADVI toggle

No behaviour change in this commit.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
