+++ src/statspai/bayes/_base.py
@@ -8,7 +8,7 @@ install recipe, leaving the rest of the package unaffected.
 from __future__ import annotations
 
 from dataclasses import dataclass, field
-from typing import Any, Dict, Optional, Tuple
+from typing import Any, Dict, List, Optional, Tuple
 
 import numpy as np
 import pandas as pd
