+++ docs/superpowers/specs/2026-04-20-v097-bayes-hte-iv-advi.md
@@ -1,0 +1,172 @@
+# StatsPAI v0.9.7 — Heterogeneous-effect Bayesian IV + ADVI toggle
+
+**Author:** Bryce Wang · **Date:** 2026-04-20 · **Status:** design → implementation
+
+## 1. Motivation
+
+v0.9.6 landed average-LATE Bayesian IV, fuzzy RD, and per-learner
+Optuna tuning. The deferred list going into this release was:
+
+1. Bayesian *heterogeneous-effect* IV — current `bayes_iv` reports a
+   single LATE number; it can't answer "is the LATE different for
+   subgroup X?".
+2. Variational inference (ADVI) toggle — NUTS is the right default
+   but for big-N / prototype mode users want a posterior in 30 s,
+   not 5 min.
+3. Bayesian bunching — queued for evaluation.
+
+## 2. Scope
+
+### In scope
+
+- **`sp.bayes_hte_iv(data, y, treat, instrument, effect_modifiers, ...)`**
+  — Bayesian linear IV where the LATE is linear in user-supplied
+  effect modifiers:
+
+  ```
+  D = pi_0 + pi_Z' Z + pi_X' X + v           (first stage)
+  Y = alpha + tau_0 * D + (tau_hte' M) * D + beta_X' X + rho * v_hat + eps
+  ```
+
+  where ``M = X_eff - mean(X_eff)`` is the centred effect-modifier
+  matrix. Posterior over ``(tau_0, tau_hte)`` gives:
+  - Average LATE = ``tau_0``
+  - CATE slope posteriors (one per modifier)
+  - Predicted CATE at user-supplied covariate values (optional)
+
+  Returns a new `BayesianHTEIVResult` that extends
+  `BayesianCausalResult` with a `.cate_slopes` DataFrame.
+
+- **`inference='nuts' | 'advi'`** parameter on every existing
+  Bayesian estimator (`bayes_did`, `bayes_rd`, `bayes_iv`,
+  `bayes_fuzzy_rd`) *and* the new `bayes_hte_iv`. ADVI goes through
+  `pm.fit(method='advi')` and draws `draws` samples from the
+  fitted approximation. Documented as "fast and approximate" —
+  R-hat is not meaningful for ADVI so we report `np.nan`.
+
+### Out of scope (with explicit rationale)
+
+- **Bayesian bunching** — after review, the Kleven/Saez/Chetty
+  structural bunching machinery is fundamentally a macro/public-
+  finance identification strategy that maps *behavioural elasticities
+  from kinks/notches in a tax schedule*. Turning that into a Bayesian
+  estimator requires:
+  1. A structural utility / optimisation parameterisation that
+     doesn't generalise across kink types;
+  2. Priors on taste heterogeneity that are domain-specific and
+     hard to set defaults for;
+  3. A model fit that is only as interpretable as the underlying
+     structural model, which defeats the "agent-native one-liner"
+     design thesis.
+
+  The frequentist `sp.bunching` already exists for users who need
+  this tool. **We explicitly do not ship `sp.bayes_bunching`**; the
+  pay-off per LOC is lower than every other item on the Bayesian
+  roadmap. Revisit if a user files a clear use case.
+
+- Multi-instrument interaction effects, marginal treatment effect
+  (MTE) posteriors — 0.9.8+.
+
+- Rust Phase 2 — separate branch work.
+
+## 3. API
+
+### 3.1 `sp.bayes_hte_iv`
+
+```python
+def bayes_hte_iv(
+    data: pd.DataFrame,
[…]
