### 2026-04-20 ff58277
feat(bayes): _sample_model helper + HTE-IV posterior summary
Implementation work on v0.9.7 HTE-IV spec:
- _sample_model(): unified NUTS / ADVI dispatch for all Bayesian
  estimators, parameterised by inference='nuts'|'advi'.
- HTE-IV posterior summary hook: computes CATE(modifiers) from
  tau_0 + tau_hte @ m_vec draws, with HDI + prob_positive.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
