### 2026-04-20 fb9aa6d
fix(causal): post-review bug fixes + hardening across Sprint B modules
Addresses 7 findings from a fresh-eyes code review of commit 037ddfe.
All fixes have dedicated regression tests in tests/test_review_fixes.py.

Fixes:
1. MSM _weighted_logit_cluster IRLS used the raw (unclipped) eta in
   the working response z while p_hat was derived from eta_clip. For
   rows with |eta| > 30 this mixed clipped probabilities with the
   unclipped linear predictor and could make IRLS diverge. z now
   uses eta_clip consistently.

2. DML _DoubleMLBase aggregation across n_rep > 1 splits: previously
   report
[…]
