### 2026-04-20 320ed24
fix(causal): round-2 review polish — IRM subgroup threshold, principal monotonicity, proximal bridge scaffold
Post-review round 2, addressing the remaining MAJOR/MINOR items from
the fresh-eyes review that were not in the first fix batch.

Changes:
1. DoubleMLIRM: mirror IIVM's _MIN_SUBGROUP_FIT=10 threshold. Previously
   fit a 100-tree GBM on any non-empty (D=1 or D=0) subgroup, even if
   it had 2 rows; now falls back to the subgroup mean for subgroups
   below 10. Also adds an explicit ValueError when D has no variation
   and a propensity-constant-D fallback path. Mirrors IIVM parity.

2.
[…]
