+++ tests/test_review_fixes_round2.py
@@ -1,0 +1,208 @@
+"""
+Regression tests for the second-pass post-review fixes.
+
+Covers:
+* IRM subgroup fit threshold (mirror of IIVM fix).
+* Principal score monotonicity-violation warning.
+* Proximal bridge='linear' scaffold and NotImplementedError for others.
+* front_door model_info['integrate_by_effective'] records what ran.
+"""
+
+import warnings
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+from statspai.dml import DoubleMLIRM
+
+
+# ---------------------------------------------------------------------
+# IRM subgroup fit threshold mirroring IIVM
+# ---------------------------------------------------------------------
+
+def test_irm_has_min_subgroup_fit_constant():
+    assert hasattr(DoubleMLIRM, '_MIN_SUBGROUP_FIT')
+    assert DoubleMLIRM._MIN_SUBGROUP_FIT >= 5
+
+
+def test_irm_raises_when_treatment_is_constant():
+    rng = np.random.default_rng(0)
+    n = 400
+    X = rng.normal(0, 1, n)
+    D = np.ones(n)  # no variation
+    Y = 2.0 + X + rng.normal(0, 0.3, n)
+    df = pd.DataFrame({'y': Y, 'd': D, 'x': X})
+    with pytest.raises(ValueError, match='variation in D'):
+        sp.dml(df, y='y', treat='d', covariates=['x'], model='irm')
+
+
+def test_irm_tiny_subgroup_does_not_blow_up():
+    """
+    Very unbalanced D with a small sample exercises the subgroup-fit
+    fallback. Previously the GBM ran on a few training rows and
+    produced noisy g1/g0; now subgroups below the threshold fall back
+    to the subgroup mean. Estimator should stay finite.
+    """
+    rng = np.random.default_rng(42)
+    n = 200
+    X = rng.normal(0, 1, n)
+    # Rare treatment — ~5% treated
+    D = (rng.uniform(0, 1, n) < 0.05).astype(float)
+    # At least one treated/one control required
+    D[0] = 1.0
+    D[1] = 0.0
+    Y = 1.5 * D + 0.5 * X + rng.normal(0, 0.3, n)
+    df = pd.DataFrame({'y': Y, 'd': D, 'x': X})
+    r = sp.dml(df, y='y', treat='d', covariates=['x'], model='irm',
+               n_folds=2)
+    assert np.isfinite(r.estimate)
+    assert np.isfinite(r.se) and r.se > 0
+
+
+# ---------------------------------------------------------------------
+# Principal score monotonicity violation warning
+# ---------------------------------------------------------------------
+
+def test_principal_score_warns_on_monotonicity_violation():
+    """
+    DGP where S responds OPPOSITE to D for a subset of units (i.e.
+    defiers dominate). The fitted p10(x) > p11(x) for those X values,
+    raw e_complier goes negative, and we expect a RuntimeWarning.
+    """
+    rng = np.random.default_rng(7)
+    n = 3000
+    X = rng.normal(0, 1, n)
+    D = rng.binomial(1, 0.5, n).astype(float)
+    # For rows with X > 0, S = 1 - D (defiers); for X <= 0, S = D (compliers).
+    S = np.where(X > 0, 1 - D, D).astype(float)
+    Y = 0.5 * D + 0.2 * X + rng.normal(0, 0.3, n)
+    df = pd.DataFrame({'y': Y, 'd': D, 's': S, 'x': X})
+
+    with warnings.catch_warnings(record=True) as caught:
[…]
