### 2026-04-20 28a6d81
feat(registry): register 6 Sprint-B causal functions + update dml schema
Post-sprint audit found that although the 6 new top-level functions
(g_computation, front_door, msm, mediate_interventional, proximal,
principal_strat) were exported from sp.__init__, none of them appeared
in sp.list_functions() / sp.describe_function() / sp.search_functions()
— the machine-readable registry that LLM agents use to discover what
StatsPAI can do. Fixing that discoverability gap here.

Changes:
- registry.py: add FunctionSpec for each of the 6 new functions with
  full parameter descriptions, enum values, de
[…]
