+++ tests/test_registry_new_modules.py
@@ -1,0 +1,197 @@
+"""
+Registry + integration tests for Sprint-B modules.
+
+Pins behaviours discovered in the post-sprint audit:
+* All new public functions are listed by sp.list_functions().
+* sp.describe_function(name) returns a populated spec for each.
+* sp.search_functions() surfaces each by its primary tag.
+* The dml registry entry reflects the IIVM addition (model enum).
+* CausalResult.tidy() / .glance() work on new-module outputs.
+* Class wrappers (MarginalStructuralModel, ProximalCausalInference) fit.
+* Integer-typed treatment columns don't break the estimators.
+"""
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+from statspai.core.results import CausalResult
+
+
+# ---------------------------------------------------------------------
+# Registry coverage
+# ---------------------------------------------------------------------
+
+NEW_FUNCTIONS = (
+    'g_computation',
+    'front_door',
+    'msm',
+    'mediate_interventional',
+    'proximal',
+    'principal_strat',
+)
+
+
+@pytest.mark.parametrize('fn_name', NEW_FUNCTIONS)
+def test_registry_lists_new_function(fn_name):
+    assert fn_name in sp.list_functions(), (
+        f'{fn_name} missing from sp.list_functions()'
+    )
+
+
+@pytest.mark.parametrize('fn_name', NEW_FUNCTIONS)
+def test_registry_describe_function_populated(fn_name):
+    spec = sp.describe_function(fn_name)
+    assert spec['name'] == fn_name
+    assert spec['category'] == 'causal'
+    assert len(spec['description']) > 50
+    assert len(spec['params']) >= 3
+    # Every param must have a non-empty description (we pay for LLM discoverability here)
+    for p in spec['params']:
+        assert p['description'] or p['name'] == 'data', (
+            f'{fn_name} param {p["name"]!r} missing description'
+        )
+
+
+def test_dml_registry_enum_includes_iivm():
+    """After the sprint the dml model= enum must list iivm."""
+    spec = sp.describe_function('dml')
+    model_param = next(p for p in spec['params'] if p['name'] == 'model')
+    assert 'iivm' in model_param['enum']
+    assert {'plr', 'irm', 'pliv', 'iivm'}.issubset(set(model_param['enum']))
+
+
+def test_registry_search_picks_up_new_modules():
+    """Tag-based search should locate each new module by a salient keyword."""
+    assert 'proximal' in [f['name'] for f in sp.search_functions('bridge')]
+    assert 'msm' in [f['name'] for f in sp.search_functions('iptw')]
+    assert 'front_door' in [f['name'] for f in sp.search_functions('pearl')]
+    assert 'principal_strat' in [f['name'] for f in sp.search_functions('sace')]
+
+
+# ---------------------------------------------------------------------
+# CausalResult integration (tidy / glance)
+# ---------------------------------------------------------------------
+
+@pytest.fixture
+def simple_binary_data():
+    rng = np.random.default_rng(42)
+    n = 500
+    X = rng.normal(0, 1, n)
+    D = rng.binomial(1, 0.5, n).astype(float)
+    Y = 1.0 * D + 0.5 * X + rng.normal(0, 0.3, n)
+    return pd.DataFrame({'y': Y, 'd': D, 'x': X})
+
+
+def test_g_computation_tidy_glance(simple_binary_data):
[…]
