+++ src/statspai/registry.py
@@ -304,17 +304,25 @@ def _build_registry():
     register(FunctionSpec(
         name="dml",
         category="causal",
-        description="Double/Debiased Machine Learning for treatment effect estimation.",
+        description=(
+            "Double/Debiased Machine Learning for treatment effect estimation. "
+            "Supports partially linear (PLR), interactive regression (IRM, binary D), "
+            "partially linear IV (PLIV), and interactive IV (IIVM, binary D/binary Z → LATE)."
+        ),
         params=[
             ParamSpec("data", "DataFrame", True),
             ParamSpec("y", "str", True, description="Outcome"),
-            ParamSpec("treatment", "str", True, description="Treatment variable"),
-            ParamSpec("controls", "list", True, description="List of control variable names"),
+            ParamSpec("treat", "str", True, description="Treatment variable"),
+            ParamSpec("covariates", "list", True, description="List of control variable names"),
+            ParamSpec("model", "str", False, "plr", "DML model family",
+                      ["plr", "irm", "pliv", "iivm"]),
+            ParamSpec("instrument", "str", False, description="Instrument (required for pliv/iivm)"),
             ParamSpec("n_folds", "int", False, 5, "Cross-fitting folds"),
+            ParamSpec("n_rep", "int", False, 1, "Repeated cross-fitting splits (median aggregation)"),
         ],
-        returns="EconometricResults",
-        example='sp.dml(data=df, y="outcome", treatment="treat", controls=["x1","x2"])',
-        tags=["dml", "ml", "causal", "semiparametric"],
+        returns="CausalResult",
+        example='sp.dml(df, y="wage", treat="training", covariates=["age","edu"], model="plr")',
+        tags=["dml", "ml", "causal", "semiparametric", "iivm", "plr", "irm", "pliv"],
         reference="Chernozhukov et al. (2018)",
     ))
 
@@ -714,6 +722,166 @@ def _build_registry():
         returns="Dict with method_type and checks list",
         example='sp.diagnose_result(result)',
         tags=["diagnostics", "robustness", "battery", "auto"],
+    ))
+
+    # -- G-methods family ------------------------------------------------- #
+    register(FunctionSpec(
+        name="g_computation",
+        category="causal",
+        description=(
+            "Parametric g-formula (standardization) estimator. "
+            "ATE/ATT for binary D, or dose-response curve for continuous D. "
+            "Consistent under correctly-specified outcome model; not doubly robust."
+        ),
+        params=[
+            ParamSpec("data", "DataFrame", True),
+            ParamSpec("y", "str", True, description="Outcome"),
+            ParamSpec("treat", "str", True, description="Treatment variable"),
+            ParamSpec("covariates", "list", True, description="Baseline covariates"),
+            ParamSpec("estimand", "str", False, "ATE", "Target estimand",
+                      ["ATE", "ATT", "dose_response"]),
+            ParamSpec("treat_values", "list", False, description="Dose grid (required for dose_response)"),
+            ParamSpec("n_boot", "int", False, 500, "Bootstrap replications for SE"),
+        ],
+        returns="CausalResult",
+        example='sp.g_computation(df, y="wage", treat="trained", covariates=["age","edu"])',
+        tags=["g-computation", "g-formula", "standardization", "causal", "robins"],
+        reference="Robins (1986); Hernán & Robins (2020) ch. 13",
+    ))
+
+    register(FunctionSpec(
+        name="front_door",
+        category="causal",
+        description=(
+            "Pearl's front-door adjustment: identifies ATE with unmeasured "
+            "confounding when a mediator fully transmits the effect of D on Y. "
[…]
