### 2026-04-20 075852e
feat: v0.9.7 — heterogeneous-effect Bayesian IV + ADVI toggle on all Bayesian estimators
Closes two v0.9.6 follow-ons; Bayesian bunching is explicitly
declined with documented reasoning.

Added:
- sp.bayes_hte_iv — Bayesian IV with linear CATE-by-covariate model.
  Returns BayesianHTEIVResult carrying avg LATE posterior, per-
  modifier slope HDIs (cate_slopes DataFrame), and
  predict_cate(values) for out-of-sample CATE queries. Uses control-
  function plug-in so NUTS samples cleanly. Multi-instrument +
  multi-modifier + exogenous controls supported.

- inference='nuts' | 'advi' parameter o
[…]
