+++ src/statspai/bayes/hte_iv.py
@@ -1,0 +1,279 @@
+"""Heterogeneous-effect Bayesian IV.
+
+Extends :func:`bayes_iv` with a linear CATE-by-covariate model:
+
+.. code-block:: text
+
+    D_i = pi_0 + pi_Z' Z_i + pi_X' X_i + v_i                        (first stage)
+    tau(M_i) = tau_0 + tau_hte' (M_i - M_bar)                       (CATE)
+    Y_i = alpha + tau(M_i) * D_i + beta_X' X_i + rho * v_hat_i + eps_i   (structural)
+
+The posterior gives the average LATE (``tau_0``) and a table of
+slopes on each effect modifier. ``prob_positive`` on any individual
+slope answers "is the CATE systematically higher when this variable
+is higher?".
+"""
+from __future__ import annotations
+
+from typing import List, Optional, Sequence, Tuple, Union
+
+import numpy as np
+import pandas as pd
+
+from ._base import (
+    BayesianHTEIVResult,
+    _require_pymc,
+    _sample_model,
+    _summarise_posterior,
+)
+
+
+def _prepare_hte_iv_frame(
+    data: pd.DataFrame,
+    y: str,
+    treat: str,
+    instrument: Union[str, Sequence[str]],
+    effect_modifiers: Sequence[str],
+    covariates: Optional[List[str]],
+) -> dict:
+    if isinstance(instrument, str):
+        iv_cols = [instrument]
+    else:
+        iv_cols = list(instrument)
+    mod_cols = list(effect_modifiers)
+    cov_cols = list(covariates) if covariates else []
+
+    for c in [y, treat] + iv_cols + mod_cols + cov_cols:
+        if c not in data.columns:
+            raise ValueError(f"Column '{c}' not found in data")
+    if not mod_cols:
+        raise ValueError("effect_modifiers must list at least one covariate.")
+
+    # Reject overlap between effect_modifiers and covariates. If the
+    # same column appears in both, the structural equation at fit
+    # time includes it as (M - M_bar) * D *and* as plain beta_X * X,
+    # while the first-stage plug-in v_hat only includes it through X.
+    # That asymmetry is confusing and near-certain to be a user bug.
+    # Force the caller to pick a lane.
+    overlap = set(mod_cols) & set(cov_cols)
+    if overlap:
+        raise ValueError(
+            f"effect_modifiers and covariates must not overlap. "
+            f"Columns in both: {sorted(overlap)}. Pass the column as "
+            "one role only — either as an effect modifier (the LATE "
+            "slopes on it) or as a control (it enters both stages "
+            "but does not modify the LATE)."
+        )
+
+    all_cols = [y, treat] + iv_cols + mod_cols + cov_cols
+    # Deduplicate while preserving order (modifiers may overlap with
+    # covariates; that's fine, we just don't want NA-drop redundancy)
+    seen = set()
+    uniq_cols = [c for c in all_cols if not (c in seen or seen.add(c))]
+    clean = data[uniq_cols].dropna().reset_index(drop=True)
+    n = len(clean)
+    if n < 40:
+        raise ValueError(
+            f"bayes_hte_iv needs at least 40 observations; got {n}."
+        )
+
+    Y = clean[y].to_numpy(dtype=float)
+    D = clean[treat].to_numpy(dtype=float)
+    Z = clean[iv_cols].to_numpy(dtype=float)
+    M = clean[mod_cols].to_numpy(dtype=float)
[…]
