+++ src/statspai/bayes/__init__.py
@@ -20,16 +20,19 @@ Available estimators:
 """
 from __future__ import annotations
 
-from ._base import BayesianCausalResult
+from ._base import BayesianCausalResult, BayesianHTEIVResult
 from .did import bayes_did
 from .rd import bayes_rd
 from .iv import bayes_iv
 from .fuzzy_rd import bayes_fuzzy_rd
+from .hte_iv import bayes_hte_iv
 
 __all__ = [
     'bayes_did',
     'bayes_rd',
     'bayes_iv',
     'bayes_fuzzy_rd',
+    'bayes_hte_iv',
     'BayesianCausalResult',
+    'BayesianHTEIVResult',
 ]
