### 2026-04-20 fd0dea9
feat: v0.9.8 — Bayesian MTE (Heckman-Vytlacil) + Pathfinder + SMC samplers
Closes v0.9.7's next-batch list. Ships Python's first first-class
Bayesian Marginal Treatment Effect estimator and extends the sampler
dispatch from 2 backends (NUTS, ADVI) to 4.

Added:
- sp.bayes_mte — Heckman-Vytlacil (2005) MTE via plug-in logit
  first stage + polynomial MTE posterior. Returns
  BayesianMTEResult with .mte_curve (posterior on a user-supplied
  U_D grid), integrated .ate / .att / .atu, and .plot_mte(). Full
  joint first-stage + MTE posterior is explicitly deferred.

- inference='pathfinder' — route
[…]
