+++ src/statspai/bayes/mte.py
@@ -1,0 +1,287 @@
+"""Bayesian Marginal Treatment Effects (MTE) via PyMC.
+
+Heckman-Vytlacil (2005) MTE measures how the LATE varies along the
+propensity-to-be-treated distribution ``U_D``. The posterior over the
+MTE curve ``tau(u) = E[Y_1 - Y_0 | U_D = u]`` is obtained by fitting
+a polynomial on ``U_D`` inside a Bayesian linear-separable outcome
+model:
+
+.. code-block:: text
+
+    D_i = 1{ pi_0 + pi_Z Z_i + pi_X' X_i + U_D_i > 0 }
+    U_D_i ~ U(0, 1)  (on propensity scale)
+    tau(u) = b_0 + b_1 u + b_2 u^2 + ...
+    Y_i = alpha + beta_X' X_i + D_i * tau(U_D_i) + eps_i
+
+**Pragmatic shortcut (same trick as `sp.bayes_iv`)**: we estimate the
+propensity `P(D=1|Z,X)` by a *plug-in* logit and compute the induced
+``U_D_i`` pseudo-observations; the Bayesian layer then lies over the
+``tau(u)`` polynomial coefficients only. This is asymptotically
+correct when the first stage is correctly specified; it does not
+propagate first-stage uncertainty into the MTE posterior. Document
+this in the docstring.
+
+References:
+- Heckman, J. J., & Vytlacil, E. J. (2005). Structural equations,
+  treatment effects, and econometric policy evaluation.
+  *Econometrica*, 73(3), 669-738.
+"""
+from __future__ import annotations
+
+from typing import List, Optional, Tuple, Union
+
+import numpy as np
+import pandas as pd
+
+from ._base import (
+    BayesianMTEResult,
+    _require_pymc,
+    _sample_model,
+    _summarise_posterior,
+)
+
+
+def _logit_propensity(Z: np.ndarray, X: Optional[np.ndarray], D: np.ndarray):
+    """Plug-in logit first stage returning fitted P(D=1|Z,X).
+
+    Uses scikit-learn's LogisticRegression; falls back to a scipy
+    Newton-Raphson if sklearn is missing (unlikely in practice since
+    it's a core dep). Guarantees probabilities in ``[1e-4, 1 - 1e-4]``
+    so the induced quantile function is numerically well-behaved.
+    """
+    from sklearn.linear_model import LogisticRegression
+    W = Z.reshape(-1, 1) if Z.ndim == 1 else Z
+    if X is not None:
+        W = np.hstack([W, X])
+    clf = LogisticRegression(max_iter=500, solver='lbfgs')
+    clf.fit(W, D.astype(int))
+    ps = clf.predict_proba(W)[:, 1]
+    return np.clip(ps, 1e-4, 1 - 1e-4)
+
+
+def bayes_mte(
+    data: pd.DataFrame,
+    y: str,
+    treat: str,
+    instrument: str,
+    covariates: Optional[List[str]] = None,
+    *,
+    u_grid: Optional[np.ndarray] = None,
+    poly_u: int = 2,
+    prior_coef_sigma: float = 10.0,
+    prior_mte_sigma: float = 5.0,
+    prior_noise: float = 5.0,
+    rope: Optional[Tuple[float, float]] = None,
+    hdi_prob: float = 0.95,
+    inference: str = 'nuts',
+    advi_iterations: int = 20000,
+    draws: int = 2000,
+    tune: int = 1000,
+    chains: int = 4,
+    target_accept: float = 0.9,
+    random_state: int = 42,
+    progressbar: bool = False,
+) -> BayesianMTEResult:
[…]
