### 2026-04-20 7825d2d
feat(causal): escape-hatches + ROADMAP for 5 deferred causal items
Instead of silently punting on 5 items previously listed as "deferred",
ship lightweight escape hatches (new kwargs, helper functions, clearer
error messages) + a docs/ROADMAP.md entry for each. Every escape hatch
has a contract test that locks the user-facing behaviour so future
refactors can't silently break the documented workaround.

1. front_door continuous D: the rejection error now explicitly points
   users to sp.g_computation(..., estimand='dose_response') for the
   common case where the front-door criterion itself is
[…]
