+++ src/statspai/inference/front_door.py
@@ -155,8 +155,15 @@ def front_door(
 
     if not set(np.unique(D)).issubset({0, 1}):
         raise ValueError(
-            "Front-door estimator currently supports binary D (0/1). "
-            f"Got values {sorted(set(np.unique(D)))}."
+            "Front-door estimator currently supports binary D (0/1); "
+            f"got values {sorted(set(np.unique(D)))[:5]}"
+            f"{'...' if len(set(np.unique(D))) > 5 else ''}. "
+            "Continuous-D front-door is not shipped in this release — "
+            "see docs/ROADMAP.md. If the front-door criterion is not "
+            "required (i.e. no unmeasured confounder blocking the "
+            "back-door), use sp.g_computation(..., estimand='dose_response', "
+            "treat_values=[d1, d2, ...]) for a dose-response curve under "
+            "the standard unconfoundedness assumption."
         )
 
     if mediator_type == 'auto':
