### 2026-04-20 10c72b4
docs(reference): ship mkdocs pages for Sprint B causal modules
Closes the last ecosystem-integration gap from the post-Sprint-B
audit: six new top-level functions (g_computation, front_door, msm,
mediate_interventional, proximal, principal_strat) had zero coverage
in the rendered mkdocs site. Users browsing
https://statspai.../reference/ could not find them.

New reference pages:
- docs/reference/g_methods.md: unified doc for the Robins g-methods
  family — g_computation, front_door, msm, mediate_interventional —
  with identification-condition table, API snippets, and links to
  the DR cousin
[…]
