+++ docs/reference/g_methods.md
@@ -1,0 +1,231 @@
+# G-methods family
+
+Robins' **g-methods** family of causal-inference estimators, adapted
+to the StatsPAI `CausalResult` API. These methods all share the
+conceptual goal of evaluating counterfactual outcome distributions
+under hypothetical interventions, but differ in which nuisance is
+estimated (outcome model, treatment model, or both) and which
+identifying assumption is leveraged.
+
+| Estimator | When to reach for it | Identification |
+| --- | --- | --- |
+| `sp.g_computation` | Outcome model is trusted. Clean dose-response curves. | Unconfoundedness |
+| `sp.front_door` | Unobserved back-door confounder, but a mediator fully transmits D's effect on Y. | Pearl's front-door criterion |
+| `sp.msm` | Time-varying treatment with time-varying confounders affected by past treatment. | Sequential exchangeability + positivity |
+| `sp.mediate_interventional` | Mediation with treatment-induced mediator-outcome confounders. | Interventional effects (no cross-world) |
+
+Related modules (covered elsewhere):
+
+* `sp.aipw`, `sp.tmle`, `sp.dml` — doubly-robust cousins of
+  `g_computation` with orthogonal / efficient-influence-function
+  inference.
+* `sp.mediate` — classical natural (in)direct effects (Imai-Keele-
+  Tingley 2010); see the note in the mediation section below.
+
+---
+
+## `sp.g_computation` — parametric g-formula / standardisation
+
+Given an outcome regression :math:`Q(d, X) = E[Y | D=d, X]`, the
+g-formula marginal counterfactual is
+
+$$
+E[Y(d)] = E_X[\, Q(d, X) \,].
+$$
+
+```python
+# Binary treatment — ATE
+r = sp.g_computation(df, y='wage', treat='trained',
+                     covariates=['age', 'edu', 'exp'])
+
+# Binary treatment — ATT
+r = sp.g_computation(df, y='wage', treat='trained',
+                     covariates=['age', 'edu'],
+                     estimand='ATT')
+
+# Continuous / multi-valued treatment — dose-response curve
+r = sp.g_computation(df, y='bp', treat='dose',
+                     covariates=['age', 'bmi'],
+                     estimand='dose_response',
+                     treat_values=[0, 10, 20, 30])
+r.detail   # pd.DataFrame: dose / estimate / se / ci_lower / ci_upper / pvalue
+```
+
+**Learners**: defaults to OLS via statsmodels. Pass any sklearn-
+compatible regressor via `ml_Q=` for flexible fits:
+
+```python
+from sklearn.ensemble import GradientBoostingRegressor
+r = sp.g_computation(df, y='y', treat='d', covariates=[...],
+                     ml_Q=GradientBoostingRegressor(n_estimators=200))
+```
+
+**Inference**: nonparametric bootstrap with NaN-based failure
+tracking. Use `r.model_info['n_boot_failed']` to audit.
+
+**Relationship to AIPW/TMLE**: `g_computation` is consistent when the
+outcome model is correctly specified, but **not doubly robust**. Use
+`sp.aipw` or `sp.tmle` when you want coverage guarantees from either
+the outcome or the propensity model.
+
+---
+
+## `sp.front_door` — Pearl's front-door adjustment
+
+Identifies `E[Y | do(D)]` when an unmeasured confounder blocks the
+back-door criterion, but a mediator `M` fully transmits the effect of
+D on Y:
+
+```text
+U ──┬──► D ──► M ──► Y
+    │             ▲
+    └─────────────┘     (U unobserved)
+```
+
+Front-door formula:
+
+$$
[…]
