### 2026-04-20 58cc4d2
feat(api): article-facing aliases + auto_did / auto_iv race estimators
Fills the public-facing API gap exposed by the 2026-04-20 causal-
inference survey blog post: nine top-level aliases the article
advertises now resolve, and two new race-style estimators give
staggered-DiD and weak-IV users the canonical 'report all three
estimators' pattern in one call.

P0 — article aliases (src/statspai/_article_aliases.py)
  sp.rdd / sp.frontdoor / sp.xlearner / sp.conformal_ite / sp.psm /
  sp.partial_identification / sp.anderson_rubin_ci /
  sp.conditional_lr_ci / sp.tF_adjustment
  Every wrapper dele
[…]
