+++ tests/test_auto_estimators.py
@@ -1,0 +1,226 @@
+"""Tests for the P2 differentiating surface: sp.auto_did / sp.auto_iv
+plus a coverage property test for sp.conformal_ite.
+
+These cover the 'race multiple estimators' pattern that distinguishes
+StatsPAI from per-method Stata / R packages, and pin the empirical
+coverage of conformal intervals — both properties the blog post
+advertises but that are easy to regress on silently.
+"""
+
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+from statspai._auto_estimators import AutoDIDResult, AutoIVResult
+
+
+# =====================================================================
+# sp.auto_did
+# =====================================================================
+
+@pytest.fixture(scope="module")
+def did_panel():
+    return sp.dgp_did(
+        n_units=80,
+        n_periods=8,
+        staggered=True,
+        n_groups=3,
+        effect=0.5,
+        seed=2026,
+    )
+
+
+def test_auto_did_returns_leaderboard(did_panel):
+    result = sp.auto_did(
+        did_panel, y="y", g="first_treat", t="time", i="unit",
+    )
+    assert isinstance(result, AutoDIDResult)
+    assert set(result.leaderboard["method"]) == {"CS", "SA", "BJS"}
+    # All methods must populate numeric columns
+    for col in ["estimate", "std_error", "ci_lower", "ci_upper"]:
+        assert result.leaderboard[col].notna().all(), f"{col} has NaN"
+
+
+def test_auto_did_winner_points_into_candidates(did_panel):
+    result = sp.auto_did(
+        did_panel, y="y", g="first_treat", t="time", i="unit",
+    )
+    # The winner must be one of the candidate CausalResult objects,
+    # not a copy — this keeps `result.candidates['cs'] is result.winner`
+    # a meaningful assertion for downstream code.
+    assert any(result.winner is v for v in result.candidates.values())
+
+
+def test_auto_did_recovers_effect_in_band(did_panel):
+    """All three estimators should land within 0.5 of the true 0.5 effect
+    on this DGP.  The band is wide on purpose — we're testing that *no*
+    estimator has a catastrophic bug, not parity with a tabulated truth.
+    BJS can produce the most different estimate on staggered DGPs with
+    small groups because its imputation is more demanding on pretreatment
+    periods; see Borusyak, Jaravel & Spiess (2024)."""
+    result = sp.auto_did(
+        did_panel, y="y", g="first_treat", t="time", i="unit",
+    )
+    true_effect = 0.5
+    for _, row in result.leaderboard.iterrows():
+        assert abs(row["estimate"] - true_effect) < 0.5, row.to_dict()
+    # At least one estimator should get within 0.2 of the truth
+    close_enough = result.leaderboard["estimate"].sub(true_effect).abs() < 0.2
+    assert close_enough.any(), (
+        "no DiD estimator recovered the true effect within 0.2 — "
+        "likely a regression"
+    )
+
+
+def test_auto_did_respects_methods_filter(did_panel):
+    result = sp.auto_did(
+        did_panel, y="y", g="first_treat", t="time", i="unit",
+        methods=["cs", "sa"],
+    )
+    assert set(result.leaderboard["method"]) == {"CS", "SA"}
+    assert set(result.candidates.keys()) == {"cs", "sa"}
+
+
+def test_auto_did_select_by_specific_method(did_panel):
+    result = sp.auto_did(
+        did_panel, y="y", g="first_treat", t="time", i="unit",
+        select_by="sa",
+    )
[…]
