+++ tests/test_article_aliases.py
@@ -1,0 +1,323 @@
+"""Tests for the article-facing aliases (sp.rdd / sp.frontdoor / ...).
+
+These tests pin the alias → implementation mapping exposed by
+``src/statspai/_article_aliases.py``.  They guard against regressions in
+the public README / blog API and ensure the blog snippets actually run.
+
+We deliberately keep the assertions *structural* (right return type,
+reasonable magnitude) rather than numerical — exact numerical
+correctness is covered by the per-method parity tests in
+``tests/reference_parity/``.
+"""
+
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+from statspai.core.results import CausalResult
+
+
+# ---------------------------------------------------------------------------
+# Existence pin: every advertised alias resolves at the top level.
+# ---------------------------------------------------------------------------
+
+ADVERTISED_ALIASES = [
+    "rdd",
+    "frontdoor",
+    "xlearner",
+    "conformal_ite",
+    "psm",
+    "partial_identification",
+    "anderson_rubin_ci",
+    "conditional_lr_ci",
+    "tF_adjustment",
+]
+
+
+@pytest.mark.parametrize("name", ADVERTISED_ALIASES)
+def test_alias_is_exported(name):
+    assert hasattr(sp, name), f"sp.{name} missing from top-level namespace"
+    assert callable(getattr(sp, name)), f"sp.{name} is not callable"
+
+
+def test_alias_in_dunder_all():
+    for name in ADVERTISED_ALIASES:
+        assert name in sp.__all__, f"{name} missing from sp.__all__"
+
+
+# ---------------------------------------------------------------------------
+# sp.rdd — sharp RD integration smoke test
+# ---------------------------------------------------------------------------
+
+def test_rdd_sharp():
+    df = sp.dgp_rd(n=500, effect=0.4, cutoff=0.0, seed=7)
+    result = sp.rdd(df, y="y", running="x", cutoff=0.0)
+    assert isinstance(result, CausalResult)
+    assert np.isfinite(result.estimate)
+    # the DGP has effect ≈ 0.4; allow a wide band because bandwidth
+    # selection + bias correction introduces variance in small samples
+    assert abs(result.estimate - 0.4) < 0.5
+
+
+def test_rdd_matches_rdrobust():
+    """sp.rdd should be identical to sp.rdrobust with x/c naming."""
+    df = sp.dgp_rd(n=400, effect=0.3, cutoff=0.0, seed=11)
+    a = sp.rdd(df, y="y", running="x", cutoff=0.0)
+    b = sp.rdrobust(df, y="y", x="x", c=0.0)
+    assert a.estimate == pytest.approx(b.estimate, rel=1e-10)
+
+
+# ---------------------------------------------------------------------------
+# sp.xlearner — X-Learner CATE integration smoke test
+# ---------------------------------------------------------------------------
+
+def _make_xlearner_df(seed=0, n=400):
+    rng = np.random.default_rng(seed)
+    X1 = rng.normal(size=n)
+    X2 = rng.normal(size=n)
+    d = (rng.uniform(size=n) < 0.5).astype(int)
+    y = 1.0 + 0.5 * X1 + 0.3 * X2 + 0.8 * d + rng.normal(size=n) * 0.5
+    return pd.DataFrame({"y": y, "d": d, "x1": X1, "x2": X2})
+
+
+def test_xlearner_delegates_to_metalearner():
+    df = _make_xlearner_df(seed=0)
+    result = sp.xlearner(df, y="y", d="d", X=["x1", "x2"])
+    assert isinstance(result, CausalResult)
+    assert np.isfinite(result.estimate)
+    assert "cate" in result.model_info
[…]
