### 2026-04-20 b312306
feat: v0.9.9 — joint first-stage MTE + policy-relevant weights + honesty pass
Closes v0.9.8's two explicit follow-ons and ships a semantic
correction on the MTE labelling that survived two rounds of code
review.

Added:
- sp.bayes_mte(..., first_stage='plugin' | 'joint') — joint mode
  puts the first-stage logit coefficients inside the PyMC graph so
  first-stage uncertainty propagates into the effect-at-propensity
  curve. 2-4× slower than plug-in; default stays 'plugin' for
  backward compat.
- BayesianMTEResult.policy_effect(weight_fn, label, rope=None) —
  posterior summary of int w(u) g(u
[…]
