### 2026-04-21 bbdb0e4
feat(causal): wire Sprint-B estimators into sp.causal / recommend / diagnose
Closes the full ecosystem-integration gap from the post-Sprint-B audit.
After this pass the six new causal-inference modules (g_computation,
front_door, msm, mediate_interventional, proximal, principal_strat) are
not just importable and documented — they are *discoverable* from the
platform's end-to-end workflow, recommendation engine, and diagnostic
battery. All 142 focused regression tests pass.

Changes:

1. sp.recommend() — added five opt-in kwargs (mediator, tv_confounders,
   proxy_z, proxy_w, post_treat_strata)
[…]
