+++ src/statspai/smart/recommend.py
@@ -379,6 +379,13 @@ def recommend(
     cutoff: float = None,
     design: str = None,
     dag=None,
+    # --- Sprint-B / 0.9.6 causal-method extensions (all opt-in) ---
+    mediator: str = None,
+    tv_confounders: List[str] = None,
+    proxy_z: List[str] = None,
+    proxy_w: List[str] = None,
+    post_treat_strata: str = None,
+    # --- verification (pre-existing) ---
     verify: bool = False,
     verify_B: int = 50,
     verify_budget_s: float = 30.0,
@@ -420,6 +427,23 @@ def recommend(
         'observational', 'panel', 'cross-section'.
     dag : DAG, optional
         Causal DAG for identification analysis.
+    mediator : str, optional
+        Mediator variable name. Triggers mediation / front-door
+        recommendations (Imai-Keele-Tingley 2010,
+        VanderWeele et al. 2014, Pearl 1995).
+    tv_confounders : list of str, optional
+        Time-varying confounders (must be pre-treatment at each period).
+        Triggers `sp.msm` — Marginal Structural Model via IPTW
+        (Robins, Hernán & Brumback 2000).
+    proxy_z : list of str, optional
+        Treatment-side proxy variables. Triggers `sp.proximal` when
+        `proxy_w` is also supplied (Tchetgen Tchetgen et al. 2020).
+    proxy_w : list of str, optional
+        Outcome-side proxy variables (see `proxy_z`).
+    post_treat_strata : str, optional
+        Binary post-treatment variable defining principal strata
+        (take-up, survival, employment, …). Triggers
+        `sp.principal_strat` (Frangakis & Rubin 2002).
     verify : bool, default False
         If True, run *resampling-stability* checks on the top-k
         recommendations (bootstrap CV, permutation placebo, 50%-subsample
@@ -677,6 +701,184 @@ def recommend(
             'reason': 'Cross-sectional data with continuous outcome.',
             'code': f"sp.regress('{y} ~ {ctrl_str or '...'}', data=df, robust='hc1')",
             'params': {'formula': formula, 'data': data, 'robust': 'hc1'},
+        })
+
+    # ====================================================================== #
+    #  Sprint-B causal extensions (0.9.6): opt-in via the new kwargs.        #
+    #  These append to the candidate list — primary design-based             #
+    #  recommendations still drive the top slot.                             #
+    # ====================================================================== #
+
+    # Proximal Causal Inference — unobserved confounding with twin proxies
+    if proxy_z and proxy_w and treatment:
+        exog = [c for c in (covariates or []) if c not in proxy_z + proxy_w]
+        recommendations.append({
+            'method': 'Proximal Causal Inference (linear bridge 2SLS)',
+            'function': 'proximal',
+            'reason': 'Unmeasured confounder U with a treatment-side '
+                      'proxy Z and outcome-side proxy W available; '
+                      'linear bridge 2SLS identifies ATE under the '
+                      'proxy completeness conditions.',
+            'assumptions': [
+                'Z ⊥ Y | (D, U, X)  — treatment proxy',
+                'W ⊥ D | (U, X)  — outcome proxy',
+                'Linear outcome bridge h(W, D, X) (current release)',
+            ],
+            'robustness': 'Inspect first_stage_F in r.model_info; '
+                          'compare to sp.dml/sp.aipw for sensitivity.',
+            'code': f"sp.proximal(df, y='{y}', treat='{treatment}', "
+                    f"proxy_z={proxy_z!r}, proxy_w={proxy_w!r})",
+            'params': {
+                'data': data, 'y': y, 'treat': treatment,
+                'proxy_z': list(proxy_z), 'proxy_w': list(proxy_w),
+                'covariates': exog,
[…]
