### 2026-04-21 8a4cbd0
feat(api): article-surface round-2 — namespace fixes + kwarg alignment
Closes the six article-promised APIs that round-1 missed, hardens the
auto_* estimators, and walks back one experimental addition that broke
workflow dispatch.

Namespace-collision fixes (sp.X was a submodule, now a function):
  - sp.matrix_completion(df, y, d, unit, time)   → mc_panel
  - sp.causal_discovery(df, method='notears'|...) → dispatcher
  - sp.mediation(df, y, d, m, X)                  → mediate
  Uses the same late-bind-in-__init__.py trick the package already
  uses for sp.iv (line 127-129).  Dispatcher handles
[…]
