+++ tests/test_article_aliases_round2.py
@@ -1,0 +1,305 @@
+"""Round-2 fixes: namespace-collision aliases, kwarg alignment, and the
+tidy-attribute aliases on ``EconometricResults``.
+
+These tests pin the article-promised surface that the first alias pass
+missed (``sp.matrix_completion`` / ``sp.causal_discovery`` / ``sp.mediation``
+all used to resolve to modules, not functions), the ``depth=`` / ``model_y=``
+kwarg renames, the ``sp.evalue_rr`` convenience wrapper, and the
+cross-estimator ``.estimate`` / ``.se`` / ``.pvalue`` / ``.ci`` aliases on
+``EconometricResults``.
+"""
+
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+from statspai.core.results import CausalResult, EconometricResults
+
+
+# =====================================================================
+# Namespace-collision fixes
+# =====================================================================
+
+def test_matrix_completion_is_callable_function():
+    assert callable(sp.matrix_completion), (
+        "sp.matrix_completion should now be a function, not a module"
+    )
+    assert type(sp.matrix_completion).__name__ == "function"
+
+
+def test_causal_discovery_is_callable_function():
+    assert callable(sp.causal_discovery)
+    assert type(sp.causal_discovery).__name__ == "function"
+
+
+def test_mediation_is_callable_function():
+    assert callable(sp.mediation)
+    assert type(sp.mediation).__name__ == "function"
+
+
+def _make_mc_panel(seed=0, n_units=20, n_periods=8):
+    """Small panel with one treated unit & a few treated periods."""
+    rng = np.random.default_rng(seed)
+    rows = []
+    for u in range(n_units):
+        treated_unit = (u == 0)
+        for t in range(n_periods):
+            post = t >= n_periods // 2
+            d = int(treated_unit and post)
+            y = (
+                0.5 * u / n_units
+                + 0.1 * t
+                + 0.8 * d
+                + rng.normal() * 0.3
+            )
+            rows.append({"unit": u, "time": t, "y": y, "d": d})
+    return pd.DataFrame(rows)
+
+
+def test_matrix_completion_end_to_end():
+    df = _make_mc_panel(seed=1)
+    r = sp.matrix_completion(
+        df, y="y", d="d", unit="unit", time="time",
+        n_bootstrap=50,
+    )
+    assert isinstance(r, CausalResult)
+    assert np.isfinite(r.estimate)
+
+
+def test_mediation_end_to_end():
+    rng = np.random.default_rng(11)
+    n = 300
+    x1 = rng.normal(size=n)
+    d = (rng.uniform(size=n) < 0.5).astype(int)
+    m = 0.3 * d + 0.2 * x1 + rng.normal(size=n) * 0.3
+    y = 0.5 * m + 0.2 * d + 0.1 * x1 + rng.normal(size=n) * 0.3
+    df = pd.DataFrame({"y": y, "d": d, "m": m, "x1": x1})
+
+    r = sp.mediation(df, y="y", d="d", m="m", X=["x1"], n_boot=100)
+    assert isinstance(r, CausalResult)
+    # `mediate` reports a total effect by default; just check finiteness.
+    assert np.isfinite(r.estimate)
+
+
+def test_causal_discovery_dispatch_notears():
+    rng = np.random.default_rng(7)
+    n = 200
+    x1 = rng.normal(size=n)
+    x2 = 0.6 * x1 + rng.normal(size=n) * 0.5
[…]
