### 2026-04-21 56af79b
feat: v0.9.10 — HV-latent MTE (textbook Heckman-Vytlacil via latent U_D)
Closes the semantic debt v0.9.9 flagged: the polynomial-in-propensity
fit equals textbook MTE only under HV-2005 linear-separable +
bivariate-normal errors. v0.9.10 adds an opt-in fully HV-faithful
model that samples a latent U_D_i ~ Uniform(0,1) per unit via a
truncated-uniform reparameterisation, making the fitted polynomial
a genuine posterior over tau(u) = E[Y_1 - Y_0 | U_D = u].

Added:
- sp.bayes_mte(..., mte_method='polynomial' | 'hv_latent') — new
  kwarg orthogonal to first_stage. All four combinations run.
- hv_
[…]
