+++ src/statspai/bayes/mte.py
@@ -25,8 +25,42 @@ Model:
     # Selection (first stage)
     # 'plugin' mode: logit MLE gives fixed p_i
     # 'joint'  mode: pi ~ Normal priors, D_i ~ Bernoulli(sigmoid(pi'W_i))
-    g(p) = b_0 + b_1 p + b_2 p^2 + ...  (polynomial in propensity)
-    Y_i = alpha + beta_X' X_i + D_i * g(p_i) + eps_i
+
+    # MTE structural equation:
+    # 'polynomial' mode: g(p) = b_0 + b_1 p + ... (polynomial in propensity)
+    #                    Y_i = alpha + beta_X' X_i + D_i * g(p_i) + eps_i
+    # 'hv_latent'  mode: tau(u) = b_0 + b_1 u + ... (polynomial in latent U_D)
+    #                    raw_U_i ~ Uniform(0, 1)
+    #                    U_D_i = raw_U_i * p_i               if D_i = 1
+    #                          = p_i + raw_U_i * (1 - p_i)   if D_i = 0
+    #                    Y_i = alpha + beta_X' X_i + D_i * tau(U_D_i) + eps_i
+
+HV-augmentation factorisation (``hv_latent`` + ``joint``)
+---------------------------------------------------------
+This uses the *standard Form-2 data-augmentation factorisation*:
+
+    p(Y, D, U_D | p, θ) = p(Y | U_D, D, θ) · p(U_D | D, p) · p(D | p)
+
+    where p(U_D | D=1, p) = Uniform(0, p)
+          p(U_D | D=0, p) = Uniform(p, 1)
+          p(D=1 | p)       = p   (Bernoulli)
+
+The truncated-uniform augmentation enforces the HV indicator
+``D_i = 1{U_D_i < p_i}``, and the Bernoulli(D|p) term is the
+correct marginal selection likelihood. BOTH are needed: dropping
+the Bernoulli loses ~half the selection information and biases
+``pi`` (empirically verified — `piZ` flipped sign in a counter-
+factual test). Under this factorisation, marginalising out ``raw_U``
+exactly recovers the standard HV likelihood
+``p(Y, D | p, θ) = ∫ p(Y | U_D, D, θ) · p(U_D | D, p) dU_D · p(D|p)``
+where the integral runs over ``[0, p]`` (if D=1) or ``[p, 1]``
+(if D=0) against the truncated-uniform density.
+
+**Memory note**: ``hv_latent`` registers a latent ``raw_U`` of
+shape ``(n,)``, stored in the posterior as ``(chains, draws, n)``.
+For large ``n`` this can exceed hundreds of megabytes; the function
+emits a ``UserWarning`` when the product exceeds ~50M floats
+(~400MB at f64).
 
 References:
 - Heckman, J. J., & Vytlacil, E. J. (2005). Structural equations,
@@ -76,6 +110,7 @@ def bayes_mte(
     covariates: Optional[List[str]] = None,
     *,
     first_stage: str = 'plugin',
+    mte_method: str = 'polynomial',
     u_grid: Optional[np.ndarray] = None,
     poly_u: int = 2,
     prior_coef_sigma: float = 10.0,
@@ -113,6 +148,23 @@ def bayes_mte(
           polynomial sees it directly, so first-stage uncertainty
           propagates into the MTE curve. 2-4× slower than plugin but
           honest about uncertainty.
+    mte_method : {'polynomial', 'hv_latent'}, default ``'polynomial'``
+        MTE parameterisation.
+
+        - ``'polynomial'`` : fit a polynomial in the propensity
+          ``p_i`` (v0.9.9 behaviour). Under Heckman-Vytlacil 2005
[…]
