+++ tests/test_bayes_mte_hv_latent.py
@@ -1,0 +1,256 @@
+"""Tests for ``sp.bayes_mte(mte_method='hv_latent')`` — textbook HV MTE."""
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+from statspai.bayes import BayesianMTEResult, bayes_mte
+
+pymc = pytest.importorskip(
+    "pymc",
+    reason="PyMC is an optional dependency; skip Bayesian tests if missing.",
+)
+
+
+# ---------------------------------------------------------------------------
+# DGP: genuine Heckman-Vytlacil where outcome = MTE(U_D_i) * D_i + eps
+# ---------------------------------------------------------------------------
+
+def _hv_dgp(n, mte_coefs, seed):
+    """Data-generating process where the outcome equation evaluates
+    the MTE polynomial at the TRUE latent ``U_D_i`` (not at the
+    propensity). ``hv_latent`` mode should recover ``mte_coefs``.
+    """
+    rng = np.random.default_rng(seed)
+    Z = rng.normal(size=n)
+    U_D = rng.uniform(0, 1, size=n)
+    p_true = 1.0 / (1.0 + np.exp(-(0.8 * Z)))
+    D = (p_true > U_D).astype(float)
+    u_powers = np.column_stack([U_D ** k for k in range(len(mte_coefs))])
+    tau = u_powers @ np.asarray(mte_coefs, dtype=float)
+    Y = 1.0 + tau * D + 0.3 * rng.normal(size=n)
+    return pd.DataFrame({'y': Y, 'd': D, 'z': Z})
+
+
+@pytest.fixture
+def hv_decreasing_data():
+    # tau(u) = 2 - 2u: decreasing selection-on-gains
+    return _hv_dgp(600, mte_coefs=(2.0, -2.0), seed=801)
+
+
+# ---------------------------------------------------------------------------
+# API
+# ---------------------------------------------------------------------------
+
+
+def test_bayes_mte_hv_latent_runs(hv_decreasing_data):
+    r = bayes_mte(hv_decreasing_data, y='y', treat='d', instrument='z',
+                  mte_method='hv_latent', poly_u=1,
+                  draws=300, tune=300, chains=2, progressbar=False)
+    assert isinstance(r, BayesianMTEResult)
+    assert r.model_info['mte_method'] == 'hv_latent'
+    assert 'HV-latent' in r.method
+
+
+def test_bayes_mte_polynomial_method_label_unchanged(hv_decreasing_data):
+    r = bayes_mte(hv_decreasing_data, y='y', treat='d', instrument='z',
+                  mte_method='polynomial', poly_u=1,
+                  draws=200, tune=200, chains=2, progressbar=False)
+    assert r.model_info['mte_method'] == 'polynomial'
+    assert 'treatment-effect-at-propensity' in r.method
+
+
+# ---------------------------------------------------------------------------
+# Recovery vs bias: the whole point of v0.9.10
+# ---------------------------------------------------------------------------
+
+
+def test_hv_latent_recovers_true_mte_polynomial(hv_decreasing_data):
+    """On the true HV DGP, hv_latent should cover (b_0=2.0, b_1=-2.0)
+    inside the posterior 95% HDI. The test DGP uses n=600 so the
+    posterior is concentrated."""
+    r = bayes_mte(hv_decreasing_data, y='y', treat='d', instrument='z',
+                  mte_method='hv_latent', poly_u=1,
+                  draws=500, tune=500, chains=2, progressbar=False,
+                  random_state=42)
+    b_mte = r.trace.posterior['b_mte'].values.reshape(-1, 2)
[…]
