### 2026-04-21 3934820
feat(mediation): sp.mediate() pvalue_method kwarg + workflow fail-loud + deep battery tests
Three improvements from the post-ecosystem-integration code review.

1. sp.mediate() — pvalue_method kwarg
   Mirrors the kwarg already on sp.mediate_interventional so a single
   mediation study can use the same p-value convention across natural
   and interventional effects. Default 'bootstrap_sign' preserves the
   existing behaviour byte-for-byte (verified by regression test).
   'wald' uses 2·(1−Φ(|θ̂/ŝe|)) matching sp.aipw / sp.dml. The choice
   is recorded in model_info['pvalue_method'].
   Docs
[…]
