+++ tests/test_diagnose_batteries_sprint_b.py
@@ -1,0 +1,429 @@
+"""
+In-depth tests for the Sprint-B diagnostic batteries.
+
+Earlier round-3 tests only checked that the dispatcher lights up the
+right battery. This file exercises the actual diagnostic *logic* —
+each test verifies that a specific check fires with the correct
+pass/fail verdict in response to a specific condition in the input
+(weak first-stage F, extreme IPTW weight, monotonicity violation,
+heavy bootstrap failure rate, decomposition drift, etc.).
+"""
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+
+
+# ---------------------------------------------------------------------
+# Proximal — bridge / proxy rank / first-stage F
+# ---------------------------------------------------------------------
+
+@pytest.fixture
+def strong_proxy_data():
+    """High-F DGP: Z is a clean proxy for U."""
+    rng = np.random.default_rng(42)
+    n = 2000
+    U = rng.normal(0, 1, n)
+    D = 0.8 * U + rng.normal(0, 0.3, n)
+    Z = 0.95 * U + rng.normal(0, 0.2, n)
+    W = 0.9 * U + rng.normal(0, 0.3, n)
+    Y = 1.5 * D + 0.7 * U + rng.normal(0, 0.3, n)
+    return pd.DataFrame({'y': Y, 'd': D, 'z': Z, 'w': W})
+
+
+@pytest.fixture
+def weak_proxy_data():
+    """Low-F DGP: Z barely tracks U."""
+    rng = np.random.default_rng(42)
+    n = 500
+    U = rng.normal(0, 1, n)
+    D = 0.8 * U + rng.normal(0, 0.5, n)
+    Z = 0.1 * U + rng.normal(0, 1.0, n)   # very noisy proxy
+    W = 0.9 * U + rng.normal(0, 0.3, n)
+    Y = 1.5 * D + 0.7 * U + rng.normal(0, 0.3, n)
+    return pd.DataFrame({'y': Y, 'd': D, 'z': Z, 'w': W})
+
+
+def _check_by_name(diag, name_substr):
+    for c in diag['checks']:
+        if name_substr in c.get('test', ''):
+            return c
+    return None
+
+
+def test_proximal_battery_passes_on_strong_proxy(strong_proxy_data):
+    r = sp.proximal(strong_proxy_data, y='y', treat='d',
+                    proxy_z=['z'], proxy_w=['w'])
+    diag = sp.diagnose_result(r, print_results=False)
+    f_check = _check_by_name(diag, 'First-stage F')
+    assert f_check is not None
+    assert f_check['F'] >= 10.0
+    assert f_check['pass'] is True
+
+
+def test_proximal_battery_fails_on_weak_proxy(weak_proxy_data):
+    r = sp.proximal(weak_proxy_data, y='y', treat='d',
+                    proxy_z=['z'], proxy_w=['w'])
+    diag = sp.diagnose_result(r, print_results=False)
+    f_check = _check_by_name(diag, 'First-stage F')
+    assert f_check is not None
+    assert f_check['F'] < 10.0
+    assert f_check['pass'] is False
+
+
+def test_proximal_battery_order_condition_violated():
+    """Feed more W proxies than Z proxies — should fail the order check."""
+    rng = np.random.default_rng(0)
+    n = 800
+    U = rng.normal(0, 1, n)
+    D = 0.8 * U + rng.normal(0, 0.5, n)
+    Z1 = 0.9 * U + rng.normal(0, 0.3, n)
+    W1 = 0.9 * U + rng.normal(0, 0.3, n)
[…]
