### 2026-04-21 7d6ce12
feat(parity): paper-level parity + HDFE speed bench + Mixtape guide
Four parallel tracks closing the "benchmarks + docs" debt flagged
in the 2026-04-20 audit.  All four add verification-first surface
that will catch regressions on the canonical identification
strategies going forward.

Track A — Honest DiD paper-level parity
  tests/external_parity/test_honest_did_paper_parity.py (NEW, 9 tests)
  Pins sp.breakdown_m to Rambachan-Roth (2023) Definition 2 closed
  form M* = max((|θ̂| - z·SE) / n_drift, 0) at 5 parametric points,
  plus clamp-at-zero, α-sensitivity, missing-e error path, and an
 
[…]
