+++ tests/external_parity/test_honest_did_paper_parity.py
@@ -1,0 +1,152 @@
+"""Paper-level parity for Honest DiD (Rambachan & Roth, 2023, RES).
+
+The ``breakdown_m`` function implements Definition 2 of the paper under
+the smoothness restriction:
+
+    M* = (|θ̂| - z_{α/2} · SE) / n_drift
+
+where ``θ̂`` and ``SE`` are the event-study estimate and standard error
+at relative time ``e``, and ``n_drift = max(e + 1, 1)``.
+
+This module pins the implementation to the closed-form paper formula by
+constructing synthetic event-study tables with known ``(θ̂, SE)`` and
+comparing ``sp.breakdown_m`` against the analytical answer.  It is
+complementary to the structural smoke tests in ``tests/test_honest_did.py``.
+
+References
+----------
+Rambachan, A. & Roth, J. (2023). "A More Credible Approach to Parallel
+Trends." *Review of Economic Studies*, 90(5), 2555-2591.  Definition 2.
+"""
+
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+from scipy import stats
+
+import statspai as sp
+from statspai.core.results import CausalResult
+
+
+# ---------------------------------------------------------------------------
+# Helper: build a minimal CausalResult whose event study we control.
+# ---------------------------------------------------------------------------
+
+def _event_study_result(
+    relative_times: list,
+    atts: list,
+    ses: list,
+    *,
+    method: str = "Test DiD",
+) -> CausalResult:
+    """Wrap a hand-crafted (t, att, se) triple in a CausalResult so that
+    ``sp.breakdown_m`` sees exactly the numbers we supply — no DGP noise."""
+    assert 0 in relative_times, (
+        "_event_study_result expects a relative_time=0 row; other layouts "
+        "would make the headline-estimate fallback silently pick the wrong row."
+    )
+    es = pd.DataFrame({
+        "relative_time": relative_times,
+        "att": atts,
+        "se": ses,
+    })
+    # Pick the relative_time=0 row as the headline estimate.  Fallback to
+    # index 0 only if the assertion above is bypassed by a subclass;
+    # kept for defence in depth.
+    idx = next((i for i, t in enumerate(relative_times) if t == 0), 0)
+    return CausalResult(
+        method=method,
+        estimand="ATT(0)",
+        estimate=float(atts[idx]),
+        se=float(ses[idx]),
+        pvalue=0.0,
+        ci=(float(atts[idx] - 1.96 * ses[idx]),
+            float(atts[idx] + 1.96 * ses[idx])),
+        alpha=0.05,
+        n_obs=1000,
+        model_info={"event_study": es},
+    )
+
+
+# ---------------------------------------------------------------------------
+# Pin breakdown_m to the paper's closed form at several points.
+# ---------------------------------------------------------------------------
+
+@pytest.mark.parametrize("theta, se, e", [
+    (0.50, 0.10, 0),
+    (0.30, 0.05, 1),
+    (0.80, 0.20, 2),
+    (0.15, 0.03, 0),
+    (1.20, 0.25, 3),
+])
+def test_breakdown_m_matches_closed_form(theta, se, e):
+    """M* = (|θ̂| - z_{0.025} · SE) / n_drift where n_drift = max(e+1, 1)."""
+    # Construct an event study with a single positive-relative-time row.
+    # Include a pre-period row so the table shape is realistic.
+    rel_times = [-1, 0, 1, 2, 3]
[…]
