+++ benchmarks/bench_hdfe_scaled.py
@@ -1,0 +1,214 @@
+"""Scaled HDFE benchmark: sp.absorb_ols + sp.feols vs linearmodels PanelOLS.
+
+Extends ``bench_hdfe.py`` with 100k-1M observation scales, where the
+StatsPAI native alternating-projection HDFE and the pyfixest-backed
+``sp.feols`` pull decisively ahead of the dummy-variable baseline.
+
+Usage
+-----
+
+    # Run all three sizes (10k, 100k, 1M) and print wallclocks:
+    python benchmarks/bench_hdfe_scaled.py
+
+    # Quick mode (10k, 100k only — skip 1M):
+    python benchmarks/bench_hdfe_scaled.py --quick
+
+The 1M-row case takes ~5-10 s per backend on a modern laptop; keep the
+``--quick`` flag for CI.  Pass ``--json-out <path>`` to dump the
+result dict as JSON (parent dirs are created automatically).
+"""
+from __future__ import annotations
+
+import argparse
+import json
+import sys
+from pathlib import Path
+from typing import Dict, List, Optional
+
+import numpy as np
+import pandas as pd
+
+import statspai as sp
+
+# Allow "python bench_hdfe_scaled.py" from repo root or benchmarks/.
+sys.path.insert(0, str(Path(__file__).resolve().parent))
+from _utils import bench, fmt_ms, speedup_label  # noqa: E402
+
+
+# ---------------------------------------------------------------------
+# Data generation
+# ---------------------------------------------------------------------
+
+def _make_two_way_panel(n_units: int, n_periods: int,
+                        seed: int = 2026) -> pd.DataFrame:
+    """Realistic two-way-FE DGP (unit + time effects, 2 covariates)."""
+    rng = np.random.default_rng(seed)
+    n = n_units * n_periods
+    i_ = np.repeat(np.arange(n_units), n_periods)
+    t_ = np.tile(np.arange(n_periods), n_units)
+    x1 = rng.normal(size=n)
+    x2 = rng.normal(size=n)
+    unit_fe = rng.normal(size=n_units)[i_]
+    time_fe = rng.normal(size=n_periods)[t_]
+    y = 1.0 + 0.5 * x1 - 0.3 * x2 + unit_fe + time_fe + rng.normal(size=n)
+    return pd.DataFrame({
+        "i": i_, "t": t_, "x1": x1, "x2": x2, "y": y,
+    })
+
+
+# ---------------------------------------------------------------------
+# One benchmark row (per size)
+# ---------------------------------------------------------------------
+
+def _bench_size(n_units: int, n_periods: int, with_feols: bool,
+                with_linearmodels: bool) -> Dict:
+    n = n_units * n_periods
+    df = _make_two_way_panel(n_units, n_periods)
+
+    # sp.absorb_ols (native HDFE, Numba-JIT'd)
+    y_arr = df["y"].values
+    X_arr = df[["x1", "x2"]].values
+    fe_df = df[["i", "t"]]
+    sp_native = bench(
+        lambda: sp.absorb_ols(y_arr, X_arr, fe_df),
+        n_runs=3,
+    )
+
+    row = {
+        "n_units": n_units,
+        "n_periods": n_periods,
+        "n_obs": n,
+        "sp_absorb_ols_s": sp_native["mean_s"],
+    }
+
+    # sp.feols (pyfixest backend) — only if pyfixest is installed.
+    if with_feols:
+        try:
+            feols_result = bench(
+                lambda: sp.feols("y ~ x1 + x2 | i + t", data=df),
+                n_runs=2,
+            )
+            row["sp_feols_s"] = feols_result["mean_s"]
+            row["feols_vs_absorb"] = speedup_label(
[…]
