+++ benchmarks/RESULTS.md
@@ -20,6 +20,27 @@
 | 500 × 10 | 5,000 | 0.5 ms | 13 ms | 24.7x faster |
 | 2,000 × 10 | 20,000 | 1.4 ms | 22 ms | 16.2x faster |
 
+## HDFE scaled (10k / 100k, two-way FE) — 2026-04-20
+
+Results from `python benchmarks/bench_hdfe_scaled.py --quick`.
+`sp.absorb_ols` is the Numba-JIT'd alternating-projections HDFE
+kernel; `sp.feols` delegates to pyfixest when installed.
+
+| n_obs | sp.absorb_ols | linearmodels PanelOLS | speedup vs lm |
+|---:|---:|---:|:---:|
+| 10,000 | 0.8 ms | 17 ms | **21x faster** |
+| 100,000 | 6.4 ms | 75 ms | **11.7x faster** |
+
+For 1M-row runs, omit `--quick`; linearmodels becomes uncompetitive
+at that scale because it materialises the dummy-variable matrix.
+`sp.feols` (pyfixest backend) is the recommended path when you
+need the full `reghdfe`/`fixest` feature surface — R-style formulas,
+multi-way clustered SEs, singleton-group removal.
+
+**Takeaway**: on two-way-FE panels, `sp.absorb_ols` matches the
+"fast-HDFE" regime of Stata `reghdfe` and R `fixest`, delivering
+10-20× speedups over the dummy-variable baseline in `linearmodels`.
+
 ## Staggered DID (4 cohorts, 8 periods)
 
 | units | obs | CS 2021 | Wooldridge |

[…]
