### 2026-04-21 ae4eac1
feat(smart): Sprint-B coverage in assumption_audit + pub_ready + fail-loud tests
Two Smart-tool gaps for the Sprint-B causal modules, closed after a
fresh-eyes code review of this session's work.

1. sp.pub_ready — single-result acceptance + array/Series guard
   Pre-fix, sp.pub_ready(my_causal_result) crashed with
   "'CausalResult' object is not iterable". Now auto-wraps bare
   scalars; numpy.ndarray and pandas.Series pass through the iterable
   guard (so pub_ready(np.array([r1, r2])) iterates elements rather
   than wrapping the whole array). The truthy check was replaced by
   an explici
[…]
