+++ src/statspai/smart/assumptions.py
@@ -136,41 +136,118 @@ def assumption_audit(
     """
     checks = []
 
-    # Detect method type
-    model_info = getattr(result, 'model_info', {})
-    method = model_info.get('model_type', '')
+    # Detect method type. CausalResult stores the human-readable method
+    # on ``.method`` (e.g. "Proximal Causal Inference (linear 2SLS)");
+    # legacy EconometricResults put it in ``model_info['model_type']``.
+    # Read both so Sprint-B modules actually match on their keyword
+    # rather than falling silently into the OLS branch (which triggers
+    # on the *empty string* under the old detection — the "''" kw list
+    # member always matched).
+    model_info = getattr(result, 'model_info', {}) or {}
+    method_str = ' '.join([
+        str(model_info.get('model_type', '') or ''),
+        str(getattr(result, 'method', '') or ''),
+        str(model_info.get('estimator', '') or ''),
+    ]).lower()
 
     # Try to get data from result
     if data is None:
         data = getattr(result, '_data', None) or getattr(result, 'data', None)
 
-    # ===== OLS / LINEAR REGRESSION =====
-    if any(kw in method.lower() for kw in ['ols', 'linear', 'regress', '']):
-        checks.extend(_audit_linear(result, data, alpha))
+    # ----- Sprint-B (0.9.6) methods: explicit match BEFORE the generic
+    # OLS branch so they don't fall through into "looks like
+    # regression → run RESET" by accident.
+    sprint_b_matched = False
+    if any(kw in method_str for kw in ('proximal',)):
+        checks.extend(_audit_proximal(result, data, alpha))
+        sprint_b_matched = True
+    if any(kw in method_str
+           for kw in ('marginal structural', 'iptw')):
+        checks.extend(_audit_msm(result, data, alpha))
+        sprint_b_matched = True
+    # PrincipalStratResult stores 'monotonicity' or 'principal_score' on
+    # .method (not a CausalResult label), so match both on the class name
+    # and on the Zhang-Rubin / Ding-Lu keywords.
+    if (type(result).__name__ == 'PrincipalStratResult'
+            or any(kw in method_str for kw in (
+                'principal stratif', 'zhang-rubin',
+                'principal score', 'monotonicity'))):
+        checks.extend(_audit_principal_strat(result, data, alpha))
+        sprint_b_matched = True
+    if any(kw in method_str
+           for kw in ('g-computation', 'g computation',
+                      'g-formula', 'g formula',
+                      'standardization', 'standardisation')):
+        checks.extend(_audit_g_computation(result, data, alpha))
+        sprint_b_matched = True
+    if any(kw in method_str
+           for kw in ('front-door', 'front door', 'front_door')):
+        checks.extend(_audit_front_door(result, data, alpha))
+        sprint_b_matched = True
+    if any(kw in method_str
+           for kw in ('interventional mediation', 'iie')):
+        checks.extend(_audit_mediation_interventional(result, data, alpha))
+        sprint_b_matched = True
+    # Natural mediation branch is guarded: if a future estimator labels
+    # itself "Causal Mediation Analysis (interventional)" the
+    # interventional branch above fires first and this `and not
+    # sprint_b_matched` prevents double-firing the natural-effects
+    # stub on the same result.
+    if (not sprint_b_matched
+            and 'causal mediation' in method_str):
+        checks.extend(_audit_mediation_natural(result, data, alpha))
+        sprint_b_matched = True
 
-    # ===== IV / 2SLS =====
-    if any(kw in method.lower() for kw in ['iv', '2sls', 'liml', 'jive', 'gmm']):
-        checks.extend(_audit_iv(result, data, alpha))
[…]
