+++ tests/test_smart_tools_sprint_b.py
@@ -1,0 +1,338 @@
+"""
+Smart-tool integration tests for Sprint-B causal modules.
+
+Covers the second-round coverage fixes:
+
+1. ``sp.pub_ready`` now accepts a single ``CausalResult`` (was only
+   list/tuple).
+2. ``sp.assumption_audit`` recognises Sprint-B methods and fires the
+   method-specific assumption checks (proximal, msm, principal_strat,
+   g_computation, front_door, interventional/natural mediation).
+3. ``sp.assumption_audit`` does NOT leak panel / OLS / IV checks into
+   Sprint-B results via keyword-substring collisions
+   (e.g. "fe"/"re"/"iv" inside "Proximal Causal Inference (linear 2SLS)").
+"""
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+
+
+# ---------------------------------------------------------------------
+# Fixtures
+# ---------------------------------------------------------------------
+
+@pytest.fixture
+def proximal_result():
+    rng = np.random.default_rng(0)
+    n = 600
+    U = rng.normal(0, 1, n)
+    D = 0.8 * U + rng.normal(0, 0.5, n)
+    Z = 0.9 * U + rng.normal(0, 0.3, n)
+    W = 0.9 * U + rng.normal(0, 0.3, n)
+    Y = 1.5 * D + 0.7 * U + rng.normal(0, 0.3, n)
+    df = pd.DataFrame({'y': Y, 'd': D, 'z': Z, 'w': W})
+    return sp.proximal(df, y='y', treat='d', proxy_z=['z'], proxy_w=['w'])
+
+
+@pytest.fixture
+def msm_result():
+    rng = np.random.default_rng(0)
+    rows = []
+    for i in range(150):
+        V = rng.normal()
+        L = rng.normal()
+        A_prev = 0.0
+        for t in range(3):
+            L = L + 0.3 * A_prev + rng.normal(0, 0.3)
+            A = float(rng.binomial(1, 1 / (1 + np.exp(-0.4 * L))))
+            A_prev = A
+            rows.append({'id': i, 'time': t, 'A': A,
+                         'L_lag': L, 'V': V})
+    panel = pd.DataFrame(rows)
+    panel['Y'] = (
+        0.5 * panel.groupby('id')['A'].cumsum()
+        + panel['V'] * 0.3
+        + rng.normal(0, 0.3, len(panel))
+    )
+    return sp.msm(panel, y='Y', treat='A', id='id', time='time',
+                  time_varying=['L_lag'], baseline=['V'])
+
+
+# ---------------------------------------------------------------------
+# sp.pub_ready accepts single result (bug fix)
+# ---------------------------------------------------------------------
+
+def test_pub_ready_accepts_single_causal_result(proximal_result):
+    """Pre-fix, this raised TypeError: 'CausalResult' not iterable."""
+    pub = sp.pub_ready(proximal_result)
+    assert pub is not None
+    assert 0 <= pub.score <= 100
+
+
+def test_pub_ready_single_and_list_agree(proximal_result):
+    """
+    Passing ``r`` vs ``[r]`` must produce the same report. The fix
+    auto-wraps a scalar result; this contract test pins that
+    equivalence.
+    """
+    pub_single = sp.pub_ready(proximal_result)
+    pub_list = sp.pub_ready([proximal_result])
+    assert pub_single.score == pub_list.score
+    assert pub_single.present == pub_list.present
+    assert pub_single.missing == pub_list.missing
+
+
+def test_pub_ready_none_still_works():
+    """Backward-compat: sp.pub_ready() without results still returns a report."""
+    pub = sp.pub_ready()
+    assert pub is not None
+
+
+# ---------------------------------------------------------------------
[…]
