### 2026-04-21 612a3f1
feat(smart): g_computation in compare_estimators + Sprint-B method labels in sensitivity_dashboard
Round-3 Smart-tool integration pass for Sprint-B.

1. sp.compare_estimators — adds 'g_computation' as a first-class
   comparison method. Each branch now has an explicit binary-treat
   guard that emits a compact warning on continuous D rather than
   falling through to sp.g_computation's own 20 KB value-dump
   ValueError (which the outer try/except would otherwise swallow
   inside a multi-kilobyte UserWarning). The docstring explicitly
   lists the four Sprint-B methods that are NOT supported 
[…]
