+++ tests/test_smart_tools_sprint_b_round3.py
@@ -1,0 +1,194 @@
+"""
+Round-3 Smart-tool coverage for Sprint-B: compare_estimators registers
+g_computation, sensitivity_dashboard reports the correct method label
+for Sprint-B results, and ROADMAP §6 documents the compare_estimators
+gap for the remaining hint-driven estimators.
+"""
+
+from pathlib import Path
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+
+
+# ---------------------------------------------------------------------
+# compare_estimators — g_computation is now a first-class option
+# ---------------------------------------------------------------------
+
+@pytest.fixture
+def binary_dgp():
+    rng = np.random.default_rng(0)
+    n = 500
+    X = rng.normal(0, 1, n)
+    D = rng.binomial(1, 0.5, n).astype(float)
+    Y = 1.0 * D + 0.5 * X + rng.normal(0, 0.3, n)
+    return pd.DataFrame({'y': Y, 'd': D, 'x': X})
+
+
+def test_compare_estimators_accepts_g_computation(binary_dgp):
+    comp = sp.compare_estimators(
+        data=binary_dgp, y='y', treatment='d',
+        methods=['ols', 'g_computation'],
+        covariates=['x'],
+    )
+    # Both methods ran (agreement diagnostics populated)
+    assert hasattr(comp, 'summary')
+    txt = comp.summary()
+    assert 'G-computation' in txt
+    assert 'OLS' in txt
+    # Sign agreement 100 % on a linear, binary-D, positive-effect DGP.
+    assert 'Sign agreement:        100%' in txt
+
+
+def test_compare_estimators_g_computation_recovers_truth(binary_dgp):
+    """Both OLS and g_computation should hug the true effect ≈ 1.0."""
+    comp = sp.compare_estimators(
+        data=binary_dgp, y='y', treatment='d',
+        methods=['g_computation'],
+        covariates=['x'],
+    )
+    r = comp.results['G-computation']
+    assert abs(r.estimate - 1.0) < 0.15
+
+
+def test_compare_estimators_unsupported_method_does_not_crash(binary_dgp):
+    """
+    Sprint-B methods with extra required args (proximal, msm, ...) are
+    documented as NOT supported. Requesting one must not silently
+    drop the supported methods too.
+
+    ``compare_estimators`` wraps each method in its own try/except +
+    ``warnings.warn`` (see src/statspai/smart/compare.py), so the call
+    itself MUST NOT raise to the caller. The contract we're pinning:
+    the supported method (OLS) still appears in ``.results`` even
+    when an unsupported method ('proximal') sits beside it in the
+    ``methods=`` list.
+    """
+    import warnings
+    with warnings.catch_warnings():
+        # compare_estimators emits UserWarning for the unsupported
+        # branch; suppress it so the test output is clean while still
+        # exercising the real path.
+        warnings.simplefilter('ignore')
+        comp = sp.compare_estimators(
+            data=binary_dgp, y='y', treatment='d',
+            methods=['ols', 'proximal'],   # 'proximal' unsupported
+            covariates=['x'],
+        )
+    # OLS must still be present regardless of what the unsupported
+    # method did. No permissive try/except — a true regression
+    # (OLS silently disappearing) should fail this assertion.
+    assert 'OLS' in comp.results
+    # 'proximal' must NOT appear (never dispatched)
+    assert 'Proximal' not in comp.results
+
+
+# ---------------------------------------------------------------------
+# sensitivity_dashboard — method label now reads from .method first
[…]
