### 2026-04-21 577f66d
feat: v0.9.11 — multi-instrument MTE + true CHV-2011 PRTE weights
Closes two long-standing API gaps and an empirical math debt.

Added:
- sp.bayes_mte(instrument: str | Sequence[str], ...) — multi-IV
  support matching sp.bayes_iv / sp.bayes_hte_iv. Scalar calls
  unchanged. pi_Z ~ Normal(shape=k_iv) in joint mode; post-hoc
  propensity recomputation handles (k_iv,) coefficient vector.
- sp.policy_weight_observed_prte(propensity_sample, shift) — true
  CHV-2011 PRTE weight via gaussian_kde.integrate_box_1d — the
  exact CDF-difference form [F_P(u) - F_P(u-Δ)]/Δ, closing the
  v0.9.9 docstring 
[…]
