+++ src/statspai/bayes/policy_weights.py
@@ -159,9 +159,125 @@ def policy_weight_marginal(
     return _w
 
 
+def policy_weight_observed_prte(
+    propensity_sample: np.ndarray,
+    shift: float,
+    *,
+    bw_method=None,
+) -> Callable[[np.ndarray], np.ndarray]:
+    """True **CHV-2011 PRTE** weights from the observed propensity
+    distribution via Gaussian KDE.
+
+    Implements the policy-relevant treatment effect weighting
+    (Carneiro-Heckman-Vytlacil 2011, Theorem 1):
+
+    .. code-block:: text
+
+        w(u) ∝ [F_P(u) - F_{P + Δ}(u)] / Δ
+             = [F_P(u) - F_P(u - Δ)] / Δ
+             = ∫_{u - Δ}^{u} f_P(s) ds / Δ
+
+    where ``F_P`` is the CDF of the observed propensity sample, ``f_P``
+    is its density (Gaussian KDE here), and ``Δ = shift`` is the
+    scalar propensity-scale policy shift. Intuition: ``w(u)`` is the
+    population density of policy-induced compliers at propensity
+    level ``u``.
+
+    Compared to :func:`policy_weight_prte` (a stylised rectangle),
+    this uses the *actual* sample distribution of propensity, which is
+    what CHV 2011 describes as the correct weighting kernel.
+
+    Parameters
+    ----------
+    propensity_sample : np.ndarray
+        1-D array of observed propensity scores in ``[0, 1]``.
+        Typical source: ``sp.bayes_mte(...).model_info['propensity']``
+        or a direct logit fit on your sample.
+    shift : float
+        Policy-scale propensity shift. Must be non-zero and in
+        ``(-1, 1)``. Positive = expand treatment uptake by ``shift``
+        at every propensity level; negative = contraction.
+    bw_method : str | float | callable | None, default ``None``
+        Passed to :class:`scipy.stats.gaussian_kde`. ``None`` uses
+        Scott's rule, which is a good default for smooth propensity
+        densities.
+
+    Returns
+    -------
+    callable
+        A function ``weight_fn(u: np.ndarray) -> np.ndarray`` that
+        returns the (non-negative) PRTE weight at each grid point.
+        Negative-density-difference values are clipped at 0 — the
+        integral ``∫ w(u) MTE(u) du`` is ill-defined for negative
+        weights and clipping aligns with the CHV-2011 interpretation
+        when the kernel returns slightly negative tail values due to
+        the shift placing density outside ``[0, 1]``.
+
+    References
+    ----------
+    Carneiro, P., Heckman, J. J., & Vytlacil, E. J. (2011).
+    Estimating marginal returns to education. *AER*, 101(6),
+    2754-2781.
+    """
+    from scipy.stats import gaussian_kde
+
+    sample = np.asarray(propensity_sample, dtype=float).ravel()
+    if sample.size < 2:
+        raise ValueError(
+            "propensity_sample must contain at least 2 values to "
+            "estimate a density; got size "
+            f"{sample.size}."
+        )
+    if np.any((sample < 0.0) | (sample > 1.0)):
+        raise ValueError(
+            "propensity_sample values must all lie in [0, 1]."
+        )
+    if not (-1.0 < shift < 1.0):
+        raise ValueError(
+            f"shift must be in (-1, 1); got {shift}"
+        )
+    if shift == 0.0:
+        raise ValueError(
+            "shift must be non-zero; shift=0 yields zero-weight "
+            "grids."
+        )
+
+    kde = gaussian_kde(sample, bw_method=bw_method)
[…]
