### 2026-04-21 1a8d89d
feat(smart): compare_estimators method_hints + sensitivity Sprint-B dims + g_computation error compaction
Three smart-layer enhancements closing the remaining Sprint-B
ergonomic gaps, plus five fixes from a fresh-eyes code review.

1. sp.compare_estimators(method_hints={...})
   New kwarg forwards per-method kwargs to each estimator when the
   shared (data, y, treatment, covariates, id, time, instrument)
   signature is too narrow. Enables six Sprint-B methods in batch
   comparison:
     - proximal            — needs {proxy_z, proxy_w}
     - msm                 — needs {time_varying} + id +
[…]
