+++ tests/test_smart_tools_sprint_b_round4.py
@@ -1,0 +1,353 @@
+"""
+Round-4 Smart-tool extensions:
+
+1. compare_estimators(method_hints={...}) routes proximal / msm /
+   principal_strat / mediate / mediate_interventional / front_door
+   using per-method kwargs, with the ROADMAP §6 collision rule
+   (hint wins + UserWarning on conflict).
+
+2. sensitivity_dashboard gains Sprint-B-aware dimensions:
+   'first_stage_f' (proximal), 'trim_sweep' (msm),
+   'monotonicity' (principal_strat).
+
+3. sp.g_computation rejects non-binary D with a COMPACT message
+   (truncated value list), not a 400-float dump.
+"""
+
+import warnings
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+
+
+# ---------------------------------------------------------------------
+# C) g_computation compact error message
+# ---------------------------------------------------------------------
+
+def test_g_computation_binary_error_is_compact():
+    rng = np.random.default_rng(0)
+    n = 400
+    df = pd.DataFrame({
+        'y': rng.normal(0, 1, n),
+        'd': rng.normal(0, 1, n),   # continuous
+        'x': rng.normal(0, 1, n),
+    })
+    with pytest.raises(ValueError) as exc_info:
+        sp.g_computation(df, y='y', treat='d', covariates=['x'], n_boot=5)
+    msg = str(exc_info.value)
+    # Old behaviour dumped all 400 unique floats into the error
+    # (~20 KB). The fixed version truncates to 5 values + an
+    # "(N more)" suffix. Keep the cap generous enough for the suffix
+    # but tight enough to catch the regression.
+    assert len(msg) < 400, (
+        f'error message is {len(msg)} chars; expected <400. Got: {msg}'
+    )
+    # Still actionable
+    assert 'binary' in msg.lower()
+    assert 'dose_response' in msg
+    # Preview truncation suffix
+    assert 'more' in msg
+
+
+# ---------------------------------------------------------------------
+# B) sensitivity_dashboard Sprint-B-aware dimensions
+# ---------------------------------------------------------------------
+
+@pytest.fixture
+def proximal_dgp():
+    rng = np.random.default_rng(0)
+    n = 500
+    U = rng.normal(0, 1, n)
+    D = 0.8 * U + rng.normal(0, 0.5, n)
+    Z = 0.9 * U + rng.normal(0, 0.3, n)
+    W = 0.9 * U + rng.normal(0, 0.3, n)
+    Y = 1.5 * D + 0.7 * U + rng.normal(0, 0.3, n)
+    return pd.DataFrame({'y': Y, 'd': D, 'z': Z, 'w': W})
+
+
+def test_sensitivity_dashboard_proximal_first_stage_f_dimension(proximal_dgp):
+    r = sp.proximal(proximal_dgp, y='y', treat='d',
+                    proxy_z=['z'], proxy_w=['w'])
+    dash = sp.sensitivity_dashboard(
+        r, data=proximal_dgp, verbose=False,
+        dimensions=['first_stage_f'],
+    )
+    dims = [d['dimension'] for d in dash.dimensions]
+    assert 'Proximal first-stage F (weak-IV)' in dims
+    # Strong proxies → should be 'stable'
+    fs_dim = next(d for d in dash.dimensions
+                  if d['dimension'] == 'Proximal first-stage F (weak-IV)')
+    assert fs_dim['stable'] is True
+
+
+def test_sensitivity_dashboard_proximal_weak_F_flagged_unstable():
+    rng = np.random.default_rng(1)
+    n = 500
+    U = rng.normal(0, 1, n)
+    D = 0.8 * U + rng.normal(0, 0.5, n)
[…]
