+++ src/statspai/smart/sensitivity.py
@@ -97,10 +97,28 @@ def sensitivity_dashboard(
     >>> dash = sp.sensitivity_dashboard(result, data=df)
     >>> print(dash.summary())
     """
-    # Extract baseline
-    if hasattr(result, 'estimate'):
-        baseline_est = result.estimate
-        baseline_se = result.se
+    # Extract baseline. Order: (1) CausalResult-like `.estimate` /
+    # `.se`, (2) EconometricResults `.params` / `.std_errors`,
+    # (3) PrincipalStratResult (which has neither) — pull the top
+    # row of `.effects` instead.
+    if hasattr(result, 'estimate') and not isinstance(
+        getattr(result, 'estimate', None), pd.Series
+    ):
+        baseline_est = float(result.estimate)
+        baseline_se = float(getattr(result, 'se', 0.0))
+    elif type(result).__name__ == 'PrincipalStratResult':
+        # Use the complier row explicitly (LATE) rather than
+        # .iloc[0], which depends on the effects DataFrame ordering
+        # and would silently select an always-taker effect if the
+        # table is ever sorted differently upstream.
+        effects = getattr(result, 'effects', None)
+        if effects is not None and len(effects):
+            _complier_mask = effects['stratum'].astype(str).str.lower().str.contains('complier')
+            _row = effects[_complier_mask].iloc[0] if _complier_mask.any() else effects.iloc[0]
+            baseline_est = float(_row['estimate'])
+            baseline_se = float(_row['se']) if 'se' in effects.columns else 0.0
+        else:
+            baseline_est, baseline_se = 0.0, 0.0
     elif hasattr(result, 'params') and len(result.params) > 1:
         # Use first non-constant coefficient
         non_const = [k for k in result.params.index if k != '_cons']
@@ -129,17 +147,39 @@ def sensitivity_dashboard(
     # Inference (linear 2SLS)"); older EconometricResults use
     # ``model_info['model_type']``. Read both so the dashboard
     # doesn't show "Unknown" on proximal / msm / g_computation / etc.
+    # PrincipalStratResult likewise has a ``.method`` ('monotonicity'
+    # or 'principal_score') plus an 'estimator' label in model_info;
+    # prefer the latter because it's more descriptive.
     _model_info = getattr(result, 'model_info', {}) or {}
-    method = (
-        str(getattr(result, 'method', '') or '')
-        or str(_model_info.get('model_type', '') or '')
-        or str(_model_info.get('estimator', '') or '')
-        or 'Unknown'
-    )
+    if type(result).__name__ == 'PrincipalStratResult':
+        # PrincipalStratResult: prefer the verbose estimator label.
+        method = (
+            str(_model_info.get('estimator', '') or '')
+            or 'Principal Stratification'
+        )
+    else:
+        method = (
+            str(getattr(result, 'method', '') or '')
+            or str(_model_info.get('model_type', '') or '')
+            or str(_model_info.get('estimator', '') or '')
+            or 'Unknown'
+        )
     dim_results = []
 
     if dimensions is None:
         dimensions = ['sample', 'outliers', 'unobservables']
+        # Auto-append Sprint-B-specific dimensions when the result
+        # type matches. Users who pass an explicit ``dimensions=``
+        # list opt out of this expansion; the explicit list is taken
+        # verbatim.
+        _ml_lower = method.lower()
+        if 'proximal' in _ml_lower:
+            dimensions.append('first_stage_f')
+        if 'marginal structural' in _ml_lower:
+            dimensions.append('trim_sweep')
+        if (type(result).__name__ == 'PrincipalStratResult'
+                or 'principal' in _ml_lower):
+            dimensions.append('monotonicity')
 
     if data is not None:
         n = len(data)
[…]
