### 2026-04-21 a546ead
feat: v0.9.12 — probit-scale MTE (Heckman frame) + statspai CLI + sp.help introspection
- bayes_mte gains selection='uniform'|'normal' kwarg; 'normal' fits the
  MTE polynomial on V = Φ⁻¹(U_D) so (poly_u=1, hv_latent, normal)
  exactly recovers the linear Heckman MTE slope on a bivariate-normal DGP.
- Shared PROBIT_CLIP in bayes/_base keeps fit / ATT-ATU integrator /
  policy_effect on the same numerical footing; BayesianMTEResult carries
  the selection tag so policy_effect transforms the grid to V scale.
- mte_curve adds a 'v' column under selection='normal'.
- Round-C review fixed 2 BLOCKER
[…]
