+++ tests/test_help.py
@@ -1,0 +1,235 @@
+"""Tests for sp.help() unified entry point and CLI."""
+from __future__ import annotations
+
+import io
+import json
+import sys
+
+import pytest
+
+
+# --------------------------------------------------------------------------- #
+#  sp.help() — Python API
+# --------------------------------------------------------------------------- #
+
+class TestTopLevelHelp:
+    def test_help_no_args_returns_overview(self):
+        import statspai as sp
+        r = sp.help()
+        # HelpResult prints as text; should mention the package and version.
+        text = str(r)
+        assert "StatsPAI" in text
+        assert sp.__version__ in text
+        assert "CATEGORIES" in text
+        assert "sp.help" in text
+
+    def test_help_overview_dict_format(self):
+        import statspai as sp
+        d = sp.help(format="dict")
+        assert d["kind"] == "overview"
+        assert d["version"] == sp.__version__
+        assert isinstance(d["categories"], dict)
+        assert sum(d["categories"].values()) >= 40  # at least hand-written specs
+
+    def test_help_bad_format_raises(self):
+        import statspai as sp
+        with pytest.raises(ValueError):
+            sp.help(format="xml")
+
+
+class TestFunctionLookup:
+    def test_help_by_function_name(self):
+        import statspai as sp
+        r = sp.help("did")
+        text = str(r)
+        assert "sp.did" in text
+        assert "causal" in text
+
+    def test_help_function_dict_format(self):
+        import statspai as sp
+        d = sp.help("regress", format="dict")
+        assert d["name"] == "regress"
+        assert d["category"] == "regression"
+        assert any(p["name"] == "formula" for p in d["params"])
+
+    def test_help_scoped_category_dot_name(self):
+        import statspai as sp
+        r = sp.help("causal.did")
+        assert "sp.did" in str(r)
+
+    def test_help_callable_object(self):
+        import statspai as sp
+        r = sp.help(sp.regress)
+        text = str(r)
+        # Should fall through to registry detail (has one).
+        assert "sp.regress" in text or "regress" in text
+
+
+class TestCategoryListing:
+    def test_help_by_category_name(self):
+        import statspai as sp
+        r = sp.help("causal")
+        text = str(r)
+        assert "Category: causal" in text
+        assert "did" in text
+
+
+class TestSearch:
+    def test_help_search_keyword(self):
+        import statspai as sp
+        r = sp.help(search="treatment")
+        text = str(r)
+        assert "Search:" in text
+        assert "match" in text
+
+    def test_help_search_dict_format(self):
+        import statspai as sp
+        d = sp.help(search="synthetic control", format="dict")
+        assert d["query"] == "synthetic control"
+        assert isinstance(d["results"], list)
+
+
+class TestNotFound:
+    def test_help_unknown_topic_returns_suggestions(self):
+        import statspai as sp
+        r = sp.help("diid")  # typo
+        text = str(r)
+        assert "No match" in text
+        # Should suggest 'did'
+        assert "did" in text.lower()
+
+    def test_help_unknown_dict_format(self):
+        import statspai as sp
+        d = sp.help("totallynosuchthing", format="dict")
+        assert d["kind"] == "not_found"
+        assert isinstance(d["function_suggestions"], list)
+
+    def test_help_bad_topic_type_raises(self):
+        import statspai as sp
+        with pytest.raises(TypeError):
+            sp.help(42)
+
+
+# --------------------------------------------------------------------------- #
+#  Registry auto-registration
+# --------------------------------------------------------------------------- #
+
+class TestRegistryExpansion:
+    def test_list_functions_covers_more_than_hand_written(self):
+        import statspai as sp
[…]
