+++ src/statspai/help.py
@@ -1,0 +1,511 @@
+"""Unified help system for StatsPAI.
+
+Aggregates four help layers:
+    1. Registry metadata (sp.describe_function, sp.search_functions)
+    2. Python docstrings (help(sp.xxx))
+    3. Smart recommender pointers (sp.recommend, sp.compare_estimators)
+    4. Submodule method catalogs (sp.decomposition.available_methods)
+
+Usage
+-----
+>>> import statspai as sp
+>>> sp.help()                       # top-level overview
+>>> sp.help('did')                  # function detail
+>>> sp.help('causal')               # category listing
+>>> sp.help('causal.did')           # scoped detail
+>>> sp.help(sp.regress)             # docstring fallback
+>>> sp.help(search='treatment')     # keyword search
+>>> sp.help('did', format='dict')   # programmatic access
+"""
+from __future__ import annotations
+
+import inspect
+import textwrap
+from typing import Any, Dict, List, Optional, Union
+
+
+# ====================================================================== #
+#  Category index — groups of related functions for hierarchical help.
+# ====================================================================== #
+#
+# NOTE: kept in sync with registry categories, but additionally covers
+# areas the registry does not yet hand-register.  Used by sp.help()
+# category listings and by the auto-registration fallback.
+
+CATEGORY_DESCRIPTIONS: Dict[str, str] = {
+    "regression": "OLS / IV / GLM / quantile / Tobit / Heckman / count / survival",
+    "causal": "DID, RD, Synth, Matching, DML, IPW, Causal Forest, Meta-learners, DAG, g-methods",
+    "panel": "Fixed/random effects, dynamic panel GMM, HDFE, interactive FE, unit roots",
+    "timeseries": "VAR, ARIMA, GARCH, local projections, cointegration, structural breaks",
+    "spatial": "SAR/SEM/SDM/SAC, GWR, spatial panels, LISA, Moran's I, spatial weights",
+    "survey": "svydesign, svymean/glm, raking, calibration",
+    "survival": "Cox, Kaplan-Meier, AFT, frailty, logrank",
+    "decomposition": "Oaxaca, RIF, DFL, Machado-Mata, Kitagawa, Shapley, Gelbach",
+    "diagnostics": "Sensitivity (Oster, sensemakr, E-value), specification tests, heterogeneity",
+    "robustness": "Spec curve, subgroup analysis, sensitivity dashboards",
+    "inference": "Bootstrap, wild cluster, Romano-Wolf, conformal, CR2/CR3, Conley",
+    "output": "outreg2, modelsummary, regtable, paper_tables, coefplot",
+    "plots": "binscatter, event-study plots, theme management",
+    "smart": "Recommender, compare_estimators, assumption_audit, pub_ready, workflow",
+    "power": "Power analysis for RCT/DID/RD/IV/cluster-RCT and MDE",
+    "experimental": "Randomization, balance checks, attrition, optimal design",
+    "missing": "MICE, multiple imputation, mi_estimate",
+    "bayesian": "bayes_did / bayes_rd / bayes_iv / bayes_mte / BCF / policy weights",
+    "postestimation": "margins, contrast, pwcompare, test, lincom",
+    "agent": "LLM tool-definition surface + JSON schema export",
+    "utils": "Labels, winsor, DGP simulators, read_data, describe, pwcorr",
+    "datasets": "Canonical datasets (Prop99, German reunification, CPS wage, ...)",
+}
+
+# Module-path-prefix → category.  First-match wins; order matters for
+# overlapping prefixes (e.g. 'causal_discovery' before 'causal').
+_MODULE_CATEGORY_PREFIXES: List[tuple] = [
[…]

[…]
