### 2026-04-21 2fd88de
feat: v0.9.13 — ArviZ HDI compat shim + ATT/ATU uncertainty
Small-but-load-bearing cleanup release closing two items deferred
across v0.9.10/11/12 code reviews.

Added:
- `_az_hdi_compat(samples, hdi_prob)` in statspai.bayes._base:
  routes every az.hdi() call in the Bayesian sub-package through
  one place so the inevitable arviz >= 0.18 kwarg rename
  (hdi_prob -> prob) is a one-line future change. Try hdi_prob
  first, fall back to prob on TypeError.
- ATT/ATU uncertainty on BayesianMTEResult:
  att_sd/att_hdi_lower/att_hdi_upper + atu_*. posterior_sd covers
  ATE (primary estimand), no red
[…]
