### 2026-04-21 a8fbd1f
feat: v0.9.14 — summary() renders ATT/ATU block (completes v0.9.13 spec §3.3)
Tiny patch release closing the remaining thread on v0.9.13: the six
att_sd / att_hdi_* / atu_sd / atu_hdi_* fields landed but
summary() never printed them. v0.9.14 wires the rendering through a
BayesianMTEResult.summary() override.

Added:
- BayesianMTEResult.summary() override appends a
  'Population-integrated effects' block inside the framing = ruler:

      ATT: 0.2407 (sd 0.0370, 95% HDI [0.1693, 0.3136])
      ATU: 0.2147 (sd 0.0435, 95% HDI [0.1341, 0.2947])

  Silently skipped when SDs are NaN (empty subpopul
[…]
