### 2026-04-21 78a6d08
fix(bartik): proper leave-one-out with regional_shocks panel + fail-loud warning
Pre-fix, `bartik(leave_one_out=True)` (the default!) silently fell
through to the simple Bartik instrument because proper LOO requires
per-region industry growth data that the basic API didn't carry. The
docstring promised "exclude own region from national average" but the
implementation was `pass`. Users relying on LOO got lied to.

Fix:
- Add optional `regional_shocks: pd.DataFrame` (n_units x n_industries)
  parameter. When provided with leave_one_out=True, compute exact
  Borusyak-Hull-Jaravel 2022 LOO: g_k^{-
[…]
