+++ src/statspai/bartik/shift_share.py
@@ -16,6 +16,8 @@ Supports:
 """
 
 from typing import Optional, List, Dict, Any
+import warnings
+
 import numpy as np
 import pandas as pd
 from scipy import stats
@@ -31,6 +33,7 @@ def bartik(
     shocks: pd.Series,
     covariates: Optional[List[str]] = None,
     leave_one_out: bool = True,
+    regional_shocks: Optional[pd.DataFrame] = None,
     robust: str = 'hc1',
     alpha: float = 0.05,
 ) -> EconometricResults:
@@ -53,7 +56,23 @@ def bartik(
     covariates : list of str, optional
         Exogenous control variables.
     leave_one_out : bool, default True
-        Compute leave-one-out shocks (exclude own region from national average).
+        Compute leave-one-out shocks (exclude own region from national
+        average). Only takes effect when ``regional_shocks`` is also
+        supplied — without the per-region industry growth panel there
+        is not enough information to reconstruct ``g_k`` excluding
+        region ``i``. When ``leave_one_out=True`` but
+        ``regional_shocks`` is not provided, a ``UserWarning`` is
+        raised and the estimator falls back to the simple Bartik
+        instrument.
+    regional_shocks : pd.DataFrame, optional
+        Regional industry growth matrix (n_units x n_industries). Row
+        ``i``, column ``k`` is the realised growth of industry ``k``
+        in region ``i``. When provided with ``leave_one_out=True``,
+        the instrument uses
+        ``g_k^{-i} = (sum_j g_{jk} - g_{ik}) / (n - 1)`` (Borusyak-
+        Hull-Jaravel 2022-style exact leave-one-out). Row index must
+        align with ``shares``; columns must be a superset of
+        ``shocks.index``.
     robust : str, default 'hc1'
         Standard error type.
     alpha : float, default 0.05
@@ -74,6 +93,7 @@ def bartik(
     estimator = BartikIV(
         data=data, y=y, endog=endog, shares=shares, shocks=shocks,
         covariates=covariates, leave_one_out=leave_one_out,
+        regional_shocks=regional_shocks,
         robust=robust, alpha=alpha,
     )
     return estimator.fit()
@@ -93,6 +113,7 @@ class BartikIV:
         shocks: pd.Series,
         covariates: Optional[List[str]] = None,
         leave_one_out: bool = True,
+        regional_shocks: Optional[pd.DataFrame] = None,
         robust: str = 'hc1',
         alpha: float = 0.05,
     ):
@@ -103,6 +124,7 @@ class BartikIV:
         self.shocks = shocks
         self.covariates = covariates or []
         self.leave_one_out = leave_one_out
+        self.regional_shocks = regional_shocks
         self.robust = robust
         self.alpha = alpha
 
@@ -127,17 +149,66 @@ class BartikIV:
         self.shares = self.shares[common]
         self.shocks = self.shocks[common]
 
+        if self.regional_shocks is not None:
+            if self.regional_shocks.shape[0] != len(self.data):
+                raise ValueError(
+                    f"regional_shocks has {self.regional_shocks.shape[0]} "
+                    f"rows but data has {len(self.data)} rows"
+                )
+            missing = set(common) - set(self.regional_shocks.columns)
+            if missing:
+                raise ValueError(
+                    "regional_shocks is missing industries present in "
+                    f"shares/shocks: {sorted(missing)}"
+                )
+            self.regional_shocks = self.regional_shocks[common]
+
     def _construct_instrument(self) -> np.ndarray:
-        """Construct the Bartik instrument: B_i = sum_k(s_ik * g_k)."""
+        """Construct the Bartik instrument.
+
[…]
