+++ tests/test_bartik.py
@@ -2,6 +2,8 @@
 Tests for Shift-Share (Bartik) IV module.
 """
 
+import warnings
+
 import pytest
 import numpy as np
 import pandas as pd
@@ -144,6 +146,110 @@ class TestBartikErrors:
         with pytest.raises(ValueError, match="No common"):
             bartik(data, y='y', endog='x_endog',
                    shares=shares, shocks=bad_shocks)
+
+
+class TestBartikLeaveOneOut:
+    """v0.9.13: proper leave-one-out requires `regional_shocks` panel.
+
+    Before v0.9.13 `leave_one_out=True` silently fell through to the
+    simple Bartik instrument. Now it either computes the true LOO
+    instrument or warns the user about the fallback.
+    """
+
+    @pytest.fixture
+    def bartik_loo_data(self, bartik_data):
+        """Attach a regional_shocks panel consistent with bartik_data."""
+        data, shares, shocks = bartik_data
+        rng = np.random.default_rng(2026)
+        # Regional industry growth — national shock + regional noise.
+        # Mean across regions is close to the aggregated shock but
+        # leave-one-out must subtract the region's own contribution.
+        G = shocks.values[None, :] + rng.normal(0, 0.01,
+                                                (len(data), len(shocks)))
+        regional = pd.DataFrame(G, columns=shares.columns)
+        return data, shares, shocks, regional
+
+    def test_loo_warns_when_regional_shocks_missing(self, bartik_data):
+        data, shares, shocks = bartik_data
+        with pytest.warns(UserWarning, match="regional_shocks"):
+            bartik(data, y='y', endog='x_endog',
+                   shares=shares, shocks=shocks, leave_one_out=True)
+
+    def test_loo_silent_when_disabled(self, bartik_data):
+        data, shares, shocks = bartik_data
+        with warnings.catch_warnings():
+            warnings.simplefilter('error', UserWarning)
+            bartik(data, y='y', endog='x_endog',
+                   shares=shares, shocks=shocks, leave_one_out=False)
+
+    def test_loo_uses_panel_when_provided(self, bartik_loo_data):
+        data, shares, shocks, regional = bartik_loo_data
+        with warnings.catch_warnings():
+            warnings.simplefilter('error', UserWarning)
+            # Must NOT warn because regional_shocks covers the LOO path
+            result = bartik(data, y='y', endog='x_endog',
+                            shares=shares, shocks=shocks,
+                            regional_shocks=regional,
+                            leave_one_out=True)
+        assert result is not None
+
+    def test_loo_instrument_differs_from_simple_bartik(self, bartik_loo_data):
+        """B_i^{LOO} should differ from B_i^{simple} when regional
+        growth differs from the aggregated shock."""
+        data, shares, shocks, regional = bartik_loo_data
+        model_simple = BartikIV(data=data, y='y', endog='x_endog',
+                                shares=shares, shocks=shocks,
+                                leave_one_out=False)
+        model_simple._validate()
+        B_simple = model_simple._construct_instrument()
+
+        model_loo = BartikIV(data=data, y='y', endog='x_endog',
+                             shares=shares, shocks=shocks,
+                             regional_shocks=regional,
+                             leave_one_out=True)
+        model_loo._validate()
+        B_loo = model_loo._construct_instrument()
+
+        # They should be close (LOO ≈ national when noise is small)
+        # but not identical.
+        assert not np.allclose(B_simple, B_loo)
+        assert np.corrcoef(B_simple, B_loo)[0, 1] > 0.5
+
+    def test_loo_formula_matches_manual_computation(self, bartik_loo_data):
[…]
