### 2026-04-21 db050de
feat(bayes): multi-term tidy() for BayesianMTEResult (v0.9.15 spec)
v0.9.13 added ATT/ATU posterior SD+HDI on BayesianMTEResult, and
v0.9.14 wired them into summary(). But .tidy() still returned only
the single ATE row, so broom-style meta-analysis pipelines couldn't
pick up ATT/ATU through pd.concat(r.tidy() for r in results). This
closes the gap.

- BayesianMTEResult.tidy(terms=None | str | sequence) override:
  - terms=None → single ATE row (byte-identical to v0.9.14, back-compat).
  - terms='att'/'atu'/'ate' → single row.
  - terms=[...] → multi-row, order preserved.
  - Unknown names rais
[…]
