+++ tests/test_bayes_mte_tidy.py
@@ -1,0 +1,203 @@
+"""Tests for v0.9.15 BayesianMTEResult.tidy(terms=...) multi-term output."""
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+
+pymc = pytest.importorskip(
+    "pymc",
+    reason="PyMC is an optional dependency; skip Bayesian tests if missing.",
+)
+
+from statspai.bayes import bayes_mte
+from statspai.bayes._base import BayesianMTEResult
+
+
+def _hv_dgp(n, slope, seed):
+    rng = np.random.default_rng(seed)
+    Z = rng.normal(size=n)
+    U_D = rng.uniform(0, 1, size=n)
+    p = 1.0 / (1.0 + np.exp(-0.8 * Z))
+    D = (p > U_D).astype(float)
+    Y = 1.0 + slope * U_D * D + 0.3 * rng.normal(size=n)
+    return pd.DataFrame({'y': Y, 'd': D, 'z': Z})
+
+
+@pytest.fixture
+def fit_result():
+    df = _hv_dgp(300, slope=1.5, seed=11)
+    return bayes_mte(df, y='y', treat='d', instrument='z',
+                     mte_method='hv_latent', poly_u=1,
+                     draws=200, tune=200, chains=2, progressbar=False,
+                     random_state=11)
+
+
+EXPECTED_COLUMNS = {
+    'term', 'estimate', 'std_error', 'statistic', 'p_value',
+    'conf_low', 'conf_high', 'prob_positive', 'hdi_prob',
+}
+
+
+# ---------------------------------------------------------------------------
+# Back-compat: default path unchanged
+# ---------------------------------------------------------------------------
+
+
+def test_tidy_default_is_single_row_ate(fit_result):
+    df = fit_result.tidy()
+    assert isinstance(df, pd.DataFrame)
+    assert len(df) == 1
+
+
+def test_tidy_default_schema_unchanged(fit_result):
+    df = fit_result.tidy()
+    assert set(df.columns) == EXPECTED_COLUMNS
+
+
+# ---------------------------------------------------------------------------
+# Single-term paths
+# ---------------------------------------------------------------------------
+
+
+def test_tidy_ate_single_row(fit_result):
+    """Round-B review-H1 fix: ``tidy(terms='ate')`` emits the same
+    long label as the parent default (``estimand.lower()``), so a
+    user mixing ``tidy()`` and ``tidy(terms='ate')`` calls inside a
+    concat pipeline gets uniform ``term`` values."""
+    df = fit_result.tidy(terms='ate')
+    assert len(df) == 1
+    assert df['term'].iloc[0] == fit_result.estimand.lower()
+
+
+def test_tidy_ate_matches_default_path(fit_result):
+    """terms='ate' must be byte-identical to the default tidy() call."""
+    default_row = fit_result.tidy().iloc[0].to_dict()
+    explicit_row = fit_result.tidy(terms='ate').iloc[0].to_dict()
+    # Same dict keys
+    assert set(default_row.keys()) == set(explicit_row.keys())
+    # Same term label
+    assert default_row['term'] == explicit_row['term']
+    # Same numeric values (NaN-safe comparison)
+    for k in ('estimate', 'std_error', 'conf_low', 'conf_high',
+              'prob_positive', 'hdi_prob'):
+        a, b = default_row[k], explicit_row[k]
+        if np.isnan(a) and np.isnan(b):
+            continue
+        assert a == b, f"{k}: default={a} vs explicit={b}"
+
+
+def test_tidy_att_single_row(fit_result):
+    df = fit_result.tidy(terms='att')
+    assert len(df) == 1
+    assert df['term'].iloc[0] == 'att'
[…]
