### 2026-04-21 6d59fe8
feat: v0.9.15 — BayesianMTEResult.tidy(terms=[...]) multi-row output
Completes the broom-pipeline integration of v0.9.13's per-population
ATT/ATU uncertainty. Users can now pd.concat ATE/ATT/ATU rows across
fits in a single call.

Added:
- BayesianMTEResult.tidy(conf_level=None, terms=None) override.
  terms=None (default) unchanged → single ATE row. Single-string
  terms='ate'/'att'/'atu' → single row of that term. List like
  ['ate', 'att', 'atu'] → multi-row DataFrame with 9-column schema.
  Invalid term names raise ValueError.

- BayesianMTEResult now carries att_prob_positive and
  atu_pr
[…]
