### 2026-04-21 012e4de
feat: v0.9.16 — breadth expansion (new-methods pass across 15 sub-packages)
Ships 30+ new modules across 15 sub-packages plus 7 dedicated test
files, bumping StatsPAI's coverage of the applied-causal-inference
landscape closer to v1.0. This is a breadth release in the spirit
of the earlier v1.0 upgrade pass (see memory project_v1_upgrade),
not an architectural change.

Spec: docs/superpowers/specs/2026-04-20-v0916-breadth-expansion.md

New estimators and analyses by area
  Causal forests:      multi_arm_forest, iv_forest,
                       survival/causal_forest (Cui-Kosorok 2023)
  Causa
[…]
