### 2026-04-21 b07b68d
feat(registry): wire v0.9.16 breadth-expansion modules into sp.help / sp.list_functions
Adds Sprint 1-6 new public APIs to sp.__all__ so auto-registration
picks them up, plus 11 hand-written FunctionSpec entries carrying
parameter schemas, tags, and canonical references:

  target_trial_protocol   — Hernan-Robins / JAMA 2022 TTE framework
  clone_censor_weight     — CCW for sustained-treatment strategies
  ipcw                    — Robins-Finkelstein censoring weights
  identify                — Shpitser-Pearl ID algorithm on DAGs
  swig                    — Richardson-Robins SWIG construction
[…]
