+++ src/statspai/registry.py
@@ -884,6 +884,246 @@ def _build_registry():
         reference="Frangakis & Rubin (2002); Zhang & Rubin (2003); Ding & Lu (2017)",
     ))
 
+    # -- v0.9.16 breadth-expansion: Target Trial Emulation ----------- #
+    register(FunctionSpec(
+        name="target_trial_protocol",
+        category="target_trial",
+        description=(
+            "Create a 7-component target trial protocol (Hernan-Robins / "
+            "JAMA 2022 framework). Formalizes eligibility, treatment "
+            "strategies, time zero, follow-up, outcome, causal contrast, "
+            "and analysis plan before any estimation."
+        ),
+        params=[
+            ParamSpec("eligibility", "str | list | callable", True),
+            ParamSpec("treatment_strategies", "list", True),
+            ParamSpec("assignment", "str", True,
+                      description="'randomization' or 'observational emulation'"),
+            ParamSpec("time_zero", "str", True),
+            ParamSpec("followup_end", "str", True),
+            ParamSpec("outcome", "str", True),
+            ParamSpec("causal_contrast", "str", False, "ITT",
+                      enum=["ITT", "per-protocol", "as-treated", "observational-analogue"]),
+            ParamSpec("analysis_plan", "str", False),
+            ParamSpec("baseline_covariates", "list", False),
+            ParamSpec("time_varying_covariates", "list", False),
+        ],
+        returns="TargetTrialProtocol",
+        example='proto = sp.target_trial_protocol(eligibility="age >= 50", ...)',
+        tags=["target_trial", "epidemiology", "observational", "JAMA"],
+        reference="Hernan & Robins (2016); JAMA (2022)",
+    ))
+    register(FunctionSpec(
+        name="clone_censor_weight",
+        category="target_trial",
+        description=(
+            "Clone-Censor-Weight (CCW) for sustained-treatment target "
+            "trials. Clones each subject per strategy, artificially "
+            "censors on deviation, and re-weights via IPCW."
+        ),
+        params=[
+            ParamSpec("data", "DataFrame", True),
+            ParamSpec("id_col", "str", True),
+            ParamSpec("time_col", "str", True),
+            ParamSpec("treatment_col", "str", True),
+            ParamSpec("strategies", "dict[str, callable]", True),
+            ParamSpec("censor_covariates", "list", False),
+            ParamSpec("stabilize", "bool", False, True),
+        ],
+        returns="CloneCensorWeightResult",
+        tags=["target_trial", "ccw", "longitudinal", "dynamic_strategy"],
+        reference="Cain et al. 2010; Hernan et al. 2016",
+    ))
+    register(FunctionSpec(
+        name="ipcw",
+        category="censoring",
+        description=(
+            "Inverse Probability of Censoring Weights -- corrects for "
+            "informative censoring under conditional independent "
+            "censoring given covariates."
+        ),
+        params=[
+            ParamSpec("data", "DataFrame", True),
+            ParamSpec("time", "str", True),
+            ParamSpec("event", "str", True),
+            ParamSpec("censor_covariates", "list", True),
+            ParamSpec("treatment_covariates", "list", False),
+            ParamSpec("stabilize", "bool", False, True),
+            ParamSpec("method", "str", False, "pooled_logistic",
+                      enum=["pooled_logistic", "cox_ph"]),
+            ParamSpec("truncate", "tuple", False, (0.01, 0.99)),
+        ],
+        returns="IPCWResult",
+        tags=["censoring", "weighting", "survival", "What If"],
+        reference="Robins & Finkelstein (2000); Cole & Hernan (2008)",
+    ))
+
[…]
