### 2026-04-21 34189cf
feat: v0.9.16 — bayes_did(cohort=...), bayes_iv(per_instrument=True), Rust Phase-2 CI scaffold
Wires the BayesianDIDResult / BayesianIVResult classes (already in
_base.py) into the DID and IV entry points, adds per-cohort τ
posteriors with treated-size-weighted pooled ATT, adds per-instrument
just-identified LATE sub-fits, and lands an inert cibuildwheel matrix
that activates on feat/rust-hdfe. Full back-compat when cohort= and
per_instrument= are omitted.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
