### 2026-04-21 85a6dd8
docs(changelog): merge breadth-expansion into v0.9.16 entry
The v0.9.16 CHANGELOG section had been drafted for the Bayesian
family work (bayes_did cohort, bayes_iv per_instrument, Rust Phase-2
CI scaffold) before the breadth-expansion commit 012e4de landed.
This adds the 27+ breadth modules (target_trial, DAG SCM machinery,
transport, ICP/FCI, OPE, CEVAE, parametric g-formula, LTMLE,
Q/A-learning, multi_arm_forest, iv_forest, survival causal forest,
negative controls, network_exposure, genmatch, rosenbaum_bounds,
spatial DiD/IV, ITS, balke_pearl, four_way_decomposition, ...) +
registry wiring 
[…]
