+++ CHANGELOG.md
@@ -2,12 +2,95 @@
 
 All notable changes to StatsPAI will be documented in this file.
 
-## [0.9.16] - 2026-04-20 — Textbook Heckman HV + multi-term tidy() for DID/IV + Rust Phase-2 CI
+## [0.9.16] - 2026-04-20 — v1.0 breadth expansion + Bayesian family polish + Rust Phase-2 CI
 
-Three additions that close long-standing gaps in the Bayesian
-family, plus a CI scaffold for the Rust HDFE spike.
+The largest release since the v1.0 breadth pass. Maps StatsPAI onto
+the full Mixtape + What If + Elements of Causal Inference curriculum:
+Hernan-Robins target-trial emulation, Pearl-Bareinboim SCM machinery,
+modern off-policy / neural-causal estimators, plus three additions
+that close long-standing gaps in the Bayesian family, plus a CI
+scaffold for the Rust HDFE spike.
 
-### Added (0.9.16)
+### Added (0.9.16) — v1.0 breadth expansion (27+ new modules)
+
+**Target trial emulation & censoring (`sp.target_trial`, `sp.ipcw`)**
+
+- `target_trial_protocol`, `target_trial_emulate`, `clone_censor_weight`,
+  `immortal_time_check` — JAMA 2022 7-component TTE framework with
+  explicit eligibility / time-zero / per-protocol contrast support.
+- `ipcw` — Robins-Finkelstein inverse probability of censoring weights
+  (pooled-logistic or Cox hazard) with stabilization + truncation.
+
+**SCM / DAG machinery (`sp.dag` extended)**
+
+- `identify` — Shpitser-Pearl ID algorithm; returns do-free estimand
+  when identifiable, witness hedge `(F, F')` otherwise.
+- `do_rule1 / do_rule2 / do_rule3`, `do_calculus_apply` — mechanized
+  do-calculus with d-separation on mutilated graphs `G_{bar X}`,
+  `G_{underline Z}`, and `G_{bar Z(W)}`.
+- `swig` — Richardson-Robins Single-World Intervention Graphs via
+  node-splitting of intervened variables.
+- `SCM` — abduction-action-prediction counterfactual runner with
+  rejection sampling fallback for non-Gaussian structural equations.
+- `llm_dag` — LLM-backed DAG extraction from free-form descriptions.
+
+**Causal discovery with latents (`sp.causal_discovery`)**
+
+- `fci` — FCI for PAGs with unobserved confounders (Zhang 2008):
+  skeleton + v-structures + FCI rules R1-R4.
+- `icp`, `nonlinear_icp` — Peters-Bühlmann-Meinshausen invariant
+  causal prediction; linear F-test / K-S nonlinear invariance.
+
+**Transportability (`sp.transport`)**
+
+- `transport_weights_fn` / `transport_generalize` — Stuart / Dahabreh
+  density-ratio transport with inverse odds of sampling weighting.
+- `identify_transport` — Bareinboim-Pearl s-admissibility; enumerates
+  adjustment sets on selection diagrams, returns transport formula.
+
+**Off-policy evaluation (`sp.ope`)**
+
+- `ips`, `snips`, `doubly_robust`, `switch_dr`, `direct_method`,
+  `evaluate` — Dudik-Langford-Li DR family plus Swaminathan-Joachims
+  SNIPS and Wang-Agarwal-Dudík Switch-DR for bandits / RL.
+
+**Deep causal & latent-confounder models (`sp.neural_causal`)**
+
+- `cevae` — Louizos et al. CEVAE with PyTorch path + numpy
+  variational fallback so import never fails.
+
[…]
