### 2026-04-21 ac92e9a
fix(bayes): v0.9.16 self-review — cohort dropna alignment + poly_u sentinel
Two round-2 fixes caught in self-review after v0.9.16 landed:

1. `bayes_did(cohort=...)` used two independent `.dropna()` passes
   (one in `_prepare_did_frame`, one in the cohort factoriser) which
   silently misaligned `cohort_codes` with `DID` whenever the cohort
   column had NaN rows that y/treat/post/unit/time did not. Synthetic
   test data never triggered this; real panels with missing cohort
   info would have either shape-errored in PyMC or, worse, fit with
   shuffled cohort labels. Fix threads `cohort` thr
[…]
