### 2026-04-21 8755996
fix(tests): explicit utf-8 encoding on Path.read_text() — Windows CI
Windows CI test matrix (windows-latest × Python 3.10/3.11/3.12) was
crashing on v0.9.16 with UnicodeDecodeError at
tests/test_workflow_sprint_b.py::test_index_md_has_no_stale_did_guide_link
because Path.read_text() without encoding= uses the platform default
locale, and on Windows that is cp1252/charmap, which cannot decode
bytes >= 0x80 that appear in docs/index.md (CJK / en-dashes / curly
quotes). macOS and Ubuntu default to utf-8 so they passed silently.

Fixes 6 call sites across 5 test files:

  tests/test_workflow_sprin
[…]
