### 2026-04-21 2a5f368
feat: v0.9.17 — weakrobust panel + SBW + MC g-formula + end-to-end sp.causal()
Closes four gaps identified in the v1.0 gap-analysis pass:

* sp.weakrobust — Stata-style unified panel bundling Anderson-Rubin,
  Moreira CLR, Kleibergen K, Kleibergen-Paap rk LM+F, Olea-Pflueger
  effective F, and LMMP tF critical values.
* sp.sbw — Zubizarreta (2015) stable balancing weights (variance /
  entropy objective, ATT/ATC/ATE, SLSQP solver, balance table, ESS).
* sp.gformula_mc — Robins (1986) Monte-Carlo parametric g-formula
  with static or dynamic treatment strategies; complements the
  existing Bang
[…]
