+++ tests/test_v0917_additions.py
@@ -1,0 +1,422 @@
+"""Tests for v0.9.17 additions: weakrobust, sbw, gformula_mc, enhanced causal()."""
+
+from __future__ import annotations
+
+import warnings
+
+import numpy as np
+import pandas as pd
+import pytest
+
+
+# ═══════════════════════════════════════════════════════════════════════
+#  Fixtures
+# ═══════════════════════════════════════════════════════════════════════
+
+@pytest.fixture
+def iv_dataset():
+    rng = np.random.default_rng(0)
+    n = 300
+    z1 = rng.normal(size=n)
+    z2 = rng.normal(size=n)
+    u = rng.normal(size=n)
+    d = 0.6 * z1 + 0.4 * z2 + u + rng.normal(size=n)
+    y = 0.5 * d + u + rng.normal(size=n)
+    return pd.DataFrame({"y": y, "d": d, "z1": z1, "z2": z2})
+
+
+@pytest.fixture
+def obs_dataset():
+    rng = np.random.default_rng(0)
+    n = 400
+    x1 = rng.normal(size=n)
+    x2 = rng.normal(size=n)
+    p = 1 / (1 + np.exp(-(0.3 * x1 - 0.2 * x2)))
+    T = rng.binomial(1, p)
+    Y = 0.5 + 0.8 * T + 0.5 * x1 + 0.3 * x2 + rng.normal(size=n)
+    return pd.DataFrame({"Y": Y, "T": T, "x1": x1, "x2": x2})
+
+
+@pytest.fixture
+def panel_gformula_data():
+    """Two-period longitudinal with time-varying confounding."""
+    rng = np.random.default_rng(0)
+    n = 600
+    L0 = rng.normal(size=n)
+    A0 = (rng.random(n) < 1 / (1 + np.exp(-0.5 * L0))).astype(float)
+    L1 = 0.3 * L0 + 0.4 * A0 + rng.normal(size=n)
+    A1 = (rng.random(n) < 1 / (1 + np.exp(-(0.3 * L0 - 0.2 * A0 + 0.4 * L1)))).astype(float)
+    Y = 0.2 + 0.4 * A0 + 0.6 * A1 + 0.3 * L0 + 0.2 * L1 + rng.normal(size=n) * 0.3
+    return pd.DataFrame({"L0": L0, "A0": A0, "L1": L1, "A1": A1, "Y": Y})
+
+
+# ═══════════════════════════════════════════════════════════════════════
+#  sp.weakrobust — unified weak-IV panel
+# ═══════════════════════════════════════════════════════════════════════
+
+class TestWeakRobust:
+
+    def test_basic_runs(self, iv_dataset):
+        import statspai as sp
+        panel = sp.weakrobust(
+            iv_dataset, y="y", endog="d", instruments=["z1", "z2"],
+            h0=0.0, clr_simulations=2000, grid_size=101, random_state=1,
+        )
+        assert panel.n == 300
+        assert len(panel.instruments) == 2
+
+    def test_contains_all_statistics(self, iv_dataset):
+        import statspai as sp
+        panel = sp.weakrobust(
+            iv_dataset, y="y", endog="d", instruments=["z1", "z2"],
+            clr_simulations=2000, grid_size=101, random_state=1,
+        )
+        frame = panel.to_frame()
+        labels = frame["statistic"].tolist()
+        assert any("First-stage" in l for l in labels)
+        assert any("effective F" in l for l in labels)
+        assert any("Kleibergen–Paap" in l for l in labels)
+        assert any("Anderson–Rubin" in l for l in labels)
+        assert any("CLR" in l for l in labels)
+        assert any("Kleibergen K" in l for l in labels)
+
[…]
