+++ src/statspai/gformula/__init__.py
@@ -21,5 +21,9 @@ Public API
 """
 
 from .ice import ice, gformula_ice, ICEResult
+from .mc import gformula_mc, MCGFormulaResult
 
-__all__ = ["ice", "gformula_ice", "ICEResult"]
+__all__ = [
+    "ice", "gformula_ice", "ICEResult",
+    "gformula_mc", "MCGFormulaResult",
+]
