### 2026-04-21 7888c24
v0.9.17: three-school completion + clinical/MR depth + preregistration + review-fix pass
Three-school completion driven by the "Causal Inference Knowledge Map"
article (Econometrics / Epidemiology / ML), adds the entry-level
primitives, DSL, and QA layer that closed the last gaps vs. Stata / R.

New subpackages
- sp.epi           — OR / RR / RD / AF / IRR / NNT / PR, Mantel-Haenszel
                     (+ Tarone-corrected Breslow-Day), direct + indirect
                     standardization, Bradford-Hill rubric, and the
                     clinical-diagnostics toolkit (sensitivity/
         
[…]
