+++ src/statspai/conformal_causal/counterfactual.py
@@ -1,0 +1,404 @@
+"""
+Conformal intervals for counterfactual outcomes and individual
+treatment effects (Lei & Candès 2021, JRSS-B).
+
+Provides three layers on top of the existing :func:`conformal_cate`:
+
+* :func:`conformal_counterfactual` — prediction intervals for the
+  per-unit potential outcomes ``Y(1) | X`` and ``Y(0) | X`` with
+  finite-sample coverage under covariate-shift-aware weighting.
+* :func:`conformal_ite` — prediction intervals for the individual
+  treatment effect ``τ(x) = Y(1) - Y(0)``, obtained by combining
+  per-arm counterfactual intervals via a Lei-Candès nested bound.
+* :func:`weighted_conformal_prediction` — generic covariate-shift
+  weighted split conformal prediction when the test distribution
+  differs from the training distribution (Tibshirani et al. 2019).
+
+All three are doubly-robust with respect to propensity / outcome
+misspecification and enjoy marginal coverage guarantees.
+
+References
+----------
+Lei, L. & Candès, E.J. (2021). "Conformal inference of
+counterfactuals and individual treatment effects." *JRSS-B*, 83(5),
+911-938.
+
+Tibshirani, R.J., Barber, R.F., Candès, E.J. & Ramdas, A. (2019).
+"Conformal prediction under covariate shift." *NeurIPS*.
+
+Romano, Y., Patterson, E. & Candès, E.J. (2019). "Conformalized
+quantile regression." *NeurIPS*.
+"""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Callable, Optional
+
+import numpy as np
+import pandas as pd
+
+from sklearn.base import BaseEstimator, clone
+from sklearn.ensemble import RandomForestRegressor
+from sklearn.linear_model import LogisticRegression
+
+
+# ═══════════════════════════════════════════════════════════════════════
+#  Result containers
+# ═══════════════════════════════════════════════════════════════════════
+
+@dataclass
+class ConformalCounterfactualResult:
+    """Counterfactual prediction intervals under each potential outcome."""
+
+    X: np.ndarray
+    lower_Y1: np.ndarray
+    upper_Y1: np.ndarray
+    lower_Y0: np.ndarray
+    upper_Y0: np.ndarray
+    alpha: float
+    marginal_coverage_estimate: float
+    method: str = "Lei-Candès-2021-split-CQR"
+
+    def to_frame(self) -> pd.DataFrame:
+        return pd.DataFrame({
+            "Y1_lower": self.lower_Y1,
+            "Y1_upper": self.upper_Y1,
+            "Y0_lower": self.lower_Y0,
+            "Y0_upper": self.upper_Y0,
+        })
+
+    def summary(self) -> str:
+        return (
+            "Conformal counterfactual prediction intervals\n"
+            f"  method              : {self.method}\n"
+            f"  coverage target     : {100*(1-self.alpha):.0f}%\n"
+            f"  empirical coverage  : "
+            f"{100*self.marginal_coverage_estimate:.2f}%  (calibration arm)\n"
+            f"  n test points       : {len(self.X)}\n"
+            f"  mean width Y(1) band: "
+            f"{np.mean(self.upper_Y1 - self.lower_Y1):.4f}\n"
+            f"  mean width Y(0) band: "
+            f"{np.mean(self.upper_Y0 - self.lower_Y0):.4f}"
+        )
+
+    def __repr__(self) -> str:  # pragma: no cover
+        return self.summary()
+
+
+@dataclass
+class ConformalITEResult:
[…]
