### 2026-04-21 9dfaa36
docs(readme): restore image-1.png and fix broken screenshot link
Raw GitHub URL 404'd since fbbc5e8 removed the file. Restore from 37b1245
and switch both READMEs to relative path (./image-1.png).

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
