+++ README_CN.md
@@ -31,7 +31,7 @@ StatsPAI 是一个**面向 AI Agent** 的 Python 因果推断与应用计量经�
 
 **v0.6 新功能**：`sp.interactive(fig)` —— 类似 Stata Graph Editor 的 WYSIWYG 图表编辑器，支持 29 种学术主题、实时预览、自动生成可复现代码。
 
-![StatsPAI 交互式图表编辑器](https://raw.githubusercontent.com/brycewang-stanford/statspai/main/image-1.png)
+![StatsPAI 交互式图表编辑器](./image-1.png)
 
 > 由 [CoPaper.AI](https://copaper.ai) 团队构建 · 斯坦福 REAP 项目
 
