### 2026-04-21 32dd393
v1.0.0: research-frontier capstone — bridging theorems, fairness, surrogates, MVMR, PCMCI, beyond-average QTE
The capstone release. Builds on the v0.9.17 three-school completion
(sp.epi, sp.longitudinal, sp.question, unified sensitivity, DAG
recommender, preregistration) and ships the 2025-2026 research-
frontier modules that Stata / R have not yet caught up with. Every
scaffolded subpackage is now wired into the top-level sp.* namespace
and registered in sp.registry for agent discovery.

Added — v1.0 research-frontier modules

- sp.bridge — 6 bridging theorems (DiD≡SC, EWM≡CATE, CB≡IPW, KinkR
[…]
