### 2026-04-21 05b0933
v1.0.0 post-review fix pass: 3 Critical + 5 High + 3 Medium correctness issues
Independent code-review-expert pass on the v1.0 frontier modules
surfaced 3 Critical + 5 High + 6 Medium + 2 Low findings. This commit
closes every one that represents a real correctness, security, or
shape bug, plus adds 8 pinning regression tests that fail the instant
any of these bugs come back.

Critical fixes

- pcmci Fisher-z SE: use sqrt(n - k - 3) effective sample size, not
  sqrt(df - 1). The old formula was off by one, understating partial-
  correlation p-values and systematically missing edges in PCMCI.

[…]
