+++ src/statspai/ope/__init__.py
@@ -3,7 +3,9 @@ Off-Policy Evaluation (``sp.ope``): estimate the value of a target
 policy from data collected under a different behaviour policy. Covers
 contextual bandits and off-policy reinforcement learning evaluation.
 
-Implemented: DM, IPS, SNIPS, DR, Switch-DR.
+Implemented: DM, IPS, SNIPS, DR, Switch-DR, sharp OPE under
+unobserved confounding (Kallus-Mao-Uehara 2025), causal-policy forest
+(arXiv:2512.22846).
 """
 
 from .estimators import (
@@ -15,6 +17,10 @@ from .estimators import (
     evaluate,
     OPEResult,
 )
+from .sharp_confounding import (
+    sharp_ope_unobserved, causal_policy_forest,
+    SharpOPEResult, CausalPolicyForestResult,
+)
 
 __all__ = [
     "direct_method",
@@ -24,4 +30,6 @@ __all__ = [
     "switch_dr",
     "evaluate",
     "OPEResult",
+    "sharp_ope_unobserved", "causal_policy_forest",
+    "SharpOPEResult", "CausalPolicyForestResult",
 ]
