+++ src/statspai/transport/__init__.py
@@ -17,6 +17,10 @@ Quick start
 from .weighting import transport_weights as weights, TransportWeightResult
 from .generalize import generalize
 from .identify import identify_transport, TransportIdentificationResult
+from .evidence_synthesis import (
+    synthesise_evidence, heterogeneity_of_effect, rwd_rct_concordance,
+    EvidenceSynthesisResult, HeterogeneityResult, ConcordanceResult,
+)
 
 __all__ = [
     "weights",
@@ -24,4 +28,6 @@ __all__ = [
     "generalize",
     "identify_transport",
     "TransportIdentificationResult",
+    "synthesise_evidence", "heterogeneity_of_effect", "rwd_rct_concordance",
+    "EvidenceSynthesisResult", "HeterogeneityResult", "ConcordanceResult",
 ]
