### 2026-04-21 16c1f1f
fix(registry): wire conformal_continuous + conformal_interference
Add two missing registry entries so they surface in sp.list_functions()
and sp.help() alongside the other P1.3 conformal extensions.

Brings visible roadmap functions to 36/36 and total registered to 836.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
