### 2026-04-21 1126f9f
v1.0.1: close NEEDS_VERIFICATION items — Abadie complier QTE + real dual-path PCI bridge
Patch release that turns the two items v1.0.0 shipped with
``# NEEDS_VERIFICATION`` markers into paper-grounded implementations,
with dedicated pinning tests that fail the instant either bug returns.

Fixed — NEEDS_VERIFICATION closeout

- ``beyond_average_late`` (qte/beyond_average.py): replaced the ad-hoc
  ``cdf_diff * (Y95 - Y05)`` rescaling with the Abadie (2002)
  κ-weighted complier-CDF construction.  For each target quantile the
  complier CDFs

    F_{Y1|c}(y) = [P(Y ≤ y, D = 1 | Z = 1) − P(Y ≤ y,
[…]
