### 2026-04-21 e8a7f91
docs: align README / README_CN / docs/index.md to v1.0.1 for PyPI release
Replaces stale v0.9.3 headline banners and citations with v1.0.1
release messaging. Function-count claim bumped 550+ → 800+ to match
the v1.0 registry size. Adds a compact release-highlights history
table (v1.0.1 → v0.9.0) in docs/index.md.

What changed

- README.md and README_CN.md: headline section rewritten around the
  v1.0 stable-API release and the independent code-review pass
  (3 Critical + 5 High + 6 Medium + 2 Low, all fixed and pinned).
- docs/index.md: "Current release" callout updated, release-highlights
  
[…]
