### 2026-04-21 9e4a8f2
fix(ci): add TYPE_CHECKING guard for Path in CausalQuestion.save()
CI flake8 F821 failed because ``"Path"`` was used as a string-forward-
reference return annotation in ``question.py:140`` without the class
being imported. With ``from __future__ import annotations`` the string
evaluates correctly at runtime, but ``flake8 --select=F82`` still
flags it as an undefined name.

Adds a ``TYPE_CHECKING`` guard so ``Path`` is importable for static
analysers without adding a runtime import cost.

Confirmed locally:
  flake8 src/ tests/ --count --select=E9,F63,F7,F82 --show-source
  # → 0

Co-Authored-B
[…]
