+++ src/statspai/rd/rdrobust.py
@@ -48,6 +48,9 @@ def rdrobust(
     donut: float = 0,
     weights: Optional[str] = None,
     alpha: float = 0.05,
+    bootstrap: Optional[str] = None,
+    n_boot: int = 999,
+    random_state: Optional[int] = None,
 ) -> CausalResult:
     """
     Local polynomial RD estimation with robust bias-corrected inference.
@@ -102,12 +105,27 @@ def rdrobust(
         excluded. Useful when manipulation near the cutoff is suspected.
     alpha : float, default 0.05
         Significance level for confidence intervals.
+    bootstrap : {'rbc', None}, default None
+        If ``'rbc'``, augment the output with a robust-bias-corrected
+        percentile bootstrap CI following
+        Cattaneo, Jansson & Ma (arXiv:2512.00566, 2026).  The rbc
+        bootstrap studentises the bias-corrected statistic using the
+        robust variance and resamples observations within the estimation
+        bandwidth.  Empirically produces CIs ~15–20% shorter than the
+        analytic robust CI at the same coverage (Table 3, Cattaneo et al.
+        2026).
+    n_boot : int, default 999
+        Number of bootstrap replicates when ``bootstrap='rbc'``.
+    random_state : int, optional
+        Seed for the rbc bootstrap.
 
     Returns
     -------
     CausalResult
         Results with conventional and robust inference, bandwidth info,
-        and all standard CausalResult methods.
+        and all standard CausalResult methods.  When ``bootstrap='rbc'``
+        the ``model_info`` dict contains a ``'rbc_bootstrap'`` block with
+        the studentised CI and length-ratio vs the analytic robust CI.
 
     Examples
     --------
@@ -142,6 +160,13 @@ def rdrobust(
         raise ValueError(f"deriv must be non-negative, got {deriv}")
     if donut < 0:
         raise ValueError(f"donut must be non-negative, got {donut}")
+    if bootstrap is not None and bootstrap not in ('rbc',):
+        raise ValueError(
+            f"bootstrap must be None or 'rbc', got {bootstrap!r}. "
+            "See Cattaneo, Jansson & Ma (arXiv:2512.00566, 2026)."
+        )
+    if bootstrap is not None and n_boot < 99:
+        raise ValueError("rbc bootstrap needs n_boot >= 99 (recommended 999).")
     # For RKD (deriv >= 1), polynomial order must be at least deriv + 1
     if deriv > 0 and p < deriv + 1:
         p = deriv + 1
@@ -288,12 +313,38 @@ def rdrobust(
         },
     }
 
+    # --- rbc bootstrap (Cattaneo-Jansson-Ma 2026) -----------------------
+    if bootstrap == 'rbc':
+        rbc = _rbc_bootstrap(
+            Y=Y, X_c=X_c, D=D, Z=Z,
+            left=left, right=right,
+            h=h, b=b, p=p, q=q,
+            kernel=kernel, deriv=deriv,
+            cluster_vals=cl_vals_all,
+            alpha=alpha,
+            n_boot=n_boot,
+            random_state=random_state,
+            tau_bc=tau_bc,
+            se_robust=se_robust,
+        )
+        ci_robust_len = ci_robust[1] - ci_robust[0]
+        rbc_len = rbc['ci'][1] - rbc['ci'][0]
+        rbc['length_ratio'] = (
+            float(rbc_len / ci_robust_len) if ci_robust_len > 0 else float('nan')
+        )
+        rbc['n_boot_effective'] = int(rbc.pop('_n_ok'))
+    else:
[…]
