### 2026-04-21 07dfc92
fix(ci): add tabulate to core dependencies
Windows Python 3.10 CI runs failed 4 tests in TestReportGeneration
and test_report_contains_new_sections with:

  ImportError: Missing optional dependency 'tabulate'.

pandas ``.to_markdown()`` dispatches to ``tabulate`` which pandas
lists as optional.  Our ``sp.causal(...).report(fmt='markdown'|'html')``
path uses ``.to_markdown()`` internally, so the dependency has to be
declared core on the StatsPAI side — otherwise a user who does

    r = sp.causal(...).report('markdown')

hits an ImportError at runtime on systems that happen not to pull
tabulate
[…]
