+++ pyproject.toml
@@ -49,6 +49,11 @@ dependencies = [
     "openpyxl>=3.0.0",
     "xlsxwriter>=3.0.0",
     "python-docx>=1.0.0",
+    # Required by ``sp.causal(...).report(fmt='markdown'|'html')`` because
+    # pandas ``.to_markdown()`` dispatches to ``tabulate``.  Ship it as a
+    # core dep so published artifacts work on Linux/macOS/Windows without
+    # users having to discover the optional-dep pitfall at runtime.
+    "tabulate>=0.9.0",
 ]
 
 [project.optional-dependencies]
