### 2026-04-21 d94a219
v1.2.0: doc-alignment sprint — close v3 reference document gaps
Aligns the public API to the *Causal-Inference Method Family 万字剖析 v3*
(2026-04-20) reference document. Most v3 frontier methods were already
implemented in v1.0.x but lived inside sub-packages without top-level
exposure or curated registry specs; this release wires them up, adds the
remaining classical/frontier gaps, and upgrades 14 frontier estimators
from auto-generated to hand-written registry specifications.

Added — new estimators (this overnight pass)

* sp.gardner_did / sp.did_2stage — Gardner (2021) two-stage DID
  (arXiv:
[…]
