+++ src/statspai/registry.py
@@ -2488,6 +2488,621 @@ def _build_registry():
         reference="Imbens, Kallus, Mao (arXiv:2601.17712, 2026).",
     ))
 
+    # ------------------------------------------------------------------
+    # v1.1 additions (doc-alignment sprint — Gardner, Ahrens MA-DML,
+    # Kernel IV, Continuous LATE, HAL-TMLE, Synth Survival, RD aliases)
+    # ------------------------------------------------------------------
+
+    register(FunctionSpec(
+        name="gardner_did",
+        category="causal",
+        description=(
+            "Gardner (2021) two-stage DID. Stage-1 fits two-way FEs on "
+            "untreated observations; Stage-2 regresses the residualised "
+            "outcome on treatment dummies (ATT or event study). Numerically "
+            "close to Borusyak-Jaravel-Spiess imputation with unit-clustered SEs."
+        ),
+        params=[
+            ParamSpec("data", "DataFrame", True),
+            ParamSpec("y", "str", True, description="Outcome column"),
+            ParamSpec("group", "str", True, description="Unit/panel-id column"),
+            ParamSpec("time", "str", True, description="Time column"),
+            ParamSpec("first_treat", "str", True,
+                      description="First-treatment-period column; 0/NaN/inf = never treated"),
+            ParamSpec("controls", "list", False, None, "Additional covariates"),
+            ParamSpec("event_study", "bool", False, False,
+                      "If True, report coefficients by relative time k = t - first_treat"),
+            ParamSpec("horizon", "list", False, None,
+                      "Relative-time leads/lags to report (default range(-5, 6))"),
+            ParamSpec("cluster", "str", False, None,
+                      "Cluster variable for Stage-2 SEs (defaults to group)"),
+            ParamSpec("alpha", "float", False, 0.05),
+        ],
+        returns="CausalResult",
+        example='sp.gardner_did(df, y="wage", group="county", time="year", first_treat="first_treat", event_study=True)',
+        tags=["did", "causal", "staggered", "two-stage", "did2s"],
+        reference="Gardner (2021), arXiv:2207.05943. Butts & Gardner (2022), R Journal 14(3).",
+    ))
+
+    register(FunctionSpec(
+        name="dml_model_averaging",
+        category="causal",
+        description=(
+            "Model-averaging DML (PLR) per Ahrens et al. (2025, JAE). Fits "
+            "DML-PLR under multiple candidate nuisance learners and reports "
+            "a risk-weighted (or equal/single-best) average of their θ "
+            "estimates with a covariance-adjusted SE."
+        ),
+        params=[
+            ParamSpec("data", "DataFrame", True),
+            ParamSpec("y", "str", True, description="Outcome column"),
+            ParamSpec("treat", "str", True, description="Treatment column"),
+            ParamSpec("covariates", "list", True, description="Covariate columns X"),
+            ParamSpec("candidates", "list", False, None,
+                      "List of (ml_g, ml_m, label) sklearn triples; defaults to Lasso/Ridge/RF/GBM"),
+            ParamSpec("n_folds", "int", False, 5),
+            ParamSpec("seed", "int", False, 0),
+            ParamSpec("weight_rule", "str", False, "inverse_risk",
+                      "Weighting of candidate estimators",
+                      ["inverse_risk", "equal", "single_best"]),
+            ParamSpec("alpha", "float", False, 0.05),
+        ],
+        returns="DMLAveragingResult",
+        example=('sp.dml_model_averaging(df, y="y", treat="d", '
+                 'covariates=[f"x{j}" for j in range(10)])'),
[…]
