+++ src/statspai/bartik/__init__.py
@@ -18,5 +18,9 @@ Borusyak, K., Hull, P., and Jaravel, X. (2022).
 
 from .shift_share import bartik, BartikIV
 from .adao_correction import ssaggregate, shift_share_se
+from .political import shift_share_political, ShiftSharePoliticalResult
 
-__all__ = ['bartik', 'BartikIV', 'ssaggregate', 'shift_share_se']
+__all__ = [
+    'bartik', 'BartikIV', 'ssaggregate', 'shift_share_se',
+    'shift_share_political', 'ShiftSharePoliticalResult',
+]
