### 2026-04-21 0859763
docs(changelog): record v1.3.0 CI fixes (F821 + tabulate dep)
Extends the [1.3.0] "Fixed" section so the post-release bugfixes that
went into this tag are visible to users browsing the CHANGELOG:

  - TYPE_CHECKING import for pathlib.Path in question.save's return
    annotation (closes the flake8 F821 that turned CI red for four
    commits running).
  - tabulate>=0.9.0 added to core deps so pandas.to_markdown() works
    out of the box on Windows / minimal envs.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
