### 2026-04-21 59eb3fc
docs: add v1.2/v1.3 frontier guides + cross-reference from choosers
New guides
----------

* docs/guides/v1_2_frontier.md — walks through every estimator added in
  v1.2 (gardner_did / harvest_did / dml_model_averaging / kernel_iv /
  continuous_iv_late / hal_tmle / synth_survival / RD aliases / BCF
  ordinal + factor-exposure / shift_share_political / causal_mas /
  evidence_without_injustice / causal_kalman) with when-to-use-which
  decision tree.
* docs/guides/harvest_did.md — standalone reference for
  sp.harvest_did (MIT/NBER WP 34550, 2025).
* docs/guides/synth_experimental.md — Abadie-Z
[…]
