+++ docs/guides/assimilative_ci.md
@@ -1,0 +1,157 @@
+# Assimilative Causal Inference
+
+> *Nature Communications* 2026: bridging Bayesian data assimilation
+> (weather forecasting, oceanography) with streaming causal inference.
+
+## 1. What problem does this solve?
+
+Most causal-inference pipelines assume the data arrive in one shot.
+But in A/B testing, pharmacovigilance, and policy evaluation you
+typically get **a batch at a time**, and the treatment effect itself
+may drift with seasons, user segments, or policy regimes.
+
+Assimilative causal inference treats the causal effect as a
+time-varying latent state and updates its posterior every time a new
+batch arrives — exactly the same mathematics that a weather model
+uses to fuse satellite observations into a global atmospheric state.
+
+## 2. State-space formulation
+
+$$
+\begin{aligned}
+\theta_t &= \theta_{t-1} + w_t,
+   \qquad w_t \sim \mathcal{N}(0, Q_t) \quad \text{(dynamics)}\\[2pt]
+\widehat{\theta}_t &= \theta_t + e_t,
+   \qquad e_t \sim \mathcal{N}(0, \sigma_t^2) \quad \text{(observation)}
+\end{aligned}
+$$
+
+Under Gaussianity the Kalman filter gives closed-form posteriors
+`θ_t | y_{1:t} ~ N(m_t, P_t)`.  For non-Gaussian / nonlinear settings
+StatsPAI ships a bootstrap-SIR particle filter with systematic
+resampling.
+
+## 3. Two-level API
+
+### Low-level: fuse a pre-computed stream
+
+If you have already produced a stream of `(θ̂_t, σ_t)` from, say,
+running `sp.did` or `sp.dml` on each batch:
+
+```python
+import statspai as sp
+
+res = sp.causal_kalman(
+    estimates=[θ̂_1, θ̂_2, ...],
+    standard_errors=[σ_1, σ_2, ...],
+    prior_mean=0.0,
+    prior_var=1.0,
+    process_var=0.0,    # 0 = static effect, >0 = random-walk drift
+    alpha=0.05,
+)
+print(res.summary())
+```
+
+For heavy-tailed observation noise or non-Gaussian priors swap in the
+particle filter:
+
+```python
+res = sp.assimilation.particle_filter(
+    estimates, standard_errors,
+    n_particles=3000,
+    process_sd=0.05,     # random-walk SD
+    random_state=0,
+)
+```
+
+### High-level: fuse directly from raw batches
+
+Pass the batches and a `(df → (θ̂, σ))` estimator callback and the
+pipeline handles both per-batch estimation and assimilation:
+
+```python
+def est(df):
+    r = sp.regress('y ~ d + x', data=df)
+    return float(r.params['d']), float(r.std_errors['d'])
+
+res = sp.assimilative_causal(
+    batches=[batch_jan, batch_feb, ...],
+    estimator=est,
+    prior_mean=0.0,
+    prior_var=1.0,
+    process_var=0.01,       # allow small month-to-month drift
+    backend='kalman',       # or 'particle' for non-Gaussian
+)
+```
+
+## 4. Outputs
+
+Every backend returns the same `AssimilationResult` dataclass:
+
+| Field              | Shape  | Meaning                             |
+|--------------------|--------|-------------------------------------|
+| `posterior_mean`   | (T,)   | Running posterior mean `m_t`        |
+| `posterior_sd`     | (T,)   | Running posterior SD `sqrt(P_t)`    |
+| `posterior_ci`     | (T, 2) | Per-step CI at level `alpha`        |
+| `innovations`      | (T,)   | Observation surprise `θ̂_t − m_{t|t-1}` |
+| `ess`              | (T,)   | Effective sample size               |
[…]
