### 2026-04-21 88658a4
release: bump to v1.4.0 — docs + CausalForest deflake + CHANGELOG wire-up
Doc / release-note alignment for the v1.4.0 tag plus the CI flake fix
that v1.4.0 carries forward:

- pyproject.toml + __version__ bumped 1.3.0 → 1.4.0 (owner pre-set)
- CHANGELOG.md: the pre-existing [1.4.0] section gains a Fixed entry
  recording the CausalForest parity-test deflake
- README.md / README_CN.md: headline banner rewritten around Sprint-2
  (panel shift-share, real-LLM adapters, particle-filter assimilation,
  3 new guides), with "v1.3 stable foundation (carried forward)" row
- docs/index.md: "Current rele
[…]
