### 2026-04-21 3eefa71
docs(guides): add bridging-theorems guide + sp.bridge full test suite
Bridging theorems are the differentiator vs Stata / R: six 2025-2026
results (Shi-Athey, Ferman, Zhao-Percival, Lu-Wang-Xie, Zhang,
Kallus-Mao) proving that pairs of superficially different causal
estimators identify the same estimand. Running both and comparing
turns assumption uncertainty into a data-visible diagnostic.

docs/guides/bridging_theorems.md
  Full walkthrough: dispatcher usage, BridgeResult fields, per-bridge
  explanation + identification story + "when to use / how to read
  disagreement" for all six bridges.
[…]
