+++ tests/test_bridge_full.py
@@ -1,0 +1,163 @@
+"""End-to-end smoke + correctness tests for all 6 bridging theorems.
+
+Each ``sp.bridge(kind=...)`` dispatcher runs two independent identification
+paths and reports a doubly-robust combined estimate plus an agreement
+test.  These tests ensure every kind is dispatchable, produces finite
+numbers, and — on a DGP that satisfies both identifying assumptions —
+the two paths agree within normal sampling error.
+"""
+
+from __future__ import annotations
+
+import warnings
+import numpy as np
+import pandas as pd
+import pytest
+
+warnings.filterwarnings("ignore")
+
+import statspai as sp
+
+
+# ═══════════════════════════════════════════════════════════════════════
+#  did_sc — Shi-Athey 2025
+# ═══════════════════════════════════════════════════════════════════════
+
+def _did_sc_panel(seed=0, units=11, T=15, treat_time=10, tau=2.0):
+    rng = np.random.default_rng(seed)
+    rows = []
+    for u in range(units):
+        u_fe = rng.normal(0, 0.3)
+        for t in range(T):
+            y = u_fe + rng.normal(0, 0.1)
+            if u == 0 and t >= treat_time:
+                y += tau
+            rows.append({"unit": u, "year": t, "y": y})
+    return pd.DataFrame(rows)
+
+
+def test_bridge_did_sc_recovers_truth():
+    df = _did_sc_panel(seed=0)
+    r = sp.bridge(kind="did_sc", data=df, y="y", unit="unit", time="year",
+                   treated_unit=0, treatment_time=10)
+    assert abs(r.estimate_dr - 2.0) < 0.5
+    assert r.kind == "did_sc"
+    assert r.path_a_name and r.path_b_name
+    assert r.se_a >= 0 and r.se_b >= 0
+
+
+# ═══════════════════════════════════════════════════════════════════════
+#  cb_ipw — Zhao-Percival 2025
+# ═══════════════════════════════════════════════════════════════════════
+
+def _binary_treatment_data(n=400, tau=1.5, seed=0):
+    rng = np.random.default_rng(seed)
+    X = rng.normal(size=(n, 3))
+    p = 1 / (1 + np.exp(-X[:, 0]))
+    D = (rng.uniform(size=n) < p).astype(float)
+    Y = tau * D + X[:, 0] + rng.normal(0, 0.3, n)
+    return pd.DataFrame({"y": Y, "d": D,
+                          **{f"x{j}": X[:, j] for j in range(3)}})
+
+
+def test_bridge_cb_ipw_recovers_truth():
+    df = _binary_treatment_data(seed=0, tau=1.5)
+    r = sp.bridge(kind="cb_ipw", data=df, y="y", treat="d",
+                   covariates=["x0", "x1", "x2"])
+    # CB and IPW should both recover ~1.5
+    assert abs(r.estimate_a - 1.5) < 0.4
+    assert abs(r.estimate_b - 1.5) < 0.4
+    # Under correct specification the two paths should agree
+    assert r.diff_p > 0.05, f"paths disagree (p={r.diff_p}) on correctly-specified DGP"
+
+
+def test_bridge_cb_ipw_dr_combination_finite():
+    df = _binary_treatment_data(seed=1)
+    r = sp.bridge(kind="cb_ipw", data=df, y="y", treat="d",
+                   covariates=["x0", "x1", "x2"])
+    assert np.isfinite(r.estimate_dr) and np.isfinite(r.se_dr)
+
+
+# ═══════════════════════════════════════════════════════════════════════
+#  dr_calib — Zhang et al. 2025
[…]
