+++ docs/guides/bridging_theorems.md
@@ -1,0 +1,238 @@
+# Bridging theorems — when two methods target the same estimand
+
+> **Six 2025-2026 results proving that pairs of superficially different
+> estimators identify the same causal quantity under distinct
+> assumptions.** Running both and comparing is how you turn assumption
+> uncertainty into data-visible diagnostics.
+
+A bridging theorem takes two popular causal-inference methods and shows
+that — when their identifying assumptions hold — they target the *same*
+estimand. Empirically they should give numerically close point
+estimates.  Disagreement is then a signal that at least one of the two
+assumptions fails in your data.
+
+This is the spiritual descendant of the classic "OLS = IV under no
+endogeneity" sanity check, but lifted to the 2020s frontier of DiD,
+synthetic controls, welfare maximisation, covariate balancing, and
+proximal/surrogate long-term effects.
+
+`sp.bridge(kind=..., **kwargs)` is the unified dispatcher. It runs both
+paths, computes an agreement test on the difference, and returns a
+doubly-robust combined estimate — plus a warning if the two paths
+disagree beyond sampling error.
+
+---
+
+## The six bridges
+
+| `kind`            | Theorem                                          | Reference                                 |
+|-------------------|--------------------------------------------------|-------------------------------------------|
+| `did_sc`          | DiD ≡ Synthetic Control                          | Shi & Athey, arXiv:2503.11375 (2025)       |
+| `ewm_cate`        | EWM ≡ CATE-max policy                            | Ferman et al., arXiv:2510.26723 (2025)     |
+| `cb_ipw`          | Covariate Balancing ≡ IPW ≡ DR                   | Zhao & Percival, arXiv:2310.18563 v6 (2025)|
+| `kink_rdd`        | Bunching ≡ Kink RDD first-order expansion        | Lu, Wang, Xie, arXiv:2404.09117 (2025)     |
+| `dr_calib`        | Doubly-robust ≡ outcome + Riesz joint calibration | Zhang et al., arXiv:2411.02771 (2025)     |
+| `surrogate_pci`   | Long-term Surrogate Index ≡ PCI                  | Kallus & Mao, arXiv:2601.17712 (2026)      |
+
+Each bridge is importable at top level as `sp.bridge(kind="..." ...)` or
+via the per-module function in `statspai.bridge.*`.
+
+---
+
+## How to read a `BridgeResult`
+
+```python
+>>> r = sp.bridge(kind="cb_ipw", data=df, y="y", treat="d",
+...               covariates=["x1", "x2", "x3"])
+>>> r.summary()
+```
+
+The key fields:
+
+- `estimate_a`, `estimate_b` — the two path point estimates.
+- `se_a`, `se_b` — their standard errors.
+- `diff`, `diff_se`, `diff_p` — difference and formal test
+  `H0: estimate_a - estimate_b = 0`.
+- `estimate_dr`, `se_dr` — precision-weighted combination, the
+  recommended final number **when the agreement test does not reject**.
+- `path_a_name` / `path_b_name` — human-readable method names.
+- `reference` — the citation for this bridge.
+
+**Rule of thumb**: if `diff_p > 0.10`, trust the DR estimate. If
+`diff_p < 0.05`, at least one identifying assumption is violated —
+inspect the detail dict (`r.detail`) for method-specific diagnostics
+before trusting either path.
+
+---
+
+## 1. DiD ≡ Synthetic Control — Shi-Athey (2025)
+
+The result: whenever (a) the classical DiD parallel-trends assumption
+*or* (b) unit-level synthetic-control weights identify the ATT, then
[…]
