### 2026-04-21 14056b6
fix(dml): √n scaling bug in dml_model_averaging SE — CI was ~20× too wide
The cross-candidate variance aggregator multiplied influence functions
by 1/√n in the outer-product, giving Σ φ_k φ_l = (1/n)·Σ_i φ_ki φ_li,
but then used that directly as Var(θ̂_avg) instead of the correct
(1/n)·Σ / n = (1/n²)·Σ.  Net effect: reported SE was √n larger than
the truth.

On the canonical n=400 DGP with true θ=1.5:

              before fix      after fix
  CI width     4.20            0.21     (×20 = √400)
  Coverage    100% (over)      82% (~nominal)

82% vs the nominal 95% is driven by a small finite-sam
[…]
