### 2026-04-21 c78b68f
fix(did): gardner_did event-study reference-category contamination
The Stage-2 dummy regression pooled all non-event rows (never-treated +
treated units outside the event-study horizon) into a single baseline
category, dragging every event-time coefficient toward the mean of that
pool.  On a synthetic panel with true τ=2 and strict parallel trends,
this produced:

  before: pre-trends ≈ -0.30 (should be 0), post ≈ +1.72 (should be 2.0)
  after:  pre-trends ≈ +0.01 (≈0),           post ≈ +2.02 (≈2.0)

The fix replaces the Stage-2 dummy regression in event-study mode with a
direct Borusyak-Jarav
[…]
