### 2026-04-21 bb1e72e
release: v1.4.1 — Sprint 3 additive follow-up
New surface (additive kwargs only; all v1.4.0 APIs stable):

- sp.shift_share_political_panel(cluster='shock') — panel-extended
  Adão-Kolesár-Morales (2019) shock-clustered SE recommended by
  Park-Xu (2026) §4.2; exposes diagnostics['akm_se'] + human-readable
  diagnostics['cluster'] label.
- sp.causal_llm.anthropic_client(thinking_budget=N) — Claude 4.5 /
  Opus 4.7 extended-thinking API.  Reasoning trace stored on
  client.history[-1]['thinking'] for auditability; NOT returned as
  part of the public answer.  Handles thinking + redacted_thinkin
[…]
