### 2026-04-21 1598f8b
release: v1.4.2 — correctness patches + Proximal/QTE/Causal-RL family guides
Patch release on top of v1.4.1.  Two silent-wrong-numbers bug fixes
already landed on main (commits 14056b6 and c78b68f) plus three new
family guides closing the last v3-document documentation gaps.

Version bump
------------

pyproject.toml + src/statspai/__init__.py: 1.4.1 → 1.4.2

Fixed (previously landed on main, now formally shipped)
-------------------------------------------------------

* sp.dml_model_averaging — √n SE scaling bug.  CI width went from
  4.20 (coverage 100%) to 0.21 (coverage 82% ≈ nominal) on 
[…]
