+++ docs/guides/qte_family.md
@@ -1,0 +1,240 @@
+# Quantile & distributional treatment effects — the full family
+
+> **"The ATE is misleading when treatment redistributes outcomes."**
+> A policy can leave the mean unchanged while compressing or spreading
+> the distribution — only QTE and distributional-TE catch that.
+> StatsPAI v1.4 ships the full 2025-2026 frontier.
+
+The Average Treatment Effect answers: "how much did the mean outcome
+move?"  But for policy, clinical, and fairness work you usually care about
+*who* got moved.  Did the bottom decile benefit?  Did treatment widen
+inequality?  Those questions need quantile or distributional effects.
+
+Every function below is at top level: `sp.qdid`, `sp.qte`,
+`sp.distributional_te`, `sp.dist_iv`, `sp.kan_dlate`,
+`sp.beyond_average_late`, `sp.qte_hd_panel`, `sp.cic`.
+
+---
+
+## The three levels of granularity
+
+| Level                  | What it answers                                  | Estimator(s)                     |
+|------------------------|--------------------------------------------------|----------------------------------|
+| Average                | "Did the mean go up?"                            | `sp.did`, `sp.iv`, `sp.dml`      |
+| Quantile (τ-th)         | "Did the bottom 10%? The median? The top 10%?"  | `sp.qte`, `sp.qdid`, `sp.cic`    |
+| Whole distribution     | "How did the *shape* change?"                    | `sp.distributional_te`, `sp.dist_iv` |
+
+---
+
+## Conditional QTE under unconfoundedness — `sp.qte`
+
+The default when you have observational data with `X` rich enough for
+selection-on-observables:
+
+```python
+r = sp.qte(
+    data=df, y="wage", treatment="job_program",
+    controls=["age", "education", "experience"],
+    quantiles=[0.1, 0.25, 0.5, 0.75, 0.9],
+    method="quantile_regression",   # or "ipw"
+    n_boot=500,
+)
+r.plot()    # QTE curve with bootstrap uniform band
+```
+
+Reference: Firpo (2007), *Econometrica* 75(1).  IPW variant: Firpo (2007)
+§4 with the propensity score.
+
+---
+
+## Quantile DiD — `sp.qdid`
+
+Two-period, two-group DiD at each quantile.  Parallel-trends is replaced
+by the **quantile parallel trends** assumption: "in the absence of
+treatment, the τ-th quantile of the treated group would have moved the
+same way as the τ-th quantile of the control group."
+
+```python
+r = sp.qdid(
+    data=df, y="y", group="treated", time="post",
+    quantiles=[0.1, 0.25, 0.5, 0.75, 0.9],
+    n_boot=500,
+)
+r.summary()     # QTE(τ) for each τ with placebo-based CI
+```
+
+Reference: Athey & Imbens (2006), *Econometrica* 74(2).
+
+---
+
+## Changes-in-changes (CiC) — `sp.cic`
+
+Weaker than QDiD: CiC requires only that the **rank** of each unit within
+its group's outcome distribution is stable over time in the absence of
+treatment.  No parallel-trends.  Works when outcome distributions have
+different shapes across groups.
+
+```python
+r = sp.cic(
+    data=df, y="earnings", group="treated", time="post",
+    quantiles=np.linspace(0.05, 0.95, 19),
+    n_boot=500,
+)
+```
+
+Reference: Athey & Imbens (2006), *Econometrica*.  The same paper proves
+CiC nests QDiD when the distributions happen to share scale.
+
+---
+
[…]
