+++ docs/guides/causal_rl_family.md
@@ -1,0 +1,220 @@
+# Causal reinforcement learning — the full family
+
+> **When your decisions today change the data you learn from tomorrow**,
+> classical causal methods break: propensity scores mutate, confounders
+> drift, reward is delayed.  Causal RL bridges the gap.  StatsPAI v1.4
+> ships bandit, MDP, offline-safe, confounding-robust deep-Q, and a
+> benchmark suite — all under a consistent API.
+
+This guide is for applied economists and decision-support engineers
+deciding between classical causal inference and RL-style sequential
+policy learning.  Every function below lives at top level:
+`sp.causal_bandit`, `sp.causal_dqn`, `sp.counterfactual_policy_optimization`,
+`sp.structural_mdp`, `sp.offline_safe_policy`, `sp.causal_rl_benchmark`.
+
+---
+
+## When to reach for causal RL (vs one-shot causal inference)
+
+| Your problem looks like…                                              | Right tool                      |
+|-----------------------------------------------------------------------|---------------------------------|
+| "Did treatment X affect outcome Y once?" (one time period, no follow-up) | Classical causal (`sp.did`, `sp.iv`, …) |
+| "Under which context should we assign A or B?" (contextual, one shot) | `sp.causal_bandit`              |
+| "Best multi-step policy, offline data, no confounders"                | Standard RL (outside StatsPAI)  |
+| "Multi-step policy + unobserved confounders"                          | `sp.causal_dqn`                 |
+| "Multi-step policy + safety constraint"                               | `sp.offline_safe_policy`        |
+| "Off-policy evaluation of a stochastic policy"                        | `sp.counterfactual_policy_optimization` |
+| "General structural MDP with causal interpretation"                   | `sp.structural_mdp`             |
+
+---
+
+## Causal bandit — `sp.causal_bandit` (Bareinboim-Pearl 2015)
+
+Contextual bandit that respects a causal DAG: certain arms are
+causally linked to context variables, so pulling arm `A` with context
+`X` reveals information about arm `A'` in that same context.  Standard
+contextual bandits ignore this and waste samples.
+
+```python
+def reward_fn(arm, context):
+    # user-supplied: returns noisy reward for this arm × context
+    ...
+
+r = sp.causal_bandit(
+    arms=["A", "B", "C"],
+    reward_fn=reward_fn,
+    context={"age_group": "young"},
+    n_samples=1000,
+)
+r.best_arm              # arm with highest posterior mean reward
+r.regret_curve          # cumulative regret over samples
+```
+
+Reference: Bareinboim, Forney, Pearl (2015), *NeurIPS*.
+
+---
+
+## Causal DQN — `sp.causal_dqn` (Fu-Zhou 2025)
+
+Deep Q-learning with a confounding-robust update rule.  Given offline
+trajectories `(state, action, reward, next_state)` possibly collected
+under an *unknown* logging policy, `causal_dqn` learns a Q-function
+that is robust to bias from unobserved confounders in the logging
+policy, bounded by `gamma_bound`.
+
+```python
+r = sp.causal_dqn(
+    data=offline_trajectories,
+    state="s_vec", action="a", reward="r", next_state="s_next_vec",
+    gamma_bound=0.1,        # upper bound on confounding strength
+    discount=0.95,
+    n_iter=200,
+    lr=0.05,
+)
+r.policy                     # learned greedy policy
+r.q_values                   # Q̂(s, a)
+r.confounding_sensitivity    # worst-case Q-value shift
+```
+
+**Why this over vanilla DQN**: vanilla DQN trained on confounded offline
[…]
