### 2026-04-21 60815e3
docs(readme): surface v1.4.2 release notes in README + README_CN
Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
