+++ README_CN.md
@@ -15,7 +15,18 @@ StatsPAI 是一个**面向 AI Agent** 的 Python 因果推断与应用计量经�
 
 它将 R 的 [Causal Inference Task View](https://cran.r-project.org/web/views/CausalInference.html)（fixest、did、rdrobust、gsynth、DoubleML、MatchIt、CausalImpact、sfaR、lme4、oaxaca、ddecompose……）和 Stata 的核心计量命令（`frontier`、`xtfrontier`、`mixed`、`meglm`、`mixlogit`、`ivqreg`……），统一到一个一致的 Python API 中。
 
-**🎉 v1.4.1 新版本 — v3 前沿 Sprint 3：AKM 冲击聚类 SE、Claude 扩展思考、对齐与集成测试套件、2 篇新指南**
+**🎉 v1.4.2 新版本 — 正确性补丁 + Proximal / QTE / 因果 RL 家族指南**
+
+StatsPAI 1.4.2 是 patch 版本。无破坏性改动、无新增公开签名——只修了两个"悄悄算错"的 bug，补上三篇家族指南，关闭 v3 参考文档和公开文档之间的最后几处缺口。
+
+- **⚠️ 正确性修复 — `sp.dml_model_averaging` 的 √n SE 尺度 bug**：候选间方差聚合器把"样本均值影响函数的外积"当成了 `Var(θ̂_avg)` 本身，漏掉了最后的 `/ n`。报告的 SE 因此比真值大 `√n` 倍；在典型 n=400 DGP 上，95% CI 宽度 4.20（理论 ≈ 0.21），实证覆盖率 100%（名义 95%）。修复后 CI 宽度 0.21，覆盖率回到名义水平。回归守卫加在 `tests/test_dml_model_averaging.py::test_se_on_correct_scale`。
+- **⚠️ 正确性修复 — `sp.gardner_did` event-study 参照组污染**：Stage-2 dummy 回归把"从未处理单元"和"处在 event-study 视野之外的已处理单元"合并成同一个基线，把每个 event-time 系数都往这个混合基线的均值拖。τ=2 且严格平行趋势的合成面板上，pre-trend ≈ -0.30（应为 0），post ≈ +1.72（应为 2.0）。Event-study 模式下改用 Borusyak-Jaravel-Spiess 风格的 (cohort × relative-time) 分箱直接聚合 imputed gap。修复后 pre-trend ≈ +0.01，post ≈ +2.02。非 event-study 的单 ATT 路径本来就对，保持不变。
+- **新增家族指南**：`docs/guides/proximal_family.md`（Proximal 因果推断家族完整走查，覆盖 `sp.proximal`、`sp.fortified_pci`、`sp.bidirectional_pci`、`sp.pci_mtp`、`sp.double_negative_control`、`sp.proximal_surrogate_index`、`sp.select_pci_proxies`，含决策树与 PCI 分析必报的 4 项诊断）；`docs/guides/qte_family.md`（均值 → 分位数 → 整分布三级粒度，配 cross-section / DiD / IV / 面板四条决策路径，覆盖 `sp.qte`、`sp.qdid`、`sp.cic`、`sp.distributional_te`、`sp.dist_iv`、`sp.kan_dlate`、`sp.beyond_average_late`、`sp.qte_hd_panel`）；`docs/guides/causal_rl_family.md`（何时用 causal RL 而非经典因果推断，覆盖 `sp.causal_bandit`、`sp.causal_dqn`、`sp.offline_safe_policy`、`sp.counterfactual_policy_optimization`、`sp.structural_mdp`、`sp.causal_rl_benchmark`，附 4 项 causal-RL 专属 sanity check）。
[…]
