### 2026-04-21 d1719bb
release: v1.5.0 — MR/conformal/interference family consolidation
- Unified family dispatchers: sp.mr(method=...), sp.conformal(kind=...),
  sp.interference(design=...) mirroring sp.synth / sp.decompose / sp.dml
- ⚠️ Breaking: sp.mr is now a dispatcher function, not a module alias
  (use sp.mendelian for module-level access; sp.mr_* top-level helpers
  unchanged)
- ⚠️ Correctness: mr_egger slope inference now uses t(n−2) instead of
  Normal (anti-conservative at small n_snps); mr_presso MC p-values now
  use (k+1)/(B+1) convention (no more exact-zero p-values)
- Fixed: dead var= line in network
[…]
