### 2026-04-21 7502e6a
docs(v1.5.0): sync README_CN, docs/index, 3 family-guide footers to 1.5.0
- README_CN.md: add v1.5.0 "NEW" block; demote v1.4.2 to "Previously in"
- docs/index.md: Current release banner + release-highlights table top
  row bumped to v1.5.0; BibTeX version bump; drop stale migration link
- docs/guides/{proximal,qte,causal_rl}_family.md: version footer
  "≥ 1.4.2" → "≥ 1.5.0"

Companion to d1719bb which shipped the code, README, CHANGELOG, and the
three new interference / conformal / MR guides but missed these
downstream doc-sync files.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@ant
[…]
