+++ README_CN.md
@@ -15,16 +15,27 @@ StatsPAI 是一个**面向 AI Agent** 的 Python 因果推断与应用计量经�
 
 它将 R 的 [Causal Inference Task View](https://cran.r-project.org/web/views/CausalInference.html)（fixest、did、rdrobust、gsynth、DoubleML、MatchIt、CausalImpact、sfaR、lme4、oaxaca、ddecompose……）和 Stata 的核心计量命令（`frontier`、`xtfrontier`、`mixed`、`meglm`、`mixlogit`、`ivqreg`……），统一到一个一致的 Python API 中。
 
-**🎉 v1.4.2 新版本 — 正确性补丁 + Proximal / QTE / 因果 RL 家族指南**
+**🎉 v1.5.0 新版本 — Interference / Conformal / Mendelian 三家族合并升级**
 
-StatsPAI 1.4.2 是 patch 版本。无破坏性改动、无新增公开签名——只修了两个"悄悄算错"的 bug，补上三篇家族指南，关闭 v3 参考文档和公开文档之间的最后几处缺口。
+StatsPAI 1.5.0 是 minor 版本，一次性完成 interference、conformal 因果推断、Mendelian 随机化三个家族的三项联动升级：完整家族文档指南、与 `sp.synth` / `sp.decompose` / `sp.dml` 同构的统一 dispatcher，以及一次针对性的正确性审计修掉两处"悄悄算错"。
+
+- **三篇新家族指南**：`docs/guides/interference_family.md`（`sp.spillover`、`sp.network_exposure`、`sp.peer_effects`、`sp.network_hte`、`sp.inward_outward_spillover`、`sp.cluster_matched_pair`、`sp.cluster_cross_interference`、`sp.cluster_staggered_rollout`、`sp.dnc_gnn_did` 完整走查，含决策树 + 5 项 interference 分析必报诊断）；`docs/guides/conformal_family.md`（10 个 conformal 估计器按"边际覆盖保证的含义"重新组织，明确 marginal 与 conditional 的区别）；`docs/guides/mendelian_family.md`（17 个 MR 函数按 IV1 / IV2 / IV3 假设层级重新组织 + 每次 MR 分析必报的 4 项 sanity check + BMI → T2D 完整示例）。
+- **三个新统一 dispatcher**：`sp.mr(method=...)` 覆盖 33 个别名（IVW / Egger / median / mode / MVMR / mediation / BMA / PRESSO / radial / Steiger / F 统计量 / LOO / 多效性 / 异质性）；`sp.conformal(kind=...)` 覆盖 29 个别名（CATE / counterfactual / ITE / density / 多阶段 / debiased-ML / fair / continuous / interference）；`sp.interference(design=...)` 覆盖 29 个别名（partial / network-exposure / peer-effects / network-HTE / inward-outward / 4 种 cluster-RCT 变体）。kwargs 原样透传；每个 dispatcher 的输出与直调结果逐字节一致（30 个新测试守卫）。
[…]
