### 2026-04-21 b3c5fe4
docs(readme): add causal inference coverage matrix + ecosystem stats page
Adds a 15-row coverage matrix at the top of README.md and README_CN.md
positioning StatsPAI against Stata (official + SSC), R (CRAN),
statsmodels+linearmodels, and DoubleML across the core causal-inference
families (DiD, IV, RD, Synth, DML, TMLE, neural causal, discovery,
Proximal, QTE, MR, Conformal, BCF, Spatial).

New docs/stats.md carries the full 23-row matrix, cross-ecosystem LOC
comparison (StatsPAI 188K vs statsmodels 382K vs Stata official 1.04M vs
R ecosystem ~150M), per-module LOC/function breakdown for all 78
[…]
