+++ docs/stats.md
@@ -1,0 +1,162 @@
+# StatsPAI — Ecosystem & Code Statistics
+
+> This page tracks StatsPAI's size and coverage against the broader statistical ecosystem, in the spirit of *evidence-based* positioning. Numbers marked **"measured"** are reproducible on a standard install; numbers marked **"estimated"** are extrapolated from public ecosystem statistics with the reasoning shown.
+>
+> *Last measured: 2026-04-21 on macOS arm64 against `statspai` 0.9.16, `statsmodels` 0.14.x, `linearmodels` (latest), Stata 18, R 4.5.2.*
+
+---
+
+## 1 · Cross-ecosystem lines-of-code
+
+| Ecosystem / Project                  | Method     |  Files | Lines of code | Primary focus                      |
+| ------------------------------------ | ---------- | -----: | ------------: | ---------------------------------- |
+| **StatsPAI** `src/statspai/`         | measured   |    528 |   **188,244** | full-stack causal inference        |
+| StatsPAI tests (`tests/`)            | measured   |    255 |        42,768 | —                                  |
+| statsmodels 0.14.x                   | measured   |    948 |   **381,981** | GLM / time series / general        |
+| linearmodels                         | measured   |    131 |        36,607 | panel / IV                         |
+| **Python causal-inference subtotal** |            |  1,079 |   **418,588** |                                    |
+| Stata 18 — official `.ado`           | measured   |  3,767 |       937,543 | command layer above closed kernel  |
+| Stata 18 — official `.mata`          | measured   |    411 |       103,822 | Mata numerical layer               |
+| **Stata official executable code**   | measured   |  4,178 | **1,041,365** | (+ 738,543 lines of `.sthlp` help text, not counted as code) |
+| Stata SSC (third-party)              | estimated  |  ~3,500 pkgs | **2M – 4M** | community extensions; local sample (reghdfe + winsor2 + 50 others) = 33,296 LOC |
+| R base interpreter (C + R + Fortran) | estimated  |      — |        ~1.5M | language itself                    |
+| R base library (73 recommended pkgs) | measured   |    509 |        62,321 | shipped with R on this machine     |
+| CRAN (~22,000 packages, 2026)        | estimated  |      — |  **80M–120M** (R-only; >200M incl. C/C++/Fortran) | main R package universe |
+| Bioconductor (~2,300 packages)       | estimated  |      — |       30M–50M | bioinformatics                     |
+| **R ecosystem total**                | estimated  |      — |    **≈ 150M+** |                                   |
+
+**How to read this table**
+
+LOC is a vanity metric in isolation — what matters is **coverage density** within a target domain. StatsPAI is deliberately scoped at causal inference and applied econometrics; it is **not** trying to match R's 150M+ lines because ~90% of CRAN is bioinformatics, visualization, text mining, and general-purpose ML that is out of scope. The relevant comparison is the coverage matrix in §3 below.
+
+---
+
+## 2 · StatsPAI per-module breakdown
+
+Sorted by LOC. All numbers measured on `src/statspai/` at the commit listed in the timestamp above.
+
[…]
