### 2026-04-22 e0d2fd3
docs(specs): P0 agent-native coverage + recovery design
Design doc for P0: auto-generated MCP tool manifest (coverage 8→150+),
result.to_dict()/for_agent() on CausalResult + EconometricResults,
expanded remediation registry with ~15 causal-specific entries,
server version sync.
