+++ docs/superpowers/specs/2026-04-21-p0-agent-native-design.md
@@ -1,0 +1,151 @@
+# P0 — Agent-Native Coverage + Recovery (2026-04-21)
+
+> Goal: take StatsPAI from "agent-native in principle" to "agent-native with
+> measurable coverage" along two axes — **what's reachable** (MCP surface)
+> and **what's recoverable** (failure → fix loop).
+
+## Context (探测事实)
+
+- `sp.list_functions()` = **889 registered functions**
+- `sp.agent.tool_manifest()` = **8 hand-curated tools** → 99.1% coverage gap
+- `sp.function_schema(name)` already emits OpenAI-format schemas for any
+  registered function (field-complete — name/description/parameters with
+  enum + required)
+- `sp.all_schemas()` already bulk-exports all 889 schemas
+- `CausalResult` / `EconometricResults` have `.summary / .tidy / .glance /
+  .cite / .to_latex / .to_docx / .next_steps` but **no `to_dict()` /
+  `to_json()` / `for_agent()`** — serialization sits externally in
+  `agent/tools.py::_default_serializer` and is fragile
+- `REMEDIATIONS` in `agent/remediation.py` = 11 entries, mostly generic
+  Python errors — no causal-specific recovery paths
+- `mcp_server.py::SERVER_VERSION = "0.9.17"` — version drift vs project
+  v1.5.0 (release flow doesn't propagate)
+
+## Scope
+
+Approach **A+B hybrid**:
+
+- **A (coverage lever)**: drive MCP tool count from 8 → ~150 via auto-
+  generation on top of the existing `all_schemas()` surface. Use a
+  category-based whitelist so we don't expose every utility function.
+- **B (depth)**: add agent-readable serialization to the result classes,
+  expand remediation registry to cover causal-specific failures, sync
+  server version, and make `sp.recommend` / `sp.compare_estimators` first-
+  class MCP tools.
+
+## Deliverables
+
+### D1 — `to_dict()` / `for_agent()` on result classes
+
+- `CausalResult.to_dict()` → flat JSON-safe dict with `method`, `estimand`,
+  `estimate`, `se`, `pvalue`, `ci`, `alpha`, `n_obs`, `diagnostics`,
+  `citation_key`, optional `detail_head` (first N rows of detail DataFrame)
+- `CausalResult.for_agent()` → adds `warnings`, `next_steps` (compact),
+  `assumption_checks` (from model_info), and `suggested_functions` (other
+  `sp.xxx` to call next)
+- `EconometricResults.to_dict()` → adds `coefficients` keyed by term with
+  estimate/se/t/p/CI
+- `EconometricResults.for_agent()` → same + glance()-style model summary
+- Existing `_default_serializer` in `agent/tools.py` reduced to a
+  one-liner that calls `result.to_dict()` with backward-compat fallback
+
+### D2 — `auto_tool_manifest()` + merged `tool_manifest()`
+
+- New `agent/auto_tools.py`:
+  - `auto_tool_manifest(categories=None, exclude=None)` → walks the
+    registry, returns MCP tool specs for every function whose `category`
+    is in the whitelist AND which is not explicitly excluded
+  - Default whitelist: `{causal, regression, panel, inference, diagnostics,
+    smart, decomposition, robustness, postestimation, survival, survey}`
+  - Default exclude: plotting, I/O, CLI helpers, private underscore names,
+    known deep-learning estimators that need torch (excluded to keep
+    manifest loadable without optional deps)
+- `tool_manifest()` (in `agent/tools.py`) becomes a merger: hand-curated
+  8 tools win on name collision; auto-generated tools fill the rest
[…]
