### 2026-04-22 86d3736
docs(specs): P1 design — LLM-DAG loop + sp.paper + causal_text MVP
Three subprojects under one P1 umbrella:
- P1-A: LLM-DAG closed loop (sp.llm_dag_constrained, sp.llm_dag_validate)
- P1-C: sp.paper(data, question) end-to-end pipeline
- P1-B: sp.causal_text MVP (text-as-treatment + LLM-annotator MEC)

Builds on P0 agent-native infra (registry agent_card, exception
taxonomy, result.to_dict/for_agent). Marks B as experimental; defers
text-as-confounder/outcome to v1.7.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
