+++ docs/superpowers/specs/2026-04-21-p1-agent-native-frontier-design.md
@@ -1,0 +1,318 @@
+# P1 — Agent-Native × Methodological Frontier (2026-04-21)
+
+> Goal: push StatsPAI's two competitive axes — **Agent-native** and
+> **methodological frontier** — by closing the LLM-DAG loop, shipping an
+> end-to-end `sp.paper()` orchestrator, and opening the `causal_text`
+> frontier (MVP).
+
+## Context (探测事实)
+
+- P0 (`2026-04-21-p0-agent-native-design.md`) already shipped:
+  - MCP server (`agent/mcp_server.py`, 355 LOC, full JSON-RPC stdio)
+  - Tool manifest + `execute_tool` (`agent/tools.py`, 576 LOC)
+  - Remediation registry (`agent/remediation.py`, 270 LOC)
+  - `recommend` / `verify_recommendation` rule engine
+  - LLM single-shot DAG propose (`causal_llm/llm_dag.py`, in
+    `llm_dag_propose`) + LLM evaluator (`dag/llm_evaluator.py`)
+  - `sp.causal()` workflow object (`workflow/causal_workflow.py`, 1075 LOC,
+    diagnose → recommend → estimate → robustness → report)
+- 9 causal-discovery algorithms exist in `causal_discovery/`
+  (PC, GES, FCI, NOTEARS, DyNOTEARS, LiNGAM, ICP, PCMCI, LPCMCI), but
+  none accept *prior knowledge* (forbidden / required edges) as a
+  constraint
+- `dag/llm_dag.py` only does *single-shot* merge of an oracle with a CI
+  skeleton — no iteration, no per-edge confidence
+- `sp.causal()` produces an HTML/MD report but **does not** produce a
+  publication-ready TeX/Word section, and is not callable from a single
+  user-facing entry point that takes data + a natural-language question
+- No `causal_text` module exists (text/embedding as treatment / outcome /
+  confounder / measurement-error correction is unaddressed)
+
+## Scope
+
+Three subprojects under one P1 umbrella, sized for one release:
+
+| ID | Item | Depth | Why |
+|----|------|-------|-----|
+| **P1-A** | LLM-DAG closed loop | full | shared infra: `sp.paper()` and `recommend()` both consume DAGs |
+| **P1-C** | `sp.paper(data, question)` end-to-end | full | flagship demo of agent-native value |
+| **P1-B** | `sp.causal_text` MVP | MVP | greenfield — open the door, ship 2 methods, mark `experimental` |
+
+Out of scope (deferred):
+- `text-as-confounder` (Roberts-Stewart-Nielsen STM-style) — needs topic-
+  model dependency story
+- `text-as-outcome` (Egami et al. 2018) — needs annotation pipeline
+- LLM ensemble DAG aggregation (multiple LLMs voting) — depends on stable
+  closed-loop first
+
+## Deliverables
+
+### D1 — P1-A: LLM-DAG closed loop
+
+**New module**: `src/statspai/causal_llm/llm_dag_loop.py`
+
+**New public functions** (registered):
+
+1. `sp.llm_dag_constrained(data, variables, descriptions, oracle=None,
+   algo='pc', alpha=0.05, max_iter=3, ci_validate=True,
+   verbose=False) -> LLMConstrainedDAGResult`
+   - **Loop**: (a) call `oracle` (or `llm_dag_propose`) to get candidate
+     directed edges + confidences; (b) split into `forbidden` (LLM says
+     no edge) and `required` (LLM says edge); (c) run `algo` (PC variant
+     that respects forbidden/required edges as background knowledge);
+     (d) for each LLM-required edge, run a CI test on data — if rejected
[…]
