### 2026-04-22 8155ada
feat(agent): unified exception taxonomy + per-function agent cards
Foundation for agent-native failure recovery and self-describing tool
catalog. Extends the registry schema and adds three layered surfaces:

- `statspai.exceptions` — 7-class taxonomy (`StatsPAIError` base with
  `AssumptionViolation`, `IdentificationFailure`, `DataInsufficient`,
  `ConvergenceFailure`, `NumericalInstability`, `MethodIncompatibility`)
  plus `StatsPAIWarning` / `ConvergenceWarning` / `AssumptionWarning`
  and a `warn()` helper. Every exception carries `recovery_hint`,
  `diagnostics`, and `alternative_functions
[…]
