+++ src/statspai/registry.py
@@ -33,8 +33,39 @@ class ParamSpec:
 
 
 @dataclass
+class FailureMode:
+    """One failure mode for agent-native recovery.
+
+    Parameters
+    ----------
+    symptom : str
+        What the agent observes (exception class, warning text, pattern).
+    exception : str
+        Fully-qualified exception name (``"statspai.AssumptionViolation"``
+        or ``"ValueError"``). Agents should ``except`` on this.
+    remedy : str
+        One-sentence, actionable recovery hint.
+    alternative : str, optional
+        ``sp.xxx`` to try next when this failure mode triggers.
+    """
+    symptom: str
+    exception: str
+    remedy: str
+    alternative: str = ""
+
+    def to_dict(self) -> Dict[str, Any]:
+        return asdict(self)
+
+
+@dataclass
 class FunctionSpec:
-    """Machine-readable specification for a StatsPAI function."""
+    """Machine-readable specification for a StatsPAI function.
+
+    Agent-native fields (``assumptions`` / ``failure_modes`` /
+    ``alternatives`` / ``typical_n_min`` / ``pre_conditions``) are
+    optional — any entry without them still renders correctly; only
+    the agent-card layer surfaces the extras.
+    """
     name: str
     category: str
     description: str
@@ -43,6 +74,19 @@ class FunctionSpec:
     example: str = ""
     tags: List[str] = field(default_factory=list)
     reference: str = ""  # paper / method reference
+    # ------------------------------------------------------------------ #
+    #  Agent-native metadata (all optional; populate per-estimator)
+    # ------------------------------------------------------------------ #
+    assumptions: List[str] = field(default_factory=list)
+    """Identifying / statistical assumptions, human-readable one-liners."""
+    pre_conditions: List[str] = field(default_factory=list)
+    """Data-shape preconditions the agent should verify before calling."""
+    failure_modes: List[FailureMode] = field(default_factory=list)
+    """Common failures + recovery paths (see :class:`FailureMode`)."""
+    alternatives: List[str] = field(default_factory=list)
+    """Ranked ``sp.xxx`` fallbacks when this estimator is a poor fit."""
+    typical_n_min: Optional[int] = None
+    """Rule-of-thumb minimum sample size; ``None`` if not applicable."""
 
     def to_openai_schema(self) -> Dict[str, Any]:
         """Export as OpenAI function-calling compatible JSON schema."""
@@ -82,6 +126,30 @@ class FunctionSpec:
     def to_dict(self) -> Dict[str, Any]:
         return asdict(self)
 
+    def agent_card(self) -> Dict[str, Any]:
+        """Return the agent-native view of this function.
+
+        This is the structured payload rendered into guide
+        ``## For Agents`` sections, surfaced by :func:`sp.describe_function`
+        and consumed by :func:`sp.recommend` / agent code. It is a
+        superset of :meth:`to_openai_schema` plus the agent-native
+        fields (``assumptions`` / ``failure_modes`` / ``alternatives`` /
+        ``typical_n_min`` / ``pre_conditions``).
+        """
+        return {
+            "name": self.name,
+            "category": self.category,
+            "description": self.description,
+            "signature": self.to_openai_schema(),
+            "pre_conditions": list(self.pre_conditions),
+            "assumptions": list(self.assumptions),
+            "failure_modes": [fm.to_dict() for fm in self.failure_modes],
+            "alternatives": list(self.alternatives),
+            "typical_n_min": self.typical_n_min,
+            "reference": self.reference,
+            "example": self.example,
+        }
+
 
 # ====================================================================== #
 #  Registry
@@ -115,6 +183,39 @@ def _build_registry():
         returns="EconometricResults",
         example='sp.regress("wage ~ education + experience", data=df, robust="hc1")',
[…]
