+++ src/statspai/core/_agent_summary.py
@@ -1,0 +1,481 @@
+"""
+Agent-native structured views of StatsPAI result objects.
+
+This module hosts the logic behind ``result.to_agent_summary()`` and
+``result.violations()`` for both :class:`EconometricResults` and
+:class:`CausalResult`.  Kept separate from ``results.py`` to avoid
+bloating that file and to let the per-method rules evolve
+independently of the core data model.
+
+Design principles
+-----------------
+
+* **Non-invasive.** Neither method alters the underlying result
+  object.  Callers can invoke them any number of times.
+* **Structured, not prose.**  ``to_agent_summary()`` returns a
+  plain ``dict`` suitable for ``json.dumps`` or direct consumption
+  by an LLM tool loop.  ``violations()`` returns a list of dicts.
+* **Pattern-matching, not fitting.**  Violation detection only
+  *inspects* diagnostics that the estimator already stored
+  (``pretrend_test``, ``rhat``, ``first_stage_f``, …). It never
+  re-runs a test.
+"""
+
+from __future__ import annotations
+
+from typing import Any, Dict, List, Optional
+
+import numpy as np
+import pandas as pd
+
+
+# ====================================================================== #
+#  Thresholds (literature-based rules of thumb)
+# ====================================================================== #
+
+#: Pre-trend test p-value below which we flag a DID parallel-trends
+#: concern.  Deliberately *not* 0.05: agents should treat 0.10 as
+#: "warrants follow-up" per Roth (2022) on low-power pre-trend tests.
+_PRETREND_ALPHA = 0.10
+
+#: Stock-Yogo 5% bias threshold (single endogenous regressor, 2SLS).
+_WEAK_IV_F = 10.0
+
+#: Gelman-Rubin / MCMC convergence thresholds (mirrors PyMC / arviz).
+_RHAT_MAX = 1.01
+_ESS_MIN = 400
+_DIVERGENCES_MAX = 0
+
+#: Maximum tolerable standardized mean difference after matching (SMD).
+_SMD_MAX = 0.10
+
+#: Propensity score overlap: treated weight share below this → bad
+#: common support.
+_OVERLAP_MIN = 0.05
+
+
+# ====================================================================== #
+#  CausalResult helpers
+# ====================================================================== #
+
+
+def _safe_get(obj: Dict[str, Any], *keys: str, default: Any = None) -> Any:
+    """Get ``obj[keys[0]][keys[1]]...`` or ``default`` if any step
+    is missing / not a dict."""
+    cur: Any = obj
+    for k in keys:
+        if not isinstance(cur, dict) or k not in cur:
+            return default
+        cur = cur[k]
+    return cur
+
+
+def _as_float(x: Any) -> Optional[float]:
+    """Coerce to float or return ``None`` on failure / NaN."""
+    try:
+        f = float(x)
+    except (TypeError, ValueError):
+        return None
+    if not np.isfinite(f):
+        return None
+    return f
+
+
+def causal_violations(result) -> List[Dict[str, Any]]:
+    """Detect assumption / diagnostic violations on a ``CausalResult``.
+
+    Each violation is a dict with keys ``kind`` / ``severity`` /
+    ``test`` / ``value`` / ``threshold`` / ``message`` /
+    ``recovery_hint`` / ``alternatives``.
+
+    ``severity`` is one of ``"error"`` (identifying assumption
+    clearly rejected), ``"warning"`` (borderline / low-power signal),
+    or ``"info"`` (worth mentioning but unlikely to change the
+    conclusion).
+    """
+    from .next_steps import _detect_family  # lazy to avoid cycle
+
+    mi: Dict[str, Any] = result.model_info or {}
+    method_family = _detect_family((result.method or "").lower())
[…]

[…]
