### 2026-04-22 f7f8a15
feat(core): result.to_dict() / for_agent() / to_json() on both result classes
Adds the flat canonical JSON-safe surface that LLM agents and tool
pipelines actually want — complements the existing nested
to_agent_summary() with a shape closer to OpenAI tool-output.

- `CausalResult.to_dict(detail_head=5)` — flat keys (method, estimand,
  estimate, se, pvalue, ci, alpha, n_obs, diagnostics, citation_key),
  plus optional detail_head. Always round-trips through json.dumps.
- `EconometricResults.to_dict()` — includes a term-keyed coefficients
  dict with estimate/std_error/t_statistic/p_value/conf
[…]
