+++ src/statspai/core/results.py
@@ -8,6 +8,72 @@ import numpy as np
 from scipy import stats
 
 
+# ----------------------------------------------------------------------
+# JSON-safe coercion used by to_dict() / for_agent() on the result
+# classes below.  Agents consume these methods; the output MUST round-trip
+# through json.dumps without raising — numpy scalars, pandas objects,
+# tuples, and numpy arrays all need primitive-form conversion.
+# ----------------------------------------------------------------------
+
+def _to_jsonable(value: Any, *, _depth: int = 0, _max_depth: int = 6) -> Any:
+    """Return a JSON-safe representation of ``value``.
+
+    Handles numpy scalars/arrays, pandas Series/DataFrame, tuples/lists,
+    dicts, NaN/Inf (converted to None) and bails out gracefully on
+    anything exotic by falling back to ``str(value)``.
+    """
+    if _depth > _max_depth:
+        return str(value)
+    if value is None:
+        return None
+    if isinstance(value, bool):
+        return bool(value)
+    if isinstance(value, (int, np.integer)):
+        return int(value)
+    if isinstance(value, (float, np.floating)):
+        v = float(value)
+        if np.isnan(v) or np.isinf(v):
+            return None
+        return v
+    if isinstance(value, str):
+        return value
+    if isinstance(value, (list, tuple)):
+        return [_to_jsonable(v, _depth=_depth + 1) for v in value]
+    if isinstance(value, dict):
+        return {str(k): _to_jsonable(v, _depth=_depth + 1)
+                for k, v in value.items()}
+    if isinstance(value, np.ndarray):
+        return _to_jsonable(value.tolist(), _depth=_depth + 1)
+    if isinstance(value, pd.Series):
+        return _to_jsonable(value.to_dict(), _depth=_depth + 1)
+    if isinstance(value, pd.DataFrame):
+        # Keep the payload bounded — agents don't need full detail tables.
+        head = value.head(20)
+        return _to_jsonable(head.to_dict(orient='records'),
+                            _depth=_depth + 1)
+    if isinstance(value, pd.Timestamp):
+        return value.isoformat()
+    # Last resort: stringify so json.dumps doesn't blow up.
+    try:
+        return str(value)
+    except Exception:
+        return None
+
+
+def _filter_jsonable_scalars(d: Dict[str, Any]) -> Dict[str, Any]:
+    """Return the subset of ``d`` whose values round-trip as JSON scalars.
+
+    Used for ``diagnostics`` payloads where agents want simple key/value
+    pairs — not nested frames.
+    """
+    out: Dict[str, Any] = {}
+    for k, v in (d or {}).items():
+        j = _to_jsonable(v)
+        if isinstance(j, (int, float, str, bool)) or j is None:
+            out[str(k)] = j
+    return out
+
+
 class EconometricResults:
     """
     Unified results class for econometric models
@@ -364,6 +430,55 @@ class EconometricResults:
         from .next_steps import econometric_next_steps, _steps_repr_html
         return _steps_repr_html(econometric_next_steps(self))
 
+    def violations(self) -> List[Dict[str, Any]]:
+        """
+        Agent-native structured list of assumption / diagnostic issues.
+
+        Inspects stored diagnostics (first-stage F, standard error
+        finiteness, …) and returns any flagged concerns as dicts with
+        keys ``kind``, ``severity``, ``test``, ``value``, ``threshold``,
+        ``message``, ``recovery_hint``, ``alternatives``.
+
+        Returns
+        -------
+        list of dict
+            Empty list if nothing flagged.
+
+        Examples
+        --------
+        >>> result = sp.iv("y ~ (x ~ z) + c", data=df)
+        >>> for v in result.violations():
+        ...     if v['severity'] == 'error':
+        ...         print(v['recovery_hint'])
+        """
[…]
