### 2026-04-22 f5cf1ae
feat(agent): auto-generated MCP manifest + 5 high-value curated tools
Closes the 99% coverage gap — tool_manifest() grows from 8 hand-curated
entries to 254 merged tools (8 curated + auto-generated from the
889-entry registry).

auto_tools.py:
- Walks the registry for every function whose category is in the
  default whitelist (causal / regression / panel / inference /
  diagnostics / smart / decomposition / robustness / postestimation /
  survival / bayesian / timeseries).
- Uses sp.function_schema() for the JSON schema, then enriches the
  description with agent_card() payload (assumptions /
[…]
