+++ src/statspai/agent/auto_tools.py
@@ -1,0 +1,347 @@
+"""Auto-generate MCP / LLM tool manifest from the StatsPAI registry.
+
+The hand-curated :data:`statspai.agent.tools.TOOL_REGISTRY` covers ~8
+flagship estimators in depth (tuned schemas, bespoke serializers, top-
+quality descriptions).  With 880+ registered functions this leaves most
+of the package invisible to agents — so this module walks the registry
+and synthesises MCP tool specs for every agent-safe estimator it finds.
+
+Design
+------
+
+* **Registry first.** ``sp.function_schema(name)`` already emits OpenAI-
+  compatible schemas for every registered function.  We start from those
+  and turn them into MCP ``{name, description, inputSchema}`` tuples.
+* **Category whitelist.**  We only expose categories that make sense as
+  agent tools — causal, regression, panel, inference, diagnostics, …
+  Plots, CLI helpers, datasets, and pure utilities are skipped so the
+  manifest stays signal-dense.
+* **Agent-card enrichment.**  If :func:`sp.agent_card` returns rich
+  metadata (assumptions, failure modes, alternatives), we append a
+  compact block to the tool description so agents see *when to use* and
+  *what to try when it breaks* right in the tool-list payload.
+* **Hand-curated wins on collision.**  The merged manifest lets every
+  hand-curated spec override its auto-generated counterpart.
+* **Deny list.**  Estimators that need optional heavy deps (torch /
+  jax / pymc) and are known to fail eagerly on ``getattr(sp, name)`` are
+  excluded by name.
+
+The public entry points are :func:`auto_tool_manifest` and
+:func:`merged_tool_manifest`.  ``tool_manifest()`` in ``tools.py``
+delegates to the latter.
+"""
+from __future__ import annotations
+
+from typing import Any, Dict, Iterable, List, Optional, Set
+
+
+# Categories we want reachable by an agent by default.  See
+# ``statspai.help.CATEGORY_DESCRIPTIONS`` for the full list.
+DEFAULT_WHITELIST: Set[str] = {
+    'causal',
+    'regression',
+    'panel',
+    'inference',
+    'diagnostics',
+    'smart',
+    'decomposition',
+    'robustness',
+    'postestimation',
+    'survival',
+    'bayesian',
+    'timeseries',
+}
+
+# Names we explicitly keep out of the auto manifest.  Reasons vary —
+# optional deep-learning dependencies, plotting helpers that slipped a
+# non-plot category, obvious internals, etc.  Keep this list short; for
+# broader exclusions, adjust the category whitelist instead.
+DEFAULT_EXCLUDE: Set[str] = {
+    # Deep-learning — heavy deps (torch / jax)
+    'dragonnet',
+    'deepiv',
+    'tarnet',
+    'cfrnet',
+    # Neural-causal estimators
+    'neural_causal_forest',
+    'cevae',
+    'ganite',
+    # Plots that happen to land in a stats category
+    'event_study_plot',
+    'coefplot',
+    'binscatter',
+    # Misc utilities that slipped into exposed API
+    'describe',
+    'pwcorr',
+    'winsor',
+    'read_data',
+}
+
+# Max length for a tool description sent to the agent.  MCP / Anthropic
+# tool-use enforces no hard limit, but large descriptions eat context
+# for every tool call the agent does, so we cap to keep the manifest
+# lean.
+_MAX_DESCRIPTION_LEN = 1200
+
+
+# ----------------------------------------------------------------------
+# Helpers
+# ----------------------------------------------------------------------
+
+def _is_agent_safe(name: str, spec: Any) -> bool:
+    """Return True iff ``name``/``spec`` should appear in auto manifest.
+
+    Rejects:
+    * leading underscore (private)
[…]
