+++ src/statspai/agent/tools.py
@@ -62,14 +62,26 @@ def _scalar_or_none(v) -> Optional[float]:
 def _default_serializer(r) -> Dict[str, Any]:
     """Serialise a ``CausalResult`` / ``EconometricResults`` to a JSON dict.
 
-    LLM tool-use protocol requires JSON-serialisable return values,
-    so convert floats / arrays / DataFrames to primitives.
-
-    ``CausalResult.estimate`` is a scalar (single ATE) while
-    ``EconometricResults.estimate`` is a Series of coefficients (via the
-    cross-estimator tidy alias).  We detect scalar semantics before
-    calling ``float(...)`` so regressions don't fail JSON serialisation.
+    LLM tool-use protocol requires JSON-serialisable return values.
+    Prefers ``result.to_dict()`` (defined on both result classes) so
+    serialisation logic lives in one place; falls back to the legacy
+    field-by-field extraction for any future result type that doesn't
+    implement ``to_dict``.
     """
+    # Canonical path: both CausalResult and EconometricResults now expose
+    # to_dict() — use it so the agent output matches result.to_dict()
+    # called directly by a user.
+    to_dict = getattr(r, 'to_dict', None)
+    if callable(to_dict):
+        try:
+            out = to_dict()
+            if isinstance(out, dict) and out:
+                return out
+        except Exception:
+            # Fall through to legacy path rather than leak an internal
+            # serialisation error back to the agent as a "tool error".
+            pass
+
     import pandas as _pd
 
     def _is_scalar(v) -> bool:
@@ -483,6 +495,224 @@ TOOL_REGISTRY: List[Dict[str, Any]] = [
             },
         },
     },
+
+    # --------------------------------------------------------------------
+    # Method advisor — one of the most valuable agent entry points
+    # --------------------------------------------------------------------
+    {
+        'name': 'recommend',
+        'description': (
+            "Method advisor: given a dataset + research question, "
+            "recommends a ranked list of estimators with reasoning, "
+            "precondition checks, and a full suggested workflow.  "
+            "This is the first call an agent should make if it "
+            "doesn't know which estimator to run. Supports DAG input, "
+            "mediator / proxy / principal-strata variables, and "
+            "optional resampling-stability verification."
+        ),
+        'input_schema': {
+            'type': 'object',
+            'properties': {
+                'y': {'type': 'string', 'description': 'Outcome column.'},
+                'treatment': {
+                    'type': 'string',
+                    'description': 'Treatment / exposure column.',
+                },
+                'covariates': {
+                    'type': 'array',
+                    'items': {'type': 'string'},
+                    'description': 'Covariate columns.',
+                },
+                'id': {'type': 'string',
+                       'description': 'Unit identifier (panel).'},
+                'time': {'type': 'string',
+                         'description': 'Time column (panel / DID).'},
+                'running_var': {
+                    'type': 'string',
+                    'description': 'Running variable (RD).',
+                },
+                'instrument': {'type': 'string',
+                               'description': 'Instrumental variable.'},
+                'cutoff': {'type': 'number',
+                           'description': 'RD cutoff value.'},
+                'design': {
+                    'type': 'string',
+                    'enum': ['rct', 'did', 'rd', 'iv',
+                             'observational', 'panel', 'cross-section'],
+                    'description': 'Override auto-detected design.',
+                },
+                'verify': {
+                    'type': 'boolean',
+                    'default': False,
+                    'description': ('If True, run resampling-stability '
+                                    'checks on top recommendations.'),
+                },
+            },
+            'required': ['y'],
+        },
+        'statspai_fn': 'recommend',
+        'serializer': lambda r: {
[…]
