### 2026-04-22 049e27b
feat(agent): causal-specific remediation + StatsPAIError bridge + version sync
Remediation registry goes from 11 generic Python-error patterns to 32
entries covering the causal-inference failure space an agent actually
encounters. Every StatsPAI exception now flows its own recovery_hint /
diagnostics / alternative_functions into the remediation payload so
agents get hand-written guidance rather than boilerplate pattern matches.

New causal-specific patterns (regex + category + diagnosis + fix):
  parallel_trends_fail, negative_weights_twfe, mccrary_reject,
  dml_ortho_fail, bayes_convergence, 
[…]
