+++ tests/test_agent.py
@@ -200,6 +200,129 @@ class TestRemediation:
 
 
 # ---------------------------------------------------------------------------
+# Result-object agent surface (to_dict / for_agent / to_json)
+# ---------------------------------------------------------------------------
+
+class TestResultToDict:
+    """EconometricResults.to_dict() / for_agent() / to_json()."""
+
+    def test_econometric_to_dict_has_required_keys(self, sample_df):
+        r = sp.regress('y ~ d + x1', data=sample_df)
+        d = r.to_dict()
+        for k in ('method', 'coefficients', 'diagnostics', 'n_obs',
+                  'glance'):
+            assert k in d, f"{k!r} missing from to_dict()"
+
+    def test_econometric_coefficients_have_full_schema(self, sample_df):
+        r = sp.regress('y ~ d + x1', data=sample_df)
+        d = r.to_dict()
+        assert 'd' in d['coefficients']
+        cell = d['coefficients']['d']
+        for k in ('estimate', 'std_error', 't_statistic', 'p_value',
+                  'conf_low', 'conf_high'):
+            assert k in cell, f"{k!r} missing from coefficient row"
+            assert cell[k] is None or isinstance(cell[k],
+                                                  (int, float, bool))
+
+    def test_econometric_to_json_round_trips(self, sample_df):
+        r = sp.regress('y ~ d + x1', data=sample_df)
+        payload = r.to_json()
+        reloaded = json.loads(payload)
+        assert reloaded['method']
+        assert 'd' in reloaded['coefficients']
+
+    def test_econometric_for_agent_extends_to_dict(self, sample_df):
+        r = sp.regress('y ~ d + x1', data=sample_df)
+        base = r.to_dict()
+        fa = r.for_agent()
+        assert set(base.keys()).issubset(set(fa.keys()))
+        for k in ('violations', 'warnings', 'next_steps',
+                  'suggested_functions'):
+            assert k in fa, f"for_agent missing {k!r}"
+            assert isinstance(fa[k], list)
+
+
+class TestCausalResultToDict:
+    """CausalResult.to_dict() / for_agent() / to_json()."""
+
+    @pytest.fixture(scope='class')
+    def did_result(self):
+        rng = np.random.default_rng(5)
+        rows = []
+        for i in range(200):
+            tr = 1 if i < 100 else 0
+            for t in [0, 1]:
+                y = 1.0 + 0.3*t + 0.5*tr + 2.0*tr*t + rng.normal(scale=0.5)
+                rows.append({'i': i, 't': t, 'treated': tr,
+                             'post': t, 'y': y})
+        df = pd.DataFrame(rows)
+        return sp.did(df, y='y', treat='treated', time='t', post='post')
+
+    def test_causal_to_dict_has_required_keys(self, did_result):
+        d = did_result.to_dict()
+        for k in ('method', 'estimand', 'estimate', 'se', 'pvalue',
+                  'ci', 'alpha', 'n_obs', 'diagnostics'):
+            assert k in d, f"{k!r} missing"
+        assert isinstance(d['ci'], list) and len(d['ci']) == 2
+
+    def test_causal_estimate_close_to_truth(self, did_result):
+        # DGP: ATT = 2.0 with treatment-period interaction
+        d = did_result.to_dict()
+        assert abs(d['estimate'] - 2.0) < 0.2
+
+    def test_causal_to_json_round_trips(self, did_result):
+        payload = did_result.to_json()
+        reloaded = json.loads(payload)
+        assert reloaded['method']
+        assert reloaded['estimate'] is not None
+
+    def test_causal_for_agent_includes_violations(self, did_result):
+        fa = did_result.for_agent()
[…]
