+++ src/statspai/agent/remediation.py
@@ -203,12 +203,386 @@ REMEDIATIONS = [
             "re-running."
         ),
     },
+
+    # ==================================================================
+    # Causal-specific failure modes (P0 expansion, 2026-04-21)
+    # ==================================================================
+
+    # --- DID: parallel trends ---
+    {
+        'match': r"(parallel trend|pre-trend|pretrend).{0,40}(reject|violat|fail)",
+        'category': 'parallel_trends_fail',
+        'diagnosis': (
+            "The parallel-trends assumption is in doubt — a pre-trend "
+            "joint test rejects.  Two-way fixed-effects ATT estimates "
+            "are likely biased."
+        ),
+        'fix': (
+            "Run sp.sensitivity_rr(result) for Rambachan-Roth (2023) "
+            "honest CIs.  Or switch to sp.callaway_santanna / "
+            "sp.did_imputation which are robust to heterogeneous "
+            "effects.  Consider dropping the longest-pre-period cohorts."
+        ),
+    },
+
+    # --- DID: negative weights under TWFE ---
+    {
+        'match': r"(negative weight|forbidden comparison|bacon.*decomp)",
+        'category': 'negative_weights_twfe',
+        'diagnosis': (
+            "Two-way FE gives negative weights to some 2x2 comparisons "
+            "(Goodman-Bacon / de Chaisemartin).  Under heterogeneous "
+            "effects the aggregate ATT estimate can flip sign."
+        ),
+        'fix': (
+            "Switch to sp.callaway_santanna (group-time ATT) or "
+            "sp.did_multiplegt (de Chaisemartin).  Run sp.bacon_decomposition "
+            "to quantify how bad the weighting is."
+        ),
+    },
+
+    # --- RD: McCrary / manipulation ---
+    {
+        'match': r"(mccrary|manipulation|density discontinuity).{0,40}(reject|significant|detect)",
+        'category': 'mccrary_reject',
+        'diagnosis': (
+            "Running-variable density has a discontinuity at the "
+            "cutoff (McCrary test rejects).  Sorting/manipulation "
+            "undermines the continuity assumption — estimates may be "
+            "confounded with selection."
+        ),
+        'fix': (
+            "Try sp.rd_donut to exclude units just around the cutoff, "
+            "or sp.bounds for partial identification.  Investigate the "
+            "mechanism — administrative rounding often causes benign "
+            "discontinuities."
+        ),
+    },
+
+    # --- DML: orthogonality / score mean ≠ 0 ---
+    {
+        'match': r"(orthogonal.{0,20}(fail|violat)|score mean.{0,10}≠|score mean.{0,10}!=|dml.*bias)",
+        'category': 'dml_ortho_fail',
+        'diagnosis': (
+            "Double-ML nuisance estimates did not yield a zero-mean "
+            "orthogonal score — the debiasing step is compromised."
+        ),
+        'fix': (
+            "Increase cross-fitting folds (n_folds=10+), use richer "
+            "nuisance learners (GradientBoosting / XGBoost), and check "
+            "propensity-score overlap.  For small n consider sp.tmle."
+        ),
+    },
+
+    # --- Bayesian convergence ---
+    {
+        'match': r"(rhat.{0,6}>|max.?rhat|ess.{0,8}<|divergen|target_accept)",
+        'category': 'bayes_convergence',
+        'diagnosis': (
+            "MCMC convergence diagnostics failed: rhat > 1.01, low ESS, "
+            "or post-warmup divergences.  Posterior summaries are "
+            "unreliable."
+        ),
+        'fix': (
+            "Raise tune and draws (tune=4000, draws=4000).  Increase "
+            "target_accept to 0.95+.  For funnel geometries use non-"
[…]
