### 2026-04-22 0c7413c
docs(guides): ## For Agents blocks rendered from registry agent cards
Each choosing-* guide now carries a structured ## For Agents block
generated by `_agent_docs.render_agent_block(name)` from the per-
function agent_card. Agents parse the block; humans read it in
MkDocs. Single source of truth — the rendered block stays in sync
with the card via tests/test_agent_blocks_drift.py.

Sections in each block:
  - Pre-conditions (data-shape requirements the agent should verify)
  - Identifying assumptions (parallel trends, exclusion, overlap, …)
  - Failure modes → recovery (symptom / exception / r
[…]
