### 2026-04-22 5d0c082
chore(agent): scripts/sync_agent_blocks.py + CI drift guard
Keeps docs/guides/*.md ## For Agents blocks in lockstep with the
registry agent cards. `--check` mode (used by the CI test) fails if
any block is stale; bare invocation rewrites the blocks in place.

tests/test_agent_blocks_drift.py runs the script in --check mode —
guards against drift between registry metadata and the human-facing
guide pages.
