### 2026-04-22 ef5358d
docs(readme): fix PyPI broken image + surface v1.5.0 highlights table
- image-1.png switched from relative (`./image-1.png`) to absolute
  raw.githubusercontent URL so PyPI renders it instead of showing a
  broken image icon; GitHub continues to render it fine.
- v1.5.0 section now carries a "v1.5 Highlights" table with visual
  weight matching the existing v1.4 Highlights table, so the current
  release is the README's visual subject instead of being outweighed
  by the previous-version section immediately below it.
- README_CN.md updated in lockstep.

Co-Authored-By: Claude Opus 4.7 (1M cont
[…]
