+++ README_CN.md
@@ -49,12 +49,14 @@ StatsPAI 聚焦**因果推断**——在这条主线上，我们的目标是成�
 
 StatsPAI 1.5.0 是 minor 版本，一次性完成 interference、conformal 因果推断、Mendelian 随机化三个家族的三项联动升级：完整家族文档指南、与 `sp.synth` / `sp.decompose` / `sp.dml` 同构的统一 dispatcher，以及一次针对性的正确性审计修掉两处"悄悄算错"。
 
-- **三篇新家族指南**：`docs/guides/interference_family.md`（`sp.spillover`、`sp.network_exposure`、`sp.peer_effects`、`sp.network_hte`、`sp.inward_outward_spillover`、`sp.cluster_matched_pair`、`sp.cluster_cross_interference`、`sp.cluster_staggered_rollout`、`sp.dnc_gnn_did` 完整走查，含决策树 + 5 项 interference 分析必报诊断）；`docs/guides/conformal_family.md`（10 个 conformal 估计器按"边际覆盖保证的含义"重新组织，明确 marginal 与 conditional 的区别）；`docs/guides/mendelian_family.md`（17 个 MR 函数按 IV1 / IV2 / IV3 假设层级重新组织 + 每次 MR 分析必报的 4 项 sanity check + BMI → T2D 完整示例）。
-- **三个新统一 dispatcher**：`sp.mr(method=...)` 覆盖 33 个别名（IVW / Egger / median / mode / MVMR / mediation / BMA / PRESSO / radial / Steiger / F 统计量 / LOO / 多效性 / 异质性）；`sp.conformal(kind=...)` 覆盖 29 个别名（CATE / counterfactual / ITE / density / 多阶段 / debiased-ML / fair / continuous / interference）；`sp.interference(design=...)` 覆盖 29 个别名（partial / network-exposure / peer-effects / network-HTE / inward-outward / 4 种 cluster-RCT 变体）。kwargs 原样透传；每个 dispatcher 的输出与直调结果逐字节一致（30 个新测试守卫）。
-- **⚠️ 正确性修复 — `sp.mr_egger` slope 推断错用 Normal 而非 t(n−2)**。配套的 `sp.mr_pleiotropy_egger` 用的是 `t(n−2)`，但 `mr_egger` 的 slope p 值和 CI 临界值都用 `stats.norm`，两处口径不一致。小样本下反而过松——`n_snps = 5`、`t ≈ 1.5` 时 Normal 给 `p = 0.134`，正确的 `t(3)` 给 `p = 0.231`。修复后两处都用 `t(n − 2)`。`n_snps ≥ ~100` 时不可见。回归守卫：`tests/test_correctness_v150.py::TestMREggerUsesTDistribution`。
-- **⚠️ 正确性修复 — `sp.mr_presso` MC p 值可能等于 0**。全局检验和逐 SNP outlier p 值都用原始 `mean(null ≥ obs)`，当实测统计量大于所有模拟 null 时会塌缩为 `0.0`。MC 估计 p 值的真实下界是 `1 / (B + 1)`，不可能为零。修复改用标准 `(k + 1) / (B + 1)`（与 R 的 MR-PRESSO 一致）。下游影响：`log(p)` 和灵敏度变换不会再静默返回 `-inf`。回归守卫：`tests/test_correctness_v150.py::TestMRPressoMCPvalueConvention`。
[…]
