+++ README.md
@@ -49,12 +49,14 @@ StatsPAI's focus is **causal inference** — and on this axis we aim to be the m
 
 StatsPAI 1.5.0 is a minor release bundling three concurrent improvements to the interference, conformal causal inference, and Mendelian Randomization families: full-family documentation guides, unified dispatchers matching the `sp.synth` / `sp.decompose` / `sp.dml` pattern, and a targeted correctness audit that fixed two silent-wrong-numbers issues.
 
-- **Three new family guides** — `docs/guides/interference_family.md` (complete walkthrough of `sp.spillover`, `sp.network_exposure`, `sp.peer_effects`, `sp.network_hte`, `sp.inward_outward_spillover`, `sp.cluster_matched_pair`, `sp.cluster_cross_interference`, `sp.cluster_staggered_rollout`, `sp.dnc_gnn_did` with decision tree + 5 diagnostics every interference analysis should report), `docs/guides/conformal_family.md` (all 10 conformal estimators organised around what the marginal-coverage guarantee actually buys you), `docs/guides/mendelian_family.md` (all 17 MR functions structured around the IV1 / IV2 / IV3 assumption hierarchy + 4 sanity checks every MR analysis should report + a worked BMI → T2D example).
-- **Three new unified dispatchers** — `sp.mr(method=...)` (33 aliases covering IVW / Egger / median / mode / MVMR / mediation / BMA / PRESSO / radial / Steiger / F-stat / LOO / pleiotropy / heterogeneity), `sp.conformal(kind=...)` (29 aliases across CATE / counterfactual / ITE / density / multi-DP / debiased-ML / fair / continuous / interference), `sp.interference(design=...)` (29 aliases across partial / network-exposure / peer-effects / network-HTE / inward-outward / 4× cluster-RCT variants). Kwargs pass through unchanged; each dispatcher's output is byte-for-byte identical to the direct call (30 new tests guarantee this).
-- **⚠️ correctness fix — `sp.mr_egger` slope inference used Normal, not t(n−2).** The companion `sp.mr_pleiotropy_egger` correctly used `t(n−2)` for the Egger intercept p-value, but `mr_egger` itself used `stats.norm` for both the slope p-value and CI critical value. Anti-conservative at small `n_snps` — e.g. `n_snps = 5` and `t ≈ 1.5` gave `p = 0.134` instead of the correct `p = 0.231`. The fix switches both to `t(n − 2)`. For `n_snps ≥ ~100` the change is numerically invisible. Regression guard: `tests/test_correctness_v150.py::TestMREggerUsesTDistribution`.
-- **⚠️ correctness fix — `sp.mr_presso` MC p-value could equal 0.** Both the global test and the per-SNP outlier p-values used the raw `mean(null ≥ obs)` form, which collapses to `0.0` when the observed statistic exceeds every simulated null. An MC-estimated p-value cannot be zero — its true lower bound is `1 / (B + 1)`. Fix switches to the standard `(k + 1) / (B + 1)` convention (matching R's MR-PRESSO). Downstream effect: `log(p)` and sensitivity transforms no longer silently produce `-inf`. Regression guard: `tests/test_correctness_v150.py::TestMRPressoMCPvalueConvention`.
[…]
