### 2026-04-22 e217576
release: v1.5.1 — PyPI readme hotfix + unshipped-import cleanup
Patch release. No code / behaviour changes vs. v1.5.0.

- Fix PyPI project-page broken image: README's interactive-plot-editor
  screenshot was referenced by relative path (./image-1.png), which
  GitHub resolves but PyPI cannot, so the PyPI v1.5.0 page showed a
  broken-image icon. Both README.md and README_CN.md now use the
  absolute raw.githubusercontent.com URL so both renderers work.

- Remove premature v1.6 imports from src/statspai/__init__.py that
  were introduced by commit 8155ada but whose implementations had not
  yet
[…]
