### 2026-04-22 c87af29
feat(v1.6): P1 agent-native × frontier — LLM-DAG closed loop, sp.paper(), causal_text MVP, MR frontier
Pure-additive release pushing two competitive axes at once.

Agent-native:
- P1-A `sp.llm_dag_constrained` — closed loop propose → constrained PC →
  CI-test validate → demote → iterate, with per-edge llm_score +
  ci_pvalue + source.
- P1-A `sp.llm_dag_validate` — per-edge CI-test audit of any DAG.
- P1-A `sp.pc_algorithm(forbidden=, required=)` — background-knowledge
  constraints in PC, default-None preserves prior contract bit-for-bit.
- P1-C `sp.paper(data, question, ...)` — end-to-end "
[…]
