+++ docs/guides/paper_pipeline.md
@@ -1,0 +1,166 @@
+# `sp.paper()` — Data → Publication-Draft Pipeline
+
+> One call from a DataFrame + a natural-language question to a
+> publication-quality draft (markdown, LaTeX, or Word) with EDA,
+> identification check, recommended estimator, fitted result, and
+> robustness suite — pre-assembled.
+
+## When to use
+
+- You have data and a research question; you want **a draft you can
+  edit**, not a slide deck.
+- You're orchestrating an **autonomous agent** that needs to produce
+  empirical reports without holding the analyst's hand through each
+  stage.
+- You want a single artifact that captures *why* the estimator was
+  chosen (identification + assumptions), not just the point estimate.
+
+If you only need the fitted model → use the matched estimator directly
+(`sp.regress`, `sp.did`, `sp.iv`, ...). If you want the full workflow
+without rendering → use `sp.causal()`.
+
+## Quick start
+
+```python
+import statspai as sp
+import pandas as pd
+
+df = pd.read_stata("nlsw88.dta")
+
+draft = sp.paper(
+    df,
+    "effect of training on wage",
+    treatment="training",
+    y="wage",
+    covariates=["education", "experience"],
+)
+
+print(draft.to_markdown())   # human readable
+draft.write("nlsw88_draft.tex")  # LaTeX
+draft.write("nlsw88_draft.docx") # Word (requires python-docx)
+```
+
+## What you get
+
+A `PaperDraft` with **seven sections**:
+
+| Section | Contents |
+|---------|----------|
+| Question | The natural-language question + parsed outcome / treatment / design. |
+| Data | Sample size, missingness top-5, outcome moments, treatment distribution, mean covariates by arm, std-diff. |
+| Identification | Verdict from `sp.check_identification` plus per-finding severity / fix. |
+| Estimator | Recommended method + `sp.<fn>()` call + rationale + key assumptions. |
+| Results | Estimate, SE, 95% CI, p-value, n_obs (or coefficient table for regressions). |
+| Robustness | Method-specific suite (e.g. parallel trends + e-value for DiD; first-stage F + AR-CI for IV). |
+| References | BibTeX-style entries pulled from `result.cite()`. |
+
+Plus an extra **Pipeline notes** section if any stage failed (so the
+draft never silently hides errors).
+
+## The question parser
+
+`sp.paper()` runs a lightweight parser to extract column hints from the
+natural-language question — **but explicit kwargs always win**.
+
+| Pattern | Example | Hints |
+|---------|---------|-------|
+| "effect of X on Y" | "effect of training on wage" | treatment=training, y=wage |
+| "Y ~ X" | "wage ~ training" | y=wage, treatment=training |
+| "difference-in-differences" / "did" | "DiD effect of ..." | design='did' |
+| "regression discontinuity" / "rd" | "RD with running variable score" | design='rd', running_var='score' |
+| "instrument(ing)? <Z>" | "using distance as an instrument" | design='iv', instrument='distance' |
+| "discontinuity at <c>" / "threshold <c>" | "threshold at 0.5" | cutoff=0.5, design='rd' |
+
+The parser **never overrides** explicit args — it only fills in what
+you didn't pass. To bypass it entirely, pass everything explicitly.
+
+```python
+draft = sp.paper(
+    df,
+    question="(any string)",  # parser ignored when args complete
+    y="wage", treatment="training", design="did",
+    time="year", id="worker_id",
+)
+```
+
+## Rendering formats
+
+### Markdown
+The default. Easy to render in Jupyter, GitHub, or as a starting point
+for a PR description.
+
+```python
+draft.to_markdown()
+```
+
+### LaTeX
[…]
