### 2026-04-22 1833274
release: v1.6.0 — P1 Agent-Native × Frontier + agent-native infrastructure
Minor release pushing two competitive axes simultaneously.

Agent-native:
- sp.paper() end-to-end publication pipeline (PaperDraft → md/tex/docx)
- sp.llm_dag_constrained / sp.llm_dag_validate closed-loop LLM-DAG
- sp.pc_algorithm(forbidden=, required=) constrained PC
- sp.text_treatment_effect / sp.llm_annotator_correct (causal_text MVP)
- Typed exception taxonomy: StatsPAIError + 6 subclasses, each with
  recovery_hint / diagnostics / alternative_functions; 13 call-site
  migrations across DID/IV/matching/DML-IRM/synt
[…]
