### 2026-04-22 d29eef9
docs(contributing): add authorship & acknowledgments policy
Every merged PR is automatically credited in CONTRIBUTORS.md and the
JOSS paper's Acknowledgments. Paper authorship is separate and
extended by invitation against four objective criteria (new estimator
family, reference-parity ownership, sustained core contribution, or
paper writing), aligned with ICMJE / JOSS norms.

Adds CONTRIBUTORS.md seed file and a prominent callout in the "Types
of Contributions" block so new contributors see the policy before
opening a PR.
