+++ CONTRIBUTORS.md
@@ -1,0 +1,22 @@
+# Contributors
+
+StatsPAI is built by the people listed here. Every merged pull request
+earns a line in this file and a slot in the Acknowledgments section of
+the project's JOSS paper.
+
+> Paper **authorship** is separate and extended by invitation against
+> objective criteria — see the "Authorship & Acknowledgments" section
+> of [`CONTRIBUTING.md`](CONTRIBUTING.md#-authorship--acknowledgments).
+
+## Core Maintainers
+
+- **Biaoyue (Bryce) Wang** ([@brycewang-stanford](https://github.com/brycewang-stanford)) — project lead, architecture, most estimator modules.
+
+## Contributors
+
+<!--
+Format: - **Name** ([@handle](https://github.com/handle)) — area of contribution (e.g., "did module", "docs", "reference parity for sp.rd").
+New entries go in alphabetical order by handle. Added automatically on PR merge.
+-->
+
+_Be the first! Open a PR and your name lands here._
