### 2026-04-22 d9565a0
docs(contributing): simplify authorship clause to maintainer discretion
Drop the four-criterion rubric; keep authorship as an invitation
extended at the maintainers' discretion, case by case. Acknowledgment
tier (auto-credit for every merged PR) unchanged. Rename the second
CONTRIBUTORS.md heading to avoid MD024 duplicate-heading warning.
