### 2026-04-22 5e4152a
docs(contributing): signal intent behind authorship invitations
Keep maintainer sole discretion unchanged; add one clause naming the
spirit of the bar ("substantial intellectual contribution to the
library or to the paper itself") so potential collaborators know what
direction to aim without turning it into a rubric.
