### 2026-04-22 0eca784
docs(citations): correct upstream attributions for 5 estimators
Verified each reference via arXiv / NBER primary sources. Changes are
metadata-only — algorithm paths, signatures, and numerics unchanged.

* bcf_longitudinal (arXiv:2508.08418): "Alessi, Zorzetto et al." ->
  Prevot, Haring, Nichols, Holmes & Ganjgahi (2025)
* surrogate_index (NBER WP 26463): "Pollmann, Taubinsky" ->
  Athey, Chetty, Imbens & Kang (2019)
* proximal_surrogate_index + bridge(surrogate_pci): Imbens-Kallus-Mao
  arXiv:2601.17712 (2026) -> Imbens, Kallus, Mao & Wang,
  JRSS-B 87(2) 2025, arXiv:2202.07234 (fix arXiv ID
[…]
