+++ src/statspai/bayes/dml.py
@@ -1,4 +1,4 @@
-"""Bayesian Double Machine Learning (arXiv:2508.12688, 2025).
+"""Bayesian Double Machine Learning (DiTraglia & Liu, arXiv:2508.12688, 2025).
 
 Wraps :func:`sp.dml` cross-fitted residuals in a Normal-Normal conjugate
 posterior, turning frequentist DML into a fully Bayesian analogue with
@@ -120,7 +120,10 @@ def bayes_dml(
 
     References
     ----------
-    Chernozhukov et al. (arXiv:2508.12688, 2025).
+    DiTraglia, F. J., & Liu, L. (2025). "Bayesian Double Machine
+    Learning for Causal Inference." arXiv:2508.12688.
+    (The underlying orthogonal-moments DML construction is due to
+    Chernozhukov et al. 2018, Econometrics Journal.)
     """
     if mode not in ("conjugate", "full"):
         raise ValueError(f"mode must be 'conjugate' or 'full'; got {mode!r}.")

[…]
