### 2026-04-22 4990800
docs(citations): fix 49 wrong attributions + add automated auditor
Systematic audit of all arXiv / NBER / DOI references in src/ and docs/
found 49 unique IDs attributed to the wrong authors — predominantly in
post-2024 citations where LLM-generated attributions diverged from
arXiv / NBER primary sources.

Build: new tools/audit_citations.py (stdlib-only, ~790 lines)

  - EXTRACT  regex over src/ + docs/ (arXiv IDs, NBER WP, DOIs)
  - VERIFY   batch arXiv API, NBER meta scrape, Crossref JSON, all
             cached on disk under tools/.citation_cache/
  - DIFF     unicode-aware last-name matc
[…]
