+++ src/statspai/causal_text/__init__.py
@@ -23,7 +23,7 @@ methods alongside topic-model integration.
 
 References
 ----------
-- Veitch, V., Wang, Y., & Blei, D. M. (2020). "Adapting text embeddings
+- Veitch, V., Sridhar, D., & Blei, D. M. (2019). "Adapting text embeddings
   for causal inference."  *UAI*.  arXiv:1905.12741.
 - Egami, N., Hinck, M., Stewart, B., & Wei, H. (2024). "Using
   imperfect surrogates for downstream inference: Design-based
