+++ src/statspai/registry.py
@@ -2705,7 +2705,7 @@ def _build_registry():
             "surrogates_waves=[['s1'],['s2','s3']], long_term_outcome='Y')"
         ),
         tags=["surrogate", "long_term", "multi_wave"],
-        reference="Ghassami, Yang, Shpitser, Tchetgen Tchetgen (arXiv:2311.08527, 2024).",
+        reference="Tran, Bibaut & Kallus (arXiv:2311.08527, 2023).",
     ))
 
     # -- Next-gen evidence synthesis (RCT + RWD + AI/ML) ------------ #
@@ -2949,7 +2949,7 @@ def _build_registry():
         ],
         returns="SharpOPEResult",
         tags=["ope", "sensitivity", "sharp", "bandit"],
-        reference="Kallus, Mao, Uehara (arXiv:2502.13022, 2025).",
+        reference="Hess, Frauen, Melnychuk & Feuerriegel (arXiv:2502.13022, 2025).",
     ))
     register(FunctionSpec(
         name="causal_policy_forest",
@@ -2991,7 +2991,7 @@ def _build_registry():
         ],
         returns="NetworkHTEResult",
         tags=["interference", "network", "hte", "orthogonal"],
-        reference="Parmigiani et al. (arXiv:2509.18484, 2025).",
+        reference="Wu & Yuan (arXiv:2509.18484, 2025).",
     ))
     register(FunctionSpec(
         name="inward_outward_spillover",
@@ -3009,7 +3009,7 @@ def _build_registry():
         ],
         returns="InwardOutwardResult",
         tags=["interference", "spillover", "directional"],
-        reference="Li, Ratkovic et al. (arXiv:2506.06615, 2025).",
+        reference="Fang, Airoldi & Forastiere (arXiv:2506.06615, 2025).",
     ))
 
     # -- Bayesian Double Machine Learning ---------------------------- #
@@ -3567,7 +3567,7 @@ def _build_registry():
             "protected='gender')"
         ),
         tags=["fairness", "preprocessing", "residualize"],
-        reference="Marchesin et al. (arXiv:2403.17852v3, 2025).",
+        reference="Chen & Zhu (arXiv:2403.17852v3, 2024).",
     ))
 
     register(FunctionSpec(
@@ -3716,7 +3716,7 @@ def _build_registry():
         example=('sp.dml_model_averaging(df, y="y", treat="d", '
                  'covariates=[f"x{j}" for j in range(10)])'),
         tags=["dml", "causal", "model_averaging", "ensemble", "plr"],
-        reference="Ahrens, Hansen, Kurz, Schaffer, Wiemann (2025). JAE 40(3):381-402. DOI 10.1002/jae.3103.",
+        reference="Ahrens, Hansen, Schaffer & Wiemann (2025). JAE 40(3):249-269. DOI 10.1002/jae.3103.",
     ))
 
     # -- v1.7 long-panel DML (Semenova-Chernozhukov 2023) -------------- #
@@ -3852,7 +3852,7 @@ def _build_registry():
         returns="ContinuousLATEResult",
         example='sp.continuous_iv_late(df, y="y", treat="d", instrument="z", n_quantiles=5)',
         tags=["iv", "late", "continuous-instrument", "complier"],
-        reference="Xie et al. (2025). arXiv:2504.03063.",
+        reference="Zeng et al. (2025). arXiv:2504.03063.",
     ))
 
     register(FunctionSpec(
@@ -3884,7 +3884,7 @@ def _build_registry():
         returns="CausalResult",
         example='sp.hal_tmle(df, y="y", treat="d", covariates=["x1","x2","x3"])',
         tags=["tmle", "hal", "semiparametric", "causal", "double-robust"],
[…]
