### 2026-04-23 28cbdfc
docs(citations): Phase 2 bibkey annotations + paper.bib Crossref expansion
Annotate docstring References across 227 modules with [@bibkey] links
pointing back to paper.bib, and expand paper.bib with Phase 2 auto-
verified entries. Rename cattaneo2024multi → cattaneo2021extrapolating
(JASA, DOI 10.1080/01621459.2020.1751646) and add cattaneo2024binscatter
(AER, DOI 10.1257/aer.20221576) to match the two distinct Cattaneo et al.
papers the docstrings cite.

Refs verified via Crossref + doi.org per CLAUDE.md §10.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
