+++ paper.bib
@@ -386,16 +386,28 @@
   doi={10.1016/j.jeconom.2007.05.005}
 }
 
-@article{cattaneo2024multi,
-  title={Analysis of regression-discontinuity designs with multiple cutoffs or multiple scores},
+@article{cattaneo2021extrapolating,
+  title={Extrapolating treatment effects in multi-cutoff regression discontinuity designs},
   author={Cattaneo, Matias D. and Keele, Luke and Titiunik, Roc{\'\i}o and Vazquez-Bare, Gonzalo},
-  journal={Journal of Econometrics},
-  volume={238},
-  number={1},
-  pages={105546},
+  journal={Journal of the American Statistical Association},
+  volume={116},
+  number={536},
+  pages={1941--1952},
+  year={2021},
+  publisher={Taylor \& Francis},
+  doi={10.1080/01621459.2020.1751646}
+}
+
+@article{cattaneo2024binscatter,
+  title={On binscatter},
+  author={Cattaneo, Matias D. and Crump, Richard K. and Farrell, Max H. and Feng, Yingjie},
+  journal={American Economic Review},
+  volume={114},
+  number={5},
+  pages={1488--1514},
   year={2024},
-  publisher={Elsevier},
-  doi={10.1016/j.jeconom.2023.105546}
+  publisher={American Economic Association},
+  doi={10.1257/aer.20221576}
 }
 
 @article{keele2015geographic,
@@ -514,5 +526,3120 @@
   number={1},
   pages={1--48},
   year={2015},
-  doi={10.18637/jss.v67.i01}
+  doi={10.18637/jss.v067.i01}
+}
+
+% === Phase 2 auto-verified entries from bulk Crossref verification ===
+
+@article{cameron2008bootstrap,
+  title={Bootstrap-Based Improvements for Inference with Clustered Errors},
+  author={Cameron, A. Colin and Gelbach, Jonah B. and Miller, Douglas L.},
+  journal={Review of Economics and Statistics},
+  year={2008},
+  doi={10.1162/rest.90.3.414}
+}
+
+@techreport{doudchenko2016balancing,
+  title={Balancing, Regression, Difference-In-Differences and Synthetic Control Methods: A Synthesis},
+  author={Doudchenko, Nikolay and Imbens, Guido},
+  year={2016},
+  doi={10.3386/w22791}
+}
+
+@article{mundlak1978pooling,
+  title={On the Pooling of Time Series and Cross Section Data},
+  author={Mundlak, Yair},
+  journal={Econometrica},
+  year={1978},
+  doi={10.2307/1913646}
+}
+
+@article{benmichael2022synthetic,
+  title={Synthetic Controls with Staggered Adoption},
+  author={Ben-Michael, Eli and Feller, Avi and Rothstein, Jesse},
+  journal={Journal of the Royal Statistical Society Series B: Statistical Methodology},
+  year={2022},
+  doi={10.1111/rssb.12448}
+}
+
+@article{breusch1980lagrange,
+  title={The Lagrange Multiplier Test and its Applications to Model Specification in Econometrics},
+  author={Breusch, T. S. and Pagan, A. R.},
+  journal={The Review of Economic Studies},
+  year={1980},
+  doi={10.2307/2297111}
+}
+
+@article{calonico2020optimal,
+  title={Optimal bandwidth choice for robust bias-corrected inference in regression discontinuity designs},
+  author={Calonico, Sebastian and Cattaneo, Matias D and Farrell, Max H},
+  journal={The Econometrics Journal},
+  year={2020},
+  doi={10.1093/ectj/utz022}
+}
+
+@article{chamberlain1982multivariate,
+  title={Multivariate regression models for panel data},
+  author={Chamberlain, Gary},
[…]
