### 2026-04-23 2d0c381
docs(citations): fix wrong attributions + expand paper.bib with verified entries
- fairness/core.py: replace Marchesin et al. 2025 with Chen & Zhu 2024 (arXiv:2403.17852)
- diagnostics/estat.py: fix Breusch 1978 author initials (L.G. -> T.S.) + full title
- registry.py: correct conformal_dose (Schroeder et al. 2024), network_hte (Wu & Yuan 2025),
  inward_outward (Fang/Airoldi/Forastiere 2025), synth_survival (Han & Shah 2025),
  causal_dqn (Li/Zhang/Bareinboim 2025) with arXiv IDs
- synth/__init__.py: align synth_survival attribution with registry
- paper.bib: +860 lines of Crossref/arXiv-ver
[…]
