+++ paper.bib
@@ -3643,3 +3643,863 @@
   howpublished={arXiv preprint arXiv:2305.00050},
   doi={10.48550/arXiv.2305.00050}
 }
+
+@article{belloni2012sparse,
+  title={Sparse Models and Methods for Optimal Instruments With an Application to Eminent Domain},
+  author={Belloni, Alexandre and Chen, Daniel and Chernozhukov, Victor and Hansen, Christian},
+  journal={Econometrica},
+  volume={80},
+  number={6},
+  pages={2369--2429},
+  year={2012},
+  doi={10.3982/ECTA9626}
+}
+
+@article{chernozhukov2013inference,
+  title={Inference on Counterfactual Distributions},
+  author={Chernozhukov, Victor and Fern{\'a}ndez-Val, Iv{\'a}n and Melly, Blaise},
+  journal={Econometrica},
+  volume={81},
+  number={6},
+  pages={2205--2268},
+  year={2013},
+  doi={10.3982/ECTA10582}
+}
+
+@misc{long2023causal,
+  title={Causal Discovery with Language Models as Imperfect Experts},
+  author={Long, Stephanie and Pich{\'e}, Alexandre and Zantedeschi, Valentina and Schuster, Tibor and Drouin, Alexandre},
+  year={2023},
+  howpublished={arXiv preprint arXiv:2307.02390},
+  doi={10.48550/arXiv.2307.02390}
+}
+
+@misc{chen2024counterfactual,
+  title={Counterfactual Fairness through Transforming Data Orthogonal to Bias},
+  author={Chen, Shuyi and Zhu, Shixiang},
+  year={2024},
+  howpublished={arXiv preprint arXiv:2403.17852},
+  doi={10.48550/arXiv.2403.17852}
+}
+
+@misc{prevot2025hierarchical,
+  title={A hierarchical modelling approach for Bayesian Causal Forests on longitudinal data: A Case Study in Multiple Sclerosis Clinical Trials},
+  author={Prevot, Emma and H{\"a}ring, Dieter A. and Nichols, Thomas E. and Holmes, Chris C. and Ganjgahi, Habib},
+  year={2025},
+  howpublished={arXiv preprint arXiv:2508.08418},
+  doi={10.48550/arXiv.2508.08418}
+}
+
+@misc{abadie2025synthetic,
+  title={Synthetic Controls for Experimental Design},
+  author={Abadie, Alberto and Zhao, Jinglong},
+  year={2025},
+  howpublished={arXiv preprint arXiv:2108.02196 (v5, April 2025)},
+  doi={10.48550/arXiv.2108.02196}
+}
+
+@article{acemoglu2001consequences,
+  title={Consequences of Employment Protection? The Case of the {Americans with Disabilities Act}},
+  author={Acemoglu, Daron and Angrist, Joshua D.},
+  journal={Journal of Political Economy},
+  volume={109},
+  number={5},
+  pages={915--957},
+  year={2001},
+  doi={10.1086/322836}
+}
+
+@misc{han2025synthetic,
+  title={Synthetic Survival Control: Extending Synthetic Controls for ``When-If'' Decision},
+  author={Han, Jessy Xinyi and Shah, Devavrat},
+  year={2025},
+  howpublished={arXiv preprint arXiv:2511.14133},
+  doi={10.48550/arXiv.2511.14133}
+}
+
+@misc{peng2024local,
+  title={Local Balance Calibration for Nonparametric Propensity Score Estimation},
+  author={Peng, Maosen and Li, Yan and Wu, Chong and Li, Liang},
+  year={2024},
+  howpublished={arXiv preprint arXiv:2404.04794},
+  doi={10.48550/arXiv.2404.04794}
+}
+
+@article{austin2011introduction,
[…]
