### 2026-04-23 fa35662
ci(citations): gate PR/push on auditor + 3 auditor fixes
Closes the §10 zero-hallucination loop: every push to main / PR that
touches src/ docs/ paper.bib paper.md or the auditor now runs
tools/audit_citations.py against arXiv / NBER / Crossref live. Mismatch
or unresolved id (via --strict) fails the build; full report is
uploaded as the citation-audit-report artifact for reviewers.

Three auditor bugs surfaced while verifying the pre-CI state and fixed
inline:

1. DOI regex ate the closing '}' when a DOI sat inside a bibtex-style
   'doi={10.xxxx/...}' literal (e.g. src/statspai/rd/extrapolat
[…]
