### 2026-04-23 9e46226
docs(readme): add Quick Start with 4 verified causal examples
Inserted a "Quick Start — 60 seconds" section between the positioning
block and the feature list. Four snippets, each locally verified:

- DiD via sp.callaway_santanna + sp.aggte on mpdta (ATT ≈ -0.033)
- IV via sp.ivreg on card_1995 (educ ≈ 0.142, first-stage F ≈ 160)
- RD via sp.rdrobust on lee_2008_senate (estimate ≈ 0.062)
- SCM via sp.synth on california_prop99 (ATT ≈ -13.3 packs/capita)

Uses bundled sp.datasets (mpdta / card_1995 / lee_2008_senate /
california_prop99) so snippets run offline with no data setup.
