### 2026-04-23 5bc6f15
docs(readme): hoist Quick Start to top + per-result diagnostics recap
Moved the "Quick Start — 60 seconds" block from its prior position
(below "What StatsPAI is — and is not") to the very top of the README,
immediately after the intro paragraphs — before the coverage matrix and
version history — so new readers hit runnable code within one scroll.

Added a new sub-section "What each result object gives you for free"
with four bullet points recapping the inference scaffolding that every
.summary() call prints beyond the point estimate:

- DiD: uniform crit val, balanced/min/max e bookkeeping, 1
[…]
