### 2026-04-23 8ce7574
fix(bib): dedup paper.bib (2 dup keys + 17 redundant) + add dedup auditor
paper.bib had accumulated two integrity bugs that CLAUDE.md §10
"paper.bib single source" was silently violating:

  1. Two keys defined twice — biblatex error class:
       - correia2020fast  (lines 1124 + 4496)
       - imai2014covariate (lines 1507 + 3933)
     For each we kept the richer entry (volume/number/pages/note) and
     dropped the shorter one.

  2. Seventeen pairs of different keys pointing at the same DOI. Each
     causes citation drift: half the codebase cites @a, half cites @b,
     for the same paper.
[…]
