+++ paper.bib
@@ -34,18 +34,6 @@
   doi={10.1016/j.jeconom.2020.12.001}
 }
 
-@article{goodman2021difference,
-  title={Difference-in-differences with variation in treatment timing},
-  author={Goodman-Bacon, Andrew},
-  journal={Journal of Econometrics},
-  volume={225},
-  number={2},
-  pages={254--277},
-  year={2021},
-  publisher={Elsevier},
-  doi={10.1016/j.jeconom.2021.03.014}
-}
-
 @article{sun2021estimating,
   title={Estimating dynamic treatment effects in event studies with heterogeneous treatment effects},
   author={Sun, Liyang and Abraham, Sarah},
@@ -529,8 +517,6 @@
   doi={10.18637/jss.v067.i01}
 }
 
-% === Phase 2 auto-verified entries from bulk Crossref verification ===
-
 @article{cameron2008bootstrap,
   title={Bootstrap-Based Improvements for Inference with Clustered Errors},
   author={Cameron, A. Colin and Gelbach, Jonah B. and Miller, Douglas L.},
@@ -921,14 +907,6 @@
   doi={10.32614/cran.package.rdhte}
 }
 
-@article{colincameron2015practitioner,
-  title={A Practitioner’s Guide to Cluster-Robust Inference},
-  author={Colin Cameron, A. and Miller, Douglas L.},
-  journal={Journal of Human Resources},
-  year={2015},
-  doi={10.3368/jhr.50.2.317}
-}
-
 @article{card2015inference,
   title={Inference on Causal Effects in a Generalized Regression Kink Design},
   author={Card, David and Pei, Zhuan and Lee, David and Weber, Andrea},
@@ -1119,14 +1097,6 @@
   journal={Technometrics},
   year={1977},
   doi={10.1080/00401706.1977.10489570}
-}
-
-@article{correia2020fast,
-  title={Fast Poisson estimation with high-dimensional fixed effects},
-  author={Correia, Sergio and Guimarães, Paulo and Zylkin, Tom},
-  journal={The Stata Journal: Promoting communications on statistics and Stata},
-  year={2020},
-  doi={10.1177/1536867x20909691}
 }
 
 @article{cragg1993testing,
@@ -1502,14 +1472,6 @@
   booktitle={Proceedings of the National Academy of Sciences},
   year={2010},
   doi={10.1073/pnas.1009510107}
-}
-
-@article{imai2014covariate,
-  title={Covariate Balancing Propensity Score},
-  author={Imai, Kosuke and Ratkovic, Marc},
-  journal={Journal of the Royal Statistical Society Series B: Statistical Methodology},
-  year={2014},
-  doi={10.1111/rssb.12027}
 }
 
 @article{imbens2008regression,
@@ -1931,14 +1893,6 @@
   doi={10.1214/14-sts485}
 }
 
-@article{robins1986approach,
-  title={A new approach to causal inference in mortality studies with a sustained exposure period—application to control of the healthy worker survivor effect},
-  author={Robins, James},
-  journal={Mathematical Modelling},
-  year={1986},
-  doi={10.1016/0270-0255(86)90088-6}
-}
-
 @article{romano2005exact,
   title={Exact and Approximate Stepdown Methods for Multiple Hypothesis Testing},
   author={Romano, Joseph P and Wolf, Michael},
@@ -2250,10 +2204,6 @@
   doi={10.1080/01621459.2015.1023805}
 }
 
-
-% === Pass 2 additions (relaxed year ±1) ===
[…]
