### 2026-04-23 742ec43
test(citations): 49 tests pin auditor regex + stopword behaviour
Both §10 auditors (tools/audit_citations.py — 790 LOC, and
tools/audit_bib_duplicates.py — 159 LOC) are now the quality command
for the whole bibliography, but they had zero tests. Any future tweak
to the regex or stopword set could silently break citation enforcement
without anyone noticing until wrong attributions leak into a release.

Added two stdlib-only test suites (<1s combined) that pin the
regressions we fixed in commits fa35662 and 8ce7574, plus the core
invariants both tools rely on.

tests/test_audit_citations.py (32 
[…]
