+++ tests/test_audit_citations.py
@@ -1,0 +1,389 @@
+"""Tests for ``tools/audit_citations.py`` — the §10 citation auditor.
+
+These tests never touch the network or the real arXiv / Crossref / NBER
+endpoints. They exercise the pure-Python layer only:
+  * regex extraction (``ARXIV_RE``, ``NBER_RE``, ``DOI_RE``)
+  * name normalisation (``_normalise``, ``_last_name``)
+  * the ``diff_citation`` classifier against a hand-built ``PaperMeta``
+  * ``extract_citations`` over a tmp_path tree
+  * CLI exit code semantics (via ``--kinds`` + a tmp source tree so the
+    test never depends on the real paper.bib / src tree)
+
+The two regressions fixed in commits ``fa35662`` and earlier are pinned
+explicitly so a future regex / stopword tweak doesn't silently break:
+  * DOI regex must NOT swallow a trailing ``}`` from a bibtex
+    ``doi={10.x/y}`` literal.
+  * ``Q-network`` must NOT be flagged as a phantom author surname.
+"""
+
+from __future__ import annotations
+
+import subprocess
+import sys
+from pathlib import Path
+
+import pytest
+
+REPO_ROOT = Path(__file__).resolve().parent.parent
+TOOLS_DIR = REPO_ROOT / "tools"
+sys.path.insert(0, str(TOOLS_DIR))
+
+import audit_citations as ac  # noqa: E402
+
+
+# ---------------------------------------------------------------------------
+# Regex: arXiv
+# ---------------------------------------------------------------------------
+
+
+def test_arxiv_regex_matches_standard_form():
+    m = ac.ARXIV_RE.search("See arXiv:2408.12345 for details.")
+    assert m is not None
+    assert m.group("id") == "2408.12345"
+
+
+def test_arxiv_regex_matches_5digit_suffix():
+    m = ac.ARXIV_RE.search("arXiv:2408.12345 is fine but arXiv:2408.123456 too")
+    assert m is not None
+    assert m.group("id") == "2408.12345"
+
+
+def test_arxiv_regex_strips_version_suffix():
+    m = ac.ARXIV_RE.search("arXiv:2008.12345v3")
+    assert m is not None
+    assert m.group("id") == "2008.12345"  # v3 dropped
+
+
+def test_arxiv_regex_ignores_bare_year_number():
+    # "2008.12345" without the arXiv prefix must NOT match: the tool
+    # only audits explicitly-flagged arXiv references.
+    assert ac.ARXIV_RE.search("version 2008.12345 of the library") is None
+
+
+# ---------------------------------------------------------------------------
+# Regex: NBER
+# ---------------------------------------------------------------------------
+
+
+def test_nber_regex_requires_wp_marker():
+    m = ac.NBER_RE.search("NBER Working Paper 26463")
+    assert m is not None
+    assert m.group("id") == "26463"
+
+
+def test_nber_regex_accepts_w_prefix():
+    m = ac.NBER_RE.search("NBER WP w34550 provides …")
+    assert m is not None
+    assert m.group("id") == "34550"
+
+
+def test_nber_regex_ignores_bare_number():
+    # A standalone 5-digit number is NOT an NBER reference.
+    assert ac.NBER_RE.search("room 26463 at the conference") is None
+
+
+# ---------------------------------------------------------------------------
+# Regex: DOI — REGRESSION for the bibtex closing-brace bug
+# ---------------------------------------------------------------------------
+
+
+def test_doi_regex_excludes_closing_brace():
+    """Regression for commit fa35662: the DOI regex used to greedy-eat
+    the ``}`` that closes a bibtex ``doi={10.x/y}`` field. Real paper
[…]
