+++ tests/test_audit_bib_duplicates.py
@@ -1,0 +1,322 @@
+"""Tests for ``tools/audit_bib_duplicates.py``.
+
+These tests are stdlib-only — they never touch paper.bib itself, never
+hit the network, and never shell out (except for the CLI smoke-test).
+They guard the two integrity invariants the tool enforces:
+  * ``parse_bib`` correctly locates entry keys + DOIs + arXiv ids even
+    when titles carry nested braces.
+  * ``find_duplicates`` counts different-key collisions but not the
+    trivial "key defined N times" case as a DOI duplicate.
+"""
+
+from __future__ import annotations
+
+import subprocess
+import sys
+import textwrap
+from pathlib import Path
+
+import pytest
+
+REPO_ROOT = Path(__file__).resolve().parent.parent
+TOOLS_DIR = REPO_ROOT / "tools"
+sys.path.insert(0, str(TOOLS_DIR))
+
+import audit_bib_duplicates as abd  # noqa: E402
+
+
+# ---------------------------------------------------------------------------
+# parse_bib
+# ---------------------------------------------------------------------------
+
+
+def test_parse_bib_single_entry_extracts_key_doi_arxiv():
+    text = textwrap.dedent("""
+        @article{smith2020example,
+          title={A Study of X},
+          author={Smith, Alice},
+          year={2020},
+          doi={10.1234/abcd.5678}
+        }
+    """).strip()
+    entries = abd.parse_bib(text)
+    assert len(entries) == 1
+    e = entries[0]
+    assert e.key == "smith2020example"
+    assert e.kind == "article"
+    assert e.doi == "10.1234/abcd.5678"
+    assert e.arxiv is None
+
+
+def test_parse_bib_handles_nested_title_braces():
+    """Bibtex titles often protect capitalisation with nested braces.
+
+    The parser must balance braces to find the entry's closing ``}``,
+    not the first one after the opening.
+    """
+    text = textwrap.dedent("""
+        @article{jones2021nested,
+          title={Deep {RL} for {HTE}: Methods and Evidence},
+          author={Jones, Bob},
+          year={2021},
+          doi={10.5555/xxxx.yyyy}
+        }
+
+        @article{doe2022next,
+          title={Simple},
+          author={Doe, Jane},
+          year={2022},
+          doi={10.6666/zzzz}
+        }
+    """).strip()
+    entries = abd.parse_bib(text)
+    assert [e.key for e in entries] == ["jones2021nested", "doe2022next"]
+    assert entries[0].doi == "10.5555/xxxx.yyyy"
+    assert entries[1].doi == "10.6666/zzzz"
+
+
+def test_parse_bib_doi_lowercased_and_trailing_period_stripped():
+    text = textwrap.dedent("""
+        @article{upper2023case,
+          title={Test},
+          doi={10.1177/1536867X20909691.}
+        }
+    """).strip()
+    entry = abd.parse_bib(text)[0]
+    # Lowercased so `10.X/Y` and `10.x/y` dedupe together.
+    assert entry.doi == "10.1177/1536867x20909691"
+
+
+def test_parse_bib_arxiv_from_journal_field():
+    text = textwrap.dedent("""
+        @misc{foo2024preprint,
+          title={Whatever},
+          author={Foo, Bar},
+          year={2024},
+          journal={arXiv:2408.12345}
+        }
+    """).strip()
+    entry = abd.parse_bib(text)[0]
+    assert entry.arxiv == "2408.12345"
+
+
+def test_parse_bib_arxiv_from_eprint_field():
+    text = textwrap.dedent("""
+        @misc{bar2024preprint,
+          title={Whatever},
+          eprint={2501.09876},
+          archivePrefix={arXiv}
+        }
+    """).strip()
[…]
