### 2026-04-23 9cfc3b3
fix(ci): override pyproject addopts so auditor tests don't need pytest-cov
The citation-audit workflow only installs ``pytest`` (no coverage
toolchain) because the auditor tests are stdlib-only. But pytest
auto-applies the project's ``pyproject.toml`` addopts, which include
``--cov=statspai`` — so the CI run died with
``ERROR: usage: ... unrecognized arguments --cov=statspai`` and
never reached the live auditor steps.

Override addopts with ``-o 'addopts='`` to run the tests in isolation.
Local ``pytest`` (with the full [dev] extras) keeps its coverage
reporting unchanged.

Verification: ran t
[…]
