### 2026-04-23 5d3ed21
feat(citations): add bib coverage auditor (gate 2) + fix dangling ref in paper.md
Third and final auditor in the §10 enforcement stack, closing the
``[@bibkey]``-graph invariants the other two couldn't see:

  * audit_citations.py       — are the AUTHORS / YEAR right?
  * audit_bib_duplicates.py  — is paper.bib internally clean?
  * audit_bib_coverage.py ←  — does every citation resolve, and is
                               every bib entry actually cited?

Immediate find
--------------
Running the new auditor on HEAD surfaced one real regression my
earlier consolidation (commit 8ce7574) misse
[…]
