+++ tests/test_audit_bib_coverage.py
@@ -1,0 +1,293 @@
+"""Tests for ``tools/audit_bib_coverage.py``.
+
+Guards the bibkey-graph invariants:
+  * ``[@bibkey]`` pandoc citations are collected from ``.py`` / ``.md`` /
+    ``.rst`` / ``.txt`` files, including the multi-key bracket form
+    ``[@a; @b, @c]``.
+  * ``dangling`` = cited keys missing from paper.bib.
+  * ``orphan``   = paper.bib keys never cited.
+  * CLI exit code semantics: dangling always hard-fails under
+    ``--strict-dangling``; orphan is advisory by default.
+"""
+
+from __future__ import annotations
+
+import subprocess
+import sys
+import textwrap
+from pathlib import Path
+
+import pytest
+
+REPO_ROOT = Path(__file__).resolve().parent.parent
+TOOLS_DIR = REPO_ROOT / "tools"
+sys.path.insert(0, str(TOOLS_DIR))
+
+import audit_bib_coverage as abc  # noqa: E402
+
+
+# ---------------------------------------------------------------------------
+# extract_citations
+# ---------------------------------------------------------------------------
+
+
+def test_extract_citations_simple_bracket(tmp_path, monkeypatch):
+    monkeypatch.setattr(abc, "REPO_ROOT", tmp_path)
+    (tmp_path / "mod.py").write_text(
+        '"""See [@smith2020example] for proof."""\n',
+        encoding="utf-8",
+    )
+    sites = abc.extract_citations([tmp_path])
+    assert [s.key for s in sites] == ["smith2020example"]
+    assert sites[0].line == 1
+
+
+def test_extract_citations_multi_key_bracket(tmp_path, monkeypatch):
+    """Pandoc allows ``[@a; @b]`` and ``[@a, @b]`` for stacked refs.
+    Both forms must yield one CitationSite per key."""
+    monkeypatch.setattr(abc, "REPO_ROOT", tmp_path)
+    (tmp_path / "mod.md").write_text(
+        "Method [@abadie2010synthetic; @abadie2015comparative, @doudchenko2016balancing].\n",
+        encoding="utf-8",
+    )
+    sites = abc.extract_citations([tmp_path])
+    keys = sorted(s.key for s in sites)
+    assert keys == [
+        "abadie2010synthetic",
+        "abadie2015comparative",
+        "doudchenko2016balancing",
+    ]
+
+
+def test_extract_citations_ignores_email(tmp_path, monkeypatch):
+    """``user@domain.com`` must not be mistaken for a citation."""
+    monkeypatch.setattr(abc, "REPO_ROOT", tmp_path)
+    (tmp_path / "mod.py").write_text(
+        "# Contact user@domain.com for questions.\n",
+        encoding="utf-8",
+    )
+    sites = abc.extract_citations([tmp_path])
+    assert sites == []
+
+
+def test_extract_citations_ignores_python_decorator(tmp_path, monkeypatch):
+    """``@staticmethod`` and similar bare decorators must not match.
+    Bracket form is the only one we accept; bare ``@key`` is skipped."""
+    monkeypatch.setattr(abc, "REPO_ROOT", tmp_path)
+    (tmp_path / "mod.py").write_text(
+        "class X:\n    @staticmethod\n    def foo(): pass\n",
+        encoding="utf-8",
+    )
+    sites = abc.extract_citations([tmp_path])
+    assert sites == []
+
+
+def test_extract_citations_honours_extension_filter(tmp_path, monkeypatch):
+    monkeypatch.setattr(abc, "REPO_ROOT", tmp_path)
+    (tmp_path / "data.json").write_text('{"ref": "[@smith2020]"}', encoding="utf-8")
[…]
