+++ tools/audit_bib_coverage.py
@@ -1,0 +1,262 @@
+#!/usr/bin/env python3
+"""
+Bibliography coverage auditor for StatsPAI.
+
+Closes the §10 red line from the opposite direction of the other two
+auditors:
+
+  * ``audit_citations.py``        — every cited paper must match primary
+                                    source (authors / year).
+  * ``audit_bib_duplicates.py``   — paper.bib must have no duplicates.
+  * ``audit_bib_coverage.py`` ←   — the ``[@bibkey]`` graph must be
+                                    consistent: every code citation
+                                    resolves to a bib entry (no
+                                    dangling refs) and optionally every
+                                    bib entry is cited somewhere (no
+                                    orphans).
+
+Dangling refs are a hard build failure for ``paper.md`` (pandoc will
+emit ``???`` instead of the citation), so ``--strict-dangling`` is the
+default CI gate. Orphan entries are a softer code-smell and kept as an
+advisory report by default.
+
+Usage
+-----
+    python tools/audit_bib_coverage.py                 # report only
+    python tools/audit_bib_coverage.py --strict-dangling
+    python tools/audit_bib_coverage.py --strict        # dangling + orphan
+    python tools/audit_bib_coverage.py --bib paper.bib --roots src docs
+    python tools/audit_bib_coverage.py --json report.json
+
+Stdlib only — safe to run in CI without extra dependencies.
+"""
+
+from __future__ import annotations
+
+import argparse
+import json
+import re
+import sys
+from collections import defaultdict
+from dataclasses import dataclass, field
+from pathlib import Path
+from typing import Optional
+
+REPO_ROOT = Path(__file__).resolve().parent.parent
+TOOLS_DIR = Path(__file__).resolve().parent
+
+# Re-use the bib parser from the duplicate auditor so we have a single
+# source of truth for what a bib entry is.
+sys.path.insert(0, str(TOOLS_DIR))
+from audit_bib_duplicates import parse_bib  # noqa: E402
+
+DEFAULT_ROOTS = ("src", "docs", "paper.md")
+DEFAULT_BIB = REPO_ROOT / "paper.bib"
+DEFAULT_EXTENSIONS = {".py", ".md", ".rst", ".txt"}
+
+# Pandoc citation syntax: [@key] or [@key1; @key2]. The bracket form is
+# how both docstrings and paper.md cite papers in this repo. We do NOT
+# match bare ``@key`` — it collides with Python decorators and YAML.
+CITE_RE = re.compile(r"\[@([A-Za-z0-9][A-Za-z0-9_\-]*)(?:[;,]\s*@[A-Za-z0-9][A-Za-z0-9_\-]*)*\]")
+# For capturing all keys inside a multi-key bracket like [@a; @b; @c]:
+MULTI_KEY_RE = re.compile(r"@([A-Za-z0-9][A-Za-z0-9_\-]*)")
+
+
+@dataclass
+class CitationSite:
+    """One ``[@bibkey]`` occurrence in a source file."""
+
+    key: str
+    file: str
+    line: int
+
+
+@dataclass
+class CoverageReport:
+    bib_keys: set[str]
+    cited_keys: set[str]
+    citations: list[CitationSite]
+    # key -> list of (file, line) where it's cited
+    citations_by_key: dict[str, list[tuple[str, int]]] = field(default_factory=dict)
+
+    @property
+    def dangling(self) -> set[str]:
+        """Cited but not defined in paper.bib. HARD bug — breaks pandoc."""
+        return self.cited_keys - self.bib_keys
+
+    @property
+    def orphan(self) -> set[str]:
+        """Defined in paper.bib but never cited. Advisory: may be a
+        deliberate reference placeholder, or stale."""
[…]

[…]
