### 2026-04-23 ea064be
ci: opt into Node 24 early for all 3 workflows (blog-documented env)
GitHub Actions runners deprecate Node 20 on 2026-06-02 and auto-switch
to Node 24. Every run of our workflows currently emits the reminder
annotation on actions/checkout@v4, setup-python@v5, upload-artifact@v4.

Rather than chase six separate major-version bumps (checkout v4->v6,
setup-python v5->v6, upload-artifact v4->v7, cache v4->v5, codecov v4->v6,
download-artifact v3->?) — each a potential breaking change — use
GitHub's documented escape hatch: set FORCE_JAVASCRIPT_ACTIONS_TO_NODE24
as a workflow-level env var. Keeps e
[…]
