### 2026-04-23 89665af
feat(citations): suggest_bibkey_backfills tool for orphan entries
Closes the last §10 feedback loop: paper.bib has 177 orphan entries
(defined but never cited via [@bibkey]). audit_bib_coverage reports
the count, but doesn't tell you which orphans are actually already
referenced in prose — just missing the annotation.

tools/suggest_bibkey_backfills.py cross-indexes each orphan's arXiv id
and DOI against src/ + docs/ + paper.md. If a match is found on a
line that does NOT already carry a [@bibkey], it's a back-fill
candidate.

Dry-run on HEAD reports 21 of 177 orphans as instant back-fill
cand
[…]
