+++ tools/suggest_bibkey_backfills.py
@@ -1,0 +1,332 @@
+#!/usr/bin/env python3
+"""
+Suggest ``[@bibkey]`` back-fills for paper.bib orphan entries.
+
+An orphan is a paper.bib entry that no ``[@bibkey]`` citation in
+src/ / docs/ / paper.md refers to. Many orphans are NOT dead weight —
+they are papers whose **arXiv id or DOI is already quoted in a docstring
+References section**, just without the canonical ``[@bibkey]`` link.
+
+This tool finds those instant back-fill candidates by cross-indexing
+identifiers: for each orphan,
+
+  1. extract its arXiv id (from ``journal={arXiv:...}`` or ``eprint={...}``
+     or bare ``arXiv:...`` text in the bib entry) and its DOI;
+  2. grep src/ + docs/ + paper.md for each identifier;
+  3. for every match, check whether a ``[@key]`` annotation is already
+     present on the SAME line — if so, the line already has a (different
+     or same) bibkey and we skip; otherwise this is a back-fill candidate.
+
+By default the tool only prints the plan. Pass ``--apply`` to actually
+append ``[@orphan_key]`` to each candidate line in place.
+
+Usage
+-----
+    python tools/suggest_bibkey_backfills.py                # dry-run
+    python tools/suggest_bibkey_backfills.py --apply        # mutate files
+    python tools/suggest_bibkey_backfills.py --json out.json
+    python tools/suggest_bibkey_backfills.py --bib custom.bib
+
+Stdlib only — safe to run in any repo snapshot with the other two
+auditors present.
+"""
+
+from __future__ import annotations
+
+import argparse
+import json
+import re
+import subprocess
+import sys
+from dataclasses import dataclass, asdict
+from pathlib import Path
+from typing import Optional
+
+REPO_ROOT = Path(__file__).resolve().parent.parent
+TOOLS_DIR = Path(__file__).resolve().parent
+
+sys.path.insert(0, str(TOOLS_DIR))
+from audit_bib_duplicates import parse_bib  # noqa: E402
+from audit_bib_coverage import build_report  # noqa: E402
+
+DEFAULT_BIB = REPO_ROOT / "paper.bib"
+DEFAULT_ROOTS = ("src", "docs", "paper.md")
+
+# Capture the arXiv id from a bib entry's own body. Accepts the
+# common StatsPAI patterns: ``journal={arXiv:2408.12345}``,
+# ``eprint={2408.12345}``, or bare ``arXiv:2408.12345`` prose.
+ARXIV_IN_BIB = re.compile(
+    r"(?:arXiv[:\s]*|eprint\s*=\s*[{\"]?)(\d{4}\.\d{4,5})",
+    re.IGNORECASE,
+)
+
+# A line already has a bibkey annotation if it contains ``[@xxx]``.
+HAS_BIBKEY = re.compile(r"\[@[A-Za-z0-9][A-Za-z0-9_\-]*\]")
+
+# Lines that look like Python string-literal fragments being
+# concatenated into a bibtex entry (e.g. inside ``_CITATIONS`` dicts).
+# Annotating these would PUSH ``[@bibkey]`` INTO the generated bibtex
+# string, breaking biblatex downstream. We look for two clear signals:
+#
+#   1. the line ends with ``\n"`` or ``\n',`` — the escape-newline +
+#      quote pattern is the canonical fragment-concatenation tail;
+#   2. the line contains ``field={`` where ``field`` is a bibtex entry
+#      key (``journal``, ``eprint``, ``doi``, ``author``, ``title``,
+#      ``year``, ``volume``, ``number``, ``pages``) — that's a bibtex
[…]
