+++ tests/test_suggest_bibkey_backfills.py
@@ -1,0 +1,332 @@
+"""Tests for ``tools/suggest_bibkey_backfills.py``.
+
+Guards the three invariants the tool needs to be safe:
+  * ``find_candidates`` surfaces orphan bib entries whose arXiv/DOI is
+    in prose without a ``[@bibkey]`` annotation.
+  * Lines that already have ``[@someotherkey]`` are skipped (idempotent).
+  * Unsafe contexts — Python string literals, bibtex-synthesis code,
+    ``reference=...`` kwargs — are FILTERED OUT so ``--apply`` cannot
+    corrupt the codebase by injecting ``[@bibkey]`` into a generated
+    bibtex string or causing Python syntax errors.
+  * ``apply_candidates`` actually mutates files when called, but is
+    defensive: it checks the current line still matches the recorded
+    ``original`` before overwriting.
+"""
+
+from __future__ import annotations
+
+import subprocess
+import sys
+import textwrap
+from pathlib import Path
+
+import pytest
+
+REPO_ROOT = Path(__file__).resolve().parent.parent
+TOOLS_DIR = REPO_ROOT / "tools"
+sys.path.insert(0, str(TOOLS_DIR))
+
+import suggest_bibkey_backfills as sbb  # noqa: E402
+
+
+# ---------------------------------------------------------------------------
+# _append_bibkey — placement logic
+# ---------------------------------------------------------------------------
+
+
+def test_append_bibkey_to_plain_line():
+    out = sbb._append_bibkey("    See arXiv:2408.12345 for details.\n", "smith2024")
+    assert out == "    See arXiv:2408.12345 for details. [@smith2024]\n"
+
+
+def test_append_bibkey_to_line_without_trailing_newline():
+    out = sbb._append_bibkey("Citation: arXiv:2408.12345", "smith2024")
+    assert out == "Citation: arXiv:2408.12345 [@smith2024]"
+
+
+def test_append_bibkey_before_docstring_close():
+    # Python _CITATIONS dicts often build bibtex strings line-by-line
+    # with ``"..."`` fragments. We don't annotate these (UNSAFE filter),
+    # but if called directly _append_bibkey must insert BEFORE the
+    # closing quote to at least avoid trivial breakage.
+    src = '    "  arXiv:2408.12345\\n"\n'
+    out = sbb._append_bibkey(src, "smith2024")
+    # The bibkey goes before the \n" tail, preserving the trailing newline.
+    assert "[@smith2024]" in out
+    assert out.endswith('\n')
+
+
+# ---------------------------------------------------------------------------
+# find_candidates — core scan
+# ---------------------------------------------------------------------------
+
+
+def _setup_tmp_repo(tmp_path: Path, monkeypatch) -> Path:
+    """Create an isolated repo-like tree and point the tool at it."""
+    monkeypatch.setattr(sbb, "REPO_ROOT", tmp_path)
+    # build_report + audit_bib_* modules also dereference REPO_ROOT for
+    # relative-path reporting; patch those too.
+    import audit_bib_coverage
+    import audit_bib_duplicates  # noqa
+    monkeypatch.setattr(audit_bib_coverage, "REPO_ROOT", tmp_path)
+    # Initialise a minimal git repo so ``git grep`` works.
+    subprocess.run(["git", "init", "-q"], cwd=tmp_path, check=True)
+    subprocess.run(["git", "config", "user.email", "t@t"], cwd=tmp_path, check=True)
[…]
