### 2026-04-23 9f46d57
docs(robustness): first backfill batch — 3 [@bibkey] annotations in sensitivity_frontier
First real exercise of tools/suggest_bibkey_backfills.py. The suggester
identified 3 orphan bib entries whose arXiv ids already appear in this
file's References sections — the paper was cited, just without the
canonical [@bibkey] link. Appended the bibkey to the end of each
reference line:

  copula_sensitivity         -> [@balgi2025sensitivity]     (arXiv:2508.08752)
  survival_sensitivity       -> [@hu2025nonparametric]      (arXiv:2511.01412)
  calibrate_confounding_strength -> [@baitairian2025calibrati
[…]
