### 2026-04-23 5c603a0
fix(ci): Windows test parity + Bandit MD5 gate
Three CI-green-up fixes; no numerical behavior change.

- causal_text/_common.py: hashlib.md5(..., usedforsecurity=False) so
  Bandit B324 stops flagging the deterministic token-bucket hash as a
  security primitive. Digest bytes are byte-identical to the prior call.
- tools/audit_bib_coverage.py::_rel: emit POSIX paths via
  Path.as_posix() so citations_by_key reports match across Windows and
  POSIX. Fixes test_build_report_records_citation_locations.
- tests/test_suggest_bibkey_backfills.py: merge child env with
  os.environ so PATH survives an
[…]
