### 2026-04-23 4226047
docs(did): second backfill batch — 4 [@bibkey] annotations in did/ module
Second exercise of tools/suggest_bibkey_backfills.py. Cluster-picked
the did/ module's top-of-module docstring first lines — all 4 share
a uniform ``Title (authors YEAR, arXiv XXXX.YYYYY).`` pattern that
the suggester's score-3 filter identified as zero-risk end-of-reference
sites.

  cohort_anchored  -> [@cohort_anchored2025]   (arXiv 2509.01829, Liu 2025)
  design_robust    -> [@design_robust_es2026]  (arXiv 2601.18801, 2026)
  misclassified    -> [@did_misclassified2025] (arXiv 2507.20415, Augustin et al. 2025)
  did_
[…]
