### 2026-04-23 7d65682
release: v1.6.1 — CI/CD green-up
Patch release. No user-facing numerical change. Digest bytes of
hash_embed_texts are byte-identical to v1.6.0.

- hashlib.md5(..., usedforsecurity=False) so Bandit B324 security gate
  stops blocking CI.
- tools/audit_bib_coverage.py::_rel emits POSIX paths via as_posix()
  so citations_by_key report is stable across Windows/POSIX.
- tests/test_suggest_bibkey_backfills.py merges child env with
  os.environ so Windows CreateProcess can resolve git.exe.

Verified:
- TestPyPI install + import + hash_embed_texts smoke: OK
- PyPI prod upload OK (https://pypi.org/pro
[…]
