### 2026-04-23 31f6f46
perf(ci): PR-only slim matrix to cut PR wall-clock ~50%
Dynamic test matrix via set-matrix job:
- pull_request: 1 combo (ubuntu-latest × 3.10, matches Codecov upload combo)
- push / release / workflow_dispatch: full 10-combo matrix unchanged

PR CI was ~44 min wall-clock on the slowest Windows shard. Push-to-main
still runs all 10 shards before any release, so Windows parity and
cross-version regressions stay covered. Fully reversible.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
