### 2026-04-23 4bd1d29
docs(citations): 5 verified backfills + Vashishtha→Jiralerspong bib fix
Verified every arXiv ID against arxiv.org (primary source) before
touching anything. 5 orphan bib entries matched truth and got a single
clean [@bibkey] annotation each. One bib entry had the wrong arXiv
ID attributed to it — surgery below.

Verified backfills (single clean site per key)
----------------------------------------------
  olivas2025proximal    arXiv 2512.12038 → src/proximal/mtp.py:3
  wu2025estimating      arXiv 2509.18484 → src/interference/orthogonal.py:136
  fang2025inward        arXiv 2506.06615 → src/in
[…]
