+++ paper.bib
@@ -4101,9 +4101,9 @@
   year={2023}
 }
 
-@misc{vashishtha2024causal,
-  title={Causal Inference Using LLM-Guided Discovery},
-  author={Vashishtha, Aniket and others},
+@misc{jiralerspong2024efficient,
+  title={Efficient Causal Graph Discovery Using Large Language Models},
+  author={Jiralerspong, Thomas and Chen, Xiaoyin and More, Yash and Shah, Vedant and Bengio, Yoshua},
   year={2024},
   howpublished={arXiv preprint arXiv:2402.01207},
   doi={10.48550/arXiv.2402.01207}
