### 2026-04-23 a498d77
fix(proximal): eliminate 2 fabricated citations (Shi/Yang-Schwartz) — use verified Min/Yu
Per §10 red line "zero fabricated citations", two paper.bib entries
described arXiv papers that DO NOT EXIST at the claimed IDs:

  shi2025bidirectional  arXiv 2507.13965
    bib-claimed: Shi, Miao & Tchetgen Tchetgen — "Regression-Based
                 Bidirectional Proximal Causal Inference"
    arXiv truth: Min, Zhang & Luo — "A regression-based approach
                 for bidirectional proximal causal inference in the
                 presence of unmeasured confounding"

  yang2025fortified     arX
[…]
