+++ paper.bib
@@ -3041,18 +3041,20 @@
   year={2002}
 }
 
-@article{shi2025bidirectional,
-  title={Regression-Based Bidirectional Proximal Causal Inference},
-  author={Shi, Xu and Miao, Wang and Tchetgen Tchetgen, Eric J.},
+@article{min2025regression,
+  title={A regression-based approach for bidirectional proximal causal inference in the presence of unmeasured confounding},
+  author={Min, Jiaqi and Zhang, Xueyue and Luo, Shanshan},
   journal={arXiv preprint arXiv:2507.13965},
-  year={2025}
+  year={2025},
+  doi={10.48550/arXiv.2507.13965}
 }
 
-@article{yang2025fortified,
-  title={Fortified Proximal Causal Inference},
-  author={Yang, Yifan and Schwartz, Sheila},
+@article{yu2025fortified,
+  title={Fortified Proximal Causal Inference with Many Invalid Proxies},
+  author={Yu, Myeonghun and Shi, Xu and Tchetgen Tchetgen, Eric J.},
   journal={arXiv preprint arXiv:2506.13152},
-  year={2025}
+  year={2025},
+  doi={10.48550/arXiv.2506.13152}
 }
 
 @article{olivas2025proximal,
