### 2026-04-23 7d2f625
docs(citations): 6 verified backfills across bcf/causal_rl/did/synth/transport/conformal
Final sweep of remaining backfillable orphans. All 6 arXiv IDs
independently verified against arxiv.org primary source — author
lists + titles in paper.bib match the paper abstract pages exactly.
No fabrications in this batch.

Verified backfills (single clean site per key)
----------------------------------------------
  cunha2025unifying       arXiv 2512.18135  src/causal_rl/benchmarks.py:2
  han2025synthetic        arXiv 2511.14133  src/synth/survival.py:1
  peng2024local           arXiv 2404.04794  src
[…]
