### 2026-04-23 b416367
fix(tests): drop --strict from CLI smoke test to survive arXiv 429 rate limit
CI caught a real test flake introduced in commit 742ec43: the
``test_cli_on_empty_tree_exits_zero`` smoke test runs the citation
auditor as a subprocess inside a tmp_path. The auditor resolves roots
relative to its compiled ``REPO_ROOT`` (not the cwd argv), so in CI
it actually scans the real StatsPAI src/docs and fires ~77 live
arXiv queries from a cold cache. GitHub's runner shares an egress
IP with many other jobs and arxiv.org throttled it with HTTP 429,
producing "summary: 32 ok / 0 mismatch / 284 unresolved" — 
[…]
