+++ tests/test_audit_citations.py
@@ -355,8 +355,16 @@ def test_extract_citations_records_line_number(tmp_repo):
 # ---------------------------------------------------------------------------
 
 
-def test_cli_on_empty_tree_exits_zero(tmp_path):
-    """A tree with no citations must exit 0 even under --strict."""
+def test_cli_runs_without_crash_on_empty_tree(tmp_path):
+    """Smoke test: the CLI must not traceback or segfault on a trivial
+    invocation. We do NOT pass ``--strict`` here because the auditor
+    resolves roots relative to its own ``REPO_ROOT`` constant (not
+    ``cwd``), so in CI it actually scans the real src/ and makes live
+    arXiv calls — which can hit HTTP 429 rate limits on a cold runner
+    and would flip ``--strict`` into exit 1. Non-strict mode keeps
+    exit 0 whenever the script didn't crash, which is what this test
+    is really checking.
+    """
     (tmp_path / "src").mkdir()
     (tmp_path / "docs").mkdir()
     (tmp_path / "src" / "mod.py").write_text("# no citations here\n",
@@ -364,16 +372,10 @@ def test_cli_on_empty_tree_exits_zero(tmp_path):
     result = subprocess.run(
         [sys.executable, str(TOOLS_DIR / "audit_citations.py"),
          "--roots", "src", "docs",
-         "--out", str(tmp_path / "report.md"),
-         "--strict"],
+         "--out", str(tmp_path / "report.md")],
         capture_output=True, text=True, check=False,
         cwd=tmp_path,
     )
-    # The auditor resolves roots relative to its own REPO_ROOT constant,
-    # so 'src'/'docs' under tmp_path won't be found; check that the
-    # exit-code path still behaves.  The real repository CI already
-    # exercises the happy path end-to-end, so here we only assert the
-    # script executes without crashing.
     assert result.returncode in (0, 2), (
         f"unexpected exit {result.returncode}: {result.stderr}"
     )
