### 2026-04-23 b127aa7
feat(skill): add StatsPAI_skill Claude Code skill + bilingual README
Agent-native end-to-end empirical-analysis skill for Claude Code:
SKILL.md (Step 0–6 workflow + Method Catalog, 49 sp.* symbols
verified against current registry of 925 functions) + EN/中文
README with install and scope. Ported from upstream
Awesome-Agent-Skills-for-Empirical-Research with name aligned to
folder, stale "390+" bumped to "900+", and two natural-language
triggers for "complete empirical analysis" flows.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
