### 2026-04-24 1d71986
docs(skill): rewrite StatsPAI_skill — data contract, power, pre-reg, sequential robustness
Replace the v1 skill with a verified end-to-end empirical pipeline (Step −1 → 7)
that keeps sp.* as the single entry point:

- Step −1: sp.power(design, ...) design & sample-size planning
- Step  0: data_contract() — 5 checks + Rubin MCAR hint (go/no-go gate)
- Step  3: IdentificationPlan framed as pre-registration artifact (frozen to disk)
- Step  6: sequential confounder blocks via sp.regtable (Oster-style stability)
- Step  7: deliverables (to_latex / to_word / cite + reproducibility stamp)
- Common M
[…]
