+++ StatsPAI_skill/SKILL.md
@@ -1,369 +1,432 @@
 ---
 name: StatsPAI_skill
-description: Agent-native one-stop toolkit for the full empirical data-analysis pipeline in Python (v1.6+). 900+ functions, one import (`import statspai as sp`), unified API. Covers the complete loop after data cleaning — descriptive stats & EDA (sp.sumstats, sp.balance_table, sp.balance_panel), estimand-first research-question DSL (sp.causal_question), LLM-assisted DAG discovery (sp.llm_dag_propose/validate/constrained), one-call orchestration (sp.causal), classical estimators (OLS, IV, DID, staggered DID, RDD, PSM, SCM), ML causal (DML, Causal Forest, Meta-Learners, TMLE), neural causal, text causal (sp.causal_text), and diagnostics + robustness (sp.diagnose, sp.spec_curve, sp.honest_did). Use when the user asks to run a full empirical analysis, decide which estimator to use ("DID vs RD vs IV?"), explore models via DAG, estimate treatment effects, evaluate policy, run observational studies, or apply any of the listed econometric methods in Python. Every function returns structured result objects with self-describing schemas for LLM-driven workflows. Data cleaning (missing values, type coercion, merges) is *not* covered — handle that with pandas first, then enter StatsPAI.
+description: Use when the user asks to run a full empirical / causal analysis in Python, pick between DID / RD / IV / SCM / DML / matching, estimate treatment effects, build a DAG, evaluate policy, or produce regression / robustness tables for a paper. Also triggers on keywords "StatsPAI", "statspai", "causal_question", "estimand-first", "honest_did", "spec_curve", "callaway_santanna", "dragonnet", or "text as treatment".
 triggers:
-  - implement causal inference
-  - run a DID analysis
-  - instrumental variables regression
-  - regression discontinuity
-  - propensity score matching
+  - causal inference in python
+  - DID IV RD SCM
+  - callaway_santanna
   - synthetic control
   - double machine learning
   - causal forest
-  - panel data regression
-  - econometric analysis
-  - treatment effect estimation
-  - causal analysis
-  - policy evaluation
-  - observational study
+  - honest_did
+  - spec_curve
   - estimand-first DSL
-  - causal question to estimator
   - LLM-assisted DAG discovery
-  - complete empirical pipeline
-  - data analysis workflow
-  - descriptive statistics and balance table
-  - panel balance check
-  - pre-flight diagnostics
   - text as treatment
-  - automate a complete empirical analysis
-  - end-to-end empirical workflow in Python
   - StatsPAI
   - statspai
 ---
 
 # StatsPAI: Agent-Native Causal Inference & Econometrics
 
-StatsPAI is the agent-native Python package for causal inference and applied econometrics. One `import statspai as sp`, 900+ functions, covering the complete empirical research workflow.
+StatsPAI is the agent-native Python package for causal inference and applied econometrics: one `import statspai as sp`, 900+ functions behind a self-describing API, and `CausalResult` objects that export to LaTeX / Word / Excel / BibTeX.
 
-**Source**: https://github.com/brycewang-stanford/StatsPAI  
-**PyPI**: `pip install statspai`  
-**Paper**: Submitted to Journal of Open Source Software (JOSS) — under review
+- **Source**: https://github.com/brycewang-stanford/StatsPAI
+- **Install**: `pip install statspai` (>= 1.6)
+- **Paper**: submitted to JOSS (under review)
 
-## Why StatsPAI for Agents?
+## Why for Agents
 
[…]
