### 2026-04-24 74f8582
fix(synth): rho2024cluster was fabricated — real paper is Rho et al. 2025 ClusterSC
Systematic orphan-audit sprint (156 entries against primary sources)
found one confirmed bib metadata fabrication.

Investigation
-------------
paper.bib claimed:
  @article{rho2024cluster,
    title={Cluster Synthetic Control Methods},
    author={Rho, Saeyoung},
    journal={arXiv preprint},    # <- NO arXiv id, NO DOI
    year={2024}
  }

Evidence that this paper doesn't exist as described:
  1. No arXiv ID in the bib entry (red flag — every other synth/ arxiv
     reference has one).
  2. Semantic Scholar s
[…]
