+++ paper.bib
@@ -4048,11 +4048,12 @@
   year={2025}
 }
 
-@article{rho2024cluster,
-  title={Cluster Synthetic Control Methods},
-  author={Rho, Saeyoung},
-  journal={arXiv preprint},
-  year={2024}
+@article{rho2025clustersc,
+  title={{ClusterSC}: Advancing Synthetic Control with Donor Selection},
+  author={Rho, Saeyoung and Tang, Andrew and Bergam, Noah and Cummings, Rachel and Misra, Vishal},
+  journal={arXiv preprint arXiv:2503.21629},
+  year={2025},
+  doi={10.48550/arXiv.2503.21629}
 }
 
 @article{richardson2013single,
