### 2026-04-24 fd63ef6
docs(bib): add verified arXiv IDs to santanna2020doubly + runge2020discovering
Orphan-audit follow-up: 2 high-risk bib entries verified against
arxiv.org primary source, enriched with their arXiv identifier so the
audit_citations.py pipeline can independently check them on future runs.

Both were in the Phase 3 "rate-limited high-risk pending" bucket; I
worked around API throttling by WebFetch'ing the arxiv abstract page
directly.

Verified:
  santanna2020doubly  Sant'Anna & Zhao (2020) "Doubly Robust
                      Difference-in-Differences Estimators" JoE 219(1):
                     
[…]
