+++ paper.bib
@@ -3014,7 +3014,9 @@
   number={1},
   pages={101--122},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  eprint={1812.01723},
+  archivePrefix={arXiv}
 }
 
 @article{cohort_anchored2025,
@@ -4076,7 +4078,10 @@
   title={Discovering contemporaneous and lagged causal relations in autocorrelated nonlinear time series datasets},
   author={Runge, Jakob},
   journal={Conference on Uncertainty in Artificial Intelligence (UAI 2020)},
-  year={2020}
+  year={2020},
+  eprint={2003.03685},
+  archivePrefix={arXiv},
+  doi={10.48550/arXiv.2003.03685}
 }
 
 @article{schwab2014ltmle,
