### 2026-04-24 9e45135
fix(bib): 4 fabrication surgeries from systematic orphan audit
Phase 3 high-risk verification (API cooldown retry) surfaced four more
paper.bib entries that violated §10 "zero fabricated citations":

1. borusyak2025harvesting
   -------------------------
   bib-claimed:  Borusyak, Kirill and others (2025)
                 "Harvesting Differences-in-Differences and Event-Study
                 Designs" — NBER Working Paper 34550
   verified:     Abadie, Angrist, Frandsen & Pischke (2025)
                 "Harvesting Differences-in-Differences and Event-Study
                 **Evidence**" — NBE
[…]
