+++ paper.bib
@@ -4029,11 +4029,14 @@
   note={reporting guideline consensus statement}
 }
 
-@article{padh2024metalearners,
+@article{schweisthal2024metalearners,
   title={Meta-Learners for Partially-Identified Treatment Effects Across Multiple Environments},
-  author={Padh, Kirtan and Oprescu, Miruna and van der Schaar, Mihaela},
-  journal={arXiv preprint},
-  year={2024}
+  author={Schweisthal, Jonas and Frauen, Dennis and van der Schaar, Mihaela and Feuerriegel, Stefan},
+  journal={arXiv preprint arXiv:2406.02464},
+  year={2024},
+  eprint={2406.02464},
+  archivePrefix={arXiv},
+  doi={10.48550/arXiv.2406.02464}
 }
 
 @article{phillips2018jackknife,
@@ -4041,13 +4044,6 @@
   author={Phillips, Garry D. A. and Hale, C.},
   journal={Cardiff Economics Working Papers},
   year={2018}
-}
-
-@article{qian2025hal,
-  title={{HAL-TMLE}: Two Variants for Semiparametric Efficient Estimation},
-  author={Qian, Yi and van der Laan, Mark J.},
-  journal={arXiv preprint},
-  year={2025}
 }
 
 @article{rho2025clustersc,
@@ -4117,13 +4113,6 @@
   doi={10.48550/arXiv.2402.01207}
 }
 
-@misc{zorzetto2026bayesian,
-  title={Bayesian Causal Forest for Ordered Treatments},
-  author={Zorzetto, Dafne L. and Molina, Jos{\'e} and Bargagli-Stoffi, Falco J. and Ortiz, Javier J. and Dominici, Francesca},
-  year={2026},
-  howpublished={arXiv preprint},
-}
-
 @misc{goodmanbacon2019bacondecomp,
   title={{BACONDECOMP}: {Stata} Module to Perform the Goodman-Bacon Decomposition},
   author={Goodman-Bacon, Andrew and Goldring, Thomas and Nichols, Austin},
@@ -4179,11 +4168,13 @@
   year={2022}
 }
 
-@misc{borusyak2025harvesting,
-  title={Harvesting Differences-in-Differences and Event-Study Designs},
-  author={Borusyak, Kirill and others},
-  year={2025},
-  howpublished={NBER Working Paper 34550 / MIT}
+@techreport{abadie2025harvesting,
+  title={Harvesting Differences-in-Differences and Event-Study Evidence},
+  author={Abadie, Alberto and Angrist, Joshua and Frandsen, Brigham and Pischke, J{\"o}rn-Steffen},
+  institution={National Bureau of Economic Research},
+  type={NBER Working Paper},
+  number={34550},
+  year={2025}
 }
 
 @article{correia2020fast,
