### 2026-04-24 89607c7
docs(readme): make Quick Example runnable with bundled datasets
The `## Quick Example` block used an undefined `df`, so paste-running it
hit `NameError: name 'df' is not defined` on the very first line.
Rewrote both the English and Chinese versions to draw from
`sp.datasets` (card_1995, mpdta, lee_2008_senate, nsw_dw) — verified
end-to-end that every call (regress, ivreg, did, rdrobust, dml,
causal_forest, margins, test, oster_bounds, modelsummary, outreg2,
sumstats, spec_curve, robustness_report, subgroup_analysis, estat,
recommend) runs cleanly on the bundled data with no wrangling.
