### 2026-04-24 fd5a9fe
feat(registry): v1.6.2 — register continuous_did + did_multiplegt
Pure-additive release. Closes a registry-coverage gap for two already-
shipping DiD estimators (`sp.continuous_did`, `sp.did_multiplegt`) that
were callable but invisible to `sp.list_functions()` / `sp.describe_function()`
/ agent discovery per CLAUDE.md §4.

Adds supporting RFC docs under `docs/rfc/` so the new registry entries'
`reference` / `remedy` pointers resolve (CGS 2024 continuous-DiD,
dCDH 2024 intertemporal event-study, DiD v1.7 gap audit).

No numerical behaviour changes. Registered count 923 → 925. Targeted
tests (r
[…]
