+++ src/statspai/registry.py
@@ -5372,6 +5372,215 @@ def _build_registry():
         typical_n_min=200,
     ))
 
+    # ------------------------------------------------------------------
+    # DiD frontier: continuous treatment + on/off switching
+    # ------------------------------------------------------------------
+    register(FunctionSpec(
+        name="continuous_did",
+        category="causal",
+        description=(
+            "DiD with continuous treatment intensity. Currently exposes three "
+            "modes: (i) TWFE with dose×post interaction, (ii) dose-quantile "
+            "group-time ATT versus the untreated (dose=0) arm with bootstrap "
+            "SE, and (iii) dose-response via local-linear regression of "
+            "ΔY=Y_post−Y_pre on baseline dose. A full Callaway-Goodman-Bacon-"
+            "Sant'Anna (2024) ATT(d|g,t) / ACRT implementation with their "
+            "influence-function inference is on the roadmap — see "
+            "docs/rfc/continuous_did_cgs.md."
+        ),
+        params=[
+            ParamSpec("data", "DataFrame", True),
+            ParamSpec("y", "str", True, description="Outcome variable"),
+            ParamSpec("dose", "str", True,
+                      description="Continuous treatment / dose variable"),
+            ParamSpec("time", "str", True, description="Time period column"),
+            ParamSpec("id", "str", True, description="Unit identifier"),
+            ParamSpec("post", "str", False, None,
+                      "Binary post-treatment indicator "
+                      "(inferred from t_pre / t_post if omitted)"),
+            ParamSpec("t_pre", "int", False, None,
+                      "Last pre-treatment period"),
+            ParamSpec("t_post", "int", False, None,
+                      "First post-treatment period"),
+            ParamSpec("method", "str", False, "att_gt",
+                      "Estimation mode",
+                      ["att_gt", "twfe", "dose_response"]),
+            ParamSpec("n_quantiles", "int", False, 5,
+                      "Number of dose quantiles for discretisation"),
+            ParamSpec("controls", "list", False, None, "Control variables"),
+            ParamSpec("cluster", "str", False, None,
+                      "Cluster variable for SE (TWFE mode)"),
+            ParamSpec("n_boot", "int", False, 500,
+                      "Bootstrap replications for SE"),
+            ParamSpec("alpha", "float", False, 0.05),
+            ParamSpec("seed", "int", False, None),
+        ],
+        returns="CausalResult",
+        example=(
+            'sp.continuous_did(df, y="wage", dose="training_hours", '
+            'time="year", id="worker_id", t_pre=2019, t_post=2020)'
+        ),
+        tags=["did", "continuous_treatment", "dose_response", "causal",
+              "acrt", "frontier"],
+        reference=(
+            "Callaway, Goodman-Bacon & Sant'Anna (2024) "
+            "[@callaway2024difference]; de Chaisemartin & D'Haultfœuille "
+            "(2018) [@dechaisemartin2018fuzzy]."
+        ),
+        pre_conditions=[
+            "panel data with unit × time × outcome × continuous dose",
+            "at least one unit with dose == 0 acts as untreated control "
+            "(or the lowest dose quantile is used as control)",
+            "both a pre and a post period per unit",
+        ],
+        assumptions=[
+            "Parallel trends in potential outcomes across dose levels",
+            "No anticipation of treatment",
+            "Strong parallel trends (CGS 2024) required for ATT(d|g,t) "
+            "interpretation in att_gt mode",
+            "Overlap: positive density of dose in the treated support",
+        ],
+        failure_modes=[
+            FailureMode(
+                symptom="No units with dose == 0",
+                exception="DataInsufficient",
+                remedy=("Lowest-dose quantile is auto-used as the control arm; "
+                        "pass an explicit never-treated indicator via post= if "
+                        "this is not intended."),
[…]
