+++ docs/rfc/multiplegt_dyn.md
@@ -1,0 +1,159 @@
+# RFC — `sp.did_multiplegt_dyn`: de Chaisemartin-D'Haultfœuille (2024) intertemporal event-study DiD
+
+> **Status**: draft 2026-04-23. No code written yet. All paper-specific formulas below carry `[待核验]` until two-source verification per CLAUDE.md §10. Verified anchors: `paper.bib` entry `dechaisemartin2024difference` (DOI 10.1162/rest_a_01414, *ReStat* 2024). ArXiv identifier for working-paper version left unstated until confirmed.
+
+## 1. 动机
+
+当前 `sp.did_multiplegt` 实现的是 **dCDH 2020 (AER) 的 DID_M 估计器**：consecutive-period switcher-vs-stayer 配对 DID 的样本加权平均。它做得已经不错——2024 joint placebo + avg cumulative overlay 已经落地。
+
+但 Stata/R 社区里"did_multiplegt"这个名字的 muscle memory，今天指向的是 **dCDH 2024 ReStat**（arXiv:2007.04267）里的 `_dyn` 包 —— 一个 **intertemporal event-study** 估计器。它：
+
+- 不是 pair 级别的 rollup，而是 first-switch 时段 `F` 的 long-difference 事件研究：对每个 horizon `l`，比较 `Y_{F+l} − Y_{F−1}` 在 switchers vs "not-yet-treated-at-horizon-l" 之间。
+- 有**独立的影响函数方差**（非 bootstrap-only）。
+- 处理 treatment reversal 的方式是 control group 在 horizon `l` 上仍保持未处理（稳定限制），不是简单的 sign-flip。
+
+当前 `sp.did_multiplegt` 的 `dynamic=H` 参数虽然返回了 horizon-l 效应，但其内部路径是 pair rollup 的延展——**不等价于** dCDH 2024 的 `_dyn` 估计量。这是 `gap_audit.md` 标记的另一个头等抓手。
+
+## 2. 目标估计量 `[待核验]`
+
+> dCDH 2024 *ReStat* §2—§3（DOI 10.1162/rest_a_01414，已在 `paper.bib` 核验），以及 R `DIDmultiplegtDYN` 包 reference manual。arXiv working-paper 版本号 [待核验]。所有公式核验前不得落到 docstring。
+
+对每个 horizon `l ≥ 0`：
+
+- **Switchers at F**: units first switching treatment at time `F`.
+- **Not-yet-treated at F+l**: units whose treatment stays at its pre-F value from `F−1` through `F+l` inclusive.
+  [待核验 — dCDH 2024 Definition 2.x]
+
+**Dynamic effect at horizon `l`**:
+[待核验 — dCDH 2024 eq. (2.x)]
+`δ_l = E[ Y_{F+l} − Y_{F−1} | switch ] − E[ Y_{F+l} − Y_{F−1} | not-yet-treated-at-F+l ]`,
+averaged over `F` with appropriate weights [待核验 — 2024 eq. (2.y)].
+
+**Placebo effect at lag `l`**:
+[待核验 — dCDH 2024 eq. (3.x)]
+same structure but comparing `Y_{F−1−l} − Y_{F−1−l−1}` pre-treatment.
+
+**Sign convention for off-switchers**: [待核验]. The current `sp.did_multiplegt` uses a sign flip; `_dyn` handles directionality via the stable-control restriction instead.
+
+### 2.1 识别假设 `[待核验]`
+
+1. **Parallel trends** between switchers at `F` and the not-yet-treated control at horizon `l`, for each `l`.
+2. **No anticipation** prior to `F`.
[…]
