### 2026-04-24 1a434ce
feat(did): v1.6.3 DiD frontier sprint — 5 new estimators + shared primitives
Adds five new estimators (all strictly additive, no numerical changes to
existing code) and the shared-primitives module that future DiD work
will build on. Paper-specific identification details carry [待核验]
markers per CLAUDE.md §10 until reference-parity anchoring is in place.

New estimators
--------------
- sp.lp_did — Local-Projections DiD (Dube-Girardi-Jordà-Taylor 2023)
- sp.ddd_heterogeneous — heterogeneity-robust DDD (Olden-Møen 2022)
- sp.did_timevarying_covariates — Caetano et al. (2022) baseline-frozen
- sp
[…]
