+++ tests/test_did_core_primitives.py
@@ -1,0 +1,196 @@
+"""Unit tests for statspai.did._core — shared DiD primitives."""
+
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+
+from statspai.did import _core as dc
+
+# ----------------------------------------------------------------------
+# cluster_bootstrap_draw
+# ----------------------------------------------------------------------
+
+
+class TestClusterBootstrapDraw:
+    def _panel(self, n_clust=8, n_periods=4, seed=0):
+        rng = np.random.default_rng(seed)
+        rows = []
+        for g in range(n_clust):
+            for t in range(n_periods):
+                rows.append({"g": g, "t": t, "y": rng.normal()})
+        return pd.DataFrame(rows)
+
+    def test_preserves_row_count(self):
+        df = self._panel()
+        rng = np.random.default_rng(42)
+        draw = dc.cluster_bootstrap_draw(df, cluster_col="g", rng=rng)
+        assert len(draw) == len(df)
+
+    def test_relabel_prevents_collisions(self):
+        """When a cluster is drawn twice, its two copies must have distinct
+        relabelled values, so downstream groupby doesn't collapse them."""
+        df = self._panel(n_clust=4)
+        rng = np.random.default_rng(0)
+        # Force collisions by sampling many times
+        for _ in range(10):
+            draw = dc.cluster_bootstrap_draw(df, cluster_col="g", rng=rng)
+            # Each chunk has n_periods rows with identical relabelled id
+            per_id = draw.groupby("g").size()
+            # All chunk sizes should equal the periods-per-cluster constant
+            assert (per_id == 4).all(), "Bootstrap relabel merged independent draws"
+
+    def test_raises_on_missing_cluster_col(self):
+        df = self._panel()
+        rng = np.random.default_rng(0)
+        with pytest.raises(ValueError, match="cluster_col"):
+            dc.cluster_bootstrap_draw(df, cluster_col="nonexistent", rng=rng)
+
+
+# ----------------------------------------------------------------------
+# event_study_frame
+# ----------------------------------------------------------------------
+
+
+class TestEventStudyFrame:
+    def test_empty_input_returns_empty_frame(self):
+        df = dc.event_study_frame([])
+        assert list(df.columns) == list(dc.EVENT_STUDY_COLUMNS)
+        assert len(df) == 0
+
+    def test_fills_missing_optional_keys(self):
+        rows = [
+            {"relative_time": -1, "att": 0.1, "se": 0.05},
+            {"relative_time": 0, "att": 0.3, "se": 0.06, "pvalue": 0.01},
+        ]
+        df = dc.event_study_frame(rows)
+        assert set(dc.EVENT_STUDY_COLUMNS) <= set(df.columns)
+        assert df["type"].tolist() == ["", ""]
+        # First row lacks pvalue → should be NaN
+        assert pd.isna(df.iloc[0]["pvalue"])
+        assert df.iloc[1]["pvalue"] == 0.01
+
+    def test_preserves_extra_columns(self):
+        rows = [{"relative_time": 0, "att": 1.0, "se": 0.1, "n_switchers": 50}]
+        df = dc.event_study_frame(rows)
+        assert "n_switchers" in df.columns
+
+
+# ----------------------------------------------------------------------
+# influence_function_se
+# ----------------------------------------------------------------------
+
+
+class TestInfluenceFunctionSE:
+    def test_scalar_case_matches_manual_var(self):
+        rng = np.random.default_rng(1)
+        ifv = rng.normal(size=500)
[…]
