+++ src/statspai/did/did_multiplegt_dyn.py
@@ -1,0 +1,471 @@
+"""de Chaisemartin & D'Haultfœuille (2024) intertemporal event-study DiD.
+
+Differs from ``sp.did_multiplegt`` (dCDH 2020 DID_M): the 2020 estimator
+is a consecutive-period switcher-vs-stayer pair rollup, while this 2024
+estimator is a **long-difference event study** — at each horizon l ≥ 0,
+compares ``Y_{F+l} − Y_{F−1}`` between units first switching at F and a
+"not-yet-treated at F+l" control group held stable across the horizon.
+
+Verified anchor
+---------------
+- de Chaisemartin & D'Haultfœuille (2024) "Difference-in-Differences
+  Estimators of Intertemporal Treatment Effects", DOI
+  ``10.1162/rest_a_01414`` (bib key ``dechaisemartin2024difference``).
+
+[待核验] identification details
+-------------------------------
+The following are based on my best-effort reading of the paper and the
+companion R ``DIDmultiplegtDYN`` package logic; every item flagged here
+must be verified against the paper's published equations before moving
+this estimator from RFC/experimental status to production / paper-
+faithful claim:
+
+1. **Switcher definition**: units first switching from d=0 to d=1 at
+   period F. First-cut skips switch-off events; the paper handles both
+   directions via a sign convention that I have NOT implemented here.
+   [待核验 — paper §2.x]
+
+2. **Control group per horizon l**: "not-yet-treated at F+l" = units
+   whose d stays at its pre-F value through F+l inclusive. MVP uses
+   this definition; the paper also supports never-treated-only controls
+   which I expose as ``control='never_treated'``. [待核验 — exact window
+   definition]
+
+3. **Per-horizon estimate**:
+
+       δ_l = Σ_F w_F × {E[Y_{F+l} − Y_{F−1} | switchers at F]
+                        − E[Y_{F+l} − Y_{F−1} | not-yet-treated at F+l]}
+
+   with weights ``w_F`` proportional to the number of switchers at F.
+   [待核验 — paper may use a different weighting scheme; heteroskedastic
+   weights variant (dCDH 2023 EJ survey) is not implemented here.]
+
+4. **Placebo lag l < 0**: same structure but the comparison period
+   moves to ``Y_{F-1-|l|} − Y_{F-1-|l|-1}``. [待核验]
+
+5. **Inference**: analytical influence-function variance per horizon is
+   not implemented in this MVP — SE comes from cluster bootstrap on
+   the panel unit. The paper's IF variance is [待核验] and is the clear
+   next step.
+
+Scope for this first cut
+------------------------
+- Never-treated and not-yet-treated control variants.
+- Placebo + dynamic horizons with cluster bootstrap SE.
+- Joint Wald tests for placebo and overall (placebo + dynamic) via the
+  ``_core.joint_wald`` helper on the bootstrap covariance.
+- NO switch-off events, NO heteroskedastic weights, NO analytical IF.
+
+Users who need paper-faithful numerics should wait for the next
+iteration when the paper's equations are in-hand and reference parity
+vs. R ``DIDmultiplegtDYN`` is in place. In the interim, the function
+raises its method label so no user can mistake this for a paper-
+faithful implementation.
+"""
+
+from __future__ import annotations
+
+from typing import Any, Dict, List, Optional
+
+import numpy as np
+import pandas as pd
+from scipy import stats
+
+from ..core.results import CausalResult
+from . import _core as _dc
+
+
+def did_multiplegt_dyn(
+    data: pd.DataFrame,
+    y: str,
+    *,
+    group: str,
+    time: str,
+    treatment: str,
+    placebo: int = 0,
+    dynamic: int = 3,
[…]
