### 2026-04-24 a90875f
release: v1.6.3 — docs, CHANGELOG, MIGRATION, paper.bib upgrade
Release documentation for the 1.6.3 DiD frontier sprint (code commit
precedes). Strictly additive release — no migration required for
existing user code aside from two label-level docstring truth-up
entries called out in MIGRATION.md v1.6.2 → v1.6.3.

Docs
----
- CHANGELOG.md: full 1.6.3 entry covering 5 new estimators, shared
  _core.py, registry enrichment, docstring truth-up, test coverage
- MIGRATION.md: v1.6.2 → v1.6.3 section documents the two user-
  visible label changes (continuous_did att_gt method / estimand
  strings) 
[…]
