+++ paper.bib
@@ -980,11 +980,15 @@
 }
 
 @article{dechaisemartin2022fixed,
-  title={Two-Way Fixed Effects and Differences-in-Differences with Heterogeneous Treatment Effects: A Survey},
-  author={de Chaisemartin, Clément and d'Haultfoeuille, Xavier},
-  journal={SSRN Electronic Journal},
-  year={2022},
-  doi={10.2139/ssrn.4034184}
+  title={Two-way fixed effects and differences-in-differences with heterogeneous treatment effects: a survey},
+  author={de Chaisemartin, Cl{\'e}ment and D'Haultf{\oe}uille, Xavier},
+  journal={The Econometrics Journal},
+  volume={26},
+  number={3},
+  pages={C1--C30},
+  year={2023},
+  publisher={Oxford University Press},
+  doi={10.1093/ectj/utac017}
 }
 
 @article{dechaisemartin2024difference,
