+++ docs/rfc/HANDOFF_2026-04-24.md
@@ -1,0 +1,150 @@
+# Handoff — DiD frontier sprint (A + D + C + B) — 2026-04-24
+
+> 你说完"ABCD 全做，带 `[待核验]` 标记"之后，我一口气把四档全部落完了。下面是验收清单。全部改动**仅在本地 `main` 上，未 push**。测试最终**329 passed, 1 skipped**；没有任何已存在的测试回归。
+
+## 0. TL;DR
+
+| 档位 | 产出 | 测试 |
+|---|---|---|
+| **A — 安全修补** | 5 件事全做：docstring 纠偏、测试补洞、参考对齐 skeleton、guide 补 frontier section、`paper.bib` `dechaisemartin2022fixed` Crossref 核验并升级到发表版 | +11 + 3 (skip 1) |
+| **D — 基础设施** | 18 个已有 DiD 函数补 rich registry spec；新建 `did/_core.py` + 16 个单元测试 | +16 |
+| **C — 已识别的 DiD 前沿** | 新增三个函数：`sp.lp_did` / `sp.ddd_heterogeneous` / `sp.did_timevarying_covariates` | +9 + 7 + 6 |
+| **B — 最难的两块** | `sp.did_multiplegt_dyn` (dCDH 2024 **MVP**) + `sp.continuous_did(method='cgs')` (CGS 2024 **MVP**)，都带 `[待核验]` 和 MVP warning | +10 + 8 |
+| **发布** | 版本 bump 到 **1.6.3**，CHANGELOG + MIGRATION 完整更新 | — |
+
+**所有涉及论文公式的地方一律 `[待核验]`**；我没让任何一个新估计器以"paper-faithful"的名义进入生产。新估计器的 method 标签里都有显式的 MVP / `[待核验]` 标识，方便用户在论文里不会误引。
+
+## 1. 文件变更
+
+```
+修改 (M):
+  M  CHANGELOG.md                                +110 LOC (1.6.3 entry)
+  M  MIGRATION.md                                +43  LOC (v1.6.2 → v1.6.3 section)
+  M  docs/guides/choosing_did_estimator.md       +17  LOC (§4.5 Frontier section)
+  M  docs/rfc/multiplegt_dyn.md                  小修，核验后升级 SSRN → EJ
+  M  paper.bib                                   dechaisemartin2022fixed 升级
+  M  pyproject.toml                              1.6.2 → 1.6.3
+  M  src/statspai/__init__.py                    版本 + 5 new exports
+  M  src/statspai/did/continuous_did.py          docstring 纠偏 + method='cgs' MVP
+  M  src/statspai/did/did_multiplegt.py          docstring 纠偏（_dyn 非等价说明）
+  M  src/statspai/registry.py                    +1300 LOC 登记 (18 旧 + 5 新)
+
+新建 (??):
+  ?? src/statspai/did/_core.py                   共享基元（~270 LOC）
+  ?? src/statspai/did/ddd_heterogeneous.py       Olden-Møen 2022 (~250 LOC)
+  ?? src/statspai/did/did_multiplegt_dyn.py      dCDH 2024 MVP (~320 LOC)
+  ?? src/statspai/did/lp_did.py                  Dube-Girardi-Jordà-Taylor 2023 (~310 LOC)
+  ?? src/statspai/did/timevarying_covariates.py  Caetano et al. 2022 (~250 LOC)
[…]
