### 2026-04-24 9086415
fix(iv): v1.6.4 — ⚠️ correctness — 2SLS/LIML/Fuller cluster & robust SE use projected X̂ in meat
_k_class_fit passed unprojected X_actual to _cluster_cov/_robust_cov,
inflating reported cluster/HC SE by a factor that depends on first-stage
fit (up to ~2.5× on the audit DGP). Fixed by computing AX = A @ X_actual
and passing it as the sandwich meat, matching Cameron-Miller (2015),
Stata ivregress, ivreg2 (Baum-Schaffer-Stillman 2007), and linearmodels.

Point estimates unchanged. GMM and JIVE paths already used the correct
formula and are unaffected. Nonrobust default SE unchanged.

Adds tests/r
[…]
