+++ tests/reference_parity/test_iv_se_parity.py
@@ -1,0 +1,207 @@
+"""IV robust / clustered SE parity tests.
+
+Validates that ``sp.iv.iv`` / ``sp.ivreg`` produce the correct 2SLS
+sandwich covariance using the **projected** regressor matrix
+:math:`\\hat X = P_W X` in the meat, not the original :math:`X`.
+
+References
+----------
+- Cameron & Miller (2015). A Practitioner's Guide to Cluster-Robust
+  Inference. *Journal of Human Resources*, 50(2), 317-372.
+- Baum, Schaffer, Stillman (2007). Enhanced routines for IV/GMM
+  estimation and testing. *Stata Journal*, 7(4), 465-506.
+- linearmodels.iv.covariance.ClusteredCovariance (docstring &
+  implementation at :pycode:`xhat_e = z @ (pinvz @ x) * eps`).
+
+The correct 2SLS cluster "meat" is
+
+.. math::
+
+   \\hat S = \\sum_c \\Big(\\sum_{i\\in c}\\hat x_i \\hat u_i\\Big)
+                      \\Big(\\sum_{i\\in c}\\hat x_i \\hat u_i\\Big)^{\\top}
+
+with :math:`\\hat X = P_W X`, where :math:`P_W` is the projection onto
+the full instrument set (exog + excluded).  Using the unprojected X in
+the meat gives a consistent estimator for OLS, but inflates the 2SLS
+cluster SE by a factor depending on how poorly the first stage fits
+the endogenous regressor.
+
+The "bread" of the sandwich is :math:`(X'AX)^{-1}` which for k-class
+with :math:`A = (1-\\kappa)I + \\kappa P_W` equals
+:math:`(\\hat X'\\hat X)^{-1}` when :math:`\\kappa = 1`.  That part
+has always been correct in StatsPAI; this test targets only the meat.
+"""
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+
+
+# ---------------------------------------------------------------------------
+# Deterministic DGP with cluster structure and weak-ish first stage so that
+# X_hat and X differ substantively in the endog column.
+# ---------------------------------------------------------------------------
+
+@pytest.fixture(scope="module")
+def iv_cluster_data():
+    rng = np.random.default_rng(20260424)
+    n_clusters = 40
+    cluster_size = 25
+    n = n_clusters * cluster_size
+    cl = np.repeat(np.arange(n_clusters), cluster_size)
+    alpha = rng.normal(size=n_clusters)[cl]
+    eta = rng.normal(size=n_clusters)[cl]
+    z1 = rng.normal(size=n)
+    z2 = rng.normal(size=n)
+    x = rng.normal(size=n)
+    d = 0.7 * z1 + 0.4 * z2 + 0.3 * x + 0.6 * eta + rng.normal(size=n)
+    u = 0.8 * alpha + rng.normal(size=n)
+    y = 1.0 + 2.0 * d + 0.5 * x + u
+    return pd.DataFrame({"y": y, "d": d, "z1": z1, "z2": z2, "x": x, "cl": cl})
+
+
+def _hand_2sls_sandwich(df, cov_type: str):
+    """Closed-form 2SLS sandwich using projected X̂ in the meat.
+
+    Returns the vector of SEs in the order ``[const, x, d]``.  Uses
+    Cameron–Miller (2015) cluster correction ``(G/(G-1))((n-1)/(n-k))``
+    and Wooldridge HC1 correction ``n/(n-k)`` — matching StatsPAI's
+    existing finite-sample conventions so alignment is byte-exact.
+    """
+    y = df["y"].to_numpy()
[…]
