### 2026-04-24 e756901
fix(iv): v1.6.5 — ⚠️ correctness — standalone sp.liml κ solver + cluster/robust meat
Two independent bugs fixed in regression/advanced_iv.py::liml, an orphan
standalone LIML implementation separate from the v1.6.4-fixed
_k_class_fit dispatcher:

1. κ_LIML solver switched from np.linalg.eigvals(inv(A) @ B) (non-
   symmetric product, can return complex eigenvalues and a biased κ)
   to scipy.linalg.eigh(S_exog, S_full) (proper generalized symmetric
   eigenvalue problem). Aligns with the canonical iv.py::_liml_kappa
   implementation. Point estimates β̂ now correct.
2. Cluster / robust sandwich
[…]
