+++ tests/reference_parity/test_liml_se_parity.py
@@ -1,0 +1,186 @@
+"""Standalone LIML (``sp.liml`` / ``sp.iv.liml``) SE parity tests.
+
+The standalone LIML estimator in ``statspai.regression.advanced_iv`` is a
+separate code path from the ``_k_class_fit`` dispatcher (fixed in v1.6.4):
+users reach it via ``sp.liml(...)`` directly, not via
+``sp.ivreg(method='liml')``. This test pins its cluster / robust SE to:
+
+1. A hand-computed sandwich using the k-class transformed regressor
+   ``AX = (I − κ M_Z) X``, per the k-class FOC
+   ``X' (I − κ M_Z) (y − X β) = 0``.
+2. ``linearmodels.iv.IVLIML`` with ``cov_type='clustered', debiased=True``.
+
+Before v1.6.5 the standalone LIML used raw ``X`` in the cluster / robust
+meat, identical in character to the 2SLS bug fixed in v1.6.4 but in a
+different module.
+"""
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+
+
+@pytest.fixture(scope="module")
+def iv_cluster_data():
+    rng = np.random.default_rng(20260424)
+    n_clusters = 40
+    cluster_size = 25
+    n = n_clusters * cluster_size
+    cl = np.repeat(np.arange(n_clusters), cluster_size)
+    alpha = rng.normal(size=n_clusters)[cl]
+    eta = rng.normal(size=n_clusters)[cl]
+    z1 = rng.normal(size=n)
+    z2 = rng.normal(size=n)
+    x = rng.normal(size=n)
+    d = 0.7 * z1 + 0.4 * z2 + 0.3 * x + 0.6 * eta + rng.normal(size=n)
+    u = 0.8 * alpha + rng.normal(size=n)
+    y = 1.0 + 2.0 * d + 0.5 * x + u
+    return pd.DataFrame({"y": y, "d": d, "z1": z1, "z2": z2, "x": x, "cl": cl})
+
+
+def _hand_liml_cluster_sandwich(df):
+    """Hand-computed LIML cluster SE using (I−κM_Z)X in the meat.
+
+    Uses ``scipy.linalg.eigh`` on the symmetric generalized eigenvalue
+    problem to pick κ — more numerically stable than the raw
+    ``np.linalg.eigvals(inv(A) @ B)`` path and matches ``linearmodels``
+    to machine precision on well-conditioned DGPs.
+
+    Returns (SE vector in order [const, x, d], params, kappa).
+    """
+    from scipy.linalg import eigh as _eigh
+    y = df["y"].to_numpy()
+    X_exog = np.column_stack([np.ones(len(df)), df["x"].to_numpy()])
+    X_endog = df["d"].to_numpy().reshape(-1, 1)
+    Z = np.column_stack([df["z1"].to_numpy(), df["z2"].to_numpy()])
+    Z_all = np.column_stack([X_exog, Z])
+    X_all = np.column_stack([X_exog, X_endog])
+    cl = df["cl"].to_numpy()
+
+    n, k = X_all.shape
+    Px = X_exog @ np.linalg.solve(X_exog.T @ X_exog, X_exog.T)
+    Mx = np.eye(n) - Px
+    Pz = Z_all @ np.linalg.solve(Z_all.T @ Z_all, Z_all.T)
+    Mz = np.eye(n) - Pz
+
+    # κ_LIML = smallest root of S_exog v = κ S_full v, both symmetric PSD.
+    W = np.column_stack([y.reshape(-1, 1), X_endog])
[…]
