### 2026-04-24 b552a3e
release: v1.6.6 — ⚠️ Heckman SE correctness fix + HDFE LSMR/LSQR solver
Heckman (1979) two-step analytical variance
--------------------------------------------
regression/heckman.py::heckman previously reported an HC1-style sandwich
that the source itself flagged as "Heckman SEs are complex; robust is
conservative". This release replaces it with the textbook Heckman
(1979) / Greene (2003, eq. 22-22) / Wooldridge (2010, §19.6)
two-step analytical variance:

  V(β̂) = σ̂² (X*'X*)⁻¹ [X*'(I − ρ̂² D_δ) X* + ρ̂² F V̂_γ F'] (X*'X*)⁻¹

where δ_i = λ̂_i(λ̂_i + Z_iγ̂), D_δ = diag(δ_i), F = X*' D_δ Z, a
[…]
