+++ tests/reference_parity/test_heckman_se_parity.py
@@ -1,0 +1,202 @@
+"""Heckman (1979) two-step SE parity tests.
+
+Before v1.6.6 ``sp.heckman`` reported an ad-hoc HC1-style sandwich that
+ignored both (a) the heteroskedasticity induced by selection,
+``Var(y|X, D=1) = σ² (1 − ρ² δ_i)``, and (b) the uncertainty in γ̂
+(the probit first-stage estimator) propagated through the inverse
+Mills ratio λ̂ — the "generated regressor" problem.
+
+This release restores the textbook Heckman (1979) / Greene (2003) /
+Wooldridge (2010) two-step analytical variance:
+
+.. math::
+
+   V(\\hat β) = \\hat σ²\\, (X_*'X_*)^{-1}
+                \\big[ X_*'(I − \\hat ρ² D_δ) X_*
+                       + \\hat ρ²\\, F V̂_γ F' \\big]
+                (X_*'X_*)^{-1}
+
+where
+
+- ``X_*`` is the second-stage design including λ̂ as its last column,
+- ``δ_i = λ̂_i (λ̂_i + Z_i γ̂) ≥ 0`` (Heckman 1979, eq. 17),
+- ``D_δ = diag(δ_i)``,
+- ``F = X_*'\\, D_δ\\, Z`` (``k × q``),
+- ``V̂_γ = (Z'\\,diag(w_i)\\,Z)^{-1}`` with probit info weights
+  ``w_i = φ(Z_iγ̂)²/[Φ(Z_iγ̂)(1 − Φ(Z_iγ̂))]``,
+- ``σ̂² = RSS / n + β̂_λ² · mean(δ_i)`` (Greene 2003, eq. 22-21).
+
+References
+----------
+- Heckman, J. J. (1979). "Sample Selection Bias as a Specification
+  Error." *Econometrica*, 47(1), 153–161.
+- Greene, W. H. (2003). *Econometric Analysis*, 5th ed., §22.4.
+- Wooldridge, J. M. (2010). *Econometric Analysis of Cross Section and
+  Panel Data*, 2nd ed., §19.6.
+"""
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+from scipy import stats
+
+import statspai as sp
+
+
+# ---------------------------------------------------------------------------
+# DGP with known selection structure and valid exclusion restriction
+# ---------------------------------------------------------------------------
+
+@pytest.fixture(scope="module")
+def heckman_data():
+    """n=2000 units. z1 enters both equations (kept in x), z2 enters only
+    selection (valid exclusion restriction). Errors correlated ρ=0.5.
+    """
+    rng = np.random.default_rng(20260424)
+    n = 2000
+    z1 = rng.normal(size=n)
+    z2 = rng.normal(size=n)
+    rho = 0.5
+    u1 = rng.normal(size=n)
+    u2 = rho * u1 + np.sqrt(1 - rho ** 2) * rng.normal(size=n)
+    # Selection equation (probit latent)
+    d_star = 0.2 + 0.6 * z1 + 0.7 * z2 + u1
+    d = (d_star > 0).astype(int)
+    # Outcome (observed only when d == 1)
+    y = 1.0 + 2.0 * z1 + u2
+    return pd.DataFrame({"y": y, "z1": z1, "z2": z2, "d": d})
+
+
+# ---------------------------------------------------------------------------
+# Hand-computed Heckman (1979) analytical variance
+# ---------------------------------------------------------------------------
+
+def _hand_heckman_variance(df):
+    """Compute (β̂, analytical SE) using the Heckman (1979) / Greene
[…]
