### 2026-04-24 43ff5be
docs(migration): fix pyreghdfe API in migration table — absorb= -> fe=
The API mapping table for pyreghdfe -> StatsPAI in MIGRATION.md's
"Migrating from pyreghdfe" section showed the left column as
`reghdfe(y, x, absorb=['firm','year'], data=df, ...)`. That is Stata's
reghdfe syntax, not the Python port's: pyreghdfe uses `fe=[...]` as
the keyword, not `absorb=[...]`. Corrected to
`reghdfe(data=df, y='y', x=['x'], fe=['firm','year'], cluster=['firm'])`
to match the actual 0.2.1 / 0.3.0 API users will migrate from.

Documentation-only fix; no code, no version bump.

Co-Authored-By: Claude Opus 4
[…]
