### 2026-04-25 214e26e
docs(skill): rewrite empirical-analysis skill in AER style + multi-regression cookbook
Rename StatsPAI_skill → StatsPAI_full_data_analysis_skill and rewrite from a
generic 7-step pipeline into the canonical 8-section pipeline of an applied
AER / QJE / AEJ paper:

§-1 Pre-Analysis Plan · §0 Data construction + sample log · §1 Multi-panel
Table 1 + Figure 1 trends/rollout · §2 Empirical strategy with explicit
estimating equation × identifying assumption table · §3 Identification graphs
(event-study + numerical pre-trends, Bacon weights, CS-DID dynamics, RD plot,
McCrary, love plot, SCM trajector
[…]
