+++ StatsPAI_full_data_analysis_skill/README.md
@@ -1,0 +1,132 @@
+# StatsPAI Skill for Claude Code
+
+This folder is a **Claude Code Skill** that teaches Claude (or any
+compatible agent harness) how to drive [StatsPAI](https://github.com/brycewang-stanford/StatsPAI)
+end-to-end through the **canonical pipeline of an applied AER /
+QJE / AEJ empirical paper** — pre-analysis plan, Table 1 summary
+statistics, written-out estimating equation and identifying
+assumption, identification graphics (event-study / first-stage F /
+McCrary / SCM trajectory / love plot), the AER multi-regression
+"Table 2" gauntlet (progressive controls / design horse-race /
+multi-outcome / IV first-stage triplet), heterogeneity (subgroup
+table + CATE figure), mechanisms, and a full robustness battery
+(placebo / Oster / honest_did / E-value / Conley / two-way SE /
+spec curve / sensitivity dashboard) — ending in a paper-ready
+replication package.
+
+## Install
+
+```bash
+# Option 1: copy
+cp -r StatsPAI_skill ~/.claude/skills/StatsPAI_skill
+
+# Option 2: symlink (auto-follow upstream updates)
+ln -s "$(pwd)/StatsPAI_skill" ~/.claude/skills/StatsPAI_skill
+```
+
+Then install the Python package itself:
+
+```bash
+pip install statspai
+```
+
+## Activate
+
+The skill auto-activates on natural-language triggers such as
+*"run a DID analysis"*, *"AER empirical analysis"*, *"applied
+microeconomics pipeline"*, *"instrumental variables regression"*,
+*"event-study plot"*, *"first-stage F"*, *"Oster bound"*,
+*"estimand-first DSL"*, etc. See the `triggers:` block in
+`SKILL.md` for the full list.
+
+## Scope
+
+**In scope** — the canonical AER 8-section empirical pipeline:
+
+```text
+§-1 Pre-Analysis Plan      →  §0 Data construction & contract
+§1  Table 1 (descriptives) →  §2 Empirical strategy (equation + identifying assumption)
+§3  Identification graphics (event-study / first-stage / McCrary / SCM trajectory / love plot)
+§4  Multi-regression main tables (progressive controls / design horse-race / multi-outcome /
+                                  panel A-B / IV first-stage triplet) + coefplot
+§5  Heterogeneity (subgroup table + CATE / dose-response figure)
+§6  Mechanisms (mediation / decomposition)
+§7  Robustness gauntlet (placebo / Oster / honest_did / E-value / 2-way & Conley SE /
+                         spec curve / sensitivity dashboard) + robustness master table
+§8  Replication package (LaTeX tables + figures + reproducibility stamp)
+```
+
+**Out of scope** — data cleaning (use pandas first) and end-to-end
+paper drafting (`sp.paper()`; the skill stops at diagnostics and
+hands the `CausalResult` back to you).
+
+## Files
+
[…]
