+++ StatsPAI_full_data_analysis_skill/SKILL.md
@@ -1,0 +1,1139 @@
+---
+name: StatsPAI_skill
+description: Use when the user asks to run a full empirical / causal analysis in Python in the style of an applied economics paper (AER / QJE / JPE / ReStud / AEJ), pick between DID / RD / IV / SCM / DML / matching, write down an estimating equation and identifying assumption, produce Table 1 / Table 2 / event-study figure / robustness gauntlet, or generate a paper-ready replication package. Also triggers on keywords "StatsPAI", "statspai", "AER empirical analysis", "applied micro pipeline", "Table 1 balance", "event study", "first-stage F", "Oster bound", "honest_did", "spec_curve", "callaway_santanna", "dragonnet", "text as treatment".
+triggers:
+  - causal inference in python
+  - applied microeconomics pipeline
+  - AER empirical analysis
+  - QJE style robustness
+  - DID IV RD SCM
+  - callaway_santanna
+  - synthetic control
+  - double machine learning
+  - causal forest
+  - event study plot
+  - first stage F-statistic
+  - Oster bound
+  - honest_did
+  - spec_curve
+  - estimand-first DSL
+  - LLM-assisted DAG discovery
+  - text as treatment
+  - StatsPAI
+  - statspai
+---
+
+# StatsPAI: Agent-Native Causal Inference & AER-Style Empirical Workflow
+
+StatsPAI is the agent-native Python package for causal inference and applied econometrics: one `import statspai as sp`, 900+ functions behind a self-describing API, and `CausalResult` objects that export to LaTeX / Word / Excel / BibTeX.
+
+This skill drives StatsPAI through the **canonical pipeline of an applied AER empirical paper**. Each step maps to a section of the published paper and emits a paper-ready artifact (Table 1, event-study figure, Table 2 main results, robustness panel, replication stamp).
+
+- **Source**: https://github.com/brycewang-stanford/StatsPAI
+- **Install**: `pip install statspai` (>= 1.6)
+- **Paper**: submitted to JOSS (under review)
+
+## Why for Agents
+
+1. **Self-describing**: `sp.list_functions()` / `sp.describe_function(name)` / `sp.function_schema(name)` — every public symbol is discoverable without doc lookup.
+2. **Unified result**: every estimator returns `CausalResult` with `.summary()`, `.plot()`, `.diagnostics`, `.to_latex()`, `.to_word()`, `.cite()`.
+3. **One import, full pipeline**: data contract → Table 1 → estimand-first DSL → identification graphs → main table → heterogeneity → mechanisms → robustness → replication package.
+4. **Estimand-first**: `sp.causal_question(...).identify()` forces the "DID vs RD vs IV?" decision *before* estimation, with the identifying assumption written down — the way a referee expects to read it.
+
+## The AER-style empirical pipeline
+
+The skill mirrors the canonical sections of an applied AER / QJE / AEJ paper. Each step below is one paper section and one set of artifacts on disk.
+
+```
+Paper section               Step  StatsPAI moves
+─────────────────────────── ───── ────────────────────────────────────────────────
+Pre-Analysis Plan           −1    sp.power.* + freeze IdentificationPlan to disk
+§1. Data                     0    data_contract + sample-construction log (footnote 4)
+§1.1 Descriptives (Table 1)  1    sp.sumstats · sp.balance_table · sp.describe
+§2. Empirical Strategy       2    write equation + identifying assumption + sp.causal_question
+   (LLM-DAG addendum)        2.5  sp.llm_dag_propose · validate · constrained
[…]
