### 2026-04-25 5f3bb07
docs(readme): rename screenshot image-1.png → StatsPAI-interactive.png
Replace cryptic image-1.png filename with descriptive
StatsPAI-interactive.png (fixes earlier 'StasPAI' typo on the
incoming screenshot) and update both README.md and README_CN.md
GitHub raw image links accordingly. Renders identically; just a
clearer filename.
