### 2026-04-25 329353c
docs(readme): move hero screenshot into docs/assets/
Relocate StatsPAI-interactive.png from repo root to docs/assets/
following the MkDocs convention (docs/ is auto-served, assets/
is the canonical media folder). Both README.md and README_CN.md
GitHub raw URLs updated; image renders identically. Keeps repo
root tidy (only meta + code files).
