### 2026-04-25 5b4fcdd
docs(skill): fix 9 broken sp.* call signatures in AER pipeline skill
Previous rewrite (commit 214e26e) claimed "All API calls signature-checked"
but 9 call patterns failed at runtime against installed StatsPAI 1.6.6.
Replayed every code block end-to-end (10/10 smoke tests pass after fix).

Surgical corrections:
- §-1 power: power_did/power_cluster_rct(power_target=) → sp.power(design,
  power_target=, ...) dispatcher (wrappers don't auto-solve)
- §1.1 Table 1: sumstats(groups=panels) (no such kwarg) → loop sumstats
  per panel and concat with %-Panel markers
- §2.3 freeze pre-registration: pla
[…]
