+++ StatsPAI_full_data_analysis_skill/SKILL.md
@@ -90,11 +90,18 @@ A modern AER paper has **5–7 figures** and **3–5 main tables** + an appendix
 `sp.power(design, n=..., effect_size=..., power_target=...)` is a unified dispatcher — leave one argument `None` to solve for it (sample size, MDE, or power). Convenience wrappers: `sp.power_rct`, `sp.power_did`, `sp.power_rd`, `sp.power_iv`, `sp.power_cluster_rct`, `sp.power_ols`.
 
 ```python
-sp.power("rct", effect_size=0.3, power_target=0.80)                # → PowerResult(n=349, power=0.80)
-sp.power("did", n=200, effect_size=0.15, power_target=0.80)        # DID: solves MDE / n / power
-sp.power_cluster_rct(n_clusters=30, cluster_size=50,
-                     icc=0.05, effect_size=0.2, power_target=0.80) # Cluster RCT
-sp.pretrends_power(result)                                          # Roth (2022) pre-trends power
+# Always go through the dispatcher when you want auto-solve. The
+# `sp.power_<design>` wrappers (power_rct / power_did / power_rd /
+# power_iv / power_cluster_rct / power_ols) accept *only* the design's
+# native arguments — they will NOT solve for power_target / n / effect
+# unless you go via `sp.power(design, ..., power_target=...)`.
+
+sp.power("rct", effect_size=0.3, power_target=0.80)                  # → PowerResult(n=349, power=0.80)
+sp.power("did", n=200, effect_size=0.15, power_target=0.80,
+         n_periods=4, n_treated_periods=2)                            # DID: solves MDE / n / power
+sp.power("cluster_rct", cluster_size=50, icc=0.05,
+         effect_size=0.2, power_target=0.80)                          # Cluster RCT: solves n_clusters
+sp.pretrends_power(result)                                            # Roth (2022) pre-trends power
 ```
 
 Persist the `PowerResult` next to `data_contract.json` and `empirical_strategy.md` — a referee will ask whether the design was powered before data collection, not after.
@@ -198,7 +205,7 @@ sp.describe(df).to_markdown("references/codebook.md")              # auto-codebo
 ```
 
 ### 1.1 Multi-panel Table 1 (AER convention)
-Group rows into **Panel A: Outcomes**, **Panel B: Treatment intensity**, **Panel C: Controls**, **Panel D: Sample composition**. `sp.sumstats` accepts `groups=` for panel-style layout, and the result can be merged into one tex file:
+Group rows into **Panel A: Outcomes**, **Panel B: Treatment intensity**, **Panel C: Controls**, **Panel D: Sample composition**. `sp.sumstats` does not accept `groups=` directly — emit one Panel per call and concatenate. The result is one `.tex` file with `% Panel X` markers you can replace with `\midrule \multicolumn{...}{l}{\textit{Panel X. ...}} \\` during paper assembly.
 
 ```python
 panels = {
@@ -207,9 +214,14 @@ panels = {
     "C. Demographic controls": ["age", "edu", "female", "married"],
     "D. Labor market":         ["tenure", "firm_size", "industry_id"],
 }
-sp.sumstats(df, groups=panels, by="training",
-            stats=["mean", "sd", "n"],
-            output="tables/table1_summary.tex")
+import io
+buf = io.StringIO()
+for label, vs in panels.items():
+    buf.write(f"\n% Panel {label}\n")
+    buf.write(sp.sumstats(df, vars=vs, by="training",
+                          stats=["mean", "sd", "n"],
[…]
