### 2026-04-25 6aa5222
docs(skill): integrate Word/Excel export stack (regtable/paper_tables/Collection)
Wire StatsPAI's three-tier export stack into the AER pipeline skill so the
agent ships .docx + .xlsx + .tex (not just .tex) for every numbered table.
Agent-native equivalent of Stata outreg2 / esttab / collect and R
modelsummary / summary_col / gtsave.

Three tiers documented + threaded through every regtable example:
  Tier 1  rt = sp.regtable(M1..M5, template='aer', title=...)
          rt.to_word('table2.docx'); rt.to_excel('table2.xlsx')
  Tier 2  sp.paper_tables(main=, heterogeneity=, robustness=, placebo=,

[…]
