+++ StatsPAI_full_data_analysis_skill/SKILL.md
@@ -64,24 +64,61 @@ Pre-Analysis Plan           −1    sp.power.* + freeze IdentificationPlan to di
 
 ## Paper-ready figure & table inventory (what to produce by section)
 
-A modern AER paper has **5–7 figures** and **3–5 main tables** + an appendix robustness table. Every step below should leave at least one numbered artifact on disk. Default file names (use these so the agent can reference them later):
+A modern AER paper has **5–7 figures** and **3–5 main tables** + an appendix robustness table. Every step below should leave at least one numbered artifact on disk. Default file names assume parallel `.tex` / `.docx` / `.xlsx` exports (the agent should produce all three so co-authors can edit in Word / Excel and the build system can use LaTeX):
 
-| § | Artifact | StatsPAI primitive | Filename |
+| § | Artifact | StatsPAI primitive | Filenames (write all three) |
 |---|---|---|---|
 | §1 | **Figure 1**: raw trends / treatment rollout | `sp.parallel_trends_plot` · `sp.treatment_rollout_plot` | `figures/fig1_trends.png` |
-| §1 | **Table 1**: summary stats (full / treated / control + Δ) | `sp.sumstats(by=)` · `sp.balance_table` | `tables/table1_summary.tex` |
+| §1 | **Table 1**: summary stats (full / treated / control + Δ) | `sp.sumstats` + `sp.mean_comparison(...).to_word()/.to_excel()` (or `sp.collect().add_summary().add_balance()`) | `tables/table1_summary.{tex,docx,xlsx}` |
 | §3 | **Figure 2**: identification graphic (event-study / first-stage / McCrary / RD scatter / SCM trajectory) | `sp.enhanced_event_study_plot` · `sp.binscatter` · `sp.rdplot` · `sp.rddensity().plot()` · `sp.synthdid_plot` | `figures/fig2_identification.png` |
-| §4 | **Table 2**: main results — progressive controls | `sp.regtable(M1, M2, M3, M4, ...)` | `tables/table2_main.tex` |
-| §4 | **Table 2-bis**: design horse-race (OLS / IV / DID / DML) | `sp.regtable(ols, iv, did, dml, ...)` | `tables/table2b_designs.tex` |
+| §4 | **Table 2**: main results — progressive controls | `rt = sp.regtable(M1...M5, template="aer"); rt.to_word(...); rt.to_excel(...)` | `tables/table2_main.{tex,docx,xlsx}` |
+| §4 | **Table 2-bis**: design horse-race (OLS / IV / DID / DML) | `sp.regtable(ols, iv, did, dml, ...).to_word/.to_excel` | `tables/table2b_designs.{tex,docx,xlsx}` |
 | §4 | **Figure 3** (optional): coefficient plot across specs | `sp.coefplot(M1, M2, M3, M4)` | `figures/fig3_coef.png` |
-| §5 | **Table 3**: heterogeneity by subgroup | `sp.regtable(g_full, g_male, g_fem, g_q1...q4)` | `tables/table3_heterogeneity.tex` |
+| §5 | **Table 3**: heterogeneity by subgroup | `sp.regtable(g_full, g_male, g_fem, g_q1...q4).to_word/.to_excel` | `tables/table3_heterogeneity.{tex,docx,xlsx}` |
 | §5 | **Figure 4**: dose-response / CATE | `sp.dose_response(...).plot()` · `sp.cate_plot` · `sp.cate_group_plot` | `figures/fig4_cate.png` |
-| §6 | **Table 4**: mechanisms (mediation / decomposition) | `sp.regtable(total, direct, indirect)` | `tables/table4_mechanisms.tex` |
-| §7 | **Table A1**: robustness master (one row per check) | `sp.regtable(rob1...robN, panel_labels=[...])` | `tables/tableA1_robustness.tex` |
[…]
