### 2026-04-25 4571592
feat(output): output-layer overhaul — AER/QJE DOCX, paper_tables export, sp.collect, regtable.alpha, Quarto cross-refs
Five additive output-layer enhancements; no breaking changes, no
numerical changes to any estimator.

- regtable(..., quarto_label=, quarto_caption=) emits Quarto
  ":caption {#tbl-<id>}" cross-ref blocks via .to_quarto() /
  .to_markdown(quarto=True) / save("table.qmd").
- regtable(..., alpha=) now actually controls CI width (90/95/99).
- to_word/to_docx output adopts AER/QJE book-tab style (heavy top &
  bottom rules, thin mid rule, no internal borders, Times New Roman).
  H
[…]
