+++ src/statspai/output/paper_tables.py
@@ -21,53 +21,29 @@ from dataclasses import dataclass, field
 from typing import Any, Dict, List, Optional, Sequence
 
 from .regression_table import regtable, RegtableResult
+from ._journals import JOURNALS
 
 
 # ---------------------------------------------------------------------------
-# Journal templates
+# Journal templates (re-exported from ._journals for backwards compatibility)
 # ---------------------------------------------------------------------------
 
+# Build a TEMPLATES dict with the same shape as the historical
+# paper_tables.TEMPLATES (with ``se_type`` and ``notes_default`` as a list)
+# while delegating the actual content to the single-source-of-truth
+# :data:`statspai.output._journals.JOURNALS`. Adding a journal in
+# ``_journals.py`` automatically exposes it here.
 TEMPLATES = {
-    'aer': {
-        'label': 'AER style',
-        'star_levels': (0.10, 0.05, 0.01),
-        'se_type': 'se',
-        'stats': ('N', 'R-squared'),
-        'notes_default': [
-            "Standard errors in parentheses.",
-            "*** p<0.01, ** p<0.05, * p<0.10.",
-        ],
-    },
-    'qje': {
-        'label': 'QJE style',
-        'star_levels': (0.10, 0.05, 0.01),
-        'se_type': 'se',
-        'stats': ('N',),
-        'notes_default': [
-            "Robust standard errors in parentheses.",
-            "*** p<0.01, ** p<0.05, * p<0.10.",
-        ],
-    },
-    'econometrica': {
-        'label': 'Econometrica style',
-        'star_levels': (0.01, 0.05),     # only 5% and 1%
-        'se_type': 'se',
-        'stats': ('N', 'R-squared'),
-        'notes_default': [
-            "Standard errors in parentheses.",
-            "** p<0.05, *** p<0.01.",
-        ],
-    },
-    'restat': {
-        'label': 'Review of Economics and Statistics style',
-        'star_levels': (0.10, 0.05, 0.01),
-        'se_type': 'se',
-        'stats': ('N', 'R-squared'),
-        'notes_default': [
-            "Standard errors in parentheses.",
-            "*** p<0.01, ** p<0.05, * p<0.10.",
-        ],
-    },
+    name: {
+        "label": preset["label"],
+        "star_levels": tuple(preset["star_levels"]),
+        "se_type": "se",
+        "se_label": preset.get("se_label", "Standard errors"),
+        "stats": tuple(preset["stats"]),
+        "notes_default": list(preset["notes_default"]),
+        "font_name": preset.get("font_name", "Times New Roman"),
+    }
+    for name, preset in JOURNALS.items()
 }
 
 
@@ -142,6 +118,184 @@ class PaperTables:
             parts.append("")
         return "\n".join(parts)
 
+    # ------------------------------------------------------------------
+    # DOCX / XLSX export
+    # ------------------------------------------------------------------
+
+    def to_docx(self, path: str) -> str:
+        """Write every populated panel into a single ``.docx`` file.
+
+        Each panel renders as an AER/QJE book-tab table (heavy top rule,
+        thin mid rule, heavy bottom rule, no internal borders). Panels
+        are separated by a page break so the file lands on the
+        co-author's desk ready for direct insertion into a manuscript.
+
+        Returns the file path that was written.
+        """
+        try:
+            from docx import Document
+            from docx.shared import Pt
+            from docx.enum.text import WD_ALIGN_PARAGRAPH
+        except ImportError as e:
+            raise ImportError(
+                "python-docx is required for .docx export. "
+                "Install with: pip install python-docx"
+            ) from e
+
+        from ._aer_style import (
+            apply_word_booktab_rules,
+            style_word_table_typography,
+            add_word_notes_paragraph,
+        )
+
+        doc = Document()
+        panels = list(self.panels().items())
[…]
