### 2026-04-25 b2e042d
docs(skill): default regtable to full coef list (drop=Intercept, not keep=focal)
User-reported issue: exported Word/Excel Table A1 only showed the focal
coefficient (educ / training), hiding all controls. Root cause: every
regtable example in SKILL.md aggressively used keep=[focal_var], which
filters down to one coefficient row.

AER convention is the opposite: show ALL coefficients (controls included)
so the reader can verify the full spec, then let the user manually drop
rows in Word/Excel if they want a focal-only view.

Default policy across the skill:
  - drop=['Intercept']  → suppress on
[…]
