### 2026-04-25 226e07b
docs(skill): preserve Intercept in default regtable output too
User requested: keep the Intercept in the default regression table.
Previously the skill defaulted to drop=['Intercept'] (suppress constant);
now it defaults to NO filter at all so every estimated parameter — controls
AND the intercept — is exported verbatim. AER/QJE convention: show what was
actually estimated; let the user post-process if they want to hide rows.

Behavior:
  Default              regtable(*models, template='aer', ...)
                       → all coefs incl. Intercept (the new norm)
  Suppress constant    regtable
[…]
