### 2026-04-25 7abc62d
docs(skill): route FE regressions through sp.feols (sp.regress | fe is broken)
CRITICAL FIX: sp.regress is a thin statsmodels OLS wrapper that does NOT
parse '|' as a fixed-effects separator. Calls like
  sp.regress('wage ~ x | firm_id', df, cluster='firm_id')
silently produce wrong numbers — statsmodels treats 'x | firm_id' as a
single garbage variable name. The skill had ~15 such calls in §4 main /
§5 heterogeneity / §7.11 robustness master, all silently broken.

The correct API is sp.feols (pyfixest backend, AER/QJE workhorse) with
pyfixest's vcov dict syntax for clustering:
  sp.feols('wag
[…]
