+++ StatsPAI_full_data_analysis_skill/SKILL.md
@@ -516,6 +516,15 @@ This is the densest section of an applied paper. A modern AER §4 typically cont
 - **Table 2-ter** (multi-outcome): same treatment, several outcomes side-by-side
 - **Figure 3** (coefplot): visual summary of β̂ and 95% CI across specs
 
+> **Estimator routing** (memorize this — getting it wrong silently produces nonsense):
+> - **No FE** → `sp.regress("y ~ x1 + x2", df, cluster="firm_id")`
+> - **High-dim FE** → `sp.feols("y ~ x1 + x2 | fe1 + fe2", df, vcov={"CRV1":"firm_id"})`
+> - **Two-way cluster** → `sp.feols(..., vcov={"CRV1":"firm_id+year"})`
+> - **2SLS / IV** → `sp.ivreg("y ~ (x ~ z) + controls", df, cluster=...)`
+> - **DID / event-study** → `sp.callaway_santanna(...)` / `sp.sun_abraham(...)`
+>
+> **Never** write `sp.regress("y ~ x | firm_id")` — `sp.regress` does not parse `|` and silently treats `x | firm_id` as a single variable name. Use `sp.feols` for any formula containing `|`.
+
 `sp.regtable(*models, ...)` is the workhorse. Useful kwargs:
 
 ```
@@ -539,13 +548,16 @@ Stable β̂ across columns ⇒ less concern that selection on observables is dri
 | Controls | none | age, edu | + tenure, firm_size | high-dim FE | individual FE |
 
 ```python
-M1 = sp.regress("wage ~ training",                                     df, cluster="firm_id")
-M2 = sp.regress("wage ~ training + age + edu",                         df, cluster="firm_id")
-M3 = sp.regress("wage ~ training + age + edu + tenure + firm_size",    df, cluster="firm_id")
-M4 = sp.regress("wage ~ training + age + edu + tenure + firm_size | "
-                "region + industry + year",                            df, cluster="firm_id")
-M5 = sp.regress("wage ~ training + age + edu + tenure + firm_size | "
-                "worker_id + year",                                    df, cluster="firm_id")
+# RULE: pure OLS → sp.regress; high-dim FE absorption → sp.feols
+# (sp.regress does NOT parse `|` as FE — it's a thin OLS wrapper. Use
+# `sp.feols("y ~ x | fe1 + fe2", df, vcov={"CRV1":"firm_id"})` for FE.)
+M1 = sp.regress("wage ~ training",                                  df, cluster="firm_id")
+M2 = sp.regress("wage ~ training + age + edu",                      df, cluster="firm_id")
+M3 = sp.regress("wage ~ training + age + edu + tenure + firm_size", df, cluster="firm_id")
+M4 = sp.feols  ("wage ~ training + age + edu + tenure + firm_size | region + industry + year",
+                df, vcov={"CRV1": "firm_id"})
+M5 = sp.feols  ("wage ~ training + age + edu + tenure + firm_size | worker_id + year",
+                df, vcov={"CRV1": "firm_id"})
 
 # Consolidate 5 models into ONE table (= Stata `outreg2 [M1..M5] using ..., replace`).
 # **Default = show ALL coefficients verbatim — controls AND the intercept**
@@ -574,8 +586,8 @@ open("tables/table2_main.tex", "w").write(rt.to_latex())
 Show the same coefficient of interest under multiple identification strategies. This is *the* AER credibility move: convergent evidence across designs each making different identifying assumptions.
 
 ```python
-ols  = sp.regress("wage ~ training + age + edu + tenure | industry + year",
-                   df, cluster="firm_id")                                                 # OLS / FE
[…]
