### 2026-04-25 822075b
docs(skill): refresh README + extend SKILL frontmatter triggers for new export stack
The README had drifted across 5 dimensions versus 5 SKILL.md commits this
sprint (Word/Excel export stack, sp.feols routing, default = full coef
list, journal templates, install path). The frontmatter triggers also
omitted every new-export keyword (regtable / collect / paper_tables /
feols / outreg2 in Python / etc.), so users who say 'export regression
to Word' would not even activate the skill.

README.md (+127 / -23):
- New 'What this skill ships' section: 3-tier export table (regtable /
  paper_tables / sp
[…]
