+++ StatsPAI_full_data_analysis_skill/README.md
@@ -11,23 +11,62 @@ McCrary / SCM trajectory / love plot), the AER multi-regression
 multi-outcome / IV first-stage triplet), heterogeneity (subgroup
 table + CATE figure), mechanisms, and a full robustness battery
 (placebo / Oster / honest_did / E-value / Conley / two-way SE /
-spec curve / sensitivity dashboard) — ending in a paper-ready
-replication package.
+spec curve / sensitivity dashboard) — ending in a single editable
+Word / Excel / LaTeX replication bundle.
+
+## What this skill ships (paper-ready, in three formats)
+
+Every numbered table in the AER pipeline emits parallel
+**`.docx` + `.xlsx` + `.tex`** (Markdown / HTML on demand) so
+co-authors can edit in Word, journal editors can review in Excel,
+and the build system can assemble LaTeX — all from one source.
+
+| Tier | When | API |
+| --- | --- | --- |
+| **1. Single multi-column table** (Stata `outreg2` / R `modelsummary` equivalent) | One Table 2 / Table 3 / Table A1 with progressive columns | `rt = sp.regtable(M1, M2, ..., template="aer", title=...)` then `rt.to_word("table2.docx") · rt.to_excel("table2.xlsx") · rt.to_latex()` |
+| **2. Multi-panel paper format** | Tables 2 + 3 + A1 + A2 in one document | `sp.paper_tables(main=, heterogeneity=, robustness=, placebo=, template="aer").to_docx("paper_tables.docx")` |
+| **3. Full session bundle** (Stata 15 `collect` equivalent) | Replication appendix mixing summary + balance + multiple regression tables + headings + prose | `sp.collect("Paper").add_summary(...).add_balance(...).add_regression(...)...save("paper.docx")` (auto-detects `.docx`/`.xlsx`/`.tex`/`.md`/`.html`) |
+
+Journal templates auto-apply the right SE label, star levels, and
+footer notes:
+
+```python
+sp.list_journal_templates()
+# → ('aer', 'qje', 'econometrica', 'restat', 'jf', 'aeja', 'jpe', 'restud')
+
+rt = sp.regtable(M1, M2, M3, template="qje")     # QJE styling
+rt.to_word("table2_qje.docx")
+```
+
+Inline citations drop a coefficient straight into prose:
+
+```python
+sp.cite(M3, "training")           # → "1.239*** (0.153)"
+```
+
+## Default behavior the agent enforces
+
+| Convention | Default action | When to override |
+| --- | --- | --- |
+| **Show every estimated parameter verbatim** (controls AND intercept) | `sp.regtable(*models, template="aer", ...)` with NEITHER `keep=` NOR `drop=` | Add `drop=["Intercept"]` to suppress the constant; `keep=[focal]` for intentionally focal-only tables (IV first-stage triplet, interaction-form heterogeneity) |
+| **Route FE regressions through `sp.feols`** (pyfixest backend) | `sp.feols("y ~ x \| firm_id + year", df, vcov={"CRV1": "firm_id"})` | `sp.regress` for pure OLS only — it does **NOT** parse `\|` as FE |
[…]
