### 2026-04-26 e0115fe
release: v1.7.0 — Phase 2 output overhaul
Cuts CHANGELOG [Unreleased] → [1.7.0] (2026-04-25). Bumps pyproject.toml
and src/statspai/__init__.py to 1.7.0.

Highlights (additive; no numerical changes to any estimator):
- Journal presets (8): aer, qje, econometrica, restat, jf, aeja, jpe, restud
- Auto-diagnostic rows: FE / Cluster SE / IV first-stage F / DiD pretrend
  / RD bandwidth & kernel — emitted from result.model_info
- Multi-SE side-by-side via regtable(multi_se={...})
- sp.cite() inline coefficient reporter (text/latex/markdown/html)
- Reproducibility footer via repro={"data": df, "seed
[…]
