### 2026-04-26 e274e7c
fix(tests): skip FE-row test when pyfixest absent
CI installs [dev] extras only — pyfixest lives in [fixest] (lazy
import per CLAUDE.md §4). Aligns with the existing
test_fixest.py importorskip pattern so the v1.7.0 output-overhaul
matrix turns green again.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
