### 2026-04-26 5c57508
release: v1.7.1 — fmt="auto" magnitude-adaptive formatting + unified book-tab xlsx style
Pure-additive output-layer patch on top of v1.7.0. No numerical changes
to any estimator. Two themes — both close gaps that AER/QJE referees
and production editors flag in practice.

1. fmt="auto" — sp.regtable / sp.modelsummary pick decimal precision
   per-cell so a single table can mix dollar-magnitude coefficients
   (1,521) with elasticity-magnitude coefficients (0.288) without one
   side rounding to bare 0. Bucketing: |x|≥1000 → comma int; ≥100 →
   int; ≥10 → 1 dp; ≥1 → 2 dp; <1 → 3 dp. Closes the 
[…]
