### 2026-04-26 9e08153
docs(skill): fix sp.xtreg drift — point to sp.feols / sp.panel
The Stata→Python migration row claimed sp.xtreg as a same-name
equivalent, but hasattr(sp, 'xtreg') is False — statspai never
shipped that symbol. Stata xtreg's actual equivalents are
sp.feols("y ~ x | id + year", df) for OLS-with-FE and
sp.panel(..., method="fe"/"re") for the panel estimator family.

Caught by a sp.* symbol cross-check against the installed v1.7.1
package (98/98 other references resolve).

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
