### 2026-04-26 59d67a6
docs(skill): add epi (§A) + ML-causal (§B) sub-pipelines on top of AER default
The skill was AER-only — fine for applied econ users, but two of the
biggest causal-inference traditions (epi / public health and ML
causal inference) had to be hand-wired by users from sp.dml /
sp.tmle / sp.target_trial / sp.metalearner / sp.causal_forest /
sp.conformal_causal etc. without a paper-shaped scaffolding.

This commit adds two parallel sub-pipelines that mirror the AER
8-section skeleton and reuse the same export stack (sp.regtable /
sp.collect / sp.paper_tables) and CausalResult interface — only
the St
[…]
