+++ StatsPAI_full_data_analysis_skill/SKILL.md
@@ -1,6 +1,6 @@
 ---
 name: StatsPAI_skill
-description: Use when the user asks to run a full empirical / causal analysis in Python in the style of an applied economics paper (AER / QJE / JPE / ReStud / AEJ), pick between DID / RD / IV / SCM / DML / matching, write down an estimating equation and identifying assumption, produce Table 1 / Table 2 / event-study figure / robustness gauntlet, OR export a multi-column regression table to Word / Excel / LaTeX (Stata outreg2 / esttab / R modelsummary equivalent), OR bundle an entire replication appendix into one .docx / .xlsx / .tex file. Also triggers on keywords "StatsPAI", "statspai", "AER empirical analysis", "applied micro pipeline", "Table 1 balance", "event study", "first-stage F", "Oster bound", "honest_did", "spec_curve", "callaway_santanna", "dragonnet", "text as treatment", "outreg2 in Python", "regression table to Word/Excel", "sp.regtable", "sp.collect", "sp.paper_tables", "sp.feols", "summary_col", "modelsummary", "AER style table", "QJE style table".
+description: Use when the user asks to run a full empirical / causal analysis in Python — by default in the style of an applied economics paper (AER / QJE / JPE / ReStud / AEJ) with DID / RD / IV / SCM / DML / matching, written-out estimating equation + identifying assumption, Table 1 / Table 2 / event-study figure / robustness gauntlet — OR in epidemiology / public health style (target-trial emulation, IPTW + g-formula + TMLE triplet, Mendelian randomization, KM/AFT survival, E-value sensitivity, STROBE/TRIPOD reporting) — OR in ML causal inference style (DML, S/T/X/R/DR meta-learners, causal forest, Dragonnet/TARNet/CEVAE, BCF, CATE distribution, policy learning, conformal causal, fairness audit, causal discovery). Also covers exporting multi-column regression tables to Word / Excel / LaTeX (Stata outreg2 / esttab / R modelsummary equivalent) and bundling an entire replication appendix into one .docx / .xlsx / .tex file. Triggers on keywords "StatsPAI", "statspai", "AER empirical analysis", "applied micro pipeline", "Table 1 balance", "event study", "first-stage F", "Oster bound", "honest_did", "spec_curve", "callaway_santanna", "dragonnet", "text as treatment", "outreg2 in Python", "regression table to Word/Excel", "sp.regtable", "sp.collect", "sp.paper_tables", "sp.feols", "summary_col", "modelsummary", "AER style table", "QJE style table", "epidemiology pipeline", "target trial emulation", "g-formula", "IPTW", "TMLE", "Mendelian randomization", "STROBE", "TRIPOD", "公共健康", "流行病学", "DML", "double machine learning", "causal forest", "meta-learner", "CATE", "conformal causal", "policy learning", "因果机器学习", "ML causal".
 triggers:
   - causal inference in python
   - applied microeconomics pipeline
@@ -45,6 +45,33 @@ triggers:
   - mixed magnitude coefficients
   - sumstats by_labels
   - Control Treated auto labels
+  - epidemiology pipeline
+  - public health causal inference
+  - target trial emulation
+  - g-formula
+  - IPTW marginal structural model
+  - TMLE doubly robust
+  - HAL-TMLE
+  - Mendelian randomization
+  - MR-Egger weighted median
+  - STROBE TRIPOD reporting
+  - E-value sensitivity
+  - Kaplan-Meier AFT survival
+  - 流行病学
+  - 公共健康
+  - ML causal inference
+  - double machine learning DML
+  - meta-learner S T X R DR
+  - causal forest GRF
+  - Dragonnet TARNet CEVAE
+  - Bayesian causal forest BCF
+  - CATE distribution
+  - policy tree
[…]
