+++ StatsPAI_full_data_analysis_skill/README.md
@@ -2,17 +2,36 @@
 
 This folder is a **Claude Code Skill** that teaches Claude (or any
 compatible agent harness) how to drive [StatsPAI](https://github.com/brycewang-stanford/StatsPAI)
-end-to-end through the **canonical pipeline of an applied AER /
-QJE / AEJ empirical paper** — pre-analysis plan, Table 1 summary
-statistics, written-out estimating equation and identifying
-assumption, identification graphics (event-study / first-stage F /
-McCrary / SCM trajectory / love plot), the AER multi-regression
-"Table 2" gauntlet (progressive controls / design horse-race /
-multi-outcome / IV first-stage triplet), heterogeneity (subgroup
-table + CATE figure), mechanisms, and a full robustness battery
-(placebo / Oster / honest_did / E-value / Conley / two-way SE /
-spec curve / sensitivity dashboard) — ending in a single editable
-Word / Excel / LaTeX replication bundle.
+end-to-end through a full causal-inference / empirical analysis,
+covering **three domain modes** that share the same export stack
+and `CausalResult` interface:
+
+- **Default — Applied Econ (AER / QJE / AEJ).** The canonical 8-section
+  empirical paper: pre-analysis plan, Table 1 summary statistics,
+  written-out estimating equation + identifying assumption,
+  identification graphics (event-study / first-stage F / McCrary /
+  SCM trajectory / love plot), the AER multi-regression Table 2
+  gauntlet (progressive controls / design horse-race / multi-outcome
+  / IV first-stage triplet), heterogeneity (subgroup table + CATE
+  figure), mechanisms, and a full robustness battery (placebo /
+  Oster / honest_did / E-value / Conley / two-way SE / spec curve /
+  sensitivity dashboard) — ending in a single editable Word / Excel
+  / LaTeX replication bundle.
+- **Mode A — Epidemiology / public health (§A).** Target-trial
+  emulation, IPTW + g-formula + TMLE doubly-robust triplet,
+  Mendelian randomization (IVW / Egger / weighted median), KM / AFT
+  survival, E-value sensitivity, principal stratification — under
+  STROBE / TRIPOD reporting conventions.
+- **Mode B — ML causal inference (§B).** DML, S/T/X/R/DR
+  meta-learners, causal forest (GRF), Dragonnet / TARNet / CEVAE
+  neural causal, BCF, matrix completion, CATE distribution + policy
+  tree + off-policy evaluation, conformal causal prediction,
+  fairness audit, and DAG learning (PC / NOTEARS / LLM-assisted).
+
+All three modes reuse the same Word / Excel / LaTeX export stack
+and the estimand-first `sp.causal_question(...).identify()` DSL —
+switching modes only changes which Step-4 estimators you reach
+for, not the surrounding scaffolding.
 
 ## What this skill ships (paper-ready, in three formats)
 
@@ -71,19 +90,40 @@ pip install statspai          # >= 1.6.6 (Word/Excel export stack)
 
 ## Activate
 
[…]
