### 2026-04-27 8dc99ff
docs(skill): add CJK font + retina DPI notebook setup section
Adds a self-contained "Notebook setup" section between the Export
cookbook and Step −1, providing a drop-in setup_plot() that fixes
two recurring issues for matplotlib-backed sp.* plot output:
1. CJK labels rendering as ▢▢▢ tofu under default DejaVu Sans
2. Plots looking fuzzy on hi-DPI screens (default figure.dpi=100)

Includes cross-platform CJK font fallback (macOS/Windows/Linux),
axes.unicode_minus=False, figure.dpi=144 + savefig.dpi=300, and
auto-enables Jupyter retina backend when running in IPython.
Adds smoke test, troublesh
[…]
