### 2026-04-27 b4ba4a3
feat(fast): native HDFE stack — Phase 0..8 + post-ship audit + 5 follow-ups
New ``sp.fast`` namespace, additive (no public API change in existing
paths). All existing tests still pass; 231 new fast-path tests added.

Phase 1 — K-way alternating-projection demean kernel in Rust with
  Irons-Tuck (vector Aitken) acceleration, iterative singleton drop.
  Public: ``sp.fast.demean(X, fe, accel=, backend="auto"|"rust"|
  "numpy"|"jax")``. Numeric parity vs R ``fixest::demean`` to 1.12e-14.

Phase 2 — Native ``sp.fast.fepois`` PPML-HDFE estimator (Correia-
  Guimarães-Zylkin 2020), independent of pyf
[…]
