+++ src/statspai/fast/event_study.py
@@ -1,0 +1,214 @@
+"""``sp.fast.event_study`` — fast event-study (TWFE) on the Phase 1+ HDFE stack.
+
+Phase 6 deliverable. This is a homogeneous-effects event-study estimator
+(post-treatment leads/lags interacted with treatment), built on top of:
+
+* ``sp.fast.within``    — the cached HDFE residualizer.
+* ``sp.fast.i``         — event-time dummies with an explicit reference.
+* ``sp.fast.crve``      — cluster-robust standard errors.
+
+It is **not** a replacement for ``sp.callaway_santanna`` /
+``sp.sun_abraham`` / ``sp.borusyak_jaravel_spiess`` — those handle
+heterogeneous treatment effects with the appropriate group-specific
+weights, and they are unchanged. ``sp.fast.event_study`` ships the
+TWFE-style estimator on the new Rust path, suitable for designs with
+homogeneous effects and as a baseline against which the heterogeneous
+estimators can be diff-checked.
+
+Design assumptions
+------------------
+
+* Two-way fixed effects: unit + time.
+* Single binary treatment turning on at unit-specific event time.
+* Symmetric event window around the event time; missing event-times
+  outside the window are pooled into never-event indicators.
+* Reference period defaults to ``t = -1`` (one period before treatment)
+  per the canonical event-study convention (Borusyak-Jaravel-Spiess
+  2024 §3.1).
+
+References
+----------
+de Chaisemartin, C., D'Haultfœuille, X. (2020). Two-way fixed effects
+estimators with heterogeneous treatment effects. AER 110(9): 2964–2996.
+"""
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import List, Optional, Tuple
+
+import numpy as np
+import pandas as pd
+
+from .within import within as _within
+from .dsl import i as _i
+from .inference import crve as _crve
+
+
+@dataclass
+class EventStudyResult:
+    """Outcome of :func:`event_study`."""
+
+    formula: str
+    event_times: np.ndarray
+    coefs: np.ndarray
+    ses: np.ndarray
+    n_obs: int
+    n_kept: int
+    n_clusters: Optional[int]
+    cluster_var: Optional[str]
+    reference_event_time: int
+
+    def tidy(self) -> pd.DataFrame:
+        return pd.DataFrame({
+            "event_time": self.event_times,
+            "Estimate": self.coefs,
+            "Std. Error": self.ses,
+            "ci_lower": self.coefs - 1.96 * self.ses,
+            "ci_upper": self.coefs + 1.96 * self.ses,
+        })
+
+    def plot(self, ax=None):  # pragma: no cover  - cosmetic
+        try:
+            import matplotlib.pyplot as plt
+        except ImportError as exc:
+            raise RuntimeError("matplotlib is required for .plot()") from exc
+        if ax is None:
+            _, ax = plt.subplots(figsize=(8, 4))
+        td = self.tidy()
+        ax.errorbar(
+            td["event_time"], td["Estimate"],
+            yerr=1.96 * td["Std. Error"], fmt="o-", capsize=3,
+        )
+        ax.axhline(0, color="black", linewidth=0.5)
+        ax.axvline(-0.5, color="grey", linestyle="--", linewidth=0.5)
+        ax.set_xlabel("Event time")
+        ax.set_ylabel("Estimated effect")
+        ax.set_title("Event study (TWFE)")
+        return ax
+
+    def summary(self) -> str:
+        return (
+            f"sp.fast.event_study  |  N={self.n_obs:,}, kept={self.n_kept:,}, "
+            f"event_times={list(self.event_times)}\n" +
[…]

[…]
