### 2026-04-27 613a26c
docs(superpowers): add Native Rust IRLS for sp.fast.fepois design spec
Two-phase plan to close the 4× wall-clock gap to fixest::fepois on the
medium HDFE benchmark (1M rows, fe1=100k, fe2=1k):

- Phase A (~1w, v1.8.0): native Rust weighted demean kernel
  (demean_2d_weighted) wired into _weighted_ap_demean. Target wall ≤
  1.5 s on medium (merge blocker).
- Phase B (~2w, v1.8.1): native Rust IRLS outer loop (irls.rs +
  fepois_irls). vcov stays Python. Target wall ≤ 0.95 s = ≤1.5×
  fixest (merge blocker).

Spec covers FFI shapes, ownership, error handling, parity gates
(coef ≤ 1e-13 vs pyfixe
[…]
