+++ docs/superpowers/specs/2026-04-27-native-rust-irls-fepois-design.md
@@ -1,0 +1,455 @@
+# Native Rust IRLS for `sp.fast.fepois` — Design Spec
+
+> **Status**: design (pre-implementation). Authored 2026-04-27. Targets
+> v1.8.0 (Phase A) and v1.8.1 (Phase B). Implementation gated on user
+> approval of this spec and the follow-up `writing-plans` output.
+
+## 1. Goal
+
+Close the wall-clock gap between `sp.fast.fepois` (StatsPAI's native
+PPML-HDFE estimator, shipped in v1.8 RC under `feat(fast): native HDFE
+stack`, commit `b4ba4a3`) and R `fixest::fepois` from the current ~4×
+to ≤ 1.5× on the project's medium benchmark dataset (n=1M, fe1=100k,
+fe2=1k), without breaking any existing public API or numeric guarantee.
+
+## 2. Non-goals
+
+The following are explicitly out of scope for this work and remain on
+the existing tracked list in `benchmarks/hdfe/SUMMARY.md`:
+
+- GLM families other than Poisson (logit / negbin / gamma — `feglm`).
+- Anderson(m) acceleration in the demean inner loop.
+- GPU / CUDA benchmarks (no hardware on the dev box).
+- Cluster-robust SE inside `sp.fast.fepois` beyond the current CR1
+  (CR2 / CR3 / wild bootstrap stay on the Phase 4 inference primitives
+  surface; wiring `vcov="cr2"|"cr3"` into `fepois()` is a separate PR).
+- IM (Imbens-Kolesar) Satterthwaite DOF correction.
+- Word/DOCX export from `sp.fast.etable`.
+- Wiring `sp.callaway_santanna` / `sp.sun_abraham` through `backend="fast"`.
+
+## 3. Background — current state
+
+### 3.1 Wall-clock measurement
+
+Source: `benchmarks/hdfe/SUMMARY.md` and `PHASE2_VERIFY.md`. Medium
+dataset, n=1M, fe1=100k, fe2=1k.
+
+| Backend                          | Wall    | Iters | vs fixest |
+|----------------------------------|--------:|------:|----------:|
+| R `fixest::fepois`               | 0.64 s  | 5     | 1.0×      |
+| `sp.fast.fepois` (Phase 0 ship)  | 2.61 s  | 6     | **4.08×** |
+| `pyfixest.fepois`                | 4.16 s  | —     | 6.5×      |
+
+### 3.2 Where the gap lives
+
+The IRLS outer loop in `src/statspai/fast/fepois.py` (lines 548–601) is
+in pure Python. Each of the 6 outer iters does:
+
+1. `z = eta + (y - mu) / mu`, `w = mu * obs_weights` — O(n) NumPy.
+2. **Weighted alternating-projection demean** of `z` (1 column) and
+   `X` (k columns) via `_weighted_ap_demean` (line 204). This path
+   uses Python `np.bincount(weights=...)` for the weighted group sum;
+   the Phase 1 Rust kernel `demean_2d` is **unweighted only** and
+   never gets called from inside IRLS. **This step is ~80% of wall
+   time on the medium dataset.**
+3. WLS solve: `(X̃ᵀ W X̃) β = X̃ᵀ W z̃`, k×k SPD system. Trivial.
+4. `eta_new`, `mu_new`, deviance, step-halving. O(n) NumPy.
+
+Closing the 4× gap therefore boils down to: get the inner weighted
+demean off Python, and (further) get the entire IRLS off Python so
+there is no Python ↔ Rust FFI boundary inside the hot loop.
+
+## 4. Decomposition
+
+Two independently-shippable phases:
+
+| Phase | Length | Rust delta | Python delta | Target wall (medium) | Target version |
+|-------|-------:|------------|--------------|---------------------:|----------------|
[…]
