### 2026-04-27 4f60838
feat(output): add lineage + replication_pack + paper.qmd export trinity
Closes three concrete gaps between StatsPAI's export stack and the
R / Posit publication tooling. No numerical changes to any estimator.

- sp.replication_pack(): one-liner that bundles a PaperDraft / fitted
  result into a journal-ready zip with MANIFEST.json (versions, git
  SHA, per-file SHA-256), data CSV + schema, code, pip freeze,
  rendered paper, citations, and lineage.json. Tolerant: failed
  sub-steps log into manifest warnings, archive is always valid.

- sp.Provenance + sp.attach_provenance(): traceability data
[…]
