+++ src/statspai/workflow/paper.py
@@ -33,12 +33,15 @@ Notes on design
 """
 from __future__ import annotations
 
+import datetime as _dt
 import re
 from dataclasses import dataclass, field
 from typing import Any, Dict, List, Optional
 
 import numpy as np
 import pandas as pd
+
+from ..output._lineage import format_provenance, get_provenance
 
 
 __all__ = ["paper", "PaperDraft", "parse_question"]
@@ -263,15 +266,151 @@ class PaperDraft:
                     doc.add_paragraph(line)
         doc.save(path)
 
+    def to_qmd(
+        self,
+        *,
+        title: str = "Causal Analysis Draft",
+        author: Optional[str] = None,
+        formats: Optional[List[str]] = None,
+        bibliography: Optional[str] = None,
+        csl: Optional[str] = None,
+        include_provenance: bool = True,
+    ) -> str:
+        """Render to a Quarto (``.qmd``) document.
+
+        Quarto is the publication-grade default: a single source compiles
+        to PDF / HTML / DOCX / Beamer with cross-refs, citations (CSL),
+        and embedded code chunks. ``sp.paper()`` already produces all
+        the prose; this method just wraps it in the correct YAML
+        frontmatter so ``quarto render paper.qmd`` Just Works.
+
+        Parameters
+        ----------
+        title : str
+            ``title:`` field in the YAML frontmatter.
+        author : str, optional
+            ``author:``. When omitted, no author line is emitted (Quarto
+            handles that fine).
+        formats : list of str, optional
+            Output formats Quarto should support. Default
+            ``["pdf", "html", "docx"]`` covers the common journal
+            workflows. Pass e.g. ``["pdf", "beamer"]`` for slide decks.
+        bibliography : str, optional
+            Path Quarto should resolve for citation lookup, e.g.
+            ``"paper.bib"``. When omitted, the YAML omits the field
+            entirely (so Quarto won't error if no .bib file exists);
+            but if ``self.citations`` is non-empty we default to
+            ``"paper.bib"`` because :func:`sp.replication_pack` writes
+            citations there alongside the rendered draft.
+        csl : str, optional
+            CSL style file (e.g. ``"american-economic-association.csl"``).
+            Pure pass-through.
+        include_provenance : bool, default True
+            Append a Reproducibility appendix with
+            :func:`format_provenance` when ``self.workflow.result`` carries
+            a ``_provenance`` record.
+
+        Returns
+        -------
+        str
+            The complete ``.qmd`` document as a single string.
+
+        Notes
+        -----
+        - The body sections are the same as :meth:`to_markdown` —
+          standard markdown with ``## H2`` headers, which Quarto will
+          render natively.
+        - Code chunks are *not* injected by default. When the calling
+          script wants the ``.qmd`` to re-execute the analysis on each
+          render, pass it through :func:`sp.replication_pack` which
+          writes both the ``.qmd`` and a ``code/script.py`` reproducer.
+        """
+        formats = formats or ["pdf", "html", "docx"]
+        bib_path = bibliography
+        if bib_path is None and self.citations:
+            bib_path = "paper.bib"
+
+        yaml_lines: List[str] = ["---", f"title: {_yaml_str(title)}"]
+        if author:
+            yaml_lines.append(f"author: {_yaml_str(author)}")
+        yaml_lines.append(f"date: \"{_dt.date.today().isoformat()}\"")
+        if self.question:
+            yaml_lines.append(
+                f"subtitle: {_yaml_str(self.question)}"
+            )
+        # ``format:`` block.
+        if len(formats) == 1:
[…]
