+++ src/statspai/output/_lineage.py
@@ -1,0 +1,405 @@
+"""Numerical lineage / provenance for causal-inference estimates.
+
+Every number a paper reports is the output of *some* function call on
+*some* data. Stata, R, and most Python libraries lose that link the
+moment the estimate hits a docx/xlsx/tex file: the reader has the value
+but no machine-readable trail back to the call that produced it.
+
+This module attaches a small ``Provenance`` record to estimator results
+so downstream exporters (`sp.paper`, `sp.replication_pack`,
+`sp.regtable`) can stamp each number with:
+
+- The function/method name that produced it.
+- The arguments (filtered for non-serialisable objects).
+- A short SHA-256 fingerprint of the input DataFrame.
+- A run id (uuid4) — disambiguates two identical calls in the same
+  paper.
+- StatsPAI + Python versions and a wall-clock timestamp.
+
+The record is **additive**: we attach it as ``result._provenance``
+without touching the result's public surface. Estimators that don't
+opt in still work; estimators that do gain free traceability into
+every downstream export.
+
+Design choices
+--------------
+- **Hash the DataFrame, not serialise the params verbatim.** Hashes
+  are stable, small, and safe to print. Params are filtered to
+  JSON-able types before being captured (lists/dicts/scalars/strings);
+  arrays/frames/Series get a fingerprint summary (shape + dtype +
+  hash) instead of being serialised verbatim.
+- **No external deps.** Lives in ``output/`` because it's a sibling
+  of ``_repro.py`` (the version/data-hash footer); the two modules
+  share the same fingerprint primitive so ``replication_pack`` and
+  ``regtable``'s footer agree on what "the same data" means.
+- **Lazy by default.** ``attach_provenance()`` does nothing if the
+  caller passes ``enabled=False``. Free for users who don't care.
+
+This module is the foundation that ``sp.replication_pack`` and the
+Quarto emitter both build on — every downstream artifact can read
+``result._provenance`` to traceably wire a number back to a call.
+"""
+from __future__ import annotations
+
+import datetime as _dt
+import hashlib
+import sys
+import uuid
+from dataclasses import dataclass, field, asdict
+from typing import Any, Dict, Mapping, Optional
+
+
+__all__ = [
+    "Provenance",
+    "attach_provenance",
+    "get_provenance",
+    "compute_data_hash",
+    "format_provenance",
+    "lineage_summary",
+]
+
+
+# Reuse the same hashing semantics as _repro.py so a paper-tables footer
+# and a provenance record always agree on "same data".
+MAX_HASH_ROWS = 1_000_000
+
+
+def _statspai_version() -> str:
+    try:
+        from .. import __version__
+        return str(__version__)
+    except Exception:  # pragma: no cover — defensive
+        return "unknown"
+
+
+def _python_version() -> str:
+    v = sys.version_info
+    return f"{v.major}.{v.minor}.{v.micro}"
+
+
+def compute_data_hash(data: Any, length: int = 12) -> Optional[str]:
+    """Return a short SHA-256 fingerprint of *data*, or ``None``.
+
+    Accepts:
+    - ``pandas.DataFrame`` — order- and column-name-sensitive hash.
+    - ``pandas.Series`` — hashed via ``hash_pandas_object``.
+    - ``numpy.ndarray`` — bytes-hashed (shape-included).
+    - bytes / bytearray — direct SHA-256.
+
+    Anything else returns ``None`` rather than raising — provenance
+    must never break the calling estimator.
+    """
+    if data is None:
+        return None
+    try:
+        import pandas as pd
+        import numpy as np
+
+        if isinstance(data, pd.DataFrame):
+            if len(data) > MAX_HASH_ROWS:
+                return None
+            # ``hash_pandas_object`` is row-value sensitive but does NOT
[…]
