### 2026-04-27 ccd94fb
docs(superpowers): add Phase A implementation plan (Rust weighted demean)
13 bite-sized TDD tasks covering the Phase A surface from the design
spec: 4 Rust unit tests + 5 Python parity tests, dispatcher + IRLS
rewire in fepois.py, medium-dataset benchmark gate at ≤ 1.5 s.

Each task is 2-5 minutes, with explicit failing-test → impl → green
→ commit cycle. Self-review at the bottom confirms full spec coverage
and signature consistency across tasks.

Phase B plan deferred until Phase A merges.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
