+++ docs/superpowers/plans/2026-04-27-phase-a-rust-weighted-demean.md
@@ -1,0 +1,1380 @@
+# Phase A — Rust Weighted Demean for `sp.fast.fepois` Implementation Plan
+
+> **For agentic workers:** REQUIRED SUB-SKILL: Use superpowers:subagent-driven-development (recommended) or superpowers:executing-plans to implement this plan task-by-task. Steps use checkbox (`- [ ]`) syntax for tracking.
+
+**Goal:** Replace the Python `_weighted_ap_demean` inside `sp.fast.fepois`'s IRLS loop with a native Rust kernel call, closing ~50% of the wall-clock gap to `fixest::fepois` on the medium HDFE benchmark, while preserving every existing public API and numeric guarantee.
+
+**Architecture:** Add a new Rust function family (`weighted_group_sweep` → `weighted_demean_column_inplace` → `weighted_demean_matrix_fortran_inplace`) that mirrors the existing unweighted `demean_*` chain bit-for-bit but uses `Σ w_i x_i / Σ w_i` for the per-group sweep. Expose it via PyO3 as `demean_2d_weighted`. Python `fepois.py` gets a thin dispatcher: calls Rust when available, falls back to the existing pure-Python `_weighted_ap_demean_numpy` otherwise. The `wsum_list` (one `Σ w_i` array per FE dim) is precomputed in Python via `np.bincount(weights=...)` per IRLS iter — Python's small upstream overhead is negligible vs the Rust sweep dominating the inner loop.
+
+**Tech Stack:** Rust 1.76+ · PyO3 0.21 (abi3-py39) · Rayon 1.10 · numpy crate 0.21 · maturin · Python 3.9-3.13 · NumPy · pytest.
+
+**Spec:** [`docs/superpowers/specs/2026-04-27-native-rust-irls-fepois-design.md`](../specs/2026-04-27-native-rust-irls-fepois-design.md) §5 covers Phase A. This plan implements §5.1–5.8 in order.
+
+---
+
+## File Structure
+
+| File | Action | Responsibility |
+|------|--------|----------------|
+| `rust/statspai_hdfe/src/demean.rs` | modify | + `weighted_group_sweep` (private), `weighted_sweep_all_fe` (private), `weighted_demean_column_inplace` (pub), `weighted_demean_matrix_fortran_inplace` (pub), 4 unit tests |
+| `rust/statspai_hdfe/src/lib.rs` | modify | + `#[pyfunction] demean_2d_weighted` + register in `#[pymodule]`; bump `__version__` to `"0.3.0"` |
+| `rust/statspai_hdfe/Cargo.toml` | modify | bump crate version `0.1.0-alpha.1 → 0.2.0-alpha.1` (no new deps) |
+| `src/statspai/fast/fepois.py` | modify | rename `_weighted_ap_demean → _weighted_ap_demean_numpy`; add `_weighted_ap_demean_dispatcher`; rewire 2 IRLS call sites |
+| `tests/test_fast_fepois.py` | modify | + 5 new tests (Rust kernel parity, fepois coef/SE parity, weights, fallback) |
+| `benchmarks/hdfe/run_fepois_phase_a.py` | create | medium-dataset wall-clock harness, JSON output, asserts ≤ 1.5 s |
+| `benchmarks/hdfe/SUMMARY.md` | modify | update headline numbers + close Phase A row |
+| `benchmarks/hdfe/AUDIT.md` | modify | append Phase A round (if any anomaly) |
+| `CHANGELOG.md` | modify | + `## [1.8.0]` Performance entry |
+| `pyproject.toml` | modify | version `1.6.0 → 1.8.0` |
+| `src/statspai/__init__.py` | modify | `__version__ = "1.8.0"` |
+
[…]
