### 2026-04-27 cfa9afa
feat(rust-hdfe): add weighted_group_sweep + weighted_demean_column_inplace
Crate-internal weighted variants of group_sweep and
demean_column_inplace. Same Aitken / safeguard / convergence layout
as the unweighted path; differs only in the inner sweep math
(Σ w_i x_i / Σ w_i instead of Σ x_i / count).

Phase B's Rust IRLS will call these directly without going through
PyO3. Phase A's PyO3 binding lands in the next commit.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
