+++ rust/statspai_hdfe/src/demean.rs
@@ -75,6 +75,56 @@ fn group_sweep(x: &mut [f64], codes: &[i64], counts: &[f64], scratch: &mut [f64]
     }
 }
 
+/// In-place weighted group de-mean: x[i] -= Σ_g(w_i * x_i) / Σ_g(w_i).
+///
+/// ``wsum[g]`` must be the precomputed weighted group sum
+/// (``Σ_{i ∈ g} weights[i]``). ``scratch`` must have length
+/// ``wsum.len()`` and is zero-filled on entry. Caller owns the
+/// allocation; this function never reallocates on the hot path.
+#[inline]
+fn weighted_group_sweep(
+    x: &mut [f64],
+    codes: &[i64],
+    weights: &[f64],
+    wsum: &[f64],
+    scratch: &mut [f64],
+) {
+    debug_assert_eq!(scratch.len(), wsum.len());
+    debug_assert_eq!(codes.len(), x.len());
+    debug_assert_eq!(weights.len(), x.len());
+
+    for s in scratch.iter_mut() {
+        *s = 0.0;
+    }
+    for i in 0..x.len() {
+        let g = codes[i] as usize;
+        scratch[g] += weights[i] * x[i];
+    }
+    for g in 0..scratch.len() {
+        let w = wsum[g];
+        if w > 0.0 {
+            scratch[g] /= w;
+        }
+    }
+    for i in 0..x.len() {
+        let g = codes[i] as usize;
+        x[i] -= scratch[g];
+    }
+}
+
+/// Sweep all K FE dimensions once, weighted variant.
+fn weighted_sweep_all_fe(
+    x: &mut [f64],
+    fe_codes: &[&[i64]],
+    weights: &[f64],
+    wsum: &[&[f64]],
+    scratch: &mut [Vec<f64>],
+) {
+    for k in 0..fe_codes.len() {
+        weighted_group_sweep(x, fe_codes[k], weights, wsum[k], &mut scratch[k]);
+    }
+}
+
 /// Sweep through all K FEs once.
 fn sweep_all_fe(
     x: &mut [f64],
@@ -165,6 +215,105 @@ pub fn demean_column_inplace(
     for it in 0..max_iter {
         before.copy_from_slice(x);
         sweep_all_fe(x, fe_codes, counts, scratch);
+
+        let mut local_max = 0.0_f64;
+        for i in 0..x.len() {
+            let d = (x[i] - before[i]).abs();
+            if d > local_max {
+                local_max = d;
+            }
+        }
+        max_dx = local_max;
+        iters = it + 1;
+
+        if max_dx <= stop {
+            converged = true;
+            break;
+        }
+
+        if accelerate {
+            hist.push(x.to_vec());
+            if hist.len() >= 3 && (it + 1) % accel_period == 0 {
+                let n = hist.len();
+                let acc = aitken_step(&hist[n - 3], &hist[n - 2], &hist[n - 1]);
+
+                let mut max_abs_acc = 0.0_f64;
+                for &v in &acc {
+                    let av = v.abs();
+                    if av > max_abs_acc {
+                        max_abs_acc = av;
+                    }
+                }
+                if max_abs_acc < 10.0 * base_scale {
+                    x.copy_from_slice(&acc);
+                }
+                hist.clear();
+            }
+        }
+    }
+
+    DemeanInfo {
+        iters,
+        converged,
+        max_dx,
+    }
+}
+
+/// Weighted demean a single column in place. ``scratch`` is a per-FE
+/// workspace owned by the caller; pre-allocating lets us reuse across
[…]
