### 2026-04-27 ff6dbc5
feat(output): Phase 2 — gt adapter + CSL pipeline + paper auto-provenance
Layered on top of the v1.7.2 export trinity. Three additions, all
opt-in, no numerical changes to any estimator. Stays orthogonal to
the parallel sp.fast / Rust HDFE work.

great_tables adapter (sp.gt):
  - Soft dep — sp.is_great_tables_available() reports installation;
    sp.gt(...) raises a friendly ImportError when missing.
  - Dispatches: RegtableResult / PaperTables / MeanComparisonResult /
    DataFrame / any object with to_dataframe().
  - Journal preset → gt theme (font + footnote marks). All 8 presets
    (AER/
[…]
