+++ tests/test_bibliography.py
@@ -1,0 +1,281 @@
+"""Tests for ``sp.output._bibliography`` — CSL hub + paper.bib writer.
+
+Coverage
+--------
+- ``CSL_REGISTRY`` has all the journals StatsPAI's table presets
+  cover, plus chicago / apa for general use.
+- ``csl_url`` / ``csl_filename`` round-trip for every preset.
+- Unknown CSL name raises informative ``ValueError``.
+- Citation string parser handles the canonical AER / Callaway-style
+  forms, falls back gracefully on garbage.
+- ``make_bib_key`` is stable + collision-safe across realistic inputs.
+- ``write_bib`` produces a syntactically valid .bib (round-trip via
+  bibtexparser if available; structural checks otherwise).
+- Replication pack now writes a real bib through the pipeline.
+- ``to_qmd(csl='aer')`` resolves to the canonical filename.
+"""
+from __future__ import annotations
+
+import re
+from pathlib import Path
+
+import pytest
+
+from statspai.output._bibliography import (
+    CSL_REGISTRY,
+    citations_to_bib_entries,
+    csl_filename,
+    csl_url,
+    list_csl_styles,
+    make_bib_key,
+    parse_citation_to_bib,
+    write_bib,
+)
+from statspai.workflow.paper import PaperDraft
+
+
+# ---------------------------------------------------------------------------
+# CSL registry
+# ---------------------------------------------------------------------------
+
+class TestCslRegistry:
+    def test_aer_present(self):
+        assert "aer" in CSL_REGISTRY
+        assert CSL_REGISTRY["aer"]["filename"].endswith(".csl")
+        assert CSL_REGISTRY["aer"]["url"].startswith("https://")
+
+    def test_all_table_journals_covered(self):
+        # The journal-presets module covers these — CSL must too.
+        for j in ("aer", "qje", "econometrica", "restat",
+                  "jf", "aeja", "jpe", "restud"):
+            assert j in CSL_REGISTRY, f"missing CSL for {j!r}"
+
+    def test_aej_family_alias_to_aer_style(self):
+        # AEJ family historically uses the AER style.
+        for j in ("aeja", "aejmac", "aejmicro", "aejpol"):
+            assert (
+                CSL_REGISTRY[j]["filename"]
+                == "american-economic-association.csl"
+            )
+
+    def test_csl_url_lookup(self):
+        url = csl_url("aer")
+        assert "american-economic-association" in url
+
+    def test_csl_url_case_insensitive(self):
+        assert csl_url("AER") == csl_url("aer")
+
+    def test_unknown_csl_raises(self):
+        with pytest.raises(ValueError, match="Unknown CSL style"):
+            csl_url("nonsense")
+
+    def test_csl_filename_passthrough_for_real_filename(self):
+        # Caller already has a real filename — pass through.
+        assert csl_filename("custom-style.csl") == "custom-style.csl"
+
+    def test_list_csl_styles(self):
+        styles = list_csl_styles()
+        assert isinstance(styles, list) and len(styles) > 5
+        names = {s[0] for s in styles}
+        assert {"aer", "qje", "econometrica"} <= names
+
+
+# ---------------------------------------------------------------------------
+# Citation parser + bib key
+# ---------------------------------------------------------------------------
+
+class TestParseCitation:
+    def test_canonical_aer_format(self):
+        s = ("Callaway B, Sant'Anna PHC. (2021). "
+             "Difference-in-Differences with multiple time periods. "
+             "Journal of Econometrics.")
+        entry = parse_citation_to_bib(s)
+        assert entry["type"] == "article"
+        assert entry["fields"]["year"] == "2021"
+        assert "Callaway" in entry["fields"]["author"]
+        assert "Difference" in entry["fields"]["title"]
[…]
