+++ src/statspai/output/_replication_pack.py
@@ -55,6 +55,7 @@ import zipfile
 from pathlib import Path
 from typing import Any, Dict, Iterable, List, Mapping, Optional, Union
 
+from ._bibliography import citations_to_bib_entries
 from ._lineage import (
     Provenance,
     compute_data_hash,
@@ -563,10 +564,29 @@ def replication_pack(
 
     # ------- bibliography --------------------------------------------
     if bib and citations:
-        _put(
-            "paper/paper.bib",
-            "\n\n".join(citations).rstrip() + "\n",
-        )
+        try:
+            entries = citations_to_bib_entries(citations)
+            from ._bibliography import _format_bib_entry  # local helper
+            try:
+                from .. import __version__ as _v
+            except Exception:
+                _v = "unknown"
+            bib_text = (
+                f"% paper.bib — auto-generated by StatsPAI v{_v}\n"
+                f"% {len(entries)} entries\n\n"
+                + "\n".join(_format_bib_entry(e) for e in entries)
+            )
+            _put("paper/paper.bib", bib_text)
+        except Exception as exc:  # pragma: no cover — defensive
+            warnings.append(
+                f"bibliography emission failed: {type(exc).__name__}: {exc}; "
+                "fell back to raw citation dump."
+            )
+            _put(
+                "paper/paper.bib",
+                "% raw citations (parser failed)\n\n"
+                + "\n\n".join(f"% {c}" for c in citations) + "\n",
+            )
 
     # ------- lineage --------------------------------------------------
     if results:
