+++ src/statspai/output/__init__.py
@@ -21,6 +21,17 @@ from ._lineage import (
     lineage_summary,
 )
 from ._replication_pack import ReplicationPack, replication_pack
+from ._gt import to_gt, is_great_tables_available
+from ._bibliography import (
+    CSL_REGISTRY,
+    csl_url,
+    csl_filename,
+    list_csl_styles,
+    parse_citation_to_bib,
+    make_bib_key,
+    citations_to_bib_entries,
+    write_bib,
+)
 
 __all__ = [
     "OutReg2",
@@ -58,4 +69,16 @@ __all__ = [
     # Replication pack
     "ReplicationPack",
     "replication_pack",
+    # great_tables adapter
+    "to_gt",
+    "is_great_tables_available",
+    # Bibliography / CSL
+    "CSL_REGISTRY",
+    "csl_url",
+    "csl_filename",
+    "list_csl_styles",
+    "parse_citation_to_bib",
+    "make_bib_key",
+    "citations_to_bib_entries",
+    "write_bib",
 ]

[…]
