### 2026-04-27 a60c62c
fix(rust-hdfe): tighten weighted_sweep_all_fe length asserts (review)
Address Important findings from the Tasks 1+2 code-quality review:

1. Add debug_assert_eq!(fe_codes.len(), wsum.len()) and
   debug_assert_eq!(fe_codes.len(), scratch.len()) to
   weighted_sweep_all_fe so the K-axis length contract is documented
   at the same layer where weighted_group_sweep already asserts its
   length invariants.

2. Plan errata: Task 1's pasted test block had `&[&weights[..]]`
   (a slice-of-slices wrapper that doesn't match the function's
   `weights: &[f64]` signature). Corrected to `&weights[..]` an
[…]
