### 2026-04-27 a0d6f7e
test(rust-hdfe): add 3 weighted demean unit tests
Covers (a) parity with unweighted path when weights ≡ 1,
(b) non-trivial unequal-weights 2×2, and (c) zero-weight group
(must not panic / NaN). Closes the K≥2 / Aitken / safeguard
coverage gap that the initial weighted_oneway_exact left open.

12 tests pass total in the crate (was 9).

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
