### 2026-04-27 715549f
docs(spec): clubSandwich-equivalent HTZ Wald DOF design
Spec for v1.6.4 follow-up: cluster_wald_htz +
cluster_dof_wald_htz + WaldTestResult — Pustejovsky-Tipton 2018
moment-matching DOF, independent PR with rtol<1e-8 parity vs R
clubSandwich::Wald_test(test='HTZ').

Locks: parity bar 1e-8, working covariance Phi=I, standalone
scope (no crve/feols/event_study wiring), fail-loudly on
G<=q / eta<=q-1. Test backbone is a frozen R-generated CSV+JSON
fixture (no Python<->R RNG sync). Citation pustejovsky2018small
to be added to paper.bib only after Crossref dual-source verify
(CLAUDE.md §10).

Co-Au
[…]
