+++ docs/superpowers/specs/2026-04-27-htz-clubsandwich-parity-design.md
@@ -1,0 +1,510 @@
+# clubSandwich-equivalent HTZ Wald DOF — design spec
+
+**Date**: 2026-04-27
+**Status**: Draft → awaiting user review
+**Target version**: v1.6.4 (or v1.7 minor bump if bundled with other follow-ups)
+**Scope**: independent PR, standalone (no `crve` / `feols` / `event_study` wiring)
+**Parity bar**: `rtol < 1e-8` vs R `clubSandwich::Wald_test(..., test="HTZ")`
+
+---
+
+## 1. Why this exists
+
+### 1.1 Current gap
+
+`statspai.fast.cluster_dof_wald_bm` (added v1.8.x in HDFE Phase 4) implements
+the **BM 2002 §3 simplified** matrix Satterthwaite DOF for a multi-restriction
+Wald test under CR2 cluster-robust inference:
+
+```
+ν_W = [Σ_g tr(E_g E_g^T)]² / Σ_g ||E_g E_g^T||_F²
+```
+
+That formula is one of the valid "Bell-McCaffrey-style" approximations, and
+its current docstring (`src/statspai/fast/inference.py:743`) is honest:
+
+> **Not equivalent to** `clubSandwich::Wald_test(..., test="HTZ")`. The HTZ
+> test uses a Hotelling-T² approximation with the more elaborate
+> Pustejovsky-Tipton 2018 §3.2 moment-matching DOF, which is typically much
+> smaller (more conservative) than the BM 2002 §3 simplified [...] formula
+> implemented here. On moderate panels the two can differ by **50-100%**.
+
+For users who do paper replication / cross-language reproducibility against R
+clubSandwich, that 50-100% drift is unacceptable. The entire point of `sp.fast`
+parity claims is "drop-in replacement at machine precision".
+
+### 1.2 What this PR ships
+
+A **standalone** clubSandwich-equivalent HTZ Wald test under CR2 sandwich:
+
+| Symbol | Layer | Returns |
+|---|---|---|
+| `cluster_dof_wald_htz(X, cluster, *, R, weights, bread)` | low-level DOF helper, mirrors `cluster_dof_wald_bm` | `float` η |
+| `cluster_wald_htz(X, residuals, cluster, *, R, r, beta, weights, bread)` | full Wald test, mirrors R `Wald_test(test="HTZ")` | `WaldTestResult` |
+| `WaldTestResult` | frozen dataclass | `(test, q, eta, F_stat, p_value, Q, R, r, V_R)` |
+
+### 1.3 Non-goals (explicit YAGNI)
+
+1. ❌ Working covariance Φ ≠ I (GLS / mixed-model). Defer to v2.
+2. ❌ HTA / KZ / Naive / EDF / chi-sq variants from P-T 2018. Only HTZ.
+3. ❌ Wiring HTZ into `sp.fast.feols` / `sp.fast.crve` / `sp.fast.event_study`.
+   That is the next, separate PR.
+4. ❌ Saddlepoint approximation for the small-G regime.
+5. ❌ Wald test under sandwiches other than CR2 (no CR3, no jackknife).
+6. ❌ Performance optimisation (numba/JAX). NumPy direct is sufficient since
+   q is typically ≤ 5 and the per-cluster work is O(n_g · q) at worst.
+
+---
+
+## 2. API surface
+
+### 2.1 `WaldTestResult`
+
+```python
+from dataclasses import dataclass
+import numpy as np
+
+@dataclass(frozen=True)
+class WaldTestResult:
+    test: str            # "HTZ" (locked for v1; future: "HTA", "KZ", ...)
+    q: int               # number of restrictions
+    eta: float           # P-T 2018 §3.2 moment-matching DOF
+    F_stat: float        # Hotelling-T² scaled F: (η - q + 1)/(η · q) · Q
[…]
