### 2026-04-27 cc957fa
test(rust-hdfe): clarify weighted_unequal_weights doc + split assert (review)
Address Important + Minor findings from the Task 3 code-quality review:

- Test 2's doc comment claimed "known closed-form" but the body
  only checked finite + abs<100. Updated the comment to accurately
  describe the test as a stability/convergence sanity check, with a
  note that promoting to closed-form K=2 is a tracked hardening.

- Test 1's compound `info_w.converged && info_u.converged` assert
  with no message would have hidden which side failed if it ever
  fired. Split into two messaged asserts matching the
[…]
