### 2026-04-27 3d08270
feat(rust-hdfe): add weighted_demean_matrix_fortran_inplace (parallel)
Rayon-parallel column dispatcher for the weighted variant. Each
column gets its own scratch + Aitken history; FE codes / weights /
wsum are shared read-only across threads. Mirrors the existing
demean_matrix_fortran_inplace unweighted path; only the inner
per-column kernel differs.

This is the function the PyO3 binding in the next commit will call.
Coverage comes via Python parity tests in Task 7+.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
