### 2026-04-27 9f3bc94
docs(rust-hdfe): expand weighted_demean_matrix_fortran_inplace doc (review)
Mirror the unweighted reference's memory-layout note and explain
the wsum_lens semantics (caller-precomputed cardinality slice that
avoids repeated .len() inside the Rayon closure). Pure documentation;
no code change.

Addresses Minor finding from the Task 4 code-quality review. 12/12
tests still pass.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
