### 2026-04-27 0498bc1
feat(output): Phase 3 — estimand-first paper + estimator provenance + DAG appendix
Closes 2 of 3 outstanding deferred items from Phase 2's roadmap.
LLM-DAG closed loop remains a v1.7.3 sprint (needs API key strategy
+ DAG validator audit). No numerical changes to any estimator.

Estimand-first paper builder:
  - sp.paper(causal_question_obj) dispatch + q.paper() method.
  - Routes through q.identify() + q.estimate(), assembling a paper
    whose Question / Identification / Estimator / Results sections
    match the pre-registered Target-Trial Protocol verbatim.
  - Adds paper_from_question() a
[…]
