+++ tests/test_paper_dag_section.py
@@ -1,0 +1,192 @@
+"""Tests for the Causal DAG appendix in PaperDraft.
+
+When the user passes ``dag=`` to ``sp.paper(...)`` (or
+``CausalQuestion.paper(dag=...)``), the draft gains a *Causal DAG*
+section. The section renders fmt-aware:
+
+- ``markdown`` / ``tex`` — text-art (variables list + edges + adjustment
+  sets + back-door paths + bad controls).
+- ``qmd`` — Quarto-native ``mermaid`` code block + the same text
+  fallbacks below it.
+
+Coverage
+--------
+- ``_render_dag_section()`` direct: edges, adjustment sets, latent
+  detection, fmt='qmd' emits mermaid.
+- ``sp.paper(..., dag=g)`` populates ``sections['Causal DAG']``.
+- ``CausalQuestion.paper(dag=g)`` populates ``sections['Causal DAG']``.
+- ``draft.to_qmd()`` regenerates the section with mermaid when
+  ``draft.dag`` is set, even if the markdown body is text-art.
+- DAG with no treatment/outcome match still renders variables + edges.
+"""
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+from statspai.dag.graph import DAG
+from statspai.workflow.paper import _render_dag_section
+
+
+# ---------------------------------------------------------------------------
+# _render_dag_section direct
+# ---------------------------------------------------------------------------
+
+class TestRenderDagSection:
+    def test_basic_markdown(self):
+        g = DAG("Z -> X; Z -> Y; X -> Y")
+        out = _render_dag_section(g, treatment="X", outcome="Y",
+                                   fmt="markdown")
+        # Variables + edges + adjustment sets present.
+        assert "**Variables**" in out
+        assert "`X`" in out and "`Y`" in out and "`Z`" in out
+        assert "`Z` → `X`" in out
+        assert "**Adjustment sets**" in out
+        assert "`Z`" in out
+        # No mermaid in markdown.
+        assert "```{mermaid}" not in out
+
+    def test_qmd_emits_mermaid(self):
+        g = DAG("Z -> X; X -> Y")
+        out = _render_dag_section(g, treatment="X", outcome="Y", fmt="qmd")
+        assert "```{mermaid}" in out
+        assert "graph LR" in out
+        assert "Z --> X" in out
+        assert "X --> Y" in out
+        # Closing fence for mermaid block.
+        assert out.count("```") >= 2
+
+    def test_renders_latent_confounders(self):
+        g = DAG("Z -> X; X -> Y; U <-> Y")
+        out = _render_dag_section(g, treatment="X", outcome="Y",
+                                   fmt="markdown")
+        assert "**Latent common causes**" in out
+        # Latent node syntax `_L_U_Y` is what statspai DAG generates.
+        assert "_L_U_Y" in out
+
+    def test_backdoor_paths_listed(self):
+        g = DAG("Z -> X; Z -> Y; X -> Y")
+        out = _render_dag_section(g, treatment="X", outcome="Y",
+                                   fmt="markdown")
+        assert "**Back-door paths**" in out
+        # X — Z — Y is the canonical path.
+        assert "`X` — `Z` — `Y`" in out or "`X` — `Z` — `Y`" in out.replace("·", "—")
+
+    def test_no_treatment_outcome_still_renders_basics(self):
+        g = DAG("X -> Y")
+        out = _render_dag_section(g, fmt="markdown")
+        assert "`X`" in out and "`Y`" in out
+        assert "`X` → `Y`" in out
+        # No adjustment-set section without treatment/outcome.
+        assert "Adjustment sets" not in out
+
[…]
