+++ src/statspai/regression/ols.py
@@ -513,4 +513,22 @@ def regress(
         )
 
     model = OLSRegression(formula=formula, data=data)
-    return model.fit(robust=robust, cluster=cluster, **kwargs)
+    _result = model.fit(robust=robust, cluster=cluster, **kwargs)
+    try:
+        from ..output._lineage import attach_provenance as _attach_prov
+        _attach_prov(
+            _result,
+            function="sp.regress",
+            params={
+                "formula": formula,
+                "robust": robust,
+                "cluster": cluster,
+                **{k: v for k, v in kwargs.items()
+                   if k in ("weights", "vcov", "se_type")},
+            },
+            data=data,
+            overwrite=False,
+        )
+    except Exception:  # pragma: no cover — provenance must never break fit
+        pass
+    return _result
