+++ src/statspai/did/did_2x2.py
@@ -240,7 +240,7 @@ def did_2x2(
         'weights': weights,
     }
 
-    return CausalResult(
+    _result = CausalResult(
         method='Difference-in-Differences (2x2)',
         estimand='ATT',
         estimate=att,
@@ -253,3 +253,19 @@ def did_2x2(
         model_info=model_info,
         _citation_key='did_2x2',
     )
+    try:
+        from ..output._lineage import attach_provenance as _attach_prov
+        _attach_prov(
+            _result,
+            function="sp.did.did_2x2",
+            params={
+                "y": y, "treat": treat, "time": time,
+                "covariates": covariates, "cluster": cluster,
+                "robust": robust, "alpha": alpha, "weights": weights,
+            },
+            data=data,
+            overwrite=False,
+        )
+    except Exception:  # pragma: no cover
+        pass
+    return _result

[…]
