### 2026-04-27 7d6af4b
feat(rust-hdfe): expose demean_2d_weighted via PyO3 (v0.3.0)
New #[pyfunction] mirroring demean_2d but with per-obs weights and
caller-supplied wsum (Σ w_i per FE group). Validates only shapes;
business constraints (weights ≥ 0, finite) stay on the Python side
since fepois() already enforces them.

Releases the GIL during the parallel column dispatch via
py.allow_threads(), same as the unweighted path.

Crate __version__ bump 0.2.0 → 0.3.0; Cargo.toml version
0.1.0-alpha.1 → 0.2.0-alpha.1.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
