### 2026-04-27 6a1059a
docs(plan): clubSandwich-equivalent HTZ Wald implementation plan
9 tasks / 54 TDD steps ending in 8-acceptance-criteria sweep.
Phases: A (skeleton) -> B (eta formula, frozen R fixture first) ->
C (full Wald wrapper) -> D (test sweep) -> E (registry/bib/changelog) ->
F (full pytest + smoke import).

eta formula intentionally deferred to execution-time double-sourcing
(P-T 2018 PDF + clubSandwich Wald_test.R) gated by frozen fixture at
rtol<1e-8. R + clubSandwich verified locally so live parity is a
hard gate, not skipif.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
