+++ docs/superpowers/plans/2026-04-27-htz-clubsandwich-parity.md
@@ -1,0 +1,1743 @@
+# clubSandwich-equivalent HTZ Wald DOF Implementation Plan
+
+> **For agentic workers:** REQUIRED SUB-SKILL: Use superpowers:subagent-driven-development (recommended) or superpowers:executing-plans to implement this plan task-by-task. Steps use checkbox (`- [ ]`) syntax for tracking.
+
+**Goal:** Ship `cluster_wald_htz` + `cluster_dof_wald_htz` + `WaldTestResult` in `statspai.fast.inference`, numerically equivalent to R `clubSandwich::Wald_test(test="HTZ")` to `rtol < 1e-8` per the Pustejovsky-Tipton 2018 §3.2 moment-matching DOF.
+
+**Architecture:** New code goes into `src/statspai/fast/inference.py`, immediately after the existing `cluster_dof_wald_bm` block and before `__all__`. Standalone — no wiring into `crve` / `feols` / `event_study`. Single working covariance Φ = I (OLS+CR2 path). Parity validated three ways: (1) frozen R-generated CSV+JSON fixture (CI backbone, no R needed), (2) live R `subprocess` parity tests (skipif when R missing), (3) q=1 documented-drift unit test against existing `cluster_dof_bm` to detect regressions.
+
+**Tech Stack:** NumPy + SciPy (existing). pytest. R + clubSandwich (verified installed at `/usr/local/bin/Rscript`) only as black-box reference for fixture generation and parity testing — no GPL code copied.
+
+**Spec:** [`docs/superpowers/specs/2026-04-27-htz-clubsandwich-parity-design.md`](../specs/2026-04-27-htz-clubsandwich-parity-design.md)
+
+**Pre-flight context:**
+- WIP state: `src/statspai/fast/inference.py` has uncommitted changes adding `cluster_dof_bm` + `cluster_dof_wald_bm` + `extra_df=` plumbing to `boottest`. HTZ work builds on this WIP and will land *after* it (or alongside, as a separate commit in the same push).
+- HEAD commit: `b4ba4a3` (`feat(fast): native HDFE stack`).
+- R + `clubSandwich` package both installed locally — live parity is a hard gate, not a skipif.
+
+---
+
+## Phase A — Foundation
+
+### Task 1: WaldTestResult dataclass + first failing test
+
+**Files:**
+- Modify: `src/statspai/fast/inference.py` (add dataclass after `cluster_dof_wald_bm`, before `__all__`)
+- Modify: `src/statspai/fast/__init__.py` (export)
+- Create: `tests/test_fast_htz.py`
+
+- [ ] **Step 1: Create `tests/test_fast_htz.py` with first failing test**
+
+```python
+"""Tests for clubSandwich-equivalent HTZ Wald DOF (Pustejovsky-Tipton 2018).
+
+See docs/superpowers/specs/2026-04-27-htz-clubsandwich-parity-design.md.
+"""
+from __future__ import annotations
+
+import json
+import subprocess
+import shutil
+import warnings
+from pathlib import Path
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+
+
+# ---------------------------------------------------------------------------
+# Shared panel fixture (mirrors tests/test_fast_inference.py::_ols_panel)
+# ---------------------------------------------------------------------------
+
+def _ols_panel(n_clusters=20, m=30, seed=0, beta=(0.30, -0.20), unbalanced=False):
+    rng = np.random.default_rng(seed)
[…]
