### 2026-04-27 58884e2
feat(fast): WaldTestResult dataclass + HTZ function stubs
Task 1 of HTZ implementation plan
(docs/superpowers/plans/2026-04-27-htz-clubsandwich-parity.md).

Adds frozen WaldTestResult dataclass and NotImplementedError stubs for
cluster_dof_wald_htz / cluster_wald_htz. Wires into sp.fast namespace
and __all__. Tests: 1 passing (frozen-dataclass identity).

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
