+++ src/statspai/fast/inference.py
@@ -224,6 +224,7 @@ def boottest(
     B: int = 9999,
     seed: Optional[int] = None,
     obs_weights: Optional[np.ndarray] = None,
+    extra_df: int = 0,
 ) -> BootTestResult:
     """Wild cluster bootstrap of H0: β[null_coef] = null_value (Davidson-Flachaire / MacKinnon 2019).
 
@@ -257,6 +258,13 @@ def boottest(
         RNG seed for reproducibility.
     obs_weights : ndarray, shape (n,), optional
         Observation weights for WLS. Default uniform.
+    extra_df : int, default 0
+        Additional residual-DOF charge in the CR1 small-sample factor
+        (used both for the observed t-stat and uniformly across the
+        bootstrap replicas). Pass ``Σ(G_k - 1)`` when ``X`` is
+        FE-residualised — same convention as :func:`crve`. Without it,
+        bootstrap p-values for HDFE designs are biased downward because
+        the t-pivot uses an under-corrected SE.
 
     Returns
     -------
@@ -264,6 +272,8 @@ def boottest(
     """
     if weights not in ("rademacher", "webb"):
         raise ValueError(f"weights={weights!r}; expected 'rademacher' or 'webb'")
+    if extra_df < 0:
+        raise ValueError(f"extra_df={extra_df}; must be >= 0")
 
     n, k = X.shape
     if y.shape[0] != n:
@@ -289,7 +299,10 @@ def boottest(
 
     # Observed t-statistic
     resid_full = y - X @ beta_full
-    V_full = crve(X, resid_full, cluster, weights=w, bread=bread, type="cr1")
+    V_full = crve(
+        X, resid_full, cluster, weights=w, bread=bread, type="cr1",
+        extra_df=extra_df,
+    )
     se_obs = float(np.sqrt(V_full[null_coef, null_coef]))
     t_obs = float((beta_full[null_coef] - null_value) / se_obs)
 
@@ -318,7 +331,7 @@ def boottest(
     boot_t = np.empty(B)
     weight_fn = _rademacher_weights if weights == "rademacher" else _webb6_weights
 
-    cr1_factor = (G / (G - 1.0)) * ((n - 1.0) / max(n - k, 1))
+    cr1_factor = (G / (G - 1.0)) * ((n - 1.0) / max(n - k - extra_df, 1))
     bread_row = bread[null_coef]               # (k,) row of bread for this null
 
     # Pre-bake the score basis ``X * w``: doesn't change across B, and
@@ -401,6 +414,7 @@ def boottest_wald(
     B: int = 9999,
     seed: Optional[int] = None,
     obs_weights: Optional[np.ndarray] = None,
+    extra_df: int = 0,
 ) -> BootWaldResult:
     """Wild cluster bootstrap of the joint linear hypothesis ``R β = r``.
 
@@ -424,6 +438,11 @@ def boottest_wald(
     B : int
     seed : int, optional
     obs_weights : ndarray, optional
+    extra_df : int, default 0
+        Additional residual-DOF charge in the CR1 small-sample factor
+        (used both for the observed Wald and uniformly across bootstrap
+        replicas). Pass ``Σ(G_k - 1)`` for FE-residualised ``X`` — same
+        convention as :func:`crve` and :func:`boottest`.
 
     Returns
     -------
@@ -431,6 +450,8 @@ def boottest_wald(
     """
     if weights not in ("rademacher", "webb"):
         raise ValueError(f"weights={weights!r}; expected 'rademacher' or 'webb'")
+    if extra_df < 0:
[…]
