### 2026-04-27 b860935
docs(rust-hdfe): list demean_2d_weighted in module header + Returns line (review)
- Add a "Phase A additions" bullet to the crate-level //! doc so the
  module inventory matches what the #[pymodule] block exports.
- Add a Returns sentence to demean_2d_weighted's leading doc comment,
  mirroring demean_2d's "Returns a list of dicts (one per column)".

Pure documentation; no code change. Addresses the two Minor findings
from the Task 5 code-quality review.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
