### 2026-04-27 37bf9e1
feat(fast): _htz_per_cluster_quantities internal helper
Task 2 of HTZ implementation plan. Provides:
  G_g = R · bread · X_g^T · diag(sqrt(w_g))   (q × n_g, NO A_g)
  A_g_sqrtW = diag(sqrt(w_g)) · A_g           (n_g × n_g)
  Omega = sum_g G_g G_g^T                     (q × q)

Verified: V_R = sum_g (G_g · A_g_sqrtW · u_g)(...)^T matches
crve(type='cr2') sandwich sliced to R-subspace at rtol<1e-10.
TDD caught a formula bug where A_g was double-applied — fixed before
the eta implementation lands on top.

Tests: 3/3 passing (frozen-dataclass + V_R parity + Omega PSD).

Co-Authored-By: Claude Opus
[…]
