+++ src/statspai/fast/inference.py
@@ -28,6 +28,7 @@ MacKinnon, J. G., Webb, M. D. (2018). The wild bootstrap for few
 """
 from __future__ import annotations
 
+import warnings
 from dataclasses import dataclass
 from typing import Optional, Tuple, Union
 
@@ -891,6 +892,142 @@ class WaldTestResult:
             "R": self.R.tolist(), "r": self.r.tolist(),
             "V_R": self.V_R.tolist(),
         }
+
+
+def _htz_per_cluster_quantities(
+    X: np.ndarray,
+    cluster: np.ndarray,
+    *,
+    R: np.ndarray,
+    weights: Optional[np.ndarray] = None,
+    bread: Optional[np.ndarray] = None,
+) -> dict:
+    """Internal helper: per-cluster lifted contributions for the HTZ test.
+
+    Computes the matrices that both :func:`cluster_dof_wald_htz` and
+    :func:`cluster_wald_htz` need:
+
+    Per cluster ``g``:
+        H_gg      = X_g · bread · X_g^T · diag(w_g)            (n_g × n_g)
+        A_g       = (I_g - 0.5(H_gg + H_gg^T))^{-1/2}          (n_g × n_g)
+        G_g       = R · bread · X_g^T · diag(√w_g)             (q × n_g)
+                    # No A_g — Ω = Σ_g G_g G_g^T = R · bread · R^T
+                    # under the working covariance Φ = I.
+        A_g_sqrtW = diag(√w_g) · A_g                           (n_g × n_g)
+                    # Applied to raw residuals: ẽ_g = A_g_sqrtW · u_g.
+                    # Then v_g = G_g · ẽ_g = R bread X_g^T diag(w_g) A_g u_g
+                    # — exactly the weighted CR2 score lifted to R-subspace,
+                    # matching :func:`crve` (CR2 path).
+
+    Aggregate:
+        Ω = Σ_g G_g G_g^T                                       (q × q)
+                    # Unweighted: Ω = R · (X'X)^{-1} · R^T.
+                    # Weighted:   Ω = R · (X'WX)^{-1} · R^T.
+
+    Eigenvalue floor 1e-12 on ``(I - Hsym)`` matches the convention in
+    :func:`crve` (CR2 path) and :func:`cluster_dof_wald_bm`.
+
+    Returns
+    -------
+    dict
+        Keys: ``"G_g"`` (list of q×n_g arrays per cluster, in cluster-code
+        order from ``pd.factorize(sort=False)``), ``"A_g_sqrtW"`` (list of
+        n_g×n_g arrays — same ordering), ``"Omega"`` (q×q),
+        ``"cluster_codes"`` (n-vector), ``"G"`` (cluster count), ``"q"``
+        (R rank), ``"bread"``.
+
+    Raises
+    ------
+    ValueError
+        If ``R`` is rank-deficient, ``G ≤ q``, or ``weights`` non-positive.
+    """
+    R = np.atleast_2d(np.asarray(R, dtype=np.float64))
+    n, k = X.shape
+    if R.shape[1] != k:
+        raise ValueError(f"R has {R.shape[1]} cols but X has k={k}")
+    q = R.shape[0]
+    if q < 1:
+        raise ValueError("R must have at least one row")
+    if q > k:
+        raise ValueError(f"R has {q} rows > k={k}; over-determined")
+    if np.linalg.matrix_rank(R) < q:
+        raise ValueError("R must have full row rank")
+
+    if weights is None:
+        weights = np.ones(n)
+    else:
+        weights = np.asarray(weights, dtype=np.float64).ravel()
+        if (weights <= 0).any():
[…]
