### 2026-04-27 8432c0d
test(fast-fepois): Rust weighted demean matches NumPy reference
5 random seeds × (n=5000, K=2 FEs G=200/30, p=3 cols, random weights)
exercise the new statspai_hdfe.demean_2d_weighted PyO3 entry point
against the existing pure-NumPy _weighted_ap_demean. atol=1e-14.

This is the first end-to-end FFI test for Phase A. Confirms numerical
parity between the Rust path (used when statspai_hdfe is loadable)
and the NumPy fallback path (used when it isn't).

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
