+++ tests/test_fast_fepois.py
@@ -263,3 +263,235 @@ def test_coefs_match_r_fixest(tmp_path):
     for k in ("x1", "x2"):
         diff_coef = abs(float(fit.coef()[k]) - float(r_out["coefs"][k]))
         assert diff_coef < 1e-6, f"vs fixest::fepois coef[{k}] diff {diff_coef:.3e}"
+
+
+# ---------------------------------------------------------------------------
+# Cluster-robust SE (CR1) — Phase 4 follow-up wired through fepois
+# ---------------------------------------------------------------------------
+
+def test_fepois_cluster_validation():
+    df = _poisson_panel(seed=40)
+    # vcov='cr1' without cluster
+    with pytest.raises(ValueError, match="cluster"):
+        sp.fast.fepois("y ~ x1 + x2 | fe1 + fe2", df, vcov="cr1")
+    # cluster with vcov='iid' (mismatched)
+    with pytest.raises(ValueError, match="vcov='iid'"):
+        sp.fast.fepois("y ~ x1 + x2 | fe1 + fe2", df, vcov="iid", cluster="fe1")
+    # cluster with vcov='hc1' (mismatched)
+    with pytest.raises(ValueError, match="vcov='hc1'"):
+        sp.fast.fepois("y ~ x1 + x2 | fe1 + fe2", df, vcov="hc1", cluster="fe1")
+    # missing cluster column
+    with pytest.raises(KeyError):
+        sp.fast.fepois(
+            "y ~ x1 + x2 | fe1 + fe2", df, vcov="cr1", cluster="not_a_col",
+        )
+
+
+def test_fepois_cluster_nan_rejected():
+    df = _poisson_panel(seed=41)
+    df["cl"] = df["fe1"].astype(float)
+    df.loc[df.index[5], "cl"] = np.nan
+    with pytest.raises(ValueError, match="NaN"):
+        sp.fast.fepois("y ~ x1 + x2 | fe1 + fe2", df, vcov="cr1", cluster="cl")
+
+
+def test_fepois_cluster_iid_sanity():
+    """Cluster-robust SE on a non-clustered DGP should be in the same
+    ballpark as IID SE (small G inflation aside). Sanity check that the
+    new path produces finite, positive SEs and a valid result."""
+    df = _poisson_panel(seed=42, n_units=80, n_periods=20)
+    fit_iid = sp.fast.fepois("y ~ x1 + x2 | fe1 + fe2", df, vcov="iid")
+    fit_cr1 = sp.fast.fepois(
+        "y ~ x1 + x2 | fe1 + fe2", df, vcov="cr1", cluster="fe1",
+    )
+    # Same coefs (same IRLS path)
+    for k in ("x1", "x2"):
+        assert abs(float(fit_iid.coef()[k]) - float(fit_cr1.coef()[k])) < 1e-12
+    # Cluster SE positive and finite
+    se_cr1 = fit_cr1.se()
+    assert (se_cr1 > 0).all() and np.isfinite(se_cr1).all()
+    assert fit_cr1.vcov_type == "cr1"
+
+
+def test_fepois_cluster_closed_form():
+    """Pin the cluster CR1 sandwich to its closed-form expression:
+
+        V_CR1 = (X̃' diag(μ) X̃)^{-1}
+                @ Σ_g (X̃_g'(y_g - μ_g))(X̃_g'(y_g - μ_g))'
+                @ (X̃' diag(μ) X̃)^{-1}
+                * (G/(G-1)) * (n-1)/(n - p - Σ(G_k - 1))
+
+    Reconstructs the system from the IRLS-converged ``mu`` and verifies
+    bit-for-bit identity with what fepois reports.
[…]
