### 2026-04-27 d4315ba
feat(output): Phase 4 — synth refactor + 5 more estimator provenance
Continues the v1.7.2 provenance rollout. Closes the deferred sp.synth
dispatcher and adds 4 more high-leverage estimators. Coverage moves
from 4/925 (Phase 3) to 9/925. No numerical changes to any estimator.

sp.synth refactor (deferred from Phase 3):
  - The 13-method dispatcher's 13+ return sites previously made a
    naive attach_provenance pass maintenance-hostile.
  - Split synth() into outer wrapper + _dispatch_synth_impl(): impl
    holds the original dispatcher logic verbatim, wrapper captures
    kwargs, calls impl, 
[…]
