### 2026-04-27 fbdeb85
feat(fast): cluster_dof_wald_htz + cluster_wald_htz with rtol<1e-8 R parity
Task 3+4 of HTZ implementation plan. Implements:

* _htz_eta_from_quantities: η = q(q+1)/sum(var_mat), from clubSandwich's
  P_array + total_variance_mat formulation. Critical Ω quantity is the
  *trace* of P_array over clusters (Σ_g G_g G_g^T - Σ_g H_g H_g^T), not
  Σ_g G_g G_g^T alone — that 18% drift was the trap.
* cluster_dof_wald_htz: thin η wrapper (validation + degenerate guard).
* cluster_wald_htz: full Wald test, V_R via crve(type='cr2') decomp,
  Hotelling-T² scaling F = (η - q + 1)/(η q) · Q.

Verified pari
[…]
