+++ src/statspai/fast/inference.py
@@ -979,6 +979,8 @@ def _htz_per_cluster_quantities(
 
     G_list: list = []
     Asw_list: list = []
+    A_raw_list: list = []
+    X_g_list: list = []
     Omega = np.zeros((q, q))
 
     for cg in range(G_clusters):
@@ -1014,6 +1016,8 @@ def _htz_per_cluster_quantities(
 
         G_list.append(G_g)
         Asw_list.append(A_g_sqrtW)
+        A_raw_list.append(A_g)
+        X_g_list.append(X_g)
         Omega += G_g @ G_g.T
 
     # Symmetrize Ω against floating drift
@@ -1022,12 +1026,139 @@ def _htz_per_cluster_quantities(
     return {
         "G_g": G_list,
         "A_g_sqrtW": Asw_list,
+        "A_g_raw": A_raw_list,        # for HTZ η formula (clubSandwich convention)
+        "X_g_list": X_g_list,         # for H_g_cs = G_cs · X_g · L^T
+        "R": R,                       # cached for η helper
+        "weights": weights,           # for η to detect w!=1 (raises in v1)
         "Omega": Omega,
         "cluster_codes": cluster_codes,
         "G": G_clusters,
         "q": q,
         "bread": bread,
     }
+
+
+def _htz_eta_from_quantities(qty: dict) -> float:
+    """Compute the HTZ moment-matching DOF η.
+
+    1:1 translation of clubSandwich (R, v0.6.2) ``Wald_testing`` HTZ branch
+    + ``get_P_array(all_terms=TRUE)`` + ``total_variance_mat``. Source
+    reconciliation: ``/tmp/htz_reconciliation.txt`` (developer-side note).
+
+    Reconstructs clubSandwich-style cluster quantities from the helper's
+    raw outputs:
+
+        G_cs[g] = R · bread · X_g^T · A_g                  (q × n_g)
+        H_cs[g] = G_cs[g] · X_g · L^T                      (q × k)
+        Ω_cs    = Σ_g G_cs[g] · G_cs[g]^T                  (q × q)
+
+    where L is the lower-triangular Cholesky factor of ``bread`` (so
+    ``L · L^T = bread``).
+
+    Then for each cluster pair (k, l) (with J = number of clusters):
+        if k == l:  P[:, :, k, k] = G_cs[k] · G_cs[k]^T  −  H_cs[k] · H_cs[k]^T
+        else:       P[:, :, k, l] =                       −  H_cs[k] · H_cs[l]^T
+
+    With Ω_nsqrt = Ω_cs^{-1/2}:
+        B[:, :, k, l] = Ω_nsqrt · P[:, :, k, l] · Ω_nsqrt
+
+    var_mat[s, t] = Σ_{k, l} ( B[s, t, k, l] · B[t, s, k, l]
+                                + B[s, s, k, l] · B[t, t, k, l] )
+
+    η = q · (q + 1) / Σ_{s, t} var_mat[s, t]
+
+    Note: the helper's ``Omega`` (V_R-decomposition convention, no A_g)
+    equals ``R · bread · R^T``. The HTZ formula uses a **different** Ω
+    that includes A_g — ``Ω_cs = Σ_g R · bread · X_g^T · A_g^2 · X_g · bread · R^T``.
+    These two Ω matrices coincide only in degenerate cases.
+    """
+    A_raw_list = qty["A_g_raw"]
+    X_g_list = qty["X_g_list"]
+    bread = qty["bread"]
+    R_arr = np.atleast_2d(np.asarray(qty["R"], dtype=np.float64))
+    weights = qty["weights"]
+    J = qty["G"]
+    q = qty["q"]
+
[…]
