### 2026-04-27 19759ab
test(fast): HTZ validation + invariance + edge case unit tests
Task 5 of HTZ implementation plan. Adds 14 tests covering:
- helper <-> full-wrapper eta self-consistency at q=1
- F = (eta-q+1)/(eta q) * Q + p = sf(F) Hotelling-T scaling
- zero-residual short-circuit (Q=0 -> p=1)
- R shape / rank validation, G<=q rejection
- design-equivariance: X column rescaling + cluster ID relabel
- bread-arg path independence
- singleton cluster RuntimeWarning (no crash)
- r shape mismatch + negative weights + non-uniform weights guards
- q=1 documented drift vs cluster_dof_bm (BM 2002 simplified)

Also: sh
[…]
