+++ tests/test_fast_htz.py
@@ -162,3 +162,235 @@ def test_htz_frozen_fixture_matches_clubsandwich(panel_name):
         res.V_R, np.atleast_2d(np.array(fx["V_R"], dtype=np.float64)),
         rtol=1e-9, err_msg=f"V_R mismatch on panel {panel_name}",
     )
+
+
+# ---------------------------------------------------------------------------
+# Task 4: helper ↔ full-wrapper self-consistency
+# ---------------------------------------------------------------------------
+
+def test_htz_q1_helper_eta_matches_full_wrapper():
+    """The DOF helper and the full wrapper share Step 4 — η must be bit-equal."""
+    df = _ols_panel(seed=70, n_clusters=20, m=25)
+    X = df[["x1", "x2"]].to_numpy()
+    y = df["y"].to_numpy()
+    g = df["g"].to_numpy()
+    beta = np.linalg.solve(X.T @ X, X.T @ y)
+    e = y - X @ beta
+    contrast = np.array([1.0, 0.0])
+
+    nu_helper = sp.fast.cluster_dof_wald_htz(X, g, R=contrast.reshape(1, -1))
+    res = sp.fast.cluster_wald_htz(
+        X=X, residuals=e, cluster=g, R=contrast.reshape(1, -1), beta=beta,
+    )
+    assert abs(nu_helper - res.eta) < 1e-12
+
+
+def test_htz_F_and_p_value_internal_consistency():
+    """Hotelling-T² scaling consistency: F = (η-q+1)/(η q) · Q, p = sf(F)."""
+    from scipy.stats import f as scipy_f
+
+    df = _ols_panel(seed=71, n_clusters=25, m=20)
+    X = df[["x1", "x2"]].to_numpy()
+    y = df["y"].to_numpy()
+    g = df["g"].to_numpy()
+    beta = np.linalg.solve(X.T @ X, X.T @ y)
+    e = y - X @ beta
+
+    R = np.eye(2)
+    res = sp.fast.cluster_wald_htz(
+        X=X, residuals=e, cluster=g, R=R, beta=beta,
+    )
+    # F = (η - q + 1) / (η · q) · Q
+    F_recomputed = (res.eta - res.q + 1) / (res.eta * res.q) * res.Q
+    assert abs(F_recomputed - res.F_stat) < 1e-10
+
+    # p = 1 - F.cdf(F_stat)  with df1=q, df2=η-q+1
+    p_recomputed = scipy_f.sf(res.F_stat, res.q, res.eta - res.q + 1)
+    assert abs(p_recomputed - res.p_value) < 1e-10
+
+
+def test_htz_zero_residuals_returns_p_one():
+    """e ≡ 0 ⇒ Q = 0 ⇒ F = 0 ⇒ p = 1 exactly."""
+    df = _ols_panel(seed=72, n_clusters=20, m=15)
+    X = df[["x1", "x2"]].to_numpy()
+    g = df["g"].to_numpy()
+    beta = np.zeros(2)
+    e = np.zeros(len(df))
+    R = np.eye(2)
+    res = sp.fast.cluster_wald_htz(
+        X=X, residuals=e, cluster=g, R=R, beta=beta,
+    )
+    assert res.Q == 0.0
+    assert res.F_stat == 0.0
+    assert res.p_value == 1.0
+
+
+# ---------------------------------------------------------------------------
+# Task 5: Validation + edge cases
+# ---------------------------------------------------------------------------
+
+def test_htz_validates_R_shape():
+    df = _ols_panel(seed=80)
+    X = df[["x1", "x2"]].to_numpy()
[…]
