### 2026-04-27 4976fc2
test(fast): live R clubSandwich HTZ parity (skipif belt-and-suspenders)
Task 6 of HTZ implementation plan. 3 parametrized live-R tests:
  (seed=1010, G=15, q=1)  rtol < 1e-8
  (seed=1011, G=25, q=2)  rtol < 1e-8
  (seed=1012, G=30, q=2)  rtol < 1e-8

Each spawns R via subprocess, runs clubSandwich Wald_test + dumps
nu_Z=q(q+1)/sum(Var_mat), F, p, df_denom to JSON. Compares to
sp.fast.cluster_wald_htz outputs at rtol<1e-8 (eta, F) / 1e-7 (p).

skipif Rscript not on PATH OR clubSandwich missing — frozen-fixture
test (Phase B Task 3) is the CI backbone for environments without R.

Tests: 23/23 pa
[…]
