### 2026-04-27 39c94d0
fix(fast-fepois): restore CR1 cluster-robust SE integration (HTZ track WIP recovery)
Restores 54 lines of WIP that were destroyed by an earlier
`git reset --hard d4315ba` while recovering from an unrelated
over-commit (8432c0d). Recovered from a Claude Code auto-checkpoint
dangling commit (49ce051b6b @ 2026-04-27 15:41:11) via
`git fsck --lost-found`.

Restored functionality (8 additive hunks, all in fepois()):

- Add `cluster: Optional[str] = None` parameter.
- Expand `vcov` to accept `"cr1"` (Liang-Zeger one-way cluster-robust)
  with FE-rank-aware small-sample factor (G/(G-1)) * (n-1)/dof_r
[…]
