### 2026-04-27 d073caa
feat(causal_llm): Phase 5 — LLM-DAG closed loop + layered credential resolver
Closes the LLM-DAG closed-loop deferred from Phases 2-4. No numerical
changes to any estimator. The export pipeline can now auto-propose a
DAG via Anthropic Claude or OpenAI GPT without users having to
pre-build one. Credential resolution follows the industry-standard
layered fallback pattern (Anthropic SDK / OpenAI SDK / AWS CLI /
kubectl all do this).

Layered LLM client resolver (sp.causal_llm.get_llm_client):
  1. Explicit client= pass-through.
  2. Explicit provider= + api_key= construct.
  3. Env var auto-detec
[…]
