+++ src/statspai/causal_llm/_config.py
@@ -1,0 +1,189 @@
+"""``~/.config/statspai/llm.toml`` — persistent LLM preferences.
+
+Stores **provider** + **model** preferences only. **Never stores API
+keys** — those live in environment variables (``ANTHROPIC_API_KEY`` /
+``OPENAI_API_KEY``) so secrets aren't committed to a config file the
+user might accidentally share or check into a dotfiles repo.
+
+XDG-Base-Directory compliant:
+- Linux / macOS: ``${XDG_CONFIG_HOME:-~/.config}/statspai/llm.toml``
+- Windows: ``%APPDATA%/statspai/llm.toml`` (best-effort)
+
+File format::
+
+    [llm]
+    provider = "anthropic"        # or "openai"
+    model = "claude-sonnet-4-5"   # provider-specific default
+
+The resolver in :mod:`statspai.causal_llm._resolver` reads this file
+exactly once per call (no global cache — easier to test, and the
+overhead is negligible compared to an LLM round-trip).
+"""
+from __future__ import annotations
+
+import os
+import sys
+from pathlib import Path
+from typing import Any, Dict, Optional
+
+
+__all__ = [
+    "config_path",
+    "load_config",
+    "save_config",
+    "set_preferences",
+    "DEFAULT_MODELS",
+]
+
+
+#: Provider → recommended default model. Updated whenever Anthropic /
+#: OpenAI ship a new flagship; users can override per-call.
+DEFAULT_MODELS: Dict[str, str] = {
+    "anthropic": "claude-sonnet-4-5",
+    "openai": "gpt-4o-mini",
+}
+
+
+def config_path() -> Path:
+    """Return the platform-appropriate config file path.
+
+    Honours ``XDG_CONFIG_HOME`` on Linux/macOS; falls back to
+    ``~/.config`` if unset. Windows uses ``%APPDATA%``.
+    """
+    if sys.platform == "win32":
+        appdata = os.environ.get("APPDATA")
+        base = Path(appdata) if appdata else Path.home() / "AppData" / "Roaming"
+    else:
+        xdg = os.environ.get("XDG_CONFIG_HOME")
+        base = Path(xdg) if xdg else Path.home() / ".config"
+    return base / "statspai" / "llm.toml"
+
+
+def _parse_toml(text: str) -> Dict[str, Any]:
+    """Minimal TOML parser sufficient for our flat ``[llm]`` schema.
+
+    We deliberately don't import ``tomllib`` (3.11+ only) at module
+    import time — Python 3.9/3.10 users still need this to work. The
+    schema is so simple that a 30-line manual parser beats pulling in
+    ``tomli`` as a dep.
+    """
+    out: Dict[str, Any] = {}
+    section = None
+    for raw_line in text.splitlines():
+        line = raw_line.strip()
+        if not line or line.startswith("#"):
+            continue
+        if line.startswith("[") and line.endswith("]"):
+            section = line[1:-1].strip()
+            out.setdefault(section, {})
+            continue
+        if "=" not in line:
+            continue
+        key, value = line.split("=", 1)
+        key = key.strip()
+        value = value.strip()
+        # Strip inline comment.
+        if "#" in value and not (value.startswith('"') or value.startswith("'")):
+            value = value.split("#", 1)[0].strip()
+        # Strip surrounding quotes.
+        if (value.startswith('"') and value.endswith('"')) or \
+           (value.startswith("'") and value.endswith("'")):
+            value = value[1:-1]
+        # Cast booleans / ints minimally — provider+model are strings,
+        # but future fields might be bools.
+        if value.lower() in ("true", "false"):
+            value = value.lower() == "true"
+        elif value.isdigit():
+            value = int(value)
+        if section is None:
+            out[key] = value
+        else:
+            out[section][key] = value
[…]
