### 2026-04-27 68b99df
docs: HTZ shipping — paper.bib (Crossref-verified) + CHANGELOG + SUMMARY + cross-link
Tasks 7+8 of HTZ implementation plan.

paper.bib: pustejovsky2018small added after Crossref dual-source
verification (DOI 10.1080/07350015.2016.1247004 — authors / 2018 /
JBES 36(4) / 672-683 / title — all four elements per CLAUDE.md §10).

CHANGELOG.md: top-of-Unreleased entry covering Added / Verification /
Scope (v1) / References. Notes the explicit limit Φ=I and the closed
50-100% BM-vs-HTZ gap.

SUMMARY.md: appended a 2026-04-27 independent-PR blurb to the
post-audit update header.

inference.py: cross-l
[…]
