+++ paper.bib
@@ -4191,3 +4191,15 @@
   year={2020},
   doi={10.1177/1536867X20909691}
 }
+
+@article{pustejovsky2018small,
+  author    = {Pustejovsky, James E. and Tipton, Elizabeth},
+  title     = {Small-Sample Methods for Cluster-Robust Variance Estimation
+               and Hypothesis Testing in Fixed Effects Models},
+  journal   = {Journal of Business \& Economic Statistics},
+  volume    = {36},
+  number    = {4},
+  pages     = {672--683},
+  year      = {2018},
+  doi       = {10.1080/07350015.2016.1247004}
+}
