### 2026-04-27 daa035f
feat(fast-fepois): route weighted AP demean to Rust when available
- Rename existing _weighted_ap_demean → _weighted_ap_demean_numpy
  (canonical reference + Rust-unavailable fallback).
- Add _weighted_ap_demean dispatcher: routes K≥2 to
  statspai_hdfe.demean_2d_weighted (Phase A Rust kernel) when the
  extension is loadable, else delegates to _numpy.
- IRLS call sites unchanged (same name, same signature) — the
  cluster-robust SE path lands intact.

K=0 (no-op) and K=1 (closed form) stay on the NumPy path because FFI
overhead would dominate. K≥2 is where the Rust win lives. Rust ↔
NumPy par
[…]
