### 2026-04-27 cf28aa3
docs(guides): Phase 6 — replication_workflow + llm_dag_setup user guides
Surface the v1.7.2 export trinity (Phases 1-5) in the mkdocs site.
Without these guides, users had to read CHANGELOG entries to discover
sp.replication_pack / sp.Provenance / paper.qmd / sp.gt / sp.csl_url /
paper_from_question / llm='auto'. Now there's a single end-to-end
walkthrough plus a focused LLM credential guide.

docs/guides/replication_workflow.md (~530 lines):
  - Quickstart (4 lines: paper -> replication_pack).
  - Two entry points: natural-language vs estimand-first.
  - Output formats with method/extension/u
[…]
