+++ docs/guides/llm_dag_setup.md
@@ -1,0 +1,306 @@
+# LLM-DAG Setup — Credential Resolution & Auto-Propose
+
+> Three layers of fallback so `sp.paper(..., llm='auto')` "just works":
+> environment variable → config file → terminal prompt → fail with
+> concrete remediation. API keys never live in a plaintext config
+> file (industry-standard split).
+
+## TL;DR
+
+```bash
+# Anthropic Claude:
+export ANTHROPIC_API_KEY=sk-ant-...
+
+# OR OpenAI GPT:
+export OPENAI_API_KEY=sk-...
+```
+
+Then in Python:
+
+```python
+import statspai as sp
+draft = sp.paper(df, "effect of trained on wage",
+                 treatment="trained", y="wage",
+                 llm="auto")
+```
+
+That's it. The resolver detects whichever env var is set, picks a
+sensible default model, calls `llm_dag_propose`, and attaches the
+returned DAG to your draft. See [Resolution order](#resolution-order)
+for the full layered fallback.
+
+## Why a layered resolver
+
+StatsPAI's LLM credential handling follows the same pattern as the
+Anthropic SDK, OpenAI SDK, AWS CLI, `kubectl`, and `huggingface_hub`:
+
+1. **API keys live in environment variables**, never in a plaintext
+   config file. Plaintext keys leak — committed to dotfiles repos
+   by accident, synced to cloud backups without encryption, sometimes
+   world-readable when users forget `chmod 600`.
+2. **Provider + model preferences live in a config file**
+   (`~/.config/statspai/llm.toml`), so you don't have to retype them
+   every call.
+3. **Interactive prompts only on TTY**, never inside a Jupyter kernel
+   or agent script (where `input()` blocks indefinitely).
+
+The result: you set `ANTHROPIC_API_KEY` once at machine setup, and
+every StatsPAI call that needs an LLM finds it without further
+configuration.
+
+## Resolution order
+
+`sp.causal_llm.get_llm_client()` walks the layers below in order;
+**first match wins**.
+
+### 1. Explicit `client=`
+
+Pass an already-built `LLMClient` instance:
+
+```python
+client = sp.causal_llm.anthropic_client(
+    model="claude-sonnet-4-5",
+    api_key="sk-ant-...",
+)
+client = sp.causal_llm.get_llm_client(client=client)  # pass-through
+```
+
+Useful for tests (inject a mock / `echo_client`) and for advanced
+users who roll their own retry / caching wrapper.
+
+### 2. Explicit `provider=` + `api_key=`
+
+```python
+client = sp.causal_llm.get_llm_client(
+    provider="anthropic",
+    api_key="sk-ant-...",
+    model="claude-sonnet-4-5",   # optional; defaults per provider
+)
+```
+
+Forces a specific provider regardless of env state. Use this when
+you want one Python process to talk to one provider while another
+talks to another.
+
+### 3. Environment variable auto-detect
+
+```bash
+export ANTHROPIC_API_KEY=sk-ant-...    # → Anthropic, default model
+# OR
+export OPENAI_API_KEY=sk-...           # → OpenAI, default model
+```
+
+```python
+client = sp.causal_llm.get_llm_client()   # picks whichever is set
+```
+
+When **both** are set, the resolver tie-breaks via the config file's
+`[llm].provider`. If no config file exists, it tie-breaks to
+Anthropic.
+
+### 4. Config file `~/.config/statspai/llm.toml`
+
+Stores **provider** and **model** preferences (XDG-Base-Directory
+compliant). API keys are never written here.
+
+```toml
+[llm]
+provider = "anthropic"
+model = "claude-sonnet-4-5"
[…]
