+++ docs/guides/replication_workflow.md
@@ -1,0 +1,530 @@
+# Replication Workflow — Question → Estimate → Paper → Archive
+
+> One call to bundle data, code, environment, paper, citations, and
+> per-number provenance into a journal-ready replication archive.
+> Built for the AEA / AEJ data-editor checklist out of the box.
+
+This guide ties together the v1.7.2 export trinity:
+
+- `sp.paper(...)` / `q.paper()` — **data → draft** pipeline (Markdown,
+  LaTeX, Quarto `.qmd`, Word).
+- `sp.replication_pack(...)` — **draft → journal-ready zip** with
+  manifest, hashes, environment lock, and lineage.
+- `sp.Provenance` / `sp.attach_provenance()` — **per-number traceability**
+  back to the call that produced it.
+
+Plus the surrounding glue:
+
+- `sp.gt(result)` — `great_tables` adapter for publication-grade HTML
+  / LaTeX tables.
+- `sp.csl_url(...)` / `sp.write_bib(...)` — CSL hub + `paper.bib`
+  writer for Quarto citation rendering.
+- `sp.paper(..., llm='auto')` — auto-propose a Causal DAG via LLM
+  (see [LLM-DAG setup guide](llm_dag_setup.md)).
+
+## When to use
+
+- You're submitting to AER / AEJ / Econometrica / QJE / RestStat /
+  RestUd / JF / JPE and the data editor wants a self-contained
+  replication archive.
+- You're an agent that needs to produce **audit-grade** empirical
+  reports — every number traceable to a function call + parameter
+  set + input data hash.
+- You want **one source** that compiles to PDF / HTML / DOCX / Beamer
+  via Quarto, with auto-generated citations and an embedded
+  Reproducibility appendix.
+
+## Quickstart — full pipeline in 4 lines
+
+```python
+import statspai as sp
+import pandas as pd
+
+df = pd.read_csv("training_panel.csv")
+
+# 1. Question → estimate → draft
+draft = sp.paper(df, "effect of trained on wage",
+                 treatment="trained", y="wage",
+                 fmt="qmd")  # Quarto-native output
+
+# 2. Draft → journal-ready replication archive
+sp.replication_pack(draft, "submission.zip",
+                    code="analysis.py")
+```
+
+Open `submission.zip` and you'll find:
+
+```
+submission.zip
+├── MANIFEST.json          versions, timestamp, git SHA, per-file SHA-256
+├── README.md              replication instructions
+├── data/
+│   ├── dataset.csv        the analysis frame
+│   └── manifest.json      shape + dtypes + SHA-256
+├── code/
+│   └── script.py          your analysis script
+├── env/
+│   └── requirements.txt   from pip freeze (or importlib.metadata fallback)
+├── paper/
+│   ├── paper.qmd          Quarto source — `quarto render paper.qmd`
+│   └── paper.bib          auto-emitted from estimator citations
+└── lineage.json           per-result Provenance (function + params + data hash)
+```
+
+Hand the zip to a co-author or upload it to the journal's data
+repository — `quarto render paper/paper.qmd` reproduces your draft
+verbatim, and `MANIFEST.json` lets anyone verify the data is byte-
+identical to what you analyzed.
+
+## Two entry points
+
+### A. Natural-language path
+
+When you want StatsPAI to infer the design from prose:
+
+```python
+draft = sp.paper(df,
+    "effect of training on wages, controlling for education",
+    treatment="trained", y="wage",
[…]
