+++ mkdocs.yml
@@ -78,6 +78,8 @@ nav:
       - "v1.6 P1: LLM-assisted causal discovery — closed loop (constrained PC + CI validate)": guides/llm_dag_family.md
       - "v1.6 P1: sp.paper() — data → publication-draft pipeline": guides/paper_pipeline.md
       - "v1.6 P1: Causal text MVP (text-as-treatment / LLM-annotator MEC) — experimental": guides/causal_text_family.md
+      - "v1.7.2 export trinity: replication_pack + Provenance + Quarto + great_tables + CSL": guides/replication_workflow.md
+      - "v1.7.2 LLM-DAG setup: providers, env vars, configure_llm(), sp.paper(llm='auto')": guides/llm_dag_setup.md
   - Reference:
       - "Overview": reference/index.md
       - "Difference-in-differences": reference/did.md
