### 2026-04-27 0bb244f
fix(fast-fepois): Task 8 review fixes — parity test + module docstring
Address Important findings from the Task 8 code-quality review:

1. test_rust_weighted_demean_matches_numpy_kernel imported the public
   _weighted_ap_demean symbol, which after Task 8's rename is the
   dispatcher (routes to Rust). This made the parity test compare
   Rust ↔ Rust instead of Rust ↔ NumPy — providing zero regression
   protection. Now imports _weighted_ap_demean_numpy explicitly,
   restoring the intended Rust-vs-NumPy parity check.

2. Module docstring claimed "the Rust kernel does not yet expose a
   weigh
[…]
