### 2026-04-27 2f1cb8b
test(fast-fepois): end-to-end Phase A parity vs pyfixest (3 new tests)
Adds three parity tests on a 50k-row, 500/50 FE synthetic Poisson
panel + a private _make_synthetic_panel helper:

- test_fepois_rust_path_coef_parity_vs_pyfixest — coef atol 1e-13.
- test_fepois_rust_path_with_weights — weighted coef atol 1e-13.
- test_fepois_falls_back_when_rust_unavailable — monkeypatch
  _HAS_RUST_HDFE=False; Rust ↔ NumPy fallback agree at atol 1e-12.

These complement Task 7's kernel-level parity test (Rust ↔ NumPy at
1e-14) by exercising the FULL fepois() pipeline (formula parsing,
pre-passes, IRLS, v
[…]
