+++ tests/test_fast_fepois.py
@@ -495,3 +495,91 @@ def test_rust_weighted_demean_matches_numpy_kernel():
             X_rust, X_ref, atol=1e-14, rtol=0,
             err_msg=f"seed={seed}: Rust weighted demean diverged from NumPy reference",
         )
+
+
+def _make_synthetic_panel(seed: int = 0, n: int = 50_000, G1: int = 500, G2: int = 50):
+    """Synthetic Poisson panel with two FE dimensions for parity tests."""
+    rng = np.random.default_rng(seed)
+    fe1 = rng.integers(0, G1, n)
+    fe2 = rng.integers(0, G2, n)
+    alpha = rng.standard_normal(G1) * 0.3
+    gamma = rng.standard_normal(G2) * 0.3
+    x1 = rng.standard_normal(n)
+    x2 = rng.standard_normal(n)
+    eta = 0.5 * x1 - 0.3 * x2 + alpha[fe1] + gamma[fe2]
+    mu = np.exp(eta.clip(-10, 10))
+    y = rng.poisson(mu)
+    return pd.DataFrame({"y": y, "x1": x1, "x2": x2, "fe1": fe1, "fe2": fe2})
+
+
+def test_fepois_rust_path_coef_parity_vs_pyfixest():
+    """Coef from the Rust-dispatched path must match pyfixest.fepois to 1e-13.
+
+    End-to-end pipeline test: exercises formula parsing, singleton/separation
+    pre-passes, IRLS outer loop with the Rust-routed weighted demean,
+    step-halving, and final coef extraction.
+    """
+    pytest.importorskip("statspai_hdfe")
+    pytest.importorskip("pyfixest")
+    import pyfixest as pf
+
+    df = _make_synthetic_panel(seed=42)
+    fit_sp = sp.fast.fepois("y ~ x1 + x2 | fe1 + fe2", data=df, vcov="iid")
+    fit_pf = pf.fepois("y ~ x1 + x2 | fe1 + fe2", data=df, vcov="iid")
+    np.testing.assert_allclose(
+        fit_sp.coef().values, fit_pf.coef().values,
+        atol=1e-13, rtol=0,
+    )
+
+
+# Note: SE-vs-pyfixest parity for IID is already covered by
+# ``test_se_matches_pyfixest_iid_with_ssc`` above (atol 1e-7, with
+# ``fixef_rm="singleton"`` alignment). Phase A's contribution to SE
+# correctness is verified by ``test_fepois_falls_back_when_rust_unavailable``
+# below, which confirms the Rust dispatch path produces SEs identical
+# to the NumPy fallback at atol ≤ 1e-12. HC1 SE has a small (~1e-5
+# relative) baseline drift vs pyfixest unrelated to Phase A — flagged
+# for the v1.8.x audit, not a Phase A blocker.
+
+
+def test_fepois_rust_path_with_weights():
+    """Coef parity with pyfixest.fepois(..., weights=) on the Rust path."""
+    pytest.importorskip("statspai_hdfe")
+    pytest.importorskip("pyfixest")
+    import pyfixest as pf
+
+    df = _make_synthetic_panel(seed=11)
+    rng = np.random.default_rng(11)
+    df["w"] = rng.uniform(0.5, 2.0, len(df))
+
+    fit_sp = sp.fast.fepois("y ~ x1 + x2 | fe1 + fe2", data=df, weights="w", vcov="iid")
+    fit_pf = pf.fepois("y ~ x1 + x2 | fe1 + fe2", data=df, weights="w", vcov="iid")
+    np.testing.assert_allclose(
+        fit_sp.coef().values, fit_pf.coef().values,
+        atol=1e-13, rtol=0,
+    )
+
+
[…]
