### 2026-04-28 d44e5b0
bench(hdfe): Phase A wall-clock harness + collapse dispatcher double-alloc
Adds run_fepois_phase_a.py (3-rep wall measurement on the medium
dataset, asserts ≤ 1.5 s gate) and the resulting medium_statspai_phase_a.json.
Collapses the dispatcher's astype-then-asfortranarray pair into a
single np.array(..., dtype=f64, order='F') allocation
(Task 8 review minor recommendation; saves one (n, p) buffer copy
per dispatch call, ~10ms total at n=1M / 12 calls — code-quality win,
does NOT close the perf gate).

Phase A merge gate result: medium wall median = 2.449s, gate ≤ 1.5s — FAIL.
Findings document
[…]
