### 2026-04-28 e48d4cc
docs(hdfe): record Phase A round 1 — wall-clock gate FAILED
Per spec §5.7 ("If Phase A wall-clock ≥ 1.5 s, the PR does not
merge: record findings ... return to brainstorming ... do NOT
silently widen the threshold"), this commit documents the failed
benchmark gate before any merge / release decision.

Findings:

- Medium wall: 2.449 s (gate ≤ 1.5 s) — FAIL.
- Speedup vs Phase 0 baseline: 6.1 % (2.61 → 2.45 s), not the
  projected ~50 %.
- Profile: Rust kernel itself is 78.5 % of wall (2.19 s of 2.79 s);
  per-call wall 155-318 ms with AP iters in [14, 16].

What broke — the assumption: NumPy's
[…]
