+++ benchmarks/hdfe/AUDIT.md
@@ -140,3 +140,82 @@ measured directly because the audit fix is in the same patch series.
 
 These remain in [`SUMMARY.md`](SUMMARY.md)'s "what deliberately did NOT
 ship" list with the same priorities.
+
+---
+
+## Phase A round 1 — wall-clock gate FAILED (2026-04-27)
+
+The Phase A spec (`docs/superpowers/specs/2026-04-27-native-rust-irls-fepois-design.md` §5.7) set a **merge blocker** of medium wall ≤ 1.5 s. The actual measurement was **2.449 s** (median of 3 reps). Per spec we MUST NOT silently widen the gate; instead we record findings here and surface to the user for a brainstorm-level decision.
+
+### Numbers
+
+| stage                                   | medium wall | iters | vs fixest 0.64 s |
+| --------------------------------------- | ----------: | ----: | ---------------: |
+| Phase 0 baseline (Python `np.bincount`) |      2.61 s |     6 |            4.08× |
+| Phase A (Rust `demean_2d_weighted`)     | **2.45 s**  |     6 |       **3.83×**  |
+| Phase A gate target                     |   ≤ 1.50 s  |     — |          ≤ 2.34× |
+| R `fixest::fepois`                      |      0.64 s |     5 |            1.00× |
+
+Phase A delivers a **6.1 % wall reduction** (2.61 → 2.45 s), not the projected ~50 % (which would have landed at ~1.4 s).
+
+### What broke — the assumption
+
+The plan assumed the IRLS-internal weighted demean was ~80 % of wall and that a Rust port would deliver ~3-5× over Python. With those numbers, end-to-end wall was projected at ~1.4 s.
+
+What the profile actually shows (cProfile on a single timed `fepois` call, 2.789 s total):
+
+```
+ncalls   tottime  cumtime  filename:lineno(function)
+     1   0.216    2.789    fepois.py:404(fepois)
+    12   0.088    2.293    fepois.py:300(_weighted_ap_demean)        ← dispatcher
+    12   2.190    2.190    {built-in: statspai_hdfe.demean_2d_weighted}  ← Rust kernel
+     6   0.113    0.115    fepois.py:751(_poisson_deviance)
+```
+
+Two findings:
+
+1. **The Rust kernel is 78.5 % of wall (2.19 s of 2.79 s).** The dispatcher overhead (`np.asfortranarray`, `np.bincount` for `wsum`, list packing) is only 0.088 s — i.e., the projected "Python overhead" we'd save by going to Rust was already small. The dominant cost is the Rust AP loop itself, not the Python glue.
+
+2. **Per-Rust-call wall is ~155-318 ms** with AP iters in [14, 16]. For 1M rows × K=2 FEs × p ∈ {1,2}, that is **~5-10× slower than the fixest C++ kernel** (which does similar work in ~50 ms on similar hardware). The bottleneck is the random-scatter inner loop (`scratch[codes[i]] += weights[i] * x[i]`) hitting cache-miss territory on the G1=100k FE bucket array (800 KB scratch, exceeds L1, lives in L2 with mostly-random access pattern).
+
+The honest reframe of the gap:
+
+```
+fixest C++   ----  hand-tuned scatter-gather + likely SIMD + sort-by-FE-code
+np.bincount  ----  hand-tuned C scatter-gather (Python NumPy)
+Phase A Rust ----  straight Rust scatter-gather, Rayon column-parallel
+```
+
[…]
