### 2026-04-28 1397b87
feat(output): Phase 7 — provenance rollout to 21/925 (DiD long-tail + RD)
Continues the v1.7.2 provenance rollout from Phases 3-4. No numerical
changes to any estimator. 12 more estimators instrumented (DiD
long-tail × 10 + RD × 2); sp.etwfe refactored into wrapper +
dispatcher (mirrors Phase 4's sp.synth move). Coverage 9/925 → 21/925.

DiD long-tail (10):
  - sp.cic — Athey-Imbens (2006) Changes-in-Changes.
  - sp.cohort_anchored_event_study — staggered-robust ES
    (arXiv:2509.01829).
  - sp.design_robust_event_study — orthogonalised TWFE
    (arXiv:2601.18801).
  - sp.gardner_did / sp.did
[…]
