+++ src/statspai/did/design_robust.py
@@ -155,7 +155,7 @@ def design_robust_event_study(
     z = att_avg / se_avg if se_avg > 0 else 0.0
     pvalue = float(2 * (1 - stats.norm.cdf(abs(z))))
 
-    return CausalResult(
+    _result = CausalResult(
         method="Design-Robust Event-Study (TWFE, orthogonalised)",
         estimand="ATT (avg post)",
         estimate=att_avg,
@@ -173,6 +173,22 @@ def design_robust_event_study(
         },
         _citation_key='design_robust_es',
     )
+    try:
+        from ..output._lineage import attach_provenance as _attach_prov
+        _attach_prov(
+            _result,
+            function="sp.did.design_robust_event_study",
+            params={
+                "y": y, "treat": treat, "time": time, "id": id,
+                "leads": leads, "lags": lags,
+                "cluster": cluster, "alpha": alpha,
+            },
+            data=data,
+            overwrite=False,
+        )
+    except Exception:  # pragma: no cover
+        pass
+    return _result
 
 
 CausalResult._CITATIONS['design_robust_es'] = (
