### 2026-04-28 3886d3f
docs(htz): drop inline DOI/journal-name citations, single-source via bibkey
audit_citations CLI surfaced 4 false-positive mismatches where the
auditor's surname-proximity heuristic picked up non-author surface
words near the inline DOI string:

* inference.py:1219     "Journal of Business..." adjacency -> "business"
* plan:809              same docstring template       -> "business"
* spec:351              "Wald" near DOI in §6.1 line   -> "wald"
* spec:506              References section "Journal of Business..." -> "business"

Per CLAUDE.md §10 single-source convention, docstrings and design 
[…]
