### 2026-04-28 83deb12
docs(superpowers): add Phase B implementation plan (B0 spike + B1 IRLS)
Two-stage delivery designed around Phase A's failure mode (the
"50 % wall reduction" projection that turned out to deliver 6 %):

- Stage B0 (~1-3 days): Validate sort-by-primary-FE in isolation by
  adding a sequential weighted sweep variant and rewiring the
  dispatcher. Gate: medium wall ≤ 1.5 s (the original Phase A gate).
  If B0 fails, STOP and brainstorm; do NOT auto-proceed to B1.
- Stage B1 (~1-2 weeks, gated on B0 PASS): Native Rust IRLS in
  irls.rs with persistent FePoisIRLSWorkspace (eliminates per-iter
  allo
[…]
