+++ docs/superpowers/plans/2026-04-27-phase-b-rust-irls.md
@@ -1,0 +1,829 @@
+# Phase B — Native Rust IRLS for `sp.fast.fepois` Implementation Plan
+
+> **For agentic workers:** REQUIRED SUB-SKILL: Use superpowers:subagent-driven-development to implement this plan task-by-task. Steps use checkbox (`- [ ]`) syntax for tracking.
+
+**Goal:** Close the wall-clock gap between `sp.fast.fepois` and R `fixest::fepois` from the post-Phase-A 3.83× to ≤ 1.5× (target: medium wall ≤ 0.95 s vs fixest's 0.64 s) via two algorithmic + architectural changes — **sort observations by primary FE code** for sequential L1-cache-friendly sweeps, and **port the IRLS outer loop into Rust** to eliminate 12 FFI round-trips per `fepois` call.
+
+**Architecture:** Two-stage delivery designed to mitigate the failure mode that broke Phase A — committing weeks of work to an unvalidated speedup assumption. Stage **B0** (spike, ~1-3 days) validates sort-by-primary-FE in isolation by adding a sort-aware variant of `weighted_demean_matrix_fortran_inplace` and re-running the Phase A benchmark; if wall drops to ≤ ~1.4 s, Stage **B1** (full Rust IRLS, ~1-2 weeks) proceeds. If B0 misses target, the plan returns to brainstorming before any further investment. The B0 / B1 split is **the structural counter-measure** to Phase A's "assumption broke 6 lines into the post-mortem" failure.
+
+**Tech Stack:** Rust 1.76+ · PyO3 0.21 (abi3-py39) · Rayon 1.10 · numpy crate 0.21 · ndarray 0.15 · maturin · Python 3.9-3.13 · NumPy · pytest.
+
+**Spec:** [`docs/superpowers/specs/2026-04-27-native-rust-irls-fepois-design.md`](../specs/2026-04-27-native-rust-irls-fepois-design.md) §6 sets the original Phase B surface. **This plan supersedes §6.1's "no new dependency / hand-coded Cholesky" assumption with the AUDIT.md finding** that sort-by-primary-FE is load-bearing — without it, even a perfect Rust IRLS port plateaus near Phase A's 2.45 s. The audit at [`benchmarks/hdfe/AUDIT.md`](../../../benchmarks/hdfe/AUDIT.md) "Phase A round 1" section is the design's ground truth.
+
+---
+
+## File Structure
+
+| File | Stage | Action | Responsibility |
+|------|:-----:|--------|----------------|
+| `rust/statspai_hdfe/src/sort_perm.rs`            | B0 | create  | counting-sort `primary_fe_sort_perm`; pure Rust, no PyO3 |
+| `rust/statspai_hdfe/src/demean.rs`               | B0 | modify  | + `weighted_group_sweep_sorted` (sequential O(n)); + `weighted_demean_matrix_fortran_inplace_sorted` driver |
+| `rust/statspai_hdfe/src/lib.rs`                  | B0 | modify  | + `#[pyfunction] demean_2d_weighted_sorted`; reuse helpers |
+| `rust/statspai_hdfe/Cargo.toml`                  | B0 | modify  | bump `0.2.0-alpha.1 → 0.3.0-alpha.1`; `__version__` `"0.3.0" → "0.4.0"` |
+| `src/statspai/fast/fepois.py`                    | B0 | modify  | dispatcher routes K≥2 to the sorted path |
+| `tests/test_fast_fepois.py`                      | B0 | modify  | + sort-aware kernel parity test + sort-aware fepois parity test |
[…]
