### 2026-04-28 98fbc22
docs(did): citations near arXiv 2601.18801 — add Wright author, drop title-surface
Independent small PR addressing the pre-existing audit_citations
mismatch the HTZ rollout sweep surfaced (NOT introduced by HTZ).

Background: arXiv 2601.18801 is "Design-Robust Event-Study Estimation
under Staggered Adoption..." by Craig S Wright (2026), verified via
arXiv API. The auditor's surname-proximity heuristic was picking up
nearby surface words ('twfe', 'diagnostics', 'sensitivity') as
non-author surnames because the citation context had the title and
TWFE keyword adjacent to the bare arXiv ID.

Fix: 
[…]
