### 2026-04-28 f512273
feat(rust-hdfe): add sort_perm — counting-sort permutation by FE code
Phase B0 prerequisite. Returns a permutation π such that codes[π[k]] is
non-decreasing, in O(n + G). Caller applies π once to weights and
columns before IRLS; the resulting sorted layout enables a sequential
weighted sweep that replaces the random-scatter inner loop currently
dominating the Rust demean kernel's wall on the medium benchmark.

3 unit tests cover non-trivial perm / empty groups / round-trip via
group_starts_from_codes_sorted helper. 15 total crate tests pass
(12 from Phase A + 3 new).

This is Stage B0 of the t
[…]
