+++ rust/statspai_hdfe/src/lib.rs
@@ -23,6 +23,7 @@
 
 mod demean;
 mod singletons;
+mod sort_perm;
 
 use numpy::{
     PyArray1, PyReadonlyArray1, PyReadwriteArray1, PyReadwriteArray2, PyUntypedArrayMethods,
