### 2026-04-28 aee6583
review(fast-htz): fix doc drift + early-fail non-uniform weights
Self-review surfaced 3 P1 issues in the HTZ helper after Phase F sweep:

1. Returns docstring missed 4 of the 11 dict keys actually exported
   (A_g_raw, X_g_list, R, weights). Now lists every key with shape
   and consumer.

2. weights validation only checked > 0; non-uniform weights would
   pass here, do one O(G·n_g²·k) cluster loop, and only fail later
   inside _htz_eta_from_quantities with NotImplementedError. Pulled
   the check up to the helper boundary so the stack trace points at
   the user's call, and removed the redu
[…]
