### 2026-04-28 d921141
test(rust-hdfe): B0.1 review fixes — empty-group test + docstring + comment
Address Minor findings from the B0.1 code-quality review:

- Add test ``group_starts_handles_empty_interior_group`` covering the
  ``[0, 0, 2, 2] → [0, 2, 2, 4]`` case where group 1 is empty. The
  implementation handled this correctly already; this test pins the
  contract before B0.2 / B1 callers depend on it.
- Fix docstring on ``group_starts_from_codes_sorted`` — the previous
  "(start, len) per-group offsets" phrasing implied a tuple return
  type; replaced with the accurate "length-(n_groups + 1) start-index
  ar
[…]
