### 2026-04-28 8ad6b6e
feat(output): Phase 8 — provenance rollout to 36/925 (IV + matching + DML)
Continues the v1.7.2 provenance rollout from Phases 3-4-7. No
numerical changes to any estimator. 12 instrumentation points added
(15 user-facing functions, since the JIVE family of 4 share a single
_run instrumentation). Coverage 21/925 → 36/925.

IV family (9 user-facing names, 6 instrumentation points):
  - sp.liml — Limited Information Maximum Likelihood / Fuller.
  - sp.jive — legacy single-method JIVE (regression/advanced_iv).
  - sp.lasso_iv — Belloni-Chen-Chernozhukov-Hansen (2012). Pre-existing
    iv() API dri
[…]
