### 2026-04-28 40f0a33
feat(rust-hdfe): add weighted_group_sweep_sorted (sequential O(n))
Phase B0.2 — sort-aware variant of weighted_group_sweep that exploits
the contiguous per-group layout produced by sort_perm::primary_fe_sort_perm.
Replaces the random-scatter inner loop's L2-cache-miss pattern (the
bottleneck Phase A's wall-clock gate failure surfaced in AUDIT.md
"Phase A round 1") with a sequential sweep that fits in L1 for any
reasonable group size.

For the medium benchmark (n=1M, G1=100k), the random-scatter scratch
buffer is 800 KB — exceeds L1, lives in L2 with random access. The
sorted variant per-group 
[…]
