### 2026-04-28 f1b1bde
test(rust-hdfe): B0.2 review fixes — wsum formula + length assert + empty-group test
Address Important + Minor findings from B0.2 code-quality review:

- Append the formula ``Σ_{i ∈ g} weights[i]`` to the wsum docstring,
  matching the random-scatter parent's documentation. Avoids
  integration-time confusion at B0.3 / B0.4 where the matrix driver
  and PyO3 binding will read this contract.
- Add ``debug_assert!(group_starts.last().copied().unwrap_or(0)
  <= x.len(), …)`` so a malformed group_starts surfaces a clear
  invariant message in debug builds, not just a generic bounds-check
  panic i
[…]
