### 2026-04-28 23035ad
feat(rust-hdfe): add weighted_demean_matrix_fortran_inplace_sorted
Phase B0.3 — Rayon-parallel column dispatcher that uses
weighted_group_sweep_sorted (B0.2) for the primary (high-cardinality)
FE and the existing random-scatter weighted_group_sweep for secondary
FEs. AP loop / Aitken / safeguard layout mirrors the random-scatter
parent function exactly; only the inner primary-FE sweep changes
algorithmically (sequential vs random scatter).

Operates on PRE-PERMUTED inputs — the caller (Python dispatcher in
B0.4) applies π once before this call and applies π⁻¹ once after.
Per-thread scratch is 
[…]
