### 2026-04-28 3f4924e
feat(output): Phase 9 — provenance rollout to 48/925 (TMLE + forest + DR)
Continues the v1.7.2 provenance rollout. No numerical changes to any
estimator. 12 ML-causal + classical-identification estimators
instrumented. Coverage 36/925 → 48/925.

ML-causal (8):
  - sp.tmle — van der Laan & Rose Targeted MLE.
  - sp.tmle.ltmle — Longitudinal TMLE.
  - sp.tmle.hal_tmle — TMLE with Highly Adaptive Lasso.
  - sp.causal_forest — GRF causal forest factory.
  - sp.multi_arm_forest — multi-arm forest.
  - sp.iv_forest — IV causal forest.
  - sp.metalearner — S/T/X/R/DR dispatcher.
  - sp.bcf — Bayesian
[…]
