+++ tests/test_estimator_provenance_round5.py
@@ -1,0 +1,251 @@
+"""Round-5 estimator provenance instrumentation.
+
+Layered on Phases 3+4+7+8 (36 estimators). This round adds 12 ML-causal
++ identification-strategy estimators. Coverage 36/925 → **48/925**.
+
+ML-causal:
+- ``sp.tmle`` — van der Laan & Rose Targeted MLE.
+- ``sp.tmle.ltmle`` — Longitudinal TMLE.
+- ``sp.tmle.hal_tmle`` — TMLE with HAL nuisance.
+- ``sp.causal_forest`` — GRF causal forest.
+- ``sp.multi_arm_forest`` — multi-arm causal forest (Athey et al.).
+- ``sp.iv_forest`` — Athey-Tibshirani-Wager IV forest.
+- ``sp.metalearner`` — S/T/X/R/DR meta-learner dispatcher.
+- ``sp.bcf`` — Hahn-Murray-Carvalho Bayesian Causal Forest.
+
+Classical identification:
+- ``sp.aipw`` — Augmented IPW (doubly robust, cross-fit).
+- ``sp.ipw`` — Inverse Probability Weighting.
+- ``sp.g_computation`` — parametric g-formula.
+- ``sp.front_door`` — Pearl front-door adjustment.
+"""
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+
+
+@pytest.fixture
+def binary_df():
+    rng = np.random.default_rng(0)
+    n = 250
+    X1 = rng.normal(size=n)
+    X2 = rng.normal(size=n)
+    ps = 1 / (1 + np.exp(-(X1 + X2)))
+    d = (rng.uniform(size=n) < ps).astype(int)
+    y = 0.5 * d + 0.3 * X1 + rng.normal(size=n)
+    return pd.DataFrame({"y": y, "d": d, "x1": X1, "x2": X2})
+
+
+@pytest.fixture
+def iv_df():
+    rng = np.random.default_rng(2)
+    n = 250
+    z = rng.normal(size=n)
+    d = rng.binomial(1, 0.5, size=n)
+    y = 0.5 * d + 0.3 * z + rng.normal(size=n)
+    return pd.DataFrame({"y": y, "d": d, "z": z,
+                          "x1": rng.normal(size=n),
+                          "x2": rng.normal(size=n)})
+
+
+@pytest.fixture
+def front_door_df():
+    rng = np.random.default_rng(3)
+    n = 200
+    d = rng.binomial(1, 0.5, size=n)
+    m = 0.5 * d + rng.normal(size=n)
+    y = 0.7 * m + rng.normal(size=n)
+    return pd.DataFrame({"y": y, "d": d, "m": m})
+
+
+# ---------------------------------------------------------------------------
+# ML-causal
+# ---------------------------------------------------------------------------
+
+class TestTmleProvenance:
+    def test_attached(self, binary_df):
+        r = sp.tmle(binary_df, y="y", treat="d",
+                    covariates=["x1", "x2"], n_folds=3)
+        prov = sp.get_provenance(r)
+        assert prov is not None
+        assert prov.function == "sp.tmle"
+        assert prov.params["n_folds"] == 3
+        assert prov.params["estimand"] == "ATE"
+
+
+class TestHalTmleProvenance:
+    def test_attached(self, binary_df):
+        from statspai.tmle.hal_tmle import hal_tmle
+        r = hal_tmle(binary_df, y="y", treat="d",
+                      covariates=["x1", "x2"],
+                      max_anchors_per_col=5, n_folds=3)
+        prov = sp.get_provenance(r)
+        assert prov is not None
[…]
