### 2026-04-28 ba8fb43
fix(audit): DOI regex captures Elsevier-style + trim trailing period
DOI_RE lookahead was stopping at the first `.letter` boundary, truncating
DOIs like 10.1016/j.econlet.2009.04.026 down to 10.1016/j and dropping
the .x suffix from 10.1111/j.1467-9469.2007.00585.x. That made the
Citation Audit workflow red on every push that mentioned Elsevier
DOIs in docs/.

Replace the `\.(?:$|\s|[^\d])` alternative with an outer `\.?` that
strips a single trailing period only when followed by a hard delimiter
or end-of-string. Now captures full DOIs while still discarding
sentence-ending periods.

Add five
[…]
