### 2026-04-28 547b4db
fix(fast): commit missing feols.py + tests to repair CI/CD
src/statspai/fast/__init__.py was committed referencing
`from .feols import feols, FeolsResult` but the feols.py module itself
was never staged, so every push since aee6583 broke the CI/CD pytest
collection step on all five OS/Python shards with:

    ModuleNotFoundError: No module named 'statspai.fast.feols'

Ship the missing files:

- src/statspai/fast/feols.py — native OLS HDFE estimator (closed-form
  WLS on FE-residualised X/y, vcov: iid/hc1/cr1, optional weights → WLS).
  Mirrors fepois result shape so downstream code (sp.fast.ev
[…]
