+++ tests/test_fast_feols.py
@@ -1,0 +1,348 @@
+"""Tests for ``sp.fast.feols`` — native OLS HDFE estimator."""
+from __future__ import annotations
+
+import json
+import shutil
+import subprocess
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+
+
+# ---------------------------------------------------------------------------
+# Synthetic OLS DGP (panel with within-cluster correlation)
+# ---------------------------------------------------------------------------
+
+def _ols_panel(n=4_000, n_units=100, n_periods=20, seed=0,
+               beta=(0.30, -0.20)):
+    rng = np.random.default_rng(seed)
+    i = np.repeat(np.arange(n_units), n_periods)[:n]
+    t = np.tile(np.arange(n_periods), n_units)[:n]
+    n = i.size
+    x1 = rng.normal(size=n)
+    x2 = rng.normal(size=n)
+    a = rng.normal(0, 0.5, size=n_units)[i]      # unit FE
+    g = rng.normal(0, 0.3, size=n_periods)[t]    # time FE
+    cl = rng.normal(0, 0.4, size=n_units)[i]      # cluster shock
+    eps = cl + rng.normal(size=n)
+    y = beta[0] * x1 + beta[1] * x2 + a + g + eps
+    return pd.DataFrame({
+        "y": y, "x1": x1, "x2": x2,
+        "fe1": i.astype(np.int32), "fe2": t.astype(np.int32),
+    })
+
+
+# ---------------------------------------------------------------------------
+# Coefficient parity vs pyfixest
+# ---------------------------------------------------------------------------
+
+@pytest.mark.parametrize("seed", [0, 3, 11])
+def test_feols_coef_matches_pyfixest_iid(seed):
+    pf = pytest.importorskip("pyfixest")
+    df = _ols_panel(seed=seed)
+    fit = sp.fast.feols("y ~ x1 + x2 | fe1 + fe2", df, vcov="iid")
+    pf_fit = pf.feols(fml="y ~ x1 + x2 | fe1 + fe2", data=df,
+                       vcov="iid", fixef_rm="singleton")
+    for k in ("x1", "x2"):
+        diff = abs(float(fit.coef()[k]) - float(pf_fit.coef()[k]))
+        assert diff < 1e-10, f"coef[{k}] diff {diff:.3e}"
+
+
+def test_feols_se_iid_close_to_pyfixest():
+    """Order-of-magnitude SE parity vs pyfixest IID — 1% tolerance.
+
+    pyfixest mirrors fixest's default ``ssc(fixef.K="nested")`` which
+    for non-nested 2-FE counts ``Σ G_k - K + 1`` (the true matrix rank,
+    accounting for the shared-intercept collinearity). StatsPAI uses
+    ``Σ (G_k - 1)`` consistently across fast/* — off by 1 DOF for K=2,
+    which translates to a ~3e-4 SE drift on this panel. That's well
+    inside 1% (factor of 30×) but we don't promise machine epsilon.
+    """
+    pf = pytest.importorskip("pyfixest")
+    df = _ols_panel(seed=4)
+    fit = sp.fast.feols("y ~ x1 + x2 | fe1 + fe2", df, vcov="iid")
+    pf_fit = pf.feols(fml="y ~ x1 + x2 | fe1 + fe2", data=df,
+                       vcov="iid", fixef_rm="singleton")
+    for k in ("x1", "x2"):
+        rel = abs(float(fit.se()[k]) - float(pf_fit.se()[k])) / max(
[…]
