+++ src/statspai/fast/feols.py
@@ -1,0 +1,427 @@
+"""Native StatsPAI OLS HDFE estimator (``sp.fast.feols``).
+
+Closes the ``sp.fast.*`` symmetry: ``fepois`` (Phase 2) ships native
+Poisson HDFE; ``feols`` ships the linear analogue. Pure-Python
+orchestration on top of the Phase 1 demean kernel + Phase 4 inference
+primitives — independent of pyfixest.
+
+What this is
+------------
+- ``y ~ x1 + x2 | fe1 + fe2`` minimal formula DSL (same as ``fepois``).
+- Closed-form WLS solve on FE-residualised regressors / outcome.
+- Full DOF accounting: ``fe_dof = sum(G_k - 1)`` (matches ``fepois``,
+  matches ``fixest::ssc(fixef.K='full')`` to a uniform 1-DOF off-by-true-
+  rank for K≥2 — same convention as the rest of fast/*).
+- vcov: ``iid`` (homoscedastic), ``hc1`` (heteroscedasticity-robust),
+  ``cr1`` (one-way cluster-robust, FE-rank-aware via ``extra_df``).
+- Optional observation weights → WLS.
+
+What this is NOT (yet)
+----------------------
+- IV / 2SLS. Use ``sp.iv`` / ``sp.fast.fepois`` for analogues; ``feols``
+  here is single-equation OLS only.
+- ``i()`` / ``sw()`` / ``csw()`` formula expansion. Build the dummies
+  yourself with the Phase 3 DSL helpers and pass them in as columns.
+- CR2 / CR3 cluster SE. Wired through :func:`crve` directly if needed.
+
+References
+----------
+Bergé, L. (2018). Efficient estimation of maximum likelihood models with
+multiple fixed-effects: the R package FENmlm. CREA DP 13.
+Correia, S. (2017). Linear models with high-dimensional fixed effects.
+"""
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import List, Optional, Tuple
+
+import numpy as np
+import pandas as pd
+
+
+# ---------------------------------------------------------------------------
+# Result type
+# ---------------------------------------------------------------------------
+
+@dataclass
+class FeolsResult:
+    """Outcome of :func:`feols`.
+
+    Attribute names mirror :class:`FePoisResult` so downstream code
+    using ``.coef()`` / ``.se()`` / ``.tidy()`` keeps working.
+    """
+
+    formula: str
+    coef_names: List[str]
+    coef_vec: np.ndarray
+    vcov_matrix: np.ndarray
+    n_obs: int
+    n_kept: int
+    n_dropped_singletons: int
+    rss: float                  # residual sum of squares (in residualised space)
+    tss: float                  # total sum of squares (after FE residualisation)
+    r_squared_within: float     # within R²
+    fe_names: List[str]
+    fe_cardinality: List[int]
+    df_resid: int               # n_kept - p - sum(G_k - 1)
+    vcov_type: str
+    cluster_var: Optional[str] = None
+    backend: str = "statspai-native"
+
+    def coef(self) -> pd.Series:
+        return pd.Series(self.coef_vec, index=self.coef_names, name="Estimate")
+
+    def se(self) -> pd.Series:
+        return pd.Series(
+            np.sqrt(np.diag(self.vcov_matrix)),
+            index=self.coef_names, name="Std. Error",
+        )
+
+    def vcov(self) -> pd.DataFrame:
+        return pd.DataFrame(
+            self.vcov_matrix, index=self.coef_names, columns=self.coef_names,
+        )
+
+    def tidy(self) -> pd.DataFrame:
+        b = self.coef_vec
+        s = np.sqrt(np.diag(self.vcov_matrix))
+        with np.errstate(divide="ignore", invalid="ignore"):
[…]

[…]
