### 2026-04-28 126bcdb
docs(rust-hdfe): B0.3 review fixes — assertions + FP-order docstring note
Address Important + Minor findings from B0.3 code-quality review:

- Add ``debug_assert_eq!(primary_starts.len(), primary_wsum.len() + 1)``
  at the matrix-driver entry, matching the layered assertions on
  weighted_group_sweep_sorted. Misconfigured callers now surface a
  clear panic message at the call-site rather than from inside the
  Rayon thread pool with no context.
- Add ``debug_assert!(!primary_wsum.is_empty())`` to make the
  "sorted dispatch requires at least one primary FE" precondition
  self-documenting (th
[…]
