### 2026-04-28 4f15105
feat(output): Phase 10 — provenance rollout to 54/925 (panel + decomp + mediation)
Continues the v1.7.2 provenance rollout. No numerical changes to any
estimator. 6 estimators instrumented; sp.panel refactored into outer
wrapper + _dispatch_panel_impl (parallel to Phase 4 sp.synth and
Phase 7 sp.etwfe moves). Coverage 48/925 → 54/925.

Estimators:
  - sp.panel — multi-method dispatcher (FE/RE/BE/FD/pooled/twoway/
    CRE/GMM). Public signature bit-identical; impl moved to
    _dispatch_panel_impl.
  - sp.causal_impact — Brodersen et al. (2015) BSTS-style impact.
  - sp.mediate — Imai-Keele-Tin
[…]
