### 2026-04-28 c8d7ec1
feat(fast-fepois,rust-hdfe): wire sort-aware demean into IRLS dispatcher (v0.4.0)
Phase B0.4 — completes the spike infrastructure that B0.5 will measure.

Rust side:
- New #[pyfunction] demean_2d_weighted_sorted exposes
  weighted_demean_matrix_fortran_inplace_sorted (B0.3) to Python.
  Validates shapes (n, p, primary_starts.len == primary_wsum.len+1,
  per-secondary-FE lengths) and releases the GIL during the parallel
  Rayon column dispatch, mirroring demean_2d_weighted's pattern.
- Crate __version__ 0.3.0 → 0.4.0; Cargo crate 0.2.0-alpha.1 →
  0.3.0-alpha.1.

Python side (fepois.py):
- _wei
[…]
