+++ rust/statspai_hdfe/src/lib.rs
@@ -308,6 +308,138 @@ fn demean_2d_weighted<'py>(
     Ok(out)
 }
 
+/// Sort-aware weighted demean of a Fortran-order (n, p) matrix in place.
+/// Caller has applied the primary-FE sort permutation π to ``x`` (rows),
+/// ``weights``, and the secondary FE codes; this function does NOT
+/// permute. Result is in π-order; caller applies π⁻¹ on return.
+///
+/// Parameters
+/// ----------
+/// x : 2-D float64 ndarray, shape (n, p), Fortran-contiguous, in π order.
+/// primary_starts : ndarray[int64, shape (G1+1,)]
+///     Group-start offsets for the primary FE (caller computes once via
+///     ``primary_fe_sort_perm`` + cumulative count).
+/// primary_wsum : ndarray[float64, shape (G1,)]
+///     Weighted group sums for the primary FE, computed in π order
+///     (i.e., ``np.bincount(codes_perm, weights=weights_perm)``).
+/// secondary_codes : list[ndarray[int64, shape (n,)]]
+///     K-1 arrays, one per non-primary FE; codes are under π.
+/// secondary_wsum : list[ndarray[float64, shape (G_k,)]]
+///     Weighted group sums for non-primary FEs.
+/// weights_sorted : ndarray[float64, shape (n,)]
+///     Per-obs weights in π order.
+/// max_iter, tol_abs, tol_rel, accelerate, accel_period :
+///     Same semantics as ``demean_2d_weighted``.
+#[pyfunction]
+#[pyo3(signature = (x, primary_starts, primary_wsum, secondary_codes, secondary_wsum, weights_sorted, max_iter, tol_abs, tol_rel, accelerate, accel_period))]
+#[allow(clippy::too_many_arguments)]
+fn demean_2d_weighted_sorted<'py>(
+    py: Python<'py>,
+    mut x: PyReadwriteArray2<'py, f64>,
+    primary_starts: PyReadonlyArray1<'py, i64>,
+    primary_wsum: PyReadonlyArray1<'py, f64>,
+    secondary_codes: &Bound<'py, PyList>,
+    secondary_wsum: &Bound<'py, PyList>,
+    weights_sorted: PyReadonlyArray1<'py, f64>,
+    max_iter: u32,
+    tol_abs: f64,
+    tol_rel: f64,
+    accelerate: bool,
+    accel_period: u32,
+) -> PyResult<Bound<'py, PyList>> {
+    if secondary_codes.len() != secondary_wsum.len() {
+        return Err(pyo3::exceptions::PyValueError::new_err(format!(
+            "len(secondary_codes)={} but len(secondary_wsum)={}",
+            secondary_codes.len(),
+            secondary_wsum.len()
+        )));
+    }
+
+    let primary_starts_view = primary_starts.as_slice()?;
+    let primary_wsum_view = primary_wsum.as_slice()?;
+    let weights_view = weights_sorted.as_slice()?;
+    let sec_code_views = py_list_to_i64_views(secondary_codes)?;
+    let sec_wsum_views = py_list_to_f64_views(secondary_wsum)?;
+
+    let arr = x.as_array();
+    let shape = arr.shape();
+    if shape.len() != 2 {
+        return Err(pyo3::exceptions::PyValueError::new_err("x must be 2-D"));
+    }
+    let n = shape[0];
+    let p = shape[1];
+
+    if weights_view.len() != n {
+        return Err(pyo3::exceptions::PyValueError::new_err(format!(
+            "weights_sorted length {} but n={}",
+            weights_view.len(),
+            n
+        )));
+    }
+    if primary_starts_view.len() != primary_wsum_view.len() + 1 {
[…]
