### 2026-04-28 1b171bb
feat(output): Phase 11 — provenance rollout to 61/925 (spatial + qte + bootstrap + conformal)
Continues the v1.7.2 provenance rollout. No numerical changes to any
estimator. 7 estimators spanning spatial econometrics / quantile /
distributional / bootstrap / conformal inference. Coverage 54/925 →
61/925.

Estimators:
  - sp.spatial.spatial_did — spatial-lag DiD with spillover.
  - sp.spatial.spatial_iv — spatial 2SLS.
  - sp.qte.dist_iv — distributional IV / quantile LATE.
  - sp.qte.beyond_average_late — quantile LATE under fuzzy compliance.
  - sp.qte.qte_hd_panel — HD panel QTE via LASSO co
[…]
