+++ src/statspai/qte/beyond_average.py
@@ -177,7 +177,7 @@ def beyond_average_late(
     ci_low = late_q - z_crit * se_q
     ci_high = late_q + z_crit * se_q
 
-    return BeyondAverageResult(
+    _result = BeyondAverageResult(
         quantiles=quantiles,
         late_q=late_q,
         se_q=se_q,
@@ -186,3 +186,21 @@ def beyond_average_late(
         complier_share=complier_share,
         n_obs=n,
     )
+    try:
+        from ..output._lineage import attach_provenance as _attach_prov
+        _attach_prov(
+            _result,
+            function="sp.qte.beyond_average_late",
+            params={
+                "y": y, "treat": treat, "instrument": instrument,
+                "quantiles": list(quantiles) if quantiles is not None
+                              and hasattr(quantiles, "__iter__") else None,
+                "alpha": alpha,
+                "n_boot": n_boot, "seed": seed,
+            },
+            data=data,
+            overwrite=False,
+        )
+    except Exception:  # pragma: no cover
+        pass
+    return _result
