### 2026-04-28 9fb6973
feat(output): Phase 12 — provenance rollout to 66/925 (bounds + randomization + imputation)
Continues the v1.7.2 provenance rollout. No numerical changes to any
estimator. 5 estimators spanning bounds / randomization inference /
imputation. Coverage 61/925 → 66/925.

Estimators:
  - sp.balke_pearl — Balke-Pearl ATE bounds.
  - sp.lee_bounds — Lee (2009) trimming bounds.
  - sp.manski_bounds — Manski (1990) worst-case bounds.
  - sp.fisher_exact — Fisher randomization test (permutation).
  - sp.imputation.mice — Multiple Imputation by Chained Equations.

Tests: 6 new (5 per-estimator + 1 multi-
[…]
