### 2026-04-28 2e21a57
perf(fast-fepois): cache FE-only setup in dispatcher (B0.4 review R2)
Address the load-bearing R2 finding from the B0.4 code-quality review:
the dispatcher was recomputing the sort permutation, inverse perm,
secondary-code permutations, and primary group_starts on every IRLS
iter — work that depends only on the FE codes (which don't change
across IRLS iters), not on weights.

Measured per-call setup cost on the medium benchmark (n=1M, G1=100k):
  np.argsort + inv_perm + searchsorted + secondary perm = 127 ms/call
  × 12 dispatcher calls per fepois = 1.525 s of pure overhead.

Without caching, 
[…]
