+++ src/statspai/fast/fepois.py
@@ -297,6 +297,107 @@ def _weighted_ap_demean_numpy(
     return arr, iters, converged_all
 
 
+@dataclass
+class _SortedDemeanPlan:
+    """FE-code-only setup for the sort-aware Rust dispatcher path.
+
+    Caches the work that depends solely on the FE codes (the permutation,
+    inverse permutation, primary group_starts, π-order secondary codes).
+    The IRLS outer loop changes ``weights`` every iter but the codes are
+    fixed, so this plan is built once per ``fepois`` call and reused
+    across all ~12 dispatcher calls.
+
+    Without caching, ``np.argsort`` + per-FE indexing + ``searchsorted``
+    cost ~127 ms per call on the medium benchmark (n=1M, G1=100k); 12
+    calls = ~1.5 s, which alone exceeds the Phase A wall-clock target.
+    """
+
+    primary_idx: int
+    primary_G: int
+    perm: np.ndarray             # int64, length n
+    inv_perm: np.ndarray         # int64, length n
+    primary_starts: np.ndarray   # int64, length primary_G + 1
+    secondary_codes_p: List[np.ndarray]  # each int64, in π order
+    secondary_G: List[int]       # cardinality of each non-primary FE
+
+
+# Module-level single-slot fingerprint cache for ``_SortedDemeanPlan``.
+# Key is a tuple of (data-pointer, size, first-elem, last-elem) per FE
+# code array — the combination is bulletproof against false positives
+# within a Python session because matching data pointer + size + boundary
+# elements is astronomically unlikely after garbage-collected reuse.
+_PLAN_CACHE_KEY: Optional[Tuple[Tuple[int, int, int, int], ...]] = None
+_PLAN_CACHE_VALUE: Optional[_SortedDemeanPlan] = None
+
+
+def _fe_codes_fingerprint(
+    fe_codes_list: List[np.ndarray],
+) -> Tuple[Tuple[int, int, int, int], ...]:
+    """Lightweight fingerprint identifying an FE-codes list.
+
+    Combines (data-pointer, size, first-elem, last-elem) per array.
+    O(K) with no per-element work — safe to call on every dispatcher hit.
+    """
+    return tuple(
+        (
+            int(c.ctypes.data),
+            int(c.size),
+            int(c[0]) if c.size > 0 else -1,
+            int(c[-1]) if c.size > 0 else -1,
+        )
+        for c in fe_codes_list
+    )
+
+
+def _build_sorted_demean_plan(
+    fe_codes_list: List[np.ndarray],
+    counts_list: List[np.ndarray],
+) -> _SortedDemeanPlan:
+    """Build the FE-only setup for the sorted dispatcher path.
+
+    Picks the highest-cardinality FE as primary, computes the stable
+    sort permutation, the inverse permutation, primary group_starts via
+    searchsorted, and the π-order secondary codes.
+    """
+    K = len(fe_codes_list)
+    primary_idx = int(np.argmax([c.size for c in counts_list]))
+    primary_codes = fe_codes_list[primary_idx]
+    primary_G = int(counts_list[primary_idx].size)
+
+    perm = np.argsort(primary_codes, kind="stable")
+    inv_perm = np.empty_like(perm)
+    inv_perm[perm] = np.arange(perm.size, dtype=perm.dtype)
+
+    primary_codes_p = primary_codes[perm]
+    primary_starts = np.empty(primary_G + 1, dtype=np.int64)
+    primary_starts[:primary_G] = np.searchsorted(
+        primary_codes_p,
+        np.arange(primary_G, dtype=primary_codes_p.dtype),
+        side="left",
+    )
+    primary_starts[primary_G] = primary_codes_p.size
+
+    secondary_codes_p: List[np.ndarray] = []
[…]
