### 2026-04-28 0ca2aaf
bench(hdfe): Phase B0 wall-clock gate PASSED — sort-by-FE validates
Median wall on the medium benchmark (n=1M, G1=100k, G2=1k, 3 reps
warmup x2 + 3 timed):

  Phase 0 (Python np.bincount):     2.61 s
  Phase A (Rust scatter):           2.45 s   gate FAILED
  Phase B0 (Rust sequential+cache): 1.441 s  ← gate PASSED ✅
  Gate target:                    ≤ 1.50 s
  R fixest::fepois:                 0.64 s

  All 3 reps:    1.450, 1.441, 1.426
  Iterations:    6 (converged)
  Speedup:       1.81× vs Phase 0
  vs fixest:     2.25× (was 3.83× — gap closed by 41 %)

Sort-by-primary-FE delivered the pro
[…]
