### 2026-04-28 f41ccbb
docs(hdfe): record Phase B0 round 1 — wall-clock gate PASSED
Companion to ``Phase A round 1`` failure entry. Documents:

- Numbers: median 1.441 s on the medium benchmark (gate ≤ 1.50 s),
  44 % wall reduction vs Phase 0 baseline, 41 % gap-closure to
  fixest (3.83× → 2.25×).
- Decomposition of the win: ~0.5 s from sort-aware sequential primary
  sweep (B0.2 / B0.3), ~0.5 s from FE-only-plan caching in the
  dispatcher (B0.4 review R2). Both are independently load-bearing.
- Verification of the spike's risk-management discipline: the
  B0 measurement validated the assumptions before B1 invests
[…]
