+++ benchmarks/hdfe/AUDIT.md
@@ -219,3 +219,67 @@ The honest options are (per spec §5.7):
 - **C. Invest in Rust-kernel optimization (sort-by-FE-code, SIMD intrinsics) before Phase B.** Multi-day effort with diminishing returns; might or might not close to 1.5 s. Defers Phase B.
 
 The spec explicitly forbids silently widening the gate. The user's call.
+
+---
+
+## Phase B0 round 1 — wall-clock gate PASSED (2026-04-27)
+
+After the Phase A round 1 audit identified sort-by-primary-FE as the load-bearing missing piece, Phase B0 was implemented as a 5-task spike to validate the assumption in isolation before committing to the full B1 (native Rust IRLS) investment.
+
+### Numbers
+
+| stage                                              | medium wall | iters | vs fixest 0.64 s | gate         |
+| -------------------------------------------------- | ----------: | ----: | ---------------: | ------------ |
+| Phase 0 baseline (Python `np.bincount`)            |     2.61 s  |     6 |            4.08× | —            |
+| Phase A (Rust scatter, no cache)                   |     2.45 s  |     6 |            3.83× | ❌ FAIL      |
+| **Phase B0 (Rust sequential + dispatcher cache)**  | **1.441 s** |     6 |        **2.25×** | **✅ PASS**  |
+| Phase A / B0 gate target                           |   ≤ 1.50 s  |     — |          ≤ 2.34× | —            |
+| R `fixest::fepois`                                 |     0.64 s  |     5 |            1.00× | —            |
+
+Phase B0 delivers a **44 % wall reduction over Phase 0** (2.61 → 1.44 s) and **closes the gap to fixest by 41 %** (3.83× → 2.25×). The gate target ≤ 1.50 s is met by 4 % margin (median 1.441 s, all 3 reps in [1.426, 1.450]).
+
+### Where the win came from
+
+Decomposition of the ~1.0 s improvement vs Phase A (2.45 s → 1.44 s):
+
+1. **Sort-aware sequential primary sweep** (~0.5 s saved). The B0.2 / B0.3 kernels replace the random-scatter inner loop's L2-cache-miss pattern (the bottleneck Phase A round 1 surfaced) with an O(n) sequential sweep that fits in L1 for any reasonable group size. On the medium benchmark (G1 = 100k → 800 KB scratch), this is the single largest line item.
+
+2. **FE-only-plan caching in the dispatcher** (~0.5 s saved). The first B0.4 implementation recomputed `np.argsort(primary_codes, kind="stable")` + inverse perm + secondary-code permutations + `np.searchsorted(primary_starts)` on every IRLS iter — work that depends solely on FE codes, not on weights. Direct measurement at n=1M, G1=100k showed 127 ms per dispatcher call × 12 calls = **1.525 s of pure overhead**, which alone exceeded the entire 1.5 s gate budget. The B0.4-review-driven refactor moves this work into a `_SortedDemeanPlan` dataclass cached at module level via a fingerprint (data pointer + size + first/last element per FE array) and rebuilt only when the caller passes new arrays. The plan is rebuilt naturally on every `fepois()` call's first dispatcher hit, since fingerprints differ across calls; cache hits within a single `fepois` reuse the plan across all ~12 IRLS dispatcher calls.
+
+The two changes are complementary: the sequential sweep is the algorithmic primitive; the cache makes it economically usable inside a Python-orchestrated IRLS without paying setup overhead per iter.
+
+### What the assumption-was-actually-correct-this-time check looks like
+
[…]
