### 2026-04-28 0e4039c
fix(ci): repair two pre-existing failures unblocked by feols.py fix
Once 547b4db unblocked pytest collection, two latent failures surfaced
on every shard of the CI/CD matrix:

1. tests/test_agent_blocks_drift.py — docs/guides/paper_pipeline.md was
   missing `sp.replication_pack` from its alternatives block. The
   registry had it but `scripts/sync_agent_blocks.py` had not been re-run
   after the function was added. Re-synced (one drift block, zero edits
   to other guides).

2. tests/test_fast_within_dsl.py::test_sw_drives_multiple_regressions —
   calls `sp.feols` (the pyfixest wrapper, NOT
[…]
