### 2026-04-28 f0e172c
feat(output): regtable eform + column spanners + coef_map + depvar stats + N-mismatch warning
Closes the residual gap between sp.regtable and Stata `esttab` /
R `modelsummary` / R `fixest::etable` for empirical-paper writing.
Output-layer only — no numerical changes to any estimator.

Five new regtable parameters:

- eform : bool | List[bool]
  Report exp(b) (logit OR / poisson IRR / cox HR) with delta-method
  SE (exp(b)*SE) and exp-of-endpoints CI. t and p unchanged because
  H0: b=0 is equivalent to H0: exp(b)=1. Per-model list mixes
  transformed/untransformed columns; multi_se cells are r
[…]
