+++ tests/test_regtable_publication_extensions.py
@@ -1,0 +1,309 @@
+"""
+Publication-quality extensions to ``regtable``.
+
+Covers six feature additions designed to close the gap to Stata ``esttab`` /
+R ``modelsummary`` / R ``fixest::etable`` for empirical-paper workflows:
+
+1. ``eform`` — exponentiate coefficients (logit OR / poisson IRR / cox HR)
+   with delta-method SE and exp-of-endpoints CI; ``bool`` or per-model
+   ``List[bool]``.
+2. ``column_spanners`` — multi-row header with ``\\multicolumn`` spans
+   (LaTeX) / ``colspan`` (HTML) / repeated header (text/markdown).
+3. ``stats=["depvar_mean", "depvar_sd"]`` — auto rows for the dependent
+   variable's sample mean and SD, populated from ``data_info['y']`` when
+   available.
+4. ``coef_map`` — single dict that simultaneously renames, reorders, and
+   drops (mirrors R ``modelsummary``).
+5. N-consistency and SE-type-mixing warnings.
+6. LaTeX special-character escape correctness on user-supplied labels.
+
+These tests are pure black-box: they construct OLS/Logit/Poisson models
+with fixed seeds and inspect ``regtable`` output strings.
+"""
+
+from __future__ import annotations
+
+import re
+import warnings
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+
+
+# ---------------------------------------------------------------------------
+# Fixtures
+# ---------------------------------------------------------------------------
+
+@pytest.fixture
+def ols_models():
+    rng = np.random.default_rng(2026)
+    n = 600
+    x1 = rng.normal(0, 1, n)
+    x2 = rng.normal(0, 1, n)
+    y = 1.0 + 0.5 * x1 + 0.25 * x2 + rng.normal(0, 1, n)
+    df = pd.DataFrame({"y": y, "x1": x1, "x2": x2})
+    m1 = sp.regress("y ~ x1", data=df)
+    m2 = sp.regress("y ~ x1 + x2", data=df)
+    return m1, m2, df
+
+
+@pytest.fixture
+def logit_model():
+    rng = np.random.default_rng(2027)
+    n = 800
+    x = rng.normal(0, 1, n)
+    p = 1.0 / (1.0 + np.exp(-(0.3 + 0.7 * x)))
+    y = (rng.uniform(size=n) < p).astype(int)
+    df = pd.DataFrame({"y": y, "x": x})
+    return sp.logit("y ~ x", data=df)
+
+
+# ---------------------------------------------------------------------------
+# 1. eform
+# ---------------------------------------------------------------------------
+
+class TestEform:
+
+    def test_eform_true_exponentiates_coefs_text(self, logit_model):
+        raw = sp.regtable(logit_model).to_text()
+        eform = sp.regtable(logit_model, eform=True).to_text()
+        # Coefficient changes after exp transform — expect different rendered cell
+        assert raw != eform
+        # x odds ratio should be e^0.7 ≈ 2.01 → cell starting "2." in eform
+        m = re.search(r"\bx\b\s+(-?\d+\.\d+)", eform)
+        assert m, f"Could not find coefficient row for x in:\n{eform}"
+        assert float(m.group(1)) > 1.0, "exp(0.7) should be > 1"
+
+    def test_eform_se_uses_delta_method(self, logit_model):
+        """SE_exp = exp(b) * SE(b) under the delta method."""
+        eform_txt = sp.regtable(logit_model, eform=True).to_text()
+        # Cell layout: "x   <OR>***\n     (<SE>)\n"
+        # Just verify SE row exists and SE > 0 by parsing
[…]
