+++ src/statspai/output/regression_table.py
@@ -130,6 +130,8 @@ class RegtableResult:
         template: Optional[str] = None,
         quarto_label: Optional[str] = None,
         quarto_caption: Optional[str] = None,
+        eform_flags: Optional[List[bool]] = None,
+        column_spanners: Optional[List[Tuple[str, int]]] = None,
     ):
         if not 0.0 < alpha < 1.0:
             raise ValueError(f"alpha must be in (0, 1), got {alpha!r}")
@@ -170,6 +172,38 @@ class RegtableResult:
         # print(result).
         self._output = output
 
+        # Per-model eform flag (length n_models). When True for a model,
+        # the rendered point estimate becomes exp(b), the SE becomes
+        # exp(b)·SE(b) (delta method), and CI bounds become (exp(lo), exp(hi))
+        # of the original endpoints. t and p-values are unchanged because
+        # H0: b=0 is equivalent to H0: exp(b)=1.
+        if eform_flags is None:
+            eform_flags = [False] * self.n_models
+        if len(eform_flags) != self.n_models:
+            raise ValueError(
+                f"eform_flags has {len(eform_flags)} entries but "
+                f"there are {self.n_models} models."
+            )
+        self.eform_flags = [bool(f) for f in eform_flags]
+
+        # Column spanners: a list of ``(label, span)`` tuples where the
+        # spans must sum to n_models. Renders as a multi-row header above
+        # the model labels (LaTeX ``\multicolumn``, HTML ``colspan``,
+        # text-mode centered ASCII).
+        if column_spanners is not None:
+            total_span = sum(int(s) for _, s in column_spanners)
+            if total_span != self.n_models:
+                raise ValueError(
+                    f"column_spanners total span = {total_span} but there are "
+                    f"{self.n_models} models. The spans must partition the "
+                    f"columns exactly (consecutive grouping)."
+                )
+        self.column_spanners = (
+            [(str(lbl), int(s)) for lbl, s in column_spanners]
+            if column_spanners
+            else None
+        )
+
         # Resolve stat keys once
         self._stat_keys = self._resolve_stat_keys()
 
@@ -206,29 +240,55 @@ class RegtableResult:
             all_vars = ordered
         return all_vars
 
-    def _coef_cell(self, model: _ModelData, var: str) -> str:
+    def _model_eform(self, flat_idx: int) -> bool:
+        """Return whether model at flat position ``flat_idx`` uses eform."""
+        if not self.eform_flags:
+            return False
+        if 0 <= flat_idx < len(self.eform_flags):
+            return self.eform_flags[flat_idx]
+        return False
+
+    def _coef_cell(self, model: _ModelData, var: str, flat_idx: int = 0) -> str:
         if var not in model.params.index:
             return ""
-        val = model.params[var]
+        val = float(model.params[var])
+        if self._model_eform(flat_idx) and np.isfinite(val):
+            val = float(np.exp(val))
         txt = _fmt_val(val, self.fmt)
         if self.show_stars and var in model.pvalues.index:
             txt += _format_stars(model.pvalues[var], self.star_levels)
         return txt
 
-    def _se_cell(self, model: _ModelData, var: str) -> str:
+    def _se_cell(self, model: _ModelData, var: str, flat_idx: int = 0) -> str:
         if var not in model.params.index:
             return ""
+        eform = self._model_eform(flat_idx)
         if self.se_type == "ci":
             lo_v, hi_v = _ci_bounds(model, var, self.alpha)
+            if eform:
+                if not (lo_v is None or pd.isna(lo_v)):
+                    lo_v = float(np.exp(lo_v))
+                if not (hi_v is None or pd.isna(hi_v)):
[…]

[…]
