### 2026-04-28 c259dc1
feat(rust-hdfe): B1.1+B1.2 — IRLS structs + hand-coded SPD Cholesky
Foundational pieces for the native Rust IRLS port. Both modules are
crate-internal Rust with unit tests; no PyO3 surface yet (B1.4 wires
the entry point ``fepois_irls``).

irls.rs:
- FePoisIRLSConfig — knobs (maxiter, tol, fe_tol, fe_maxiter, eta_clip,
  accel_period, max_halvings). Default matches the Python path's
  current values.
- FePoisIRLSResult — outputs (beta, x_tilde F-order, w, eta, mu,
  deviance, log_likelihood, iters, converged, n_halvings, max_inner_dx).
- FePoisIRLSWorkspace::new — pre-allocates all per-iter bu
[…]
