+++ rust/statspai_hdfe/src/irls.rs
@@ -1,0 +1,250 @@
+//! Native Rust IRLS state machine for ``sp.fast.fepois``.
+//!
+//! Phase B1: this module hosts ``FePoisIRLSConfig`` (knobs),
+//! ``FePoisIRLSResult`` (outputs), ``FePoisIRLSWorkspace`` (per-fepois
+//! persistent buffers), and ``fepois_loop`` (the main body, lands in
+//! B1.3). The Python ``fepois()`` parses the formula + runs the
+//! singleton/separation pre-passes, then calls the PyO3 binding
+//! ``fepois_irls`` (B1.4) which delegates here.
+//!
+//! Design intent: eliminate the 12 PyO3 round-trips per fepois call
+//! (one per IRLS iter × 2 — z and X) by keeping the entire IRLS state
+//! machine in Rust. The workspace pre-allocates all per-iter buffers
+//! once at construction so the inner loop is allocation-free in the
+//! happy path.
+
+/// Knobs for the IRLS outer loop.
+#[derive(Clone, Debug)]
+pub struct FePoisIRLSConfig {
+    /// Cap on outer IRLS iterations.
+    pub maxiter: u32,
+    /// Relative-deviance convergence tolerance for the outer loop.
+    pub tol: f64,
+    /// Relative tolerance for the inner AP demean.
+    pub fe_tol: f64,
+    /// Cap on inner AP iterations per call.
+    pub fe_maxiter: u32,
+    /// Symmetric clip applied to ``eta`` to keep ``mu = exp(eta)`` in
+    /// double-precision range; matches the Python path's hard-coded 30.
+    pub eta_clip: f64,
+    /// Aitken extrapolation period inside the inner AP demean.
+    pub accel_period: u32,
+    /// Cap on step-halving attempts when deviance fails to decrease.
+    pub max_halvings: u32,
+}
+
+impl Default for FePoisIRLSConfig {
+    fn default() -> Self {
+        Self {
+            maxiter: 50,
+            tol: 1e-8,
+            fe_tol: 1e-10,
+            fe_maxiter: 1000,
+            eta_clip: 30.0,
+            accel_period: 5,
+            max_halvings: 10,
+        }
+    }
+}
+
+/// Result of one ``fepois_loop`` call. ``x_tilde`` and ``w`` are the
+/// final-iteration demeaned X and IRLS working weight; the Python
+/// caller uses them to compute the requested vcov (IID / HC1 / CR1)
+/// without re-running the within transform.
+#[derive(Debug)]
+pub struct FePoisIRLSResult {
+    pub beta: Vec<f64>,           // length p
+    pub x_tilde: Vec<f64>,        // F-order (n, p)
+    pub w: Vec<f64>,              // final IRLS working weights, length n
+    pub eta: Vec<f64>,            // final eta, length n
+    pub mu: Vec<f64>,             // final mu, length n
+    pub deviance: f64,
+    pub log_likelihood: f64,
+    pub iters: u32,
+    pub converged: bool,
+    pub n_halvings: u32,
+    pub max_inner_dx: f64,
+}
+
+/// Persistent per-fepois scratch. Allocated once at construction and
+/// reused across all IRLS iters; the inner loop does no further heap
+/// allocation in the happy path.
+///
+/// ``primary_starts`` and ``sec_codes_p`` are populated by the
+/// constructor (sort permutation applied once). Per-iter buffers
+/// (``weights_p``, ``primary_wsum``, ``sec_wsums``, ``z_buf``,
+/// ``x_buf``, ``before``, ``aitken_hist``) are sized once and
[…]
