+++ rust/statspai_hdfe/src/cholesky.rs
@@ -1,0 +1,171 @@
+//! Hand-coded k×k SPD Cholesky factorization + triangular solves.
+//!
+//! Used by the native Rust IRLS in ``irls.rs`` for the WLS step
+//! ``(X'WX) β = X'Wz``. For the typical fepois case k ≤ 30, an
+//! in-place Rust loop beats a BLAS dispatch by avoiding FFI / setup
+//! overhead, and avoids adding a new wheel-build dependency.
+//!
+//! The matrix is stored row-major as a flat `&mut [f64]` of length
+//! `k*k`. After `cholesky_factor`, the lower triangle holds L such
+//! that A = L L^T; the strict-upper triangle is left untouched (the
+//! caller must not read it after factorization).
+//!
+//! `cholesky_solve` does forward + back substitution to solve
+//! `L L^T x = b` in-place on `b`.
+
+const RIDGE: f64 = 1e-12;
+
+/// In-place Cholesky factor of a k×k SPD matrix stored row-major.
+/// On return, the lower triangle of `a` holds L (the factor); the
+/// strict upper triangle is unchanged. If a diagonal pivot would
+/// underflow below `RIDGE`, a small ridge is added (matches the
+/// `lapack_lite` policy in the Python NumPy fallback).
+///
+/// Returns `Err(())` if the input is too poorly conditioned to factor
+/// even with ridge regularization (currently unreachable on
+/// well-formed Σ_i w_i x_i x_i' for the IRLS use case).
+#[allow(clippy::needless_range_loop)]
+pub fn cholesky_factor(a: &mut [f64], k: usize) -> Result<(), ()> {
+    debug_assert_eq!(a.len(), k * k);
+
+    for j in 0..k {
+        // Diagonal element: a[j,j] -= Σ_{m<j} L[j,m]^2; sqrt; ridge guard.
+        let mut diag = a[j * k + j];
+        for m in 0..j {
+            let ljm = a[j * k + m];
+            diag -= ljm * ljm;
+        }
+        if diag <= RIDGE {
+            diag = diag.max(RIDGE);
+        }
+        let l_jj = diag.sqrt();
+        if !l_jj.is_finite() || l_jj <= 0.0 {
+            return Err(());
+        }
+        a[j * k + j] = l_jj;
+        // Sub-diagonal elements in column j: L[i,j] = (a[i,j] - Σ_{m<j} L[i,m] L[j,m]) / L[j,j]
+        for i in (j + 1)..k {
+            let mut s = a[i * k + j];
+            for m in 0..j {
+                s -= a[i * k + m] * a[j * k + m];
+            }
+            a[i * k + j] = s / l_jj;
+        }
+    }
+    Ok(())
+}
+
+/// In-place solve `L L^T x = b` given the lower-triangular factor `L`
+/// produced by `cholesky_factor`. Forward + back substitution. The
+/// strict upper triangle of `l` is ignored.
+pub fn cholesky_solve(l: &[f64], k: usize, b: &mut [f64]) {
+    debug_assert_eq!(l.len(), k * k);
+    debug_assert_eq!(b.len(), k);
+
+    // Forward substitution: solve L y = b → y = b after this loop.
+    for i in 0..k {
+        let mut s = b[i];
+        for m in 0..i {
+            s -= l[i * k + m] * b[m];
+        }
+        b[i] = s / l[i * k + i];
+    }
+    // Back substitution: solve L^T x = y → x = b after this loop.
+    for i in (0..k).rev() {
+        let mut s = b[i];
+        for m in (i + 1)..k {
+            s -= l[m * k + i] * b[m];
+        }
+        b[i] = s / l[i * k + i];
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
[…]
