### 2026-04-28 f2107ec
feat(output): Phase 13 — provenance rollout to 76/925 (survival + IV variants)
Continues v1.7.2 provenance rollout. No numerical changes. 10
estimators instrumented spanning survival analysis + remaining IV
variants. Coverage 66 → 76/925.

Survival (5):
  - sp.kaplan_meier — KM non-parametric survival.
  - sp.cox — Cox proportional hazards.
  - sp.survival.aft — Accelerated Failure Time MLE.
  - sp.survival.cox_frailty — Cox with shared gamma frailty.
  - sp.survival.causal_survival_forest — RMST ATE forest.

IV variants (5):
  - sp.iv.kernel_iv — kernel IV regression.
  - sp.iv.npiv — sieve n
[…]
