### 2026-04-28 fde5b58
feat(output): Phase 14 — provenance 82/925 (timeseries + diagnostics)
Added: arima, garch, its, local_projections (timeseries) + mccrary_test,
rddensity (diagnostics). 6 new tests pass.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
