### 2026-04-28 fe412f6
feat(rust-hdfe): B1.3 — fepois_loop core IRLS state machine
The central deliverable of Phase B1: the entire Poisson IRLS outer
loop runs in Rust against the FePoisIRLSWorkspace (B1.1), using
weighted_demean_matrix_fortran_inplace_sorted (B0.3) for the within-
transform and the hand-coded SPD Cholesky (B1.2) for the WLS step.

Algorithm (mirrors the Python path bit-for-bit):
1. Initialise mu = max(y, 1) + 0.1, eta = ln(mu).
2. For each outer iter:
   a. z = eta + (y - mu)/mu (working response, original row order).
   b. w = mu * obs_weights (working weight, π applied to weights_p).
   c. Build 
[…]
