+++ rust/statspai_hdfe/src/irls.rs
@@ -198,6 +198,293 @@ impl FePoisIRLSWorkspace {
     }
 }
 
+/// Poisson deviance: 2 · Σ w_i [ y_i log(y_i / μ_i) - (y_i - μ_i) ].
+/// Convention: y log y → 0 at y = 0; mu is clamped at 1e-30 to avoid
+/// log(0). ``obs_weights`` defaults to 1 if all-1 vector is passed.
+fn poisson_deviance(y: &[f64], mu: &[f64], obs_weights: &[f64]) -> f64 {
+    debug_assert_eq!(y.len(), mu.len());
+    debug_assert_eq!(y.len(), obs_weights.len());
+    let mut acc = 0.0_f64;
+    for i in 0..y.len() {
+        let yi = y[i];
+        let mu_i = mu[i].max(1e-30);
+        let r = if yi > 0.0 {
+            yi * (yi / mu_i).ln()
+        } else {
+            0.0
+        };
+        let contrib = r - (yi - mu_i);
+        acc += obs_weights[i] * contrib;
+    }
+    2.0 * acc
+}
+
+/// Native Poisson IRLS outer loop. The entire state machine runs in
+/// Rust against the persistent ``FePoisIRLSWorkspace`` — no allocation
+/// in the happy path beyond what `weighted_demean_matrix_fortran_inplace_sorted`
+/// internally manages for Aitken history.
+///
+/// Parameters
+/// ----------
+/// `y`              — outcome vector, length n.
+/// `x`              — F-order (n, p) regressor matrix.
+/// `obs_weights`    — per-obs weights, length n. Pass an all-1 slice
+///                    for unweighted MLE.
+/// `cfg`            — IRLS knobs.
+/// `ws`             — pre-built workspace from `FePoisIRLSWorkspace::new`.
+///
+/// Returns
+/// -------
+/// `FePoisIRLSResult` with `beta`, `x_tilde` (F-order, in original row
+/// order), `w` (final IRLS working weight), `eta`, `mu`, `deviance`,
+/// `log_likelihood`, `iters`, `converged`, `n_halvings`, `max_inner_dx`.
+#[allow(clippy::too_many_arguments)]
+pub fn fepois_loop(
+    y: &[f64],
+    x: &[f64],
+    obs_weights: &[f64],
+    cfg: &FePoisIRLSConfig,
+    ws: &mut FePoisIRLSWorkspace,
+) -> FePoisIRLSResult {
+    let n = ws.n;
+    let p = ws.p;
+    debug_assert_eq!(y.len(), n);
+    debug_assert_eq!(x.len(), n * p);
+    debug_assert_eq!(obs_weights.len(), n);
+
+    // Initialise: mu = max(y, 1) + 0.1, eta = ln(mu).
+    for i in 0..n {
+        let m = y[i].max(1.0) + 0.1;
+        ws.mu[i] = m;
+        ws.eta[i] = m.ln();
+    }
+    let mut deviance = f64::INFINITY;
+    let mut iters_used: u32 = 0;
+    let mut converged = false;
+    let mut n_halvings: u32 = 0;
+    let mut max_inner_dx: f64 = 0.0;
+
+    // Pre-allocate WLS scratch (small — k = p ≤ ~30 typical).
+    let mut xt_w_x = vec![0.0_f64; p * p];
+    let mut xt_w_z = vec![0.0_f64; p];
+    let mut beta = vec![0.0_f64; p];
+    let mut eta_new = vec![0.0_f64; n];
+    let mut mu_new = vec![0.0_f64; n];
+    // Reusable z buffer in original order (for eta_new computation).
+    let mut z_orig = vec![0.0_f64; n];
+
+    let n_col = 1 + p;
+
[…]
