### 2026-04-28 b158807
feat(output): Phase 15 — provenance 85/925 (pate + jackknife_se + cr2_se)
Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
