### 2026-04-28 619ddba
feat(output): Phase 16 — provenance 91/925 (proximal + gformula + msm)
Added: proximal, bidirectional_pci, pci_mtp, ice, gformula_mc, msm.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
