### 2026-04-28 1c9d2f3
feat(rust-hdfe): B1.4 — expose fepois_irls via PyO3 (v0.5.0)
Single PyO3 entry point that runs the entire Poisson IRLS state
machine (B1.3 fepois_loop) in Rust — eliminating the 12 FFI
round-trips per IRLS iter that B0.4 still has via the per-iter
demean_2d_weighted_sorted dispatch.

Surface:
- fepois_irls(y, x, fe_codes, g_per_fe, obs_weights, config_dict)
  → dict with beta / x_tilde_flat / x_tilde_n / x_tilde_p /
  w / eta / mu / deviance / log_likelihood / iters / converged /
  n_halvings / max_inner_dx.
- x_tilde returned as flat F-order 1-D array; caller reconstructs
  via result["x_tild
[…]
