+++ rust/statspai_hdfe/src/lib.rs
@@ -28,7 +28,8 @@ mod cholesky;
 mod irls;
 
 use numpy::{
-    PyArray1, PyReadonlyArray1, PyReadwriteArray1, PyReadwriteArray2, PyUntypedArrayMethods,
+    PyArray1, PyReadonlyArray1, PyReadonlyArray2, PyReadwriteArray1,
+    PyReadwriteArray2, PyUntypedArrayMethods,
 };
 use pyo3::prelude::*;
 use pyo3::types::{PyDict, PyList};
@@ -442,6 +443,163 @@ fn demean_2d_weighted_sorted<'py>(
     Ok(out)
 }
 
+/// Native Rust Poisson IRLS for ``sp.fast.fepois``.
+///
+/// Single PyO3 entry point that runs the entire IRLS state machine
+/// in Rust — no per-iter FFI round-trips. The Python side parses the
+/// formula and runs the singleton/separation pre-passes; this function
+/// runs the IRLS body and returns the final β plus the demeaned X̃ and
+/// working weights so the caller can compute the requested vcov in
+/// Python.
+///
+/// Parameters
+/// ----------
+/// y : ndarray[float64, shape (n,)]
+///     Outcome vector (after pre-passes).
+/// x : ndarray[float64, shape (n, p)], Fortran-contiguous
+///     Regressor matrix. The function does NOT mutate `x`.
+/// fe_codes : list[ndarray[int64, shape (n,)]]
+///     K dense FE code arrays.
+/// g_per_fe : ndarray[int64, shape (K,)]
+///     Cardinality of each FE.
+/// obs_weights : ndarray[float64, shape (n,)]
+///     Per-observation weights. Pass an all-1 array for unweighted MLE.
+/// config : dict
+///     IRLS knobs: maxiter / tol / fe_tol / fe_maxiter / eta_clip /
+///     accel_period / max_halvings. Missing keys fall back to
+///     `FePoisIRLSConfig::default()`.
+///
+/// Returns
+/// -------
+/// dict with keys: ``beta`` (ndarray, shape (p,)),
+/// ``x_tilde_flat`` (ndarray, shape (n*p,), F-order data),
+/// ``x_tilde_n`` (int), ``x_tilde_p`` (int),
+/// ``w`` (shape (n,)), ``eta`` (shape (n,)), ``mu`` (shape (n,)),
+/// ``deviance`` (float), ``log_likelihood`` (float), ``iters`` (int),
+/// ``converged`` (bool), ``n_halvings`` (int), ``max_inner_dx`` (float).
+///
+/// Notes
+/// -----
+/// ``x_tilde`` is returned as a flat 1-D array in Fortran (column-major)
+/// order together with shape scalars ``x_tilde_n`` and ``x_tilde_p``.
+/// Reconstruct on the Python side with:
+/// ``x_tilde = result["x_tilde_flat"].reshape(result["x_tilde_p"],
+///              result["x_tilde_n"]).T``
+/// which produces a row-major (n, p) view of the F-order data.
+#[pyfunction]
+#[pyo3(signature = (y, x, fe_codes, g_per_fe, obs_weights, config))]
+fn fepois_irls<'py>(
+    py: Python<'py>,
+    y: PyReadonlyArray1<'py, f64>,
+    x: PyReadonlyArray2<'py, f64>,
+    fe_codes: &Bound<'py, PyList>,
+    g_per_fe: PyReadonlyArray1<'py, i64>,
+    obs_weights: PyReadonlyArray1<'py, f64>,
+    config: &Bound<'py, PyDict>,
+) -> PyResult<Bound<'py, PyDict>> {
+    let y_view = y.as_slice()?;
+    let obs_w_view = obs_weights.as_slice()?;
+    let g_view = g_per_fe.as_slice()?;
[…]
