+++ rust/statspai_hdfe/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "statspai_hdfe"
-version = "0.3.0-alpha.1"
+version = "0.4.0-alpha.1"
 edition = "2021"
 authors = ["Bryce Wang <brycew6m@stanford.edu>"]
 description = "PyO3 + Rayon HDFE group-demean kernel for StatsPAI"
