### 2026-04-28 d40ed63
feat(output): Phase 17 — provenance 99/925 (conformal + transport + target_trial + vcnet)
Added: conformal_debiased_ml, conformal_density_ite, conformal_fair_ite,
conformal_continuous (conformal_causal); transport_weights; emulate,
clone_censor_weight (target_trial); vcnet (dose_response).

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
