### 2026-04-28 f8ec439
feat(output): regtable Round 2 — estimate/statistic templates, notation, apply_coef, escape, Word/Excel spanners
Continues commit f0e172c. **No numerical changes** to any
estimator; output-layer only.

Five new regtable parameters:

- estimate / statistic : str
  Per-line template strings mirroring R modelsummary. Placeholders
  {estimate} {stars} {std_error} {t_value} {p_value} {conf_low}
  {conf_high}. Unknown placeholders raise a KeyError at the
  regtable() call site, not buried in the renderer.

- notation : "stars" | "symbols" | tuple
  Significance-marker family. "stars" (default) = ("*
[…]
