+++ src/statspai/output/regression_table.py
@@ -132,6 +132,12 @@ class RegtableResult:
         quarto_caption: Optional[str] = None,
         eform_flags: Optional[List[bool]] = None,
         column_spanners: Optional[List[Tuple[str, int]]] = None,
+        estimate_template: Optional[str] = None,
+        statistic_template: Optional[str] = None,
+        notation: Union[str, Tuple[str, ...]] = "stars",
+        apply_coef: Optional[Any] = None,
+        apply_coef_deriv: Optional[Any] = None,
+        escape: bool = True,
     ):
         if not 0.0 < alpha < 1.0:
             raise ValueError(f"alpha must be in (0, 1), got {alpha!r}")
@@ -204,6 +210,43 @@ class RegtableResult:
             else None
         )
 
+        # ── estimate / statistic templates ─────────────────────────────────
+        # When set, ``_coef_cell`` and ``_se_cell`` build a context dict and
+        # use ``str.format_map`` to produce the cell. Backward compat: when
+        # both are None, the legacy "{val}{stars}" / "({se})" path runs.
+        self.estimate_template = estimate_template
+        self.statistic_template = statistic_template
+
+        # ── notation (significance markers) ────────────────────────────────
+        # ``"stars"``  → ("*", "**", "***") — backward compatible
+        # ``"symbols"`` → ("†", "‡", "§")     — AER / JPE alternative
+        # tuple        → custom 3-tuple of strings
+        self._notation_symbols = self._resolve_notation(notation)
+
+        # ── apply_coef ────────────────────────────────────────────────────
+        # Generalised eform. Conflicts with eform_flags are caught up at the
+        # public regtable() entry. Optional ``apply_coef_deriv`` enables
+        # delta-method SE rescaling: SE → |f'(b)| · SE(b).
+        if apply_coef is not None and not callable(apply_coef):
+            raise TypeError(
+                f"apply_coef must be callable, got {type(apply_coef).__name__}"
+            )
+        if apply_coef_deriv is not None and not callable(apply_coef_deriv):
+            raise TypeError(
+                f"apply_coef_deriv must be callable, got "
+                f"{type(apply_coef_deriv).__name__}"
+            )
+        self.apply_coef = apply_coef
+        self.apply_coef_deriv = apply_coef_deriv
+
+        # ── escape ─────────────────────────────────────────────────────────
+        # When False, the LaTeX/HTML renderers skip ``_latex_escape`` /
+        # ``_html_escape`` on user-supplied label paths so users can pass
+        # raw markup (e.g. ``"$\\beta_1$"``). Cell content (numeric
+        # estimates, computed stat values) is still safe — it never
+        # contains user-controlled metacharacters.
+        self.escape = bool(escape)
+
         # Resolve stat keys once
         self._stat_keys = self._resolve_stat_keys()
 
@@ -248,21 +291,172 @@ class RegtableResult:
             return self.eform_flags[flat_idx]
         return False
 
+    @staticmethod
+    def _resolve_notation(
+        notation: Union[str, Tuple[str, ...]],
+    ) -> Tuple[str, str, str]:
+        """Map a notation spec to a 3-tuple ``(low, mid, high)`` symbols.
+
+        Mirrors the canonical Top-5 ladder where the *high* symbol marks
+        the strictest significance threshold. Three valid input shapes:
+
+        - ``"stars"`` (default) → ``("*", "**", "***")``
+        - ``"symbols"``        → ``("†", "‡", "§")`` (AER / JPE family)
+        - 3-tuple of strings   → custom (must be ordered low → high)
+        """
+        if isinstance(notation, str):
+            key = notation.lower()
+            if key == "stars":
+                return ("*", "**", "***")
+            if key == "symbols":
[…]
