### 2026-04-28 29fa63f
feat(output): Phase 18 — provenance 103/925 (mendelian + bunching + censoring + surrogate)
Added: mr_mode (Mendelian), kink_unified (bunching), ipcw (censoring),
surrogate_index (surrogate). Crosses 100/925 (>10× original baseline).

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
