+++ docs/guides/replication_workflow.md
@@ -378,7 +378,7 @@ So `lineage.json` traces the full chain: aggregate → producing CS run
 
 ## Provenance scorecard
 
-As of v1.7.2, **99 estimators** are instrumented:
+As of v1.7.2, **103 estimators** are instrumented (>10× original 9-baseline):
 
 | Estimator                                                | Phase    |
 |---                                                       |---       |
@@ -481,8 +481,12 @@ As of v1.7.2, **99 estimators** are instrumented:
 | `sp.target_trial.emulate`                                | **P17**  |
 | `sp.target_trial.clone_censor_weight`                    | **P17**  |
 | `sp.dose_response.vcnet` (Varying-coefficient DR)        | **P17**  |
+| `sp.mendelian.mr_mode` (Mendelian Randomization mode)    | **P18**  |
+| `sp.bunching.kink_unified` (RDD+RKD+bunching)            | **P18**  |
+| `sp.censoring.ipcw` (IPCW weights)                       | **P18**  |
+| `sp.surrogate.surrogate_index` (Athey-Chetty-Imbens)     | **P18**  |
 
-The remaining ~826 estimators are scheduled for v1.7.3+ rollouts. To
+The remaining ~821 estimators are scheduled for v1.7.3+ rollouts. To
 check whether a specific estimator is instrumented:
 
 ```python
