### 2026-04-28 38ff8d1
test+bench(fast-fepois): B1.6 native-vs-Python parity test + B1.8 wall-clock gate PASS
B1.6 — Add test_fepois_native_irls_vs_python_irls_parity:
  Verifies native Rust IRLS (single PyO3 call via fepois_irls) matches
  the Python IRLS for-loop (which uses the Phase B0 dispatcher
  internally) at coef atol ≤ 1e-10 and SE atol ≤ 1e-7. Uses
  monkeypatch to delete the fepois_irls attribute on the loaded
  _rust_hdfe module, forcing the Python IRLS path while keeping
  _HAS_RUST_HDFE=True so the B0 sort-aware kernel still runs from
  Python. This is the explicit B1 parity gate — confirms the IRLS
 
[…]
