### 2026-04-28 8e8c485
release(1.8.0): native Rust IRLS for sp.fast.fepois — 2.97× faster, 1.37× of fixest
Bumps StatsPAI to v1.8.0. Headline: Phase B1's native Rust IRLS for
sp.fast.fepois closes the long-standing wall-clock gap to R
fixest::fepois from 4.08× (v1.7.x) to 1.37× — under the ≤ 1.5×
target set in the v1.8 design spec, on the project's standard
medium HDFE benchmark (n=1M, fe1=100k, fe2=1k).

Documentation rolled up in this commit:
- benchmarks/hdfe/SUMMARY.md — wall-clock table updated with the
  full Phase A → B0 → B1 trail (4.08× → 3.83× → 2.25× → 1.37×).
  "Native Rust IRLS for fepois" promoted from
[…]
