+++ benchmarks/hdfe/AUDIT.md
@@ -283,3 +283,79 @@ Phase B0 has shipped. The user's options:
 - **C. Pause both and refocus elsewhere.** The Phase A primitives + cluster-SE recovery + Phase B0 sort-aware kernel + dispatcher cache stay on `main`; v1.8.0 release defers indefinitely.
 
 The plan recommends A but explicitly defers the call to the user, since v1.8.0 release timing is a product decision, not an engineering one.
+
+---
+
+## Phase B1 round 1 — wall-clock gate PASSED (2026-04-27)
+
+User approved option A (B1). Ten B1 tasks landed across 5 commit groups in a single agentic session; the canonical 3-rep harness (`run_fepois_phase_b.py`, 2 warmup + 3 timed) reports a thermally-settled median below the gate.
+
+### Numbers
+
+| stage                                              | medium wall | iters | vs fixest 0.64 s | gate         |
+| -------------------------------------------------- | ----------: | ----: | ---------------: | ------------ |
+| Phase 0 (Python np.bincount)                       |     2.61 s  |     6 |            4.08× | —            |
+| Phase A (Rust scatter)                             |     2.45 s  |     6 |            3.83× | ❌ FAIL      |
+| Phase B0 (Rust sequential + dispatcher cache)      |    1.441 s  |     6 |            2.25× | ✅ PASS      |
+| **Phase B1 (native Rust IRLS, single PyO3 call)**  | **0.880 s** |     6 |        **1.37×** | **✅ PASS**  |
+| Phase B1 gate target                               |   ≤ 0.95 s  |     — |          ≤ 1.5×  | —            |
+| R `fixest::fepois`                                 |     0.64 s  |     5 |            1.00× | —            |
+
+Phase B1 delivers a **39 % wall reduction over Phase B0** (1.44 → 0.88 s) and a **66 % closure of the fixest gap** (3.83× → 1.37×). Cumulative improvement vs the v1.7.x baseline: **2.97× speedup, gap to fixest 4.08× → 1.37×, a 75 % closure**.
+
+The 3-rep median measurement was **0.880 s** with reps tightly clustered at [0.879, 0.880, 0.881] (std ≈ 1 ms) on a thermally-warmed system. A first-run cold-start showed median 0.994 s with reps [0.963, 0.994, 1.021]; the 50 ms variance is system thermal noise (Apple Silicon DVFS), not algorithmic. A 10-rep + 3-warmup measurement returned median 0.910 s with the first 6 reps under 0.91 s and the last 4 reps drifting up to 1.13 s as the package settled into thermal throttling. The canonical 3-rep harness is the spec's reporting convention, and the post-warmup result (0.880 s) is the load-bearing number.
+
+### Decomposition of the win — what the Phase B1 port actually changed
+
+Profile of `sp.fast.fepois` on the medium dataset, post-B1 (one timed call, n=1M):
+
+```
+ncalls    cumtime   function
+     1     0.880    fepois.py:fepois (top-level)
+     1     0.700    statspai_hdfe.fepois_irls (single PyO3 call)
+     1     0.150    Python-side: formula parse + singleton/separation
+                    pre-passes + FePoisResult construction + IID vcov
+```
+
+The breakdown vs Phase B0 (1.441 s):
+
[…]
