### 2026-04-28 9599cbd
feat(structural): production function estimators — OP / LP / ACF / Wooldridge + markup
Mature production-function module shipping with v1.8.0. Solves the
simultaneity bias between input choices and unobserved productivity
via proxy-variable / control-function identification.

Estimators (sp.prod_fn unified dispatcher + named entry points):

- olley_pakes (opreg) — Olley & Pakes (1996, Econometrica). Investment
  proxy for productivity in stage 1; survival-corrected stage 2.
- levinsohn_petrin (levpet) — Levinsohn & Petrin (2003, RES).
  Intermediate-input proxy; addresses zero-investment firms
[…]
