+++ paper.bib
@@ -4203,3 +4203,66 @@
   year      = {2018},
   doi       = {10.1080/07350015.2016.1247004}
 }
+
+% --- Production function estimators (proxy-variable identification) ----------
+% All five entries verified via Crossref API (api.crossref.org/works/<DOI>)
+% on 2026-04-27 — title, authors, journal, year, volume/issue match.
+
+@article{olley1996dynamics,
+  author  = {Olley, G. Steven and Pakes, Ariel},
+  title   = {The Dynamics of Productivity in the Telecommunications
+             Equipment Industry},
+  journal = {Econometrica},
+  volume  = {64},
+  number  = {6},
+  pages   = {1263--1297},
+  year    = {1996},
+  doi     = {10.2307/2171831}
+}
+
+@article{levinsohn2003estimating,
+  author  = {Levinsohn, James and Petrin, Amil},
+  title   = {Estimating Production Functions Using Inputs to Control
+             for Unobservables},
+  journal = {Review of Economic Studies},
+  volume  = {70},
+  number  = {2},
+  pages   = {317--341},
+  year    = {2003},
+  doi     = {10.1111/1467-937X.00246}
+}
+
+@article{ackerberg2015identification,
+  author  = {Ackerberg, Daniel A. and Caves, Kevin and Frazer, Garth},
+  title   = {Identification Properties of Recent Production Function
+             Estimators},
+  journal = {Econometrica},
+  volume  = {83},
+  number  = {6},
+  pages   = {2411--2451},
+  year    = {2015},
+  doi     = {10.3982/ECTA13408}
+}
+
+@article{wooldridge2009estimating,
+  author  = {Wooldridge, Jeffrey M.},
+  title   = {On Estimating Firm-Level Production Functions Using Proxy
+             Variables to Control for Unobservables},
+  journal = {Economics Letters},
+  volume  = {104},
+  number  = {3},
+  pages   = {112--114},
+  year    = {2009},
+  doi     = {10.1016/j.econlet.2009.04.026}
+}
+
+@article{deloecker2012markups,
+  author  = {De Loecker, Jan and Warzynski, Frederic},
+  title   = {Markups and Firm-Level Export Status},
+  journal = {American Economic Review},
+  volume  = {102},
+  number  = {6},
+  pages   = {2437--2471},
+  year    = {2012},
+  doi     = {10.1257/aer.102.6.2437}
+}
