+++ src/statspai/structural/production/_dispatcher.py
@@ -1,0 +1,154 @@
+"""
+Unified ``sp.prod_fn(method=...)`` dispatcher.
+
+Parallels ``sp.synth(method=...)`` and ``sp.decompose(method=...)`` —
+one entry point selects between Olley-Pakes, Levinsohn-Petrin,
+Ackerberg-Caves-Frazer, and Wooldridge.
+"""
+
+from __future__ import annotations
+
+from typing import Optional, Sequence
+
+import pandas as pd
+
+from ._result import ProductionResult
+from .op_lp_acf import (
+    ackerberg_caves_frazer,
+    levinsohn_petrin,
+    olley_pakes,
+)
+from .wooldridge import wooldridge_prod
+
+
+_METHOD_DISPATCH = {
+    "op": olley_pakes,
+    "olley_pakes": olley_pakes,
+    "lp": levinsohn_petrin,
+    "levinsohn_petrin": levinsohn_petrin,
+    "acf": ackerberg_caves_frazer,
+    "ackerberg_caves_frazer": ackerberg_caves_frazer,
+    "wrdg": wooldridge_prod,
+    "wooldridge": wooldridge_prod,
+}
+
+
+def prod_fn(
+    data: pd.DataFrame,
+    output: str = "y",
+    free: Sequence[str] | str | None = None,
+    state: Sequence[str] | str | None = None,
+    proxy: str | None = None,
+    panel_id: str = "id",
+    time: str = "year",
+    method: str = "acf",
+    polynomial_degree: int = 3,
+    productivity_degree: int = 1,
+    functional_form: str = "cobb-douglas",
+    boot_reps: int = 0,
+    seed: Optional[int] = None,
+    **kwargs,
+) -> ProductionResult:
+    """Production function estimation — unified interface.
+
+    Parameters
+    ----------
+    data : DataFrame
+        Long panel with one row per (firm, year).
+    output : str
+        Log output column.
+    free : str or list, default ``["l"]``
+        Free inputs (e.g. labor).
+    state : str or list, default ``["k"]``
+        State / predetermined inputs (capital).
+    proxy : str, optional
+        Productivity proxy. Defaults: ``"i"`` for ``method="op"``,
+        ``"m"`` for all others.
+    panel_id, time : str
+        Panel identifier columns.
+    method : {'op', 'lp', 'acf', 'wrdg'}, default ``'acf'``
+        Estimator. ACF is the modern default (corrects OP/LP
+        identification problem).
+    polynomial_degree : int, default 3
+        Stage-1 control function polynomial degree.
+    productivity_degree : int, default 1
+        Productivity AR polynomial degree.  Default ``1`` (linear AR(1))
+        is the most numerically robust choice — higher degrees can
+        overfit ``omega_t`` given ``omega_{t-1}`` in finite samples and
+        flatten the GMM objective surface, which makes the structural
+        parameters numerically un-identified even when they are
+        identified in population.
+    functional_form : {'cobb-douglas', 'translog'}, default 'cobb-douglas'
+        Functional form. Translog adds 0.5 * x_j**2 own-quadratic terms
+        and x_j*x_k cross terms — output elasticities then vary by
+        firm-time and ``ProductionResult.model_info["elasticities"]``
+        carries a per-row DataFrame.
+
+        Translog identification caveat: stage-2 instruments are formed
+        as polynomial transforms of the same raw set used for
+        Cobb-Douglas (``(k, l_lag)`` for ACF, ``(k, l)`` for OP/LP).
+        This is standard in the literature but the resulting moment
+        system can be near-singular when state and lagged-free inputs
+        are highly correlated, so finite-sample variance on the
[…]
