+++ src/statspai/structural/__init__.py
@@ -1,3 +1,25 @@
 """Structural estimation methods."""
 from .blp import blp, BLPResult
-__all__ = ["blp", "BLPResult"]
+from .production import (
+    prod_fn,
+    olley_pakes,
+    opreg,
+    levinsohn_petrin,
+    levpet,
+    ackerberg_caves_frazer,
+    acf,
+    wooldridge_prod,
+    markup,
+    ProductionResult,
+)
+
+__all__ = [
+    "blp", "BLPResult",
+    "prod_fn",
+    "olley_pakes", "opreg",
+    "levinsohn_petrin", "levpet",
+    "ackerberg_caves_frazer", "acf",
+    "wooldridge_prod",
+    "markup",
+    "ProductionResult",
+]
