### 2026-04-28 d65b17a
feat(output): Phase 19 — provenance 109/925 (panel + bvar + causal_discovery)
Added: panel_fgls, bvar, fci, ges, lingam, dynotears.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
