### 2026-04-28 8ad64ef
feat(output): Phase 20 — provenance 112/925 (causal_text + neural_causal + fairness)
Added: text_treatment_effect, gnn_causal, demographic_parity.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
