### 2026-04-28 f249622
docs(superpowers): add Rust Poisson-separation pre-pass design spec
Path A from the post-Phase-B1 brainstorm: only move pre-passes that
have a real Rust speedup; keep formula parsing / FePoisResult / vcov
in Python where they already live efficiently.

- Goal: 0.880 s → ~0.83 s (1.30× of fixest) by porting the iterative
  _drop_separation function to Rust. ~50 ms saved on the medium
  benchmark; the np.unique + np.isin O(n log n) passes are replaced
  with a bincount-based O(n × n_iter × K) loop.
- Surface: new pub fn separation::separation_mask + PyO3 binding
  separation_mask(y, fe_codes, g_
[…]
