+++ docs/superpowers/specs/2026-04-27-rust-separation-pre-pass.md
@@ -1,0 +1,173 @@
+# Native Rust Poisson Separation Pre-Pass — Design Spec
+
+> **Status**: design (pre-implementation). Authored 2026-04-27. Targets the
+> v1.8.0 release alongside the Phase B1 native Rust IRLS and the prod_fn
+> module. Path A from the post-B1 brainstorm: only move pre-passes that
+> have a real Rust speedup; keep formula parsing / FePoisResult / vcov in
+> Python where they already live efficiently.
+
+## 1. Goal
+
+Move the iterative Poisson-separation drop (currently `_drop_separation`
+in `src/statspai/fast/fepois.py`) from Python to Rust. Reduce medium-
+benchmark wall by ~50 ms (0.880 s → ~0.83 s = 1.30× of fixest::fepois).
+No public API change.
+
+## 2. Non-goals
+
+Explicitly out of scope (per Path A vs B/C in the brainstorm):
+
+- Formula parsing in Rust — `_parse_fepois_formula` stays Python.
+- Column extraction / intercept handling — stay Python.
+- `FePoisResult` construction / `tidy()` — stay Python.
+- IID / HC1 / CR1 vcov computation — stay Python.
+- The CR1 cluster-robust SE integration recovered in commit `39c94d0`
+  is **byte-untouched**.
+
+## 3. Background — what `_drop_separation` does
+
+Poisson regression cannot identify FE coefficients for groups whose
+outcomes are all zero (because `mu_g > 0` for all observations in the
+group, but `Σ y_g = 0`, so the score equation has no interior solution).
+The pre-pass at `fepois.py:_drop_separation` (~30 lines) iteratively:
+
+1. For each FE dimension k, compute `sums = bincount(codes_k, y)`.
+2. Find groups with `sums == 0`.
+3. Drop rows in those groups; recompute on the survivor mask.
+4. Repeat until no drops occur (one pass typically converges; pathological
+   cases may need 2-3).
+
+The cost is `O(n × n_iter × K)` where `n_iter ≤ 3` typically. On the
+medium benchmark (n=1M, K=2, n_iter=2-3), the Python path is ~50 ms.
+
+Profile breakdown of the current Python path (cProfile on a fepois call):
+
+```
+0.046 s   _drop_separation
+0.019 s     np.unique (twice for the two FE dims)
+0.019 s     np.isin
+0.005 s     np.bincount
+0.003 s     mask + indexing
+```
+
+The `np.unique` + `np.isin` are the load-bearing slow ops (each is `O(n
+log n)`). A Rust port using `bincount` + direct iteration avoids both.
+
+## 4. Surface
+
+### 4.1 Rust crate (`statspai_hdfe`)
+
+New module `rust/statspai_hdfe/src/separation.rs` with one `pub fn`:
+
+```rust
+/// Iterative Poisson separation detection. Returns a keep-mask of
+/// length n: `keep[i] = false` iff observation i lives in an FE group
+/// whose total y-sum is zero (after iterating).
+///
+/// Cost: O(n × n_iter × K) with n_iter ≤ 3 for typical inputs;
+/// allocates K group-sum buffers (sized by g_per_fe) once and reuses
+/// them across iterations. No `np.unique` / `np.isin`-like O(n log n)
+/// passes.
+pub fn separation_mask(
+    y: &[f64],
+    fe_codes: &[&[i64]],
+    g_per_fe: &[usize],
+) -> Vec<bool>;
+```
+
+Plus a PyO3 binding in `lib.rs`:
+
+```rust
+#[pyfunction]
+fn separation_mask(
+    py: Python<'py>,
+    y: PyReadonlyArray1<'py, f64>,
[…]
