### 2026-04-28 cb4386a
feat(output): Phase 21 — provenance 117/925 (epi + bridge + interference)
Added: bradford_hill, odds_ratio, did_sc_bridge, network_exposure, peer_effects.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
