### 2026-04-28 75a9b06
feat(rust-hdfe): expose separation_mask via PyO3 (v0.6.0)
Path A from the post-Phase-B1 brainstorm: move the iterative Poisson-
separation pre-pass from Python to Rust. The Python ``_drop_separation``
function (in src/statspai/fast/fepois.py) used np.unique + np.isin
which run O(n log n) per pass; the Rust port uses bincount-based
group-sum accumulation in O(n × n_iter × K) with n_iter ≤ ~3 typical.

Surface:
- New crate-internal pub fn separation::separation_mask(y, fe_codes,
  g_per_fe) -> Vec<bool>. Pre-allocates K group-sum buffers (sized by
  g_per_fe) once and reuses across iterations.
-
[…]
