### 2026-04-28 14de52b
fix(rust-hdfe): separation_mask returns all-keep on empty fe_codes
Defensive contract fix: when fe_codes is empty (k=0), there are no FE
groups → no separation drops are logically possible → return all-keep
mask of length n. The previous behavior (zeros_bound, returning all-drop
of length n) would silently drop every observation if any caller passed
an empty fe_codes list.

Not reachable from sp.fast.fepois (which guards on fe_terms before
calling), but flagged as latent bug during the Task 1+2 controller
review. The fix is a 3-line change replacing zeros_bound with an
explicit vec![1u8; n] fr
[…]
