### 2026-04-28 722b3b8
feat(output): Phase 22 — provenance 122/925 (bridge + causal_rl + matrix_completion)
Added: dr_calib_bridge, cb_ipw_bridge, causal_dqn, causal_bandit, mc_panel.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
