### 2026-04-28 b22eb2f
feat(output): regtable Round 3 — margins_table, tests= footer, fixef_sizes
Three additions on top of Rounds 1 & 2. **No numerical changes** to
any estimator (margins_table is a pure adapter; tests= formats
user-supplied hypothesis-test results; fixef_sizes reads pre-existing
model_info['n_fe_levels']).

- sp.margins_table(model)
  Adapter that wraps a sp.margins() DataFrame as a duck-typed
  result with .params / .std_errors / .tvalues / .pvalues /
  .conf_int_*. Pipes straight into sp.regtable, closing the
  "estimator → marginal-effects table" gap that previously
  required users to hand-bui
[…]
