### 2026-04-28 03f7437
feat(fast-fepois): wire Rust separation_mask into fepois pre-pass
Path A continuation: the single call to _drop_separation inside
fepois() now routes through _drop_separation_dispatcher, which
prefers the Rust statspai_hdfe.separation_mask entry point when the
loaded wheel exposes it (crate v0.6.0+). Falls back to the existing
_drop_separation NumPy reference otherwise.

The Rust path replaces the np.unique + np.isin O(n log n) passes
that dominated the Python wall (~50 ms / fepois on the medium
benchmark, n=1M, G1=100k) with an O(n × n_iter × K) bincount loop.

Surface unchanged:
- Public sp.
[…]
