### 2026-04-28 ed5eb49
docs(hdfe): record Path A round 1 — Rust separation pre-pass landed
Path A (post-Phase-B1 brainstorm) closure entry. Companion to the
"Phase A round 1" / "Phase B0 round 1" / "Phase B1 round 1" audit
trail.

Numbers (medium benchmark, n=1M, fe1=100k, fe2=1k):

  Phase 0  (Python np.bincount):                 2.61 s   4.08× fixest
  Phase A  (Rust scatter):                       2.45 s   3.83×
  Phase B0 (Rust sequential + cache):            1.441 s  2.25×
  Phase B1 (native Rust IRLS):                   0.880 s  1.37×
  Path A   (B1 + Rust separation pre-pass):      0.855 s  1.34×    ✅
  R fix
[…]
