### 2026-04-28 ef1a939
feat(output): Phase 23 — provenance 129/925 (did extras + mediation + principal_strat)
Added: sun_abraham, ddd, did_bcf, event_study, four_way_decomposition,
mediate_sensitivity, survivor_average_causal_effect.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
