### 2026-04-28 44c5ab5
feat(output): Phase 24 — provenance 135/925 (spatial models + bunching + selection)
Added: sar, sem, sdm, general_bunching, stepwise, lasso_select.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
