### 2026-04-28 30eec80
feat(output): Phase 25 — provenance 138/925 (cointegration + mr_heterogeneity)
Added: engle_granger, johansen, mr_heterogeneity.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
