### 2026-04-28 45fe46f
feat(output): regtable Round 4 — event_study_table, vcov= recompute, transpose
Three further additions on top of Rounds 1-3. **No numerical
changes** to any estimator; the vcov= recompute reuses the
fit-time X + residuals already stored on OLS results.

- sp.event_study_table(result, *, regex=None, label_fmt="t={t}",
                       include_reference=False)
  Adapter that turns an event-study fit into a regtable input.
  CausalResult fast path reads model_info['event_study'] (canonical
  shape produced by sp.event_study); regex path scans
  result.params.index for coefficient names like
[…]
