+++ src/statspai/postestimation/__init__.py
@@ -11,12 +11,13 @@ Equivalent to Stata's ``margins``, ``test``, ``lincom`` commands.
 """
 
 from .margins import (
-    margins, margins_table, marginsplot, margins_at, margins_at_plot,
-    contrast, pwcompare,
+    margins, margins_table, event_study_table, marginsplot, margins_at,
+    margins_at_plot, contrast, pwcompare,
 )
 from .hypothesis import test, lincom
 
 __all__ = [
-    'margins', 'margins_table', 'marginsplot', 'margins_at', 'margins_at_plot',
+    'margins', 'margins_table', 'event_study_table',
+    'marginsplot', 'margins_at', 'margins_at_plot',
     'contrast', 'pwcompare', 'test', 'lincom',
 ]
