+++ src/statspai/output/regression_table.py
@@ -55,6 +55,129 @@ from ._repro import build_repro_note
 _MULTI_SE_BRACKETS = (("[", "]"), ("{", "}"), ("⟨", "⟩"), ("«", "»"))
 
 
+def _hc_recompute_se(model_data: _ModelData, raw_result: Any, vcov: str) -> Optional[Dict[str, Any]]:
+    """Recompute HC0/HC1/HC2/HC3 SE/t/p/CI for an OLS-style result.
+
+    Returns ``{'std_errors', 'tvalues', 'pvalues', 'conf_int_lower',
+    'conf_int_upper'}`` keyed by variable name, or ``None`` when the
+    raw result lacks the required ``data_info`` keys (X / residuals).
+
+    Algorithm:
+
+    - HC0: Omega = diag(e_i^2)
+    - HC1: HC0 * n / (n - k)        (Stata ``robust`` default)
+    - HC2: Omega = diag(e_i^2 / (1 - h_ii))
+    - HC3: Omega = diag(e_i^2 / (1 - h_ii)^2)
+
+    The leverage diagonal ``h_ii`` is computed from
+    ``H_ij = X_ij · ((X'X)^{-1} X')_ji`` without forming the full
+    n×n hat matrix.
+    """
+    dinfo = getattr(raw_result, "data_info", None)
+    if not isinstance(dinfo, dict):
+        return None
+    X = dinfo.get("X")
+    e = dinfo.get("residuals")
+    if X is None or e is None:
+        return None
+    X_arr = np.asarray(X, dtype=float)
+    e_arr = np.asarray(e, dtype=float).ravel()
+    if X_arr.ndim != 2 or X_arr.shape[0] != e_arr.shape[0]:
+        return None
+    n, k = X_arr.shape
+    if n <= k:
+        return None
+    XtX_inv = np.linalg.pinv(X_arr.T @ X_arr)
+    if vcov.upper() in ("HC2", "HC3"):
+        # h_ii = sum_j X_ij * (X · XtX_inv)_ij
+        H_full = X_arr @ XtX_inv  # n×k
+        h_ii = np.einsum("ij,ij->i", X_arr, H_full)
+        h_ii = np.clip(h_ii, 0.0, 1.0 - 1e-10)
+    if vcov.upper() == "HC0":
+        omega = e_arr ** 2
+    elif vcov.upper() in ("HC1", "ROBUST"):
+        omega = (e_arr ** 2) * (n / (n - k))
+    elif vcov.upper() == "HC2":
+        omega = (e_arr ** 2) / (1.0 - h_ii)
+    elif vcov.upper() == "HC3":
+        omega = (e_arr ** 2) / ((1.0 - h_ii) ** 2)
+    else:
+        raise ValueError(
+            f"vcov={vcov!r} not supported for OLS recompute. Choose "
+            f"from 'HC0', 'HC1' (Stata robust), 'HC2', 'HC3', 'robust'."
+        )
+    XtOX = X_arr.T @ (omega[:, None] * X_arr)
+    V = XtX_inv @ XtOX @ XtX_inv
+    se_arr = np.sqrt(np.maximum(np.diag(V), 0.0))
+    var_names = list(model_data.params.index)
+    if len(var_names) != len(se_arr):
+        # var_cov ordering vs params ordering mismatch; use data_info if present
+        dvn = dinfo.get("var_names")
+        if dvn is not None and len(dvn) == len(se_arr):
+            var_names = list(dvn)
+        else:
+            return None
+    se_series = pd.Series(se_arr, index=var_names)
+    b = pd.Series(np.asarray(model_data.params, dtype=float), index=var_names)
+    with np.errstate(divide="ignore", invalid="ignore"):
[…]
