+++ tests/test_regtable_round4_extensions.py
@@ -1,0 +1,257 @@
+"""
+Round-4 publication extensions to ``regtable``.
+
+Three additions on top of Rounds 1-3:
+
+1. ``sp.event_study_table(result)`` — adapter that reads
+   ``model_info['event_study']`` (DiD CausalResult) or extracts
+   event-time coefs via regex from raw ``params.index``, returns a
+   regtable-renderable result with rows ordered by relative time.
+
+2. ``vcov="HC0" | "HC1" | "HC2" | "HC3" | "robust"`` — for OLS,
+   recompute SE / t / p / CI at ``regtable`` print time using the
+   stored X + residuals. Stata's ``robust`` ≈ HC1; we accept both
+   spellings. Non-OLS results raise an informative error.
+
+3. ``transpose=True`` — single-panel pivot. Rows become models,
+   columns become variables. Multi-panel / multi_se tables raise
+   to keep semantics tight.
+"""
+
+from __future__ import annotations
+
+import re
+import warnings
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+
+
+# ---------------------------------------------------------------------------
+# Fixtures
+# ---------------------------------------------------------------------------
+
+@pytest.fixture
+def ols_models():
+    rng = np.random.default_rng(2026)
+    n = 600
+    x1 = rng.normal(0, 1, n)
+    x2 = rng.normal(0, 1, n)
+    # Heteroskedastic so HC SEs differ from OLS SEs
+    y = 1.0 + 0.5 * x1 + 0.25 * x2 + (1 + np.abs(x1)) * rng.normal(0, 1, n)
+    df = pd.DataFrame({"y": y, "x1": x1, "x2": x2})
+    m1 = sp.regress("y ~ x1", data=df)
+    m2 = sp.regress("y ~ x1 + x2", data=df)
+    return m1, m2, df
+
+
+@pytest.fixture
+def event_study_result():
+    rng = np.random.default_rng(0)
+    n_units = 80
+    n_periods = 10
+    rows = []
+    for u in range(n_units):
+        treat_time = 5 if u < 40 else np.nan
+        for t in range(n_periods):
+            post = (
+                (t >= treat_time)
+                if not np.isnan(treat_time)
+                else False
+            )
+            y = (
+                (1.0 if post else 0.0) * 0.5
+                + rng.normal(0, 0.5)
+                + u * 0.02
+                + t * 0.1
+            )
+            rows.append({
+                "unit": u, "time": t, "treat_time": treat_time, "y": y,
+            })
+    df = pd.DataFrame(rows)
+    return sp.event_study(
+        df, y="y", treat_time="treat_time", time="time",
+        unit="unit", window=(-3, 3),
+    )
+
+
+# ---------------------------------------------------------------------------
+# 1. event_study_table
+# ---------------------------------------------------------------------------
+
+class TestEventStudyTable:
+
+    def test_event_study_table_exists(self):
+        assert hasattr(sp, "event_study_table"), \
+            "sp.event_study_table not registered"
+
+    def test_extracts_relative_time_rows(self, event_study_result):
+        et = sp.event_study_table(event_study_result)
+        # Rows for relative times -3..-1, 0..3 (ref_period -1 has zero estimate)
+        labels = list(et.params.index)
+        # Should contain entries spanning negative and non-negative times
+        assert any("-3" in s or "−3" in s for s in labels)
+        assert any("0" in s for s in labels)
+        assert any("3" in s and "-" not in s and "−" not in s for s in labels)
+
+    def test_event_study_table_renders(self, event_study_result):
+        et = sp.event_study_table(event_study_result)
[…]

[…]
