### 2026-04-28 25062e1
feat(output): Phase 26 — provenance 142/925 (ope + conformal counterfactual)
Added: sharp_ope_unobserved, direct_method, conformal_counterfactual,
conformal_ite_interval.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
