### 2026-04-28 0d062f8
fix(output): capture propensity_model in conformal_counterfactual/conformal_ite_interval provenance
Phase 26 instrumented these two estimators but only recorded `model`
(outcome regressor); the `propensity_model` kwarg silently dropped
from the params trail. Mirror the existing model capture so any
non-default propensity choice (e.g. user-passed LogisticRegression)
is preserved in the lineage record. No numerical path touched.
