### 2026-04-28 b4284cc
docs(release): 1.8.0 — consolidate [Unreleased] sections + README v1.8 highlights
CHANGELOG: rename 19 [Unreleased] headings (range L5..L1290) to [1.8.0],
date the canonical anchor 2026-04-28, normalize the v1.8.1-mislabeled
HDFE-completeness section to 1.8.0. Historical 0.9.3 [Unreleased] at
L4729 left intact.

README: add v1.8 highlights block above v1.7 (Native Rust IRLS / sp.prod_fn /
regtable Rounds 1-4 / provenance 142/925 / Output trinity / HTZ + LLM-DAG).

No code changes; release plumbing only.
