+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to StatsPAI will be documented in this file.
 
-## [Unreleased] — `sp.regtable` Round 4 (event_study_table, vcov= recompute, transpose)
+## [1.8.0] — `sp.regtable` Round 4 (event_study_table, vcov= recompute, transpose)
 
 Three further additions on top of Rounds 1-3. **No numerical
 changes** to any estimator; the ``vcov=`` recompute reuses the
@@ -53,7 +53,7 @@ guards on multi-panel / multi_se.
 577 targeted tests pass (Rounds 1-4 = 528 + 20 + 14 + 15, plus broad
 anchors). Zero regression.
 
-## [1.8.0] — Native Rust IRLS for `sp.fast.fepois` + production-function module
+## [1.8.0] — 2026-04-28 — Native Rust IRLS for `sp.fast.fepois` + production-function module
 
 The headline of v1.8.0 is the **3× wall-clock improvement** on the
 medium HDFE benchmark: `sp.fast.fepois` runs at **0.855 s** vs the
@@ -168,7 +168,7 @@ verified with a wall-clock spike before the next was committed
 
 ---
 
-## [Unreleased] — `sp.regtable` Round 3 (margins_table, tests= footer, fixef_sizes)
+## [1.8.0] — `sp.regtable` Round 3 (margins_table, tests= footer, fixef_sizes)
 
 Three further additions on top of Round 1 + Round 2. **No numerical
 changes** to any estimator (margins_table is a pure adapter; tests=
@@ -213,7 +213,7 @@ all three features across text / LaTeX / HTML renderers.
 anchors); zero regression on the 33 output / regression test files
 exercised.
 
-## [Unreleased] — `sp.regtable` Round 2 (templates, notation, apply_coef, escape, Word/Excel spanners)
+## [1.8.0] — `sp.regtable` Round 2 (templates, notation, apply_coef, escape, Word/Excel spanners)
 
 Five further additions on top of the Round 1 commit. **No numerical
 changes** to any estimator; output-layer only.
@@ -263,7 +263,7 @@ five features across text / LaTeX / HTML / Word / Excel renderers.
 
 548 targeted tests pass (Round 1's 528 + 20 new), zero regression.
 
-## [Unreleased] — `sp.regtable` publication-quality extensions
+## [1.8.0] — `sp.regtable` publication-quality extensions
 
 Five additions designed to close the remaining gap between
 ``sp.regtable`` and Stata ``esttab`` / R ``modelsummary`` /
@@ -318,7 +318,7 @@ all six format renderers (text / LaTeX / HTML / Markdown) plus the
 parameter validation paths. Existing 204 output-area tests
 unchanged.
 
-## [Unreleased] — Phase 12: provenance rollout to 66/925 (bounds + randomization + imputation)
+## [1.8.0] — Phase 12: provenance rollout to 66/925 (bounds + randomization + imputation)
 
 Continues the v1.7.2 provenance rollout. **No numerical changes** to
 any estimator. 5 estimators instrumented spanning bounds /
@@ -336,7 +336,7 @@ randomization inference / imputation. Coverage 61/925 → **66/925**.
 
 6 new (5 per-estimator + 1 multi-estimator integration). All pass.
 
-## [Unreleased] — Phase 11: provenance rollout to 61/925 (spatial + qte + bootstrap + conformal)
+## [1.8.0] — Phase 11: provenance rollout to 61/925 (spatial + qte + bootstrap + conformal)
 
 Continues the v1.7.2 provenance rollout. **No numerical changes** to
 any estimator. 7 estimators instrumented spanning spatial / quantile
[…]
