### 2026-04-28 21f9a74
fix(inference): pate provenance referenced undefined name 'data'
Phase 15 provenance hookup (b158807) used the boilerplate `data=data,`
without adapting to pate()'s actual parameter names — the function
takes data_experiment / data_target, not data. flake8 F821 broke the
v1.8.0 CI on every Linux/macOS matrix entry.

Fingerprint data_experiment (the analysis sample whose y/treatment/
covariates the estimator reads); data_target only contributes target
covariate distributions.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
