### 2026-04-29 0ad1ccd
feat(agent): 1.9.0 — agent-native API surface (12 modules, 422 tests)
Phase 1 — serialization & MCP polish:
- CausalResult/EconometricResults to_dict(detail="minimal"|"standard"|"agent")
- execute_tool error envelope: error_kind + error_payload for StatsPAIError
- statspai-mcp console script; per-function statspai://function/{name} resources
- typed JSON-RPC errors (-32002 / -32602); notifications/* silenced
- MCP-level detail= argument on tools/call

Phases 2-4 — workflow primitives:
- sp.audit(result)        — missing-evidence reviewer checklist (18 checks)
- sp.detect_design(data)  — heuris
[…]
