+++ src/statspai/smart/audit.py
@@ -1,0 +1,631 @@
+"""Reviewer-checklist audit of a fitted StatsPAI result.
+
+``sp.audit(result)`` returns the *missing-evidence* view of a result:
+which robustness / sensitivity / diagnostic checks a careful reviewer
+would expect for this estimator family — and which of those have
+already been run vs. are still missing on the result.
+
+Distinct from neighbouring methods:
+
+* :meth:`CausalResult.violations` — items already on ``model_info`` whose
+  values *fail* their threshold ("checked but failed").
+* :meth:`CausalResult.next_steps` — recommendations of what to *do next*,
+  oriented around action (export, robustness, alternative method).
+* :func:`statspai.smart.assumption_audit` — heavyweight: takes
+  ``(result, data)`` and *re-runs* statistical tests. ``audit`` is
+  pure introspection — it never re-runs anything, never touches data,
+  and runs in microseconds.
+
+The agent's mental model: ``audit`` answers "what evidence is missing
+for a reviewer to trust this estimate?"; ``assumption_audit`` answers
+"given the data, do the assumptions actually hold?".
+
+Returns a JSON-safe ``dict`` so MCP-mediated agents can branch on the
+``status`` field of each check without parsing prose.
+"""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Any, Dict, List, Optional, Sequence, Tuple
+
+from ..core._agent_summary import (
+    _as_float,
+    _safe_get,
+    # Threshold constants imported (not re-stated) so audit's verdict
+    # cannot drift from violations() when a future correctness fix
+    # updates a cutoff. Single source of truth for numerical thresholds.
+    _ESS_MIN,
+    _OVERLAP_MIN,
+    _PRETREND_ALPHA,
+    _RHAT_MAX,
+    _SMD_MAX,
+    _WEAK_IV_F,
+)
+
+
+# ====================================================================== #
+#  Check specification
+# ====================================================================== #
+
+
+# A single evidence path is a tuple of dict keys; ``EvidencePaths`` is
+# either one path or a tuple-of-tuples for estimators that store the
+# same diagnostic under different aliases (e.g. IV first-stage F lives
+# under ``first_stage_f`` *or* ``first_stage.f_stat`` *or*
+# ``weak_iv_f`` depending on which estimator wrote the result).
+EvidencePaths = Tuple[Tuple[str, ...], ...]
+
+
+@dataclass(frozen=True)
+class _Check:
+    """One reviewer-relevant check the result either has or lacks.
+
+    Attributes
+    ----------
+    name : str
+        Short identifier — agents branch on this.
+    question : str
+        Reviewer-style human-readable prompt.
+    applies_to : tuple[str, ...]
+        Method families (per :func:`statspai.core.next_steps._detect_family`)
+        for which this check is relevant.
+    evidence_paths : tuple[tuple[str, ...], ...]
+        Candidate paths of keys into ``result.model_info`` where the
+        evidence would live. A *tuple of paths* — the first one that
+        resolves wins — so estimators that store the same diagnostic
+        under different aliases all surface as ``passed``.
+    threshold : float | None
+        Numeric threshold used to decide passed vs. failed when the
+        evidence value is a number. ``None`` for existence-only checks.
+    compare : str
+        How to evaluate the value vs. the threshold:
+
+        - ``"greater_passes"`` — value > threshold → passed
+        - ``"less_passes"``    — value < threshold → passed
+        - ``"exists"``          — non-empty dict at the path → passed.
+                                   Strict-dict-only intentionally — a
+                                   bare ``False`` / ``0`` / ``""``
+                                   sentinel must NOT mark the check
+                                   passed. The estimator contract is
[…]

[…]
