### 2026-04-29 f64417c
fix(agent): 1.9.1 — MCP schema + JSON-RPC error polish
Two strict-schema MCP-client fixes on top of 1.9.0; no estimator
numerical paths changed.

- mcp_server: dataless tools (honest_did, sensitivity) no longer
  mark `data_path` as required in their tools/list schema. The
  property is still advertised as optional for clients that always
  send it, but `required` is now conditional on whether the
  underlying StatsPAI function consumes a DataFrame. New
  `_DATALESS_TOOLS` set in agent/mcp_server.py is the single source
  of truth (kept in sync with `agent/tools.TOOL_REGISTRY`).
- mcp_server: 
[…]
