+++ tests/test_v0917_deferred.py
@@ -468,6 +468,7 @@ class TestMCPServer:
 
     def test_tools_list_includes_data_path(self):
         import statspai as sp
+        from statspai.agent.mcp_server import _DATALESS_TOOLS
         req = json.dumps({
             "jsonrpc": "2.0", "id": 2, "method": "tools/list", "params": {},
         })
@@ -477,7 +478,15 @@ class TestMCPServer:
         for t in tools:
             schema = t["inputSchema"]
             assert "data_path" in schema["properties"]
-            assert "data_path" in schema["required"]
+            # Dataless tools (honest_did, sensitivity) advertise
+            # data_path as an OPTIONAL convenience but must not mark
+            # it required — strict-schema clients would otherwise
+            # refuse to dispatch the call without a CSV path the
+            # estimator never reads. See mcp_server._DATALESS_TOOLS.
+            if t["name"] in _DATALESS_TOOLS:
+                assert "data_path" not in schema["required"]
+            else:
+                assert "data_path" in schema["required"]
 
     def test_tools_call_runs_regress(self, mcp_csv):
         import statspai as sp
