### 2026-04-29 c4bbae0
perf(output): lazy-load openpyxl in outreg2 to cut import time
Top-level ``import openpyxl`` in ``outreg2.py`` transitively pulled
``PIL``/``Pillow`` via ``openpyxl.drawing.image`` on every
``import statspai as sp`` — even when the user never touches outreg2
(usage in repo: 4 references vs 163 for ``regtable``).

Move the import inside ``_export_with_formatting`` and drop unused
symbols (``Border``, ``Side``, ``PatternFill``, ``dataframe_to_rows``,
``write_title``).

- ``import statspai as sp`` cold start: 2066 ms → 1767 ms (-14.5%)
- Eagerly loaded heavy modules: ['openpyxl','PIL','PIL.Image'
[…]
