### 2026-04-29 517be41
feat(agent): module A+B+C — result handles, workflow tools, robust JSON encoding
Module A — quick-win polish:
- Auto-derive _DATALESS_TOOLS from registry (any spec without required
  data param), with manual override list for tools backed by stubs.
- Bump auto_tool_manifest cap 250→500 and warn loudly when truncated
  per CLAUDE.md §3 #7 ("失败要响亮").
- Loud warning when tool_manifest's auto-merge falls back to curated.
- _json_default covers np.bool_, np.complexfloating, np.datetime64,
  np.timedelta64, NaN/Inf→None, pd.Index/Timedelta/Categorical/Interval,
  set/frozenset, bytes (b64), Decimal,
[…]
