+++ tests/test_mcp_result_handle.py
@@ -1,0 +1,303 @@
+"""Tests for the result-handle / chained-workflow surface.
+
+Covers:
+* ``execute_tool(..., as_handle=True)`` caches the fitted object and
+  emits ``result_id`` / ``result_uri``.
+* ``statspai://result/<id>`` resource resolves to a JSON payload with
+  full provenance.
+* ``audit_result`` / ``brief_result`` consume a handle without re-running.
+* ``honest_did_from_result`` auto-extracts betas / sigma.
+* ``bibtex`` tool round-trips known + unknown keys.
+* LRU eviction.
+"""
+from __future__ import annotations
+
+import json
+import os
+
+import numpy as np
+import pandas as pd
+import pytest
+
+
+from statspai.agent._result_cache import RESULT_CACHE, ResultCache, CacheEntry
+from statspai.agent import (
+    execute_tool,
+    mcp_handle_request,
+)
+
+
+def _rpc(method, params=None, request_id=1):
+    msg = {"jsonrpc": "2.0", "id": request_id, "method": method,
+           "params": params or {}}
+    line = mcp_handle_request(json.dumps(msg))
+    return json.loads(line)
+
+
+# ----------------------------------------------------------------------
+# Result cache primitives
+# ----------------------------------------------------------------------
+
+class TestResultCache:
+    def test_put_and_get_round_trip(self):
+        cache = ResultCache(max_size=4)
+        rid = cache.put({"x": 1}, tool="did", arguments={"y": "wage"})
+        assert rid.startswith("r_")
+        assert cache.get(rid) == {"x": 1}
+        entry = cache.get_entry(rid)
+        assert isinstance(entry, CacheEntry)
+        assert entry.tool == "did"
+        assert entry.arguments == {"y": "wage"}
+
+    def test_lru_eviction(self):
+        cache = ResultCache(max_size=3)
+        ids = [cache.put({"i": i}, tool="t") for i in range(5)]
+        # Oldest two evicted; newest three remain.
+        assert cache.get(ids[0]) is None
+        assert cache.get(ids[1]) is None
+        for i in (2, 3, 4):
+            assert cache.get(ids[i]) == {"i": i}
+
+    def test_metadata_strips_dataframes(self):
+        df = pd.DataFrame({"y": [1, 2, 3]})
+        cache = ResultCache(max_size=2)
+        rid = cache.put({"hello": "world"}, tool="did",
+                        arguments={"data": df, "y": "wage", "post": True})
+        meta = cache.get_entry(rid).to_metadata()
+        assert "data" in meta["arguments"]
+        # DataFrame replaced by a placeholder string
+        assert isinstance(meta["arguments"]["data"], str)
+        assert meta["arguments"]["y"] == "wage"
+        assert meta["arguments"]["post"] is True
+
+
+# ----------------------------------------------------------------------
+# bibtex tool
+# ----------------------------------------------------------------------
+
+class TestBibtexTool:
+    def test_known_key_returns_entry(self):
+        out = execute_tool("bibtex", {"keys": ["abadie2003economic"]})
+        if not out["bibtex"]["abadie2003economic"]:
+            pytest.skip("paper.bib does not ship abadie2003economic in this build")
+        entry = out["bibtex"]["abadie2003economic"]
+        assert "@" in entry  # bib syntax
+        assert "abadie2003economic" in entry
+
+    def test_unknown_key_returns_empty_with_suggestions(self):
+        out = execute_tool("bibtex", {"keys": ["definitely_not_a_real_key_xyz123"]})
+        assert out["bibtex"]["definitely_not_a_real_key_xyz123"] == ""
+        assert "definitely_not_a_real_key_xyz123" in out["unknown_keys"]
+
[…]
