+++ src/statspai/agent/workflow_tools.py
@@ -1,0 +1,751 @@
+"""Hand-curated workflow / handle-based / citation tools.
+
+These are the "Tier-0" tools that close the agent feedback loop:
+
+* ``audit_result`` / ``brief_result`` / ``sensitivity_from_result`` /
+  ``honest_did_from_result`` — operate on a cached result handle
+  produced by an earlier tool call (``as_handle=True``). They eliminate
+  the LLM having to ferry back arrays and CSV paths between turns.
+* ``bibtex`` — return verified BibTeX entries from the project's
+  ``paper.bib`` (the single source of truth per CLAUDE.md §10). Closes
+  the citation-hallucination loophole.
+* ``audit`` / ``preflight`` / ``detect_design`` / ``brief`` — explicit
+  hand-curated wrappers for the smart-workflow primitives that the
+  prompt templates reference. Auto-tools used to surface these with
+  one-line descriptions; the bespoke schemas below give agents proper
+  signposting.
+
+Every workflow tool returns a dict shaped like the standard estimator
+serializer output (``estimate`` / ``method`` / ``next_calls`` / …) so
+the MCP layer doesn't need to special-case their content blocks.
+"""
+from __future__ import annotations
+
+from typing import Any, Dict, Iterable, List, Optional
+
+import pandas as pd
+
+from ._result_cache import RESULT_CACHE
+
+
+# ----------------------------------------------------------------------
+# Schema definitions surfaced via tool_manifest()
+# ----------------------------------------------------------------------
+
+def _result_id_schema(description: str) -> Dict[str, Any]:
+    return {
+        'type': 'object',
+        'properties': {
+            'result_id': {
+                'type': 'string',
+                'description': description,
+            },
+        },
+        'required': ['result_id'],
+    }
+
+
+WORKFLOW_TOOL_SPECS: List[Dict[str, Any]] = [
+    # ------------------------------------------------------------------
+    # Handle-based extensions to the curated tools — break the
+    # "LLM ferries arrays" anti-pattern.
+    # ------------------------------------------------------------------
+    {
+        'name': 'audit_result',
+        'description': (
+            "Reviewer-grade audit on a previously-fitted result. Pass "
+            "the result_id returned by an earlier tool call (with "
+            "as_handle=true). Returns the same checklist sp.audit() "
+            "produces — every robustness check the literature expects "
+            "for the design, with status='present|missing|run' and "
+            "concrete suggested_function names for the missing ones."
+        ),
+        'input_schema': _result_id_schema(
+            "Handle returned by an earlier estimator call. Must be in "
+            "the server result cache (LRU-evicted; refit if missing)."
+        ),
+    },
+    {
+        'name': 'brief_result',
+        'description': (
+            "Return the one-line agent-friendly brief for a fitted "
+            "result. Uses sp.brief(). Useful when an agent wants to "
+            "summarise a chained workflow without paying for the full "
+            "JSON payload again."
+        ),
+        'input_schema': _result_id_schema(
+            "Handle to a previously-fitted result."
+        ),
+    },
+    {
+        'name': 'sensitivity_from_result',
+        'description': (
+            "Run sp.sensitivity / sp.evalue / sp.oster_bounds / "
+            "sp.sensemakr on a cached result. Pass method='evalue' "
+            "(default) for the omitted-confounder-strength bound, "
+            "'oster' for delta/R-max, 'cinelli_hazlett' for OVB bounds."
+        ),
+        'input_schema': {
+            'type': 'object',
+            'properties': {
+                'result_id': {
+                    'type': 'string',
+                    'description': "Handle to a fitted causal result.",
+                },
+                'method': {
+                    'type': 'string',
[…]
