### 2026-04-29 7d219c2
refactor(output): consolidate numeric / star formatters into _format.py
Four backends (regtable / esttab / modelsummary / outreg2) historically
shipped their own ``_format_stars`` / ``_fmt_val`` / ``_fmt_int`` /
``_fmt_auto`` implementations — most byte-for-byte equivalent, plus a
few divergent ones with no documentation of why.

Promote the canonical helpers to a single ``output/_format.py`` and
delegate from the three legacy locations:

- ``estimates._format_stars/_fmt_val/_fmt_int/_fmt_auto`` → re-export
  from ``_format`` under the legacy underscore names so existing
  ``regression_table``
[…]
