### 2026-04-29 8249d9f
feat(agent): module D+E — output enrichment + image content blocks
Module D — output enrichment:
- New _enrichment module with build_next_calls / build_citations /
  fetch_bibtex / build_narrative / enrich_payload.
- Every tool return now carries:
  * next_calls — pre-built tools/call payloads with result_id +
    forwarded base args; agents copy-paste verbatim.
  * citations — verified bib keys from a static map (empty list ⇒
    intentionally absent, never invent) plus BibTeX bodies pulled
    from paper.bib via fetch_bibtex.
  * narrative — short markdown digest (method, point estimate, CI,
[…]
