+++ src/statspai/agent/workflow_tools.py
@@ -230,6 +230,42 @@ WORKFLOW_TOOL_SPECS: List[Dict[str, Any]] = [
     # Citation tool — the kill-switch for citation hallucination.
     # ------------------------------------------------------------------
     {
+        'name': 'plot_from_result',
+        'description': (
+            "Render the canonical diagnostic plot for a fitted result "
+            "and return it as an inline PNG image content block. "
+            "MCP clients with vision (Claude Desktop, vision-capable "
+            "agents) get the plot for free; clients that don't support "
+            "image content see only the JSON metadata. Plot kind is "
+            "auto-selected from the result type: event-study for DID, "
+            "rdplot for RD, gap plot for synth, balance plot for "
+            "matching, ROC for classification, etc."
+        ),
+        'input_schema': {
+            'type': 'object',
+            'properties': {
+                'result_id': {
+                    'type': 'string',
+                    'description': "Handle to a fitted result.",
+                },
+                'kind': {
+                    'type': 'string',
+                    'description': (
+                        "Override the auto-detected plot kind. "
+                        "Common values: 'event_study', 'rdplot', "
+                        "'synth_gap', 'love_plot', 'coef_plot'."
+                    ),
+                },
+                'figsize': {
+                    'type': 'array',
+                    'items': {'type': 'number'},
+                    'description': "Width, height in inches (default [8,5]).",
+                },
+            },
+            'required': ['result_id'],
+        },
+    },
+    {
         'name': 'bibtex',
         'description': (
             "Return verified BibTeX entries from paper.bib (StatsPAI's "
@@ -300,6 +336,9 @@ def execute_workflow_tool(
 
     if name == 'bibtex':
         return _tool_bibtex(arguments)
+
+    if name == 'plot_from_result':
+        return _tool_plot_from_result(rid_arg, arguments)
 
     if name == 'detect_design':
         return _tool_detect_design(arguments, data, detail=detail,
@@ -448,11 +487,23 @@ def _tool_sensitivity_from_result(rid: Optional[str],
     if not isinstance(out, dict):
         out = {'value': out}
     out['source_result_id'] = rid
+    new_rid: Optional[str] = None
     if as_handle:
         new_rid = RESULT_CACHE.put(result, tool='sensitivity_from_result',
                                      arguments={'source': rid, 'method': method})
         out['result_id'] = new_rid
         out['result_uri'] = f"statspai://result/{new_rid}"
+    from ._enrichment import enrich_payload
+    # Enrichment uses the underlying sensitivity method as the tool key
+    # (evalue / oster / cinelli_hazlett / sensitivity) so citations point
+    # to the correct paper.
+    enrich_key = method if method in {'evalue', 'oster_bounds',
+                                       'sensemakr', 'sensitivity'} else 'sensitivity'
+    if method == 'oster':
+        enrich_key = 'oster_bounds'
+    elif method == 'cinelli_hazlett':
+        enrich_key = 'sensemakr'
+    enrich_payload(out, tool_name=enrich_key, result_id=new_rid)
     return out
 
 
@@ -504,11 +555,14 @@ def _tool_honest_did_from_result(rid: Optional[str],
     out['source_result_id'] = rid
     out['extracted_n_pre'] = int(n_pre)
     out['extracted_n_post'] = int(n_post)
+    new_rid: Optional[str] = None
     if as_handle:
         new_rid = RESULT_CACHE.put(result, tool='honest_did_from_result',
                                      arguments={'source': rid, 'method': method})
         out['result_id'] = new_rid
[…]
