### 2026-04-29 509aa7f
fix(iv): make sp.iv() callable + unify 25+ method aliases
Phase 1 of the v1.10 API consistency pass.

The bug: sp.iv("y ~ (d ~ z)", data=df) raised TypeError:
'module' object is not callable despite being advertised in the
registry, agent summaries, MCP server docs, replication examples,
and the live call in question/question.py:505. Cause: line 45 of
statspai/__init__.py imported the function iv from regression.iv,
but later "from .iv.kernel_iv import ..." triggered a subpackage
load that reattached statspai.iv to the subpackage — silently
shadowing the function.

The fix: install a callable 
[…]
