+++ tests/test_iv_dispatcher.py
@@ -1,0 +1,208 @@
+"""Tests for the unified ``sp.iv`` dispatcher.
+
+These tests exercise the callable-module pattern installed in
+``statspai.iv.__init__`` and the ``method=`` table — they do *not*
+re-test numerical correctness of the underlying estimators, which is
+covered by the per-method test files (``test_iv.py`` etc.) and the
+reference-parity suite.
+
+The point: prevent regression of the v1.10 ``sp.iv()`` unification —
+prior to v1.10 ``sp.iv("y ~ (d ~ z)", data=df)`` silently raised
+``TypeError: 'module' object is not callable`` despite being advertised
+in the registry, agent summaries, and `question/question.py`.
+"""
+from __future__ import annotations
+
+import warnings
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+
+
+# ─── Fixtures ───────────────────────────────────────────────────────────
+
+
+@pytest.fixture(scope="module")
+def iv_data():
+    """A small over-identified IV DGP with one endogenous regressor."""
+    rng = np.random.default_rng(0)
+    n = 400
+    z1 = rng.standard_normal(n)
+    z2 = rng.standard_normal(n)
+    x1 = rng.standard_normal(n)
+    d = 0.6 * z1 + 0.5 * z2 + 0.3 * x1 + 0.5 * rng.standard_normal(n)
+    y = 1.0 + 2.0 * d + 0.5 * x1 + rng.standard_normal(n)
+    return pd.DataFrame({"y": y, "d": d, "z1": z1, "z2": z2, "x1": x1})
+
+
+# ─── Callable-module surface ────────────────────────────────────────────
+
+
+def test_sp_iv_is_callable():
+    """The headline regression: ``sp.iv`` must be callable, not a bare module."""
+    assert callable(sp.iv), (
+        "sp.iv must be callable.  If this fails, the callable-module "
+        "shim in statspai/iv/__init__.py was clobbered."
+    )
+
+
+def test_sp_iv_keeps_submodule_access():
+    """Callable trick must not break ``sp.iv.kernel_iv``-style lookup."""
+    assert callable(sp.iv.fit)
+    assert callable(sp.iv.kernel_iv)
+    assert callable(sp.iv.bayesian_iv)
+    assert callable(sp.iv.npiv)
+    assert callable(sp.iv.ivdml)
+    assert callable(sp.iv.mte)
+    # diagnostics are re-exports, also reachable
+    assert callable(sp.iv.anderson_rubin_test)
+    assert callable(sp.iv.kleibergen_paap_rk)
+
+
+def test_sp_iv_default_is_2sls(iv_data):
+    """Calling without ``method=`` should be 2SLS."""
+    r = sp.iv("y ~ (d ~ z1 + z2) + x1", data=iv_data)
+    # coefficient should be near the true value of 2.0
+    assert abs(r.params["d"] - 2.0) < 0.2
+
+
+def test_fit_alias_matches_dispatcher(iv_data):
+    """``sp.iv.fit(...)`` must equal ``sp.iv(...)`` for the same args."""
+    r1 = sp.iv("y ~ (d ~ z1 + z2) + x1", data=iv_data, method="liml")
+    r2 = sp.iv.fit("y ~ (d ~ z1 + z2) + x1", data=iv_data, method="liml")
+    assert r1.params["d"] == pytest.approx(r2.params["d"])
+
+
+# ─── Method coverage ────────────────────────────────────────────────────
+
+
+@pytest.mark.parametrize("method", ["2sls", "liml", "fuller", "gmm", "jive"])
+def test_kclass_methods(iv_data, method):
+    r = sp.iv("y ~ (d ~ z1 + z2) + x1", data=iv_data, method=method)
[…]
