+++ src/statspai/iv/__init__.py
@@ -3,51 +3,72 @@ statspai.iv — the unified Instrumental Variables namespace.
 
 The goal of this subpackage is to be the single entry point for every
 IV-flavoured workflow in StatsPAI, regardless of which sub-module the
-underlying implementation lives in:
+underlying implementation lives in.
 
-- Core point estimators (2SLS, LIML, Fuller, GMM, JIVE) live in
-  :mod:`statspai.regression.iv` — re-exported as ``sp.iv.iv``,
-  ``sp.iv.ivreg`` and ``sp.iv.IVRegression``.
-- JIVE variants (UJIVE, IJIVE, RJIVE) live here in
-  :mod:`statspai.iv.jive_variants`.
-- Weak-identification diagnostics (Olea-Pflueger effective F, Lee-McCrary
-  tF, Anderson-Rubin CI) live in :mod:`statspai.diagnostics.weak_iv`
-  — re-exported as ``sp.iv.effective_f_test`` etc.
-- New diagnostics introduced in this subpackage:
-  ``kleibergen_paap_rk``, ``sanderson_windmeijer``, ``conditional_lr_test``.
-- Plausibly exogenous sensitivity (Conley-Hansen-Rossi 2012):
-  ``plausibly_exogenous_uci``, ``plausibly_exogenous_ltz``.
-- Marginal Treatment Effects (Brinch-Mogstad-Wiswall 2017):
-  ``mte``.
-- Shift-share IV (Bartik + Adão-Kolesár-Morales correction) is re-exported
-  as ``sp.iv.bartik`` / ``sp.iv.shift_share_se``.
-- DeepIV (Hartford et al. 2017) is re-exported as ``sp.iv.deepiv``.
+The subpackage itself is **callable**::
 
-A thin :func:`fit` dispatcher ties everything together and auto-runs an
-expanded diagnostic panel (first-stage F, MOP effective F, KP rk,
-SW per-endog F, Hansen J, AR Wald).
+    sp.iv("y ~ (d ~ z) + x", data=df)                   # 2SLS (default)
+    sp.iv("y ~ (d ~ z) + x", data=df, method="liml")    # LIML
+    sp.iv(method="kernel", y=..., endog=..., instruments=..., data=df)
+    sp.iv.fit(...)         # equivalent (fit = _dispatch alias)
+    sp.iv.kernel_iv(...)   # individual estimators still reachable
+
+That callable is provided by a tiny ``ModuleType`` subclass installed via
+``sys.modules[__name__].__class__`` (the standard PEP 562-style trick).
+
+Sub-method coverage (``method=`` keyword, all aliases lowercased):
+
+- **K-class formula path** (``regression.iv``):
+  ``2sls`` / ``tsls`` / ``iv``,  ``liml``,  ``fuller``,  ``gmm``,  ``jive``.
+- **Modern JIVE variants** (``iv.jive_variants``):
+  ``jive1``, ``ujive``, ``ijive``, ``rjive``.
+- **Many-weak** (``iv.many_weak``):
+  ``jive_mw``, ``many_weak_ar``.
+- **Lasso / post-Lasso**:
+  ``lasso`` (``regression.advanced_iv.lasso_iv``),
+  ``post_lasso`` / ``bch`` (``iv.post_lasso.bch_post_lasso_iv``).
+- **ML / nonparametric**:
+  ``kernel`` (``iv.kernel_iv``), ``npiv`` (``iv.npiv``),
+  ``ivdml`` (``iv.ivdml``), ``deepiv`` (``deepiv.deepiv``, optional).
[…]

[…]
