### 2026-04-29 1e7f105
refactor(output): split MeanComparisonResult out of regression_table.py
``regression_table.py`` had grown to 3,335 lines holding two unrelated
result classes — ``RegtableResult`` (1,989 lines) and
``MeanComparisonResult`` (510 lines). They share no logic; the latter
is a balance-table renderer with its own to_text / to_html / to_latex
/ to_markdown / to_excel / to_word / to_dataframe / save pipeline.

- Move ``MeanComparisonResult`` and the public ``mean_comparison()``
  API to a new ``output/mean_comparison.py``.
- Re-export both from ``regression_table`` so existing
  ``from statspai.output.
[…]
