+++ src/statspai/output/regression_table.py
@@ -2821,515 +2821,11 @@ def regtable(
 
 
 # ═══════════════════════════════════════════════════════════════════════════
-# MeanComparisonResult
+# ``MeanComparisonResult`` / ``mean_comparison`` moved to
+# ``mean_comparison.py`` (was 510 lines wedged at the tail of this file).
+# Re-export here so existing
+# ``from statspai.output.regression_table import MeanComparisonResult``
+# imports keep working unchanged.
 # ═══════════════════════════════════════════════════════════════════════════
 
-class MeanComparisonResult:
-    """Rich result object for balance / mean comparison tables."""
-
-    def __init__(
-        self,
-        data: pd.DataFrame,
-        variables: List[str],
-        group: str,
-        group_labels: Tuple[str, str],
-        test: str,
-        fmt: str,
-        title: str,
-        weights: Optional[str],
-        output: str = "text",
-    ):
-        self.variables = variables
-        self.group = group
-        self.group_labels = group_labels
-        self.test = test
-        self.fmt = fmt
-        self.title = title
-        self.weights = weights
-        self._output = output
-
-        # Compute all stats
-        self._compute(data)
-
-    def _compute(self, data: pd.DataFrame) -> None:
-        g0_label, g1_label = self.group_labels
-        mask0 = data[self.group] == 0
-        mask1 = data[self.group] == 1
-
-        # If not binary 0/1, use the two unique values
-        unique_vals = sorted(data[self.group].dropna().unique())
-        if len(unique_vals) == 2:
-            mask0 = data[self.group] == unique_vals[0]
-            mask1 = data[self.group] == unique_vals[1]
-
-        d0 = data.loc[mask0]
-        d1 = data.loc[mask1]
-        self.n0 = len(d0)
-        self.n1 = len(d1)
-
-        rows: List[Dict[str, Any]] = []
-        for var in self.variables:
-            v0 = d0[var].dropna()
-            v1 = d1[var].dropna()
-
-            mean0, sd0 = v0.mean(), v0.std()
-            mean1, sd1 = v1.mean(), v1.std()
-            diff = mean1 - mean0
-
-            # Test
-            if self.test == "ranksum":
-                try:
-                    stat, pval = sp_stats.mannwhitneyu(v0, v1, alternative="two-sided")
-                except Exception:
-                    pval = np.nan
-            elif self.test == "chi2":
-                try:
-                    ct = pd.crosstab(data[self.group], data[var])
-                    chi2, pval, _, _ = sp_stats.chi2_contingency(ct)
-                except Exception:
-                    pval = np.nan
-            else:
-                # Default: Welch t-test
-                try:
-                    stat, pval = sp_stats.ttest_ind(v0, v1, equal_var=False)
-                except Exception:
-                    pval = np.nan
-
-            stars = _format_stars(pval, (0.10, 0.05, 0.01))
-
-            rows.append({
-                "variable": var,
-                "mean0": mean0,
-                "sd0": sd0,
-                "mean1": mean1,
-                "sd1": sd1,
-                "diff": diff,
-                "pvalue": pval,
-                "stars": stars,
-            })
-
-        self._rows = rows
-
-    # ═══════════════════════════════════════════════════════════════════════
-    # TEXT
-    # ═══════════════════════════════════════════════════════════════════════
-
-    def to_text(self) -> str:
-        g0_label, g1_label = self.group_labels
[…]
