### 2026-04-29 930c9e3
docs(output): organise __init__ exports by purpose, document canonical entry
Group ``output/__init__.py`` imports and ``__all__`` by role
(regression-table renderers / single-table helpers / multi-table
bundles / provenance / bibliography / adapters) and document in the
module docstring that ``regtable`` is the canonical regression-table
renderer — ``esttab`` / ``modelsummary`` / ``outreg2`` are
Stata/R-compatibility surfaces.

Also switches ``mean_comparison`` / ``MeanComparisonResult`` to import
directly from ``mean_comparison.py`` (its real home) rather than the
re-export hop through ``regr
[…]
