### 2026-04-29 a7cb8e8
refactor(init): dedupe redundant regression imports in top-level __init__.py
Phase 2 of the v1.10 API consistency pass.

The lines 245-247 (regression block) and 495-501 (NEW MODULES v0.6
block) were re-importing the same names: glm, GLMEstimator, logit,
probit, cloglog, poisson, nbreg, ppmlhdfe. Python silently rebinds
on the second import — no functional impact, but the redundancy was
confusing and made the file harder to maintain (it suggested two
groups had separate ownership when they didn't).

Removed the second binding; same 973 registered functions, all
top-level lookups (sp.glm, sp.lo
[…]
