### 2026-04-29 cd1166a
docs: PR-A output cleanup CHANGELOG + PR-B consolidation RFC
CHANGELOG: document the four PR-A commits as one cohesive Performance
+ Changed block (cold-start -19%, formatter consolidation, mean_comp
extraction, __init__ reorganisation).

RFC ``docs/rfc/output_pr_b_consolidation.md``: the deeper consolidation
work that needs a deprecation cycle — TableModel + renderer split for
``RegtableResult``, facade conversion of ``esttab`` / ``modelsummary``
/ ``outreg2``, application to ``MeanComparisonResult`` and
``CausalResult.to_latex``. Migration plan with snapshot-test gating
across 7 sub-PRs and 
[…]
