+++ docs/rfc/output_pr_b_consolidation.md
@@ -1,0 +1,202 @@
+# RFC: Output module consolidation (PR-B)
+
+**Status**: draft, ready for review
+**Author**: Claude (under Bryce's supervision)
+**Date**: 2026-04-29
+**Companion**: PR-A is already merged (commits `c4bbae0`...`930c9e3`).
+
+## 0. Why this RFC exists
+
+PR-A took every safe, no-deprecation cut on the output module:
+
+| Module | Change | Outcome |
+|---|---|---|
+| 1 | Lazy-load `openpyxl` in `outreg2.py` (drags in PIL transitively) | `import statspai` cold-start: **2,066 ms → 1,306 ms** (-37%); zero heavy modules eagerly loaded at top level |
+| 3 | Promote canonical formatters to `output/_format.py`, delegate from `outreg2` / `modelsummary` / `estimates` | Killed ~80 lines of duplicate `_format_stars` / `_fmt_val` / `_fmt_int`; fixed dead-code bug in `modelsummary._stars_str` |
+| 4 | Split `MeanComparisonResult` into `output/mean_comparison.py` | `regression_table.py`: **3,335 → 2,831 lines** (-15%); back-compat re-export preserved |
+| 5 | Group `output/__init__.py` by purpose, document canonical entry | Doc clarity; `sp.list_functions()` unchanged at 973 |
+
+Net: **522 passing tests** unchanged, no behavior regressions, zero deprecations.
+
+PR-B contains the cuts that need a deprecation cycle and careful design — items that risk subtly changing rendered output (numbers / formatting / row ordering) if rushed. Per CLAUDE.md §12: *"不要悄悄改现有估计器的数值输出"*.
+
+## 1. The two pieces of bloat that remain
+
+### 1.1 Four parallel regression-table backends
+
+| Backend | Public entry | Lines | Real usage (docs+tests grep) |
+|---|---|---|---|
+| `regtable` (RegtableResult) | `sp.regtable(*models)` | ~1,989 | **163** |
+| `esttab` (EstimateTable) | `sp.eststo(r); sp.esttab()` | ~1,025 | 6 |
+| `modelsummary` | `sp.modelsummary([r])` | ~845 | 7 |
+| `outreg2` (OutReg2) | `sp.outreg2(*r, filename=...)` | ~811 | 4 |
+
+Each one carries its own:
+
+- coef / SE / pvalue extractor (`_extract_model_data`, `_extract_coefs`, `OutReg2._create_regression_table`),
+- significance / number formatters — partly fixed in PR-A,
+- text / LaTeX / HTML / Markdown / Excel / Word renderers.
+
+Bug surface multiplies by 4. The user research is done — `regtable` is the de facto standard. The other three are **Stata/R muscle-memory facades** worth keeping for ergonomics, but they should be ~50-line wrappers, not 800-line reimplementations.
+
+### 1.2 `RegtableResult` is a 1,989-line god class
+
+[`output/regression_table.py:368-2208`](../../src/statspai/output/regression_table.py#L368-L2208) houses:
+
+- `_extract_model_data` (coef extraction)
+- `_resolve_vars` / `_resolve_stat_keys` (row/column resolution)
+- `_format_marker` / `_apply_coef_transform` / `_build_cell_context` (cell shaping)
+- `to_text` + `_to_text_transposed`
+- `to_html` + `_to_html_transposed`
+- `to_latex`
+- `to_markdown` + `to_quarto`
+- `to_dataframe`
+- `to_excel`
+- `to_word`
+- `_render` / `_repr_html_`
+
+Any time we add a feature (e.g., transposed mode, multi-row SE, eform), every renderer grows. This is the textbook case for **separating model from view**.
+
[…]
