### 2026-04-29 c91dc4b
feat(rd): make sp.rd() callable + 18-method dispatcher
Phase 3 of the v1.10 API consistency pass.

Mirrors the IV pattern: install a _CallableRDModule shim on
statspai.rd via PEP 562 __class__ swap so users can write
sp.rd(data, y, x, c) for the default rdrobust path or
sp.rd(..., method="honest") for any of the 18 estimator variants.
Submodule access (sp.rd.rdplot, sp.rd.rdsummary, sp.rd.rdrobust)
keeps working unchanged.

method= coverage:

  local-poly:    rdrobust/default/rd/robust/local_poly (CCT 2014)
  honest:        honest/armstrong_kolesar/ak
  randinf:       randinf/random/local_rand
[…]
