+++ src/statspai/rd/__init__.py
@@ -65,7 +65,252 @@ from ._aliases import (
     multi_cutoff_rd, geographic_rd, boundary_rd, multi_score_rd,
 )
 
+# ═══════════════════════════════════════════════════════════════════════
+#  Unified dispatcher — sp.rd(..., method=...)
+# ═══════════════════════════════════════════════════════════════════════
+#
+# Mirrors the IV pattern: ``sp.rd`` is the callable subpackage so
+# users can write ``sp.rd(data, y, x, c)`` for the default rdrobust
+# path, or ``sp.rd(..., method="honest")`` for any of the 18 estimator
+# variants below — without giving up ``sp.rd.rdplot`` and friends.
+#
+# Diagnostics-only functions (``rdbwselect``, ``rdbwsensitivity``,
+# ``rdbalance``, ``rdplacebo``, ``rdsummary``, ``rdplotdensity``,
+# ``rdpower``, ``rdsampsi``, ``rdwinselect``, ``rdsensitivity``,
+# ``rdrbounds``) are intentionally NOT in the ``method=`` table —
+# they are not estimators of treatment effects.
+
+import sys as _sys
+from types import ModuleType as _ModuleType
+from typing import Any as _Any, Dict as _Dict
+
+
+_RD_METHOD_ALIASES: _Dict[str, str] = {
+    # Local-polynomial bias-corrected (default, CCT 2014)
+    "rdrobust": "rdrobust", "local_poly": "rdrobust",
+    "default": "rdrobust", "rd": "rdrobust", "robust": "rdrobust",
+
+    # Honest CIs (Armstrong-Kolesar 2018, 2020)
+    "honest": "honest", "armstrong_kolesar": "honest", "ak": "honest",
+
+    # Local randomization (Cattaneo-Titiunik-VB 2016)
+    "randinf": "randinf", "random": "randinf",
+    "local_randomization": "randinf", "rdrandinf": "randinf",
+
+    # CATE / heterogeneous effects (Calonico et al. 2025)
+    "hte": "hte", "rdhte": "hte", "cate": "hte",
+
+    # ML + RD
+    "forest": "forest", "causal_forest": "forest", "rd_forest": "forest",
+    "boost": "boost", "gbm": "boost", "rd_boost": "boost",
+    "lasso": "lasso", "rd_lasso": "lasso",
+
+    # Bayesian HTE
+    "bayes_hte": "bayes_hte", "bayes": "bayes_hte",
+    "rd_bayes_hte": "bayes_hte",
+
+    # 2D / boundary RD
+    "rd2d": "rd2d", "2d": "rd2d", "boundary": "rd2d",
+
+    # Multi-cutoff
+    "rdmc": "rdmc", "multi_cutoff": "rdmc",
+
+    # Multi-score
+    "rdms": "rdms", "multi_score": "rdms", "geographic": "rdms",
+
+    # Kink (RKD)
+    "rkd": "rkd", "kink": "rkd",
+
+    # RD in time
+    "rdit": "rdit", "time": "rdit",
+
+    # Extrapolation
+    "extrapolate": "extrapolate", "rd_extrapolate": "extrapolate",
+    "multi_extrapolate": "multi_extrapolate",
+    "rd_multi_extrapolate": "multi_extrapolate",
+
+    # Spillover / interference
+    "interference": "interference", "spillover": "interference",
+    "rd_interference": "interference",
+
+    # Distributional treatment effects
+    "distribution": "distribution", "distributional": "distribution",
+    "rd_distribution": "distribution",
+    "distributional_design": "distributional_design",
+    "rd_distributional_design": "distributional_design",
+
+    # External validity
+    "external_validity": "external_validity",
+    "rd_external_validity": "external_validity",
+}
+
+
[…]
