### 2026-04-29 f2460f4
release(agent): 1.10.0 — pipelines, prompts, version bump, CHANGELOG
Module F — composite pipelines:
- New pipeline_tools module: pipeline_did / pipeline_iv / pipeline_rd
  run preflight + estimator + audit + sensitivity + brief in one call,
  return a markdown narrative + cached result_id + per-stage status.
- pipeline_did dispatches callaway_santanna when cohort+id supplied,
  classic 2x2 did otherwise. Auto-runs honest_did sensitivity when
  the result exposes event-study betas/sigma. Adds bacon_decomposition
  for staggered TWFE.
- pipeline_iv runs ivreg + effective_F + Anderson-Rubin + e-
[…]
