+++ src/statspai/agent/mcp_server.py
@@ -927,86 +927,246 @@ def _handle_resources_read(params: Dict[str, Any]) -> Dict[str, Any]:
 _PROMPTS: List[Dict[str, Any]] = [
     {
         "name": "audit_did_result",
-        "description": "Run a DID estimator on a CSV, surface the "
-                       "estimate, and walk through every "
-                       "reviewer-checklist gap (parallel-trends test, "
-                       "honest-DID sensitivity, Bacon decomposition, "
-                       "placebo).",
+        "description": ("Run a DID estimator on a CSV, surface the "
+                         "estimate, and walk through every "
+                         "reviewer-checklist gap. Uses pipeline_did "
+                         "to consolidate preflight + estimate + audit "
+                         "+ honest-DID + Bacon into one call."),
         "arguments": [
-            {"name": "data_path",
-             "description": "Absolute path to a CSV file.",
-             "required": True},
-            {"name": "y", "description": "Outcome column.",
-             "required": True},
-            {"name": "treat",
-             "description": "Binary 0/1 treatment column.",
-             "required": True},
-            {"name": "time", "description": "Time column.",
-             "required": True},
+            {"name": "data_path", "required": True,
+             "description": "Absolute path to the data file."},
+            {"name": "y", "required": True, "description": "Outcome column."},
+            {"name": "treat", "required": True,
+             "description": "Binary 0/1 treatment indicator."},
+            {"name": "time", "required": True, "description": "Time column."},
         ],
         "_template": (
-            "1. Call ``preflight`` with method='did' and the supplied "
-            "columns; if the verdict is FAIL, stop and report the "
-            "failed checks.\n"
-            "2. Call the ``did`` tool with data_path={data_path}, "
-            "y={y}, treat={treat}, time={time}.\n"
-            "3. Call ``audit`` on the result and list every check "
-            "with status='missing' AND importance='high'. For each, "
-            "call the function in ``suggest_function``.\n"
-            "4. Summarise: estimate, 95% CI, the violations table, "
-            "and the BibTeX citation."
+            "Call `pipeline_did` with data_path={data_path}, y={y}, "
+            "treat={treat}, time={time}. The pipeline returns a "
+            "markdown narrative with the canonical reviewer-grade "
+            "DID workflow already executed (preflight, estimator, "
+            "audit, honest-DID, Bacon decomposition, brief). Quote "
+            "the narrative verbatim; for any high-importance check "
+            "the audit flagged as missing, dispatch the corresponding "
+            "entry in `next_calls`. End with a `bibtex` lookup of the "
+            "keys in `citations.keys` so the user gets verified "
+            "references."
+        ),
+    },
+    {
+        "name": "audit_iv_result",
+        "description": ("End-to-end IV workflow: 2SLS + first-stage F + "
+                         "Anderson-Rubin CI + e-value sensitivity, all "
+                         "wrapped in pipeline_iv."),
+        "arguments": [
+            {"name": "data_path", "required": True,
+             "description": "Absolute path to the data file."},
+            {"name": "formula", "required": True,
+             "description": "'y ~ x + (d ~ z)' Wilkinson-style."},
+        ],
+        "_template": (
+            "Call `pipeline_iv` with data_path={data_path}, "
+            "formula='{formula}'. Read `effective_F` from the "
+            "response: < 10 means the Staiger-Stock weak-IV threshold "
+            "is breached and you should foreground the "
+            "Anderson-Rubin CI in your reply (it is in the "
+            "`anderson_rubin` field) instead of the 2SLS point "
+            "estimate. Cite via `bibtex(keys=...)` from the "
+            "`citations.keys` list."
+        ),
+    },
+    {
+        "name": "audit_rd_result",
[…]
