### 2026-04-29 7bb0713
feat(matching): expand sp.match method= to full matching toolkit
Phase 4 of the v1.10 API consistency pass.

sp.match was already a function with built-in method= for
classical algorithms (nearest/stratify/cem/psm/mahalanobis);
v1.10 expands the dispatcher so a single entry point reaches
every matching/weighting estimator in statspai.matching:

  classical:  nearest (default), stratify/subclass, cem/coarsened_exact,
              psm, mahalanobis
  weighting:  ebalance/entropy (Hainmueller 2012),
              cbps (Imai-Ratkovic 2014),
              sbw/stable_balancing (Zubizarreta 2015),
  
[…]
