+++ src/statspai/matching/__init__.py
@@ -9,6 +9,18 @@ Unified interface for matching estimators:
 - Propensity score stratification / subclassification
 - Abadie-Imbens (2011) bias correction
 - Entropy balancing (Hainmueller 2012)
+- Covariate Balancing Propensity Score (Imai-Ratkovic 2014)
+- Genetic Matching (Diamond-Sekhon 2013)
+- Stable Balancing Weights (Zubizarreta 2015)
+- Optimal pair / full / cardinality matching (Rosenbaum 1989, 2012)
+- Overlap weights (Li-Morgan-Zaslavsky 2018)
+
+The single entry point is :func:`match` — a method-aware dispatcher
+that routes ``method=`` to the correct estimator.  Standalone
+functions (``ebalance``, ``cbps``, ``genmatch``, ``sbw``,
+``optimal_match``, ``cardinality_match``, ``overlap_weights``) remain
+fully accessible for power users who need their estimator-specific
+parameters.
 
 References
 ----------
@@ -17,10 +29,20 @@ Abadie, A. and Imbens, G.W. (2006). Econometrica, 74(1), 235-267.
 Abadie, A. and Imbens, G.W. (2011). JBES, 29(1), 1-11.
 Iacus, S.M., King, G., and Porro, G. (2012). Political Analysis, 20(1), 1-24.
 Hainmueller, J. (2012). Political Analysis, 20(1), 25-46.
-Cunningham, S. (2021). *Causal Inference: The Mixtape*. Yale University Press. [@rosenbaum1983central]
+Imai, K. and Ratkovic, M. (2014). JRSS-B, 76(1), 243-263.
+Diamond, A. and Sekhon, J.S. (2013). REStat, 95(3), 932-945.
+Zubizarreta, J.R. (2015). JASA, 110(511), 910-922.
+Li, F., Morgan, K.L., and Zaslavsky, A.M. (2018). JASA, 113(521), 390-400.
+Rosenbaum, P.R. (2012). JASA, 107(498), 691-700.
+Cunningham, S. (2021). *Causal Inference: The Mixtape*. Yale University
+Press. [@rosenbaum1983central]
 """
 
-from .match import match, MatchEstimator, balanceplot, psplot
+from typing import Any, Dict, List, Optional
+
+# Underlying estimators — the dispatcher delegates here.
+from .match import match as _match_classical
+from .match import MatchEstimator, balanceplot, psplot
 from .ebalance import ebalance
 from .ps_diagnostics import (
     propensity_score, overlap_plot, trimming, love_plot,
@@ -34,6 +56,187 @@ from .overlap_weights import overlap_weights
 from .cbps import cbps
 from .genmatch import genmatch, GenMatchResult
 from .sbw import sbw, SBWResult
+
+
+# ═══════════════════════════════════════════════════════════════════════
+#  Unified dispatcher — sp.match(..., method=...)
+# ═══════════════════════════════════════════════════════════════════════
+#
+# ``sp.match`` was historically a function with built-in
+# ``method=`` for `nearest` / `stratify` / `cem` / `psm` /
+# `mahalanobis`.  v1.10 expands that to cover every advanced
+# matching/weighting estimator in this subpackage:
+#
+#   classical  → nearest, stratify, cem, psm, mahalanobis
+#   weighting  → ebalance, cbps, sbw, overlap
+#   genetic    → genmatch
+#   optimization-based → optimal, cardinality
+#
+# The classical methods retain their full kwarg surface
+# (caliper / replace / bias_correction / ps_poly / n_strata /
+# n_bins).  Advanced methods accept their own kwargs which we
+# forward verbatim — see each estimator's docstring for details.
+
+# These are the keyword args meaningful only to the classical
+# match() implementation.  We strip them before forwarding to
+# advanced estimators so a user explicitly invoking
[…]
