### 2026-04-29 a979248
feat(panel): expand sp.panel method= with friendly aliases + HDFE route
Phase 5 of the v1.10 API consistency pass.

The classical sp.panel already supported method= for fe / re /
be / fd / pooled / twoway / mundlak / chamberlain / ab / system.
v1.10:

  • Adds case-insensitive aliases that match what users actually
    write instead of the two-letter Stata shorthand:
      fe ← fixed / fixed_effects / within
      re ← random / random_effects
      be ← between / between_effects
      fd ← first_difference / first_diff
      pooled ← pooled_ols / pols / ols
      twoway ← two_way / two_way_fe 
[…]
