+++ src/statspai/panel/__init__.py
@@ -6,6 +6,8 @@ Provides a single entry point ``panel()`` covering all panel estimators:
 **Static models** — FE, RE, Between, First Difference, Pooled OLS, Two-way FE
 **Correlated RE** — Mundlak (1978), Chamberlain (1982)
 **Dynamic panel** — Arellano-Bond, Blundell-Bond (System GMM)
+**HDFE absorption** — high-dimensional fixed-effects OLS (Stata's reghdfe /
+R's fixest)
 
 All results return ``PanelResults`` with built-in diagnostics:
 
@@ -25,10 +27,14 @@ Blundell, R. and Bond, S. (1998). "Initial Conditions and Moment Restrictions."
 Hausman, J.A. (1978). "Specification Tests in Econometrics."
 Breusch, T.S. and Pagan, A.R. (1980). "The Lagrange Multiplier Test."
 Pesaran, M.H. (2004). "General Diagnostic Tests for Cross Section Dependence."
+Correia, S. (2017). "Linear Models with High-Dimensional Fixed Effects."
 """
 
+from typing import Any, Dict, Optional
+
+# Underlying estimators
 from .panel_reg import (
-    panel,
+    panel as _panel_classical,
     panel_compare,
     balance_panel,
     PanelResults,
@@ -39,6 +45,160 @@ from .panel_binary import panel_logit, panel_probit
 from .panel_plots import plot_within_between
 from .hdfe import Absorber, demean, absorb_ols
 from .feols import feols as hdfe_feols, hdfe_ols, FEOLSResult
+
+
+# ═══════════════════════════════════════════════════════════════════════
+#  Unified dispatcher — sp.panel(..., method=...)
+# ═══════════════════════════════════════════════════════════════════════
+#
+# The classical panel() supports method= for fe / re / be / fd / pooled /
+# twoway / mundlak / chamberlain / ab / system.  v1.10 adds:
+#
+#   • Case-insensitive aliases for the 10 classical methods
+#     (fixed → fe, random → re, between → be, first_difference → fd,
+#      pooled_ols → pooled, two_way → twoway, gmm → ab,
+#      arellano_bond → ab, blundell_bond → system, system_gmm → system).
+#
+#   • New method='hdfe' route to ``feols.hdfe_ols`` for high-dimensional
+#     fixed-effects absorption (Stata's reghdfe / R's fixest::feols).
+#
+# panel_logit / panel_probit / interactive_fe / panel_unitroot have a
+# different (data, y, x, id, time)-style signature and are intentionally
+# NOT in the dispatcher — they remain accessible as standalone functions.
+
+_PANEL_METHOD_ALIASES: Dict[str, str] = {
+    # --- Static models ---
+    "fe": "fe", "fixed": "fe", "fixed_effects": "fe", "within": "fe",
+    "re": "re", "random": "re", "random_effects": "re",
+    "be": "be", "between": "be", "between_effects": "be",
+    "fd": "fd", "first_difference": "fd", "first_diff": "fd",
+    "pooled": "pooled", "pooled_ols": "pooled", "pols": "pooled",
+    "ols": "pooled",
+    "twoway": "twoway", "two_way": "twoway", "two_way_fe": "twoway",
+    "twoway_fe": "twoway", "2way": "twoway",
+
+    # --- Correlated random effects ---
+    "mundlak": "mundlak", "mundlak_cre": "mundlak",
+    "chamberlain": "chamberlain", "chamberlain_cre": "chamberlain",
+
+    # --- Dynamic panel (GMM) ---
+    "ab": "ab", "arellano_bond": "ab", "gmm": "ab", "diff_gmm": "ab",
+    "difference_gmm": "ab",
+    "system": "system", "system_gmm": "system",
+    "blundell_bond": "system", "bb": "system",
+
[…]

[…]
