### 2026-04-29 b95d11e
refactor(forest): rename causal/ → forest/ + deprecation shim
Phase B of v1.10 architectural cleanup.

Why: src/statspai/causal/ only ever housed four forest-based
estimators (CausalForest / iv_forest / multi_arm_forest / forest
inference helpers). The "causal" name was misleading — it
suggested a top-level causal namespace when in fact it's a single
estimator family. Misled new contributors looking for "the causal
module".

What:
  • git mv src/statspai/causal → src/statspai/forest (preserves
    file history; renames detected as R in git log).
  • Updates docstring `from statspai.causal impo
[…]
