### 2026-04-29 fd28d63
refactor(agent): split tools.py monolith into agent/tools/ subpackage
CLAUDE.md §4 caps single-file agent code at ~800 lines; v1.10's
tools.py grew to 1,024 after the result-handle work. Splitting into
a subpackage keeps each file under 320 lines and clears the way for
Module 3 (from_stata translator) to add ~30 spec entries without
re-bloating the file.

Layout:
  agent/tools/
  ├── __init__.py        # public API + legacy private re-exports
  ├── _helpers.py        # _scalar_or_none / _default_serializer / _identification_serializer
  ├── _dispatch.py       # tool_manifest / execute_tool / _
[…]
