### 2026-04-29 dc35fa4
test(parity): R reference parity for dml/hdfe/grf/cs/did-variants/bcf
Phase D of v1.10 architectural pass — closes the biggest gap in
CLAUDE.md §5 ("先对齐 Stata/R 再扩展"): pre-this-commit only 9
estimators had reference parity. Adds 8 more, all gold-standard
canonical R packages.

New parity tests + fixtures (R scripts committed for reproducibility):

  • test_dml_parity.py         ↔ R DoubleML (PLR + IRM)
    PLR coef matches to 7%, SE within 25% (CV-glmnet stochasticity).
    True θ=0.5; both stacks land at ~0.53.

  • test_hdfe_parity.py         ↔ R fixest::feols
    Two-way FE absorption: x1, 
[…]
