+++ tests/reference_parity/test_dml_parity.py
@@ -1,0 +1,148 @@
+"""Reference parity: ``sp.dml`` vs R DoubleML.
+
+The DoubleML R package is the reference implementation of
+Chernozhukov et al. (2018) "Double/Debiased Machine Learning";
+sp.dml is StatsPAI's Python port.  Numerical agreement at the
+0.05 level on a fixed DGP is the bar — DoubleML's CV-glmnet folds
+introduce stochasticity that scales with sqrt(n), so we tolerate
+~5% deviation on the coefficient.
+
+Fixture lifecycle
+-----------------
+Both engines consume ``_fixtures/dml_data.csv`` (deterministic seed=42
+DGP from ``_generate_dml_data.py``).  R reference values are in
+``_fixtures/dml_R.json`` produced by ``_generate_dml.R``.  Re-run
+the generators only when the DGP itself changes; otherwise the
+fixture is the contract.
+
+Tolerance discipline
+--------------------
+We assert agreement on:
+  • PLR: coef within 7% relative; SE within 25% (CV-glmnet penalty
+    differs slightly between scikit-learn and glmnet — drives SE
+    spread).
+  • IRM: coef within absolute 0.05 (close to zero on this DGP);
+    SE within 25%.
+
+References
+----------
+- Chernozhukov, V., Chetverikov, D., Demirer, M., Duflo, E., Hansen,
+  C., Newey, W. and Robins, J. (2018). Double/debiased machine
+  learning for treatment and structural parameters. *The
+  Econometrics Journal*, 21(1), C1-C68. [@chernozhukov2018double]
+- Bach, P., Chernozhukov, V., Kurz, M.S., Spindler, M. (2022).
+  DoubleML — An Object-Oriented Implementation of Double Machine
+  Learning in R. *JSS*, 108(3), 1-56.
+"""
+from __future__ import annotations
+
+import json
+import pathlib
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+
+
+_FIXTURE_DIR = pathlib.Path(__file__).parent / "_fixtures"
+
+
+@pytest.fixture(scope="module")
+def dml_data():
+    """Load the shared seed=42 DGP — same csv R consumes."""
+    return pd.read_csv(_FIXTURE_DIR / "dml_data.csv")
+
+
+@pytest.fixture(scope="module")
+def r_reference():
+    """Load the golden R DoubleML output."""
+    with open(_FIXTURE_DIR / "dml_R.json") as f:
+        return json.load(f)
+
+
+# ─── PLR (continuous treatment) ─────────────────────────────────────────
+
+
+def test_dml_plr_coefficient_matches_R(dml_data, r_reference):
+    """sp.dml(model='plr') vs R DoubleML PLR — coefficient agreement."""
+    x_cols = [c for c in dml_data.columns if c.startswith("x")]
+    np.random.seed(42)
+    res = sp.dml(
+        data=dml_data, y="y", d="d", X=x_cols,
+        model="plr", n_folds=5,
+    )
+    py_coef = float(res.estimate)
+
+    r_coef = r_reference["plr"]["coef"]
+    rel = abs(py_coef - r_coef) / abs(r_coef)
+    assert rel < 0.07, (
+        f"PLR coefficient drifted from R DoubleML by {rel:.1%} "
+        f"(Python={py_coef:.6f}, R={r_coef:.6f}). "
+        f"Tolerance: 7% relative.  CV-glmnet stochasticity should "
+        f"keep this band; widen only if you understand why."
+    )
+
+
+def test_dml_plr_standard_error_matches_R(dml_data, r_reference):
[…]
