### 2026-04-29 b43df7a
feat(agent): from_stata / from_r translators (Tier 1, 8 + 5 commands)
The single biggest barrier for econometricians migrating to StatsPAI
is "I have a Stata do-file; how do I write this in sp?". v1.10's
agent loop handled it via LLM memory (error-prone). This module
makes translation a pure function: server-side, deterministic, never
hallucinates.

Architecture
------------

  agent/_translation/
  ├── __init__.py      # public from_stata / from_r
  ├── _stata_lexer.py  # 100-line tokenizer (no full Stata parser)
  ├── _stata.py        # Tier-1 handler dict + dispatch
  └── _r.py            #
[…]
