+++ src/statspai/agent/_translation/_r.py
@@ -1,0 +1,283 @@
+"""R command → StatsPAI tool-call translator.
+
+Targets the most common R econometrics calls: ``feols`` (fixest),
+``felm`` (lfe), ``lm`` (base), ``did`` (Callaway-Sant'Anna's R port),
+``synth`` (Synth package). The R input is parsed with a small
+regex-based scanner — we don't pull a full R parser dependency for
+five common shapes.
+"""
+from __future__ import annotations
+
+import re
+from typing import Any, Callable, Dict, List, Optional, Tuple
+
+
+def _emit(tool: str, arguments: Dict[str, Any],
+           python_code: str, notes: List[str] = None) -> Dict[str, Any]:
+    return {
+        "tool": tool,
+        "arguments": dict(arguments),
+        "python_code": python_code,
+        "notes": list(notes or []),
+        "ok": True,
+    }
+
+
+def _emit_error(message: str, **extra) -> Dict[str, Any]:
+    return {"tool": None, "ok": False, "error": message, **extra}
+
+
+# ---------------------------------------------------------------------------
+# Argument parser — handles ``foo(arg1 = "x", arg2 = c("a","b"))`` shapes
+# ---------------------------------------------------------------------------
+
+def _split_top_level(s: str, delim: str = ",") -> List[str]:
+    """Split on ``delim`` outside any parens / quotes."""
+    out: List[str] = []
+    buf: List[str] = []
+    depth = 0
+    in_q: Optional[str] = None
+    for ch in s:
+        if in_q:
+            buf.append(ch)
+            if ch == in_q:
+                in_q = None
+            continue
+        if ch in "\"'":
+            in_q = ch
+            buf.append(ch)
+            continue
+        if ch == "(":
+            depth += 1
+            buf.append(ch)
+            continue
+        if ch == ")":
+            depth = max(0, depth - 1)
+            buf.append(ch)
+            continue
+        if ch == delim and depth == 0:
+            out.append("".join(buf).strip())
+            buf = []
+            continue
+        buf.append(ch)
+    tail = "".join(buf).strip()
+    if tail:
+        out.append(tail)
+    return out
+
+
+def _parse_call(line: str) -> Optional[Tuple[str, List[str], Dict[str, str]]]:
+    """Parse ``fn(arg1, arg2, key=val, ...)`` into name + positional + kwargs.
+
+    Returns ``None`` if ``line`` doesn't match a function-call shape.
+    """
+    line = line.strip().rstrip(";").strip()
+    m = re.match(r"^([A-Za-z_][\w.]*)\s*\((.*)\)\s*$", line, flags=re.S)
+    if not m:
+        return None
+    fn = m.group(1)
+    body = m.group(2)
+    parts = _split_top_level(body, ",")
+    pos: List[str] = []
+    kw: Dict[str, str] = {}
+    for p in parts:
+        # ``key = value`` (the space around ``=`` is normal R style)
+        eq = re.match(r"^([A-Za-z_][\w.]*)\s*=\s*(.+)$", p, flags=re.S)
+        if eq:
+            kw[eq.group(1)] = eq.group(2).strip()
+        else:
+            pos.append(p)
+    return fn, pos, kw
+
+
+def _strip_quotes(s: str) -> str:
+    s = s.strip()
+    if (s.startswith('"') and s.endswith('"')) or \
+       (s.startswith("'") and s.endswith("'")):
+        return s[1:-1]
+    return s
+
+
+def _parse_c_vector(s: str) -> List[str]:
+    """``c("a", "b", "c")`` or ``c(a, b)`` → list of stripped strings."""
[…]
