### 2026-04-29 97f7313
feat(agent): from_stata Tier 2 — +12 commands (~85% coverage)
Adds the next 12 Stata commands on top of the Tier-1 flagship 8.
Same dispatcher / lexer / output schema; pure declarative additions.

New commands
------------

GLM family:
  probit, logit, poisson, nbreg
    → sp.probit / sp.logit / sp.poisson / sp.nbreg
    Shared scaffold ``_h_glm_like`` so adding mlogit / oprobit later
    is one line.

Censored / selection:
  tobit y x, ll(0) ul(80)        → sp.tobit
  heckman y x, select(d = z)     → sp.heckman with select_formula

RD ancillary:
  rdplot                         → sp.rdplot
  
[…]
