+++ src/statspai/agent/_translation/_stata.py
@@ -377,6 +377,252 @@ def _h_rdrobust(cmd: StataCommand) -> Dict[str, Any]:
 
 
 # ---------------------------------------------------------------------------
+# Tier-2 handlers — observational + RD ancillary + diagnostics
+# ---------------------------------------------------------------------------
+
+def _h_probit(cmd: StataCommand) -> Dict[str, Any]:
+    return _h_glm_like(cmd, sp_fn="probit", display_name="probit")
+
+
+def _h_logit(cmd: StataCommand) -> Dict[str, Any]:
+    return _h_glm_like(cmd, sp_fn="logit", display_name="logit")
+
+
+def _h_poisson(cmd: StataCommand) -> Dict[str, Any]:
+    return _h_glm_like(cmd, sp_fn="poisson", display_name="poisson")
+
+
+def _h_nbreg(cmd: StataCommand) -> Dict[str, Any]:
+    return _h_glm_like(cmd, sp_fn="nbreg", display_name="nbreg")
+
+
+def _h_glm_like(cmd: StataCommand, *, sp_fn: str,
+                  display_name: str) -> Dict[str, Any]:
+    """Common scaffold for probit / logit / poisson / nbreg."""
+    y, xs = _split_varlist_y_x(cmd.varlist)
+    if y is None:
+        return _emit_error(f"{display_name} requires an outcome variable",
+                            command=display_name)
+    formula = _build_formula(y, xs)
+    cluster = _vce_cluster(cmd)
+    robust = _robust_kind(cmd)
+    args: Dict[str, Any] = {"formula": formula}
+    if robust != "nonrobust":
+        args["robust"] = robust
+    if cluster:
+        args["cluster"] = cluster
+    code_pairs = ["data=df"]
+    if cluster:
+        code_pairs.append(f"cluster={cluster!r}")
+    python = (f"sp.{sp_fn}({formula!r}, "
+               + ", ".join(code_pairs) + ")")
+    return _emit(sp_fn, args, python)
+
+
+def _h_tobit(cmd: StataCommand) -> Dict[str, Any]:
+    """``tobit y x, ll(0) ul(100)`` → ``sp.tobit``."""
+    y, xs = _split_varlist_y_x(cmd.varlist)
+    if y is None:
+        return _emit_error("tobit requires an outcome variable",
+                            command="tobit")
+    args: Dict[str, Any] = {"formula": _build_formula(y, xs)}
+    ll = cmd.options.get("ll")
+    ul = cmd.options.get("ul")
+    if ll is not None:
+        try:
+            args["lower"] = float(ll)
+        except (TypeError, ValueError):
+            pass
+    if ul is not None:
+        try:
+            args["upper"] = float(ul)
+        except (TypeError, ValueError):
+            pass
+    code_pairs = ["data=df"]
+    if "lower" in args:
+        code_pairs.append(f"lower={args['lower']}")
+    if "upper" in args:
+        code_pairs.append(f"upper={args['upper']}")
+    python = f"sp.tobit({args['formula']!r}, {', '.join(code_pairs)})"
+    return _emit("tobit", args, python)
+
+
+def _h_heckman(cmd: StataCommand) -> Dict[str, Any]:
+    """``heckman y x, select(employed = age kids)`` → ``sp.heckman``."""
+    y, xs = _split_varlist_y_x(cmd.varlist)
+    if y is None:
+        return _emit_error("heckman requires an outcome variable",
+                            command="heckman")
+    formula = _build_formula(y, xs)
+    select = cmd.options.get("select")
+    if not select:
+        return _emit_error(
+            "heckman needs `select(<eq>)` (selection equation).",
+            command="heckman")
[…]
