### 2026-04-29 28cd353
feat(agent): concurrent runner — tool timeout + progress notifications
The v1.10 stdio loop processed tools/call sequentially: a 10-minute
spec_curve blocked the channel and the client got no signal during.
This module wraps each tools/call in a worker thread, drains a
progress queue from the main loop, and enforces a hard timeout.

Why threading not asyncio
-------------------------

asyncio on stdin is fragile cross-platform (Windows lacks
``connect_read_pipe``; ``anyio`` works but is a new dep). A small
``threading.Thread`` + ``queue.Queue`` keeps the public surface
unchanged and avoids the
[…]
