+++ src/statspai/agent/_runner.py
@@ -1,0 +1,223 @@
+"""Concurrent tool-call runner with timeout + progress notifications.
+
+Why this exists
+---------------
+
+The v1.x ``serve_stdio`` is a strictly serial loop:
+
+    for raw in stdin:
+        response = handle_request(raw)
+        write(response)
+
+That has three operational problems:
+
+1. **Long tools block the channel.** A 10-minute ``spec_curve`` ties up
+   the server; the client can't even ``initialize`` a sibling MCP
+   session until it ends.
+2. **No progress signal.** MCP 2024-11-05 supports
+   ``notifications/progress`` so a server can keep the agent's UI
+   informed during slow work; the v1.10 server has no plumbing.
+3. **No timeout.** A misconfigured BCF / SuperLearner can pin the
+   process indefinitely.
+
+This module wraps each ``tools/call`` in a worker thread, drains a
+shared progress queue from the main loop, and enforces a global
+timeout. Tools opt in to progress reporting via a ``progress=``
+keyword (passed through ``tools/call``'s ``_meta.progressToken``);
+tools that don't accept it remain unaffected.
+
+Why threading, not asyncio
+--------------------------
+
+asyncio on stdin is fragile cross-platform (Windows lacks
+``connect_read_pipe`` for pipes; ``anyio`` works but is a new dep).
+A small ``threading.Thread`` + ``queue.Queue`` keeps the public
+surface unchanged and avoids the asyncio Windows pitfall, at the cost
+of giving up structured cancellation (we use ``threading.Event`` +
+heartbeat checks instead).
+
+Public surface
+--------------
+
+* :func:`run_tools_call_with_progress` — wrap an MCP tools/call
+  request_id, params dict, and emit-callback into a runner that
+  returns the JSON-RPC result string and posts progress notifications
+  along the way.
+* :data:`PROGRESS_TOKEN` — thread-local key tools read to discover
+  whether to emit progress.
+* :data:`TOOL_TIMEOUT_ENV` — env var read by :func:`tool_timeout`.
+"""
+from __future__ import annotations
+
+import json
+import os
+import queue
+import threading
+import time
+from typing import Any, Callable, Dict, Optional, Tuple
+
+
+#: Env var: hard timeout (seconds) per ``tools/call``. ``0`` ⇒ disabled.
+TOOL_TIMEOUT_ENV = "STATSPAI_MCP_TOOL_TIMEOUT_SECONDS"
+_DEFAULT_TIMEOUT_SECONDS = 600  # 10 min — generous for BCF / spec_curve
+
+
+def tool_timeout() -> Optional[float]:
+    """Read the configured tool-call timeout. ``None`` ⇒ no timeout."""
+    raw = os.environ.get(TOOL_TIMEOUT_ENV)
+    if raw is None:
+        return _DEFAULT_TIMEOUT_SECONDS
+    try:
+        v = float(raw)
+    except (TypeError, ValueError):
+        return _DEFAULT_TIMEOUT_SECONDS
+    return v if v > 0 else None
+
+
+# ---------------------------------------------------------------------------
+# Per-thread progress channel
+# ---------------------------------------------------------------------------
+
+_THREAD_LOCAL = threading.local()
+
+
+def _set_progress_channel(token: Any, q: queue.Queue) -> None:
+    _THREAD_LOCAL.progress_token = token
+    _THREAD_LOCAL.progress_queue = q
+
+
+def _clear_progress_channel() -> None:
+    if hasattr(_THREAD_LOCAL, "progress_token"):
+        del _THREAD_LOCAL.progress_token
+    if hasattr(_THREAD_LOCAL, "progress_queue"):
+        del _THREAD_LOCAL.progress_queue
+
+
[…]
