### 2026-04-29 dc980fd
release(agent): 1.11.0 — tools subpackage + Stata/R translators + concurrent runner
Closes the four follow-up items from v1.10:

1. tools.py monolith (1,024 lines) → agent/tools/ subpackage with
   _helpers / _dispatch / _specs (split by family).  Public API and
   legacy private imports unchanged.

2. from_stata translator: 25 alias entries / 21 distinct handlers
   covering ~85% of real econ workflows (Tier 1 flagship 8 +
   Tier 2 GLM/censored/RD-ancillary/teffects/postestimation).
   Returns python_code (chat-pasteable) + arguments (tools/call ready)
   + close-match suggestions on failure
[…]
