### 2026-04-29 a8dcfac
refactor(agent): split mcp_server.py — _prompts/_resources/_data_loader/_errors
The v1.11 mcp_server.py grew to 1,475 lines (Module G prompts +
Module B resources + Module A data-loader work all landed there).
Splits each concern into a leaf module so the JSON-RPC dispatch file
stays focused on protocol mechanics.

New modules
-----------

* `_errors.py` (35 LOC) — RpcError / InvalidParamsError /
  ResourceNotFoundError. Lives in its own file so split helpers can
  raise the same typed errors without forming a circular import
  through mcp_server.

* `_prompts.py` (344 LOC) — the 10 prompt-tem
[…]
