+++ src/statspai/agent/_data_loader.py
@@ -1,0 +1,176 @@
+"""Data-file loading for the MCP layer.
+
+Pulled out of ``mcp_server.py`` so the JSON-RPC dispatch file stays
+focused on protocol concerns. Public surface mirrors what the server
+needs:
+
+* :func:`load_dataframe(path, columns=, sample_n=)` — entry point.
+* :func:`max_data_bytes` / :func:`is_remote_url` — config / predicate.
+
+Supported formats: ``.csv`` / ``.tsv`` / ``.txt`` / ``.parquet`` /
+``.pq`` / ``.feather`` / ``.arrow`` / ``.xlsx`` / ``.xls`` / ``.dta``
+(Stata) / ``.json`` / ``.jsonl``. Schemes: ``file://``, ``s3://``,
+``gs://``, ``https://``, ``http://``.
+
+Local loads are LRU-cached by ``(path, mtime, columns_key)`` so
+repeated tools/call invocations on the same file are O(1).
+"""
+from __future__ import annotations
+
+import functools
+import os
+from typing import List, Optional
+
+
+#: Default max file size (bytes) the server will load. A misconfigured
+#: client pointing at a 50GB parquet will OOM the host otherwise.
+#: Override via ``STATSPAI_MCP_MAX_DATA_BYTES`` (e.g. ``5_000_000_000``);
+#: set to ``0`` to disable the check.
+DEFAULT_MAX_DATA_BYTES = 2 * 1024 * 1024 * 1024  # 2 GiB
+
+
+def max_data_bytes() -> int:
+    raw = os.environ.get("STATSPAI_MCP_MAX_DATA_BYTES")
+    if raw is None:
+        return DEFAULT_MAX_DATA_BYTES
+    try:
+        return max(0, int(raw))
+    except (TypeError, ValueError):
+        return DEFAULT_MAX_DATA_BYTES
+
+
+def is_remote_url(path: str) -> bool:
+    return path.startswith(("s3://", "gs://", "https://", "http://",
+                             "file://"))
+
+
+def load_dataframe(path: str,
+                    columns: Optional[List[str]] = None,
+                    sample_n: Optional[int] = None):
+    """Load a DataFrame from a local path or remote URL.
+
+    Parameters
+    ----------
+    path : str
+        Absolute filesystem path or one of: ``file://``, ``s3://``,
+        ``gs://``, ``https://``, ``http://``.
+    columns : list of str, optional
+        Column projection. Honoured by parquet / feather / stata
+        readers; for CSV we read all columns then sub-select (read_csv's
+        ``usecols`` would also work but mismatched names raise; we want
+        the server to be permissive and let the estimator surface
+        column-name errors with rich remediation).
+    sample_n : int, optional
+        Uniform random subsample size (seed=0, deterministic).
+        Applied AFTER the projection.
+
+    Notes
+    -----
+    Caches the materialised frame keyed by ``(path, mtime, columns)``
+    so repeated tool calls on the same file are O(1) after the first
+    load.
+    """
+    if is_remote_url(path):
+        # Remote — defer all guard rails to the underlying loader; we
+        # can't ``os.path.exists`` an s3 URL, and pandas/storage_options
+        # error messages are rich enough.
+        df = _load_remote(path, columns=columns)
+    else:
+        if not os.path.isabs(path):
+            raise ValueError(
+                f"data_path must be absolute or a URL, got {path!r}")
+        if not os.path.exists(path):
+            raise FileNotFoundError(f"No such file: {path}")
+        size = os.path.getsize(path)
+        cap = max_data_bytes()
+        if cap and size > cap:
+            raise ValueError(
[…]
