+++ src/statspai/agent/_resources.py
@@ -1,0 +1,313 @@
+"""MCP ``resources/*`` plumbing — catalog text, function detail,
+result-handle reads, URI templates.
+
+Decoupled from ``mcp_server.py`` so the (large) resource bodies don't
+bloat the JSON-RPC dispatch file. The handler functions accept the
+JSON encoder + error classes as arguments rather than importing them
+from ``mcp_server`` — avoids the circular import that would otherwise
+form (mcp_server → _resources → mcp_server).
+
+Public entry points
+-------------------
+
+* :data:`FUNCTION_URI_PREFIX` / :data:`RESULT_URI_PREFIX`
+* :func:`catalog_text` — markdown catalog body
+* :func:`functions_index` — JSON ``[{name, description}, ...]`` list
+* :func:`function_detail` — full agent card for one tool name
+* :func:`handle_resources_list` — the three top-level resources
+* :func:`handle_resources_read` — URI dispatch (catalog / functions /
+  function/<name> / result/<id>)
+* :func:`handle_resources_templates_list` — the parameterised URI
+  templates
+"""
+from __future__ import annotations
+
+import json
+from typing import Any, Callable, Dict, List, Optional
+
+
+FUNCTION_URI_PREFIX = "statspai://function/"
+RESULT_URI_PREFIX = "statspai://result/"
+
+
+# ----------------------------------------------------------------------
+# Catalog / index / detail helpers
+# ----------------------------------------------------------------------
+
+def catalog_text(server_version: str) -> str:
+    """Return a Markdown catalog of every StatsPAI tool."""
+    from .tools import tool_manifest
+    manifest = tool_manifest()
+    lines = [
+        "# StatsPAI tool catalog",
+        "",
+        f"Version: {server_version}. {len(manifest)} tools registered.",
+        "",
+        "**Per-function detail**: read "
+        f"`{FUNCTION_URI_PREFIX}<name>` for the full agent card "
+        "(assumptions, failure modes, alternatives, typical_n_min, "
+        "example) of any tool listed below.",
+        "",
+        "**Machine-readable index**: read `statspai://functions` for a "
+        "JSON array of `{name, description}` entries.",
+        "",
+    ]
+    for t in manifest:
+        lines.append(f"## {t['name']}")
+        lines.append("")
+        desc = t.get("description", "").strip()
+        if desc:
+            lines.append(desc)
+            lines.append("")
+    return "\n".join(lines)
+
+
+def functions_index() -> List[Dict[str, str]]:
+    """Return a JSON-ready ``[{name, description}, …]`` list."""
+    from .tools import tool_manifest
+    return [
+        {"name": t["name"],
+         "description": (t.get("description") or "").strip()}
+        for t in tool_manifest()
+    ]
+
+
+def function_detail(name: str) -> Optional[Dict[str, Any]]:
+    """Return the rich agent card for one tool, or ``None`` if unknown.
+
+    Prefers ``statspai.registry.agent_card`` (full card with
+    assumptions / failure_modes / alternatives / typical_n_min) and
+    falls back to the manifest entry for tools that exist in the
+    auto-generated layer but lack a hand-curated registry spec.
+    """
+    # Try the registry first — it has the agent-native metadata.
+    try:
+        from ..registry import agent_card as _agent_card
+        card = _agent_card(name)
+        if card:
+            return card
+    except Exception:
+        pass
+
+    # Fallback: synthesise from the merged manifest so any registered
+    # tool — even auto-generated ones without a curated spec — still
+    # resolves to *something* readable.
+    from .tools import tool_manifest
+    for t in tool_manifest():
+        if t["name"] == name:
+            return {
+                "name": t["name"],
+                "description": (t.get("description") or "").strip(),
+                "signature": {
[…]
