### 2026-04-29 c75d96b
feat(agent): from_stata Tier 3 — long-tail (~95% coverage)
Adds 8 long-tail Stata commands. Same dispatcher / lexer / output
schema; pure declarative additions.

New commands
------------

Poisson with HDFE:
  ppmlhdfe trade gravity, absorb(orig dest year) cluster(orig)
    → sp.ppmlhdfe(formula, fe=[...], cluster=...)
  Correia-Guimarães-Zylkin "stochastically dominates" the iterative
  PPML+HDFE pattern.

Multinomial / ordinal:
  mlogit choice age income, baseoutcome(1)
    → sp.glm(family='multinomial', base_outcome=1) + translation note
      that strict multinomial diagnostics live in res
[…]
