+++ src/statspai/agent/_translation/_stata.py
@@ -623,6 +623,205 @@ def _h_xtset(cmd: StataCommand) -> Dict[str, Any]:
 
 
 # ---------------------------------------------------------------------------
+# Tier-3 handlers — long-tail GMM / multinomial / bunching / boottest /
+# Poisson HDFE / mi_estimate
+# ---------------------------------------------------------------------------
+
+def _h_ppmlhdfe(cmd: StataCommand) -> Dict[str, Any]:
+    """``ppmlhdfe y x, absorb(id year) cluster(id)`` → ``sp.ppmlhdfe`` (or
+    sp.poisson with FE if ppmlhdfe unavailable). Correia-Guimarães-Zylkin
+    Poisson PML with HDFE."""
+    y, xs = _split_varlist_y_x(cmd.varlist)
+    if y is None:
+        return _emit_error("ppmlhdfe requires an outcome variable",
+                            command="ppmlhdfe")
+    absorb = cmd.options.get("absorb") or ""
+    fe_list = [v for v in absorb.split() if v]
+    cluster = _vce_cluster(cmd) or cmd.options.get("cluster")
+    if cluster:
+        cluster = cluster.split()[0]
+    formula = _build_formula(y, xs)
+    args: Dict[str, Any] = {"formula": formula, "fe": fe_list}
+    if cluster:
+        args["cluster"] = cluster
+    code_pairs = ["data=df", f"fe={fe_list!r}"]
+    if cluster:
+        code_pairs.append(f"cluster={cluster!r}")
+    python = f"sp.ppmlhdfe({formula!r}, {', '.join(code_pairs)})"
+    notes: List[str] = []
+    if not fe_list:
+        notes.append("ppmlhdfe with no absorb() degenerates to "
+                      "sp.poisson — consider using that directly.")
+    return _emit("ppmlhdfe", args, python, notes)
+
+
+def _h_mlogit(cmd: StataCommand) -> Dict[str, Any]:
+    """``mlogit choice age income, baseoutcome(1)`` → multinomial logit.
+
+    No 1:1 sp helper exists; we translate to a ``method='mlogit'``
+    GLM call as a fallback and surface a note. Users with strict
+    multinomial needs should fall back to statsmodels via the result.
+    """
+    y, xs = _split_varlist_y_x(cmd.varlist)
+    if y is None:
+        return _emit_error("mlogit requires an outcome variable",
+                            command="mlogit")
+    formula = _build_formula(y, xs)
+    args: Dict[str, Any] = {
+        "formula": formula,
+        "family": "multinomial",
+    }
+    base = cmd.options.get("baseoutcome")
+    if base is not None:
+        args["base_outcome"] = _coerce_scalar(base)
+    notes = ["StatsPAI doesn't ship a dedicated mlogit; "
+             "sp.glm(family='multinomial') is the closest. "
+             "For full diagnostics use statsmodels.MNLogit on the "
+             "fitted result via .raw_model."]
+    code_kwargs = ", ".join(
+        ["data=df", "family='multinomial'"]
+        + ([f"base_outcome={args['base_outcome']!r}"]
+            if "base_outcome" in args else [])
+    )
+    python = f"sp.glm({formula!r}, {code_kwargs})"
+    return _emit("glm", args, python, notes)
+
+
+def _h_oprobit(cmd: StataCommand) -> Dict[str, Any]:
+    """``oprobit grade x1 x2`` → ordered probit via sp.glm(family='ordered_probit')."""
+    y, xs = _split_varlist_y_x(cmd.varlist)
+    if y is None:
+        return _emit_error("oprobit requires an outcome variable",
[…]
