### 2026-04-29 9b7c52d
feat(agent): from_r — glm/glmer/lmer/plm/matchit (~6 → 11 callables)
Expands from_r from 5 to 11 distinct R callables. Same dispatch /
output schema; declarative additions only.

New callables
-------------

GLM family with smart routing:
  glm(y ~ x, family=binomial, data=df)
    → sp.logit  (binomial w/ logit link)
  glm(y ~ x, family=binomial(link="probit"), data=df)
    → sp.probit
  glm(counts ~ x, family=poisson, data=df)
    → sp.poisson
  glm(y ~ x, family=gaussian, data=df)
    → sp.glm

Multilevel:
  lmer(y ~ x + (1|group), data=df)
    → sp.multilevel
  glmer(y ~ x + (1|group), fami
[…]
