+++ src/statspai/agent/_translation/_r.py
@@ -229,13 +229,132 @@ def _h_did(pos: List[str], kw: Dict[str, str], _) -> Dict[str, Any]:
     return _emit("callaway_santanna", args, python)
 
 
+def _h_glm(pos: List[str], kw: Dict[str, str], _) -> Dict[str, Any]:
+    """R `glm(y ~ x, family=binomial, data=df)` → ``sp.glm`` (or sp.logit
+    / sp.probit / sp.poisson when the family resolves to one of those).
+    """
+    formula = pos[0] if pos else kw.get("formula")
+    if not formula:
+        return _emit_error("glm requires a formula as the first argument")
+    formula = _strip_quotes(formula)
+    family = _strip_quotes(kw.get("family", "gaussian")) if kw.get("family") else "gaussian"
+    # Recognise common family/link combos so we can route to the
+    # specialised sp helper rather than the generic GLM.
+    fam_lower = family.lower().strip()
+    if "binomial" in fam_lower and ("logit" in fam_lower or fam_lower == "binomial"):
+        return _emit("logit", {"formula": formula},
+                      f"sp.logit({formula!r}, data=df)")
+    if "binomial" in fam_lower and "probit" in fam_lower:
+        return _emit("probit", {"formula": formula},
+                      f"sp.probit({formula!r}, data=df)")
+    if "poisson" in fam_lower:
+        return _emit("poisson", {"formula": formula},
+                      f"sp.poisson({formula!r}, data=df)")
+    args: Dict[str, Any] = {"formula": formula, "family": family}
+    python = f"sp.glm({formula!r}, data=df, family={family!r})"
+    return _emit("glm", args, python)
+
+
+def _h_glmer(pos: List[str], kw: Dict[str, str], _) -> Dict[str, Any]:
+    """R `lme4::glmer(y ~ x + (1|group), family=binomial, data=df)` →
+    ``sp.glmer`` / multilevel GLM. The ``(1|group)`` random-effect
+    syntax passes through verbatim — sp's multilevel module accepts
+    the same shape."""
+    formula = pos[0] if pos else kw.get("formula")
+    if not formula:
+        return _emit_error("glmer requires a formula as the first argument")
+    formula = _strip_quotes(formula)
+    family = _strip_quotes(kw.get("family", "gaussian")) if kw.get("family") else "gaussian"
+    args: Dict[str, Any] = {"formula": formula, "family": family}
+    notes: List[str] = []
+    code_pairs = ["data=df", f"family={family!r}"]
+    python = f"sp.glmer({formula!r}, {', '.join(code_pairs)})"
+    return _emit("glmer", args, python, notes)
+
+
+def _h_lmer(pos: List[str], kw: Dict[str, str], _) -> Dict[str, Any]:
+    """R `lme4::lmer(y ~ x + (1|group), data=df)` → ``sp.multilevel`` /
+    Gaussian random effects."""
+    formula = pos[0] if pos else kw.get("formula")
+    if not formula:
+        return _emit_error("lmer requires a formula as the first argument")
+    formula = _strip_quotes(formula)
+    args: Dict[str, Any] = {"formula": formula}
+    python = f"sp.multilevel({formula!r}, data=df)"
+    return _emit("multilevel", args, python)
+
+
+def _h_plm(pos: List[str], kw: Dict[str, str], _) -> Dict[str, Any]:
+    """R `plm(y ~ x, data=df, model='within', index=c('id','t'))` →
+    ``sp.panel`` with the chosen estimator."""
+    formula = pos[0] if pos else kw.get("formula")
+    if not formula:
+        return _emit_error("plm requires a formula as the first argument")
+    formula = _strip_quotes(formula)
[…]
