### 2026-04-29 8728cf9
feat(agent): MCP sampling/createMessage abstraction layer (opt-in)
Per MCP 2024-11-05, ``sampling/createMessage`` is a server-to-client
JSON-RPC request: the server asks the connected client to call its
LLM and return the completion. Adding this lets StatsPAI's LLM-driven
helpers (``llm_dag_propose``, ``llm_evalue``, ``llm_sensitivity``)
reuse the client's already-authenticated LLM session instead of
forcing each user to configure their own API key.

Design — fail closed
--------------------

* **Capability-gated**: nothing fires until the client advertises
  ``capabilities.sampling`` during `
[…]
