+++ tests/test_mcp_sampling.py
@@ -1,0 +1,229 @@
+"""Tests for Module 8 — MCP sampling/createMessage server-side client.
+
+The server-side helpers raise ``UnsupportedSamplingError`` when no
+client capability is advertised so callers can route to a fallback.
+When the client does advertise support, the helpers send a
+``sampling/createMessage`` request on stdout and block on the reply.
+"""
+from __future__ import annotations
+
+import io
+import json
+import threading
+import time
+
+import pytest
+
+from statspai.agent import _sampling
+from statspai.agent._sampling import (
+    UnsupportedSamplingError,
+    SamplingTimeoutError,
+    request_sampling,
+    set_capability,
+    set_writer,
+    route_response,
+    SAMPLING_TIMEOUT_ENV,
+)
+
+
+# ----------------------------------------------------------------------
+# Fail-closed when capability not set
+# ----------------------------------------------------------------------
+
+class TestFailClosed:
+    def test_raises_when_capability_unset(self):
+        # Capability resets between tests via the ``reset`` fixture below.
+        with pytest.raises(UnsupportedSamplingError):
+            request_sampling([{"role": "user",
+                                "content": {"type": "text", "text": "hi"}}],
+                               max_tokens=10)
+
+    def test_raises_when_writer_unset(self, monkeypatch):
+        set_capability(True)
+        try:
+            with pytest.raises(UnsupportedSamplingError):
+                request_sampling([{"role": "user",
+                                    "content": {"type": "text", "text": "x"}}],
+                                  max_tokens=10)
+        finally:
+            set_capability(False)
+
+
+# ----------------------------------------------------------------------
+# Round-trip via mock writer
+# ----------------------------------------------------------------------
+
+class TestRoundTrip:
+    def test_sends_request_and_returns_result(self):
+        """A fake client thread reads the outbound line, builds a
+        canned reply, and routes it back via ``route_response``."""
+        outbox = []
+        # Capture every line the server writes so we can synthesise
+        # a reply that matches the request id.
+        set_writer(lambda line: outbox.append(line))
+        set_capability(True)
+
+        def fake_client():
+            # Wait until the server has written its outbound line
+            for _ in range(100):
+                if outbox:
+                    break
+                time.sleep(0.01)
+            assert outbox, "server didn't send a sampling request"
+            sent = json.loads(outbox[0])
+            assert sent["method"] == "sampling/createMessage"
+            assert sent["params"]["maxTokens"] == 200
+            reply = {
+                "jsonrpc": "2.0",
+                "id": sent["id"],
+                "result": {
+                    "role": "assistant",
+                    "content": {"type": "text", "text": "hi back"},
+                    "model": "claude-test",
+                    "stopReason": "endTurn",
+                },
+            }
+            assert route_response(reply) is True
+
+        t = threading.Thread(target=fake_client, daemon=True)
+        t.start()
+        try:
+            result = request_sampling(
+                [{"role": "user",
+                  "content": {"type": "text", "text": "hi"}}],
+                max_tokens=200,
+                system_prompt="be concise",
+                temperature=0.2,
+                timeout=2.0,
+            )
+            assert result["content"]["text"] == "hi back"
+            assert result["model"] == "claude-test"
+        finally:
+            t.join(timeout=1)
+            set_writer(None)
+            set_capability(False)
+
+    def test_client_error_envelope_raises_runtimeerror(self):
+        outbox = []
+        set_writer(lambda line: outbox.append(line))
+        set_capability(True)
+
+        def fake_client():
+            for _ in range(100):
+                if outbox:
+                    break
+                time.sleep(0.01)
[…]
