+++ src/statspai/agent/_sampling.py
@@ -1,0 +1,260 @@
+"""MCP server-initiated LLM sampling — opt-in, capability-gated.
+
+Per MCP 2024-11-05, ``sampling/createMessage`` is a *server-to-client*
+JSON-RPC request: the server asks the connected client to call its
+LLM and return the completion. This module wraps that protocol so
+StatsPAI's LLM-driven helpers (``llm_dag_propose``, ``llm_evalue``,
+``llm_sensitivity``, …) can reuse the client's already-authenticated
+LLM session instead of forcing each user to configure their own API
+key.
+
+Design
+------
+
+* **Capability-gated**: nothing happens until the client advertises
+  ``capabilities.sampling`` during ``initialize``. Until then, calls
+  to :func:`request_sampling` fall through to the
+  ``UnsupportedSamplingError`` so the caller can route to the
+  user-API-key fallback.
+* **Stateful**: the active stdio loop registers a writer callback,
+  records pending request IDs, and matches replies. We don't try to
+  multiplex multiple in-flight sampling requests yet — the existing
+  threading runner serialises tool calls, so one-at-a-time is
+  sufficient.
+* **Bounded wait**: every sampling request has a hard timeout
+  (env: ``STATSPAI_MCP_SAMPLING_TIMEOUT_SECONDS``, default 60s) so
+  the server can't hang forever waiting on a buggy client.
+
+Public surface
+--------------
+
+* :func:`set_capability` / :func:`get_capability` — toggle the
+  client-advertised capability flag (called by ``_handle_initialize``).
+* :func:`request_sampling(messages, ...)` — server-to-client request;
+  blocks until the response or the timeout. Raises
+  :class:`UnsupportedSamplingError` when no capability is set.
+* :func:`route_response(message)` — called by the stdio reader when
+  it spots a JSON-RPC reply whose ``id`` matches a pending sampling
+  request.
+"""
+from __future__ import annotations
+
+import itertools
+import json
+import os
+import threading
+import time
+from typing import Any, Callable, Dict, List, Optional
+
+
+SAMPLING_TIMEOUT_ENV = "STATSPAI_MCP_SAMPLING_TIMEOUT_SECONDS"
+_DEFAULT_SAMPLING_TIMEOUT = 60.0
+
+
+class UnsupportedSamplingError(RuntimeError):
+    """Raised when sampling is requested but the client did not
+    advertise the capability. Callers should catch this and fall
+    back to a user-supplied API-key path."""
+
+
+class SamplingTimeoutError(TimeoutError):
+    """Raised when the client takes too long to reply to a sampling
+    request. The configured limit is read from
+    :data:`SAMPLING_TIMEOUT_ENV` (default 60 s)."""
+
+
+def _sampling_timeout() -> float:
+    raw = os.environ.get(SAMPLING_TIMEOUT_ENV)
+    if raw is None:
+        return _DEFAULT_SAMPLING_TIMEOUT
+    try:
+        v = float(raw)
+    except (TypeError, ValueError):
+        return _DEFAULT_SAMPLING_TIMEOUT
+    return v if v > 0 else _DEFAULT_SAMPLING_TIMEOUT
+
+
+# ---------------------------------------------------------------------------
+# Capability + writer registration (set by mcp_server during the loop)
+# ---------------------------------------------------------------------------
+
+_LOCK = threading.RLock()
+_CAPABILITY: bool = False
+_WRITER: Optional[Callable[[str], None]] = None
+_PENDING: Dict[Any, "_PendingRequest"] = {}
+_REQUEST_ID_COUNTER = itertools.count(1)
+
+
+class _PendingRequest:
+    """In-flight sampling request awaiting a JSON-RPC reply."""
+
[…]
