### 2026-04-29 b1bce1f
release(agent): 1.11.1 — mcp_server split + Stata Tier 3 + R deepening + sampling layer
Closes the four "留意/没做" items from v1.11.0:

1. mcp_server.py (1,475 → 817 lines) split into 5 leaf modules
   (_errors, _prompts, _resources, _data_loader, _sampling). All v1.x
   private names re-exported via shims; test fixtures unchanged.

2. from_stata Tier 3: +8 long-tail commands (ppmlhdfe / mlogit /
   oprobit / xtabond / xtdpdsys / bunching / boottest / mi estimate).
   33 alias entries → 29 distinct handlers, ~95% econ Stata coverage.

3. from_r deepening: +6 callables (glm with smart family routi
[…]
