+++ CHANGELOG.md
@@ -2,6 +2,106 @@
 
 All notable changes to StatsPAI will be documented in this file.
 
+## [1.11.1] — 2026-04-29
+
+Polish patch for the v1.11 agent surface. Closes the four
+"留意 / 没做" items from the v1.11 release notes: mcp_server.py
+1,475-line bloat, from_stata Tier 3, deeper from_r, and the MCP
+sampling abstraction layer. **No estimator numerics changed**.
+
+### Added — `from_stata` Tier 3 (long-tail, ~95% coverage)
+
+- `ppmlhdfe` → `sp.ppmlhdfe` (Correia-Guimarães-Zylkin Poisson-PML
+  with HDFE, multi-FE absorb).
+- `mlogit` / `oprobit` → `sp.glm(family='multinomial' /
+  'ordered_probit')` with a translation note pointing strict
+  diagnostics at `result.raw_model`.
+- `xtabond` / `xtdpdsys` → `sp.xtabond` / `sp.xtdpdsys` (Arellano-
+  Bond difference / Blundell-Bond system GMM).
+- `bunching` → `sp.bunching` (Saez 2010, Kleven-Waseem 2013).
+- `boottest` → `sp.wild_cluster_bootstrap` (Roodman-Webb wild-cluster
+  bootstrap; takes a fitted result).
+- `mi estimate: <inner>` → translation hint pointing at
+  `sp.mi_estimate` (Stata's nested grammar isn't auto-parsed).
+- 33 alias entries / 29 distinct handlers total.
+
+### Added — `from_r` deepening (5 → 11 callables)
+
+- `glm` with smart routing: `family=binomial` → `sp.logit`,
+  `family=binomial(link="probit")` → `sp.probit`,
+  `family=poisson` → `sp.poisson`, otherwise `sp.glm`.
+- `lmer` → `sp.multilevel`; `glmer` → `sp.glmer`.
+- `plm(formula, data=df, model='within', index=c('id','t'))` →
+  `sp.panel(method='within', id='id', time='t')`.
+- `matchit(treat ~ x, data=df, method='nearest')` → `sp.match`
+  with method-name aliasing (nearest→nn, genetic→genmatch).
+- R Synth synth() now emits a structured field-mapping note
+  (predictors → predictors, dependent → outcome, unit.variable →
+  unit, time.variable → time, treatment.identifier →
+  treated_unit, time.predictors.prior[max] + 1 → treatment_time).
+
+### Added — MCP `sampling/createMessage` abstraction (opt-in)
+
+- New `agent/_sampling.py`:
+  - `request_sampling(messages, max_tokens, ...)` — server-to-client
+    LLM request; blocks until response or timeout.
+  - `set_capability(bool)` / `get_capability()` — client capability
+    advertisement flag.
+  - `set_writer(callable)` / `route_response(message)` — stdio
+    writer registration + reply matcher.
+- Wire-up:
+  - `_handle_initialize` reads `params.capabilities.sampling`.
+  - `handle_request` routes JSON-RPC replies to pending sampling
+    requests via `route_response`.
+  - `serve_stdio` registers / clears the writer + capability flag.
+- Fail-closed: `UnsupportedSamplingError` raised when no capability
+  is advertised OR no writer is registered, so existing LLM helpers
+  (`llm_dag_propose` / `llm_evalue` / `llm_sensitivity`) keep
+  working via their user-API-key paths until clients (Claude
+  Desktop, Cursor, …) advertise `sampling`.
+- `STATSPAI_MCP_SAMPLING_TIMEOUT_SECONDS` (default `60`) caps every
+  request; `SamplingTimeoutError` on overage.
+
+### Changed — `mcp_server.py` split into leaf modules
+
+- v1.11.0: `mcp_server.py` was 1,475 lines.
[…]
