### 2026-04-29 c608528
test(output): add regtable snapshot baselines (PR-B safety net)
Lock down the byte-stable rendered output of ``regtable`` for five
representative fixtures across four text formats (text / HTML / LaTeX
/ Markdown), giving the upcoming PR-B consolidation work a reliable
drift detector.

Fixtures:
- f1_simple_ols          — single OLS
- f2_multi_model         — three nested OLS side-by-side
- f3_custom_stats        — custom stats / star_levels / fmt
- f4_with_notes          — title / notes / model_labels overrides
- f5_logit               — GLM (logit) extraction path

20 snapshots total (5 × 4).
[…]
