### 2026-04-29 5388ff9
refactor(output): collapse outreg2 to thin facade over regtable (PR-B/5a)
The Stata-style ``OutReg2`` class and ``outreg2()`` function shipped a
bespoke ~800-line renderer that re-implemented coefficient extraction,
star formatting, three-line table styling, and Excel/Word/LaTeX
export — all features ``regtable`` already provides in
publication-quality form, plus bug fixes the bespoke renderer never
received.

Collapse to ~150 lines that translate Stata-flavoured kwargs and
forward to ``sp.regtable``. The user-visible API (``OutReg2`` class,
``outreg2`` function, all keyword args) is preserved
[…]
