+++ src/statspai/output/outreg2.py
@@ -1,537 +1,201 @@
-"""
-Output regression results to Excel in publication-ready format
-Similar to Stata's outreg2 command
+"""Stata ``outreg2`` compatibility surface — thin facade over :func:`regtable`.
+
+Historically this module shipped its own bespoke renderer pipeline
+(``_create_regression_table``, ``_export_with_formatting``,
+``_apply_apa_borders``, …) totalling ~800 lines of code that
+re-implemented coefficient extraction, star formatting, three-line
+table styling and Excel / Word / LaTeX export — all features that
+:func:`statspai.output.regtable` already provided in publication-quality
+form, with bug fixes the bespoke renderer never received (broken
+``F-statistic / Trees`` row on OLS, junk ``& None & None`` LaTeX cell,
+missing star legend).
+
+In the PR-B output consolidation (see
+``docs/rfc/output_pr_b_consolidation.md``) we collapse this module to
+~150 lines that translate Stata-flavoured kwargs and forward to
+:func:`regtable`. The user-visible API (``OutReg2`` class,
+``outreg2`` function) is preserved; the *rendered output* now matches
+:func:`regtable` exactly — strictly better in every case but
+NOT byte-identical to the legacy output. A :class:`DeprecationWarning`
+points users to ``sp.regtable(...).to_excel(...)`` for full control.
+
+Migration notes
+---------------
+The following labels changed (legacy → new, all canonical to regtable):
+
+- ``"Variables"`` header column → blank (book-tab convention)
+- ``"R-squared"`` → ``"R²"``
+- ``"Adj. R-squared"`` → ``"Adj. R²"``
+- ``"Observations"`` → ``"N"``
+- ``"F-statistic / Trees"`` → ``"F"`` (bug fix: ``"/ Trees"`` only
+  applies to causal-forest results, was always wrong on OLS)
+- LaTeX gains a proper star legend below the table
+- ``show_se=False`` is no longer supported (regression tables without
+  uncertainty are pseudo-science). Pass ``se_type="t"`` etc. via
+  :func:`regtable` directly if you genuinely need an alternative cell.
 """
 
-import pandas as pd
-import numpy as np
-from typing import List, Dict, Any, Optional, Union
-from pathlib import Path
+from __future__ import annotations
+
 import warnings
+from typing import Any, Dict, List, Optional
 
 from ..core.results import EconometricResults
 
-# openpyxl is imported lazily inside ``_export_with_formatting`` —
-# importing it at module load drags in PIL/Pillow (~700 ms cold start)
-# and ``outreg2`` is rarely the entry point for a session.
+_DEPRECATION_MSG = (
+    "OutReg2 / outreg2() are now thin wrappers over sp.regtable() and "
+    "will be removed in a future minor release. Migrate to "
+    "sp.regtable(*models, ...).to_excel(filename) for the same output "
+    "with full control over labels, journal templates, and SE formats. "
+    "See docs/rfc/output_pr_b_consolidation.md for migration."
+)
+
+# Emitted-once flag so we don't spam users who call outreg2 in a loop.
+_DEPRECATION_EMITTED = False
+
+
+def _warn_once() -> None:
+    global _DEPRECATION_EMITTED
+    if not _DEPRECATION_EMITTED:
+        warnings.warn(_DEPRECATION_MSG, DeprecationWarning, stacklevel=3)
+        _DEPRECATION_EMITTED = True
+
+
+def _build_regtable(
+    *,
+    models: List[Any],
+    model_names: Optional[List[str]],
+    title: Optional[str],
+    notes: Optional[List[str]],
+    show_stars: bool,
+    show_se: bool,
+    show_tstat: bool,
+    decimal_places: int,
+    variable_labels: Optional[Dict[str, str]],
+):
[…]
