### 2026-04-29 84bf90e
refactor(output): collapse modelsummary to thin facade over regtable (PR-B/5b)
Same pattern as B-5a (outreg2 facade): the R-style ``modelsummary()``
shipped a ~700-line renderer pipeline (``_build_coef_rows`` /
``_to_text`` / ``_to_latex`` / ``_to_html`` / ``_to_excel`` /
``_to_word``) that re-implemented coefficient extraction, star
formatting, three-line table styling, and every export format —
duplicating logic already maintained by ``sp.regtable``.

Collapse to a thin facade that translates R-flavoured kwargs and
forwards to ``sp.regtable``. The user-visible ``modelsummary()``
function nam
[…]
