+++ src/statspai/output/modelsummary.py
@@ -1,154 +1,233 @@
+"""R ``modelsummary`` compatibility surface — thin facade over :func:`regtable`.
+
+Historically this module shipped its own ~700-line renderer pipeline
+(``_build_coef_rows`` / ``_build_stat_rows`` / ``_to_text`` /
+``_to_latex`` / ``_to_html`` / ``_to_excel`` / ``_to_word`` …) that
+re-implemented coefficient extraction, star formatting, three-line
+table styling and every export format — duplicating code already
+maintained by :func:`statspai.output.regtable`.
+
+In the PR-B output consolidation (see
+``docs/rfc/output_pr_b_consolidation.md``) we collapse the module to a
+thin facade that translates R-flavoured kwargs and forwards to
+:func:`regtable`. The user-visible API (``modelsummary`` function name
+and parameter spelling) is preserved; the *rendered output* now matches
+:func:`regtable` exactly — strictly cleaner (book-tab three-line table,
+publication-quality star legend, fixed labels) but NOT byte-identical
+to the legacy bespoke renderer.
+
+:func:`coefplot` is independent of the table renderer and is kept here
+unchanged.
+
+A :class:`DeprecationWarning` is emitted on first call pointing users
+to :func:`sp.regtable`. Plan to remove the facade in two minor releases
+(per CLAUDE.md §3.8).
+
+Migration notes
+---------------
+- ``stars=True`` / ``stars=False`` work the same. The dict form
+  (``stars={"*": 0.10, "**": 0.05, "***": 0.01}``) is reinterpreted —
+  only the threshold *values* are used, the symbol overrides are
+  dropped (regtable's ladder is ``*/**/***`` by convention; symbol
+  control is via ``notation='symbols'`` for ``†/‡/§``).
+- ``se_type='parentheses'`` (default) and ``se_type='none'`` work as
+  before. ``se_type='brackets'`` is no longer a separate render mode —
+  emits ``UserWarning`` and falls back to parentheses. Use
+  ``show_ci=True`` (which renders ``[lo, hi]``) if you want brackets
+  to convey actual information.
+- ``output='dataframe'`` returns a :class:`pandas.DataFrame` (same as
+  before).
+- ``output='<filename>.xlsx|.docx'`` writes the file and returns a
+  confirmation string.
+- All other ``output=`` strings (``"text"``, ``"latex"``, ``"html"``,
+  ``"markdown"``) return the rendered string.
+- Stat keys map: ``nobs`` → ``N``; ``r_squared`` / ``adj_r_squared``
+  / ``f_stat`` / ``aic`` / ``bic`` → their canonical regtable
+  equivalents.
 """
-Publication-quality multi-model comparison tables.
 
-Equivalent to R's modelsummary() / fixest::etable() / Stata's esttab.
-Accepts any mix of EconometricResults and CausalResult objects and
-produces side-by-side tables in text, LaTeX, HTML, or DataFrame format.
+from __future__ import annotations
 
-References
-----------
-Arel-Bundock, V. (2022). "modelsummary: Data and Model Summaries in R."
-*Journal of Statistical Software*, 103(1), 1-23. [@arelbundock2022modelsummary]
-"""
+import warnings
+from typing import Any, Dict, List, Optional, Union
 
-from typing import Optional, List, Dict, Any, Union, Sequence
 import numpy as np
 import pandas as pd
 
+_DEPRECATION_MSG = (
+    "modelsummary() is now a thin wrapper over sp.regtable() and will "
+    "be removed in a future minor release. Migrate to "
+    "sp.regtable(*models, ...) for the same output with full control "
+    "over labels, journal templates, and SE formats. "
+    "See docs/rfc/output_pr_b_consolidation.md for migration."
+)
+_DEPRECATION_EMITTED = False
 
[…]
