### 2026-04-29 0821c3e
refactor(output): collapse esttab/EstimateTable to facade over regtable (PR-B/5c)
Final piece of the PR-B/5 facade trio (after outreg2 in 5a and
modelsummary in 5b). The Stata-style ``esttab()`` shipped a ~500-line
``EstimateTable`` class that re-implemented the full renderer pipeline
(text / LaTeX / HTML / Markdown / CSV / DataFrame) — duplicating the
logic already maintained by ``sp.regtable``.

Collapse to a thin facade in ``output/estimates.py``:

- ``esttab()`` now translates Stata-flavoured kwargs and forwards to
  ``sp.regtable``.
- ``EstimateTableResult`` becomes a thin pass-through wr
[…]
