+++ src/statspai/output/estimates.py
@@ -261,517 +261,6 @@ from ._format import (
     fmt_val as _fmt_val,
     fmt_int as _fmt_int,
 )
-
-
-# ---------------------------------------------------------------------------
-# Core table builder
-# ---------------------------------------------------------------------------
-
-class EstimateTable:
-    """Build a comparison table from multiple model results."""
-
-    def __init__(
-        self,
-        models: Sequence[_ModelData],
-        names: Sequence[str],
-        *,
-        se: bool = True,
-        t: bool = False,
-        p: bool = False,
-        ci: bool = False,
-        stars: bool = True,
-        star_levels: Tuple[float, ...] = (0.10, 0.05, 0.01),
-        keep: Optional[Sequence[str]] = None,
-        drop: Optional[Sequence[str]] = None,
-        order: Optional[Sequence[str]] = None,
-        labels: Optional[Dict[str, str]] = None,
-        stats: Optional[Sequence[str]] = None,
-        fmt: str = "%.4f",
-        title: Optional[str] = None,
-        notes: Optional[Sequence[str]] = None,
-        alpha: float = 0.05,
-    ):
-        if not 0.0 < alpha < 1.0:
-            raise ValueError(f"alpha must be in (0, 1), got {alpha!r}")
-        self.models = list(models)
-        self.names = list(names)
-        self.se = se
-        self.t = t
-        self.p = p
-        self.ci = ci
-        self.show_stars = stars
-        self.star_levels = star_levels
-        self.keep = list(keep) if keep else None
-        self.drop = set(drop) if drop else set()
-        self.order = list(order) if order else None
-        self.labels = labels or {}
-        self.requested_stats = list(stats) if stats else ["N", "R2", "adj_R2", "F"]
-        self.fmt = fmt
-        self.title = title
-        self.notes = list(notes) if notes else []
-        self.alpha = float(alpha)
-        self.n_models = len(self.models)
-
-        # Resolve ordered variable list across all models
-        self._vars = self._resolve_vars()
-        # Resolve stat keys
-        self._stat_keys = self._resolve_stat_keys()
-
-    # --- variable ordering ---------------------------------------------------
-
-    def _resolve_vars(self) -> List[str]:
-        seen: OrderedDict[str, None] = OrderedDict()
-        for m in self.models:
-            for v in m.params.index:
-                seen[v] = None
-        all_vars = list(seen)
-
-        if self.keep is not None:
-            all_vars = [v for v in all_vars if v in set(self.keep)]
-        if self.drop:
-            all_vars = [v for v in all_vars if v not in self.drop]
-        if self.order:
-            ordered: List[str] = []
-            remaining = list(all_vars)
-            for v in self.order:
-                if v in remaining:
-                    ordered.append(v)
-                    remaining.remove(v)
-            ordered.extend(remaining)
-            all_vars = ordered
-        return all_vars
-
-    def _resolve_stat_keys(self) -> List[str]:
-        keys: List[str] = []
-        for s in self.requested_stats:
-            canonical = _STAT_ALIASES.get(s, s)
-            if canonical not in keys:
-                keys.append(canonical)
-        return keys
-
-    # --- per-cell formatting --------------------------------------------------
-
-    def _coef_cell(self, model: _ModelData, var: str) -> str:
-        if var not in model.params.index:
-            return ""
-        val = model.params[var]
-        txt = _fmt_val(val, self.fmt)
-        if self.show_stars and var in model.pvalues.index:
-            txt += _format_stars(model.pvalues[var], self.star_levels)
-        return txt
-
-    def _second_row_cell(self, model: _ModelData, var: str) -> str:
-        """Return the parenthetical / bracket row beneath the coefficient."""
[…]
