### 2026-04-29 9f53841
release: 1.11.2 — output layer facade collapse (PR-B/5a-c)
- Collapse outreg2 / modelsummary / esttab → thin facades over regtable
- Bump version to 1.11.2 in pyproject.toml and __init__.py
- Add CHANGELOG entry for [1.11.2]

Co-Authored-By: Claude Opus 4.7 <noreply@anthropic.com>
