+++ src/statspai/__init__.py
@@ -22,7 +22,7 @@ Unified API for causal inference and econometrics:
 >>> sp.outreg2(result, filename="results.xlsx")
 """
 
-__version__ = "1.11.1"
+__version__ = "1.11.2"
 __author__ = "Biaoyue Wang"
 __email__ = "brycew6m@stanford.edu"
 
