### 2026-04-30 b701c15
refactor(output): complete output layer facade collapse
Collapse remaining output facades (esttab, outreg2, modelsummary)
to thin wrappers over regression_table. Unified output dispatch
architecture now complete.

Co-Authored-By: Claude Opus 4.7 <noreply@anthropic.com>
