### 2026-04-30 218b5e5
fix(output): restore graceful degradation for to_excel/to_word
Revert optional-dependency handling from raise ImportError back to
warnings.warn() + return, restoring graceful degradation when
openpyxl / python-docx are absent. Regression introduced in v1.11.2.

Co-Authored-By: Claude Opus 4.7 <noreply@anthropic.com>
