### 2026-04-30 87cbf28
docs: resolve git conflict markers in CHANGELOG.md
