### 2026-04-30 43f3d4c
release: 1.12.0 — DML hardening + TMLE correctness pass
Multi-workstream release. Full notes in CHANGELOG.md `[1.12.0]` and
breaking-change migration guide in MIGRATION.md `v1.11 → v1.12`.

⚠️ Correctness — DML
- IRM / IIVM: KFold → StratifiedKFold (stratified by D / Z); empty
  subgroups now raise IdentificationFailure instead of silently
  filling AIPW scores with zeros.
- panel_dml(binary_treatment=True) is now a deprecated no-op (the old
  classifier path mixed within-demeaned features with raw {0,1}
  labels — no clean E[D̃|X̃] interpretation).
- dml_model_averaging: NaN rows now dropped 
[…]
