+++ paper.bib
@@ -616,6 +616,9 @@
   title={Model Averaging and Double Machine Learning},
   author={Ahrens, Achim and Hansen, Christian B. and Schaffer, Mark E. and Wiemann, Thomas},
   journal={Journal of Applied Econometrics},
+  volume={40},
+  number={3},
+  pages={249--269},
   year={2025},
   doi={10.1002/jae.3103}
 }
@@ -4265,4 +4268,46 @@
   pages   = {2437--2471},
   year    = {2012},
   doi     = {10.1257/aer.102.6.2437}
+}
+
+% --------------------------------------------------------------------
+% HAL-TMLE references (added v1.11.5 audit pass).
+% Verified independently on 2026-04-30:
+%   - li2025regularized       — arxiv.org/abs/2506.17214
+%   - vanderlaan2023efficient — degruyterbrill.com IJB 19(1):261-289,
+%                               doi 10.1515/ijb-2019-0092
+%   - benkeser2016highly      — Crossref api 10.1109/DSAA.2016.93
+% --------------------------------------------------------------------
+
+@misc{li2025regularized,
+  title={Regularized Targeted Maximum Likelihood Estimation in Highly
+         Adaptive Lasso Implied Working Models},
+  author={Li, Yi and Qiu, Sky and Wang, Zeyi and van der Laan, Mark J.},
+  year={2025},
+  eprint={2506.17214},
+  archivePrefix={arXiv},
+  primaryClass={stat.ME},
+  doi={10.48550/arXiv.2506.17214}
+}
+
+@article{vanderlaan2023efficient,
+  title={Efficient estimation of pathwise differentiable target parameters
+         with the undersmoothed highly adaptive lasso},
+  author={van der Laan, Mark J. and Benkeser, David and Cai, Weixin},
+  journal={The International Journal of Biostatistics},
+  volume={19},
+  number={1},
+  pages={261--289},
+  year={2023},
+  doi={10.1515/ijb-2019-0092}
+}
+
+@inproceedings{benkeser2016highly,
+  title={The Highly Adaptive Lasso Estimator},
+  author={Benkeser, David and van der Laan, Mark J.},
+  booktitle={2016 IEEE International Conference on Data Science and
+             Advanced Analytics (DSAA)},
+  pages={689--696},
+  year={2016},
+  doi={10.1109/DSAA.2016.93}
 }
