+++ src/statspai/metalearners/metalearners.py
@@ -1,17 +1,14 @@
 """
 Meta-Learners for heterogeneous treatment effect (CATE) estimation.
 
-Implements five canonical meta-learners that decompose CATE estimation into
-standard supervised-learning sub-problems, following:
+Implements five canonical meta-learners that decompose CATE estimation
+into standard supervised-learning sub-problems. Citations live in
+``paper.bib``; refer to the bib keys below for the canonical record:
 
-- Kunzel, S. R., Seetharam, J. S., Liang, P., & Athey, S. (2019).
-  Metalearners for estimating heterogeneous treatment effects using
-  machine learning. *PNAS*, 116(10), 4156-4165.
-- Kennedy, E. H. (2023). Towards optimal doubly robust estimation of
-  heterogeneous causal effects. *Electronic Journal of Statistics*,
-  17(2), 3008-3049.
-- Nie, X., & Wager, S. (2021). Quasi-oracle estimation of heterogeneous
-  treatment effects. *Biometrika*, 108(2), 299-319.
+- ``@kunzel2019metalearners`` — Künzel, Sekhon, Bickel & Yu (2019),
+  *PNAS* 116(10), 4156-4165.
+- ``@kennedy2023towards``     — Kennedy (2023), *EJS* 17(2).
+- ``@nie2021quasi``           — Nie & Wager (2021), *Biometrika* 108(2).
 
 All learners accept any scikit-learn compatible estimators for the
 nuisance and CATE stages.
@@ -20,7 +17,7 @@ Supported learners
 ------------------
 - **S-Learner** : single model with treatment as feature
 - **T-Learner** : separate models per treatment arm
-- **X-Learner** : two-stage imputed treatment effect (Kunzel et al.)
+- **X-Learner** : two-stage imputed treatment effect (Künzel et al.)
 - **R-Learner** : Robinson decomposition + loss minimisation (Nie & Wager)
 - **DR-Learner**: doubly robust pseudo-outcome regression (Kennedy)
 """
@@ -101,16 +98,76 @@ def _prepare_data(data, y, treat, covariates):
     return Y, D, X, len(Y)
 
 
-def _bootstrap_se(cate_values, n_bootstrap=200, rng=None):
-    """Bootstrap standard error of the ATE from individual CATE estimates."""
-    if rng is None:
-        rng = np.random.default_rng(42)
-    n = len(cate_values)
-    boot_means = np.array([
-        rng.choice(cate_values, size=n, replace=True).mean()
-        for _ in range(n_bootstrap)
-    ])
-    return float(np.std(boot_means, ddof=1))
+def _cross_fit_aipw_phi(
+    X, Y, D, outcome_model, propensity_model, n_folds=5, clip=(0.01, 0.99),
+    seed=42,
+):
+    """Cross-fit AIPW (DR) pseudo-outcome :math:`\\varphi_i`.
+
+    Returns
+    -------
+    phi : ndarray of shape (n,)
+        :math:`\\varphi_i = \\hat\\mu_1(X_i) - \\hat\\mu_0(X_i) +
+        D_i (Y_i - \\hat\\mu_1(X_i)) / \\hat e(X_i) -
+        (1-D_i)(Y_i - \\hat\\mu_0(X_i)) / (1 - \\hat e(X_i))`.
+    diagnostics : dict
+        Underlying nuisance arrays and clip counts so the caller can
+        surface overlap warnings.
+
+    Notes
+    -----
+    This is the **semiparametric efficient** estimating function for
+    :math:`E[Y(1) - Y(0)]` (van der Laan & Robins 2003, Kennedy 2023).
+    :math:`\\hat{\\rm ATE} = \\bar\\varphi` and its asymptotic SE is
+    :math:`\\sigma(\\varphi)/\\sqrt{n}` regardless of which CATE
[…]
