### 2026-04-30 c2feb59
docs(readme): announce v1.12.0 — DML hardening + TMLE correctness
Add a 1.12.0 banner to the top of README.md and README_CN.md
mirroring the existing per-release callout pattern. The v1.8 / v1.5
"NEW in" headlines further down are preserved as feature-level
landmarks; this banner specifically calls out the ⚠️ correctness
changes users should be aware of when upgrading from 1.11.x.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
