+++ README_CN.md
@@ -45,6 +45,39 @@ StatsPAI 聚焦**因果推断**——在这条主线上，我们的目标是成�
 
 ---
 
+**📦 v1.12.0（2026-04-30）— DML 加固 + TMLE 正确性审计**
+
+双线维护版本，完整发布说明见 [`CHANGELOG.md`](CHANGELOG.md) `[1.12.0]`，
+破坏性变更迁移指引见 [`MIGRATION.md`](MIGRATION.md#v111--v112--dml-module-hardening).
+
+- **⚠️ 正确性 — DML**：`sp.dml(model='irm' | 'iivm')` 改用
+  `StratifiedKFold`（之前是 `KFold`）；子组为空的折现在抛
+  `IdentificationFailure`，不再悄悄把 AIPW 分数填零。
+  `sp.dml_panel(binary_treatment=True)` 改为已弃用的 no-op（旧
+  路径在 within-demeaned 特征上拟合 raw {0,1} 标签，得到的倾向得分
+  没有 `E[D̃|X̃]` 的清晰解释）。`sp.dml_model_averaging` 默认权重
+  规则从 `"inverse_risk"` 切换为 `"short_stacking"`（Ahrens-Hansen-
+  Schaffer-Wiemann 2025 *JAE* eq. 7）— 想保持 v1.11 的数字请显式传
+  `weight_rule="inverse_risk"`。PLIV 弱工具变量 partial corr 阈值从
+  `1e-6` 收紧到 `1e-3`，并增加了残差方差比守卫以捕获工具变量与 X
+  完全共线的退化情形。
+- **⚠️ 正确性 — TMLE**：`sp.tmle.SuperLearner` 现在解约束单纯形 QP
+  （之前是 NNLS 加事后归一化，除非偶然，结果都不在单纯形上）。
+  `sp.tmle.ltmle` 截尾半成品已修；`sp.tmle.ltmle_survival` 把 RMST
+  和终止时刻 RD 的影响函数分开（之前两者共用一个非目标-函数化的
+  EIF）。`sp.hal_tmle(variant='projection')` 现在诚实抛
+  `NotImplementedError`，等 Riesz 投影步骤就位（之前是事后对 ε 做
+  ad-hoc 收缩的 silent no-op）。
+- **新增 — DML**：所有 `sp.dml(model=...)` 入口都接受 `random_state=`
+  和 `sample_weight=`。`sample_weight=` 在 PLR / IRM / `sp.dml_panel`
+  / `sp.dml_model_averaging` 上完全支持（统一走 Z-estimator
+  sandwich 方差）；PLIV / IIVM 在传入非平凡权重时抛
+  `NotImplementedError`，等加权 Wald-比率 / IV-PLR 方差推导落地。
+  每个变体的 `model_info["diagnostics"]` 现在都填好了：倾向得分分布、
+  clip 计数、子组兜底次数、partial corr、近似一阶段 F。
+
+---
+
 **🎉 v1.5.0 新版本 — Interference / Conformal / Mendelian 三家族合并升级**
 
 StatsPAI 1.5.0 是 minor 版本，一次性完成 interference、conformal 因果推断、Mendelian 随机化三个家族的三项联动升级：完整家族文档指南、与 `sp.synth` / `sp.decompose` / `sp.dml` 同构的统一 dispatcher，以及一次针对性的正确性审计修掉两处"悄悄算错"。
