### 2026-05-01 bbc1c15
docs(citation): add CITATION.cff, sp.citation(), Zenodo DOI 10.5281/zenodo.18636688
- CITATION.cff at repo root (renders GitHub "Cite this repository" button);
  ships concept DOI 10.5281/zenodo.18636688 and v1.12.0 versioned DOI
  10.5281/zenodo.18636689 under identifiers; preferred-citation block left
  commented for the JOSS article (still under review).
- New sp.citation(format=...) returning bibtex / apa / plain / cff for the
  package itself; sp.__citation__ is the default BibTeX entry as a str.
  Distinct from existing sp.cite() (inline coefficient formatter).
- .zenodo.json so future G
[…]
