### 2026-05-01 5c9abb0
release: 1.12.1 — sp.citation() + Zenodo DOI 10.5281/zenodo.18636688
No numerical or API changes to any estimator. Citation metadata polish:

- sp.citation(format=...) with bibtex / apa / plain / cff outputs (and
  sp.__citation__ as a one-liner str). Distinct from sp.cite() (inline
  coefficient formatter), no collision.
- CITATION.cff at repo root with the Zenodo concept DOI; bundled in the
  sdist via MANIFEST.in.
- Zenodo concept DOI 10.5281/zenodo.18636688 surfaces in sp.citation()
  output, README citation block, and a DOI badge alongside the JOSS
  pending status badge.
- .zenodo.json s
[…]
