+++ CITATION.cff
@@ -16,10 +16,13 @@ authors:
     given-names: "Biaoyue"
     affiliation: "Stanford REAP, Stanford University"
     email: "brycew6m@stanford.edu"
-version: "1.12.0"
+version: "1.12.1"
 date-released: "2026-04-30"
 license: MIT
 # Zenodo *concept* DOI — always resolves to the latest archived version.
+# For a specific-version DOI (e.g. for replication packages), follow the
+# concept DOI link below and pick the matching release in the version
+# history sidebar on Zenodo.
 doi: "10.5281/zenodo.18636688"
 url: "https://github.com/brycewang-stanford/StatsPAI"
 repository-code: "https://github.com/brycewang-stanford/StatsPAI"
@@ -28,9 +31,6 @@ identifiers:
   - type: doi
     value: "10.5281/zenodo.18636688"
     description: "Zenodo concept DOI — always resolves to the latest version."
-  - type: doi
-    value: "10.5281/zenodo.18636689"
-    description: "Zenodo versioned DOI for v1.12.0 (this exact release)."
 keywords:
   - econometrics
   - causal-inference
