### 2026-05-01 4f8a9d1
fix(citation): correct Zenodo DOI to StatsPAI's actual record (10.5281/zenodo.19933900)
The DOI 10.5281/zenodo.18636688 introduced in 1.12.1 (commit 5c9abb0)
was sourced from a Zenodo "DOI Badge" page that actually belonged to a
different repo (constitutional-alignment, GitHub repo ID 1142199519) —
the badge URL was https://zenodo.org/badge/1142199519.svg, which I did
not cross-check against StatsPAI's repo ID (1026640764) before wiring
it into CITATION.cff, sp.citation(), the README badges, and the
GitHub Release notes.

Replacing in this commit, verified via Zenodo API (search q=StatsPAI):


[…]
