+++ README.md
@@ -8,7 +8,7 @@
 [![Tests](https://github.com/brycewang-stanford/statspai/workflows/CI%2FCD%20Pipeline/badge.svg)](https://github.com/brycewang-stanford/statspai/actions)
 [![PyPI Downloads](https://static.pepy.tech/personalized-badge/statspai?period=total&units=INTERNATIONAL_SYSTEM&left_color=BLACK&right_color=GREEN&left_text=downloads)](https://pepy.tech/projects/statspai)
 [![status](https://joss.theoj.org/papers/9f1c837b1b1df7adfcdd538c3698e332/status.svg)](https://joss.theoj.org/papers/9f1c837b1b1df7adfcdd538c3698e332)
-[![DOI](https://zenodo.org/badge/1142199519.svg)](https://doi.org/10.5281/zenodo.18636688)
+[![DOI](https://zenodo.org/badge/1026640764.svg)](https://doi.org/10.5281/zenodo.19933900)
 
 StatsPAI is the **agent-native** Python package for causal inference and applied econometrics. One `import`, **800+ functions**, covering the complete empirical research workflow — from classical econometrics to cutting-edge ML/AI causal methods to publication-ready tables in Word, Excel, and LaTeX.
 
@@ -1217,14 +1217,14 @@ resolves to the latest version):
   title        = {StatsPAI: The Agent-Native Causal Inference \& Econometrics Toolkit for Python},
   year         = {2026},
   version      = {1.12.0},
-  doi          = {10.5281/zenodo.18636688},
-  url          = {https://doi.org/10.5281/zenodo.18636688},
+  doi          = {10.5281/zenodo.19933900},
+  url          = {https://doi.org/10.5281/zenodo.19933900},
   license      = {MIT},
 }
 ```
 
 To cite the **exact version** you used (recommended for replication packages),
-swap in the version-specific DOI shown on the [Zenodo record](https://doi.org/10.5281/zenodo.18636688)
+swap in the version-specific DOI shown on the [Zenodo record](https://doi.org/10.5281/zenodo.19933900)
 for that release.
 
 The canonical metadata lives in [`CITATION.cff`](CITATION.cff) (GitHub renders
