+++ CITATION.cff
@@ -23,13 +23,13 @@ license: MIT
 # For a specific-version DOI (e.g. for replication packages), follow the
 # concept DOI link below and pick the matching release in the version
 # history sidebar on Zenodo.
-doi: "10.5281/zenodo.18636688"
+doi: "10.5281/zenodo.19933900"
 url: "https://github.com/brycewang-stanford/StatsPAI"
 repository-code: "https://github.com/brycewang-stanford/StatsPAI"
 repository-artifact: "https://pypi.org/project/StatsPAI/"
 identifiers:
   - type: doi
-    value: "10.5281/zenodo.18636688"
+    value: "10.5281/zenodo.19933900"
     description: "Zenodo concept DOI — always resolves to the latest version."
 keywords:
   - econometrics
