### 2026-05-01 1926439
Add weighted sample_weight support for PLIV and IIVM
Implement weighted nuisance fitting and Wald variance support for PLIV and IIVM.

Update tests and docs to reflect the new weighted DML coverage.

Generated with OpenAI Codex

Co-authored-by: OpenAI Codex <codex@openai.com>
