+++ src/statspai/dml/pliv.py
@@ -9,7 +9,10 @@ Neyman-orthogonal score:
 DML1 ratio estimator:
     theta = sum(y_tilde * z_tilde) / sum(d_tilde * z_tilde)
 
-SE via influence-function variance of the ratio.
+Weighted variant (with sample_weight w_i):
+    theta = sum(w * y_tilde * z_tilde) / sum(w * d_tilde * z_tilde)
+    Var(theta) = sum(w^2 * psi_i^2) / (sum(w * d_tilde * z_tilde))^2
+where psi_i = (y_tilde_i - theta * d_tilde_i) * z_tilde_i.
 """
 
 import numpy as np
@@ -25,6 +28,7 @@ class DoubleMLPLIV(_DoubleMLBase):
     _REQUIRES_INSTRUMENT = True
     _ML_M_TARGET_BINARY = False
     _ML_R_TARGET_BINARY = False
+    _SUPPORTS_SAMPLE_WEIGHT = True
 
     # First-stage degeneracy threshold on |corr(z̃, d̃)|. Below this
     # the instrument is functionally orthogonal to the residualised
@@ -38,13 +42,17 @@ class DoubleMLPLIV(_DoubleMLBase):
     _FIRST_STAGE_CORR_FLOOR = 1e-3
 
     def _fit_one_rep(self, Y, D, X, Z, n, rng_seed, sample_weight=None):
-        # The base class refuses ``sample_weight`` upstream because
-        # ``_SUPPORTS_SAMPLE_WEIGHT`` is False on this subclass. We
-        # accept the kwarg only to keep the override signature
-        # compatible with the base.
-        del sample_weight
-        from sklearn.base import clone
         from sklearn.model_selection import KFold
+
+        if sample_weight is None:
+            w_full = None
+        else:
+            w_arr = np.asarray(sample_weight, dtype=float)
+            # Nuisance learners can be numerically sensitive to a pure
+            # rescaling of sample_weight even though the target weighted
+            # estimand is not. Normalise to mean 1 so w and c*w define
+            # the same fitting problem in practice.
+            w_full = w_arr * (len(w_arr) / float(np.sum(w_arr)))
 
         kf = KFold(n_splits=self.n_folds, shuffle=True, random_state=rng_seed)
         y_resid = np.zeros(n)
@@ -52,21 +60,35 @@ class DoubleMLPLIV(_DoubleMLBase):
         z_resid = np.zeros(n)
 
         for train_idx, test_idx in kf.split(X):
-            ml_g = clone(self.ml_g)
-            ml_g.fit(X[train_idx], Y[train_idx])
+            w_train = (
+                w_full[train_idx] if w_full is not None else None
+            )
+            ml_g = self._fit_weighted(
+                self.ml_g, X[train_idx], Y[train_idx], w_train,
+            )
             y_resid[test_idx] = Y[test_idx] - ml_g.predict(X[test_idx])
 
-            ml_m = clone(self.ml_m)
-            ml_m.fit(X[train_idx], D[train_idx])
+            ml_m = self._fit_weighted(
+                self.ml_m, X[train_idx], D[train_idx], w_train,
+            )
             d_resid[test_idx] = D[test_idx] - ml_m.predict(X[test_idx])
 
-            ml_r = clone(self.ml_r)
-            ml_r.fit(X[train_idx], Z[train_idx])
+            ml_r = self._fit_weighted(
+                self.ml_r, X[train_idx], Z[train_idx], w_train,
+            )
             z_resid[test_idx] = Z[test_idx] - ml_r.predict(X[test_idx])
 
-        denom = float(np.sum(z_resid * d_resid))
-        sum_z2 = float(np.sum(z_resid ** 2))
-        sum_d2 = float(np.sum(d_resid ** 2))
[…]
