### 2026-05-01 6c535a8
Expand causal question ML routing and robustness reporting
Add direct ML design support to causal_question, share a structured robustness battery across paper/workflow entry points, extend LLM annotator docs/registry for multiclass bootstrap correction, and add the corresponding guide/tests.

Update CHANGELOG and bibliography metadata for the new documentation and dispatch surface.

Generated with OpenAI Codex

Co-authored-by: OpenAI Codex <codex@openai.com>
