+++ paper.bib
@@ -1603,17 +1603,25 @@
   title={Targeted Maximum Likelihood Learning},
   author={van der Laan, Mark J. and Rubin, Daniel},
   journal={The International Journal of Biostatistics},
+  volume={2},
+  number={1},
   year={2006},
   doi={10.2202/1557-4679.1043}
 }
 
-@article{vanderlaan2023efficient,
-  title={Efficient estimation of pathwise differentiable target parameters with the undersmoothed highly adaptive lasso},
-  author={van der Laan, Mark J. and Benkeser, David and Cai, Weixin},
-  journal={The International Journal of Biostatistics},
-  year={2023},
-  doi={10.1515/ijb-2019-0092}
+@book{vanderlaan2003unified,
+  title={Unified Methods for Censored Longitudinal Data and Causality},
+  author={van der Laan, Mark J. and Robins, James M.},
+  series={Springer Series in Statistics},
+  year={2003},
+  publisher={Springer New York},
+  doi={10.1007/978-0-387-21700-0}
 }
+
+% NOTE: ``@article{vanderlaan2023efficient}`` lives in the
+% TMLE-citations block further down (with full volume/number/pages
+% from the IJB record). The earlier entry was a duplicate; deleting
+% it here keeps a single source of truth (CLAUDE.md §10).
 
 @article{lambert1992zero,
   title={Zero-Inflated Poisson Regression, with an Application to Defects in Manufacturing},
@@ -2644,7 +2652,11 @@
   title={Generalized random forests},
   author={Athey, Susan and Tibshirani, Julie and Wager, Stefan},
   journal={The Annals of Statistics},
+  volume={47},
+  number={2},
+  pages={1148--1178},
   year={2019},
+  publisher={Institute of Mathematical Statistics},
   doi={10.1214/18-aos1709}
 }
 
@@ -2792,9 +2804,13 @@
 
 @article{nie2021quasi,
   title={Quasi-oracle estimation of heterogeneous treatment effects},
-  author={Nie, X and Wager, S},
+  author={Nie, Xinkun and Wager, Stefan},
   journal={Biometrika},
+  volume={108},
+  number={2},
+  pages={299--319},
   year={2021},
+  publisher={Oxford University Press},
   doi={10.1093/biomet/asaa076}
 }
 
