+++ src/statspai/question/question.py
@@ -6,9 +6,11 @@ See :mod:`statspai.question` for user-facing docstring.
 
 from __future__ import annotations
 
+import warnings
 from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, Any, List, Optional, Sequence
 
+import numpy as np
 import pandas as pd
 
 if TYPE_CHECKING:
@@ -28,6 +30,8 @@ _VALID_DESIGNS = (
     "auto", "rct", "selection_on_observables", "iv", "natural_experiment",
     "policy_shock", "regression_discontinuity", "synthetic_control",
     "did", "event_study", "longitudinal_observational",
+    # ML-based selection-on-observables (v1.13+):
+    "dml", "tmle", "metalearner", "causal_forest",
 )
 
 
@@ -265,7 +269,34 @@ def causal_question(
     cohort: Optional[str] = None,
     notes: str = "",
 ) -> CausalQuestion:
-    """Declare a causal question (see :class:`CausalQuestion`)."""
+    """Declare a causal question (see :class:`CausalQuestion`).
+
+    Supported ``design`` values
+    ---------------------------
+    Classical / quasi-experimental:
+      - ``'rct'`` — randomised assignment; OLS ATE.
+      - ``'iv'`` / ``'natural_experiment'`` — 2SLS / LATE.
+      - ``'regression_discontinuity'`` — local polynomial RD.
+      - ``'did'`` / ``'event_study'`` — difference-in-differences.
+      - ``'synthetic_control'`` — convex-hull weighting.
+      - ``'longitudinal_observational'`` — MSM / g-formula.
+      - ``'selection_on_observables'`` — AIPW (default).
+
+    ML-based selection-on-observables (v1.13+):
+      - ``'dml'`` — Double/debiased ML for ATE / LATE
+        [chernozhukov2018double].
+      - ``'tmle'`` — Targeted Maximum Likelihood with Super Learner
+        [vanderlaan2006targeted].
+      - ``'metalearner'`` — S/T/X/R/DR-Learner for tau(x);
+        population ATE summary via AIPW influence function
+        [kunzel2019metalearners; nie2021quasi].
+      - ``'causal_forest'`` — honest random forest for tau(x)
+        [athey2019generalized; wager2018estimation]; population
+        ATE inference uses cross-fit AIPW
+        [vanderlaan2003unified; chernozhukov2018double].
+
+    All bib keys above resolve in ``paper.bib``.
+    """
     if estimand not in _VALID_ESTIMANDS:
         raise ValueError(
             f"estimand must be one of {_VALID_ESTIMANDS}; got {estimand!r}"
@@ -406,9 +437,22 @@ def _pick_plan(q: CausalQuestion) -> IdentificationPlan:
             fallback_estimators=["sdid", "augsynth", "scpi"],
         )
     if design == "longitudinal_observational":
+        long_warns = []
+        if q.estimand == "CATE":
+            long_warns.append(
+                "Longitudinal MSM / g-formula returns a population-level "
+                "marginal effect; declared estimand 'CATE' was coerced to "
+                "'ATE'. For heterogeneous longitudinal effects, use "
+                "design='metalearner' on a single time slice."
+            )
+        elif q.estimand == "LATE":
+            long_warns.append(
+                "Longitudinal design declared but estimand='LATE'; the "
+                "MSM / g-formula path identifies ATE, not LATE."
+            )
         return IdentificationPlan(
             estimator="longitudinal.analyze",
-            estimand=q.estimand if q.estimand != "LATE" else "ATE",
+            estimand=q.estimand if q.estimand in ("ATE", "ATT") else "ATE",
             identification_story=(
                 "Sequential exchangeability + positivity given time-"
[…]
