+++ src/statspai/workflow/_robustness.py
@@ -1,0 +1,738 @@
+"""
+Shared robustness battery used by both ``sp.paper`` entry points.
+
+The pre-1.12 split was: ``paper(data, question, ...)`` ran a thin
+:meth:`CausalWorkflow.robustness` (DID pretrend / IV first-stage F /
+ci_width / e-value), and ``paper_from_question`` ran *nothing* —
+its Robustness section was a placeholder pointing the user back to
+``sp.causal``. The user-visible promise ("parallel-trends test,
+leave-one-out, specification curve, e-value, sensitivity") was wider
+than what either entry point actually delivered.
+
+This module consolidates: any caller that has a fitted result and
+optionally the analysis frame can produce a structured
+:class:`RobustnessReport`. The battery is **design-aware** but always
+yields *something* — every sub-check is wrapped so that one failing
+diagnostic never blocks the rest. Failed checks land as
+``severity='check_failed'`` findings so the agent / human reader
+can see what *would* have been reported.
+
+Findings cover:
+
+* ``result.violations()`` — every estimator's self-reported
+  diagnostic flags (first-stage F, propensity overlap, rhat / ESS,
+  pre-trend p, McCrary, ...).
+* Estimate + CI width — universal sanity.
+* DID: ``pretrend_test``, ``bjs_pretrend_joint``.
+* IV: first-stage ``effective_f``-like metrics from
+  ``result.diagnostics``.
+* Observational / DID: e-value via :func:`statspai.evalue_from_result`.
+* Observational with covariates: Oster (2019) coefficient-stability
+  bounds via :func:`statspai.oster_bounds`.
+* Observational with covariates: Cinelli-Hazlett (2020) sensitivity
+  via :func:`statspai.sensemakr`.
+* Any result whose ``model_info["diagnostics"]`` is populated (DML,
+  panel_dml, ...) — the diagnostics block is surfaced verbatim.
+
+The same battery output is rendered into Robustness sections by
+:func:`paper_from_question` (estimand-first) and
+:meth:`CausalWorkflow.robustness` (NL-driven path) so the user gets
+the same content regardless of how they reached the draft.
+"""
+from __future__ import annotations
+
+from dataclasses import dataclass, field
+from typing import Any, Dict, List, Optional
+
+import numpy as np
+import pandas as pd
+
+
+# ---------------------------------------------------------------------- #
+#  Data types
+# ---------------------------------------------------------------------- #
+
+# severity → ordering used by markdown rendering (worst first)
+_SEVERITY_ORDER = {
+    "violation": 0,
+    "warning": 1,
+    "info": 2,
+    "ok": 3,
+    "check_failed": 4,
+}
+
+
+@dataclass
+class RobustnessFinding:
+    """One row in the robustness report.
+
+    Attributes
+    ----------
+    name : str
+        Stable identifier for programmatic access (snake_case).
+    label : str
+        Human-readable label rendered in the Robustness section.
+    value : Any
+        Numeric scalar, mapping, or short string.
+    severity : str
+        One of ``'ok'``, ``'info'``, ``'warning'``, ``'violation'``,
+        ``'check_failed'``. Drives ordering and the leading icon in
+        the markdown rendering.
+    interpretation : str, optional
+        One-line plain-English interpretation; rendered after the
+        value in markdown.
+    """
+    name: str
+    label: str
+    value: Any
+    severity: str = "info"
+    interpretation: Optional[str] = None
+
+    def to_dict(self) -> Dict[str, Any]:
+        d = {
+            "name": self.name,
+            "label": self.label,
+            "value": _coerce_jsonable(self.value),
+            "severity": self.severity,
+        }
+        if self.interpretation:
[…]

[…]
