### 2026-05-01 636fb27
chore(audit): broaden DOI/bibtex detection + DataCite fallback
DOI_RE now also stops at square brackets so Markdown link forms like
`[10.5281/zenodo.X](...)` parse cleanly. verify_crossref falls through
to DataCite on 404 (Zenodo / Figshare / Dryad register there, not
Crossref). is_bibtex marker set widened beyond `author={` / `title={`
so multi-line @software/@misc literals embedded in Python triple-quoted
strings — whose ±3-line context window often misses the author field —
still register as bibtex and skip the partial-author noise check.

Co-Authored-By: Claude Opus 4.7 (1M context) <norep
[…]
