### 2026-05-01 7f146fa
release: 1.12.2 — sp.causal_question ML routing + shared robustness battery + weighted PLIV/IIVM
Patch release on top of v1.12.0's DML hardening. Highlights:

- sp.causal_question(..., design='dml'|'tmle'|'metalearner'|'causal_forest')
  now routes directly into the right estimator with the matching
  identification story / assumptions and design-aware validation
  (DML covariates, PLIV/IIVM scalar-instrument guard, causal-forest
  binary-treatment guard).
- New shared workflow/_robustness.py + run_robustness_battery() —
  both sp.paper(data, question, ...) and sp.paper(CausalQuestion(...))
  
[…]
