+++ README_CN.md
@@ -46,6 +46,17 @@ StatsPAI 聚焦**因果推断**——在这条主线上，我们的目标是成�
 
 ---
 
+**📦 v1.12.2（2026-05-01）— `sp.causal_question` ML 路由 + 共享稳健性 battery + 加权 PLIV/IIVM**
+
+v1.12.0 DML 加固之上的补丁版：`sp.causal_question` 现在直接接受
+`design='dml'|'tmle'|'metalearner'|'causal_forest'` 标签并路由到对应
+估计器，识别故事 / 假设也按设计正确记录；`sp.paper(...)` 两条入口共享
+一份新的 design-aware 稳健性 battery（`workflow/_robustness.py`）；
+`sp.llm_annotator_correct` 升级到多类处理 + 偏差校正自助 + SE 膨胀
+诊断；`sp.dml(model='pliv'|'iivm')` 现已端到端支持 `sample_weight`。
+对 v1.12.0/1.12.1 既有调用无任何数值变化。完整发布说明见
+[`CHANGELOG.md`](CHANGELOG.md) `[1.12.2]`。
+
 **📦 v1.12.0（2026-04-30）— DML 加固 + TMLE 正确性审计**
 
 双线维护版本，完整发布说明见 [`CHANGELOG.md`](CHANGELOG.md) `[1.12.0]`，
@@ -484,7 +495,7 @@ sp.__citation__                 # 与 sp.citation("bibtex") 等价
   author       = {Wang, Biaoyue},
   title        = {StatsPAI: The Agent-Native Causal Inference \& Econometrics Toolkit for Python},
   year         = {2026},
-  version      = {1.12.0},
+  version      = {1.12.2},
   doi          = {10.5281/zenodo.19933900},
   url          = {https://doi.org/10.5281/zenodo.19933900},
   license      = {MIT},
