### 2026-05-01 e22faaf
fix(ci): clear last two CI red lights
* tools/audit_citations.py — chop the phantom-author scope at
  ``[@bibkey]`` markers (new ``_PANDOC_CITE_RE``). Prior bibliography
  entries terminated by a pandoc cite key were leaking their tail into
  the next id's surname scan, producing the lone remaining mismatch:
  ``analytics`` (from "Data Science and Advanced Analytics (DSAA)") was
  being attributed to the next entry at hal_tmle.py:35.

* src/statspai/agent/mcp_server.py — ``_json_default`` returns
  ``PurePath.as_posix()`` instead of ``str()``. On Windows ``str()``
  emits ``\\tmp\\x`` which br
[…]
