### 2026-05-01 #10387 @rich2355
Hi @brycewang-stanford, I am Richard and I am also a causal inference researcher. I will serve as an editor for this submission.

I will work on finding the reviewer soon. At the same time, you have the right to suggest the potential reviewers.

Thanks!
