### 2026-05-01 #10387 @brycewang-stanford
> Hi [@brycewang-stanford](https://github.com/brycewang-stanford), I am Richard and I am also a causal inference researcher. I will serve as an editor for this submission.
>
> I will work on finding the reviewer soon. At the same time, you have the right to suggest the potential reviewers.
>
> Thanks!

Hi @rich2355 — thanks Richard, great to have a fellow causal inference researcher as editor. I'll let you know if I think of suitable reviewer suggestions. Thanks for picking this up!
