### 2026-05-01 0aaef1e
docs(paper): refresh JOSS paper for v1.12.2
- Bump function count 550+ → 950+ (current: 974) and test count
  ~1,600 → ~1,900 (current: 1,894)
- Update date to 1 May 2026 to match v1.12.2 release
- Add v1.10–v1.12 highlights to the agent-native section: estimand-first
  DSL (sp.causal_question), Stata/R script translators behind the MCP
  server, concurrent MCP runner with per-tool timeouts, sp.citation()
  + paper.bib-checked DOI verifier

No new citations introduced; all referenced APIs verified to exist
in the public sp.* namespace before being named in the paper.

Co-Authored-By: Claude Op
[…]
