### 2026-05-04 4e8b6fb
fix: cover 6 zero-coverage modules + fix booktab/twfe bugs
- tests/test_aer_style.py (20) — AER book-tab DOCX/XLSX helpers
- tests/test_next_steps.py (41) — econometric/causal next-step suggestions
- tests/test_auto_tools.py (29) — auto-generated MCP tool manifest
- tests/test_cli.py (31) — CLI dispatch (list/describe/search/help/version)
- Existing test_mcp_runner.py (11) + test_mcp_sampling.py (10) verified
- test_results_full_suite.md — test log updated through all 6 modules

Fixes in source:
- src/statspai/output/_aer_style.py: single-row table border overwrite bug
- src/statspai/core/next
[…]
