+++ tests/test_next_steps.py
@@ -1,0 +1,406 @@
+"""Tests for core/next_steps.py — method-aware next-step recommendations.
+
+Covers
+------
+* ``Step`` dataclass construction, repr, to_dict.
+* ``_format_steps`` and ``_steps_repr_html`` rendering.
+* ``_detect_family`` classification of method strings.
+* ``econometric_next_steps`` for OLS / IV / panel results.
+* ``causal_next_steps`` for DiD / RD / IV / matching / synth / DML / HTE /
+  mediation families — including family-specific steps and universal steps.
+"""
+
+import re
+from typing import Any, Dict, List, Optional
+
+import pytest
+
+from statspai.core.next_steps import (
+    Step,
+    _format_steps,
+    _steps_repr_html,
+    econometric_next_steps,
+    causal_next_steps,
+    _detect_family,
+)
+
+
+# ======================================================================
+#  Step basics
+# ======================================================================
+
+
+class TestStep:
+    def test_defaults(self):
+        s = Step(action="sp.test()", reason="just testing")
+        assert s.action == "sp.test()"
+        assert s.reason == "just testing"
+        assert s.priority == "recommended"
+        assert s.category == "diagnostics"
+
+    def test_custom_priority(self):
+        s = Step("x", "y", priority="essential", category="robustness")
+        assert s.priority == "essential"
+        assert s.category == "robustness"
+
+    def test_repr(self):
+        s = Step("sp.estat(r)", "Post-estimation checks", priority="essential")
+        r = repr(s)
+        assert "sp.estat(r)" in r
+        assert "Post-estimation" in r
+
+    def test_to_dict(self):
+        s = Step("sp.test()", "reason", "optional", "export")
+        d = s.to_dict()
+        assert d["action"] == "sp.test()"
+        assert d["priority"] == "optional"
+        assert d["category"] == "export"
+
+
+# ======================================================================
+#  Formatters
+# ======================================================================
+
+
+class TestFormatSteps:
+    def test_empty(self):
+        output = _format_steps([])
+        assert "Suggested Next Steps" in output
+
+    def test_groups_by_priority(self):
+        steps = [
+            Step("a", "opt", priority="optional"),
+            Step("b", "ess", priority="essential"),
+            Step("c", "rec", priority="recommended"),
+            Step("d", "ess2", priority="essential"),
+        ]
+        output = _format_steps(steps)
+        # Essential first, then recommended, then optional
+        ess_idx = output.index("Essential")
+        rec_idx = output.index("Recommended")
+        opt_idx = output.index("Optional")
+        assert ess_idx < rec_idx < opt_idx
+
+
+class TestStepsReprHtml:
+    def test_contains_html_structure(self):
+        steps = [
+            Step("sp.test()", "A test step", priority="essential", category="diagnostics"),
+        ]
+        html = _steps_repr_html(steps)
+        assert "sp-ns" in html
+        assert "sp.test()" in html
+        assert "essential" in html
+
+    def test_multiple_categories(self):
+        steps = [
+            Step("a", "diag", category="diagnostics"),
+            Step("b", "rob", category="robustness"),
+            Step("c", "exp", category="export"),
+        ]
+        html = _steps_repr_html(steps)
+        assert "Diagnostics" in html
+        assert "Robustness" in html
+        assert "Export" in html
+
+    def test_empty(self):
+        html = _steps_repr_html([])
+        assert "sp-ns" in html
+
+
+# ======================================================================
+#  _detect_family
+# ======================================================================
+
+
+class TestDetectFamily:
+    def test_synth_before_did(self):
+        """sdid must match synth, not did (order matters)."""
[…]
