+++ tests/test_cli.py
@@ -1,0 +1,225 @@
+"""Tests for cli.py — StatsPAI CLI entry point.
+
+Covers
+------
+* ``_make_parser`` — argument parser structure (subcommands, flags).
+* ``main`` — dispatch logic for each subcommand, return codes, error
+  handling, JSON vs text output.
+"""
+
+from unittest.mock import MagicMock, patch
+
+import pytest
+
+import statspai
+from statspai.cli import _make_parser, main
+
+
+# ======================================================================
+#  _make_parser
+# ======================================================================
+
+
+class TestMakeParser:
+    def test_top_level_help(self):
+        parser = _make_parser()
+        assert parser.prog == "statspai"
+
+    def test_version_flag(self):
+        parser = _make_parser()
+        args = parser.parse_args(["--version"])
+        assert args.version is True
+
+    def test_version_flag_short(self):
+        parser = _make_parser()
+        args = parser.parse_args(["-V"])
+        assert args.version is True
+
+    def test_list_subcommand(self):
+        parser = _make_parser()
+        args = parser.parse_args(["list"])
+        assert args.command == "list"
+        assert args.category is None
+        assert args.json is False
+
+    def test_list_with_category(self):
+        parser = _make_parser()
+        args = parser.parse_args(["list", "--category", "causal"])
+        assert args.command == "list"
+        assert args.category == "causal"
+
+    def test_list_with_json(self):
+        parser = _make_parser()
+        args = parser.parse_args(["list", "--json"])
+        assert args.json is True
+
+    def test_describe_subcommand(self):
+        parser = _make_parser()
+        args = parser.parse_args(["describe", "did"])
+        assert args.command == "describe"
+        assert args.name == "did"
+
+    def test_describe_with_json(self):
+        parser = _make_parser()
+        args = parser.parse_args(["describe", "iv", "--json"])
+        assert args.json is True
+
+    def test_search_subcommand(self):
+        parser = _make_parser()
+        args = parser.parse_args(["search", "causal", "inference"])
+        assert args.command == "search"
+        assert args.query == ["causal", "inference"]
+
+    def test_search_with_json(self):
+        parser = _make_parser()
+        args = parser.parse_args(["search", "did", "--json"])
+        assert args.json is True
+
+    def test_help_subcommand_no_topic(self):
+        parser = _make_parser()
+        args = parser.parse_args(["help"])
+        assert args.command == "help"
+        assert args.topic is None
+
+    def test_help_subcommand_with_topic(self):
+        parser = _make_parser()
+        args = parser.parse_args(["help", "did"])
+        assert args.command == "help"
+        assert args.topic == "did"
+
+    def test_help_with_verbose(self):
+        parser = _make_parser()
+        args = parser.parse_args(["help", "iv", "--verbose"])
+        assert args.verbose is True
+
+    def test_help_short_verbose(self):
+        parser = _make_parser()
+        args = parser.parse_args(["help", "iv", "-v"])
+        assert args.verbose is True
+
+    def test_version_subcommand(self):
+        parser = _make_parser()
+        args = parser.parse_args(["version"])
+        assert args.command == "version"
+
+    def test_no_args(self):
+        """No args → command is None."""
+        parser = _make_parser()
+        args = parser.parse_args([])
+        assert args.command is None
+        assert args.version is False
+
+
+# ======================================================================
+#  main
+# ======================================================================
+
+
+class TestMain:
+    def test_version(self, capsys):
+        with patch.object(statspai, "__version__", "1.2.3", create=True):
+            rc = main(["--version"])
+            assert rc == 0
+            out, _ = capsys.readouterr()
[…]
