### 2026-05-04 14d05cf
docs(readme): switch DOI badge to direct SVG URL
GitHub camo proxy was caching a failed fetch on the redirecting
zenodo.org/badge/<repo-id>.svg URL. Use the direct
zenodo.org/badge/DOI/<doi>.svg endpoint (no 302) so the badge renders
reliably. Concept DOI link target unchanged.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
