### 2026-05-04 5a0bec6
fix(workflow): replace silent except-pass with WorkflowDegradedWarning + structured records
Phase A of the §3.7 "fail loud" cleanup.  workflow/paper/smart had 12.5
broad-except-per-kLOC density (5x the dml/tmle core), producing draft
papers and audits that looked complete but had silently dropped CI
lines, citations, DAG appendices, provenance, sensitivity panels, etc.

Adds workflow/_degradation.py exposing WorkflowDegradedWarning and
record_degradation(target, *, section, exc, detail) — warn loudly +
append a structured {section, error_type, message} entry to
target.degradations (PaperDraft 
[…]
