+++ tests/test_workflow_degradations.py
@@ -1,0 +1,279 @@
+"""Tests for the workflow-layer degradation tracking surface.
+
+CLAUDE.md §3.7 ("失败要响亮") forbids silent ``except Exception: pass``
+in best-effort orchestration code.  These tests pin down the contract:
+when a sub-step of ``sp.paper`` / ``sp.causal_workflow`` /
+``sp.assumption_audit`` fails, the user must see a
+``WorkflowDegradedWarning`` *and* be able to introspect what was skipped
+via ``draft.degradations`` / ``workflow.degradations``.
+
+The tests deliberately trigger failures by feeding pathological inputs
+(non-numeric covariates, CI shaped wrong, ``cite()`` raising, etc.)
+rather than by mocking, so they double as regression tests against
+future code paths that try to silently swallow these errors again.
+"""
+from __future__ import annotations
+
+import warnings
+
+import numpy as np
+import pandas as pd
+import pytest
+
+from statspai.workflow._degradation import (
+    WorkflowDegradedWarning,
+    record_degradation,
+)
+
+
+# --------------------------------------------------------------------- #
+#  Helper itself
+# --------------------------------------------------------------------- #
+
+
+def test_record_degradation_warns_and_returns_entry():
+    bag: list = []
+    with pytest.warns(WorkflowDegradedWarning, match="custom section"):
+        entry = record_degradation(
+            bag,
+            section="custom section",
+            exc=ValueError("boom"),
+            detail="ctx=1",
+        )
+    assert entry["section"] == "custom section"
+    assert entry["error_type"] == "ValueError"
+    assert entry["message"] == "boom"
+    assert entry["detail"] == "ctx=1"
+    assert bag == [entry]
+
+
+def test_record_degradation_appends_to_object_attr():
+    class Holder:
+        def __init__(self):
+            self.degradations: list = []
+
+    h = Holder()
+    with pytest.warns(WorkflowDegradedWarning):
+        record_degradation(h, section="x", exc=RuntimeError("y"))
+    assert len(h.degradations) == 1
+    assert h.degradations[0]["error_type"] == "RuntimeError"
+
+
+def test_record_degradation_tolerates_no_target():
+    """Passing target=None still warns; just nowhere to record."""
+    with pytest.warns(WorkflowDegradedWarning):
+        record_degradation(None, section="loose", exc=KeyError("k"))
+
+
+# --------------------------------------------------------------------- #
+#  _eda_block — covariate balance table
+# --------------------------------------------------------------------- #
+
+
+def test_eda_block_records_when_balance_table_fails():
+    """A non-numeric covariate breaks .mean(); the EDA section must
+    still render and the failure must surface."""
+    from statspai.workflow.paper import _eda_block
+
+    df = pd.DataFrame({
+        "y": [1.0, 2.0, 3.0, 4.0],
+        "treated": [0, 1, 0, 1],
+        # ``string_col`` is not numeric → groupby().mean() raises.
+        "string_col": ["a", "b", "c", "d"],
+    })
+    bag: list = []
+    with pytest.warns(
+        WorkflowDegradedWarning, match="EDA covariate balance table"
+    ):
+        out = _eda_block(
+            df, y="y", treatment="treated",
+            covariates=["string_col"], degradations=bag,
+        )
+    # Section still renders something useful (sample size + missingness).
+    assert "Sample size" in out
+    assert len(bag) == 1
+    assert bag[0]["section"] == "EDA covariate balance table"
+    # Partial markdown rows ("| covariate |") must NOT leak when the
+    # build aborted — the dropped-table fix in _eda_block guarantees
+    # all-or-nothing balance output.
+    assert "| covariate |" not in out
[…]
