### 2026-05-04 f1dae20
feat(registry): add stability tiers + per-function limitations for parity-grade vs frontier-grade clarity
Surface the parity-grade vs frontier-grade distinction as first-class
machine-readable metadata so users (and LLM agents) can tell which
functions are numerically aligned with R/Stata or analytic references
vs which are MVP / RFC-tracked frontier work, and which specific
parameter values are documented as not-yet-implemented inside an
otherwise stable function.

Two orthogonal axes on every FunctionSpec:

* stability ('stable' | 'experimental' | 'deprecated', exposed as
  sp.STABILITY_TIER
[…]
