+++ tests/test_stability.py
@@ -1,0 +1,212 @@
+"""Tests for the v1.13 stability-tier surface.
+
+Two distinct facets are exercised:
+
+* function-level ``stability`` (``stable`` / ``experimental`` / ``deprecated``)
+* per-function ``limitations`` (variant-level gaps inside otherwise
+  stable functions, e.g. ``hal_tmle(variant='projection')``)
+
+Both must surface through the registry, the OpenAI tool-schema
+serializer, the help layer, and the CLI list filter.
+"""
+from __future__ import annotations
+
+import pytest
+
+
+class TestStabilitySchema:
+    def test_tiers_constant_exposed(self):
+        import statspai as sp
+
+        assert "stable" in sp.STABILITY_TIERS
+        assert "experimental" in sp.STABILITY_TIERS
+        assert "deprecated" in sp.STABILITY_TIERS
+
+    def test_default_stability_is_stable(self):
+        from statspai.registry import FunctionSpec
+
+        spec = FunctionSpec(name="x", category="utils", description="x")
+        assert spec.stability == "stable"
+        assert spec.limitations == []
+
+    def test_invalid_stability_raises_at_construction(self):
+        from statspai.registry import FunctionSpec
+
+        with pytest.raises(ValueError, match="stability="):
+            FunctionSpec(
+                name="x", category="utils", description="x",
+                stability="bogus_tier",
+            )
+
+
+class TestStabilityFlowsThroughRegistry:
+    """to_dict / agent_card / OpenAI schema must all carry the new fields."""
+
+    def test_to_dict_includes_stability_and_limitations(self):
+        from statspai import describe_function
+
+        d = describe_function("hal_tmle")
+        assert d["stability"] == "stable"
+        assert d["limitations"], (
+            "hal_tmle should advertise its projection-variant gap in "
+            "limitations so agents can route around it without "
+            "burning a tool-call on NotImplementedError"
+        )
+        # The flagship example from the user's framing — must be visible
+        # in the schema.
+        assert any("projection" in lim for lim in d["limitations"])
+
+    def test_principal_strat_advertises_instrument_gap(self):
+        from statspai import describe_function
+
+        d = describe_function("principal_strat")
+        assert d["stability"] == "stable"
+        assert any(
+            "instrument" in lim.lower() for lim in d["limitations"]
+        ), (
+            "principal_strat must advertise that the instrument= "
+            "two-layer setup is not yet implemented"
+        )
+
+    def test_agent_card_includes_stability_and_limitations(self):
+        from statspai import agent_card
+
+        card = agent_card("hal_tmle")
+        assert card["stability"] == "stable"
+        assert "limitations" in card
+        assert card["limitations"]
+
+    def test_openai_schema_prefixes_experimental_in_description(self):
+        from statspai import function_schema
+
+        schema = function_schema("text_treatment_effect")
+        assert schema["description"].startswith("[experimental]"), (
+            "An LLM tool-caller reads only the schema description; the "
+            "experimental tag must travel with it"
+        )
+
+    def test_openai_schema_appends_limitations_to_description(self):
+        from statspai import function_schema
+
+        schema = function_schema("hal_tmle")
+        assert "Known limitations:" in schema["description"]
+        assert "projection" in schema["description"]
+
+
+class TestStabilityFilters:
+    def test_list_functions_stability_filter_stable(self):
+        from statspai import list_functions
+
+        stable = list_functions(stability="stable")
+        # Most of the catalogue is stable, so the filtered set should
+        # be much larger than the experimental bucket but still strictly
+        # smaller than the unfiltered set (since at least one function
+        # is experimental).
+        all_fns = list_functions()
+        assert len(stable) >= 100
+        assert len(stable) < len(all_fns)
[…]
