+++ src/statspai/registry.py
@@ -57,6 +57,12 @@ class FailureMode:
         return asdict(self)
 
 
+#: Allowed stability tiers (kept as a frozenset so ``in`` checks are O(1)
+#: and the canonical list lives in one place — ``sp.help``, the CLI
+#: filter, and the post-init validator all read from here).
+STABILITY_TIERS: frozenset = frozenset({"stable", "experimental", "deprecated"})
+
+
 @dataclass
 class FunctionSpec:
     """Machine-readable specification for a StatsPAI function.
@@ -65,6 +71,26 @@ class FunctionSpec:
     ``alternatives`` / ``typical_n_min`` / ``pre_conditions``) are
     optional — any entry without them still renders correctly; only
     the agent-card layer surfaces the extras.
+
+    Stability layering
+    ------------------
+    Two fields make the parity-grade vs. frontier-grade split visible
+    to humans and agents *before* a call is made:
+
+    * ``stability`` (``"stable"`` | ``"experimental"`` | ``"deprecated"``)
+      classifies the function as a whole.  ``"stable"`` (default) means
+      numerically aligned with R / Stata or an analytic reference and
+      the public signature is locked.  ``"experimental"`` means the
+      method is implemented but not (yet) parity-tested or the API may
+      shift between minor versions — agents should hold its outputs to
+      a lower trust bar.
+    * ``limitations`` enumerates **partial-implementation gaps inside
+      an otherwise stable function** — typically a parameter value that
+      raises :class:`NotImplementedError` (e.g.
+      ``hal_tmle(variant='projection')``) or a feature combination
+      that is documented as not yet supported.  This lets agents see
+      the gap from ``sp.describe_function`` instead of discovering it
+      mid-pipeline by exception.
     """
     name: str
     category: str
@@ -87,9 +113,33 @@ class FunctionSpec:
     """Ranked ``sp.xxx`` fallbacks when this estimator is a poor fit."""
     typical_n_min: Optional[int] = None
     """Rule-of-thumb minimum sample size; ``None`` if not applicable."""
+    # ------------------------------------------------------------------ #
+    #  Stability layering (parity-grade vs. frontier-grade)
+    # ------------------------------------------------------------------ #
+    stability: str = "stable"
+    """Maturity tier — see :data:`STABILITY_TIERS`."""
+    limitations: List[str] = field(default_factory=list)
+    """Known-unimplemented variants / parameter values inside an
+    otherwise stable function.  Each entry is one short sentence; the
+    canonical pattern is ``"<param>=<value>: <what's missing>"``."""
+
+    def __post_init__(self) -> None:
+        # Validate stability tier early so a typo fails at import / first
+        # ``register()`` call, not at the moment an agent filters on it.
+        if self.stability not in STABILITY_TIERS:
+            raise ValueError(
+                f"FunctionSpec(name={self.name!r}).stability={self.stability!r} "
+                f"is not one of {sorted(STABILITY_TIERS)}"
+            )
 
     def to_openai_schema(self) -> Dict[str, Any]:
-        """Export as OpenAI function-calling compatible JSON schema."""
+        """Export as OpenAI function-calling compatible JSON schema.
+
+        The ``description`` is prefixed with a stability marker
+        (``[experimental]`` / ``[deprecated]``) and any known
+        ``limitations`` are appended so an LLM tool-caller sees the
+        gap inside the same field it already reads.
+        """
         properties = {}
         required = []
         for p in self.params:
@@ -113,9 +163,16 @@ class FunctionSpec:
             if p.required:
                 required.append(p.name)
 
+        description = self.description
+        if self.stability != "stable":
+            description = f"[{self.stability}] {description}"
+        if self.limitations:
[…]
