### 2026-05-04 50bf390
docs(stats): replace stale 889/925 with live registry numbers + canonical generator
Eliminate the multi-source drift between registry.py / help.py / README.md
that was already partially fixed in HEAD (f1dae20 unified bayes/bayesian
and gformula categories, drained the 'other' bucket, and rewrote the
dishonest "kept in sync" comments). This patch finishes the job by
making the docs self-update via a single canonical generator.

- New: scripts/registry_stats.py — prints live function count, submodule
  count, per-module LOC/files/registered table, and per-category counts.
  Modes: default summar
[…]
