+++ scripts/registry_stats.py
@@ -1,0 +1,229 @@
+"""Print live registry / module statistics for StatsPAI.
+
+This script is the **single source of truth** for the function-count,
+submodule-count, and per-module breakdown numbers that appear in
+``README.md``, ``README_CN.md``, ``CLAUDE.md``, ``docs/stats.md``, and
+``docs/index.md``. Run it before a release (or when those numbers feel
+stale) and copy the relevant figures into the markdown files.
+
+Usage
+-----
+    python scripts/registry_stats.py                 # human-readable summary
+    python scripts/registry_stats.py --table         # docs/stats.md per-module table
+    python scripts/registry_stats.py --json          # machine-readable
+    python scripts/registry_stats.py --check         # exit non-zero if README.md is stale
+
+The ``--check`` mode lets CI flag drift between the live registry and
+the loose floors quoted in ``README.md`` / ``README_CN.md``. It does not
+require the floors to be exact; it only flags when the live count drops
+below the documented floor (a regression) or exceeds it by enough that
+the documented number is misleading.
+"""
+from __future__ import annotations
+
+import argparse
+import json
+import os
+import re
+import sys
+from collections import Counter
+from pathlib import Path
+from typing import Dict, Tuple
+
+
+REPO_ROOT = Path(__file__).resolve().parent.parent
+SRC_ROOT = REPO_ROOT / "src" / "statspai"
+TESTS_ROOT = REPO_ROOT / "tests"
+
+
+def _module_loc_and_files() -> Dict[str, Tuple[int, int]]:
+    """Walk src/statspai/<submodule>/ and count .py LOC + file count."""
+    out: Dict[str, Tuple[int, int]] = {}
+    for entry in sorted(SRC_ROOT.iterdir()):
+        if not entry.is_dir() or entry.name.startswith("__"):
+            continue
+        loc = files = 0
+        for path in entry.rglob("*.py"):
+            files += 1
+            try:
+                loc += sum(1 for _ in path.open("r", encoding="utf-8"))
+            except OSError:
+                pass
+        out[entry.name] = (loc, files)
+    return out
+
+
+def _tree_loc(root: Path) -> Tuple[int, int]:
+    loc = files = 0
+    for path in root.rglob("*.py"):
+        files += 1
+        try:
+            loc += sum(1 for _ in path.open("r", encoding="utf-8"))
+        except OSError:
+            pass
+    return loc, files
+
+
+def _registered_per_module() -> Tuple[Counter, int]:
+    """Map top-level submodule → number of registered ``sp.*`` symbols."""
+    sys.path.insert(0, str(REPO_ROOT / "src"))
+    import statspai as sp  # noqa: WPS433  (side-effect import is the point)
+    sp.list_functions()  # force full registry build
+    from statspai.registry import _REGISTRY  # noqa: WPS433
+
+    counts: Counter = Counter()
+    for name in _REGISTRY:
+        obj = getattr(sp, name, None)
+        mod = getattr(obj, "__module__", "") or ""
+        if mod.startswith("statspai."):
+            counts[mod.split(".")[1]] += 1
+    return counts, len(_REGISTRY)
+
+
+def collect() -> dict:
+    sys.path.insert(0, str(REPO_ROOT / "src"))
+    import statspai as sp  # noqa: WPS433
+
+    per_mod = _module_loc_and_files()
+    fn_counts, total_fns = _registered_per_module()
+    src_loc, src_files = _tree_loc(SRC_ROOT)
+    tests_loc, tests_files = _tree_loc(TESTS_ROOT) if TESTS_ROOT.exists() else (0, 0)
+
+    cat_counts: Counter = Counter()
[…]
