+++ README_CN.md
@@ -10,7 +10,7 @@
 [![status](https://joss.theoj.org/papers/9f1c837b1b1df7adfcdd538c3698e332/status.svg)](https://joss.theoj.org/papers/9f1c837b1b1df7adfcdd538c3698e332)
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.19933900.svg)](https://doi.org/10.5281/zenodo.19933900)
 
-StatsPAI 是一个**面向 AI Agent** 的 Python 因果推断与应用计量经济学工具包。一个 `import`，**800+ 个函数**，覆盖从经典计量经济学到前沿 ML/AI 因果推断方法，再到论文级 Word、Excel、LaTeX 输出表格的完整实证研究流程。
+StatsPAI 是一个**面向 AI Agent** 的 Python 因果推断与应用计量经济学工具包。一个 `import`，**950+ 个注册函数**，分布在 **80+ 个子模块**（实时数量请运行 `python scripts/registry_stats.py`），覆盖从经典计量经济学到前沿 ML/AI 因果推断方法，再到论文级 Word、Excel、LaTeX 输出表格的完整实证研究流程。
 
 **为 AI Agent 而生**：每个函数都返回结构化结果对象，附带自描述 schema（`list_functions()`、`describe_function()`、`function_schema()`），使 StatsPAI 成为首个专为 LLM 驱动的研究流程设计的计量工具包——同时对人类研究者也完全友好。
 
@@ -42,7 +42,7 @@ StatsPAI 聚焦**因果推断**——在这条主线上，我们的目标是成�
 
 **图例**：🏆 跨生态最完整 · ✅ 完整覆盖 · ⚠️ 部分 / 分散 / 单算法 · ❌ 无。
 
-**StatsPAI 一句话概览**：889 个注册函数 · 78 个模块 · 188,244 行核心代码 + 42,768 行测试。完整覆盖矩阵（23 个方法家族）、按模块拆分、以及跨生态行数对比，详见 [`docs/stats.md`](docs/stats.md)。
+**StatsPAI 一句话概览**：live agent registry 中有 950+ 个注册函数 · 80+ 个子模块 · ~230k 行核心代码 + ~70k 行测试。这四个数字都可以由唯一的生成器 (`python scripts/registry_stats.py`) 现场复算；[`docs/stats.md`](docs/stats.md) 中的按模块拆分表也由同一个脚本回写。完整覆盖矩阵（23 个方法家族）以及跨生态行数对比，详见 [`docs/stats.md`](docs/stats.md)。
 
 ---
 
