+++ docs/stats.md
@@ -2,7 +2,9 @@
 
 > This page tracks StatsPAI's size and coverage against the broader statistical ecosystem, in the spirit of *evidence-based* positioning. Numbers marked **"measured"** are reproducible on a standard install; numbers marked **"estimated"** are extrapolated from public ecosystem statistics with the reasoning shown.
 >
-> *Last measured: 2026-04-21 on macOS arm64 against `statspai` 0.9.16, `statsmodels` 0.14.x, `linearmodels` (latest), Stata 18, R 4.5.2.*
+> *Last measured: 2026-05-03 on macOS arm64 against `statspai` 1.12.2, `statsmodels` 0.14.x, `linearmodels` (latest), Stata 18, R 4.5.2.*
+>
+> The four StatsPAI numbers (function count, submodule count, source LOC, test LOC) are reproducible from a single canonical generator: `python scripts/registry_stats.py`. The per-module table in §2 is regenerated by `python scripts/registry_stats.py --table`. CI guards against drift via `python scripts/registry_stats.py --check`.
 
 ---
 
@@ -10,8 +12,8 @@
 
 | Ecosystem / Project                  | Method     |  Files | Lines of code | Primary focus                      |
 | ------------------------------------ | ---------- | -----: | ------------: | ---------------------------------- |
-| **StatsPAI** `src/statspai/`         | measured   |    528 |   **188,244** | full-stack causal inference        |
-| StatsPAI tests (`tests/`)            | measured   |    255 |        42,768 | —                                  |
+| **StatsPAI** `src/statspai/`         | measured   |    620 |   **231,858** | full-stack causal inference        |
+| StatsPAI tests (`tests/`)            | measured   |    375 |        71,729 | —                                  |
 | statsmodels 0.14.x                   | measured   |    948 |   **381,981** | GLM / time series / general        |
 | linearmodels                         | measured   |    131 |        36,607 | panel / IV                         |
 | **Python causal-inference subtotal** |            |  1,079 |   **418,588** |                                    |
@@ -33,55 +35,54 @@ LOC is a vanity metric in isolation — what matters is **coverage density** wit
 
 ## 2 · StatsPAI per-module breakdown
 
-Sorted by LOC. All numbers measured on `src/statspai/` at the commit listed in the timestamp above.
+Sorted by LOC. The four-column block (LOC / Files / Registered) is regenerated by `python scripts/registry_stats.py --table`; the **Focus** column is editorial and curated by hand.
 
 | Module              | LOC    | Files | Registered functions (`sp.*`) | Focus                         |
 | ------------------- | -----: | ----: | ----------------------------: | ----------------------------- |
-| `synth`             | 17,428 |    30 |                            50 | Synthetic Control family      |
-| `did`               | 13,926 |    28 |                            53 | DiD + event-study             |
-| `rd`                | 11,169 |    21 |                            47 | Regression Discontinuity      |
-| `regression`        | 10,597 |    18 |                            36 | OLS / GLM / Logit / Probit    |
-| `decomposition`     |  5,828 |    16 |                            29 | Oaxaca / RIF / inequality     |
-| `diagnostics`       |  5,372 |    12 |                            20 | assumption checks             |
-| `output`            |  5,072 |     8 |                            14 | Word / Excel / LaTeX / HTML   |
-| `smart`             |  5,028 |    10 |                            19 | workflow orchestration        |
-| `plots`             |  4,925 |     6 |                             — | interactive plot editor       |
-| `iv`                |  4,815 |    15 |                             4 | 2SLS / LIML / GMM / Kernel IV |
-| `panel`             |  4,511 |    12 |                            17 | FE / RE / Arellano-Bond       |
[…]

[…]
