### 2026-05-04 7eeb624
refactor(api): lazy-load optional families and surface degradation warnings
