### 2026-05-04 f9ec214
fix(api): preserve same-name function bindings under lazy submodule registry
Codex's lazy-load refactor (7eeb624) routed sp.proximal / sp.principal_strat /
sp.bridge / sp.bcf / sp.bunching / sp.dose_response / sp.multi_treatment /
sp.causal_impact / sp.frontier / sp.interference / sp.tmle / sp.msm /
sp.deepiv / sp.bartik through _LAZY_SUBMODULES, which returned the *module*
object on first access. Each of those names is also exported as a same-named
*function* (the original eager `from .X import X, ...` shadowed the module
attribute with the function). Tests doing `sp.proximal(...)` started fa
[…]
