### 2026-05-04 6197956
test(coverage): add 131-test battery for 4 worst-covered files + cross-family smoke + fix CausalResult.summary KeyError on wooldridge_did
Closes the v1.12.x ultrareview gap on parity-grade depth. Five new test
files raise per-file coverage:

- synth/report.py            4% → 81%
- workflow/paper.py         66% → 86%
- did/wooldridge_did.py     76% → 93%
- did/did_imputation.py     85% → 99%

Plus a 30-test cross-family smoke battery exercising every headline
estimator's CI/SE/point-estimate contract across rd / iv / tmle /
synth / did / bayes.

Real bug fixed in core/results.py:CausalResult.su
[…]
