+++ tests/test_did_imputation_branches.py
@@ -1,0 +1,221 @@
+"""Branch coverage for ``statspai.did.did_imputation``.
+
+Closes the validation / event-study / control-demeaning gaps surfaced by
+the v1.12.x coverage audit. Existing files (``test_bjs_joint``,
+``test_did_summary``, ``test_estimator_provenance_round2``) cover the
+default ATT path; this file adds:
+
+- All four ``ValueError`` guards (missing y/group/time/first_treat,
+  missing control column, no treated obs, no untreated obs).
+- The ``controls != []`` demeaning + beta-recovery branch.
+- The horizon / event-study aggregation branch with the joint
+  pre-trend chi-squared test.
+- The internal ``_ols_coef`` helper's empty-X early return.
+- The horizon SE helper's ``N_k == 0`` early return.
+"""
+from __future__ import annotations
+
+import warnings
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+from statspai.did.did_imputation import (
+    _cluster_se_horizon,
+    _ols_coef,
+    did_imputation,
+)
+
+
+# --------------------------------------------------------------------- #
+#  Fixtures
+# --------------------------------------------------------------------- #
+
+
+@pytest.fixture(scope="module")
+def staggered_panel():
+    df = sp.dgp_did(n_units=80, n_periods=8, staggered=True, seed=11)
+    rng = np.random.default_rng(0)
+    df = df.copy()
+    df["x"] = rng.normal(size=len(df))
+    return df
+
+
+# --------------------------------------------------------------------- #
+#  Validation guards
+# --------------------------------------------------------------------- #
+
+
+@pytest.mark.parametrize(
+    "missing_arg, kwargs",
+    [
+        ("y", {"y": "no_such_col"}),
+        ("group", {"group": "no_such_col"}),
+        ("time", {"time": "no_such_col"}),
+        ("first_treat", {"first_treat": "no_such_col"}),
+    ],
+)
+def test_did_imputation_validates_columns(staggered_panel, missing_arg, kwargs):
+    base = dict(y="y", group="unit", time="time", first_treat="first_treat")
+    base.update(kwargs)
+    with pytest.raises(ValueError, match="not found in data"):
+        did_imputation(staggered_panel, **base)
+
+
+def test_did_imputation_validates_control_columns(staggered_panel):
+    with pytest.raises(ValueError, match="Control column"):
+        did_imputation(
+            staggered_panel,
+            y="y", group="unit", time="time", first_treat="first_treat",
+            controls=["bogus"],
+        )
+
+
+def test_did_imputation_no_treated_raises(staggered_panel):
+    df = staggered_panel.copy()
+    df["first_treat"] = np.inf  # everyone never-treated
+    with pytest.raises(ValueError, match="No treated observations"):
+        did_imputation(
+            df, y="y", group="unit", time="time", first_treat="first_treat",
+        )
+
+
+def test_did_imputation_no_untreated_raises():
+    """Every obs is post-treatment (first_treat = -1) → no untreated."""
+    rng = np.random.default_rng(0)
+    n_units, n_per = 5, 4
+    rows = []
+    for u in range(n_units):
+        for t in range(n_per):
+            rows.append({
+                "unit": u, "time": t, "y": rng.normal(),
+                # first_treat smaller than every time observed → all treated
+                "first_treat": -1.0,
+            })
+    df = pd.DataFrame(rows)
+    with pytest.raises(ValueError, match="No untreated observations"):
+        did_imputation(
[…]
