+++ tests/test_paper_branches.py
@@ -1,0 +1,435 @@
+"""Branch coverage for ``statspai.workflow.paper``.
+
+Existing tests (``test_paper_pipeline.py`` / ``test_paper_quarto.py`` /
+``test_paper_tables.py`` / ``test_paper_from_question.py``) already cover
+the headline ``sp.paper(...)`` happy path. This file adds tests aimed at
+the remaining gaps in the v1.12.x coverage report:
+
+- Internal helpers: ``_yaml_str``, ``_tex_escape``, ``_md_to_tex``,
+  ``_inline_md_to_tex``, ``_render_dag_section``, ``_record_note``,
+  ``_notes_block``.
+- ``PaperDraft.to_qmd`` with explicit author / formats / bibliography /
+  csl / multi-format / single-format paths.
+- ``PaperDraft.to_docx`` fallback (markdown-on-disk when python-docx
+  is missing) and the explicit-import path.
+- ``PaperDraft.write`` extension dispatch (.md/.tex/.qmd/.docx).
+- ``paper(fmt='bogus')`` validation error.
+- ``paper`` with ``output_path=`` writing to disk.
+- DAG appendix rendering via a duck-typed DAG stub (text and qmd).
+"""
+from __future__ import annotations
+
+import os
+import sys
+import tempfile
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+from statspai.workflow.paper import (
+    PaperDraft,
+    _eda_block,
+    _inline_md_to_tex,
+    _md_to_tex,
+    _notes_block,
+    _record_note,
+    _render_dag_section,
+    _tex_escape,
+    _yaml_str,
+    paper,
+)
+
+
+# --------------------------------------------------------------------- #
+#  Fixtures
+# --------------------------------------------------------------------- #
+
+
+@pytest.fixture(scope="module")
+def rct_data() -> pd.DataFrame:
+    rng = np.random.default_rng(0)
+    n = 400
+    T = rng.binomial(1, 0.5, n)
+    x = rng.normal(size=n)
+    y = 2.0 * T + 0.5 * x + rng.normal(size=n)
+    return pd.DataFrame({
+        "wage": y, "trained": T, "edu": x,
+    })
+
+
+@pytest.fixture(scope="module")
+def draft(rct_data) -> PaperDraft:
+    return paper(
+        rct_data,
+        question="effect of trained on wage",
+        design="rct",
+    )
+
+
+# --------------------------------------------------------------------- #
+#  YAML / TeX helpers
+# --------------------------------------------------------------------- #
+
+
+def test_yaml_str_handles_none_and_quotes():
+    assert _yaml_str(None) == '""'
+    assert _yaml_str("plain") == '"plain"'
+    out = _yaml_str('he said "hi"')
+    assert out.startswith('"') and out.endswith('"')
+    # Embedded quote escaped
+    assert '\\"' in out
+
+
+def test_yaml_str_collapses_newlines():
+    """Newlines are folded to single spaces (Quarto YAML compat)."""
+    assert "\n" not in _yaml_str("line1\nline2")
+
+
+def test_yaml_str_escapes_backslash():
+    out = _yaml_str(r"path\with\backslash")
+    assert r"\\" in out
+
+
+def test_tex_escape_special_chars():
+    out = _tex_escape("a&b%c$d#e_f{g}h~i^j")
+    for tok in (r"\&", r"\%", r"\$", r"\#", r"\_",
+                r"\{", r"\}",
+                r"\textasciitilde{}", r"\textasciicircum{}"):
+        assert tok in out
+
+
+def test_tex_escape_handles_non_string():
[…]

[…]
