### 2026-05-04 5b0b95a
fix(api): un-lazify 14 collision-prone names so sp.proximal etc. stay callable
The lazy-load refactor in 7eeb624 moved 14 names that exist as both a
public function and a parent submodule (proximal, principal_strat,
bartik, bridge, causal_impact, bcf, bunching, deepiv, dose_response,
frontier, interference, msm, multi_treatment, tmle) onto the lazy
``_LAZY_ATTRS`` / ``_LAZY_SUBMODULES`` tables.  This regressed
``sp.proximal(df, ...)`` and 13 other call sites with
``TypeError: 'module' object is not callable`` whenever any prior code
path executed ``from statspai.X import Y`` (very common in th
[…]
