+++ src/statspai/__init__.py
@@ -127,14 +127,23 @@ from .dml import (
     # v1.7 long-panel DML
     dml_panel, DMLPanelResult,
 )
-# (lazy) deepiv: see _LAZY_SUBMODULES / _LAZY_ATTRS
+# Eager: ``deepiv`` is both a function (sp.deepiv(...)) and a subpackage.
+# Lazy-loading collides with the subpackage attachment — see the
+# "PEP 562 collision" note at the bottom of this file.
+from .deepiv import deepiv, DeepIV
 from .panel import (
     panel, panel_compare, balance_panel, PanelResults, PanelCompareResults, PanelRegression,
     Absorber, demean, absorb_ols, hdfe_ols, FEOLSResult,
 )
-# (lazy) causal_impact: see _LAZY_SUBMODULES / _LAZY_ATTRS
+# Eager: ``causal_impact`` collides (function + subpackage of same name).
+from .causal_impact import causal_impact, CausalImpactEstimator, impactplot
 from .mediation import mediate, MediationAnalysis, mediate_sensitivity, mediate_interventional, four_way_decomposition, FourWayResult
-# (lazy) bartik: see _LAZY_SUBMODULES / _LAZY_ATTRS
+# Eager: ``bartik`` collides (function + subpackage of same name).
+from .bartik import (
+    bartik, BartikIV, ssaggregate, shift_share_se,
+    shift_share_political, ShiftSharePoliticalResult,
+    shift_share_political_panel, ShiftSharePoliticalPanelResult,
+)
 from .output.outreg2 import OutReg2, outreg2
 from .output.modelsummary import modelsummary, coefplot
 from .output.sumstats import sumstats, balance_table
@@ -178,9 +187,21 @@ from .inference import (
     multiway_cluster_vcov, cluster_robust_se, cr3_jackknife_vcov,
     g_computation, front_door,
 )
-# (lazy) msm: see _LAZY_SUBMODULES / _LAZY_ATTRS
-# (lazy) proximal: see _LAZY_SUBMODULES / _LAZY_ATTRS
-# (lazy) principal_strat: see _LAZY_SUBMODULES / _LAZY_ATTRS
+# Eager: ``msm`` collides (function + subpackage of same name).
+from .msm import msm, MarginalStructuralModel, stabilized_weights
+# Eager: ``proximal`` collides (function + subpackage of same name).
+from .proximal import (
+    proximal, ProximalCausalInference,
+    negative_control_outcome, negative_control_exposure,
+    double_negative_control, NegativeControlResult,
+    proximal_regression, ProximalRegResult,
+    fortified_pci, bidirectional_pci, pci_mtp,
+    select_pci_proxies, ProxyScoreResult,
+)
+# Eager: ``principal_strat`` collides (function + subpackage of same name).
+from .principal_strat import (
+    principal_strat, PrincipalStratResult, survivor_average_causal_effect,
+)
 # (lazy) spatial: see _LAZY_SUBMODULES / _LAZY_ATTRS
 # NOTE: `from . import iv` would be shadowed by the `iv` function imported
 # on line 31 from `.regression.iv`, so we load the subpackage explicitly.
@@ -217,19 +238,42 @@ from .regression.glm import glm, GLMRegression, GLMEstimator
 from .regression.count import poisson, nbreg, ppmlhdfe
 # neural_causal — lazy-loaded (torch); see _LAZY_ATTRS below.
 from .causal_discovery import notears, NOTEARS, pc_algorithm, PCAlgorithm, lingam, LiNGAMResult, ges, GESResult, fci, FCIResult, icp, nonlinear_icp, ICPResult, pcmci, PCMCIResult, partial_corr_pvalue, lpcmci, LPCMCIResult, dynotears, DYNOTEARSResult
[…]
