### 2026-05-04 4fbfb43
feat(api): generated type stub so IDE / mypy see lazy-loaded names
After Codex's lazy-load refactor (7eeb624), 200+ public symbols (sp.bayes_did,
sp.target_trial_protocol, sp.tarnet, sp.gformula_ice_fn, …) are routed through
__getattr__ at runtime. Static analyzers cannot follow that path, so PyCharm /
Pyright / mypy showed red squigglies and lost autocomplete + jump-to-definition
for half the package surface.

Add src/statspai/__init__.pyi (auto-generated) which re-exports each public
name from its real source module via the standard `from .X import name as name`
re-export pattern. Static ana
[…]
