### 2026-05-04 1ffe123
feat(tests): runtime consistency for FunctionSpec.limitations + fix etwfe gap
Add tests/test_limitations_consistency.py to guard the parity-grade
"limitations" advertisements introduced in f1dae20 against drift.
Three layers of check:

1. Vocabulary — every `limitations` string must use vetted phrasing
   (`not yet implemented`, `currently requires`, `is an MVP`,
   `silently`, …); a sloppy or vague entry fails CI immediately.
2. Classification — every entry must be either (a) runtime-testable
   via a curated `(function_name, substring) -> (call, expected_exc)`
   map or (b) declared "descrip
[…]
