+++ tests/test_limitations_consistency.py
@@ -1,0 +1,197 @@
+"""Runtime consistency tests for ``FunctionSpec.limitations``.
+
+The ``limitations`` field on every :class:`statspai.registry.FunctionSpec`
+declares parameter values / variant gaps that are documented as
+not-yet-implemented inside an otherwise stable function — see
+``docs/guides/stability.md``. Without this test those advertisements
+silently rot the moment the underlying code learns a new variant or the
+function is renamed: the registry would still claim the gap exists but
+the runtime no longer enforces it (or vice versa).
+"""
+from __future__ import annotations
+
+from typing import Any, Callable, Dict, List, Tuple
+
+import numpy as np
+import pandas as pd
+import pytest
+
+
+ALLOWED_PHRASES: Tuple[str, ...] = (
+    "not yet implemented",
+    "not implemented",
+    "not yet supported",
+    "is not supported",
+    "not currently supported",
+    "raises notimplementederror",
+    "raises importerror",
+    "currently requires",
+    "currently falls back",
+    "is an mvp",
+    "mvp",
+    "only",
+    "fallback",
+    "silently",
+    "deterministic fallback",
+)
+
+
+LIMITATIONS_DESCRIPTIVE_ONLY: Dict[str, List[str]] = {
+    "did_multiplegt_dyn": [
+        "switch-on only",
+        "SE is cluster bootstrap",
+        "heteroskedastic-weights variant",
+    ],
+    "continuous_did": [
+        "method='cgs' is an MVP",
+    ],
+    "network_exposure": [
+        "design='complete' currently falls back",
+    ],
+    "text_treatment_effect": [
+        "embedder='sbert' requires the optional sentence-transformers",
+        "Veitch et al. (2020) full BERT/topic-model recipe",
+    ],
+}
+
+
+def _runtime_map() -> Dict[Tuple[str, str], Tuple[Callable[[], Any], type | tuple[type, ...]]]:
+    """Build the (function_name, limitation_substring) -> (call, exc) map."""
+    import statspai as sp
+
+    rng = np.random.default_rng(0)
+    n = 200
+
+    df_panel = pd.DataFrame({
+        "y": rng.normal(size=n),
+        "i": np.repeat(np.arange(n // 4), 4),
+        "t": np.tile(np.arange(4), n // 4),
+        "g": np.repeat(rng.choice([0, 2, 3], size=n // 4), 4),
+        "d": rng.binomial(1, 0.4, size=n),
+        "dose": rng.uniform(0, 1, size=n),
+    })
+
+    df_cs = pd.DataFrame({
+        "y": rng.normal(size=n),
+        "d": rng.binomial(1, 0.5, size=n).astype(float),
+        "s": rng.binomial(1, 0.5, size=n).astype(float),
+        "x": rng.normal(size=n),
+        "x1": rng.normal(size=n),
+        "x2": rng.normal(size=n),
+        "z": rng.binomial(1, 0.5, size=n).astype(float),
+        "w": rng.normal(size=n),
+        "score": rng.normal(size=n),
+    })
+
+    return {
+        ("principal_strat", "instrument="): (
+            lambda: sp.principal_strat(
+                df_cs, y="y", treat="d", strata="s",
+                method="monotonicity", instrument="z",
+                n_boot=10, seed=0,
+            ),
+            NotImplementedError,
+        ),
+        ("hal_tmle", "variant='projection'"): (
+            lambda: sp.hal_tmle(
+                df_cs, y="y", treat="d", covariates=["x1", "x2"],
+                variant="projection",
+            ),
+            NotImplementedError,
+        ),
+        ("callaway_santanna", "panel=False"): (
+            lambda: sp.callaway_santanna(
+                df_panel, y="y", g="g", t="t", i="i",
+                panel=False, estimator="dr",
+            ),
[…]
