### 2026-05-04 33c7592
test(api): pin late-bind + post-import-shadow contracts
Tests/test_late_bind_contracts.py adds 21 regression contracts for the
two binding hazards in src/statspai/__init__.py that the recent lazy-load
work surfaced:

1. Late-bind shadow (5 entries + sp.iv dispatcher): ``mediation``,
   ``policy_tree``, ``dml``, ``matrix_completion``, ``causal_discovery``
   are all *re-bound* by the ``_article_aliases`` block at the bottom of
   __init__.py to article-facing function wrappers (which override
   earlier eager submodule bindings).  The tests fail loudly if the
   re-bind block is removed or move
[…]
