### 2026-05-04 abc519e
feat(scripts): stability reverse-audit + tag _auto on auto-registered specs
scripts/stability_audit.py walks the registry and cross-references
every ``stability='stable'`` claim against parity-test coverage in
``tests/reference_parity/`` + ``tests/external_parity/``.  Splits the
catalogue into two populations that v1.13's "default stability"
silently merged:

* hand-written specs (203 today) — explicit ``register(FunctionSpec)``
  calls in ``_build_registry``; the maintainer wrote the spec by
  hand, so a missing parity test is a real "claim without backing".
* auto-registered specs (768 today
[…]
