+++ scripts/stability_audit.py
@@ -1,0 +1,301 @@
+"""Cross-check ``FunctionSpec.stability == 'stable'`` claims against
+parity-test coverage in ``tests/reference_parity/`` + ``tests/external_parity/``.
+
+A function classified as ``stability='stable'`` claims to be
+*parity-grade* — numerically aligned with R / Stata / a published
+reference, with a locked signature. By design that is the SLA an agent
+relies on when filtering ``sp.list_functions(stability='stable')`` for
+production use.
+
+The catch: until v1.13 every newly-registered function was *implicitly*
+``stable`` (the field's default), so the catalogue's ~970 stable
+entries currently mix two populations:
+
+* **Truly parity-backed** — at least one test in
+  ``tests/reference_parity/`` or ``tests/external_parity/`` exercises
+  the function with R / Stata / paper-replication numbers.
+* **Stability-by-default** — never explicitly classified, never
+  parity-tested; the ``stable`` label is a placeholder waiting for
+  someone to decide.
+
+This script makes the split visible so a maintainer can either (a) add
+a parity test to back the claim or (b) flip the entry to
+``stability='experimental'`` until one exists.
+
+It does **not** auto-downgrade. The decision belongs to a human:
+something can be analytically correct without a published reference,
+and we don't want a one-shot CI run to demote 700 functions overnight.
+
+Usage
+-----
+::
+
+    python scripts/stability_audit.py                  # human-readable report
+    python scripts/stability_audit.py --json           # machine-readable
+    python scripts/stability_audit.py --unbacked       # list unbacked names only
+    python scripts/stability_audit.py --hand-written   # restrict to hand-written specs
+    python scripts/stability_audit.py --check          # exit 1 if regression vs floor
+
+The ``--check`` mode is meant for CI: it succeeds as long as the count
+of *unbacked, hand-written* stable claims has not increased beyond a
+loose floor.  Auto-registered specs are excluded from the floor because
+classifying ~700 of them is a separate (large) project — see ``Step H``
+in the v1.13 stability roadmap.
+"""
+from __future__ import annotations
+
+import argparse
+import json
+import re
+import sys
+from pathlib import Path
+from typing import Dict, List, Set, Tuple
+
+REPO_ROOT = Path(__file__).resolve().parent.parent
+PARITY_DIRS: Tuple[Path, ...] = (
+    REPO_ROOT / "tests" / "reference_parity",
+    REPO_ROOT / "tests" / "external_parity",
+)
+
+#: Loose ceiling on how many *hand-written* stable specs may go without
+#: a parity test before --check fails.  Bumped when we deliberately add
+#: hand-written entries faster than parity tests.  Decrease over time as
+#: the audit gets cleaned up.
+UNBACKED_HANDWRITTEN_FLOOR = 220
+
+#: Regex matching ``sp.<name>(`` references in test source.  Used to
+#: attribute parity coverage to public ``sp.*`` symbols.
+SP_CALL_RE = re.compile(r"\bsp\.([a-zA-Z_][a-zA-Z0-9_]*)\s*\(")
+
+#: Some test files exercise multiple estimators (cross_estimator_parity,
+#: published_replications, …) — we credit every ``sp.X(`` reference
+#: found in such files even if the file's name doesn't tag a single
+#: estimator family.
+
+
+def _scan_parity_file(path: Path) -> Set[str]:
+    """Return every ``sp.<name>`` symbol referenced in a parity test file."""
+    try:
+        text = path.read_text(encoding="utf-8")
+    except OSError:
+        return set()
+    return set(SP_CALL_RE.findall(text))
+
+
+def _backed_functions() -> Tuple[Set[str], Dict[str, List[str]]]:
+    """Walk every parity test file once.
+
+    Returns
[…]
