### 2026-05-04 199edcb
fix(smart): align _audit_linear backing tests to (data, y, x) signatures
Phase A's WorkflowDegradedWarning surfacing exposed that _audit_linear
was calling sp.reset_test(result), sp.het_test(result), sp.vif(result),
sp.oster_bounds(result) — but those four take (data, y, x_names) /
(data, x_names) / (data, y, treat, controls), not a fitted result.
Every real OLS audit was silently losing all four checks (TypeError +
ValueError under except Exception: pass).

Extracts dependent_var + non-intercept regressors from result.params
and result.data_info, then calls each backing test with raw column
r
[…]
