### 2026-05-04 bc99bec
refactor(api): drop dead-code lazy entries for the 14 conflict-prone names
After f9ec214 restored eager `from .X import X, ...` imports for the 14
function names that share their submodule's name (proximal /
principal_strat / bridge / bcf / bunching / dose_response /
multi_treatment / causal_impact / frontier / interference / tmle / msm /
deepiv / bartik), every leaf those modules export is also bound eagerly.
The corresponding _LAZY_SUBMODULES entries and _register_lazy(...) blocks
were therefore unreachable — the eager bindings shadowed them on every
sp.X access — and just added noise.

This
[…]
