+++ src/statspai/__init__.py
@@ -1417,18 +1417,8 @@ _LAZY_SUBMODULES: dict = {
     "fairness":          "fairness",
     "assimilation":      "assimilation",
     "surrogate":         "surrogate",
-    "bcf":               "bcf",
-    "bunching":          "bunching",
-    "dose_response":     "dose_response",
     "bounds":            "bounds",
-    "multi_treatment":   "multi_treatment",
-    "bridge":            "bridge",
-    "causal_impact":     "causal_impact",
     "dtr":               "dtr",
-    "interference":      "interference",
-    "principal_strat":   "principal_strat",
-    "proximal":          "proximal",
-    "bartik":            "bartik",
     "spatial":           "spatial",
     "conformal_causal":  "conformal_causal",
     "ope":               "ope",
@@ -1444,13 +1434,9 @@ _LAZY_SUBMODULES: dict = {
     "imputation":        "imputation",
     "survey":            "survey",
     "survival":          "survival",
-    "tmle":              "tmle",
     "nonparametric":     "nonparametric",
-    "frontier":          "frontier",
     "multilevel":        "multilevel",
     "structural":        "structural",
-    "msm":               "msm",
-    "deepiv":            "deepiv",
     "causal_llm":        "causal_llm",
     "mht":               "mht",
     "robustness":        "robustness",
@@ -1515,65 +1501,14 @@ _register_lazy("surrogate",
     "surrogate_index", "long_term_from_short", "proximal_surrogate_index",
     "SurrogateResult",
 )
-_register_lazy("bcf",
-    "bcf", "BayesianCausalForest", "bcf_longitudinal", "BCFLongResult",
-    "bcf_ordinal", "BCFOrdinalResult",
-    "bcf_factor_exposure", "BCFFactorExposureResult",
-)
-_register_lazy("bunching",
-    "bunching", "BunchingEstimator", "notch", "NotchResult",
-    "general_bunching", "GeneralBunchingResult",
-    "kink_unified", "KinkUnifiedResult",
-)
-_register_lazy("dose_response",
-    "dose_response", "DoseResponse", "vcnet", "scigan", "VCNetResult",
-)
 _register_lazy("bounds",
     "lee_bounds", "manski_bounds", "BoundsResult", "horowitz_manski",
     "iv_bounds", "oster_delta", "selection_bounds", "breakdown_frontier",
     "balke_pearl", "BalkePearlResult", "ml_bounds", "MLBoundsResult",
 )
-_register_lazy("multi_treatment",
-    "multi_treatment", "MultiTreatment",
-)
-_register_lazy("bridge",
-    "bridge", "BridgeResult",
-)
-_register_lazy("causal_impact",
-    "causal_impact", "CausalImpactEstimator", "impactplot",
-)
 _register_lazy("dtr",
     "g_estimation", "GEstimation", "q_learning", "QLearningResult",
     "a_learning", "ALearningResult", "snmm", "SNMMResult",
-)
-_register_lazy("interference",
-    "spillover", "SpilloverEstimator",
-    "network_exposure", "NetworkExposureResult",
-    "peer_effects", "PeerEffectsResult",
-    "network_hte", "inward_outward_spillover",
-    "NetworkHTEResult", "InwardOutwardResult",
-    "cluster_matched_pair", "MatchedPairResult",
-    "cluster_cross_interference", "CrossClusterRCTResult",
-    "cluster_staggered_rollout", "StaggeredClusterRCTResult",
-    "dnc_gnn_did", "DNCGNNDiDResult",
-    "interference", "interference_available_designs",
-)
[…]
