### 2026-05-04 59fdc19
feat(smart): sp.recommend / sp.causal / sp.paper respect FunctionSpec.stability
Add ``allow_experimental: bool = False`` to the three workflow entry
points so an LLM agent or pipeline that asks for a ranked
recommendation never silently lands on a frontier MVP estimator.

* ``sp.recommend(...)`` — applies a stability filter just before
  returning, dropping any recommendation whose ``function`` is
  registered as ``stability='experimental'`` or ``'deprecated'``.
  Drops are recorded in ``RecommendationResult.warnings`` with the
  ``sp.X`` names so the user sees exactly what was excluded.  Help
[…]
