+++ tests/test_smart_stability_gating.py
@@ -1,0 +1,161 @@
+"""Tests for the v1.13 stability gate on the smart layer.
+
+``sp.recommend`` (and transitively ``sp.causal`` / ``sp.paper``) must
+default to dropping experimental/deprecated estimators from the
+ranked output so an LLM agent or pipeline never silently lands on a
+frontier MVP. ``allow_experimental=True`` is the explicit opt-in.
+
+Two layers of test:
+
+* unit: the helper ``_filter_unstable_recommendations`` drops the
+  right entries.
+* integration: ``sp.recommend(...)`` with a hand-crafted recommendation
+  list confirms the warning trail and the kept set survive the round trip.
+"""
+from __future__ import annotations
+
+import pandas as pd
+import numpy as np
+import pytest
+
+
+def _toy_panel(n_units: int = 60, n_t: int = 4) -> pd.DataFrame:
+    rng = np.random.default_rng(0)
+    n = n_units * n_t
+    return pd.DataFrame({
+        "y":          rng.normal(size=n),
+        "unit":       np.repeat(np.arange(n_units), n_t),
+        "t":          np.tile(np.arange(n_t), n_units),
+        "treatment":  rng.binomial(1, 0.5, size=n),
+        "x1":         rng.normal(size=n),
+    })
+
+
+# --------------------------------------------------------------------------- #
+#  Unit: the in-place filter
+# --------------------------------------------------------------------------- #
+
+def _import_recommend_module():
+    """Import the ``statspai.smart.recommend`` *module* (not the
+    function of the same name re-exported from ``statspai.smart``).
+
+    Same PEP 562 collision pattern documented in the root
+    ``statspai/__init__.py``: ``from .recommend import recommend``
+    rebinds ``statspai.smart.recommend`` to the function, shadowing
+    the submodule.  We use ``importlib`` to bypass that and reach
+    the actual module so we can poke private helpers.
+    """
+    import importlib
+    return importlib.import_module("statspai.smart.recommend")
+
+
+class TestUnstableFilter:
+    def test_filter_drops_experimental(self) -> None:
+        mod = _import_recommend_module()
+        recs = [
+            {"function": "regress", "method": "stable"},
+            {"function": "did_multiplegt_dyn", "method": "exp"},
+            {"function": "did", "method": "stable"},
+        ]
+        kept, dropped = mod._filter_unstable_recommendations(recs)
+        assert [r["function"] for r in kept] == ["regress", "did"]
+        assert dropped == ["did_multiplegt_dyn"]
+
+    def test_filter_keeps_unknown_function(self) -> None:
+        """Custom recs that don't appear in the registry must survive
+        the filter — backwards compatibility for downstream callers
+        that append their own entries."""
+        mod = _import_recommend_module()
+        recs = [
+            {"function": "completely_made_up_estimator"},
+            {"function": "did_multiplegt_dyn"},  # experimental
+        ]
+        kept, dropped = mod._filter_unstable_recommendations(recs)
+        assert kept == [{"function": "completely_made_up_estimator"}]
+        assert dropped == ["did_multiplegt_dyn"]
+
+    def test_filter_handles_missing_function_key(self) -> None:
+        mod = _import_recommend_module()
+        recs = [
+            {"method": "no function key"},
+            {"function": None},
+        ]
+        kept, dropped = mod._filter_unstable_recommendations(recs)
+        assert kept == recs
+        assert dropped == []
+
+
+# --------------------------------------------------------------------------- #
+#  Integration: sp.recommend default vs opt-in
+# --------------------------------------------------------------------------- #
+
+class TestRecommendStabilityDefault:
+    """Synthesize a recommendation list and round-trip via the real entry point."""
+
+    def test_recommend_default_drops_experimental_and_warns(self) -> None:
+        """When a recommendation list includes an experimental entry,
[…]
