### 2026-05-04 b7fdee0
fix(api): harden recommendation gating and iv exports
