### 2026-05-04 af65a2d
feat(estimators): implement principal_strat(instrument=...) + open hal_tmle projection RFC
Two halves of Step G in the v1.13 stability roadmap:

PRINCIPAL_STRAT (full implementation)
-------------------------------------
The previously-stubbed ``instrument=`` parameter now routes to a
proper Angrist-Imbens-Rubin (1996) §4 / Wald LATE estimator for the
encouragement-design two-layer setup ``(Z, D, S, Y)``:

* Identifies the complier share w.r.t. Z under monotonicity
  D(1)>=D(0): π_C(Z) = P(D=1|Z=1) - P(D=1|Z=0).
* Reports two Wald LATEs among Z-compliers: τ_Y for the effect of D
  on the outco
[…]
