### 2026-05-04 6c65685
feat(registry): upgrade 8 high-impact estimators from auto- to hand-written specs
Step H of the v1.13 stability roadmap.  Eight workhorse estimators
that were previously auto-registered (only ``inspect.signature`` +
the first docstring line — no agent-native metadata) now ship full
hand-written FunctionSpecs so an LLM agent calling
``sp.describe_function(name)`` sees the identification story inline,
not only via the docstring.

Upgraded:

* ``aipw``     — augmented IPW / doubly-robust ATE
                 (Robins-Rotnitzky-Zhao 1994).
* ``aggte``    — Callaway-Sant'Anna group-time ATT aggregat
[…]
