+++ src/statspai/registry.py
@@ -7437,6 +7437,452 @@ def _build_registry():
         typical_n_min=200,
     ))
 
+    # ================================================================= #
+    # v1.13 Step H: agent-native upgrades for high-impact estimators
+    # that previously shipped only as auto-registered specs (no
+    # ``assumptions`` / ``failure_modes`` / ``alternatives`` /
+    # ``typical_n_min``).  Each entry below replaces the auto-registered
+    # default with a hand-written spec carrying the canonical
+    # identification story so an agent reading
+    # ``sp.describe_function(name)`` sees the assumptions and recovery
+    # paths inline, not only via the docstring.
+    # ================================================================= #
+
+    register(FunctionSpec(
+        name="aipw",
+        category="causal",
+        description=(
+            "Augmented inverse-probability weighting (AIPW) — the "
+            "canonical doubly-robust ATE estimator.  Cross-fits an "
+            "outcome regression and a propensity model and combines "
+            "them via the efficient-influence-function formula, so the "
+            "estimate is consistent if either nuisance is correctly "
+            "specified (Robins, Rotnitzky & Zhao 1994)."
+        ),
+        params=[
+            ParamSpec("data", "DataFrame", True),
+            ParamSpec("y", "str", True, description="Outcome variable"),
+            ParamSpec("treat", "str", True, description="Binary treatment (0/1)"),
+            ParamSpec("covariates", "list", True, description="Confounders to adjust for"),
+            ParamSpec("estimand", "str", False, "ATE", "Target estimand", ["ATE", "ATT", "ATC"]),
+            ParamSpec("n_folds", "int", False, 5, "Cross-fitting folds (>= 2)"),
+            ParamSpec("alpha", "float", False, 0.05),
+            ParamSpec("seed", "int", False, None),
+        ],
+        returns="CausalResult",
+        example='sp.aipw(df, y="wage", treat="trained", covariates=["age", "edu"])',
+        tags=["aipw", "doubly-robust", "ipw", "causal", "ate", "att"],
+        reference=(
+            "Robins, Rotnitzky & Zhao (1994) JASA "
+            "[@robins1994estimation]; Glynn & Quinn (2010) [@glynn2010introduction]"
+        ),
+        pre_conditions=[
+            "binary treatment column with both arms present",
+            "covariates must contain all confounders for unconfoundedness",
+            "no perfect overlap violations (0 < propensity < 1 in support)",
+        ],
+        assumptions=[
+            "Unconfoundedness conditional on covariates (Y(0), Y(1) ⊥ D | X)",
+            "Overlap / common support: 0 < e(X) < 1 for all X with positive density",
+            "SUTVA",
+            "At least one of (outcome model, propensity model) correctly specified",
+        ],
+        failure_modes=[
+            FailureMode(
+                symptom="Propensity scores cluster near 0 or 1",
+                exception="statspai.AssumptionViolation",
+                remedy="Trim to overlap region with sp.trimming() or switch to overlap-weighted ATE.",
+                alternative="overlap_weights",
+            ),
+            FailureMode(
+                symptom="Cross-fit estimate has very wide CI",
+                exception="statspai.NumericalInstability",
+                remedy="Increase n_folds or reduce covariate dimension; check for near-empty propensity strata.",
+                alternative="dml",
+            ),
+        ],
+        alternatives=["ipw", "dml", "tmle", "matching"],
+        typical_n_min=200,
+    ))
+
+    register(FunctionSpec(
+        name="aggte",
+        category="causal",
+        description=(
+            "Aggregate Callaway-Sant'Anna group-time ATTs into "
+            "interpretable summaries — overall ATT, event-study by "
+            "relative time, group-specific ATT(g), or calendar-time "
+            "ATT(t).  Inference uses the multiplier bootstrap on the "
[…]
