### 2026-05-04 b7485e7
fix(smart): surface placebo all-reps-crashed as a single degradation warning
verify_recommendation()'s placebo battery deliberately tolerates per-rep
fit failures so one bad permutation doesn't kill the whole check.  But
when ALL n_reps crash (e.g. shared signature bug or input-shape bug
across every method call), the existing path returns score=0 — exactly
the same outcome as a method that genuinely failed placebo.  The user
can't tell the difference between "method placebo'd at zero" and
"every fit raised identically".

Capture the last exception inside the loop and, when n_crashes ==
n_reps
[…]
