### 2026-05-04 559738b
fix(core): harden principal strat instrument diagnostics
