### 2026-05-04 793c40d
chore(release): cut v1.13.0
Bump version to 1.13.0, drop the eager `OPEResult` re-export so
`sp.OPEResult` resolves to the canonical lazy class (restoring
v1.12.2 isinstance semantics for `sp.ope.*` results), refresh
paper.md to the 1.10–1.13 surface, and finish out the
`principal_strat` registry spec doc strings.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
