### 2026-05-04 b655ba1
perf(api): lazy-load statspai.forest (Step 1B cold-start budget)
Previously `import statspai` chained four eager
`from .forest.X import Y` lines at module load, transitively pulling
~245 sklearn submodules into sys.modules (~270ms cumulative on cold
cache) for every session — even ones that never touched
heterogeneous-effect forests. The eager block is removed; the 10
public leaves (CausalForest, causal_forest, calibration_test,
test_calibration, rate, honest_variance, multi_arm_forest,
MultiArmForestResult, iv_forest, IVForestResult) now resolve via
`_LAZY_ATTRS` keyed to dotted submodule pat
[…]
