+++ tests/test_late_bind_contracts.py
@@ -159,3 +159,130 @@ def test_no_conflict_function_morphs_to_module_after_lazy_traffic():
             f"Either restore the eager `from .{name} import {name}` import "
             f"or extend the __getattr__ defensive re-pin."
         )
+
+
+# ---------------------------------------------------------------------------
+# Forest cold-start budget (Step 1B).
+#
+# Previously, ``__init__.py`` eagerly imported ``forest.causal_forest``
+# / ``forest.iv_forest`` / ``forest.multi_arm_forest`` /
+# ``forest.forest_inference`` at module load, which transitively pulled
+# ~245 ``sklearn.*`` submodules into ``sys.modules`` for every
+# ``import statspai`` — even sessions that never touch heterogeneous-effect
+# forests.  ``forest`` does *not* collide with a top-level function (no
+# ``sp.forest`` callable export), so the 8 leaf names (CausalForest /
+# causal_forest / calibration_test / test_calibration / rate /
+# honest_variance / multi_arm_forest / MultiArmForestResult / iv_forest /
+# IVForestResult) live in ``_LAZY_ATTRS`` keyed to dotted submodule paths
+# (e.g. ``forest.causal_forest``) and resolve via ``__getattr__`` on first
+# touch.
+#
+# These contracts pin three things:
+#
+# 1. ``import statspai`` must NOT pre-load any ``statspai.forest.*``
+#    submodule — that's the cold-start regression contract.
+# 2. After touching a forest leaf, the public name must resolve to the
+#    callable (function/class), not the auto-bound submodule.
+# 3. A downstream ``from statspai.forest.causal_forest import CausalForest``
+#    must NOT silently re-shadow ``sp.causal_forest`` to the leaf module
+#    via Python's post-import attribute binding.
+# ---------------------------------------------------------------------------
+
+FOREST_LAZY_LEAVES = [
+    ("CausalForest",          "forest.causal_forest"),
+    ("causal_forest",         "forest.causal_forest"),
+    ("calibration_test",      "forest.forest_inference"),
+    ("test_calibration",      "forest.forest_inference"),
+    ("rate",                  "forest.forest_inference"),
+    ("honest_variance",       "forest.forest_inference"),
+    ("multi_arm_forest",      "forest.multi_arm_forest"),
+    ("MultiArmForestResult",  "forest.multi_arm_forest"),
+    ("iv_forest",             "forest.iv_forest"),
+    ("IVForestResult",        "forest.iv_forest"),
+]
+
+
+def test_forest_not_loaded_on_bare_import_statspai():
+    """``import statspai`` must not transitively import any forest leaf
+    submodule.  Regression: any future eager ``from .forest... import ...``
+    in ``__init__.py`` would re-inflate ``sys.modules`` by ~245
+    ``sklearn.*`` entries on every session.
+
+    Runs in a subprocess so the cold-state check doesn't perturb other
+    tests' ``sys.modules`` (which would change ``CausalResult`` class
+    identity and break ``isinstance`` assertions elsewhere)."""
+    import subprocess
+    import sys
+
+    code = (
+        "import sys\n"
+        "import statspai\n"
+        "leaked = sorted(m for m in sys.modules if m.startswith('statspai.forest'))\n"
+        "print('LEAKED=' + ','.join(leaked))\n"
+    )
+    proc = subprocess.run(
+        [sys.executable, "-c", code],
+        capture_output=True,
+        text=True,
+        check=True,
+    )
+    last_line = proc.stdout.strip().splitlines()[-1]
+    assert last_line.startswith("LEAKED="), proc.stdout
+    leaked_csv = last_line[len("LEAKED="):]
+    leaked = [m for m in leaked_csv.split(",") if m]
+    assert not leaked, (
[…]
