### 2026-05-04 ef410c6
perf(api): lazy-import sklearn across 18 estimator files (Step 1C)
Building on Step 1B's lazy-load of statspai.forest, every remaining
top-level `from sklearn.X import Y` in 18 estimator files
(did/overlap_did, metalearners/{auto_cate,metalearners,auto_cate_tuned},
policy_learning/{policy_tree,ope}, synth/cluster,
proximal/pci_regression, bcf/{bcf,longitudinal},
tmle/{tmle,super_learner,ltmle,ltmle_survival}, dose_response/gps,
multi_treatment/multi_ipw, mediation/four_way, interference/orthogonal)
is moved inside the function bodies that actually use it. Type
annotations referencing `BaseEsti
[…]
