### 2026-05-04 14ab633
perf(api): drop sklearn class inheritance from HAL estimators (Step 1D)
HALRegressor / HALClassifier in tmle/hal_tmle.py previously
subclassed sklearn.base.BaseEstimator + a Mixin, which pulled ~39
sklearn submodules into sys.modules at module-load time — the only
remaining sklearn footprint after Steps 1B/1C lazy-loaded forest and
the 18 estimator files. The inheritance is gratuitous: super_learner
only needs sklearn.base.clone(learner) (duck-typed —
get_params(deep=False) + cls(**params) reconstruction) plus
.fit / .predict / .predict_proba; no code path calls .score(...),
is_classifier(...)
[…]
