### 2026-05-05 f3124d1
test(stata_parity): add 21-module StatsPAI<->Stata parity harness + 3-way compare
Adds tests/stata_parity/ as the Stata sibling of the existing
R-parity dir, and extends tests/r_parity/compare.py to emit a
3-way (Python/R/Stata) parity_table_3way.{md,tex} consumed by the
JSS appendix.

Coverage: 21 of the 36 R-parity modules now have a Stata cross-check
(01_ols, 02_iv, 03_hdfe, 04_csdid, 05_sunab, 06_rd, 07_scm,
09_rddensity, 10_honest_did, 11_psm, 12_sdid, 14_ols_cluster,
15_hdfe_cluster, 16_bjs, 17_etwfe, 20_bacon, 21_honest_relmags,
22_sensemakr, 25_lmm, 28_frontier, 30_oaxaca). Each .do fi
[…]
