+++ tests/r_parity/compare.py
@@ -1,0 +1,719 @@
+"""Parity comparator: read all results/<module>_{py,R}.json pairs and
+(optionally) the sister Stata-side JSONs in
+``tests/stata_parity/results/<module>_Stata.json``, then emit:
+
+  * parity_table.md       -- human-readable Markdown (3-way when Stata available)
+  * parity_table.tex      -- LaTeX longtable, 4-col R-only baseline (legacy)
+  * parity_table_3way.tex -- LaTeX longtable, 5-col with Stata column;
+                             this is the version the JSS appendix \\input{}s.
+  * parity_summary.md     -- one-line-per-module headline + verdict
+
+Tolerance budget (pre-registered, NEXT-STEPS / JSS plan §5.2):
+
+  * closed-form estimators (OLS, 2SLS, HDFE):  rel_diff < 1e-6
+  * iterative / cross-fit (DiD, RD, SCM, DML): rel_diff < 1e-3
+  * bootstrap / placebo CI half-widths:        abs_diff < 0.05 * SE
+  * Honest-DiD CI bounds:                      abs_diff < 0.05
+
+The same tolerance applies to the StatsPAI <-> Stata comparison: we
+do not register a separate budget for the Stata side; one budget per
+module is the single source of truth.
+
+Verdict assignment is per-module, not per-row, because some rows
+(e.g. SE-with-documented-convention-gap, default-h selector) are
+expected NOT to pass at the strict tolerance and are recorded with
+an explicit rationale in the module's `extra` block.
+"""
+from __future__ import annotations
+
+import json
+from dataclasses import dataclass
+from pathlib import Path
+from typing import Any
+
+
+HERE = Path(__file__).resolve().parent
+RESULTS_DIR = HERE / "results"
+# The sister Stata-side harness lives at tests/stata_parity/. Stata
+# results are only emitted for the 21 modules with a canonical Stata
+# reference (see tests/stata_parity/README.md); the modules without
+# a canonical Stata implementation (DML, causal forest, augsynth,
+# gsynth) are flagged with explicit reasons in the 3-way table.
+STATA_RESULTS_DIR = HERE.parent / "stata_parity" / "results"
+STATA_SKIP_REASON: dict[str, str] = {
+    "08_dml":           "no canonical ref",
+    "13_causal_forest": "no canonical ref",
+    "18_augsynth":      "no canonical ref",
+    "19_gsynth":        "no canonical ref",
+    "23_evalue":        "not built",
+    "24_coxph":         "not built",
+    "26_glmm_logit":    "not built",
+    "27_glmm_aghq":     "not built",
+    "29_panel_sfa":     "not built",
+    "31_dfl":           "not built",
+    "32_rif":           "not built",
+    "33_var":           "not built",
+    "34_lp":            "not built",
+    "35_panel":         "not built",
+    "36_mediation":     "not built",
+}
+
+
+# Pre-registered tolerance per module.
+TOLERANCES: dict[str, dict[str, float]] = {
+    "01_ols":       {"rel_est": 1e-6, "rel_se": 1e-6},
+    "02_iv":        {"rel_est": 1e-6, "rel_se": 1e-6},
+    "03_hdfe":      {"rel_est": 1e-6, "rel_se": 1e-2},  # 1-df conv. gap
+    "04_csdid":     {"rel_est": 1e-3, "rel_se": 5.0},   # SE under-covers
+    "05_sunab":     {"rel_est": 1e-3, "rel_se": 0.25},
[…]
