+++ tests/stata_parity/README.md
@@ -1,0 +1,133 @@
+# `tests/stata_parity/` — cross-language parity harness against Stata
+
+This directory is the **StatsPAI ↔ Stata** sibling of
+[`tests/r_parity/`](../r_parity/): each module pair runs the same
+calibrated replica on both sides, dumps a full-precision JSON
+result, and `tests/r_parity/compare.py` joins the three sides
+(StatsPAI, R, Stata) into a single 3-way Track A parity table for
+the JSS Appendix B.
+
+The harness is read by the same `compare.py` that drives the R
+side — there is **one** comparator and **one** tolerance budget
+(`compare.py::TOLERANCES`). Parity is a property of the estimator,
+not of the reference language, so we deliberately do not register
+a separate budget for the Stata comparison.
+
+## What's here
+
+```
+tests/stata_parity/
+├── README.md
+├── _common.do            # shared scaffolding: JSON writer (file-based, survives `mata clear`)
+├── _quick_compare.py     # ad-hoc 3-way comparator while developing modules
+├── NN_<method>.do        # one .do per module
+├── logs/                 # Stata's per-run .smcl/.log + the JSON-row tmp files
+└── results/
+    └── NN_<method>_Stata.json   # full-precision results, joined by compare.py
+```
+
+Each `.do` file imports `../r_parity/data/NN_<name>.csv` (the same
+bytes the R side reads), runs the canonical Stata reference, and
+writes one row per parity statistic to
+`results/NN_<name>_Stata.json` via the helpers in `_common.do`.
+
+## Modules covered (21 of 36)
+
+| # | Method                       | StatsPAI                       | Stata reference                                              |
+| --- | --- | --- | --- |
+| 01 | OLS + HC1 SE                  | `sp.regress`                   | `regress, vce(robust)`                                       |
+| 02 | 2SLS + HC1 SE                 | `sp.iv`                        | `ivregress 2sls, vce(robust) small`                          |
+| 03 | HDFE 2-way FE                 | `sp.fast.feols`                | `reghdfe, absorb(...) vce(unadjusted)`                       |
+| 04 | CS-DiD simple ATT             | `sp.callaway_santanna`         | `csdid + estat simple, method(reg)`                          |
+| 05 | Sun-Abraham event study       | `sp.sun_abraham`               | `eventstudyinteract`                                         |
+| 06 | RD CCT bias-corrected         | `sp.rdrobust`                  | `rdrobust`                                                   |
+| 07 | Classical SCM                 | `sp.synth(method="classic")`   | `synth ..., trunit(...) trperiod(...) nested`                |
+| 09 | RD density (CJM)              | `sp.rddensity`                 | `rddensity`                                                  |
+| 10 | Honest DiD bounds (FLCI)      | `sp.honest_did`                | `honestdid, b(...) vcov(...) numpre(...) mvec(...) delta(sd)`|
+| 11 | PSM 1:1 NN                    | `sp.psm`                       | `teffects psmatch, atet nneighbor(1)`                        |
+| 12 | Synthetic DiD                 | `sp.synth(method="sdid")`      | `sdid ..., vce(placebo)`                                     |
+| 14 | OLS + cluster (CR1)           | `sp.regress(robust="cluster")` | `regress, vce(cluster ...)`                                  |
[…]
