### 2026-05-05 2e31628
test(r_parity): add 36-module R parity harness + orig/perf tracks
- tests/r_parity/: 36 paired R/Python scripts + data + JSON results +
  parity tables (md + tex); compare.py 3-way table tightened (ID column
  + small/longtable layout for paper inclusion).
- tests/orig_parity/: 4 canonical-dataset replays (card, mpdta, basque,
  lalonde) with R/Py drivers, results, and parity_table_orig.md.
- tests/perf/: Track C performance harness (HDFE, CS-DID, SCM, DML)
  with R/Py runners, results, log-log figure, and perf_table.{md,tex}.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.co
[…]
