### 2026-05-05 d34dbb1
Fix CS-DiD parity inference
Correct Callaway-Santanna REG influence-function scaling and control-regression uncertainty, refresh parity artifacts, and update coverage guards.

Co-authored-by: Codex <codex@openai.com>
