+++ tests/r_parity/compare.py
@@ -64,7 +64,7 @@ TOLERANCES: dict[str, dict[str, float]] = {
     "01_ols":       {"rel_est": 1e-6, "rel_se": 1e-6},
     "02_iv":        {"rel_est": 1e-6, "rel_se": 1e-6},
     "03_hdfe":      {"rel_est": 1e-6, "rel_se": 1e-2},  # 1-df conv. gap
-    "04_csdid":     {"rel_est": 1e-3, "rel_se": 5.0},   # SE under-covers
+    "04_csdid":     {"rel_est": 1e-3, "rel_se": 1e-2},   # R/Stata analytic SE
     "05_sunab":     {"rel_est": 1e-3, "rel_se": 0.25},
     "06_rd":        {"rel_est": 1.0,  "rel_se": 1.0,    # default-h gap
                      "_forced_rel_est": 1e-3},
@@ -194,9 +194,9 @@ def render_md(modules: list[str]) -> str:
         "Generated by `Paper-JSS/replication/parity/compare.py` on the "
         "`results/<module>_{py,R}.json` artefacts. Tolerance budget per "
         "module is pre-registered in `compare.py::TOLERANCES`. Documented "
-        "convention gaps, common-specification passes, and known SE gaps "
-        "(HDFE 1-df, CS-DiD SE under-coverage, RD default-bandwidth "
-        "selector, SCM non-uniqueness, DML fold-split noise) are flagged "
+        "convention gaps, common-specification passes, and small-sample "
+        "SE conventions (HDFE 1-df, RD default-bandwidth selector, "
+        "SCM non-uniqueness, DML fold-split noise) are flagged "
         "in the per-module `extra` block of the JSON.",
         "",
     ]
@@ -259,8 +259,8 @@ HEADLINE: dict[str, dict[str, Any]] = {
         "name": "CS-DiD simple ATT",
         "headline_filter": lambda d: d.statistic == "simple_ATT",
         "metric": "rel_est",
-        "verdict": "\\textbf{SE GAP}",
-        "gap_note": "SE under-covers; v0.9.6 roadmap",
+        "verdict": "\\textbf{PASS}",
+        "gap_note": "SE within 1\\% analytic tolerance",
     },
     "05_sunab": {
         "name": "Sun--Abraham event study",
@@ -534,9 +534,9 @@ def render_tex(modules: list[str]) -> str:
         "``Worst diff'' column reports the worst residual gap across the "
         "module's headline rows (point estimates only; per-row SE diffs "
         "and documented gap rows are reported in the Markdown source). "
-        "Verdicts use PASS, GAP, and SE GAP; common-specification "
-        "passes and convention gaps are explained in the parenthetical "
-        "notes and per-module \\code{extra} block in "
+        "Verdicts use PASS and GAP; common-specification passes, "
+        "small-sample SE conventions, and convention gaps are explained "
+        "in the parenthetical notes and per-module \\code{extra} block in "
         "\\code{tests/r\\_parity/results/}.}\n"
         "\\label{tab:track-a-parity}\\\\\n"
         "\\toprule\n"
@@ -617,9 +617,10 @@ def render_tex_3way(modules: list[str]) -> str:
         "gap across each module's headline rows (point estimates only; per-row SE diffs and "
         "documented gap rows are reported in \\code{tests/r\\_parity/results/parity\\_table\\_3way.md}). "
         "``no canonical ref'' marks modules whose \\proglang{Stata} ecosystem has no "
-        "authoritative port we can compare to without fabricating one. Verdicts use PASS, "
-        "GAP, and SE GAP; common-specification passes and convention gaps are explained "
-        "in the parenthetical notes and per-module \\code{extra} block in "
+        "authoritative port we can compare to without fabricating one. Verdicts use PASS "
[…]
