### 2026-05-05 9c9f6cf
test(parity): eighth-pass §L.5 closeout — Track B B=1000, NSW+PSID-1, agent bench
- coverage_monte_carlo: add explicit-rate run_b1000.py harness; ship
  coverage_b1000.json with measured 95% CI coverage at B=1000 for
  OLS (0.952), 2x2 DiD (0.955), strong-Z IV (0.962); all inside the
  99% Wilson band [0.935, 0.967] around nominal 0.95. Full slow
  pytest sweep at B=1000 also passes 8/8 (753.51 s).
- orig_parity: add module 04b on causalsens::lalonde.psid (true
  Dehejia-Wahba NSW+PSID-1, 2675 obs); sp.regress + sp.psm match R
  bit-equal and recover the published -15,205 to rel 1.5e-05.
- age
[…]
