### 2026-05-05 8fcaddc
docs(readme): align lead with new agent-native positioning
Mirror the GitHub repo description across README.md / README_CN.md:
foreground "first agent-native Python platform" and surface R/Stata
parity validation in the lead paragraph (was buried in the Task View
comparison further down).

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
