### 2026-05-05 e716ff0
Add JSS validation meta APIs
