+++ src/statspai/help.py
@@ -64,6 +64,7 @@ CATEGORY_DESCRIPTIONS: Dict[str, str] = {
     "bayes": "bayes_did / bayes_rd / bayes_iv / bayes_mte / BCF / policy weights",
     "postestimation": "margins, contrast, pwcompare, test, lincom",
     "agent": "LLM tool-definition surface + JSON schema export",
+    "validation": "JSS reproducibility reports, coverage matrices, table regeneration",
     "utils": "Labels, winsor, DGP simulators, read_data, describe, pwcorr",
     "datasets": "Canonical datasets (Prop99, German reunification, CPS wage, ...)",
     "epi": "Epidemiology — diagnostic tests, kappa, ROC/AUC, NNT, prevalence ratio",
@@ -188,6 +189,7 @@ _MODULE_CATEGORY_PREFIXES: List[tuple] = [
     ("statspai.registry", "agent"),
     ("statspai.help", "agent"),
     ("statspai._agent_docs", "agent"),
+    ("statspai.validation", "validation"),
     ("statspai._auto_estimators", "smart"),
     ("statspai._article_aliases", "causal"),
     ("statspai._citation", "output"),
