+++ src/statspai/validation.py
@@ -1,0 +1,863 @@
+"""Source-tree validation summaries for the StatsPAI paper workflow.
+
+The functions in this module are deliberately metadata-only by default:
+they summarize live registry state and already-materialized validation
+artifacts without running R, Stata, LaTeX, or long Monte Carlo jobs.
+Use :func:`reproduce_jss_tables` for explicit regeneration.
+"""
+
+from __future__ import annotations
+
+import csv
+import json
+import os
+import re
+import shutil
+import subprocess
+import sys
+import time
+from collections import Counter
+from dataclasses import asdict, dataclass, field
+from datetime import datetime, timezone
+from pathlib import Path
+from typing import Any, Dict, Iterable, List, Optional, Union
+
+
+_REPO_ENV = "STATSPAI_REPO_ROOT"
+
+
+@dataclass
+class ValidationReport:
+    """Structured validation evidence for the current StatsPAI checkout."""
+
+    generated_at: str
+    version: str
+    repo_root: Optional[str]
+    registry: Dict[str, Any]
+    evidence: Dict[str, Any]
+    artifacts: Dict[str, Any] = field(default_factory=dict)
+    warnings: List[str] = field(default_factory=list)
+
+    def to_dict(self) -> Dict[str, Any]:
+        """Return a JSON-serializable representation."""
+        return asdict(self)
+
+    def summary(self) -> str:
+        """Return a compact one-paragraph summary."""
+        reg = self.registry
+        ev = self.evidence
+        return (
+            f"StatsPAI {self.version}: {reg.get('total_functions', 0)} registered "
+            f"functions across {reg.get('total_categories', 0)} categories; "
+            f"{ev.get('r_parity', {}).get('matched_modules', 0)} R parity modules, "
+            f"{ev.get('stata_parity', {}).get('modules', 0)} Stata bridge modules, "
+            f"{ev.get('monte_carlo', {}).get('runs', 0)} Monte Carlo coverage rows, "
+            f"{ev.get('agent_bench', {}).get('trials', 0)} agent-benchmark trials."
+        )
+
+    def to_markdown(self) -> str:
+        """Render the report as a concise Markdown audit table."""
+        reg = self.registry
+        ev = self.evidence
+        lines = [
+            "# StatsPAI Validation Report",
+            "",
+            f"- Generated: `{self.generated_at}`",
+            f"- Version: `{self.version}`",
+            f"- Repo root: `{self.repo_root or 'not found'}`",
+            "",
+            "## Registry",
+            "",
+            "| Metric | Value |",
+            "| --- | ---: |",
+            f"| Registered functions | {reg.get('total_functions', 0)} |",
+            f"| Categories | {reg.get('total_categories', 0)} |",
+            f"| Hand-written specs | {reg.get('handwritten_specs', 0)} |",
+            f"| Auto specs | {reg.get('auto_specs', 0)} |",
+            f"| Agent cards | {reg.get('agent_cards', 0)} |",
+            "",
+            "## Validation Evidence",
+            "",
+            "| Track | Evidence |",
+            "| --- | ---: |",
+            (
+                "| R parity | "
+                f"{ev.get('r_parity', {}).get('matched_modules', 0)} matched modules |"
+            ),
+            (
+                "| Stata bridge | "
+                f"{ev.get('stata_parity', {}).get('modules', 0)} modules |"
+            ),
+            (
+                "| Monte Carlo coverage | "
+                f"{ev.get('monte_carlo', {}).get('runs', 0)} rows |"
+            ),
+            (
+                "| Reference parity tests | "
+                f"{ev.get('pytest_inventory', {}).get('reference_parity_files', 0)} files |"
+            ),
+            (
[…]
