### 2026-05-05 301d552
feat(smart): add IV first-stage F gate + adaptive recommend ranking
- preflight: new first_stage_strength check parses the Wilkinson IV
  formula, runs the first-stage OLS, and emits a warning row when the
  partial F falls below the Staiger-Stock (1997) rule of thumb (F<10)
  or the Stock-Yogo (2005) 10% max-size critical value (F<16.38 for
  1 endog/1 IV). Warning payload carries structured recovery_hints
  pointing at method='liml', inference='ar', and
  sp.anderson_rubin_ci(...) so an LLM agent can branch on the typed
  envelope without parsing prose.
- recommend(design='iv'): live first-s
[…]
