+++ src/statspai/smart/recommend.py
@@ -624,23 +624,170 @@ def recommend(
             f"{y} ~ {exog_str} + ({treatment} ~ {z})"
             if exog_str else f"{y} ~ ({treatment} ~ {z})"
         )
-        recommendations.append({
+
+        # Compute the first-stage F live so the ranking + reasons can
+        # adapt to weak instruments (Staiger-Stock 1997 rule of thumb
+        # F=10; Stock-Yogo 2005 10% max-size critical value F=16.38
+        # for one endogenous variable / one instrument).
+        first_stage_F = None
+        weak_iv = False
+        very_weak_iv = False
+        if (treatment and z and treatment in data.columns
+                and z in data.columns):
+            try:
+                d_vec = data[treatment].astype(float).to_numpy()
+                z_vec = data[z].astype(float).to_numpy()
+                exog_arrays = []
+                for c in exog_controls:
+                    if c in data.columns:
+                        try:
+                            exog_arrays.append(
+                                data[c].astype(float).to_numpy()
+                            )
+                        except (TypeError, ValueError):
+                            pass
+                n_obs = len(d_vec)
+                ones = np.ones(n_obs)
+                X_full = np.column_stack(
+                    [ones, z_vec] + exog_arrays
+                )
+                X_restricted = np.column_stack([ones] + exog_arrays)
+                beta_full, *_ = np.linalg.lstsq(
+                    X_full, d_vec, rcond=None
+                )
+                beta_rest, *_ = np.linalg.lstsq(
+                    X_restricted, d_vec, rcond=None
+                )
+                rss_full = float(
+                    np.sum((d_vec - X_full @ beta_full) ** 2)
+                )
+                rss_rest = float(
+                    np.sum((d_vec - X_restricted @ beta_rest) ** 2)
+                )
+                df_denom = n_obs - X_full.shape[1]
+                if rss_full > 0 and df_denom > 0:
+                    first_stage_F = (
+                        ((rss_rest - rss_full) / 1)
+                        / (rss_full / df_denom)
+                    )
+                    very_weak_iv = first_stage_F < 10.0
+                    weak_iv = first_stage_F < 16.38
+            except (np.linalg.LinAlgError, ValueError, KeyError, TypeError):
+                first_stage_F = None
+                weak_iv = False
+                very_weak_iv = False
+
+        # Build the 2SLS recommendation with adaptive reason.
+        twoSLS_reason = 'Standard IV estimator for endogenous treatment.'
+        twoSLS_assumptions = [
+            'Instrument relevance (F > 10)',
+            'Exclusion restriction',
+            'Monotonicity (for LATE)',
+        ]
+        if first_stage_F is not None:
+            twoSLS_reason += (
+                f' First-stage F = {first_stage_F:.2f}'
+            )
+            if very_weak_iv:
+                twoSLS_reason += (
+                    ' < 10 (Staiger-Stock 1997 rule of thumb): 2SLS '
+                    'biased toward OLS, HC1 SEs ignore weak-IV bias. '
+                    'Prefer LIML and Anderson-Rubin inference below.'
+                )
+            elif weak_iv:
+                twoSLS_reason += (
+                    ' < 16.38 (Stock-Yogo 2005 10% max size for 1 '
+                    'endog/1 IV): consider LIML or AR inference.'
+                )
+            else:
+                twoSLS_reason += ' (clears Stock-Yogo 10% max size).'
+
+        twoSLS_rec = {
             'method': '2SLS (two-stage least squares)',
             'function': 'ivreg',
-            'reason': 'Standard IV estimator for endogenous treatment.',
-            'assumptions': ['Instrument relevance (F > 10)',
[…]
