### 2026-05-05 0974aca
test(coverage_mc): add §5.3 robustness DGPs (weak IV, het CS-DiD, overlap loss)
Sibling suite to the canonical Track B Monte-Carlo coverage tests. The
canonical suite checks that 95% CIs hit nominal coverage on
well-specified DGPs; this suite checks the same estimators on DGPs that
violate or stress identification assumptions, with documented coverage
bands acting as regression guards.

- test_coverage_robustness.py (3 slow tests):
  * test_iv_weak_instrument_undercoverage: pi=0.10 first stage,
    F_med ≈ 3 (Stock-Yogo 10% size CV is 16.4); HC1 expected to under-
    cover. Documented band [0
[…]
