+++ tests/coverage_monte_carlo/run_robustness_b1000.py
@@ -1,0 +1,231 @@
+r"""Track B robustness DGPs at B=1000.
+
+Companion to ``run_b1000.py`` — that script reports B=1000 coverage on
+the canonical (well-specified) DGPs.  This script re-runs the three
+robustness DGPs from ``test_coverage_robustness.py`` at B=1000 and
+writes ``results_b1000/coverage_robustness_b1000.json`` for §5.3 of the
+JSS draft.
+
+Key difference from the canonical sweep: these rows are *expected* to
+miss nominal 0.95 calibration (e.g. weak IV under-covers at ~0.88
+because HC1 SEs ignore the weak-instrument bias of 2SLS).  The runner
+reports the actual rate; the manuscript discusses the band rather than
+asserting calibration.
+
+The DGPs are duplicated between this runner and the pytest module
+``test_coverage_robustness.py``.  That duplication is deliberate: the
+pytest module asserts a *test-friendly* band measured at B=200/300
+(wider Wilson interval), while this runner emits the *manuscript-grade*
+explicit-rate JSON at B=1000 with a tighter band.  If you change a
+DGP, change both files.
+
+Re-execute end-to-end with::
+
+    python tests/coverage_monte_carlo/run_robustness_b1000.py
+
+Total wall-clock: ~10-15 minutes on Apple M3 Pro / 36 GB.
+"""
+from __future__ import annotations
+
+import json
+import time
+import warnings
+from pathlib import Path
+
+import numpy as np
+import pandas as pd
+
+import statspai as sp
+
+
+HERE = Path(__file__).resolve().parent
+RESULTS_DIR = HERE / "results_b1000"
+RESULTS_DIR.mkdir(parents=True, exist_ok=True)
+
+B = 1000
+
+
+def _ci_covers(ci, truth) -> bool:
+    if ci is None or len(ci) != 2:
+        return False
+    lo, hi = ci
+    return lo <= truth <= hi
+
+
+# ---------------------------------------------------------------------------
+# Robustness 1: weak IV
+# ---------------------------------------------------------------------------
+
+
+def coverage_weak_iv() -> dict:
+    """2SLS HC1 on weak instrument (pi=0.10, F_med ~ 3)."""
+    truth = 1.0
+    pi = 0.10
+    eu = 1.5
+    n = 600
+    covered = 0
+    Fs = []
+    bias = []
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore")
+        for seed in range(B):
+            rng = np.random.default_rng(seed)
+            z = rng.normal(size=n)
+            u = rng.normal(size=n)
+            d = pi * z + eu * u + rng.normal(scale=0.4, size=n)
+            y = truth * d + u + rng.normal(scale=0.5, size=n)
+            df = pd.DataFrame({"y": y, "d": d, "z": z})
+            # Compute first-stage F directly for record-keeping
+            zc = z - z.mean()
+            dc = d - d.mean()
+            beta_fs = (zc * dc).sum() / (zc * zc).sum()
+            d_hat = d.mean() + beta_fs * zc
+            ssr = ((d - d_hat) ** 2).sum()
+            tss = ((d - d.mean()) ** 2).sum()
+            r2 = 1 - ssr / tss
+            F = r2 / (1 - r2) * (n - 2)
+            Fs.append(F)
+            r = sp.ivreg("y ~ (d ~ z)", data=df, robust="hc1")
+            beta = float(r.params["d"])
+            ci = r.conf_int()
+            lo, hi = ci.loc["d"].values
+            bias.append(beta - truth)
+            if lo <= truth <= hi:
+                covered += 1
+    return {
[…]
