+++ tests/coverage_monte_carlo/test_coverage_robustness.py
@@ -1,0 +1,238 @@
+r"""Monte Carlo CI coverage under *robustness* DGPs.
+
+The canonical Track B suite (``test_coverage.py``) checks that 95% CIs
+hit nominal coverage on **well-specified** DGPs.  This sibling suite
+checks the same estimators on DGPs that violate or stress the
+identification assumptions (weak instruments, heterogeneous DiD timing,
+severe propensity-score overlap loss).
+
+The contract is intentionally different from the canonical suite.  These
+tests do **not** assert that empirical coverage equals the nominal 0.95.
+They assert that empirical coverage falls within a *documented band*
+that we have measured at pilot time.  The band serves two purposes:
+
+1. ``regression``: if the band is violated (coverage drops below the
+   floor, or jumps above the ceiling), some estimator-internal change
+   has shifted the calibration of the inference procedure on this DGP.
+   Either the estimator improved (e.g. a fix lifts coverage) or it
+   regressed (e.g. a refactor broke an SE term).  Either way, a
+   maintainer should investigate.
+
+2. ``honest documentation``: weak instruments are *expected* to under-
+   cover with HC1 SEs.  AIPW with severe overlap loss is expected to
+   over-cover (because the IF blows up).  Pretending these failures
+   don't exist would be marketing, not science.  The bands let us
+   record the magnitude of the textbook failure mode while still using
+   the band as a regression guard against silent drift.
+
+Bands were measured by direct simulation at pilot time and are
+reproducible from a fixed seed schedule.  See ``run_robustness_b1000.py``
+for the explicit-rate harness used in §5.3 of the JSS draft.
+
+References
+----------
+- Stock, J. H. and Yogo, M. (2005). Testing for weak instruments in
+  linear IV regression.  In *Identification and Inference for
+  Econometric Models*, ed. D. W. K. Andrews and J. H. Stock,
+  pp. 80-108.
+- Callaway, B. and Sant'Anna, P. H. C. (2021). Difference-in-differences
+  with multiple time periods. JoE 225(2), 200-230.
+- Crump, R. K., Hotz, V. J., Imbens, G. W., and Mitnik, O. A. (2009).
+  Dealing with limited overlap in estimation of average treatment
+  effects. Biometrika 96(1), 187-199.
+"""
+from __future__ import annotations
+
+import os
+import warnings
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+
+
+# Default draws per test.  Override via env STATSPAI_MC_DRAWS=N.
+B_DEFAULT = int(os.environ.get('STATSPAI_MC_DRAWS', 300))
+
+
+def _coverage_rate(covered: int, B: int) -> float:
+    return covered / B
+
+
+def _assert_documented_band(covered: int, B: int, label: str,
+                            lo: float, hi: float) -> None:
+    """Assert empirical coverage lies in the documented [lo, hi] band.
+
+    For robustness DGPs this band reflects the *known textbook
+    failure-mode envelope* (under-cover for weak IV, over-cover for
+    AIPW under poor overlap, etc.) rather than the Wilson band around
+    nominal 0.95.
+    """
+    rate = _coverage_rate(covered, B)
+    assert lo <= rate <= hi, (
+        f"{label} coverage = {rate:.3f} outside documented "
+        f"[{lo:.3f}, {hi:.3f}] band (B={B}). If this is a true "
+        f"improvement, widen the band; if it is a regression, "
[…]

[…]
