### 2026-05-05 958c235
fix(agent): chain Honest-DiD from result handles
