### 2026-05-05 1f993aa
docs(bib): add imbens2009recent to paper.bib
Adds Imbens & Wooldridge (2009) "Recent Developments in the
Econometrics of Program Evaluation" (JEL 47(1), 5-86,
doi:10.1257/jel.47.1.5) to the curated bibliography so it can
ground the program-evaluation survey citation in the JSS
manuscript's §1 motivation and §4 worked-examples overview.

Four-element bib record (authors / year / journal+page range /
DOI) verified against Crossref (api.crossref.org) and AEAweb
(aeaweb.org/articles?id=10.1257/jel.47.1.5) per CLAUDE.md §10
zero-citation-hallucination protocol; refs verified via Crossref,
AEAweb.


[…]
