+++ paper.bib
@@ -596,6 +596,17 @@
   doi={10.1093/jrsssb/qkae095}
 }
 
+@article{imbens2009recent,
+  title={Recent Developments in the Econometrics of Program Evaluation},
+  author={Imbens, Guido W. and Wooldridge, Jeffrey M.},
+  journal={Journal of Economic Literature},
+  year={2009},
+  volume={47},
+  number={1},
+  pages={5--86},
+  doi={10.1257/jel.47.1.5}
+}
+
 @article{abadie2003economic,
   title={The Economic Costs of Conflict: A Case Study of the Basque Country},
   author={Abadie, Alberto and Gardeazabal, Javier},
