### 2026-05-05 c3d8ba2
feat(smart+agent_bench): wire weak-IV warning into RecommendationResult; agent-bench grader CLI; relax cs_report smoke
The weak-IV preflight + adaptive ranking from `sp.preflight` and
`sp.recommend(... design='iv')` already adapted the per-row reasons in
the §P pass. This pass surfaces the same signal at the top-level
`RecommendationResult.warnings` field so workflow-orchestration code
sees it without parsing per-row prose.

- recommend.py: append a top-level warning when first-stage F < 10
  (Staiger-Stock 1997) or F < 16.38 (Stock-Yogo 2005, 1 endog/1 IV);
  the wording mirrors the `sp.prefl
[…]
