### 2026-05-06 54d83b4
feat(neural): opt-in GPU/MPS routing via STATSPAI_TORCH_DEVICE env var
Centralises the previously hardcoded ``device = torch.device('cpu')``
pattern across deepiv/, neural_causal/cevae and neural_causal/models
into a single resolver in utils/_torch_device.py. Default behaviour is
CPU (preserves bit-for-bit numerics on existing pinned tests); GPU/MPS
dispatch is opt-in via STATSPAI_TORCH_DEVICE={cpu,cuda,mps,auto}.

Explicit cuda/mps requests raise loudly when the device is unavailable
rather than silently falling back to CPU — matches the "fail loudly"
rule in §3.7 of CLAUDE.md.

Also replaces
[…]
