### 2026-05-06 5920320
feat(postestimation): add postestimation_contract / postestimation_report
The rest of `sp.postestimation` provides Stata/R-style commands
(`margins`, `lincom`, `test`). This module answers the question that
comes _before_ calling them: which post-estimation actions does this
fitted result actually support?

`sp.postestimation_contract(result, data=df)` returns a structured
dict with `available` / `missing` / `recommended_next` / optional
`diagnostics`, derived from the result's attributes (`tidy`,
`glance`, `params`, `std_errors`, `effect`, `predict`, `violations`,
provenance) plus any data th
[…]
