+++ src/statspai/postestimation/__init__.py
@@ -15,9 +15,11 @@ from .margins import (
     margins_at_plot, contrast, pwcompare,
 )
 from .hypothesis import test, lincom
+from .contract import postestimation_contract, postestimation_report
 
 __all__ = [
     'margins', 'margins_table', 'event_study_table',
     'marginsplot', 'margins_at', 'margins_at_plot',
     'contrast', 'pwcompare', 'test', 'lincom',
+    'postestimation_contract', 'postestimation_report',
 ]
