+++ src/statspai/postestimation/contract.py
@@ -1,0 +1,181 @@
+"""Post-estimation capability contract for fitted StatsPAI results.
+
+The rest of :mod:`statspai.postestimation` provides Stata/R-style
+commands (``margins``, ``lincom``, ``test``). This module answers the
+agent-facing question that comes before calling any of them: which
+post-estimation actions are actually available for this result object?
+"""
+
+from __future__ import annotations
+
+from typing import Any, Dict, Iterable, List, Optional
+
+import pandas as pd
+
+
+_CORE_ACTIONS = {
+    "tidy": "long coefficient/effect table",
+    "glance": "one-row model summary",
+    "summary": "human-readable summary",
+    "to_dict": "structured result payload",
+    "to_json": "JSON serialization",
+}
+
+_EXPORT_ACTIONS = {
+    "to_latex": "LaTeX table/document export",
+    "to_docx": "Word export",
+    "to_html": "HTML/Jupyter rendering",
+}
+
+
+def postestimation_contract(
+    result: Any,
+    *,
+    data: Optional[pd.DataFrame] = None,
+    include_diagnostics: bool = True,
+) -> Dict[str, Any]:
+    """Return the post-estimation actions supported by *result*.
+
+    Parameters
+    ----------
+    result : object
+        Fitted StatsPAI result, estimator, or compatible object.
+    data : DataFrame, optional
+        Analysis frame available for data-dependent actions. When
+        supplied, the contract marks ``margins``/``predict`` paths as
+        ready rather than merely method-available.
+    include_diagnostics : bool, default True
+        Include scalar diagnostics from ``model_info`` / ``diagnostics``.
+
+    Returns
+    -------
+    dict
+        Machine-readable contract with ``available``, ``missing``,
+        ``recommended_next``, and optional ``diagnostics`` keys.
+    """
+    if result is None:
+        raise ValueError("postestimation_contract requires a fitted result object")
+
+    available: Dict[str, str] = {}
+    missing: Dict[str, str] = {}
+
+    for name, description in _CORE_ACTIONS.items():
+        _record_capability(result, name, description, available, missing)
+    for name, description in _EXPORT_ACTIONS.items():
+        _record_capability(result, name, description, available, missing)
+
+    has_params = all(hasattr(result, attr) for attr in ("params", "std_errors"))
+    has_effect = all(hasattr(result, attr) for attr in ("estimate", "se", "ci"))
+    has_model_data = data is not None
+
+    if callable(getattr(result, "predict", None)) or callable(getattr(result, "effect", None)):
+        available["predict"] = "prediction or treatment-effect prediction"
+    else:
+        missing["predict"] = "result has no predict/effect method"
+
+    if has_params:
+        available["lincom"] = "linear combinations of coefficients"
+        available["test"] = "Wald tests of linear restrictions"
+        if has_model_data:
+            available["margins"] = "marginal effects using supplied data"
+        else:
+            missing["margins"] = "requires the analysis data frame"
+    else:
+        missing["lincom"] = "requires coefficient vector and covariance/SEs"
+        missing["test"] = "requires coefficient vector and covariance/SEs"
+        missing["margins"] = "requires coefficient model plus analysis data"
+
+    if has_effect:
+        available["effect_summary"] = "single-estimand effect summary"
+    else:
+        missing["effect_summary"] = "no scalar estimate/se/ci attributes"
+
+    if callable(getattr(result, "violations", None)):
+        available["violations"] = "assumption and diagnostic violations"
+    else:
+        missing["violations"] = "result does not expose violations()"
+
+    diagnostics: Dict[str, Any] = {}
+    if include_diagnostics:
[…]
