### 2026-05-06 da9fe55
feat(paper): reviewer_mode adds Reviewer Audit section
`sp.paper(..., reviewer_mode=True)` and the `CausalQuestion.paper()`
shortcut now insert a "Reviewer Audit" section before "References"
that surfaces evidence a reviewer / replication agent needs without
re-running the workflow:

* Registry validation: estimator's `stability`,
  `validation_status`, validation notes, and known limitations.
* Identification verdict + findings from the workflow diagnostics.
* Post-estimation surface from `sp.postestimation_contract(result)`.
* Result violations from `result.violations()` if implemented.
* Pr
[…]
