+++ src/statspai/workflow/paper.py
@@ -213,7 +213,7 @@ class PaperDraft:
         order = [
             "Question", "Data", "Identification",
             "Estimator", "Results", "Robustness",
-            "Pipeline notes", "Causal DAG", "References",
+            "Reviewer Audit", "Pipeline notes", "Causal DAG", "References",
         ]
         chunks: List[str] = []
         for title in order:
@@ -405,7 +405,7 @@ class PaperDraft:
         order = [
             "Question", "Data", "Identification",
             "Estimator", "Results", "Robustness",
-            "Pipeline notes", "Causal DAG", "References",
+            "Reviewer Audit", "Pipeline notes", "Causal DAG", "References",
         ]
         # When self.dag is set, regenerate the Causal DAG body with the
         # Quarto-native mermaid block instead of the markdown text-art.
@@ -1021,6 +1021,124 @@ def _section_from_workflow(
     return sections
 
 
+def _reviewer_audit_section(
+    *,
+    workflow: Any = None,
+    result: Any = None,
+    estimator: Optional[str] = None,
+    degradations: Optional[List[Dict[str, Any]]] = None,
+) -> str:
+    """Build a reviewer-facing audit section for a fitted draft."""
+    lines: List[str] = []
+    lines.append("**Reviewer-mode audit**")
+    lines.append("")
+
+    # Estimator registry evidence.
+    est_name = estimator
+    if not est_name and workflow is not None:
+        rec = getattr(workflow, "recommendation", None)
+        try:
+            if rec is not None and rec.recommendations:
+                est_name = rec.recommendations[0].get("function")
+        except Exception:
+            est_name = None
+    if est_name:
+        try:
+            from ..registry import describe_function
+            spec = describe_function(est_name)
+            lines.append(
+                f"- **Registry**: `sp.{est_name}` is "
+                f"`stability={spec.get('stability')}`, "
+                f"`validation_status={spec.get('validation_status')}`."
+            )
+            notes = spec.get("validation_notes") or []
+            if notes:
+                shown = "; ".join(str(n) for n in notes[:3])
+                lines.append(f"- **Validation evidence**: {shown}.")
+            limitations = spec.get("limitations") or []
+            if limitations:
+                lines.append("- **Known implementation limitations**:")
+                lines.extend(f"    - {lim}" for lim in limitations[:5])
+        except Exception as exc:
+            lines.append(
+                f"- **Registry**: lookup for `sp.{est_name}` failed "
+                f"({type(exc).__name__}: {exc})."
+            )
+
+    # Identification report.
+    diag = getattr(workflow, "diagnostics", None) if workflow is not None else None
+    if diag is not None:
+        verdict = getattr(diag, "verdict", None)
+        if verdict:
+            lines.append(f"- **Identification verdict**: `{verdict}`.")
+        findings = getattr(diag, "findings", None) or []
+        if findings:
+            lines.append("- **Identification findings to defend**:")
+            for finding in findings[:6]:
+                sev = getattr(finding, "severity", "info")
+                msg = getattr(finding, "message", str(finding))
+                lines.append(f"    - [{str(sev).upper()}] {msg}")
+
+    # Post-estimation contract.
+    target = result or getattr(workflow, "result", None)
+    if target is not None:
+        try:
+            from ..postestimation import postestimation_contract
+            contract = postestimation_contract(target)
+            available = ", ".join(sorted(contract["available"].keys())[:10])
+            lines.append(f"- **Post-estimation surface**: {available}.")
+        except Exception as exc:
+            lines.append(
+                f"- **Post-estimation surface**: unavailable "
+                f"({type(exc).__name__}: {exc})."
+            )
+
+        violations = getattr(target, "violations", None)
[…]
