### 2026-05-06 1ac1ba3
feat(fast): JAX-backed feols_jax for end-to-end GPU/TPU OLS
Adds ``sp.fast.feols_jax`` — a drop-in replacement for
``sp.fast.feols`` whose WLS solve and HC1 sandwich computations execute
on JAX/XLA. FE residualisation continues to delegate to the Rust
``demean`` kernel (bincount memory pattern beats XLA on CPU); CR1
cluster sandwich continues to delegate to ``crve`` (which itself uses
the Phase 2 Rust ``cluster_meat`` when built). The unique value is on
GPU/TPU boxes — the dense linear algebra step runs at accelerator
speed without any host↔device ping-pong inside the solve.

WLS uses a sqrt-W
[…]
