+++ src/statspai/fast/jax_feols.py
@@ -1,0 +1,374 @@
+"""JAX-backed end-to-end ``feols``.
+
+This module ships a drop-in replacement for :func:`statspai.fast.feols`
+whose **WLS solve + HC1 sandwich** computations execute on JAX/XLA
+(CPU / GPU / TPU). The FE residualisation step still goes through the
+Rust ``demean`` kernel because, on typical FE cardinalities, the
+bincount-style memory pattern beats anything XLA emits on CPU. The
+``cr1`` cluster-robust path also stays on the existing
+:func:`statspai.fast.inference.crve`, which itself dispatches to the
+Phase-2 Rust ``cluster_meat`` kernel when built. The unique value of
+this module is on **GPU** boxes: the post-demean dense linear algebra
+step runs at GPU speed without any host↔device ping-pong inside the
+solve.
+
+Honest scope
+------------
+- The dev environment used to write this code has **no CUDA**, only
+  Apple-Silicon MPS. JAX-on-MPS is unofficial and the parity tests run
+  on the CPU JAX path; the GPU promise is structural (XLA auto-
+  dispatches to whatever ``jax.devices()[0]`` points at).
+- ``cr1`` parity is delegated; we test that ``feols_jax(vcov="cr1")``
+  returns the same matrix as ``feols(vcov="cr1")``.
+- Default ``dtype="float64"`` so existing pinned numerical tests stay
+  bit-comparable. Pass ``dtype="float32"`` on a GPU to trade ~1 ulp of
+  precision for the XLA float32 fast path.
+"""
+from __future__ import annotations
+
+from typing import List, Optional
+
+import numpy as np
+import pandas as pd
+
+from .feols import FeolsResult
+
+
+# ---------------------------------------------------------------------------
+# JAX availability + helpers (mirrors jax_backend.py's policy)
+# ---------------------------------------------------------------------------
+
+try:
+    import jax
+
+    # StatsPAI defaults to float64; XLA truncates unless explicitly enabled.
+    jax.config.update("jax_enable_x64", True)
+    import jax.numpy as jnp
+    from jax import jit
+
+    _HAS_JAX = True
+except ImportError:  # pragma: no cover - exercised on no-jax CI
+    jax = None  # type: ignore[assignment]
+    jnp = None  # type: ignore[assignment]
+    jit = None  # type: ignore[assignment]
+    _HAS_JAX = False
+
+
+# ---------------------------------------------------------------------------
+# JIT-compiled core: WLS solve + iid/HC1 sandwich
+# ---------------------------------------------------------------------------
+
+def _make_jax_kernels():
+    """Build JIT-compiled solvers lazily so module import is jax-free."""
+    if not _HAS_JAX:
+        raise ImportError(
+            "jax is not installed; pip install jax jaxlib to enable the JAX "
+            "feols backend."
+        )
+
+    @jit
+    def _wls_solve(X, y, w):
+        """QR-based weighted-least-squares solve.
+
+        Returns (beta, residuals, rss, XtWX_inv).
+        """
+        # Sqrt-W trick: solve (sqrt(W) X) beta = sqrt(W) y so we can use
+        # a plain QR. Numerically equivalent to the normal-equation
+        # ``inv(X' W X) X' W y`` but better-conditioned for ill-shaped X.
+        sw = jnp.sqrt(w)
+        Xw = X * sw[:, None]
+        yw = y * sw
+
+        Q, R = jnp.linalg.qr(Xw, mode="reduced")
+        beta = jnp.linalg.solve(R, Q.T @ yw)
[…]
