### 2026-05-06 c339063
feat(iv): sp.iv(..., absorb=...) HDFE Phase 3 — 2SLS with absorbed FEs
Adds high-dimensional fixed-effects absorption to sp.iv via the same
sp.fast.demean (Rust HDFE) backend used by sp.fast.feols. y, exogenous
controls, endogenous regressors, and instruments are residualised against
the absorbed FE block under the Frisch–Waugh–Lovell theorem; the intercept
is dropped because the absorbed FEs span the constant. Residual DOF is
charged sum(G_k - 1), mirroring sp.fast.feols(absorb=...), and the vcov is
rescaled by (n - k) / (n - k - fe_dof) so nonrobust, HC1, and CR1 SE all
report the correct sm
[…]
