+++ src/statspai/regression/iv.py
@@ -1033,6 +1033,243 @@ class IVRegression(BaseModel):
 #  Unified public API: iv()
 # ====================================================================== #
 
+# ====================================================================== #
+#  Absorb (HDFE) preprocessing for sp.iv(..., absorb=...)
+# ====================================================================== #
+
+def _normalise_absorb(absorb: Optional[Union[str, List[str]]]) -> List[str]:
+    """Normalise an ``absorb=`` argument to a list of column names.
+
+    Accepts ``None``, ``"firm"``, ``"firm + year"``, or ``["firm", "year"]``.
+    """
+    if absorb is None:
+        return []
+    if isinstance(absorb, str):
+        return [t.strip() for t in absorb.split('+') if t.strip()]
+    return [str(t) for t in absorb]
+
+
+def _iv_absorb_preprocess(
+    formula: str,
+    data: pd.DataFrame,
+    absorb_terms: List[str],
+    cluster_name: Optional[str] = None,
+    fe_tol: float = 1e-10,
+    fe_maxiter: int = 1_000,
+) -> Dict[str, Any]:
+    """Demean IV inputs by ``absorb_terms`` via the HDFE Phase 1 kernel.
+
+    Returns a dict with the residualised matrices, var-name dictionary,
+    cluster series (post-singleton-mask), and FE diagnostics. The
+    intercept is dropped because the absorbed FEs span the constant.
+
+    Same convention as ``sp.fast.feols``: ``fe_dof = sum(G_k - 1)``.
+    """
+    # Lazy import — keeps regression/iv.py free of fast/* dependencies
+    # at module import time.
+    from ..fast.demean import demean as _demean
+
+    parsed = parse_formula(formula)
+    if not parsed['endogenous'] or not parsed['instruments']:
+        raise ValueError(
+            "IV formula must specify endogenous variables and instruments. "
+            "Use syntax: \"y ~ (endog ~ z1 + z2) + exog\""
+        )
+
+    dependent = parsed['dependent']
+    exog_names = parsed['exogenous']
+    endog_names = parsed['endogenous']
+    instrument_names = parsed['instruments']
+
+    needed = [dependent] + exog_names + endog_names + instrument_names
+    needed += list(absorb_terms)
+    if cluster_name is not None:
+        needed.append(cluster_name)
+    missing = [v for v in needed if v not in data.columns]
+    if missing:
+        raise ValueError(f"Variables not found in data: {missing}")
+    missing_absorb = [c for c in absorb_terms if c not in data.columns]
+    if missing_absorb:
+        raise ValueError(
+            f"absorb columns not found in data: {missing_absorb}"
+        )
+
+    clean = data[needed].dropna(subset=needed)
+    n_obs = len(clean)
+    if n_obs == 0:
+        raise ValueError("No rows remain after dropping NaNs")
+
+    y = clean[dependent].to_numpy(dtype=np.float64)
+    X_exog = (
+        clean[exog_names].to_numpy(dtype=np.float64)
+        if exog_names else np.empty((n_obs, 0), dtype=np.float64)
+    )
+    if X_exog.ndim == 1:
+        X_exog = X_exog.reshape(-1, 1)
+    X_endog = clean[endog_names].to_numpy(dtype=np.float64)
+    if X_endog.ndim == 1:
+        X_endog = X_endog.reshape(-1, 1)
+    Z = clean[instrument_names].to_numpy(dtype=np.float64)
+    if Z.ndim == 1:
+        Z = Z.reshape(-1, 1)
+
+    n_exog = X_exog.shape[1]
+    n_endog = X_endog.shape[1]
+    n_z = Z.shape[1]
+
[…]
