### 2026-05-06 4782152
feat(fast): Phase 4c — wild + wild_cluster bootstrap via score formulation
Adds two new bootstrap variants to ``sp.fast.feols_jax_bootstrap``:
- ``bootstrap='wild'`` — Wu (1986) row-level wild bootstrap with
  Rademacher signs ``η_i ∈ {-1, +1}`` per row.
- ``bootstrap='wild_cluster'`` — Cameron-Gelbach-Miller (2008) §III.B
  wild cluster bootstrap with Rademacher signs drawn per cluster.
  The standard tool for few-cluster inference (G < 30).

Both variants use the **score formulation**

    β* = β̂ + (X'WX)^{-1} X'W (η ⊙ û)

instead of literally refitting OLS on the pseudo-outcome
``y* = X β̂
[…]
