+++ tests/test_jax_feols_bootstrap.py
@@ -154,15 +154,144 @@ def test_percentile_ci_contains_true_value_for_well_specified_dgp():
 def test_invalid_bootstrap_type_raises():
     df = _make_panel(seed=20)
     with pytest.raises(ValueError, match="bootstrap="):
-        feols_jax_bootstrap("y ~ x1", df, n_boot=10, bootstrap="wild")
+        feols_jax_bootstrap("y ~ x1", df, n_boot=10, bootstrap="bayesian")
 
 
 def test_cluster_bootstrap_without_cluster_raises():
     df = _make_panel(seed=21)
-    with pytest.raises(ValueError, match="cluster="):
+    with pytest.raises(ValueError, match="cluster"):
         feols_jax_bootstrap(
             "y ~ x1", df, n_boot=10, bootstrap="cluster",
         )
+
+
+def test_wild_cluster_without_cluster_raises():
+    df = _make_panel(seed=21)
+    with pytest.raises(ValueError, match="cluster"):
+        feols_jax_bootstrap(
+            "y ~ x1", df, n_boot=10, bootstrap="wild_cluster",
+        )
+
+
+# ---------------------------------------------------------------------------
+# Phase 4c: wild + wild_cluster bootstrap
+# ---------------------------------------------------------------------------
+
+def test_wild_bootstrap_se_converges_to_hc1():
+    """Wild row-level SE → HC1 SE as B → ∞ (asymptotic equivalence)."""
+    df = _make_panel(n=2_000, n_firm=80, seed=40)
+    fit_hc1 = feols("y ~ x1 + x2 | firm", df, vcov="hc1")
+    boot = feols_jax_bootstrap(
+        "y ~ x1 + x2 | firm", df, n_boot=2_000, seed=0,
+        bootstrap="wild",
+    )
+    # Wild bootstrap with Rademacher weights converges to HC SE (not
+    # HC1 — the n/(n-k) factor is asymptotically 1). 10% rtol matches
+    # the pairs-bootstrap convergence rate at B=2000.
+    np.testing.assert_allclose(
+        boot.se_boot["x1"], fit_hc1.se()["x1"], rtol=0.10,
+    )
+    np.testing.assert_allclose(
+        boot.se_boot["x2"], fit_hc1.se()["x2"], rtol=0.10,
+    )
+
+
+def test_wild_cluster_bootstrap_se_converges_to_cr1():
+    """Wild cluster SE → CR1 SE as B → ∞."""
+    df = _make_panel(n=2_000, n_firm=80, seed=41)
+    fit_cr1 = feols(
+        "y ~ x1 + x2 | firm", df, vcov="cr1", cluster="cluster",
+    )
+    boot = feols_jax_bootstrap(
+        "y ~ x1 + x2 | firm", df, n_boot=2_000, seed=0,
+        bootstrap="wild_cluster", cluster="cluster",
+    )
+    # Wild cluster convergence is slower; allow 15%.
+    np.testing.assert_allclose(
+        boot.se_boot["x1"], fit_cr1.se()["x1"], rtol=0.15,
+    )
+
+
+def test_wild_bootstrap_point_estimate_matches_feols_jax():
+    """Point estimate is from the original fit, identical across variants."""
+    df = _make_panel(seed=42)
+    fit = feols_jax("y ~ x1 + x2 | firm", df, vcov="iid")
+    for variant in ("pairs", "cluster", "wild", "wild_cluster"):
+        kw = {"n_boot": 50, "seed": 0, "bootstrap": variant}
+        if variant in ("cluster", "wild_cluster"):
+            kw["cluster"] = "cluster"
[…]
