+++ src/statspai/fast/jax_feols.py
@@ -498,7 +498,51 @@ def _make_bootstrap_kernels():
 
         return _one_cluster_boot
 
-    return _one_pairs_boot, _build_cluster_boot
+    # ─── Phase 4c: wild + wild cluster bootstrap (score formulation) ───
+    #
+    # Wild bootstrap (Wu 1986, Cameron-Gelbach-Miller 2008): pseudo-y*_i
+    # = X_i β̂ + η_i ⊙ û_i with η_i Rademacher draws. Refitting OLS on
+    # (X, y*) gives β* = β̂ + (X'WX)^{-1} X'W (η ⊙ û). The right-hand
+    # form is the "score bootstrap" — mathematically identical, much
+    # cheaper per iteration (one mat-vec instead of a full QR), and
+    # exposes the per-iteration randomness as a single Rademacher draw.
+    #
+    # ``rademacher`` is implemented as ``2 * Bernoulli(0.5) - 1`` for
+    # portability across JAX versions (``jax.random.rademacher`` only
+    # arrived in jax >= 0.4.x).
+
+    @jit
+    def _one_wild_boot(key, X, w_base, residuals, beta_hat, XtWX_inv):
+        """One row-level wild-bootstrap iteration (score form)."""
+        n = X.shape[0]
+        eta = (2 * jax.random.bernoulli(key, p=0.5, shape=(n,))
+               .astype(X.dtype) - 1)
+        score = X.T @ (w_base * eta * residuals)        # (p,)
+        return beta_hat + XtWX_inv @ score
+
+    def _build_wild_cluster_boot(n_clusters: int):
+        """Build a JIT-compiled wild-cluster-bootstrap kernel for a fixed
+        ``n_clusters``. Each iteration draws one Rademacher per cluster
+        and broadcasts to rows via ``cluster_codes``."""
+        n_clusters_int = int(n_clusters)
+
+        @jit
+        def _one_wild_cluster_boot(
+            key, X, w_base, residuals, beta_hat, XtWX_inv, cluster_codes,
+        ):
+            eta_g = (2 * jax.random.bernoulli(
+                key, p=0.5, shape=(n_clusters_int,)
+            ).astype(X.dtype) - 1)
+            eta = eta_g[cluster_codes]                  # (n,)
+            score = X.T @ (w_base * eta * residuals)   # (p,)
+            return beta_hat + XtWX_inv @ score
+
+        return _one_wild_cluster_boot
+
+    return (
+        _one_pairs_boot, _build_cluster_boot,
+        _one_wild_boot, _build_wild_cluster_boot,
+    )
 
 
 def _jax_prep_inputs(
@@ -690,19 +734,29 @@ def feols_jax_bootstrap(
         Number of bootstrap resamples.
     seed : int, default 0
         Seed for the JAX PRNG.
-    bootstrap : {"pairs", "cluster"}, default "pairs"
+    bootstrap : {"pairs", "cluster", "wild", "wild_cluster"}, default "pairs"
         - ``"pairs"`` — Efron pairs bootstrap; each iteration resamples
           rows with replacement (multinomial counts become the
-          bootstrap weights).
-        - ``"cluster"`` — cluster bootstrap (Cameron, Gelbach & Miller
-          2008 §III.A); each iteration resamples *clusters* with
+          bootstrap weights). Asymptotic target: HC1 SE.
+        - ``"cluster"`` — Cameron-Gelbach-Miller (2008) §III.A cluster
+          bootstrap; each iteration resamples *clusters* with
           replacement; observations in a cluster sampled k times get
-          weight k. Use for inference with few-but-non-tiny clusters
[…]
