### 2026-05-06 b4ae593
docs(gpu): Phase 4c follow-up — wild + wild_cluster in guide and README; drop unverified Efron 1979 attribution
Self-review pass on the v1.14 GPU docs:

1. Citation hygiene (CLAUDE.md §10). The Phase 4 GPU guide
   contained a textual ``(Efron 1979)`` attribution for pairs
   bootstrap that is **not** in paper.bib and was pulled from
   training-corpus memory. Per the zero-hallucination rule, drop
   it. ``cameron2008bootstrap`` is the only bootstrap citation
   we use; it has a verified DOI in paper.bib.

2. Phase 4c surface coverage. The ``Wild cluster bootstrap`` line
   in the GPU guide's 
[…]
