### 2026-05-06 b57e99d
feat(iv): modern reporting bundle sp.iv.iv_diag + plots + post-2022 refs
Polish pass on sp.iv to the post-2022 reporting standard. New entry
point sp.iv.iv_diag (also sp.iv_diag at top level) consolidates the
modern weak-IV / sensitivity / bootstrap toolkit in one call —
parity with R ivDiag (Lal et al. 2024 Political Analysis 32(4),
521–540) plus a TSLS-as-LATE caveat and rich plot/export helpers
that R ivDiag does not ship.

Returned IVDiagResult contains:
- 2SLS analytic + pairs / wild Rademacher bootstrap CIs (cluster-aware)
- Olea–Pflueger effective F (already in sp); LMMP 2022 tF-correct
[…]
