+++ paper.md
@@ -96,6 +96,21 @@ analysis, and publication-ready output within a single API---nor
 exposes machine-readable function schemas for LLM-driven workflows.
 `StatsPAI` fills this gap.
 
+The synthetic-control sub-ecosystem in particular is fragmented across
+languages and packages: R offers `Synth` [@abadie2010synthetic] for the
+classical estimator, `gsynth` [@xu2017generalized] for interactive
+fixed effects, `augsynth` [@benmichael2021augmented] for the augmented
+estimator, `synthdid` [@arkhangelsky2021synthetic] for synthetic DID,
+`scpi` [@cattaneo2025scpi] for prediction intervals, and `fect`
+[@liu2024practical] for unified counterfactual diagnostics. Stata
+exposes parallel implementations including the `sdid` package
+[@clarke2024synthetic]. `StatsPAI`'s `sp.synth(method=...)` dispatcher
+folds these methodological branches into a single Python entry point
+with shared input/output schema, common diagnostics, and the
+`SynthComparison` object that runs the entire menu and recommends
+an estimator under pre-RMSPE / parsimony tie-breaking---a workflow
+unavailable in any single existing package.
+
 # Software Design
 
 `StatsPAI` is organised into roughly fifty modular subpackages. All
@@ -108,28 +123,117 @@ interface: `.summary()`, `.plot()`, `.to_latex()`, `.to_docx()`, and
 
 - **Classical regression and panel:** OLS/IV/panel/GLM,
   fixed-effect high-dimensional estimation.
+- **Instrumental variables (24+ estimators, v1.14 modern reporting
+  bundle):** k-class 2SLS / LIML / Fuller / GMM / JIVE; modern JIVE
+  variants (UJIVE, IJIVE, RJIVE) and many-weak-instrument inference
+  via Mikusheva--Sun jackknife AR / LM
+  [@mikusheva2022inference; @mikusheva2024weak]; post-Lasso BCH
+  [@chernozhukov2013inference]; non-parametric IV (Newey--Powell;
+  kernel IV; DeepIV; IV-DML); Bayesian IV (Chernozhukov--Hong);
+  marginal treatment effects and the Mogstad--Santos--Torgovitsky
+  partial-identification linear program
+  [@brinch2017beyond]; quantile IV (Chernozhukov--Hansen 2005/06/08);
+  shift-share / Bartik with quasi-experimental shock-level inference
+  [@borusyak2022quasi; @borusyak2025practical] and Adão--Kolesár--
+  Morales corrected SEs [@adao2019shift]; recentered shift-share
+  under non-random exposure to exogenous shocks [@borusyak2023nonrandom];
+  Conley--Hansen--Rossi plausibly-exogenous LTZ / UCI sensitivity
+  [@conley2012plausibly; @vankippersluis2018beyond]; Masten--Poirier
+  falsification-adaptive set [@masten2021salvaging]. The unified
+  reporting bundle `sp.iv.iv_diag` (new in v1.14, port of R `ivDiag`
+  [@lal2024much]) returns 2SLS analytic + pairs / wild bootstrap CIs
+  [@young2022consistency], the Olea--Pflueger effective F
+  [@olea2013robust], the Lee--McCrary--Moreira--Porter (2022) tF
+  adjusted critical value [@lee2022valid], the Anderson--Rubin /
+  Moreira CLR / Kleibergen K confidence sets
+  [@anderson1949estimation; @moreira2003conditional; @kleibergen2002pivotal],
+  Kleibergen--Paap rk LM / Wald F [@kleibergen2006generalized],
[…]
