+++ paper.bib
@@ -806,6 +806,9 @@
   title={Quasi-Experimental Shift-Share Research Designs},
   author={Borusyak, Kirill and Hull, Peter and Jaravel, Xavier},
   journal={The Review of Economic Studies},
+  volume={89},
+  number={1},
+  pages={181--213},
   year={2022},
   doi={10.1093/restud/rdab030}
 }
@@ -846,6 +849,9 @@
   title={Beyond LATE with a Discrete Instrument},
   author={Brinch, Christian N. and Mogstad, Magne and Wiswall, Matthew},
   journal={Journal of Political Economy},
+  volume={125},
+  number={4},
+  pages={985--1039},
   year={2017},
   doi={10.1086/692712}
 }
@@ -1152,9 +1158,60 @@
 @article{delgado2015difference,
   title={Difference-in-Differences Techniques for Spatial Data: Local Autocorrelation and Spatial Interaction},
   author={Delgado, Michael S. and Florax, Raymond J.G.M.},
-  journal={SSRN Electronic Journal},
+  journal={Economics Letters},
+  volume={137},
+  pages={123--126},
   year={2015},
-  doi={10.2139/ssrn.2637764}
+  doi={10.1016/j.econlet.2015.10.035}
+}
+
+@article{roth2023whats,
+  title={What's trending in difference-in-differences? A synthesis of the recent econometrics literature},
+  author={Roth, Jonathan and Sant'Anna, Pedro H. C. and Bilinski, Alyssa and Poe, John},
+  journal={Journal of Econometrics},
+  volume={235},
+  number={2},
+  pages={2218--2244},
+  year={2023},
+  doi={10.1016/j.jeconom.2023.03.008}
+}
+
+@article{wing2024designing,
+  title={Designing Difference-in-Difference Studies with Staggered Treatment Adoption: Key Concepts and Practical Guidelines},
+  author={Wing, Coady and Yozwiak, Madeline and Hollingsworth, Alex and Freedman, Seth and Simon, Kosali I.},
+  journal={Annual Review of Public Health},
+  volume={45},
+  pages={485--505},
+  year={2024},
+  doi={10.1146/annurev-publhealth-061022-050825}
+}
+
+@techreport{fiorini2024simple,
+  title={A Simple Approach to Staggered Difference-in-Differences in the Presence of Spillovers},
+  author={Fiorini, Mario and Lee, Wooyong and Pfeifer, Gregor},
+  institution={CESifo},
+  type={CESifo Working Paper},
+  number={11011},
+  year={2024},
+  url={https://hdl.handle.net/10419/296100}
+}
+
+@misc{serenini2024spatial,
+  title={Spatial Synthetic Difference-in-Differences},
+  author={Serenini, Renan and Masek, Frantisek},
+  year={2024},
+  note={SSRN working paper, last revised January 2, 2026},
+  doi={10.2139/ssrn.4736857}
+}
+
+@misc{sun2025network,
+  title={Difference-in-Differences Under Network Interference},
+  author={Sun, Kuan and Xiao, Zhiguo},
+  year={2025},
+  eprint={2509.24259},
+  archivePrefix={arXiv},
+  primaryClass={stat.ME},
+  doi={10.48550/arXiv.2509.24259}
 }
 
 @article{diamond2013genetic,
@@ -1205,10 +1262,12 @@
   doi={10.1109/ispa.2005.195385}
 }
 
-@article{dub2014spatial,
+@article{dube2014spatial,
   title={A spatial Difference-in-Differences estimator to evaluate the effect of change in public mass transit systems on house prices},
   author={Dubé, Jean and Legros, Diègo and Thériault, Marius and Des Rosiers, François},
[…]
