### 2026-05-06 b5e340f
feat(ml-causal): DML sensitivity + RATE + BLP/mediate fixes + DAG viz
Cross-cutting polish wave on the machine-learning + causal-inference
module family so the package matches the 2024-2026 reporting frontier
set by DoubleML, EconML, grf, and lmtp.

New capabilities:
- sp.dml_sensitivity / DMLSensitivityResult — Chernozhukov-Cinelli-
  Newey-Sharma-Syrgkanis (2022) "Long Story Short" DML-OVB sensitivity
  with RV_q, RV_{q,alpha}, scenario bias bounds, benchmark covariates,
  and a sensemakr-style bias-contour plot. Refs verified via NBER + arXiv.
- sp.dml_diagnostics / DMLDiagnostics — overlap
[…]
