+++ paper.bib
@@ -4694,3 +4694,52 @@
   pages={1753--1769},
   doi={10.1080/01621459.2015.1017578}
 }
+
+% =====================================================================
+% ML+Causal module — v1.15 polish (citations verified independently
+% via NBER / arXiv / journal sites).
+% =====================================================================
+
+@techreport{chernozhukov2022long,
+  title={Long Story Short: Omitted Variable Bias in Causal Machine Learning},
+  author={Chernozhukov, Victor and Cinelli, Carlos and Newey, Whitney and Sharma, Amit and Syrgkanis, Vasilis},
+  year={2022},
+  institution={National Bureau of Economic Research},
+  type={NBER Working Paper},
+  number={30302},
+  doi={10.3386/w30302},
+  note={arXiv:2112.13398}
+}
+
+@article{semenova2021debiased,
+  title={Debiased Machine Learning of Conditional Average Treatment Effects and Other Causal Functions},
+  author={Semenova, Vira and Chernozhukov, Victor},
+  journal={The Econometrics Journal},
+  volume={24},
+  number={2},
+  pages={264--289},
+  year={2021},
+  doi={10.1093/ectj/utaa027}
+}
+
+@article{yadlowsky2025evaluating,
+  title={Evaluating Treatment Prioritization Rules via Rank-Weighted Average Treatment Effects},
+  author={Yadlowsky, Steve and Fleming, Scott and Shah, Nigam and Brunskill, Emma and Wager, Stefan},
+  journal={Journal of the American Statistical Association},
+  volume={120},
+  number={549},
+  year={2025},
+  doi={10.1080/01621459.2024.2393466},
+  note={arXiv:2111.07966}
+}
+
+@article{bach2024doubleml,
+  title={DoubleML: An Object-Oriented Implementation of Double Machine Learning in R},
+  author={Bach, Philipp and Kurz, Malte S. and Chernozhukov, Victor and Spindler, Martin and Klaassen, Sven},
+  journal={Journal of Statistical Software},
+  volume={108},
+  number={3},
+  pages={1--56},
+  year={2024},
+  doi={10.18637/jss.v108.i03}
+}
