+++ tests/test_qte.py
@@ -1,0 +1,147 @@
+"""Tests for the quantile / distributional treatment effect module.
+
+Closes the audit gap of zero direct tests for `src/statspai/qte/*`. The
+goal is API + numeric-direction coverage rather than full reference
+parity (the bootstrap inference is too expensive for the CI budget).
+
+Reference numbers we check against:
+
+- **QTE constant-shift property**: when the DGP is :math:`Y = \\tau \\cdot D
+  + \\varepsilon` for any iid :math:`\\varepsilon`, every quantile
+  treatment effect equals :math:`\\tau`. Both ``method='quantile_regression'``
+  (Firpo 2007) and ``method='distribution'`` (PS reweighting) must
+  recover this.
+- **QDID parallel-trends recovery**: when the trend is shared across
+  groups, QDID(τ) at every τ should equal the post-treatment shift
+  applied to the treated group (Athey & Imbens 2006 Prop 3).
+- **Distributional TE shape**: ``distributional_te`` should return
+  treated/control CDFs that are increasing on the support grid.
+"""
+
+from __future__ import annotations
+
+import warnings
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import statspai as sp
+
+
+@pytest.fixture
+def constant_qte_data():
+    """Y = 1.5*D + N(0,1)  ⇒  every QTE(τ) = 1.5 (Firpo benchmark)."""
+    rng = np.random.default_rng(0)
+    n = 1500
+    D = rng.binomial(1, 0.5, n)
+    Y = 1.5 * D + rng.normal(size=n)
+    return pd.DataFrame({"y": Y, "d": D})
+
+
+@pytest.fixture
+def parallel_trends_data():
+    """Two groups, two periods. Treated group gets a +2 shift in post.
+    Control follows a parallel trend (+0.5 in post). QDID(τ) should
+    therefore equal +2 - +0 = +1.5 at every τ (with bootstrap noise)."""
+    rng = np.random.default_rng(1)
+    n_per_cell = 400
+    rows = []
+    for g in (0, 1):
+        for t in (0, 1):
+            base = 0.0 + 0.5 * t          # parallel trend +0.5 per period
+            shift = 1.5 if (g == 1 and t == 1) else 0.0
+            y = base + shift + rng.normal(size=n_per_cell)
+            for yi in y:
+                rows.append({"y": yi, "g": g, "t": t})
+    return pd.DataFrame(rows)
+
+
+# --------------------------------------------------------------------- #
+# qte (Firpo 2007)
+# --------------------------------------------------------------------- #
+
+
+class TestQTE:
+
+    def test_quantile_regression_recovers_constant_shift(self, constant_qte_data):
+        res = sp.qte(
+            constant_qte_data, y="y", treatment="d",
+            quantiles=[0.25, 0.5, 0.75],
+            method="quantile_regression",
+            n_boot=50, seed=0,
+        )
+        assert res.effects.shape == (3,)
+        # Every quantile effect should be close to 1.5
+        for q, eff in zip(res.quantiles, res.effects):
+            assert abs(eff - 1.5) < 0.30, f"QTE(τ={q}) = {eff:.3f} (want ≈1.5)"
+
+    def test_distribution_method(self, constant_qte_data):
+        res = sp.qte(
+            constant_qte_data, y="y", treatment="d",
+            quantiles=[0.5],
+            method="distribution",
+            n_boot=50, seed=0,
+        )
+        assert abs(res.effects[0] - 1.5) < 0.30
+
+    def test_unknown_method_raises(self, constant_qte_data):
+        with pytest.raises(ValueError, match="Unknown QTE method"):
+            sp.qte(constant_qte_data, y="y", treatment="d",
[…]

[…]
