### 2026-05-06 3d84f1b
feat(spatial): expand spatial DID results
Signed-off-by: Codex <codex@openai.com>
