+++ src/statspai/spatial/did.py
@@ -1,48 +1,68 @@
 """
 Spatial Difference-in-Differences (Spatial DiD).
 
-Generalises TWFE DiD to account for spatial spillovers, following
-Delgado & Florax (2015) and Dubé, Legros & Thériault (2014). We fit a
-TWFE outcome equation augmented with lagged-treatment terms:
+This module implements the local-spillover DiD design of Delgado and
+Florax (2015) with StatsPAI's modern result surface: diagnostics,
+event-study support, plotting, and publication-ready exports.
+
+The baseline specification is a two-way fixed-effect outcome equation
+augmented with the spatial lag of treatment:
 
 .. math::
 
-    Y_{it} = \\alpha_i + \\gamma_t + \\tau D_{it} + \\theta\\, W\\, D_{it}
-             + X_{it} \\beta + \\varepsilon_{it}.
+    Y_{it} = \\alpha_i + \\gamma_t + \\tau D_{it}
+             + \\theta W D_{it} + X_{it}\\beta + \\varepsilon_{it}.
 
-* :math:`\\tau` — the direct (own-unit) treatment effect.
-* :math:`\\theta` — the spatial-spillover coefficient on neighbours'
-  treatment share.
-
-Standard errors are Conley-style spatial-correlation robust with a
-user-supplied distance cut-off ``conley_cutoff`` — if ``None`` we
-fall back to cluster-robust SEs at the unit level.
+``direct_effect`` estimates the own-unit treatment effect, while
+``spillover_effect`` estimates the effect of nearby units' treatment
+exposure under a user-supplied spatial weights matrix.
 
 References
 ----------
 Delgado, M. S. & Florax, R. J. G. M. (2015).
 "Difference-in-differences techniques for spatial data: local
 autocorrelation and spatial interaction." *Economics Letters*, 137,
-123-126. [@delgado2015difference]
+123-126.
 
 Dubé, J., Legros, D., & Thériault, M. (2014).
 "A spatial difference-in-differences estimator to evaluate the effect
 of change in public mass transit systems on house prices."
-*Transportation Research Part B*, 64, 24-40. [@dub2014spatial]
+*Transportation Research Part B*, 64, 24-40.
+
+Roth, J., Sant'Anna, P. H. C., Bilinski, A., & Poe, J. (2023).
+"What's trending in difference-in-differences? A synthesis of the
+recent econometrics literature." *Journal of Econometrics*, 235(2),
+2218-2244.
 """
 
 from __future__ import annotations
 
+import json
 from dataclasses import dataclass, field
-from typing import Optional, Sequence, Dict, Any
+from pathlib import Path
+from typing import Any, Dict, Optional, Sequence, Tuple
 
 import numpy as np
 import pandas as pd
 from scipy import stats
 
+from ..core.results import SummaryText, _to_jsonable
+
+
+_EPS = 1e-12
+_EARTH_RADIUS_KM = 6371.0
+
 
 @dataclass
 class SpatialDiDResult:
+    """Result object for :func:`spatial_did`.
+
+    The object intentionally exposes both a spatial-DiD-specific surface
+    (``direct_effect``, ``spillover_effect``, ``total_effect``) and a
+    broom/modelsummary-compatible surface (``tidy()``, ``glance()``,
+    ``params``, ``std_errors``).
+    """
+
     direct_effect: float
     spillover_effect: float
     se_direct: float
@@ -53,35 +73,836 @@ class SpatialDiDResult:
     pvalue_spillover: float
     coefficients: pd.DataFrame
     n_obs: int
+    total_effect: float = np.nan
+    se_total: float = np.nan
+    ci_total: tuple = (np.nan, np.nan)
[…]
