### 2026-05-10 7d88672
Polish JOSS paper draft
Revise the JOSS paper metadata and manuscript for Scott Rozelle review, update cited bibliography entries, and add the generated DOCX review draft.

Co-authored-by: Codex <noreply@openai.com>
