+++ paper.bib
@@ -473,37 +473,45 @@
   doi={10.1017/CBO9781139025751}
 }
 
-@article{sharma2020dowhy,
-  title={{DoWhy}: An end-to-end library for causal inference},
+@misc{sharma2020dowhy,
+  title={{DoWhy}: An End-to-End Library for Causal Inference},
   author={Sharma, Amit and Kiciman, Emre},
-  journal={arXiv preprint arXiv:2011.04216},
   year={2020},
-  doi={10.48550/arXiv.2011.04216}
+  eprint={2011.04216},
+  archivePrefix={arXiv},
+  primaryClass={stat.ME},
+  doi={10.48550/arXiv.2011.04216},
+  url={https://arxiv.org/abs/2011.04216}
 }
 
 @misc{econml,
   title={{EconML}: A {Python} Package for {ML}-Based Heterogeneous Treatment Effects Estimation},
-  author={{Microsoft Research}},
+  author={Battocchi, Keith and Dillon, Eleanor and Hei, Maggie and Lewis, Greg and Oka, Paul and Oprescu, Miruna and Syrgkanis, Vasilis},
   year={2019},
+  note={Version 0.x},
   url={https://github.com/py-why/EconML}
 }
 
-@article{chen2020causalml,
-  title={{CausalML}: {Python} package for causal machine learning},
-  author={Chen, Huigang and Harinen, Totte and Li, Jeong-Yoon and Yung, Liang and Zhao, Zhenyu},
-  journal={arXiv preprint arXiv:2002.11631},
+@misc{chen2020causalml,
+  title={{CausalML}: {Python} Package for Causal Machine Learning},
+  author={Chen, Huigang and Harinen, Totte and Lee, Jeong-Yoon and Yung, Mike and Zhao, Zhenyu},
   year={2020},
-  doi={10.48550/arXiv.2002.11631}
+  eprint={2002.11631},
+  archivePrefix={arXiv},
+  primaryClass={cs.CY},
+  doi={10.48550/arXiv.2002.11631},
+  url={https://arxiv.org/abs/2002.11631}
 }
 
 @article{bach2022doubleml,
-  title={{DoubleML} -- An object-oriented {DoubleML} package for the {Python} ecosystem},
+  title={{DoubleML} -- An Object-Oriented Implementation of Double Machine Learning in {Python}},
   author={Bach, Philipp and Chernozhukov, Victor and Kurz, Malte S. and Spindler, Martin},
   journal={Journal of Machine Learning Research},
   volume={23},
   number={53},
   pages={1--6},
-  year={2022}
+  year={2022},
+  url={https://jmlr.org/papers/v23/21-0862.html}
 }
 
 @article{bates2015lme4,
