+++ paper.md
@@ -1,519 +1,176 @@
 ---
-title: 'StatsPAI: A Unified, Agent-Native Python Toolkit for Causal Inference and Applied Econometrics'
+title: 'StatsPAI: An Agent-Native Python Toolkit for Causal Inference and Applied Econometrics'
 tags:
   - Python
   - causal inference
   - econometrics
-  - difference-in-differences
-  - regression discontinuity
-  - synthetic control
-  - stochastic frontier
-  - mixed-effects models
-  - decomposition analysis
+  - policy evaluation
   - machine learning
-  - heterogeneous treatment effects
+  - reproducible research
 authors:
   - name: Biaoyue Wang
     orcid: 0000-0002-1828-2208
+    email: brycew6m@stanford.edu
     corresponding: true
     affiliation: "1, 2"
+  - name: Scott Rozelle
+    email: rozelle@stanford.edu
+    affiliation: "1, 2"
 affiliations:
-  - name: StatsPAI
+  - name: Rural Education Action Program, Stanford Center on China's Economy and Institutions, Stanford University, United States
     index: 1
-  - name: Stanford REAP Program, Stanford University, United States
-    index: 2
     ror: 00f54p054
-date: 1 May 2026
+  - name: StatsPAI Inc., United States
+    index: 2
+date: 9 May 2026
 bibliography: paper.bib
 ---
 
 # Summary
 
-`StatsPAI` is an open-source Python package providing a unified API for
-causal inference and applied econometrics. A single import
-(`import statspai as sp`) exposes **950+ public functions** spanning
-classical econometric models, modern ML-based causal methods, stochastic
-frontier analysis, multilevel mixed-effects models, decomposition
-analysis, and publication-ready output generation. The package
-consolidates functionality that previously required dozens of separate
-R packages (`did`, `rdrobust`, `Synth`, `sfaR`, `lme4`, `oaxaca`,
-`ddecompose`, `mlogit`, `grf`, …) or proprietary software such as
-Stata (`frontier`, `xtfrontier`, `mixed`, `meglm`, `mixlogit`, `ivqreg`,
-…) into one coherent library. Uniquely, `StatsPAI` is *agent-native*:
-every function exposes structured result objects and machine-readable
-schemas (`list_functions()`, `describe_function()`, `function_schema()`),
-making it the first econometrics toolkit purpose-built for LLM-driven
-research workflows while remaining fully ergonomic for human researchers.
+`StatsPAI` is an open-source Python package for causal inference and
+applied econometrics. It gives empirical researchers a single interface
+for estimating, diagnosing, comparing, and reporting models that are
+usually spread across many specialized packages or proprietary
+statistical environments. The package currently exposes more than 950
+registered functions across more than 80 submodules, covering classical
+regression, instrumental variables, panel data, difference-in-differences,
+regression discontinuity, synthetic control, matching,
+stochastic frontier analysis, mixed-effects models, decomposition
+methods, sensitivity analysis, and modern machine-learning estimators
+for heterogeneous treatment effects.
+
+The package is designed for policy evaluation, social science research,
+and other empirical workflows where researchers must move between
+research design, estimation, diagnostics, robustness checks, and
+publication tables. A common result contract gives users `.summary()`,
+`.plot()`, `.to_latex()`, `.to_docx()`, and `.cite()` methods where
+appropriate. `StatsPAI` is also agent-native: registered functions
+expose machine-readable schemas and structured failure metadata so that
[…]
