### 2026-05-10 #10387 @brycewang-stanford
Hello @rich2355, I have updated the author list to add Scott Rozelle (rozelle@stanford.edu) as a co-author. Scott has agreed to join the paper, will review the manuscript, and shares responsibility for the work. I remain the first and corresponding author. Please let me know if any additional authorship-change process is needed at this stage.
