### 2026-05-10 735cee7
Fix release packaging and CI validation
- fail wheel smoke tests instead of ignoring errors
- include type marker and sdist test fixtures
- escape rich HTML result display output

Co-authored-by: Codex <noreply@openai.com>
