+++ src/statspai/core/results.py
@@ -706,6 +706,14 @@ class EconometricResults:
         f_stat = self.diagnostics.get('F-statistic', None)
         f_pv = self.diagnostics.get('F p-value', self.diagnostics.get('Prob (F-statistic)', None))
 
+        def _safe(v: Any) -> str:
+            return _html_escape(str(v))
+
+        def _fmt(v: Any, spec: str = "") -> str:
+            if isinstance(v, (int, float, np.integer, np.floating)) and not pd.isna(v):
+                return format(v, spec)
+            return _safe(v)
+
         def _s(pv):
             if pd.isna(pv): return ''
             if pv < 0.01: return '<span style="color:#E74C3C;">***</span>'
@@ -714,7 +722,9 @@ class EconometricResults:
             return ''
 
         def _val(v):
-            return f'{v:.4f}' if isinstance(v, float) else str(v)
+            if isinstance(v, (int, float, np.integer, np.floating)) and not pd.isna(v):
+                return f'{v:.4f}'
+            return _safe(v)
 
         # CSS
         S = ('<style scoped>'
@@ -745,10 +755,10 @@ class EconometricResults:
         # --- Header ---
         sub_parts = []
         if dep_var:
-            sub_parts.append(f'Y = {dep_var}')
+            sub_parts.append(f'Y = {_safe(dep_var)}')
         if method:
-            sub_parts.append(method)
-        h.append(f'<div class="sp-hdr"><h3>{model_type}</h3>')
+            sub_parts.append(_safe(method))
+        h.append(f'<div class="sp-hdr"><h3>{_safe(model_type)}</h3>')
         if sub_parts:
             h.append(f'<div class="sp-sub">{" · ".join(sub_parts)}</div>')
         h.append('</div>')
@@ -756,10 +766,10 @@ class EconometricResults:
         # --- Key Metrics Bar ---
         h.append('<div class="sp-metrics">')
         if r2 is not None:
-            h.append(f'<div class="sp-metric"><div class="sp-val">{r2:.4f}</div><div class="sp-lab">R-squared</div></div>')
+            h.append(f'<div class="sp-metric"><div class="sp-val">{_fmt(r2, ".4f")}</div><div class="sp-lab">R-squared</div></div>')
         if f_stat is not None:
-            h.append(f'<div class="sp-metric"><div class="sp-val">{f_stat:.1f}</div><div class="sp-lab">F-statistic</div></div>')
-        h.append(f'<div class="sp-metric"><div class="sp-val">{n_obs:,}</div><div class="sp-lab">Observations</div></div>')
+            h.append(f'<div class="sp-metric"><div class="sp-val">{_fmt(f_stat, ".1f")}</div><div class="sp-lab">F-statistic</div></div>')
+        h.append(f'<div class="sp-metric"><div class="sp-val">{_fmt(n_obs, ",")}</div><div class="sp-lab">Observations</div></div>')
         h.append(f'<div class="sp-metric"><div class="sp-val">{len(self.params)}</div><div class="sp-lab">Parameters</div></div>')
         h.append('</div>')
 
@@ -777,7 +787,7 @@ class EconometricResults:
             lo = self.conf_int_lower.iloc[i] if isinstance(self.conf_int_lower, pd.Series) else self.conf_int_lower[i]
             hi = self.conf_int_upper.iloc[i] if isinstance(self.conf_int_upper, pd.Series) else self.conf_int_upper[i]
             pv_color = '#DC2626' if pv < 0.01 else ('#EA580C' if pv < 0.05 else ('#D97706' if pv < 0.1 else '#64748B'))
-            h.append(f'<tr><td>{var}</td>')
+            h.append(f'<tr><td>{_safe(var)}</td>')
[…]
