+++ test_results_full_suite.md
@@ -1,5 +1,11 @@
 # StatsPAI 全量测试结果报告
 
+> **维护说明（2026-05-10）**: 本文件保留为历史全量回归记录；它不是
+> 当前 `StatsPAI v1.15.1` 的发布验收凭据。v1.15.1 的最终发布前应在
+> 已安装完整依赖的环境中重新运行全量测试矩阵并覆盖本报告。
+> 本次本地修复已创建 `.venv` 并完成针对性验证，但尚未重跑下方
+> 17 批次全量矩阵。
+
 > **自动生成**: 2026-05-03 全量回归（含 ObjSense 修复后验证）
 > **测试环境**: Python 3.13.5 · StatsPAI v1.12.2 · **scipy ObjSense 冲突已修复**
 > **全仓覆盖**: 300 个测试文件，12 个批次，并行执行
