+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools>=61.0", "wheel"]
+requires = ["setuptools>=77.0.0", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
@@ -7,7 +7,8 @@ name = "StatsPAI"
 version = "1.15.1"
 description = "The Agent-Native Causal Inference & Econometrics Toolkit for Python"
 readme = "README.md"
-license = {text = "MIT"}
+license = "MIT"
+license-files = ["LICENSE"]
 authors = [
     {name = "Biaoyue Wang", email = "brycew6m@stanford.edu"}
 ]
@@ -43,6 +44,9 @@ dependencies = [
     "scipy>=1.7.0",
     "statsmodels>=0.13.0",
     "linearmodels>=4.25",
+    # Used directly by spatial regression formula parsing; declare it
+    # explicitly instead of relying on linearmodels' transitive deps.
+    "formulaic>=0.6.0",
     "numba>=0.56.0",
     "scikit-learn>=1.0.0",
     "patsy>=0.5.0",
@@ -117,6 +121,7 @@ where = ["src"]
 "" = "src"
 
 [tool.setuptools.package-data]
+"statspai" = ["py.typed", "*.pyi"]
 "statspai.datasets" = ["data/*.csv"]
 
 [tool.black]
