### 2026-05-18 28d60da
feat(plots): auto-register CJK font fallback on import — v1.15.4
`import statspai as sp` now appends a detected CJK font (PingFang SC /
Microsoft YaHei / Noto Sans CJK / SimHei / Source Han Sans / ...) to
matplotlib's `font.family` list, enabling matplotlib 3.6+ per-glyph
fallback. Chinese text in plots renders correctly without an explicit
`sp.use_chinese()` call.

Safe-by-design:
- User's primary family stays at `font.family[0]` — Latin glyphs render
  with the original primary font unchanged.
- `axes.unicode_minus` untouched — Latin minus sign stays U+2212.
- `font.sans-serif` / `font.serif
[…]
