+++ src/statspai/plots/__init__.py
@@ -11,8 +11,15 @@ Provides publication-quality academic plots:
 """
 
 from .binscatter import binscatter
-from .themes import set_theme, list_themes, use_chinese
+from .themes import set_theme, list_themes, use_chinese, _register_cjk_fallback
 from .interactive import interactive, get_code, FigureEditor
+
+# Auto-register CJK font fallback at import time so Chinese plots work out of
+# the box. Appends detected CJK fonts to font.family list (matplotlib 3.6+
+# per-glyph fallback). The user's primary family stays at index 0, so Latin
+# glyphs are unchanged; only CJK glyphs missing from the primary fall through.
+# Opt-out: STATSPAI_NO_AUTO_CJK=1. User's later rcParams[...] assignment wins.
+_register_cjk_fallback()
 
 __all__ = [
     'binscatter',
