+++ src/statspai/plots/themes.py
@@ -139,6 +139,10 @@ _PALETTES = {
 
 _original_rcparams = None
 
+# State for auto-CJK fallback registration (see _register_cjk_fallback).
+_cjk_fallback_registered = False
+_cjk_fallback_info: Dict[str, Any] = {}
+
 
 def _get_cn_serif_fonts() -> list:
     """Auto-detect the best Chinese serif fonts on this system."""
@@ -169,6 +173,162 @@ def _get_cn_serif_fonts() -> list:
         ]
     found.extend(['Times New Roman', 'DejaVu Serif'])
     return found
+
+
+def _get_cn_sans_fonts() -> list:
+    """Auto-detect the best Chinese sans-serif fonts on this system."""
+    try:
+        from matplotlib.font_manager import fontManager
+        available = {f.name for f in fontManager.ttflist}
+    except Exception:
+        available = set()
+
+    candidates = [
+        # macOS
+        'PingFang SC', 'PingFang HK', 'Hiragino Sans GB',
+        'Heiti TC', 'STHeiti', 'Hiragino Sans',
+        # Windows
+        'Microsoft YaHei', 'Microsoft JhengHei', 'SimHei',
+        # Linux (apt install fonts-noto-cjk / fonts-wqy-*)
+        'Noto Sans CJK SC', 'Noto Sans CJK TC',
+        'Noto Sans CJK JP', 'Noto Sans CJK KR',
+        'WenQuanYi Zen Hei', 'WenQuanYi Micro Hei',
+        # Adobe / Google Source Han (cross-platform, Docker/cloud images)
+        'Source Han Sans SC', 'Source Han Sans CN', 'Source Han Sans',
+    ]
+    found = [f for f in candidates if f in available]
+    if not found:
+        _cjk_sans_kws = ('CJK', 'Han Sans', 'WenQuanYi', 'YaHei', 'PingFang', 'Heiti')
+        found = [
+            n for n in sorted(available)
+            if any(kw in n for kw in _cjk_sans_kws)
+        ]
+    return found
+
+
+def _register_cjk_fallback(force: bool = False) -> Dict[str, Any]:
+    """
+    Append detected CJK fonts to matplotlib's ``font.family`` list so per-glyph
+    fallback (mpl 3.6+) handles Chinese text automatically.
+
+    **Why ``font.family`` and not ``font.sans-serif`` / ``font.serif``?**
+    Matplotlib's per-glyph fallback walks the ``font.family`` list, not the
+    family-specific lists. Empirically verified on mpl 3.10: appending a CJK
+    font to ``font.sans-serif`` does NOT trigger fallback when ``DejaVu Sans``
+    is already first there. Appending to ``font.family`` does.
+
+    Semantics:
+
+    - The user's primary family (``'sans-serif'`` by default, resolving to
+      DejaVu Sans on a fresh install) stays at index 0 → **Latin glyphs are
+      unchanged**.
+    - CJK font names are appended → matplotlib falls through per-glyph when
+      the primary font lacks the codepoint → **Chinese renders without boxes**.
+    - The user's later ``rcParams['font.family'] = 'Times New Roman'`` (or any
+      assignment) fully replaces our list → **explicit user config wins**.
+
+    Other guarantees:
+
+    - Does **NOT** touch ``axes.unicode_minus`` (matplotlib renders U+2212
+      from the Latin primary, so the minus sign stays correct on non-CJK plots).
+    - Does **NOT** touch ``font.sans-serif`` / ``font.serif`` (theme defaults
+      and user-set fallback lists for the *primary* family are preserved).
+    - Idempotent: subsequent calls without ``force`` are no-ops.
[…]
