+++ tests/test_auto_cjk_fallback.py
@@ -1,0 +1,176 @@
+"""
+Tests for auto-CJK fallback registration.
+
+Behavior contract (see ``statspai.plots.themes._register_cjk_fallback``):
+
+1. The user's primary font family stays at ``font.family[0]`` — Latin glyphs
+   are rendered with their original primary font, unchanged.
+2. CJK font names are APPENDED to ``font.family`` (not prepended, not
+   replacing) so matplotlib's per-glyph fallback (mpl 3.6+) walks them only
+   for codepoints missing from the primary font.
+3. ``axes.unicode_minus`` is NOT touched (Latin U+2212 stays correct).
+4. ``font.sans-serif`` and ``font.serif`` are NOT touched (theme-managed
+   lists are preserved as-is).
+5. Rendering Chinese text after a plain ``import statspai as sp`` (no
+   ``use_chinese()`` call) must produce zero "Glyph missing from font"
+   warnings — provided a CJK font is installed.
+6. User's explicit ``rcParams['font.family'] = ...`` after import must
+   fully replace our list.
+7. The ``STATSPAI_NO_AUTO_CJK=1`` env var must disable registration.
+"""
+
+from __future__ import annotations
+
+import warnings
+
+import matplotlib
+matplotlib.use("Agg")
+import matplotlib as mpl
+import matplotlib.pyplot as plt
+import pytest
+
+from statspai.plots import themes
+
+
+def _has_cjk_font() -> bool:
+    """Whether the test machine has any CJK font installed."""
+    return bool(themes._get_cn_sans_fonts() or [
+        f for f in themes._get_cn_serif_fonts()
+        if f not in ('Times New Roman', 'DejaVu Serif')
+    ])
+
+
+def _glyph_warnings(records):
+    return [
+        str(r.message) for r in records
+        if 'Glyph' in str(r.message) or 'missing from font' in str(r.message)
+    ]
+
+
+@pytest.fixture(autouse=True)
+def _restore_rcparams():
+    """Snapshot rcParams around every test so user-state mutations don't leak."""
+    saved = mpl.rcParams.copy()
+    yield
+    mpl.rcParams.update(saved)
+
+
+def test_primary_family_stays_first():
+    """User's primary family stays at font.family[0]. Latin glyphs render
+    with the primary font unchanged."""
+    mpl.rcParams['font.family'] = ['sans-serif']  # matplotlib default
+
+    themes._register_cjk_fallback(force=True)
+
+    family = list(mpl.rcParams['font.family'])
+    assert family[0] == 'sans-serif', (
+        f"primary family must remain at index 0; got: {family}"
+    )
+
+
+def test_import_does_not_change_unicode_minus():
+    """We must NOT flip axes.unicode_minus globally; that degrades Latin plots."""
+    mpl.rcParams['axes.unicode_minus'] = True  # matplotlib default
+    themes._register_cjk_fallback(force=True)
+    assert mpl.rcParams['axes.unicode_minus'] is True, (
+        "axes.unicode_minus must not be touched by auto-CJK registration"
+    )
+
+
+def test_does_not_touch_family_specific_lists():
+    """font.sans-serif and font.serif (theme-managed primary-family fallbacks)
+    must be left alone by auto-registration."""
+    mpl.rcParams['font.sans-serif'] = ['DejaVu Sans']
+    mpl.rcParams['font.serif'] = ['DejaVu Serif']
+
+    themes._register_cjk_fallback(force=True)
+
+    assert list(mpl.rcParams['font.sans-serif']) == ['DejaVu Sans']
+    assert list(mpl.rcParams['font.serif']) == ['DejaVu Serif']
+
+
+@pytest.mark.skipif(not _has_cjk_font(), reason="no CJK font on this system")
[…]
