### 2026-05-24 4b8ebe0
fix(paper.bib): unescape literal @article{} in comment block
JOSS editorialbot (and strict BibTeX parsers like pybtex) reported
"Failed to parse BibTeX entry: cite-key missing" on the StatsPAI review
issue #10604.

Root cause: a `%`-prefixed note line in paper.bib (around the
vanderlaan2023efficient entry) contained the literal pattern
`@article{vanderlaan2023efficient}` inside backticks. BibTeX has no
line-comment marker — `%` is only an out-of-entry filler convention —
so strict parsers see the literal at-sign + open-brace as a real entry
start with a malformed body. Editorialbot stops at "c
[…]
