### 2026-05-24 7d88940
fix(audit): close two Citation Audit coverage holes
Two regressions surfaced when the Citation Audit GitHub workflow turned
red on the v1.15.6 release commit (run 26375546484, exit 1: "476 ok /
1 mismatch / 1 unresolved").

Root causes — both in tools/audit_citations.py, both pre-existing
(workflow has been red since at least v1.15.5):

1. DOI regex captured markdown autolink closers.
   ``docs/joss_validation_dossier.md:9`` writes the Zenodo concept DOI
   as ``<https://doi.org/10.5281/zenodo.19933900>``. The DOI character
   class ``[^\s()"'\`,;}\]\[]+`` did NOT exclude ``<`` / ``>``, so the
[…]
