### 2026-05-25 #10604 @brycewang-stanford
Thanks @rich2355 and welcome @Carol-seven, @FATelarico, @MartinSpindler.

Quick note on the `Failed to parse BibTeX entry: cite-key missing` warning above: that was caused by a `%`-prefixed note line in `paper.bib` that contained the literal `@article{...}` token inside backticks. BibTeX has no line-comment marker, so strict parsers (including the one editorialbot uses) treated that token as a real entry start. Fixed in [`4b8ebe0`](https://github.com/brycewang-stanford/StatsPAI/commit/4b8ebe0) — verified locally with pybtex (551 entries, 0 empty keys) and `pandoc --citeproc` (exit 0).

Regenerating the proof so the warning clears:
