### 2026-05-25 e34d471
docs(paper.bib): verify and update 2 of 5 editorialbot DOI suggestions
JOSS editorialbot review at openjournals/joss-reviews#10604 flagged 5 entries
as potentially having DOIs. After independent verification (doi.org resolver
+ Crossref REST API + Crossref reverse title+author search + OpenAlex index):

Updated (verified across 4 independent sources each):
- abadievives2022synthetic: @misc arXiv 2022 -> @incollection Cambridge book
  chapter "Synthetic Controls in Action" in *Advances in Economics and
  Econometrics* (2025), pages 195-224, DOI 10.1017/9781009589727.010.
  arXiv eprint 2203.062
[…]
