+++ paper.bib
@@ -2717,11 +2717,14 @@
   doi={10.1080/01621459.2021.2002600}
 }
 
-@misc{abadievives2022synthetic,
+@incollection{abadievives2022synthetic,
   title={Synthetic Controls in Action},
   author={Abadie, Alberto and Vives-i-Bastida, Jaume},
-  year={2022},
-  howpublished={arXiv preprint arXiv:2203.06279},
+  booktitle={Advances in Economics and Econometrics},
+  publisher={Cambridge University Press},
+  pages={195--224},
+  year={2025},
+  doi={10.1017/9781009589727.010},
   eprint={2203.06279},
   archivePrefix={arXiv},
   primaryClass={econ.EM}
@@ -4179,11 +4182,14 @@
 }
 
 @article{webb2014reworking,
-  title={Reworking Wild Bootstrap Based Inference for Clustered Errors},
+  title={Reworking wild bootstrap-based inference for clustered errors},
   author={Webb, Matthew D.},
-  journal={Queen's Economics Department Working Paper},
-  number={1315},
-  year={2014}
+  journal={Canadian Journal of Economics/Revue canadienne d'\'economique},
+  volume={56},
+  number={3},
+  pages={839--858},
+  year={2023},
+  doi={10.1111/caje.12661}
 }
 
 @inproceedings{romano2019conformalized,
