### 2026-05-25 f522a5c
docs(paper.bib): resolve last 3 editorialbot DOI flags
Final pass on the 3 remaining MISSING-DOI entries from JOSS issue #10604.
Each verified across doi.org resolver + Crossref REST + Crossref reverse
title-search + OpenAlex independent index.

Updated:
- schwab2013ltmle -> lendle2017ltmle: upgrade from CRAN package metadata DOI
  (10.32614/cran.package.ltmle, type=dataset) to the peer-reviewed software
  paper in Journal of Statistical Software 81(1), 2017, DOI 10.18637/jss.v081.i01.
  Same four authors; Lendle is first in the published version, hence the
  bibkey rename. Also updates the do
[…]
