+++ paper.bib
@@ -2370,12 +2370,14 @@
   doi={10.32614/cran.package.etwfe}
 }
 
-@article{schwab2013ltmle,
-  title={ltmle: Longitudinal Targeted Maximum Likelihood Estimation},
-  author={Schwab, Joshua and Lendle, Samuel and Petersen, Maya and van der Laan, Mark},
-  journal={CRAN: Contributed Packages},
-  year={2013},
-  doi={10.32614/cran.package.ltmle}
+@article{lendle2017ltmle,
+  title={ltmle: An {R} Package Implementing Targeted Minimum Loss-Based Estimation for Longitudinal Data},
+  author={Lendle, Samuel D. and Schwab, Joshua and Petersen, Maya L. and van der Laan, Mark J.},
+  journal={Journal of Statistical Software},
+  volume={81},
+  number={1},
+  year={2017},
+  doi={10.18637/jss.v081.i01}
 }
 
 @article{wang2019minimal,
@@ -2395,7 +2397,7 @@
 }
 
 @article{ado2019shift,
-  title={Shift-Share Designs: Theory and Inference*},
+  title={Shift-Share Designs: Theory and Inference},
   author={Adão, Rodrigo and Kolesár, Michal and Morales, Eduardo},
   journal={The Quarterly Journal of Economics},
   year={2019},
