### 2026-05-28 adec3d8
fix(ci/parity-guards): unblock registry-drift + soften citation-audit
Self-review of 200c9e6 surfaced three concrete defects:

1. ``registry-drift`` installed ``pip install -e .`` (no ``[dev]``) but
   then invoked ``pytest``. Two failures stacked: pyproject's pytest
   ``addopts`` injects ``--cov`` unconditionally (needs ``pytest-cov``),
   and ``pytest`` itself is only in ``[dev]``. Switched to
   ``pip install -e ".[dev]"``.

2. ``citation-audit`` ran with ``--strict``, which exits non-zero on
   unresolved citations as well as mismatches. ``--strict`` conflates
   "fabricated citation" (§1
[…]
