### 2026-05-28 d092bf6
test(did/parity): drop stale BJS xfail markers — both tests pass
commit 3607a2f ("fix(did): repair BJS imputation support") fixed the
~24% drift between sp.did_imputation and R didimputation that v1.11
had flagged. Both ``test_bjs_did_imputation_matches_R`` and
``test_bjs_did_imputation_close_to_truth`` have been quietly XPASS-ing
since then — the stale ``@pytest.mark.xfail(strict=False)`` markers
hid the green status as a soft warning instead.

Removed both markers and the surrounding note block; the comment now
just records that the v1.11 divergence was fixed in 3607a2f. A future
regression 
[…]
