### 2026-05-28 #10604 @FATelarico
> # Review checklist for FATelarico
I am currently re-running the tests, I will update my review later. For now I want mostly to notice the author's hard work on the comparison with Stata, some R packages, DoWhy, and the usual ML suspects (`EconML`, `CausalML`, and `DoubleML`).

Also, the relevance case is reasonably strong since the paper positions the software as dealing with the fragmentation in causal inference and applied econometrics using agents.

The weakest point remains independent scholarly adoption, but there is not much that can be done at this point **and** a JOSS paper may help in that sense. Also, development history is on the shorter side and the author is essentially the only contributor, although I need still to look at the forks on the repo and whether they are all or mostly Claude/Codex.
