### 2026-05-29 4a2d459
fix(xtabond): block-diagonal Arellano-Bond instruments + (Z'HZ)^-1 weight (#12)
⚠️ correctness fix. sp.xtabond difference GMM was off vs Stata by 48 % on
β_{y₋₁} and 80 % on its SE. Two root causes:

1. Instruments were a flat, fixed set of lagged-level columns
   (gmm_lags=(2,5)) followed by dropping every row missing any of them —
   which on a short panel discards most of the sample. Replaced with the
   standard block-diagonal GMM instrument matrix: every available deeper
   lag Y_{i,s} (s ≤ t-2) is a period-specific moment, missing lags are
   zero-filled, no rows dropped.
2. One-step wei
[…]
