+++ src/statspai/gmm/arellano_bond.py
@@ -8,6 +8,15 @@ Estimates dynamic panel models of the form:
 using GMM with lagged levels (Arellano-Bond) or lagged levels and
 differences (Blundell-Bond system GMM) as instruments.
 
+The Arellano-Bond first-differenced estimator removes the fixed effect
+α_i by first-differencing and then exploits the **block-diagonal**
+GMM moment conditions E[Y_{i,s} ΔU_{it}] = 0 for s ≤ t-2: every lagged
+level that is available for period t is a separate instrument for that
+period's differenced equation. The one-step weight matrix is
+(Σ_i Z_i' H Z_i)⁻¹ where H encodes the MA(1) structure of the
+first-differenced i.i.d. errors (2 on the diagonal, -1 on the first
+off-diagonals). This matches Stata's ``xtabond`` and ``xtdpd``.
+
 References
 ----------
 Arellano, M. and Bond, S. (1991).
@@ -31,7 +40,7 @@ Two-Step GMM Estimators."
 *Journal of Econometrics*, 126(1), 25-51. [@windmeijer2005finite]
 """
 
-from typing import Optional, List, Dict, Any, Tuple
+from typing import Optional, List, Tuple
 
 import numpy as np
 import pandas as pd
@@ -47,7 +56,7 @@ def xtabond(
     id: str = 'id',
     time: str = 'time',
     lags: int = 1,
-    gmm_lags: Tuple[int, int] = (2, 5),
+    gmm_lags: Tuple[int, Optional[int]] = (2, None),
     method: str = 'difference',
     twostep: bool = False,
     robust: bool = True,
@@ -65,35 +74,43 @@ def xtabond(
     y : str
         Dependent variable.
     x : list of str, optional
-        Exogenous regressors (strictly exogenous or predetermined).
+        Strictly exogenous regressors. Entered in first differences both
+        as regressors and as their own (standard) instruments.
     id : str, default 'id'
         Unit identifier.
     time : str, default 'time'
         Time period variable.
     lags : int, default 1
         Number of lags of Y to include (ρ₁ Y_{t-1} + ... + ρ_p Y_{t-p}).
-    gmm_lags : tuple (min, max), default (2, 5)
-        Range of lags used as GMM instruments.
-        E.g., (2, 5) means Y_{t-2}, ..., Y_{t-5} are instruments.
+    gmm_lags : tuple (min, max), default (2, None)
+        Range of lags of Y (in levels) used as GMM instruments. ``min``
+        must be ≥ 2 (deeper lags are orthogonal to the differenced error).
+        ``max=None`` uses **all** available deeper lags, matching Stata's
+        ``xtabond`` default. Setting ``max`` caps the instrument count
+        (Stata's ``maxldep()`` / collapse-style trimming).
     method : str, default 'difference'
-        ``'difference'`` — Arellano-Bond (first-differenced GMM).
-        ``'system'`` — Blundell-Bond style: uses longer lag structure
-        for better efficiency with persistent series. (Note: full
-        system GMM with stacked level equations is planned for a
-        future release.)
+        ``'difference'`` — Arellano-Bond (first-differenced GMM). This is
+        the validated path (machine-precision parity with Stata's
+        ``xtabond``). ``'system'`` (Blundell-Bond) currently raises
+        ``NotImplementedError``: proper system GMM requires a stacked level
+        equation and its own Stata parity reference, which is planned for a
+        future release.
     twostep : bool, default False
[…]
