+++ tests/r_parity/compare.py
@@ -35,10 +35,9 @@ from typing import Any
 HERE = Path(__file__).resolve().parent
 RESULTS_DIR = HERE / "results"
 # The sister Stata-side harness lives at tests/stata_parity/. Stata
-# results are only emitted for the 21 modules with a canonical Stata
-# reference (see tests/stata_parity/README.md); the modules without
-# a canonical Stata implementation (DML, causal forest, augsynth,
-# gsynth) are flagged with explicit reasons in the 3-way table.
+# results are emitted for modules with a canonical Stata reference;
+# modules without an authoritative Stata implementation are flagged
+# with explicit reasons in the 3-way table.
 STATA_RESULTS_DIR = HERE.parent / "stata_parity" / "results"
 STATA_SKIP_REASON: dict[str, str] = {
     "08_dml":           "no canonical ref",
@@ -91,21 +90,21 @@ TOLERANCES: dict[str, dict[str, float]] = {
     "35_panel":       {"rel_est": 1e-3, "rel_se": 1e-3},  # FE/RE only headline
     "36_mediation":   {"rel_est": 1e-2, "rel_se": 1e-2},
     # Modules added in the 2026-05-28 parity expansion session.
-    "37_ppmlhdfe":    {"rel_est": 1e-3, "rel_se": 0.6},   # sp HC1 50% inflated (finding #6)
+    "37_ppmlhdfe":    {"rel_est": 1e-3, "rel_se": 1e-2},  # post FE-score fix
     "38_drdid":       {"rel_est": 1e-3, "rel_se": 1e-1},  # bootstrap vs IF SE
     "39_arima":       {"rel_est": 1e-2, "rel_se": 1e-2},  # method gap (Kalman vs CSS-ML)
     "40_qreg":        {"rel_est": 1e-3, "rel_se": 1e-1},  # Powell SE method choice
-    "41_tobit":       {"rel_est": 1e-3, "rel_se": 3.5e-1}, # sp BFGS hess_inv (finding #9)
+    "41_tobit":       {"rel_est": 1e-3, "rel_se": 1e-3},  # observed-info Hessian
     "42_nbreg":       {"rel_est": 1e-3, "rel_se": 1e-2},
     "43_heckman":     {"rel_est": 1e-3, "rel_se": 1e-3},
-    "44_mlogit":      {"rel_est": 1e-3, "rel_se": 5e-2},  # sp SE ~2% inflated (finding #10)
-    "45_ologit":      {"rel_est": 1e-3, "rel_se": 3e-1},  # sp beta_x SE 26% off (finding #11)
+    "44_mlogit":      {"rel_est": 1e-3, "rel_se": 1e-3},  # observed-info Hessian
+    "45_ologit":      {"rel_est": 1e-3, "rel_se": 1e-3},  # observed-info Hessian
     "46_clogit":      {"rel_est": 1e-3, "rel_se": 1e-3},
     # Modules added in the second 2026-05-28 fix-and-extend pass.
     "47_ppmlhdfe_3fe":{"rel_est": 1e-3, "rel_se": 5e-2},  # post Gauss-Seidel
     "48_probit":      {"rel_est": 1e-3, "rel_se": 1e-2},
     "49_oprobit":     {"rel_est": 1e-3, "rel_se": 1e-3},
-    "50_xtabond":     {"rel_est": 0.5, "rel_se": 1.0},   # finding #12 (sp impl gap)
+    "50_xtabond":     {"rel_est": 1e-6, "rel_se": 1e-6},  # Py-Stata migration fixture
     "51_newey":       {"rel_est": 1e-3, "rel_se": 1e-2},  # post HAC fix
 }
 
@@ -200,7 +199,7 @@ def render_md(modules: list[str]) -> str:
     lines: list[str] = [
[…]
