+++ tests/r_parity/results/parity_table.tex
@@ -1,7 +1,7 @@
 % AUTO-GENERATED by tests/r_parity/compare.py
 % Re-run after any module change to refresh.
 \begin{longtable}{p{0.10\linewidth}p{0.27\linewidth}p{0.40\linewidth}p{0.16\linewidth}}
-\caption{Track A parity headline at \statspai{} 1.13.1 vs the canonical \proglang{R} reference on the calibrated replicas. The ``Worst diff'' column reports the worst residual gap across the module's headline rows (point estimates only; per-row SE diffs and documented gap rows are reported in the Markdown source). Verdicts use PASS and GAP; common-specification passes, small-sample SE conventions, and convention gaps are explained in the parenthetical notes and per-module \code{extra} block in \code{tests/r\_parity/results/}.}
+\caption{Track A parity headline at \statspai{} 1.15.6 vs the canonical \proglang{R} reference on the calibrated replicas. The ``Worst diff'' column reports the worst residual gap across the module's headline rows (point estimates only; per-row SE diffs and documented gap rows are reported in the Markdown source). Verdicts use PASS and GAP; common-specification passes, small-sample SE conventions, and convention gaps are explained in the parenthetical notes and per-module \code{extra} block in \code{tests/r\_parity/results/}.}
 \label{tab:track-a-parity}\\
 \toprule
 Module & Method & Worst headline diff & Verdict \\
@@ -50,18 +50,18 @@ Module & Method & Worst headline diff & Verdict \\
 \code{34\_lp} & Local projections & abs $\le 0.132$ {\footnotesize (Cholesky-orthogonalised vs OLS shock)} & \textit{GAP} \\
 \code{35\_panel} & Panel FE/RE + Hausman & rel $\le 1.1e-15$ {\footnotesize (FE/RE coefs match; Hausman variance differs)} & \textbf{PASS} \\
 \code{36\_mediation} & Causal mediation (IKT) & rel $\le 6.7e-15$ & \textbf{PASS} \\
-\code{37\_ppmlhdfe} & PPML + HDFE & rel $\le 4.9e-13$ {\footnotesize (sp HC1 SE 50\% larger than fixest/Stata; betas match 1e-9)} & \textbf{PASS} \\
+\code{37\_ppmlhdfe} & PPML + HDFE & rel $\le 4.9e-13$ {\footnotesize (post PPML FE-score fix; robust SE within 0.5\% of fixest)} & \textbf{PASS} \\
 \code{38\_drdid} & DR-DID (SZ 2020) & rel $\le 0.0005$ {\footnotesize (R-only (Stata drdid is Ferman's different DR formula))} & \textbf{PASS} \\
 \code{39\_arima} & ARIMA(2,0,0) & rel $\le 0.0053$ {\footnotesize (sp uses Kalman exact MLE; R/Stata use CSS-ML (0.5\% method gap))} & \textbf{PASS} \\
-\code{40\_qreg} & Quantile reg (median) & rel $\le 3.3e-15$ {\footnotesize (sp.qreg SE bug (extra n in sandwich) fixed in-session)} & \textbf{PASS} \\
-\code{41\_tobit} & Tobit (left-censored) & rel $\le 2.8e-08$ {\footnotesize (sp uses BFGS hess_inv approx; SE 13-30\% off)} & \textbf{PASS} \\
[…]
