+++ tests/r_parity/results/parity_table_3way.md
@@ -563,9 +563,9 @@ Generated by `tests/r_parity/compare.py` on the `results/<module>_{py,R}.json` a
 
 | stat | py est | R est | Stata est | rel py-R | rel py-Stata | py SE | R SE | Stata SE | rel SE py-R | rel SE py-Stata |
 |---|---:|---:|---:|---:|---:|---:|---:|---:|---:|---:|
-| `ar1` | 0.701891 | 0.698157 | 0.698216 | 0.00535 | 0.00526 | — | — | — | — | — |
-| `ar2` | -0.342893 | -0.341646 | -0.341644 | 0.00365 | 0.00366 | — | — | — | — | — |
-| `sigma2` | 0.414392 | 0.414984 | 0.410843 | 0.00143 | 0.00864 | — | — | — | — | — |
+| `ar1` | 0.701891 | 0.698157 | 0.698216 | 0.00535 | 0.00526 | 0.0478454 | — | — | — | — |
+| `ar2` | -0.342893 | -0.341646 | -0.341644 | 0.00365 | 0.00366 | 0.0537274 | — | — | — | — |
+| `sigma2` | 0.414392 | 0.414984 | 0.410843 | 0.00143 | 0.00864 | 0.0335345 | — | — | — | — |
 | `logLik` | -291.583141 | -292.528901 | -292.528902 | 0.00323 | 0.00323 | — | — | — | — | — |
 
 ## Module 40_qreg
