### 2026-05-29 291e783
fix(ci): make mypy/flake8 quality gates environment-stable
Both baseline gates were green locally but red on main in CI, for purely
environmental reasons — not a code regression:

* mypy: the count gate runs in CI's lean `.[dev]` install where optional
  heavy deps (pymc/torch/jax/sklearn/statsmodels/pandas/scipy) are absent,
  so every lazy import raised import-not-found / import-untyped and the
  error COUNT ballooned to ~4198 vs ~3497 in a dev venv where the deps are
  present (>3521 baseline -> fail). Adding `ignore_missing_imports = true`
  makes the count depend only on our own typed cod
[…]
