+++ .github/workflows/ci-cd.yml
@@ -80,9 +80,19 @@ jobs:
         python -m pip install --upgrade pip
         pip install -e ".[dev]"
 
-    - name: Lint with flake8
+    - name: Lint with flake8 (hard syntax errors — all matrix entries)
+      # E9/F63/F7/F82 are real syntax / undefined-name breakage and must be
+      # zero on every OS × Python combo.
       run: |
         flake8 src/ tests/ --count --select=E9,F63,F7,F82 --show-source --statistics
+
+    - name: flake8 baseline gate (single canonical env)
+      # The full violation COUNT drifts by Python version (3.12's PEP-701
+      # f-string tokenisation makes pyflakes report more on 3.12/3.13), so the
+      # ``scripts/quality_gate.py`` count ratchet is pinned to one canonical
+      # interpreter — same pattern as the mypy gate below — to stay stable.
+      if: matrix.os == 'ubuntu-latest' && matrix.python-version == '3.10'
+      run: |
         python scripts/quality_gate.py flake8
 
     - name: Type baseline with mypy
