### 2026-05-29 932d77f
chore(parity): land xtabond/arima parity-table refresh from PR #11
The substantive Arellano-Bond fix (block-diagonal instruments + (Z'HZ)^-1
weight + Windmeijer SE) already reached main via the xtabond branch merge.
This lands the only remaining net content from PR #11 — regenerated Track A
parity artefacts:

* tests/r_parity/compare.py: fix the report's self-reference path
  (Paper-JSS/replication/parity/ -> tests/r_parity/), tighten TOLERANCES to
  match the now-fixed estimators (37_ppmlhdfe, 41_tobit, 44_mlogit, 45_ologit,
  51_newey) and the 50_xtabond Py-Stata migration fixture (1e-6), re
[…]
