+++ tests/r_parity/results/parity_table.md
@@ -1,6 +1,6 @@
 # Track A parity report
 
-Generated by `Paper-JSS/replication/parity/compare.py` on the `results/<module>_{py,R}.json` artefacts. Tolerance budget per module is pre-registered in `compare.py::TOLERANCES`. Documented convention gaps, common-specification passes, and small-sample SE conventions (HDFE 1-df, RD default-bandwidth selector, SCM non-uniqueness, DML fold-split noise) are flagged in the per-module `extra` block of the JSON.
+Generated by `tests/r_parity/compare.py` on the `results/<module>_{py,R}.json` artefacts. Tolerance budget per module is pre-registered in `compare.py::TOLERANCES`. Documented convention gaps, common-specification passes, and small-sample SE conventions (HDFE 1-df, RD default-bandwidth selector, SCM non-uniqueness, DML fold-split noise) are flagged in the per-module `extra` block of the JSON.
 
 ## Module 01_ols
 - **formula**: `lwage ~ educ + exper + expersq + black + south + smsa`
@@ -525,9 +525,9 @@ Generated by `Paper-JSS/replication/parity/compare.py` on the `results/<module>_
 
 | stat | py est | R est | abs Δ | rel Δ | py SE | R SE | abs Δ SE | rel Δ SE |
 |---|---:|---:|---:|---:|---:|---:|---:|---:|
-| `ar1` | 0.701891 | 0.698157 | 0.00373 | 0.00535 | — | — | — | — |
-| `ar2` | -0.342893 | -0.341646 | 0.00125 | 0.00365 | — | — | — | — |
-| `sigma2` | 0.414392 | 0.414984 | 0.000593 | 0.00143 | — | — | — | — |
+| `ar1` | 0.701891 | 0.698157 | 0.00373 | 0.00535 | 0.0478454 | — | — | — |
+| `ar2` | -0.342893 | -0.341646 | 0.00125 | 0.00365 | 0.0537274 | — | — | — |
+| `sigma2` | 0.414392 | 0.414984 | 0.000593 | 0.00143 | 0.0335345 | — | — | — |
 | `logLik` | -291.583141 | -292.528901 | 0.946 | 0.00323 | — | — | — | — |
 
 ## Module 40_qreg
