+++ tests/r_parity/results/parity_table_3way.tex
@@ -4,7 +4,7 @@
 \small
 \setlength{\tabcolsep}{2pt}
 \begin{longtable}{@{}p{0.055\linewidth}p{0.205\linewidth}p{0.30\linewidth}p{0.30\linewidth}p{0.10\linewidth}@{}}
-\caption{Track A parity headline at \statspai{} 1.13.1 against the canonical \proglang{R} reference \emph{and} (where one exists) the canonical \proglang{Stata} reference, on the calibrated replicas. The ID column is the two-digit module prefix; the two diff columns report the worst residual gap across each module's headline rows (point estimates only; per-row SE diffs and documented gap rows are reported in \code{tests/r\_parity/results/parity\_table\_3way.md}). ``no canonical ref'' marks modules whose \proglang{Stata} ecosystem has no authoritative port we can compare to without fabricating one. Verdicts use PASS and GAP; common-specification passes, small-sample SE conventions, and convention gaps are explained in the parenthetical notes and per-module \code{extra} block in \code{tests/r\_parity/results/} and \code{tests/stata\_parity/results/}.}
+\caption{Track A parity headline at \statspai{} 1.15.6 against the canonical \proglang{R} reference \emph{and} (where one exists) the canonical \proglang{Stata} reference, on the calibrated replicas. The ID column is the two-digit module prefix; the two diff columns report the worst residual gap across each module's headline rows (point estimates only; per-row SE diffs and documented gap rows are reported in \code{tests/r\_parity/results/parity\_table\_3way.md}). ``no canonical ref'' marks modules whose \proglang{Stata} ecosystem has no authoritative port we can compare to without fabricating one. Verdicts use PASS and GAP; common-specification passes, small-sample SE conventions, and convention gaps are explained in the parenthetical notes and per-module \code{extra} block in \code{tests/r\_parity/results/} and \code{tests/stata\_parity/results/}.}
 \label{tab:track-a-parity}\\
 \toprule
 ID & Method & Worst diff vs \proglang{R} & Worst diff vs \proglang{Stata} & Verdict \\
@@ -53,19 +53,19 @@ ID & Method & Worst diff vs \proglang{R} & Worst diff vs \proglang{Stata} & Verd
 \code{34} & Local projections & abs $\le 0.132$ {\footnotesize (Cholesky-orthogonalised vs OLS shock)} & abs $\le 2.22e-16$ & \textit{GAP} \\
 \code{35} & Panel FE/RE + Hausman & rel $\le 1.1e-15$ {\footnotesize (FE/RE coefs match; Hausman variance differs)} & rel $\le 1.5e-15$ & \textbf{PASS} \\
 \code{36} & Causal mediation (IKT) & rel $\le 6.7e-15$ & rel $\le 3.6e-15$ & \textbf{PASS} \\
-\code{37} & PPML + HDFE & rel $\le 4.9e-13$ {\footnotesize (sp HC1 SE 50\% larger than fixest/Stata; betas match 1e-9)} & rel $\le 2.2e-15$ & \textbf{PASS} \\
[…]
