### 2026-05-29 7ce5093
fix(bootstrap): route 14 fabricated-SE-floor sites through core.bootstrap_se
Replaces the silent 'float(np.nanstd(boot, ddof=1)) or 1e-6' antipattern in
bridge/{dr_calib,ewm_cate,surrogate_pci}, bunching/general, did/{cohort_anchored,
did_bcf,misclassified}, iv/continuous_late, proximal/{bidirectional,fortified,mtp}.
Healthy runs are numerically unchanged; failed bootstrap replicates now warn
with a failure count instead of being swallowed, and SE is NaN below a success
floor rather than a fabricated tiny value (CLAUDE.md §7).

Co-Authored-By: Claude Opus 4.8 (1M context) <noreply@anthropic.co
[…]
