### 2026-05-29 d1f655c
perf(plots): defer matplotlib import via one-shot pyplot CJK hook
_register_cjk_fallback() at plots-import time was the only thing forcing an
eager matplotlib import (+font scan) on every 'import statspai', even for the
majority of agent calls that never plot. Now the CJK fallback registers via a
one-shot meta_path hook the first time matplotlib.pyplot is imported anywhere,
so import statspai stays matplotlib-free while Chinese plots still work out of
the box for every render path (statspai.plots, rd/did/synth). Zero behavior
change when plotting; self-removing hook, opt-out STATSPAI_NO_AUTO_C
[…]
