### 2026-05-29 1464b03
feat(output): sp.coefplot_tikz() — pgfplots/TikZ coefficient forest plot
The LaTeX-native counterpart to sp.coefplot (which returns a Matplotlib
(fig, ax) — PNG/PDF come free via fig.savefig). Emits an editable pgfplots
tikzpicture: one \addplot series per model with horizontal CI error bars,
variables down a reversed y-axis, dashed zero reference line. Supports
coef_labels, level, title/xlabel, and standalone=True (compilable document).
Reuses the same _extract_coefs the matplotlib coefplot uses, so the two
backends agree. Auto-registered (discoverable via sp.describe_function).

15 tests cov
[…]
