### 2026-05-29 7fe55e5
refactor(core): extract shared safe_logit_fit/predict primitive (CLAUDE.md §4)
front_door and principal_strat carried byte-identical statsmodels-logit
wrappers (fit + predict). Consolidate into core/_glm_fit.py; the module-level
_logit_fit/_logit_safe/_logit_predict names stay as thin delegators (mockable).
Behavior-preserving (28 tests incl. A1 fallback tests pass). The custom-IRLS /
sklearn logits in censoring/gformula/longitudinal/ivmte_lp/tmle are
deliberately different solvers and are left for a parity-quiet unification.

Co-Authored-By: Claude Opus 4.8 (1M context) <noreply@anthropic.com
[…]
