### 2026-05-29 debbf1e
refactor(inference): route multiway_cluster sandwich to core.sandwich_vcov
_one_way_sandwich delegates to core._vcov.sandwich_vcov with correction='cr1'
(df_adjust) or 'none'. Byte-identical to the prior CR1 sandwich for G>=2 across
df_adjust x n_params; multiway/twoway tests pass.

Co-Authored-By: Claude Opus 4.8 (1M context) <noreply@anthropic.com>
