### 2026-05-29 b2b0e43
refactor(regression): route zeroinflated robust/cluster SE to core.sandwich_vcov
_robust_se -> sandwich_vcov(correction='none'); _cluster_se ->
sandwich_vcov(clusters=, correction='stacked') (= G/(G-1)*n/(n-k)). MLE
inverse-Hessian bread. Byte-identical for G>=2; zinb/zip tests pass.

Co-Authored-By: Claude Opus 4.8 (1M context) <noreply@anthropic.com>
