### 2026-05-29 99294e5
test(r-parity): lock R reference environment + prove golden values reproduce
JSS reproducibility hardening (硬伤1 of the parity-readiness pass). The
committed R-side parity golden JSONs were not version-stamped, so a reviewer
could not tell which fixest/did/rdrobust versions produced them — the most
likely JSS reject-and-resubmit trigger. This closes that hole and proves the
golden values are genuine R output, not hand-edited constants:

- _common.R: capture R version + platform + every attached/loaded package
  version into a new top-level `provenance` block on every _R.json
  (non-invasive: co
[…]
