+++ tests/r_parity/results/parity_table.tex
@@ -19,7 +19,7 @@ Module & Method & Worst headline diff & Verdict \\
 \code{03\_hdfe} & HDFE 2-way FE & rel $\le 6.5e-15$ {\footnotesize (SE differs by 1-df convention)} & \textbf{PASS} \\
 \code{04\_csdid} & CS-DiD simple ATT & rel $\le 1.3e-15$ {\footnotesize (SE within 1\% analytic tolerance)} & \textbf{PASS} \\
 \code{05\_sunab} & Sun--Abraham event study & rel $\le 2.8e-11$ {\footnotesize (aggregation rule differs)} & \textbf{PASS} \\
-\code{06\_rd} & RD CCT bias-corrected & rel $\le 1.6e-14$ {\footnotesize (at forced common $h$; default-$h$ selector differs)} & \textbf{PASS} \\
+\code{06\_rd} & RD CCT bias-corrected & rel $\le 1.7e-14$ {\footnotesize (at forced common $h$; default-$h$ selector differs)} & \textbf{PASS} \\
 \code{07\_scm} & Classical SCM & rel $\le 0.12$ {\footnotesize (SCM non-uniqueness)} & \textit{GAP} \\
 \code{08\_dml} & DML PLR (LinReg learners) & rel $\le 0.0024$ {\footnotesize (fold-split RNG differs)} & \textbf{PASS} \\
 \code{09\_rddensity} & RD density (CJM) & abs $\le 0.0514$ {\footnotesize (bandwidth selector differs; both fail to reject)} & \textit{GAP} \\

[…]
