### 2026-05-29 9e8a0e0
docs(changelog): record JSS reproducibility hardening + 07_scm fixture refresh
Logs the Track A parity-readiness pass under [Unreleased]: R environment
lock + provenance + renv.lock, r-parity.yml CI, validation_report
collect_tests counts, the strictness-tier breakdown, and the 07_scm golden
refresh (2.18e-8, within the SCM non-uniqueness tolerance; verdicts
unchanged at 42 PASS / 10 GAP). Per CLAUDE.md §12, the parity-fixture
numeric refresh is disclosed even though it is not an estimator-output change.

Co-Authored-By: Claude Opus 4.8 (1M context) <noreply@anthropic.com>
