### 2026-05-29 cb422be
Add CausalAgentBench benchmark code under Paper-AgentBench/code
Move the standalone CausalAgentBench Python package (benchmark runner,
tasks, scoring, adapters, tests) into the companion-paper workspace.
Adjust .gitignore to keep Paper-AgentBench/ ignored but force-track the
code/ subtree; its nested .gitignore still excludes run outputs and caches.

Co-Authored-By: Claude Opus 4.8 (1M context) <noreply@anthropic.com>
