+++ Paper-AgentBench/code/causalagentbench/adapters/llm.py
@@ -1,0 +1,260 @@
+"""LLMAdapter — drives a real Claude/GPT agent over one of the three stacks.
+
+This is the cell that runs conditions C1..C6. It is intentionally a thin,
+auditable harness:
+
+1. materialise the task's data to a temp CSV;
+2. build a stack-aware prompt (statspai / pythonic / r);
+3. call the provider at ``temperature=0`` and record token usage (M4);
+4. optionally execute the returned code in a *subprocess sandbox* with a
+   wall-clock timeout, capturing exec success (M3) and a machine-readable
+   final answer;
+5. parse the final method / point estimate / diagnostics / refusal.
+
+Requires the relevant SDK (``anthropic`` / ``openai``) and an API key. The
+production tool-loop for C1/C2 plugs in ``statspai.agent.mcp_server``;
+that MCP wiring is the natural extension point and is documented in the
+README. Here we use a single-turn "write code that prints CAB_RESULT"
+protocol, which is provider-agnostic and fully reproducible.
+"""
+
+from __future__ import annotations
+
+import json
+import os
+import re
+import subprocess
+import sys
+import tempfile
+import time
+from typing import Dict, List, Optional, Tuple
+
+from ..conditions import PYTHONIC_STACK, R_STACK
+from ..schema import AgentAction, Task, Trajectory
+from .base import AgentAdapter
+
+_RESULT_SENTINEL = "CAB_RESULT="
+
+_SYSTEM = (
+    "You are a careful causal-inference analyst. You are given a research "
+    "question and a CSV dataset. Decide on a valid identification strategy, "
+    "estimate the effect, and run the standard diagnostics for that design. "
+    "If the effect is NOT identifiable from the data, say so instead of "
+    "inventing a number.\n\n"
+    "Return EXACTLY ONE fenced ```python code block that, when run, prints a "
+    "single line to stdout of the form:\n"
+    f"  {_RESULT_SENTINEL}" + '{"method": "<estimator>", "point_estimate": <float or null>, '
+    '"se": <float or null>, "diagnostics": ["<check>", ...], "refused": <true|false>}\n'
+    "The CSV path is provided to your code as the variable DATA_PATH "
+    "(already set); load it with pandas. Do not print anything else on that line."
+)
+
+_STACK_HINT = {
+    "statspai": (
+        "Use the StatsPAI package: `import statspai as sp`. Relevant entry "
+        "points include sp.regress, sp.did, sp.callaway_santanna, sp.sun_abraham, "
+        "sp.rdrobust, sp.ivreg, sp.detect_design, sp.recommend, sp.preflight. "
+        "Prefer sp.recommend(...) to pick an estimator when unsure."
+    ),
+    "pythonic": (
+        "Use only these packages (no StatsPAI): " + ", ".join(PYTHONIC_STACK) + "."
+    ),
+    "r": (
+        "Write R, not Python: emit a single ```r block instead. Use these R "
+        "packages: " + ", ".join(R_STACK) + ". Print the CAB_RESULT line via cat()."
+    ),
+}
+
+
+class LLMAdapter(AgentAdapter):
+    def __init__(self, condition, model: Optional[str] = None,
+                 execute: bool = True, timeout_s: float = 120.0, **kw):
+        super().__init__(condition, **kw)
+        self.model = model
+        self.execute = execute
+        self.timeout_s = timeout_s
+
+    # -- public ----------------------------------------------------------
+    def run(self, task: Task, seed: int = 0) -> Trajectory:
+        from .. import tasks as taskmod
+
+        traj = Trajectory()
+        t0 = time.perf_counter()
+        try:
+            df = taskmod.materialize(task, seed_offset=seed)
+            with tempfile.TemporaryDirectory() as tmp:
[…]
