### 2026-06-01 31a2bf6
docs(joss): align README quickstart numbers + paper submodule count
Fix three factual inconsistencies surfaced while cross-checking the
manuscript and README against the package's actual validated output,
ahead of JOSS review (openjournals/joss-reviews#10604):

- paper.md: "80 submodules" -> "81" to match registry_stats.py (1020
  functions / 81 submodules), the README, and the validation dossier
  (paper.md was the only source still saying 80).
- README DiD quickstart: "bootstrap SE ~ 0.004" -> "~ 0.008"; the
  actual sp.aggte(type='simple') output is 0.0078, aligned to R
  did::aggte to ~1e-
[…]
