### 2026-06-09 0acdd1b
docs: reposition StatsPAI as a library (not platform) across README + pyproject
Align README.md / README_CN.md lede and the pyproject description with the
updated GitHub repo description: lead with 'library' rather than 'platform'.
The package is first and foremost an importable Python library; the
agent-native / MCP-server surface is carried by the capability list, not by
inflating the noun. paper.md is intentionally left untouched (its 'conventional
estimator library' contrast is deliberate and it is under JOSS #10604 review).

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
