+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "setuptools.build_meta"
 [project]
 name = "StatsPAI"
 version = "1.17.0"
-description = "Validation-tiered causal inference and econometrics workflows for Python"
+description = "Validation-tiered causal inference and econometrics library for Python"
 readme = "README.md"
 license = "MIT"
 license-files = ["LICENSE"]
