### 2026-08-18 #10604 @rich2355
Hi @brycewang-stanford, since the remaining reviewer becomes unresponsive and the other two reviewers are good with your responses, I will move forward to the next step as JOSS allows this as long as there are two reviewers agreeing to move forward. Thanks for your good work!
