### 2026-08-18 #10604 @rich2355
Hi @brycewang-stanford, please fix all MISSING DOIs and check the SKIP DOIs. If there is no DOI for a given item in SKIP DOIs feel free to keep it unaddressed, but you must correct all MISSING DOIs.

In addition, please give me the DOI for your repo (by archiving the release) and the version number. Make sure that everything in your archive is the same with what shown in the github repo and paper (e.g. author list). Thanks!
