### 2026-08-24 15d0ca0
docs(refs): add 5 verified DOIs, document 3 with none, fix bib brace
JOSS #10604: the editor asked for every MISSING DOI from `check references`
to be corrected. Each candidate was checked against two independent sources
(Crossref REST API + doi.org content negotiation), with Project Euclid
citation metadata and PubMed as a third source where Crossref carried no
page range.

Added (verified):
  aalen1978nonparametric  10.1214/aos/1176344198
  gray1988class           10.1214/aos/1176350951
  schoenfeld1983sample    10.2307/2531021
  dersimonian1986meta     10.1016/0197-2456(86)90046-2
  egger19
[…]
