+++ paper.bib
@@ -4074,7 +4074,7 @@
   number={4},
   pages={772--793},
   year={1994},
-  note={NBER Working Paper 4509. Refs verified via aeaweb.org and ideas.repec.org (RePEc:aea:aecrev:v:84:y:1994:i:4:p:772-93).}
+  note={NBER Working Paper 4509. The published version has no DOI: AEA's Crossref deposits for the American Economic Review begin at vol. 89 (1999), and JSTOR 10.2307/2118030 does not resolve on doi.org or Crossref. 10.3386/w4509 is the 1993 working paper (type: report) and 10.1257/aer.90.5.1397 is the 2000 Reply -- neither is this article. Refs verified 2026-08-24 via Crossref API and doi.org; earlier via aeaweb.org and ideas.repec.org (RePEc:aea:aecrev:v:84:y:1994:i:4:p:772-93).}
 }
 
 @article{angrist1990lifetime,
@@ -4239,7 +4239,7 @@
   number={3},
   pages={450--477},
   year={1998},
-  note={NBER Working Paper 5778. Refs verified via ideas.repec.org (RePEc:aea:aecrev:v:88:y:1998:i:3:p:450-77) and nber.org.}
+  note={NBER Working Paper 5778. The published version has no DOI: AEA's Crossref deposits for the American Economic Review begin at vol. 89 (1999), and JSTOR 10.2307/116844 does not resolve on doi.org or Crossref. 10.3386/w5778 is the 1996 working paper (type: report), not this article. Refs verified 2026-08-24 via Crossref API and doi.org; earlier via ideas.repec.org (RePEc:aea:aecrev:v:88:y:1998:i:3:p:450-77) and nber.org.}
 }
 
 @article{imbens2001estimating,
@@ -4752,7 +4752,8 @@
   author={Westfall, Peter H. and Young, S. Stanley},
   publisher={Wiley-Interscience},
   year={1993},
-  isbn={978-0-471-55761-6}
+  isbn={978-0-471-55761-6},
+  note={No DOI is registered for the book itself (verified 2026-08-24 via Crossref API and doi.org). 10.2307/2533464 is Martin's Biometrics 50(4):1226 book review; Crossref indexes four separate reviews of this book (10.2307/2533464, 10.2307/1270279, 10.2307/2348369, 10.2307/2291234) but no record for the Wiley monograph.}
 }
 
 @book{greene2018econometric,
@@ -5574,7 +5575,8 @@
   volume={6},
   number={3},
   pages={534--545},
-  year={1978}
+  year={1978},
+  doi={10.1214/aos/1176344198}
 }
 
 @article{gray1988class,
@@ -5584,7 +5586,8 @@
   volume={16},
   number={3},
   pages={1141--1154},
-  year={1988}
+  year={1988},
+  doi={10.1214/aos/1176350951}
 }
 
 @article{fine1999proportional,
@@ -5604,7 +5607,8 @@
   volume={39},
   number={2},
   pages={499--503},
-  year={1983}
+  year={1983},
+  doi={10.2307/2531021}
 }
 
 @article{dersimonian1986meta,
@@ -5614,7 +5618,8 @@
   volume={7},
   number={3},
   pages={177--188},
-  year={1986}
+  year={1986},
+  doi={10.1016/0197-2456(86)90046-2}
 }
 
 @article{higgins2002quantifying,
@@ -5634,7 +5639,8 @@
   volume={315},
   number={7109},
   pages={629--634},
-  year={1997}
+  year={1997},
+  doi={10.1136/bmj.315.7109.629}
 }
 
 % ====================================================================
@@ -6018,6 +6024,8 @@
   publisher={BMJ},
   doi={10.1136/jech-2025-225449},
[…]
