### 2026-08-24 7afc9bb
chore(release): align archive metadata title with the JOSS paper
The JOSS #10604 post-review checklist requires the archived release to carry
the same title and author list as the paper. Four places disagreed:

  paper.md          StatsPAI: A Unified, Agent-Native Python Toolkit for
                    Causal Inference and Applied Econometrics  (canonical)
  .zenodo.json      StatsPAI: The Agent-Native Causal Inference &
                    Econometrics Toolkit for Python
  CITATION.cff      StatsPAI: Validation-Tiered Causal Inference and
                    Econometrics Workflows for Python

[…]
