+++ paper.bib
@@ -5516,7 +5516,7 @@
 }
 
 @software{wang2026statspai,
-  title={StatsPAI: The Agent-Native Causal Inference and Econometrics Toolkit for Python},
+  title={StatsPAI: A Unified, Agent-Native Python Toolkit for Causal Inference and Applied Econometrics},
   author={Wang, Biaoyue and Rozelle, Scott},
   year={2026},
   version={1.15.6},
