### 2026-08-24 97dbe1f
chore(release): bump to 1.23.0 and unify the citation surface
Prepares the release the JOSS #10604 post-review checklist asks for. No
release is published here -- no tag, no GitHub Release, no PyPI upload.

Version 1.22.0 -> 1.23.0 in every place that declares it:
  pyproject.toml, src/statspai/__init__.py, CITATION.cff (+ date-released
  2026-08-24), schemas/index.json and its packaged mirror (regenerated with
  scripts/dump_schemas.py, not hand-edited -- mcp_server.py checks the index
  against SERVER_VERSION at runtime), and the CHANGELOG heading, whose
  accumulated [Unreleased] section be
[…]
