+++ paper.bib
@@ -5519,7 +5519,7 @@
   title={StatsPAI: A Unified, Agent-Native Python Toolkit for Causal Inference and Applied Econometrics},
   author={Wang, Biaoyue and Rozelle, Scott},
   year={2026},
-  version={1.15.6},
+  version={1.23.0},
   doi={10.5281/zenodo.19933900},
   url={https://github.com/brycewang-stanford/StatsPAI}
 }
