+++ docs/index.md
@@ -178,10 +178,11 @@ reference — together with this package:
 ```bibtex
 @software{wang2026statspai,
   author  = {Wang, Biaoyue and Rozelle, Scott},
-  title   = {StatsPAI: Validation-Tiered Causal Inference and
-             Econometrics Workflows for Python},
+  title   = {StatsPAI: A Unified, Agent-Native Python Toolkit for
+             Causal Inference and Applied Econometrics},
   year    = {2026},
-  version = {1.21.0},
+  version = {1.23.0},
+  doi     = {10.5281/zenodo.19933900},
   url     = {https://github.com/brycewang-stanford/StatsPAI}
 }
 ```
