+++ src/statspai/CITATION.cff
@@ -16,16 +16,16 @@ type: software
 authors:
   - family-names: "Wang"
     given-names: "Biaoyue"
-    affiliation: "Stanford REAP, Stanford University"
+    affiliation: "Rural Education Action Program, Stanford Center on China's Economy and Institutions, Stanford University, United States; StatsPAI Inc., United States"
     email: "brycew6m@stanford.edu"
     orcid: "https://orcid.org/0000-0002-1828-2208"
   - family-names: "Rozelle"
     given-names: "Scott"
-    affiliation: "Stanford REAP, Stanford University"
+    affiliation: "Rural Education Action Program, Stanford Center on China's Economy and Institutions, Stanford University, United States; StatsPAI Inc., United States"
     email: "rozelle@stanford.edu"
     orcid: "https://orcid.org/0000-0001-5053-0243"
-version: "1.22.0"
-date-released: "2026-08-07"
+version: "1.23.0"
+date-released: "2026-08-24"
 license: MIT
 # Zenodo concept DOI; always resolves to the latest archived version.
 # For a specific-version DOI (e.g. for replication packages), follow the
@@ -63,12 +63,12 @@ keywords:
 #   authors:
 #     - family-names: "Wang"
 #       given-names: "Biaoyue"
-#       affiliation: "Stanford REAP, Stanford University"
+#       affiliation: "Rural Education Action Program, Stanford Center on China's Economy and Institutions, Stanford University, United States; StatsPAI Inc., United States"
 #       email: "brycew6m@stanford.edu"
 #       orcid: "https://orcid.org/0000-0002-1828-2208"
 #     - family-names: "Rozelle"
 #       given-names: "Scott"
-#       affiliation: "Stanford REAP, Stanford University"
+#       affiliation: "Rural Education Action Program, Stanford Center on China's Economy and Institutions, Stanford University, United States; StatsPAI Inc., United States"
 #       email: "rozelle@stanford.edu"
 #       orcid: "https://orcid.org/0000-0001-5053-0243"
 #   title: "StatsPAI: A Unified, Agent-Native Python Toolkit for Causal Inference and Applied Econometrics"
