+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "StatsPAI"
-version = "1.22.0"
+version = "1.23.0"
 description = "Validation-tiered causal inference and econometrics workflows for Python"
 readme = "README.md"
 license = "MIT"
