### 2026-08-24 31d6e13
docs(refs): resolve 23 more DOIs; document the ones that genuinely have none
JOSS #10604. The editor asked us to check the SKIP list from
`check references` -- entries the bot found no DOI for. "Not found by a
title search" is not the same as "does not exist", so all 62 DOI-less
entries were re-checked against Crossref and, where relevant, doi.org
content negotiation, the arXiv API, PubMed, Project Euclid and DataCite.
DOI-less entries drop from 62 to 39, and the 18 references paper.md
actually cites now carry a DOI 17 of 18 times.

Automated title matching is not sufficient here and was not t
[…]
