+++ paper.bib
@@ -52,7 +52,8 @@
   archivePrefix={arXiv},
   primaryClass={econ.EM},
   note={arXiv preprint; no journal publication as of verification},
-  url={https://arxiv.org/abs/2405.04465}
+  url={https://arxiv.org/abs/2405.04465},
+  doi={10.48550/arXiv.2405.04465}
 }
 
 @article{yatchew1999elementary,
@@ -397,7 +398,9 @@
   pages={4595--4607},
   year={2019},
   eprint={1906.00232},
-  archivePrefix={arXiv}
+  archivePrefix={arXiv},
+  doi={10.48550/arXiv.1906.00232},
+  note={NeurIPS proceedings carry no DOI; the DOI above is the arXiv preprint of this paper (title and all three authors verified identical via the arXiv API and doi.org, 2026-08-24).}
 }
 
 @article{gardner2022twostage,
@@ -688,7 +691,8 @@
   author={Battocchi, Keith and Dillon, Eleanor and Hei, Maggie and Lewis, Greg and Oka, Paul and Oprescu, Miruna and Syrgkanis, Vasilis},
   year={2019},
   note={Version 0.x},
-  url={https://github.com/py-why/EconML}
+  url={https://github.com/py-why/EconML},
+  note={No DOI: the package is not registered with Crossref and has no Zenodo or other DataCite archive (both searched 2026-08-24). Cite by URL.}
 }
 
 @misc{chen2020causalml,
@@ -710,7 +714,9 @@
   number={53},
   pages={1--6},
   year={2022},
-  url={https://jmlr.org/papers/v23/21-0862.html}
+  url={https://jmlr.org/papers/v23/21-0862.html},
+  doi={10.48550/arXiv.2104.03220},
+  note={JMLR issues no DOIs; the DOI above is this article's arXiv preprint, whose journal_ref field names exactly this paper ("Journal of Machine Learning Research 23 (53), 2022, 1-6") and whose four authors match. Do not attach 10.18637/jss.v108.i03 -- that is the Journal of Statistical Software paper on the R package, already recorded under bach2024doubleml, and arXiv:2103.09603 is that R paper's preprint (five authors). Verified 2026-08-24 via the arXiv API and doi.org.}
 }
 
 @article{bates2015lme4,
@@ -1206,7 +1212,7 @@
 @article{caudill1995frontier,
   title={Frontier Estimation and Firm-Specific Inefficiency Measures in the Presence of Heteroscedasticity},
   author={Caudill, Steven B. and Ford, Jon M. and Gropper, Daniel M.},
-  journal={Journal of Business &amp; Economic Statistics},
+  journal={Journal of Business \& Economic Statistics},
   year={1995},
   doi={10.1080/07350015.1995.10524583}
 }
@@ -1566,7 +1572,7 @@
 @article{gaure2013multiple,
   title={OLS with multiple high dimensional category variables},
   author={Gaure, Simen},
-  journal={Computational Statistics &amp; Data Analysis},
+  journal={Computational Statistics \& Data Analysis},
   year={2013},
   doi={10.1016/j.csda.2013.03.024}
 }
@@ -1657,7 +1663,7 @@
 @article{hansen1996finite,
   title={Finite-Sample Properties of Some Alternative GMM Estimators},
   author={Hansen, Lars Peter and Heaton, John and Yaron, Amir},
-  journal={Journal of Business &amp; Economic Statistics},
+  journal={Journal of Business \& Economic Statistics},
   year={1996},
   doi={10.1080/07350015.1996.10524656}
 }
@@ -2148,7 +2154,7 @@
 @article{nevo2000practitioner,
[…]
