### 2026-08-24 #10604 @brycewang-stanford
Hi @rich2355, thanks for your patience, and thanks again to @FATelarico and
@Carol-seven. I have gone through both the MISSING and the SKIP lists.

**How I checked.** For each candidate DOI I used two independent sources, the
Crossref REST API and doi.org content negotiation, and compared the author,
year, venue and page range that came back against what the entry claims. I did
not rely on title matching alone. Where Crossref had no page data or no record
at all, I used a third source: Project Euclid for the two Annals of Statistics
items, PubMed for Schoenfeld (1983), the arXiv API for preprints, DataCite for
software, and DBLP for conference venues.

**Where things stand.** Entries with a DOI went from 607 to 638. Entries without
one went from 67 to 35. Of the references the paper itself cites, all but one
now have a DOI; three were missing before. Each of the 35 remaining entries
records in its `note` field why no DOI exists, so a reader can see that the
absence was checked.

### MISSING list

Five were confirmed and added:

| Reference | DOI |
| --- | --- |
| Aalen (1978) | `10.1214/aos/1176344198` |
| Gray (1988) | `10.1214/aos/1176350951` |
| Schoenfeld (1983) | `10.2307/2531021` |
| DerSimonian & Laird (1986) | `10.1016/0197-2456(86)90046-2` |
| Egger et al. (1997) | `10.1136/bmj.315.7109.629` |

I have not added the other three, because each suggested DOI resolves to a
different work.

**Westfall & Young (1993).** `10.2307/2533464` is M. A. Martin's review of the
book in Biometrics 50(4):1226. Crossref indexes four separate reviews of this
monograph and has no record for the Wiley volume itself. The entry keeps its
ISBN.

**Card & Krueger (1994), AER 84(4):772-793.** `10.3386/w4509` is the 1993 NBER
working paper. Crossref types it as a report and titles it "Fast Food" rather
than "Fast-Food". The published article has no DOI at all: a Crossref query on
the AER's ISSN returns zero records before 1999, since AEA's deposits for that
journal begin at volume 89, and the JSTOR identifier `10.2307/2118030` does not
resolve on doi.org or Crossref. For completeness, `10.1257/aer.90.5.1397` is the
2000 Reply, which is a different paper.

**Angrist & Evans (1998), AER 88(3):450-477.** `10.3386/w5778` is the 1996
working paper. The same AEA cutoff applies, and `10.2307/116844` does not
resolve either.

### SKIP list

Since "not found by a title search" is not the same as "does not exist", I
checked all 58 entries individually. Twenty-six of them now have a verified DOI,
including two that the paper cites:

Cunningham (2021), Causal Inference: The Mixtape, is `10.2307/j.ctv1c29t27`.
Yale University Press registers the book under its main title, which is why a
title search does not find it.

Bach et al. (2022), the DoubleML Python paper, is `10.48550/arXiv.2104.03220`.
JMLR issues no DOIs of its own. The `journal_ref` field on that arXiv record
names this exact article ("Journal of Machine Learning Research 23 (53), 2022,
1-6") and the four authors match. Its sibling `arXiv:2103.09603` has five
authors and belongs to the JSS paper on the R package, which is a separate
entry.

One entry was replaced rather than annotated. The 2015 MPRA working paper
"Synthetic Control Methods: Never Use All Pre-Intervention Outcomes Together
with Covariates" is now cited as its version of record, Kaul et al. (2022),
Journal of Business & Economic Statistics 40(3):1362-1376,
`10.1080/07350015.2021.1930012`. The published article was retitled, which is
why a title search never connects the two. One unused entry was dropped.

The one cited reference still without a DOI is the EconML package, which is
registered with neither Crossref nor DataCite.

The 35 remaining entries fall into five groups: AER volumes before 89 (1999),
JMLR papers and conference proceedings, textbooks their publishers never
deposited, Stata SSC modules, and self-published working papers. Where a title
search returned a plausible but wrong candidate, I named it in the entry so that
it does not get added later. Three examples are the three JSTOR reviews of
Breslow & Day, a 2024 AJE article about SWIGs by different authors, and the CRAN
package `bacondecomp` by Flack and Jee, whose title is close enough to the Stata
SSC module of the same name to fool a matcher.

### One problem the reference check could not have found

A DOI that resolves is not necessarily the right DOI, so I also compared all 635
DOIs then in the file against the metadata each one actually resolves to. This
turned up one bad record. `chickering2002optimal` carried
`10.1162/153244303321897717`, which `check references` reported as OK because
doi.org returns a 302 for it. The redirect goes to Crossref's "Defunct DOI"
notice. Crossref files that record under "CrossRef Listing of Deleted DOIs" with
publisher "Unmaintained records". MIT Press deleted the DOI and registered no
replacement, and some legacy aggregators still carry the string with broken
metadata. This is the only entry in the file where a DOI was registered and then
withdrawn, as opposed to never being assigned, and its note now says so.

The same pass found four metadata errors, all now fixed:

1. Dahabreh et al. (2020) had a corrupted surname for Steingrimsson.
2. The TARGET Statement entry listed Hernán first, but Cashin is first author on
   both the JAMA and the BMJ version.
3. Tran et al. was filed under the 40th ICML (2023). The arXiv record is
   annotated "Accepted into ICML 2024" and DBLP indexes it under the 41st.
4. The Greathouse et al. preprint had a truncated title and a single author. Its
   title page reads "Greed is Good: Estimating Forward Difference-in-Differences
   in Stata" and lists three authors.

All of this is on `main`. I re-resolved every DOI in the file after making these
changes and did not find a broken one. The only dead DOI in the whole pass was
the Chickering record described above, which is now gone.

Two things are worth flagging if you do re-run the check. First, doi.org drops
connections intermittently when it is queried quickly, so a DOI can come back
unresolvable on one pass and fine on the next. On my last run seventeen of the
638 behaved that way, and every one of them resolved on a serial retry. Second,
27 entries cite a conference or journal version but carry the arXiv DOI of the
same paper, because NeurIPS, ICML, PMLR and JMLR issue no DOIs of their own. In
one of those cases, Hartford et al. on Deep IV, the preprint was posted under an
earlier title before being renamed for ICML, and the entry's note records that.

I will follow up shortly with the archive DOI and the version number.
