### 2026-08-24 #10604 @brycewang-stanford
Hi @rich2355, the release is done.

**Version:** 1.23.0
**Archive DOI:** [10.5281/zenodo.22085759](https://doi.org/10.5281/zenodo.22085759)

The archive metadata matches the paper: same title, both authors with their
ORCIDs (0000-0002-1828-2208 and 0000-0001-5053-0243), the same two
affiliations, and the MIT license. The GitHub release is
https://github.com/brycewang-stanford/StatsPAI/releases/tag/v1.23.0.

Thanks again to you and to @FATelarico and @Carol-seven.
