### 2026-08-24 d36e4eb
docs(claude): record the JOSS archive, and correct when it is cut
The note said to hold off on a GitHub Release until the paper is accepted.
That is backwards: JOSS asks for the archive DOI *before* recommend-accept,
and the editor's post-review checklist on 2026-08-18 asked for it outright.
A future session reading the old wording would have blocked a release the
editor was waiting on.

Rewritten around what actually triggers what: only "Publish a GitHub Release"
starts the Zenodo archive, and it fires publish-prod at the same time, so one
click is two irreversible actions. Added the pre-flig
[…]
