### 2026-08-31 #10604 @brycewang-stanford
Hi @rich2355, thanks for the follow-up. You did not miss it, but it was buried:
these three are covered in my Aug 24 comment under the "MISSING list" heading
(https://github.com/openjournals/joss-reviews/issues/10604#issuecomment-5397120629).
Condensed here, and re-verified today via the Crossref API:

1. **Card & Krueger (1994), AER 84(4):772-793.** `10.3386/w4509` is Crossref
   `type: report` (NBER Working Paper, 1993), not the journal article. The
   article itself has no DOI: AEA's Crossref deposits for the *American Economic
   Review* start at vol. 89 (1999), so nothing from 1994 is registered.

2. **Angrist & Evans (1998), AER 88(3):450-477.** `10.3386/w5778` is likewise
   `type: report` (NBER Working Paper, 1996). The same AEA cutoff applies, since
   the 1998 article also predates Crossref coverage.

3. **Westfall & Young (1993), *Resampling-Based Multiple Testing* (Wiley).**
   `10.2307/2533464` is a one-page book *review* in *Biometrics* 50(4):1226 by
   M. A. Martin, not the book. No DOI is registered for the monograph, so the
   entry carries its ISBN instead.

Adding these would attach a working paper's or a book review's identifier to a
different work. All three really belong in the **SKIP category**, since no DOI
exists, and they landed under MISSING only because a title search surfaced a
different publication. Happy to revisit if the team sees it differently.

Separately, the issue header still shows **Version:** v 0.6.0 and **Archive:**
Pending. The archive DOI should be 10.5281/zenodo.22085759 and the version
should be v1.23.0 (both posted Aug 24). Whenever you are ready, those may need
`@editorialbot set 10.5281/zenodo.22085759 as archive` and `@editorialbot set
v1.23.0 as version`.
