### 2026-09-02 #10604 @rich2355
Hi @brycewang-stanford, I have no problem for your explanation so I will move on. You may still need to revisit it if the AEiC in this track has a different opinion on your points. Thanks!
