### 2026-09-03 #10604 @crvernon
@brycewang-stanford  - I just need you to address the following before I accept this submission for publication.

In your paper:

- [ ] LINES 18-23: Define “LLM” at first use as “large language model (LLM)"
- [ ] LINES 108-109: “…against a reference R implementation, the large majority also against Stata” is grammatically off. Consider: “…against a reference R implementation, with the large majority also checked against Stata"
- [ ] LINES 116-120: The Card, LaLonde/Dehejia-Wahba, Lee, and Abadie-Diamond-Hainmueller datasets/studies are named as source studies but are not all accompanied by conventional citations. Add appropriate in-text citations and corresponding bibliography entries where the datasets derive from published research.
- [ ]  LINE 130: “(Callaway-Sant’Anna)” is not a standard author citation. Replace it with a conventional citation such as “(Callaway & Sant’Anna, 2021)” if this is referring to the cited paper.
- [ ] LINES 140-143: Please do not cite in prep. work - e.g., “(Wang, Zhang, and Hou, in preparation)”
- [ ] LINES 177-179: The Bach et al. (2022) entry cites the Journal of Machine Learning Research article but links to an arXiv DOI. The bibliographic record and identifier should refer to the same publication. Use the appropriate JMLR URL/identifier, or cite the arXiv version explicitly instead.

Let me know when these have been addressed.  Thank you.
