### 2026-09-03 0e9786d
docs(paper): address editorial citation checks
refs verified via Crossref, jmlr.org, doi.org
