+++ paper.bib
@@ -717,8 +717,7 @@
   pages={1--6},
   year={2022},
   url={https://jmlr.org/papers/v23/21-0862.html},
-  doi={10.48550/arXiv.2104.03220},
-  note={JMLR issues no DOIs; the DOI above is this article's arXiv preprint, whose journal_ref field names exactly this paper ("Journal of Machine Learning Research 23 (53), 2022, 1-6") and whose four authors match. Do not attach 10.18637/jss.v108.i03 -- that is the Journal of Statistical Software paper on the R package, already recorded under bach2024doubleml, and arXiv:2103.09603 is that R paper's preprint (five authors). Verified 2026-08-24 via the arXiv API and doi.org.}
+  note={JMLR does not register DOIs, so this record carries the journal's canonical URL and no DOI; the arXiv preprint (arXiv:2104.03220) is deliberately not attached as a DOI so that the record and identifier refer to the same publication. Do not attach 10.18637/jss.v108.i03 -- that is the Journal of Statistical Software paper on the R package, recorded under bach2024doubleml. Verified 2026-08-24 via the arXiv API and doi.org; JMLR URL and metadata re-verified 2026-09-04 via jmlr.org.}
 }
 
 @article{bates2015lme4,
@@ -1975,9 +1974,12 @@
 }
 
 @article{lee2008randomized,
-  title={Randomized experiments from non-random selection in U.S. House elections},
+  title={Randomized experiments from non-random selection in {U.S.} House elections},
   author={Lee, David S.},
   journal={Journal of Econometrics},
+  volume={142},
+  number={2},
+  pages={675--697},
   year={2008},
   doi={10.1016/j.jeconom.2007.05.004}
 }
@@ -2768,6 +2770,9 @@
   title={Causal Effects in Nonexperimental Studies: Reevaluating the Evaluation of Training Programs},
   author={Dehejia, Rajeev H. and Wahba, Sadek},
   journal={Journal of the American Statistical Association},
+  volume={94},
+  number={448},
+  pages={1053--1062},
   year={1999},
   doi={10.1080/01621459.1999.10473858}
 }
