### 2026-09-03 d02ba2b
docs(paper): tighten impact statement
