### 2026-09-03 #10604 @brycewang-stanford
@crvernon Thank you for the careful read. All six items are addressed on main (commits 0e9786d6 and d02ba2b2):

- LINES 18-23: "LLM" is now defined at first use as "large language model (LLM)".
- LINES 108-109: Reworded as suggested: "...against a reference R implementation, with the large majority also checked against Stata."
- LINES 116-120: The Card, LaLonde, Dehejia & Wahba, Lee, and Abadie, Diamond & Hainmueller source studies now carry conventional in-text citations, each pointing to a bibliography entry with a resolving DOI (Card 1993 is the NBER working paper, DOI 10.3386/w4483; LaLonde 1986 predates AER's Crossref deposits and is cited by volume and pages). While adding these I also completed the volume/issue/page fields for the Dehejia & Wahba (1999) and Lee (2008) entries.
- LINE 130: "(Callaway-Sant'Anna)" is replaced by a conventional citation to Callaway & Sant'Anna (2021).
- LINES 140-143: The in-preparation citation is removed. The paragraph keeps only a descriptive statement that the package is in use in an ongoing REAP working paper; I also tightened the impact paragraph after removing the in-preparation citation.
- LINES 177-179: The Bach et al. (2022) entry now carries only the JMLR record (volume 23, paper 53, pages 1-6) and its canonical JMLR URL; the arXiv DOI is removed so the record and identifier refer to the same publication.

References were re-verified against Crossref, doi.org, and jmlr.org. No changes to the software, version, or Zenodo archive; the regenerated proof follows below.
