### 2026-09-03 b841c31
docs(paper): clarify validation resources
