### 2026-09-03 #10604 @brycewang-stanford
@crvernon One small layout-only follow-up (commit b841c313): the citations added above pushed the last line of the reference list onto an otherwise empty sixth page, so I trimmed two redundant words from one sentence to bring the paper back to five pages. No content change; regenerating the proof below.
