### 2026-09-03 #10604 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
